system: Partial magnifier uses clip rect w/ rounded corners.
It previously used a mask, which worked until recently see [1] for more
details. But using a clip rect w/ rounded corners is more efficient
anyways, so use that.

The mask was also being used for the border, but that still works
because that layer has no backdground effects, and this probably needs
to get merged to M79.

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=1030432

Test: manual
Bug: 1023009
Change-Id: I1e910ea280701989298871f31dfdff37c09ad1c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1949293
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721315}