Maintain focus when items in a network-list are updated.
Previously, focus would jump to the first element when
the underlying network list was updated.  This fix
attempts to maintain focus on a particular network if it
is still present, otherwise it tries to stay on the same
index, and finally falls back to the first element.

Fixed: 1005934
Change-Id: I9a7e2307bc604f8ba08b659e22f80ba88ad66156
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1920188
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Jon Mann <jonmann@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721367}