Extend duration of animations/play_state test to 1s
100ms is likely not enough time to ensure that the animation can be
paused on all test configurations. Failures are more prevalent when run
in a composited test environment, but also appear when run outside of
a virtual test suite.  To avoid needing to wait an extended period of
time for the finished state.  The animation duration is reduced back to
100ms once the pause state has been detected.

Bug: 986019
Change-Id: Ib1a07cb495ef8b9f40ec4032d835ce758d3a5f70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1949111
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721469}