[Autofill Assistant] Refactor ownership of user_data and collect_user_data_options
This CL moves ownership of user_data to the Controller, and ensures that we can
resume the CollectUserDataAction (by keeping the previous user data/model).
In order to achieve this, the CL also moves ownership of the collect user data
options to the action.
When a new proto is sent by a subsequent CollectUserDataAction, the proto
options take precedence over any existing user_data values. This allows the
backend to control whether it wants to resume a previous partial action or
whether to start one from scratch.
There are some outstanding items to resolve in future CLs.

Bug: b/140019232
Change-Id: Ia177573c6ebab4cc674909f58141a8d6977350ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1949786
Commit-Queue: Marian Fechete <marianfe@google.com>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#721477}