Add logic to enterprise policy to allow ambient authentication for profile type.
Ambient authentication in private modes can currently be toggled with
feature flag. The default value of the flag is to allow for all profile
types. This default value would be set to disallowed for OTR profile
types in the future. A policy is introduced for enterprise admins
who still want the ambient authentication for the OTR profiles.

This CL adds the logic to allow ambient authentication via policy
and the existing feature flag. As a consequence for now, if either
the policy or feature flag allows ambient authentication then the
resulting behanviour is that we allow ambient authentication.

This CL also adds browser tests.

Bug: 458508
Change-Id: I488cf40ec4c178739d441fe76fb15a3e3450cb56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1944493
Commit-Queue: Rohit Agarwal <roagarwal@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721494}