Reflect sync errors on Android if missing keys
This patch is similar in spirit to the desktop equivalent changes in
https://chromium-review.googlesource.com/c/chromium/src/+/1924572.

The overall idea is that, if encryption keys are missing, the UI needs
to display this state as an error, and the user should be able to go
through a reauth flow to retrieve the keys.

The actual logic to trigger the key retrieval process is deferred to
future patches and reflected in TODOs.

Bug: 1012659
Change-Id: I0679e47f8c783fdfd2953989205964165b0f8147
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1930796
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721571}