viz: Add Shared Overlay Validator Base Class for Ozone and Android
Both Ozone's and Android's OverlayCandidateValidator implementation
goes through a list of strategies. This is different from the Mac and
Windows implementation, which instead answers simple true or false
question. This CL adds a OverlayCandidateValidatorStrategy class as a
base class to be shared by Ozone and Android.

This is moving towards separating out different platforms' validator
implementations so they can be moved into OverlayProcess. The final
goal is for OverlayProcessor to have different subclass implementation
for each platform.

This is CL 3/? of overlay processor refactor.

R=rjkroege
TBR=kylechar@chromium.org

Bug: 963597
Change-Id: I58e17a1bcff7045963ff6026e859310f37972f23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1756474
Commit-Queue: weiliangc <weiliangc@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721659}