AppCache: Add parser version, scope to AppCache classes
This is the first in a series of changes that introduce a manifest
scope in AppCache.

Prior to these changes, a manifest can specify CHROMIUM-INTERCEPT
and FALLBACK sections with resources that exist outside of the scope
of the manifest's URL.

After these changes, a manifest by default will have a scope that
is the URL of the manifest stripped to the right-most forward-slash.
Any resources specified in the CHROMIUM-INTERCEPT or FALLBACK
sections will need to reside within this scope.

This CL adds a manifest_parser_version which is currently 0 and
plumbs that through AppCache, AppCacheUpdateJob, and
AppCacheManifestParser.  manifest_parser_version is kept in 2
members: cached_manifest_parser_version_ (a const) and
fetched_manifest_parser_version_.  The manifest parser currently
gets cached_manifest_parser_version_ from the group which
initializes to 0 and always results in a
fetched_manifest_parser_version_ of 0, so this change is effectively a
no-op.

This CL also adds a manifest_scope which is currently set to the
manifest URL's root path.  The manifest_scope is retrieved in 3
stages, and there's a member for each stage: cached_manifest_scope_
and fetched_manifest_scope_.  Initially, all of these are set to the
manifest URL's empty path to make this a no-op change.  Eventually,
this will be tightened down to a stricter default of the manifest
URL without a filename.

Bug: 999001
Change-Id: I723c594e9e5abec6608c4a9eb0cd77c8434fd46e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1934316
Reviewed-by: Jonathan Metzman <metzman@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Chase Phillips <cmp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721766}