Fix & Re-enable DOMStorageBrowserTest.DataPersists
There is really no reason to expect the pre-test browser shutdown not to
race with storage commits. This changes the pre-test to effectively wait
for a commit before exiting (by queueing a usage request and waiting
for the reply), ensuring that the database is opened and pending
operations are scheduled before the browser terminates and the actual
test goes on to verify the persistence of data.

Fixed: 1027940
Change-Id: I58eac4ec5fc9d547f8e10f5f2d12cbf41c6d12ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1951765
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721800}