Add UMA histograms for proactive suggestions view impressions.
This CL adds UMA histograms for proactive suggestions views that are
shown inline in Assistant UI. Proactive suggestions views are rendered
from HTML, so we use deep links to bridge the web impression event over
to the native side for UMA logging.

The new histograms are:
- Assistant.ProactiveSuggestions.ViewImpression
- Assistant.ProactiveSuggestions.ViewImpression.ByCategory
- Assistant.ProactiveSuggestions.ViewImpression.ByVeId

The latter two histograms allow for in depth analysis of impression
events while the former histogram is recorded for all impressions.

Bug: b:144875278
Change-Id: I42cb9e784e0e495c01672177d66e872f3d66d0de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1947548
Reviewed-by: Tao Wu <wutao@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: David Black <dmblack@google.com>
Cr-Commit-Position: refs/heads/master@{#721811}