[Extensions] Add code to migrate to a new withholding pref key
This change adds a new pref key to represent if permission withholding
is enabled on an extension and code to migrate all existing extension
prefs to use the new key if it doesn't exist for an extension on
startup. It also changes the logic for setting the pref on
installation to only be set for extensions which can have permissions
withheld from them.

Bug: 984069
Change-Id: Ibbde3a7bdb0000e71f3d18f451ac9c8792714330
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1911479
Commit-Queue: Tim Judkins <tjudkins@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721823}