Extensions: Plumb navigation ID as part of WillCreateURLLoaderFactory.
This CL plumbs the navigation IDs for navigation network requests through to the
WebRequestInfo objects at the extensions layer.

This will be used in a subsequent CL to correct the behavior of
declarativeNetRequest.setActionCountAsBadgeText for main-frame requests.
Navigation IDs will make it possible to correlate the network request with
corresponding navigation event for that request.

BUG=992251
TBR=changwan@chromium.org

Change-Id: I2c11e421e6d6ad3c0dcfae7a56e117132a2793be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1923611
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721828}