Move HidConnectionImpl constructor and destructor to private
This is a follow-up CL for the review comment from [1].
HidConnectionImpl has Create method with hiding the bare call to new
HidConnectionImpl. The constructor is used to access the created
object from tests.
This CL moves the constructor and the destructor to private and adds
friend class for HidConnectionImplTest.

[1] https://crrev.com/c/1941447/2/services/device/hid/hid_connection_impl.cc#48

Change-Id: Ic9ae9883ea0eb13b9616ecb477920220e370a066
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1947619
Commit-Queue: Julie Kim <jkim@igalia.com>
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721842}