Move Allow Chrome Sign In to Sync and Google Services from Privacy
Along with the elevation of the Privacy card out of advanced, the toggle
for allowing Chrome Sign In is being moved out of Privacy and into the
Sync and Google Services section.

This toggle is disabled during the Sync setup flow if the user elects
to go to settings during Sync setup.

Additionally, the "Turn On Sync.." button is disabled when Chrome Sign
In is disabled. This removes the possibility of enabling sync before
restarting the browser after disabling sign in.

Every change in behavior is behind the privacySettingsRedesignEnabled
finch flag.

Bug: 1014143
Change-Id: I4f2f11598ac22823310623d6bd079ece27bfe3b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1943174
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/master@{#721899}