Clear style recalc root removed from flat tree.
We are not allowed to have style recalc roots outside the flat tree with
FlatTreeStyleRecalc enabled. We already made sure we don't add any, but
did not clear any when the were removed outside the flat tree as a
result of a slot re-assignment or by attaching a shadow root.

Update the style recalc root when we call Node::RemovedFromFlatTree().

We now clear the style dirty bits as part of DetachLayoutTree() when
not performing a reattach to be able to figure out if the recalc root
was part of the subtree removed from the flat tree as the is no flag
similar to isConnected() that we can use to check this. Clearing the
dirty flags should be correct in this case anyway.

Bug: 1027829
Change-Id: Ibd30ba5c16c9ccb8390e5bfa5935649d1997425a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1949788
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722022}