cbui passwords: remove PasswordPendingView CreateFootnoteView
This change removes PasswordPendingView's CreateFootnoteView override,
in favor of creating the footnote view explicitly in the constructor.
This change is only correct if model()->ShouldShowFooter() returns the
same thing during PasswordPendingView's constructor as it does when
PasswordPendingView's widget is constructed. I *believe* that this is
the case.

There is one additional change here, which is the removal of the check
against sign_in_promo_ in CreateFooterView(). This check was dead code,
since CreateFootnoteView was originally only called during widget init,
and the only way sign_in_promo_ can become non-null was via:

  PasswordPendingView::Accept calling
  PasswordPendingView::ReplaceWithPromo

and Accept obviously can't be called before the dialog's widget has
been constructed.

At first glance it appears that PasswordPendingView::ReplaceWithPromo's
call to RemoveAllChildViews() would delete the footnote view, but
actually this does not happen, because the footnote view is a child of
the DialogClientView, not of the dialog's contents view (which this
class is). Therefore the logic in ReplaceWithPromo isn't affected by
this change at all.

Bug: 1011446
Change-Id: I85cdc7c7b3b411029b8fc5e303481e8d3c45b82f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1951484
Reviewed-by: Peter Boström <pbos@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722027}