Move Password Leak detection setting from Google Services to Privacy
Along with the elevation of the Privacy card out of advanced, the toggle
for enabling Password Leak detection is being moved from the Sync and
Other Google services page into the privacy page.

This aligns the location of the Passwork Leak detection toggle with the
Safe Browsing settings already moved to the Privacy page.

Every change in behavior is behind the privacySettingsRedesignEnabled
finch flag.

Bug: 1014143
Change-Id: Icd604d38729cdae113c116c77079cad9b874cd3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1946470
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722065}