AW UMA: log uma_enabled_date
This logs the uma_enabled_date field in WebView UMA records, by
persisting the associated pref, and clears the pref if the user opts out
of UMA.

This also adds temporary code to populate the uma_enabled_date if it's
not already set. This is necessary because we previously persisted the
client_id without persisting uma_enabled_date, which put WebView prefs
in an unexpected state (the metrics component assumes both are set
consistently).

This modifies existing unittests and integration tests to assert the
uma_enabled_date is there when expected, and adds another unittest to
verify the temporary code.

Bug: 995544
Test: run_android_webview_unittests --gtest_filter=AwMetricsServiceClientTest.*
Test: run_webview_instrumentation_test_apk -f AwMetricsIntegrationTest.*
Change-Id: I70f48d09940427b8480299d7cf0f72bc97fe63fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1947948
Reviewed-by: Tao Bai <michaelbai@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722104}