Move child process UKM recording off Service Manager
This essentially eliminates the Metrics service in favor of having
Content process hosts directly bind incoming UkmRecorderInterface
requests to a metrics::UkmRecorderInterface on the IO thread.

This is equivalent to what was happening before the change, just
without a hop through the Service Manager.

Bug: 977637
Change-Id: Ibbee1cb6b5067356297931cebd6d3ebb035bf18b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1941204
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722268}