De-virtualize SecurityContext and stop using it as a base class
SecurityContext is nominally 1:1 with ExecutionContext in LocalFrames
and workers. However, each ExecutionContext subclass also has to
subclass SecurityContext. Remove the couple of virtual functions from
it, and make it an owned member of ExecutionContext (except for
RemoteFrames, where the RemoteSecurityContext will continue to be
owned by the RemoteFrame).

Added a few helpers to ExecutionContext for commonly called
SecurityContext functions. The others will require another step of
indirection.

Change-Id: I608982b553cd8162eae7085680029e0f3b1f2fbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1929468
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722320}