Handle corner case with AddToExistingGroup
See the inline comment and the attached bug for the case that this is handling. I suspect this would be extremely rare but possible, especially as we extend support to extensions post-MVP.

I opted with no-oping in this case, since I think any side effects are likely to be unintended.

Bug: 922736
Change-Id: Ie295416563c2338c0b3638ae630499814d2c8e69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1954005
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Commit-Queue: Connie Wan <connily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722342}