Suppress TopControlsSlideController for some tab-dragging tests
I got notified about some tab-dragging tests are flaky, and the
reason turned out to be TopControlsSlideController; this allows
to slide out the top-controls (i.e. tabstrip) on tablet mode on
certain situation. However, it looks like, on entering into the
tablet-mode, this controller is activated, and sets 'shown ratio'
to 0 (meaning invisible), and then sets to 1 immediately.

Sometimes the test scenario runs between them, which means the
top controls are positioned out of the screen, and therefore
tab-dragging hasn't started, and the test crashes.

This CL adds a method to BrowserView for testing which suppresses
the TopControlsSlideController itself. With that, the top controls
will never be hidden.

Bug: 1031075
Test: release build of interactive_ui_tests, passes all 20 runs
Change-Id: Ibdd5ac622533692802ecc17f3364c2e8cb20e4ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1954362
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722345}