Use base::FastHash() in //chrome (part 1)
base::Hash() is deprecated and is being replaced by base::FastHash().
The output of base::FastHash() is only fixed within the lifetime of a
single process, so if the code depends on the mapping of inputs to hash
outputs not changing over time, base::PersistentHash() must be used
instead.

Bug: 1025358
Change-Id: Idfa61dfb3162d8d5ff40e7ee7a18cf1813570b4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1952554
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722353}