Move NodeRenderingData to Oilpan
This CL moves NodeRenderingData to Oilpan and also removes
NodeRareDataBase.
NodeRareDataBase is never created directly and it was not an Oilpan
object. But its only child (i.e. NodeRareData) was on heap. Hence this
CL moves the reference of NodeRenderingData to NodeRareData and removes
NodeRareDataBase.

https://docs.google.com/document/d/1AJ1XCtiYbaw64KhPX_a_TiNDOqeYV-MO30I1C-8VzPk/edit?pli=1#heading=h.3qsfdd8uwwk3

Bug: 1030176
Change-Id: Iefeb9a0544aa781f060d38c35729a24c9b2c6450
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1948365
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Navid Zolghadr <nzolghadr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722387}