BackForwardCache: Add metrics for all sites (BackForwardCache.AllSites.*)
This CL adds new metrics for all websites, besides the existing metrics.
The existing metrics are recorded only when the domain is allowed in the
experiments, while the new metrics are recorded for all the websites.

The new metrics are useful to see e.g., the tendency of the reasons why
the page is not cached if we enable back-forward cache in all the sites,
and such information will help us to prioritize our works to enable
back-forward cache in all websites.

Bug: 1004676
Change-Id: I30cf343bb685a0e462364bb2b6d6b03302032fe3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1950731
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722427}