[Paint Previews] Enable Main Frame Capture
This CL serves to enable the ability to capture paint previews of the
main frame. Support for OOP subframe capture is not yet available. The
code for this CL is behind the buildflag ENABLE_PAINT_PREVIEW to avoid
a moderate binary size regression until the remainder of the feature
is ready.

The buildflag is setup such that tests will still run on all bots not
doing official builds.

Bug: 1014955
Change-Id: Ic16cad5a8a72af84033f00847ce2ed58e77d36d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1952392
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Mehran Mahmoudi <mahmoudi@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722463}