[Android][Mfill] Use LargeIconBridge instead of FaviconService
This CL replaces the favicon service that tried to fetch icons natively
in favor of the LargeIconBridge which provides higher quality icons if
available and simplifies creating fallbacks for non-existent or low-
resolution icons.
This causes favicons to be in line with what NTP and history use and
improves contrast in dark mode (since the black globe is gone).

In order to do that, the accessory sheet data changed in:
* always having an origin for a password
* using a IsPslMatch indicating whether the icon is a PSL match

Screenshots are in the linked bug.

Bug: 1028590
Change-Id: If27065e5695ae4514465d26f668c5315d8faaea8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1953823
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722464}