Remove embeddedExtensionOptions extension permission
<extensionoptions> is used to display extension options pages embedded
in chrome://extensions. There currently exists an extension permission
"embeddedExtensionOptions" which would allow an extension that requested
it to also use <extensionoptions>. This permission is restricted to
trunk. This was last changed in 2014. According to issue 420926, there
is no longer a need for this permission as it is enough to whitelist
usage for WebUI.

We remove this permission and the related tests that were specific to
the use of <extensionoptions> by other extensions. Note that we still
keep the <extensionoptions> tests related to the extension management
WebUI as that is separate from the permission. The tests that were
implemented in terms of extensions and are still relevant to the
extensions WebUI have been reimplemented in terms of WebUI.

Also note that there is a switch with the same name
(--embedded-extension-options), but this does not relate to the
permission. It controls the default value of |open_in_tab| if not
specified in the extension manifest.

Bug: 420926, 415949
Change-Id: I91081ca3056fdd5bb1e14e5852ec067ed26f455a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1864519
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722489}