Pre-compute indices for text_ and layout_text_ conversion
This CL is adding a pre-computed array of indices between
text_ and layout_text_. The array contains pair of indices:
  <text_ offset, layout_text_ offset>
Both sequence are ordered (e.g. text_ offsets and layout_offsets_).

Previously, a scan 0(n) was made to find the offset (e.g. rank) of the
codepoint within the text. There are an assumption that both
codepoints had the same offset. Unfortunately, this is not working
correctly with text eliding. The current implementation is 0(lg n).

This CL is a step towards having a correct support for text eliding
and removing the display_text_ phase by integrating it into the
layout_text_. The goal is to allow a mismatch between the amount of
codepoint in text_ and display_text_.

Bug: 1025561
Change-Id: I76a2343b46589f5e9260da0cc60ff6af4ccd8e81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1945277
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722566}