Provide a GUID for existing network configurations
Prior to using mojo for network configuration, when creating a new
configuration for a visible network, the GUID of the visible network
would be provided so that it was preserved.

Some UI depends on this (e.g. OOBE flow) and generally it is less
confusing for logging, debugging, etc.

This restores that behavior by providing an optional guid property
in the mojo ConfigProperties and translating it to ONC.

Bug: 1009406
Change-Id: Ia6fc22afa16b514a767dc97274bc216f28112ab8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1951558
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722577}