[iOS] Create OverlayRequest config for save passwords infobar banner.
This config will be used to trigger the display of the banner UI for
the save passwords infobar.

This CL also introduces CreateAuxilliaryData to the OverlayUserData
template so that OverlayUserData can be shared amongst OverlayRequest
configs.

The SavePasswordsOverlayRequestConfig adds an InfobarOverlayData to
its OverlayRequest exposing the InfoBar pointer.  This can also be used
by other infobar OverlayRequest configs so that their InfoBars can
be accessed as request->GetConfig<InfobarOverlayData>()->infobar().

In addition, this CL also adds IOSChromeSavePasswordInfoBarDelegate::
FromInfobarDelegate(), which allows downcasting to the appropriate type.



Bug: 1030357
Change-Id: I3b89d8516b8b25d8e32576aff200fa10d92a61f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1951130
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722590}