[DeviceSync v2] Block host verification on metadata decryption
In this CL, we do not consider a host verified until it has a public key
and beacon seeds, data required for communication between the
mult-device host and client.

This CL should only be relevant after v1 DeviceSync is deprecated,
when a public key and beacon seeds are not available until v2 DeviceSync
host device data has been decrypted. When only v1 DeviceSync is running
or when v1 and v2 DeviceSync are running in parallel, we guarantee that
the eligible host devices always have a decrypted public key and beacon
seeds, as long as that data was properly sent by the CryptAuth backend.

See http://go/cros-devicesync-v2-remote-device-id for more details.

Bug: 951969, 1019206
Change-Id: I454c79ddfae3cbf5f3bde602384e13d62ea1ffd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1935229
Commit-Queue: Josh Nohle <nohle@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722593}