split view: Fix behavior after going from tablet to portrait clamshell
The present CL addresses two blatantly broken things that together form
the behavior shown in the video linked on Monorail:
1. Even though ScreenOrientationController is only for tablet mode,
SplitViewController always gets display orientation from the
ScreenOrientationController, even in clamshell mode. If you change the
display orientation in clamshell mode, then the
ScreenOrientationController continues to indicate the display
orientation as it was at the onset of clamshell mode, and so the
SplitViewController continues to use that display orientation until you
switch to tablet mode and back to clamshell mode.
2. When the SplitViewController detects portrait orientation, clamshell
split view uses snap positions on the top and bottom, while WindowState
expects snap positions on the left and right. This disagreement between
the two classes disrupts their collaboration on clamshell split view.
SplitViewController's opinion is based on longstanding behavior of
tablet split view, and WindowState's opinion is based on longstanding
behavior of window snapping in clamshell mode. Ben McCanny agrees with
WindowState: the snap positions should be on the left and right because
that is how window snapping in clamshell mode has always worked.

Fixed: 1023845
Change-Id: Ie81346904552acae88cb36227f30de7d7fceed6b
Bug: 1023845
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1949453
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722624}