[Debug] Remove sampling logic in throughput UKM
Currently the number of throughput UKM reporting is un-expectedly low on
canary channel. We suspect that the sampling logic could cause that.
This CL removes the sampling logic. Once we confirm that the reason is
the sampling, we will likely invent another sampling logic.

Bug: 1029964
Change-Id: I104de8f63ef261a72a9d93edeae2fd62a8235e3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1946872
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722645}