Roll OpenVR SDK 1.8.19
The OpenVR SDK has made some changes that we need since the last time it
was updated. Primarily, this is the addition of the method:
GetAppContainerFilePaths and a few other updates intended to allow the
SDK to work better in the AppContainer version of the WebXr device
process.

Some prototyping will still need to be done before this can actually be
achieved, but doing this first roll better enables that and sets up a
(hopefully) smaller diff if another roll is needed due to any issues
discovered during prototyping.

Fixed:845255

Change-Id: I6f9af21df282b3fe90cd9855aefd4b4ff9cc3c56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1935411
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722658}