blink/image_bitmap: Use shared images for textures.
Replace the use of texture creation via skia with shared images in
ImageBitmap. The motivation behind this change is to ensure that the
resources created for ImageBitmap can be used by the compositor using a
different backend and to eliminate usage of skia in client code for
OOPR.

This change uses CanvasResourceProvider, instead of SkSurface, to
issue work executed using skia. This both replaces the resources
created with shared images and allows a natural transition to OOPR,
managed internally by CanvasResourceProvider, going forward.

R=kbr@chromium.org

Bug: 962630
Change-Id: Ibd9c9fc0358a2423b73c15c4e05e1926f6ebb640
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1948237
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Auto-Submit: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722707}