[signin] Fix unconsented primary account when token revoked during load
The unconsented primary account was not updated correctly when the
token is revoked during database load.
As a result, OnTokenRevoked was sent without
OnUnconsentedPrimaryAccountChanged being sent.

This fixes a startup crash in SyncAuthManager. This CL also updates
SyncAuthManager to listen to unconsented primary account changes instead
of cookie changes because it is more accurate.

Fixed: 1029318
Change-Id: I14cbf88b4c24c1a73b783c3db79ac1643bfc7069
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1944336
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Monica Basta <msalama@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722873}