Clear a removed recalc root for ShadowDOM v0.
The initial flat tree style recalc implementation was falling back on
using the shadow host as the new recalc root instead of clearing the
removed root and clear the flat tree ancestor chain like we do for Shadow
DOM v1. That caused a microbenchmark regression because it added the
need for extra traversal during style recalc, even if no extra style
recalc was done.

Try to use the same method as for Shadow DOM v1 where we find the
child-dirty parent insertion point for the removed node.

Also, there was a missing nullptr return which made us return the shadow
host as the FlatTreeParentForChildDirty() for a node outside the flat
tree.

These changes should fix the remaining performance regression in issue
1023215.

Bug: 1023215
Change-Id: I737cd0636ffc7738a3ba75a3be91147b9917d039
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1954022
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722907}