Fix WebRtcVideoQualityBrowserTest.MANUAL_TestVideoQuality
Depending on timestamps recorded in WebM files, this test which uses ffmpeg to
convert WebRTC peer connection video to .yuv video format will sometimes
misinterpret the framerate in the recorded WebM file. This causes the resulting
.yuv file to be gigantic (several GB) with loads of duplicate frames. That
confuses frame_analyzer which is used in the test to compute video Ssim.

The fix is to convert every frame in the WebM file to it's own YUV frame, and
we accomplish this by passing -vsync=passthrough to the ffmpeg invocation.

Bug: 1032031
Change-Id: I372d766561d2568ce7bd5d1c15cea27b50063202
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1953630
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Markus Handell <handellm@google.com>
Cr-Commit-Position: refs/heads/master@{#722967}