Fixed stuck inkdrop effect on download shelf item.
Fixed the issue with a stuck inkdrop on drag of a download item
due to incorrect handling of the button's OnMouseDragged() and
OnMouseCaptureLost() handlers. Prior to this change
DownloadItemView attempted to intervene in the button's inkdrop
display logic.

Removed inheritance from InkDropHost for DownloadItemView as it is
unnecessary due to the use of a transparent button overlay for the
button effects.

Bug: 1026384
Change-Id: I2436fab35f5334ca0bd947189c831f932ba1f0ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1955657
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723036}