Introduce chromecast.mojom.ServiceConnector
This new interface is analogous to a service_manager::Connector, but is
intended strictly for use with Cast-related services.

The purpose of adding this interface is twofold:

1. This makes it possible to migrate Media Service off of Service
   Manager. Instead of passing a Connector into CastRenderer,
   CastRenderer now acquires a ServiceConnector binding from Cast
   browser code via the InterfaceProvider held by Media Service.

2. Uses of Connector in the internal cast repo can start to use this
   interface in place of a Connector to help them transition off Service
   Manager without significant refactoring effort.

One of the explicit goals of this CL is also to make it so that the
ServiceConnector implementation in the browser is immediately and easily
extensible within the internal repository: CastContentBrowserClient
creates the impl instance, and its creation can be overridden with an
an internal subclass.

The internal implementation can connect directly to internal services
without bouncing through Service Manager, and it can introduce new
values for ServiceConnectorClientId to identify internal clients.

Bug: 977637
Change-Id: I013325512fef6624b435b47509a64a15ffd7d39d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1954769
Reviewed-by: Sean Topping <seantopping@chromium.org>
Reviewed-by: Yuchen Liu <yucliu@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#723044}