Update the app id checking from app type to whether app_id is empty.
This is used to resolve the issue of incorrect updates for apps opened
on startup. AppService App structure uses mojom, which is slow. So when
system starts up, App structure is not ready yet, so the instances are
missed in AppSerice instance registry. Actually even if the app is not
managed by AppService, we can still add it to instance registry,
because AppService instance registry is sync, and depends on windows,
not depend on the app info, so it should be fine even if add the app
not managed by AppService.

BUG=1011235

Change-Id: I2df54dc9d534ebf506d5cee57dec4ba7142d8f3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1957604
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723083}