[devtools] Run layout tests from devtools and chromium side-by-side
As devtools-frontend project moved to a new repo we want to
offer support for layout test migration (to the new repo
and later to karma tests). Relevant tests will be copied from
chrome repo to a new devtools-frontend root.
We want to be able to run layout tests twice:
- run with a filter only the relevant tests from chrome repo
- run all test with a root in devtools-fronted repo


Bug: 1018670
Change-Id: If7dcd7095d0d604ccf0a73dc0cf3f2392a40d15b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1940279
Reviewed-by: Robert Ma <robertma@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Liviu Rau <liviurau@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723094}