overview: Fix bounds for wide/tall windows in overview.
Ran into this issue when trying to increase the margin for the below
bug and OverviewSessionTest.ShadowBounds failed. The computations were
off by a bit but happened to worked with the given margin. So, if we
increase kWindowMargin [1] to 8, tests would fail, but that shouldn't be
the case, as the algorithm should be able to handle that.

Had to tweak the tests so that they dont run into the case the wide item's
preferred height is taller than the area given by the computations, and
then clamped. It's a rare case, and will just be squashed a bit, but
messes up the test a bit that tests the regular cases.

Also removed some unused codes.

[1] https://cs.chromium.org/chromium/src/ash/wm/overview/overview_constants.h?rcl=42a7d2fa9327a1af93d97b309f4f0b1e53c11856&l=20

Test: manual, existing tests
Bug: 989794
Change-Id: I0d56b03d9dac2a8ec1732bb8479ab90a1d5974f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1949031
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723102}