Remove GetToolbarActionsBar from BrowserWindow
The only non-test use of this was from AppMenuModel. This adds a
ToolbarActionsBar::FromBrowserWindow() method which casts
::GetExtensionsContainer() to a ToolbarActionsBar. This method DCHECKs
that it's never called when kExtensionsToolbarMenu is enabled.

The existing test uses of this method will obviously have to be updated
to work with ExtensionsToolbarContainer when ready. This was also true
before this change as ::GetToolbarActionsBar() CHECKed that the
experiment was disabled.

Bug: chromium:985834
Change-Id: I83ad087011e45209d4f4d5193077ac2005f059b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1955760
Commit-Queue: Peter Boström <pbos@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723125}