cc: Notify why a frame was not produced.
It is possible to not produce a frame because the compositor is
attempting to recover from high latency, or because there was no
damage (and so a new frame is not required). When a frame is not
produced, include the reason for it from the scheuler. This can
be used by the throughput tracker to track whether a frame was
expected to be produced or not.

BUG=1021963

Change-Id: Ide47d0afdd8924e598af23713e679d425d953315
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1955225
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723143}