Bottom container no longer listens to bottom sheet
This patch removes the bottom container's dependency on the bottom
sheet. This code path was useful when the bottom sheet was a more
persistent part of Chrome's UI, but since this is no longer the case,
it is not needed. The patch also includes the removal of an unused
method and makes another package protected.

Bug: 986310, 1011056
Change-Id: I8c23f4a1f7ca49a3548d962574e65ff6e4a262a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1726313
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723146}