Set the top level window for the Web app.
For Web apps, set the top level window as the window, because there is
only one tab in the browser, so that the window pointer could be align
with the app window in AppServiceAppWindowLauncherController. If using
the contents->GetNativeView(), there could be two instances for one app
window.

Since IsOpenedInBrowserTab is used by AppServiceInstanceRegistryHelper,
move that function to AppServiceInstanceRegistryHelper, and rename it
to IsWebApp. Update AppServiceAppWindowLauncherController to reuse the
IsWebApp from AppServiceInstanceRegistryHelper.

BUG=1011235

Change-Id: I54a5c03aa08ca4cc1c4d9c53965a2e440e379d66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1956966
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723155}