split view: Have the divider widget own the divider window
With the divider window owning the divider widget (as before the present
CL), some code is slightly more complicated than necessary. When the
divider is destroyed, it calls Widget::Close, which leads to
NativeWidgetAura::Close, which posts a task to close the divider window.
Thus, when you exit tablet mode, the divider window is closed in a
posted task, leaving plenty of opportunity for DividerView::Layout to be
called in clamshell mode, for which it needs special handling.

Bug: None
Change-Id: I029ffe4d48071d6ca204e87483b58560be6c20ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1956065
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723167}