Inner and outer viewports must be set together
https://crrev.com/c/1871329 made it so that the compositor can assume
the existence of the inner and outer viewports is linked; if one exists
the other must as well (and conversely, if one is null, the other is
too).

It appears that currently there are some edge cases where this can be
violated. These tend to revolve around root scroller during loading.
While I work those cases out in a proper way, this CL ensures that we
maintain the above invariant when pushing viewport ids from Blink to the
compositor.

This should be safe to do as the violating cases are transient, see
https://crbug.com/1030381 for details so it's better to lose some
viewport functionality for a frame than crash.

Bug: 1028083,1030381
Change-Id: Icd2d4af67d5f8c4966c46be0fc90f07776da02cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1955860
Commit-Queue: Philip Rogers <pdr@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723198}