media/gpu/vaapiVEA: Crop video frame with VPP if visible size matches
When a webrtc applications wants 16:9 resolution and camera captures 4:3
resolutions, the input frame's size is 16:9 and its visible size is
set to 4:3. For instance, the frame size is 640x480 and its visible size
is 640x360. The VEA is configured to produce 16:9 resolution as webrtc
demands the resolution, 640x360 in the above example.

VaapiVideoEncodeAccelerator wrongly scales the 640x480 input frame to
640x360 in the case. This causes a receiver observes a distortion.
VaapiVideoEncodeAccelerator should rather crop in the case, for example,
extract the area specified by the visible rectangle of the input.

Bug: Bug: b:145769650, 1031019
Test: meet on soraka
Test: appr.tc/?debug=loopback&vsc=vp8&video=maxWidth=640,maxHeight=360,minWidth=640,minHeight=360 on soraka
Change-Id: Id490f08f290aa79b2095980ec2f401973b31711b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1957911
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723222}