GuestOsSharePath FileWatcher crashes fix
There are some crashes where it seems that
GuestOsSharePath has been destroyed when
FileWatcher callbacks run.

This change fixes all code which runs in the
FileWatcher sequence to not access the
GuestOsSharePath object, but rather
post tasks to the UI thread where a WeakPtr
is used before running any code on the
GuestOsSharePath object.

Bug: 1031080
Change-Id: Ia8acae392eabf4f3d13847dbd7331c3321fba410
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1956981
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723259}