[AA Direct Actions] Static action for FindInPage.
This patch adds a static direct action for "find in page" to the
set of already existing static actions (e.g. new tab, go back, etc). In
order to be consistent the FindInPageManager is used to show the UI and
the TextField for the search term is then populated with the search
query passed in by an assist app using Android's direct action API.

There is no voice support planned for moving to the next/previous match
or close the "find in page" UI again.

Bug: b/145106813
Change-Id: Ieaee42f039cc77b7f03bdbba0b110aaa0887e039
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1944396
Reviewed-by: Stephane Zermatten <szermatt@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723313}