viz: Add OverlayProcessorInterface as a Base Class
This is a pretty straightforward splitting of OverlayProcessor interface
and implementation. This splits out the interface to be shared with
OverlayProcessor (Mac and Windows for now), and OverlayProcessorStrategy
(Ozone and Android).

The next step would to be further split subclass to be on different
platforms.

This is CL 5/? of overlay processor refactor.

R=rjkroege
TBR=kylechar

Bug: 963597
Change-Id: I7571a9e13b7fc4cc5d2dcbfbc39699e7d03d06ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1951654
Reviewed-by: weiliangc <weiliangc@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: weiliangc <weiliangc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723369}