Plumb consuming trusted vault keys on Android
Interface TrustedVaultClient is now implemented specifically for Android
where the calls to fetch keys need to reach Java, where a
embedder-specific logic should be integrated.

The write path (i.e. TrustedVaultClient::StoreKeys()) is marked as
unreachable because the codepath is unused on Android, and in fact the
only real calling site (SyncEncryptionKeysTabHelper) is not even built.

Bug: 1012659
Change-Id: I02e0b331f4cfd50cff0ff6f0af0e3dfe44601ae1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1950993
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723393}