[iOS] Reduce dependency on TabModelList for ntp metrics
To reduce usage of TabModelList, this CL removes all but two code
paths in the new_tab_page_uma:: helpers that map a browserState to a
single active WebState.

new_tab_page_uma::RecordAction now takes a WebState by default. In cases
where there isn't one to pass in, the deprecated (prior) version of the
function uses TabModelList to find the active WebState of the TabModel
associated with the current BrowserState.

The two remaining code paths for RecordAction that don't pass a WebState
are from ReadingListCoordinator (which doesn't yet have a Browser passed
in) and HistoryTableViewController (likewise).

Bug: 1032550
Change-Id: If0eee5c2fceed08357d806e973ce53e027d6ea2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1960308
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723410}