Remove some VizDisplayCompositor feature code
The VizDisplayCompositor feature is now 100% launched for all Chrome
browser platforms. Remove the feature from chrome://flags and delete
some of the code to check the feature.

We still need to check the feature flag on Android since WebView
requires it be false. As a follow up the remaining places
IsVizDisplayCompositorEnabled() is still needed should be converted to
check IsUsingVizForWebView() and features::kVizDisplayCompositor can be
deleted entirely.

Bug: 732805, 936425
Change-Id: If6894f62195b7d5593ad890be4f95f2ea8025011
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1957014
Reviewed-by: Saman Sami <samans@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723453}