[BackgroundTaskScheduler] Exact tasks receiver
The possibility of scheduling a task at an exact time is being added to
the BackgroundTaskScheduler. Previous CLs have handled refactoring,
interface and delegation changes.

This CL adds the functionality that starts a task. A task will be
started at the decided time, the scheduling being handled by
AlarmManager. The start of the task and all the additional actions are
done through a receiver class. The start and the wait for the task are
done in a detached thread, while keeping the CPU on through a wake lock.

Bug: 970160
Change-Id: I5e981ebf34c80b6fff6cdba16e4fe2811720560f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1778880
Commit-Queue: Richard Knoll <knollr@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Mugdha Lakhani <nator@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723456}