Only show permission wizard if remoting host is enabled.
The launchd helper tool shows the permission wizard for all non-root
users, even if the host was disabled. This CL fixes this, so that the
launchd service exits immediately if the host is not set up. So the
wizard is no longer shown during first installation, and will first
appear when the user sets up a host.

The wizard is still shown during upgrade if the host was enabled but
permission was not granted. This may be harder to fix, because the
helper tool (launchd service) cannot easily tell whether it was
started by the installer or some other cause.

Bug: 1032164
Change-Id: I3e8c742f4caf5520cc494b0322bd0f084d5fa881
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1960582
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Jamie Walch <jamiewalch@chromium.org>
Auto-Submit: Lambros Lambrou <lambroslambrou@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723500}