[scheduler] Remove priority-based anti-starvation logic
This CL removes the priority-based anti-starvation logic from
TaskQueueSelector. Anti-starvation has been disabled both in the
browser and for the main thread in Blink, so this CL should not
result in any additional behavior changes.

This CL also replaces the backing data structure (SmallPriorityQueue)
with a simplified ActivePriorityTracker, which is a wrapper around an
integer representing which priorities are active.

This also cleans up various unit tests:
- Remove parameterized tests (with/without anti-starvation)
- Fix PrioritizeCompositingAndLoadingInUseCaseLoadingTest in main thread
  scheduler unit tests. This was not running because it was missing an
  INSTANTIATE_TEST_SUITE_P, and the test was missing a comma causing it
  to fail.
- Replace the TaskQueueSelector starvation unit tests and test for
  different orders of posted tasks. These were modified to be a bit more
  future proof/general and to test all priorities.
- Remove the unit tests that test the starvation score logic.


Bug: 1031336
Change-Id: I1ff8d83617d33c7510c33c08b78b1abc1a148683
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1954760
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Alex Clarke <alexclarke@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723517}