[WebLayer] Simplify asset and resource handling
This change makes it so we no longer need to add asset paths to the
applications context, and instead we wrap the context in a
ContextWrapper which overrides getAssets() and getResources() to
directly return the implementation assets and resources.

This means we can no longer use WebViewFactory.getProviderClass(),
since it adds asset paths internally. Instead we load the context from
the WebView provider using WVF.getWebViewContextAndSetProvider() on N+,
which takes care of signature verification, and load manually using
createPackageContext() below N. We also need to make sure the native
library is loaded so we can use the WebView shared relro implementation.

Change-Id: I082107a71d1bf71069e498843a9ecc00470e6cb8
Bug: 1032662
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1931617
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723520}