Do full occlusion test when generating viz hit test data
When generating hit test data to send to the viz process, the renderer
will mark any iframe that has other cc::Layers overlapping it as
"inconclusive" for hit testing, which means that the browser process
will have to send the hit test to the renderer process to get a final
result.

However, just because a cc::Layer overlaps an iframe does not
necessarily mean that there is painted content overlapping the iframe.
With this patch, the renderer will perform a hit test to determine
whether any web content overlaps an iframe; and if not, then it avoids
setting the "inconclusive" flag for hit tests targeting the iframe.

BUG=1014273

Change-Id: I4812894972cfd33be52c484628005ac60fe65440
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1956385
Commit-Queue: Stefan Zager <szager@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723607}