[Structured metrics] Add API classes.
This CL adds the skeleton of the structured metrics system within
/components/metrics/structured. It's very similar to the set up of the
AppListLaunchRecorder, and contains three parts:

 - EventBase. Represents a single logging event, and will be
   subclassed by generated code for each event added to the file of
   metrics definitions.

 - StructuredMetricsProvider. This is owned by the metrics service and
   will collect, hash, and report metrics. It observes the Recorder for
   events.

 - Recorder. This is a glue singleton that allows EventBase objects to
   communicate with the StructuredMetricsProvider. It also provides a
   mechanism for the ChromeMetricsServiceClient to inform the provider
   when a profile is added (ie. a user logs in), which triggers
   initialization.

This CL instantiates the new metrics provider and adds it to the
metrics service client, so the code is running on the waterfall and
any issues are caught early.

However ProvideCurrentSessionData will be empty until the final CL, which
I'll send to the metrics team. If you'd prefer to do a full review before
the provider is instantiated, just let me know.

Unit tests checking initialization logic and that the notification
route from recorder to provider works correctly will come in a follow-
up CL.

Bug: 1016655
Change-Id: I632bd47d712c10201b2d322f3ac0a0e918334738
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1918781
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723613}