IDL: Drop [Exposed] from UserAgent dictionary
Dictionary definition in Web IDL is actually mapping strings to objects.
So it doesn't make sense to have [Exposed] on it.

Currently the spec of UA defines it as an interface NavigatorUAData,
but this CL doesn't work for the change.


Tbr: yoavweiss@chromium.org
Bug: 839389
Change-Id: If41f713b1508ec0f19b242415eb78669872275bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1958271
Reviewed-by: Hitoshi Yoshida <peria@chromium.org>
Commit-Queue: Hitoshi Yoshida <peria@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723663}