Use ui::SystemInputInjector for injecting mouse events
Previously the mouse events are injected to the WindowTreeHost,
but that doesn't actually move the mouse cursor which is visible
on the screen. This CL rewrite the event emittion based on
ozone's input injector mechanism, which does move the mouse
cursor actually, and the mouse events are processed in a more
realistic code path.

A bonus point is that it does not have to enable the mouse
events explicitly.

Bug: none
Test: tast tests
Change-Id: I1dd8980f949d300826b997a7eee0c6eb0b9102e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1949734
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723670}