ServiceWorker: Avoid accessing ServiceWorkerProviderHost on non-SW-core thread
To get RenderProcessHost, RunOrPostTaskToBindServiceWorkerReceiver() posts a
task bound with ServiceWorkerProviderHost to the UI thread. However, the
provider host is supposed to be accessed on the service worker core thread that
can be the non-UI thread (IO thread) depending on the flag configuration.

To avoid the illegal access on the non-UI thread, this CL stops binding the task
with ServiceWorkerProviderHost, and instead passes a worker process id for
taking RenderProcessHost on the UI thread.

Bug: 936482, 1032428
Change-Id: Ieb717a7df7ceffee4606be9f984947c595ccfb79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1959200
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723695}