[cros-fuzzy-app] Allow fuzzy string matching to match exact texts.
Because of the way tokenization is done, it's possible a query and a text have
different sets of tokens, but in fact they match exactly with different camel
cases only.  This cl allows exact name matches igorning the camel cases.

In this case, we interpret "hits" differently: it becomes case-insensitive and
returns the entire match.

Bug: 990684
Change-Id: Iaecd5285d5d62b38199ffed10f334561697e9143
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1958298
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Commit-Queue: Jia Meng <jiameng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723721}