[iOS] Reload instead of showing SadTab on browser crash
When the browser crash for the first time, the page should be reloaded
instead of showing the sad tab view.
Previously the SadTab had a "Reload" button when presented for the
first time, then a "Send Feedback" button if the crash was repeated.
Now there is only the "Send Feedback" button.

Bug: 1021881
Change-Id: Ie2bd38f240d19e3ba7607d1cf71ee59c86a485ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1901081
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723756}