[directorytree] Add an active directory member to DirectoryTree
When the current directory is changed to some new directory, the file-
list displays the content of the new directory: that directory becomes
the "active" directory.

Add the active directory as a DirectoryTree member, and update it when
the current directory changes. Its corresponding tree row will gain an
'active' attribute [1]. Add a test of the attribute behavior.

[1] In files-ng, the [active] .tree-row (folder) is styled differently
compared to normal Files app.

Test=browser_tests --gtest-filter="*directoryTreeActiveDirectory*"

Bug: 992819
Change-Id: I974cce636a3000cd7de4554b9752a21cc68f5f20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1957912
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723822}