split view: Fix always-on-top status of divider in certain situations
The present CL fixes state of the split view divider after a window drag
is canceled or the dragged window is destroyed. First of all, the
divider should not be set as always on top. More importantly,
is_dragging_window_ should be false so that OnWindowActivated will set
the always-on-top status correctly thereafter.

Bug: None
Change-Id: Icd3dc7405dceb832206c686cede5e6d809826637
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1959202
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723865}