Rename CreateFrontElidingTitleLabel to CreateTitleOriginLabel
This removes the policy (front eliding) from the name and changes it to
the intended use of the label.

The front eliding policy is likely to change in the future, so adding
it to the name creates unnecessary churn.

BUG=987715

Change-Id: I0ce4190f6de38cfc218984adb0204fcf4602e949
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1960969
Auto-Submit: Robert Liao <robliao@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723873}