[ozone/x11] Moved core part of XDND out of DDDClientAuraX11.
This CL is the second attempt to begin refactoring drag and drop for
Linux platforms and uniting Ozone and non-Ozone implementations.

The first attempt had been discussed in [1].

The X11 drag and drop is currently implemented in DDDClientAuraX11 [2]
that cannot be reused in Ozone.  To make it reusable, we need to move
as much as possible into some common place that doesn't depend on
platform events.

This CL moves some pieces of logic from DDDClientAuraX11 into
//ui/base/x so it can be used later by other drag and drop clients.

[1] https://crrev.com/c/chromium/src/+/1954010
[2] views::DesktopDragDropClientAuraX11

R=thomasanderson@chromium.org

Bug: 1014860
Change-Id: Ifaa81317dc5fa3da7a67d5d1c18b1626a982c794
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1960155
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723875}