CrOS app list: Ignore modal bit in |HasVisibleWindows|
Before this CL, modal windows would not count in figuring out whether
there are visible windows open (which in turn is used to check whether
the app list will be visible in tablet mode). There was no reason for
that, as modal windows are as visible as others.

This fixes an issue where modal windows could trigger the hotseat
because we believed to have shown the app list.

Bug: 1022491
Change-Id: Ib1b3c0095524ec4ca1372ee6b30cee614f55ca9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1961831
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Manu Cornet <manucornet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723908}