Add GetAvailableSize() support to LayoutManagerBase.
This makes available size a part of a ProposedLayout and reports that
size when LayoutManagerBase is queried via GetAvailableSize().

Search is currently linear but Views fan-out tends to be small (N < 10)
so using a map is probably a performance hit rather than a benefit.

Additional tests for ProposedLayout are also added.

Bug: 995448
Change-Id: Ie7aa98e398e5ea296cf2268c2a0c48549de8750e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1959685
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723912}