viz: Split OverlayProcessor into Windows and Mac Subclass.
OverlayProcessor class is used for both Windows and Mac, and the only
difference is in it uses different overlay candidate validator classes.
The validator classes implementation are simple, and achieve the same
effect with splitting OverlayProcessor into Mac and Windows subclass.

This allows us to delete OverlayCandidateValidator and its Mac subclass
and its Windows subclass.

This is CL 6/? of overlay processor refactor.

R=rjkroege
TBR=kylechar

Bug: 963597
Change-Id: I78265a8941720daa32cbd672ef85c7a547938b8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1953824
Reviewed-by: weiliangc <weiliangc@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: weiliangc <weiliangc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723971}