Add Memory.Browser.MemoryFootprint for iOS
iOS records a metric Memory.Browser. This metrics differs from the concept of
memory footprint, which we use on all other platforms. Memory footprint is much
better at measuring chrome memory leaks/bugs.

This CL adds MemoryFootprint [for which there is a syscall on iOS]. We don't
bother measuring PrivateMemoryFootprint and SharedMemoryFootprint since iOS is
single process, and MemoryFootprint is a sum of the two. It's possible to, but
requires a lot more plumbing to compute PrivateMemoryFootprint.

Change-Id: I5d330fdff57ab5d21e16f3e68c127b1063c8144e
Bug: 1033151
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1956360
Commit-Queue: Erik Chen <erikchen@chromium.org>
Auto-Submit: Erik Chen <erikchen@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723976}