overview: Nix GetDropTarget()->OnSelectorItemDragStarted(dragged_item);
There is currently no important reason for
OverviewGrid::AddDropTargetForDraggingFromOverview to call
OverviewItem::OnSelectorItemDragStarted on the drop target. I originally
did that to hide the drop target's close button, but because it was
animated, the behavior was found unsatisfactory as described in
Issue 954642. [1] fixed the bug, and also removed the need for the
function call in question.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1900127

Bug: 954642
Change-Id: Iadef44cc213158a3cc797670a46c97c4f0aa2bb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962596
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724048}