Fix the browser test LauncherPlatformAppBrowserTest.HiddenAppWindows.
Add the OnAppWindowHidden interface to ExtensionApps to clear the
running, active, visible state bits, and set the state back to the
started state, when the app window is hidden.

ExtensionApps is updated because we need the clear bits function for
the new hidden interface, so the related checking and state comparing
are moved back the app window interface.

At ui side, if the state is changed back to the started state, remove
the app window from Shelf.

BUG=1011235

Change-Id: Idae14632368758526ea5199afb2ff2aaaeaf9f39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1961349
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724157}