Use a Vector of primitive type in a test
Migrate away from a Vector of NGInlineItemResults (which
is a Vector itself) to a Vector of primitive types
to avoid having a Vector of Vector and address the test
needs.

This helps migrating some objects later to Oilpan easier
as NGInlineItemResults has some inline capacity.

Change-Id: I6db97a1bf32432faff8d757d58a9bfbc48431bc7
Bug: 1030176
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1963692
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Navid Zolghadr <nzolghadr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724160}