media: Introduce VideoEncoderInfo
libwebrtc can know the information about encoders that are not
contained in the library, by webrtc::EncoderInfo. In the current
chrome implementation, the structure stops in RtcVideoEncoder and
thus the filled information doesn't respect the hw encoders.

This Cl introduces VideoEncoderInfo. It is similar to
webrtc::EncoderInfo. The information is filled in and dealt with
the structure in chrome code.

Bug: 1031965
Test: VaapiVideoEncoder is shown in chrome://webrtc-internals
Change-Id: I7d2f5b0be6655620baf9007e984fc459287e988a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1949841
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724174}