Convert Callbacks in android_webview/ to {Once,Repeating}Callback
To do so we have to convert a number of Callbacks in components APIs.
In the process, it's clear that a callback to WebDataServiceBase
constructor is stored only to be consumed in Init(). So we move it to
be a parameter of Init(). This points out a missing call to Init which
is marked with a TODO.

We use OnceCallback where possible because the callback is only used
once and RepeatingCallback and BindRepeating otherwise.

Bug: 1007631
Change-Id: Ieb5210a1ecb7d662a7874ad4bf4501e7c24dd46c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1963106
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724219}