VulkanSwapChian: Delay destroying present wait semaphore.
Right now, present_wait_semaphore is destroyed just after submitted
GPU work is finished by GPU. It is not safe for VulkanSwapChain,since we
don't know when PE will issue command to wait on that semaphore. So to
make it safe, this CL delays releasing it until that image is returned
from PE and a new GPU work for rendering that image is finished.

Bug: 1026673
Change-Id: Ic55bf8ff7e83f711897521adb0a2233f1c5afa16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962910
Commit-Queue: Peng Huang <penghuang@chromium.org>
Reviewed-by: Jonathan Backer <backer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724231}