Clean up ATK interface initialization
Clean up the code that computes the interfaces to expose for ATK object
implementations and ensure that it doesn't interact with the rest of the
accessibility tree. Previously, it was trying to determine if a node was
the root node when deciding whether to implement the AtkHyperlink
interface. That check is redundant though, so we can safely remove it.

Bug: 1032555
Change-Id: I63aec540de7c2d0c14d8f59eb7e64ed45585c8db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1960448
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724256}