Rename link to annotation in PDF a11y action handling pipeline
Earlier we only had link in PDF a11y on which action could be performed.
Now more annotations are to be added in PDF a11y which are actionable.
e.g. highlights.
This CL renames the variables and functions from link to annotation in
action handling pipeline as the same variables and functions will be
used for all actionable annotations.
Additionally an enum PP_PdfAccessibilityAnnotationType is introduced to
disambiguate between the different annotations on which action is to be
performed.

Bug: 1008775
Change-Id: I1b6d75976c49dba5778cb3e7871df318f023779b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1958495
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724262}