workspace: For mouse dragging to snap, use same tolerance as for touch
When an edge of the work area is at an edge of the display, there shall
be a tolerance of 32px to drag a mouse to that edge and snap a window.
Before the present CL, this tolerance already exists for touch dragging,
and it is helpful because as you can drag past the edge of the display
(and onto the surrounding area of the physical device), it is difficult
to drag exactly onto the edge of the display. The present CL brings the
same tolerance to mouse dragging because mouse dragging can have the
same problem in the context of multiple displays.

Bug: 986870
Change-Id: Id8ea14adc70b14e686aab85cee65d8f8449dc139
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962998
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724279}