Remove NavigationRequest::associated_site_instance_id
The NavigationRequest was keeping track of the
|requesting_navigation_client_|'s SiteInstance ID.

There are two cases:

 * For same-site navigation, the RenderFrameHost is reused, so the
   |requesting_navigation_client_| and the
   |committing_navigation_client_| are the same.

 * For cross-site navigation, the RenderFrameHost is different, so the
   |requesting_navigation_client_| and the
   |committing_navigation_client_| are different.

Instead of checking if the navigation is same-site or not, we can simply check
if the navigation is same-RenderFrameHost or not.

This patch is not required, it is just a simplification. However, it will became
required with RenderDocument. The navigation can be same-site, but it can commit
in a different RenderFrameHost.

Bug: 511340
Change-Id: Iaec100a755f1b8e16e95f747c458cc33d487d33b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1960271
Reviewed-by: Arthur Hemery <ahemery@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724287}