Fix for scroll_latency_browsertest
Native scrollbars on Mac are overlay by default. This is a  problem to
the tests that rely on scrollbar manipulation . This CL fixes the issue
with scroll_latency_browsertest by ensuring that the scrollbars on Mac
remain solid for the duration of the test.

Bug: 1026720
Change-Id: I01c986e735cdffe64bc749a9689876a43a85c74a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962771
Commit-Queue: Rahul Arakeri <arakeri@microsoft.com>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724315}