[webaudio] Remove uses of deprecatedLengthAsUnsigned
This CL replaces calls to deprecatedLengthAsUnsigned by calls to
lengthAsSizeT. Unfortunately the content of the ArrayBuffer is copied
eventually into a WTF::Vector, whose maximum size is smaller than the
maximum size of an ArrayBuffer. Therefore I reject the incoming
ArrayBuffer with a RangeError if its size is too big.

Background: we prepare ArrayBuffers to be bigger than 4GB. Therefore we
changed the size field to size_t. Now we are changing all uses of
ByteLength to be able to deal with size_t, either by accepting a size_t,
or by throwing an exception if the size is too big.

R=rtoy@chromium.org

Bug: chromium:1008840
Change-Id: I63b9fd43a3d1afb4d06ebee105b1fc48045e1730
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962276
Reviewed-by: Raymond Toy <rtoy@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724332}