[ios] Export symbols of dependencies of MaterialComponents
When building MaterialComponents as a framework, its dependencies
will be included in the framework (in order to reduce the number
of frameworks built).

To work properly, this requires the symbols of the dependencies
to be visible (otherwise the linker complains about missing symbol).

This is done unconditionally since 1. MaterialComponents will
soon exclusively be built as a framework, 2. those libraries
only expose Objective-C symbols that are always visible anyway
(the runtime will expose them even if they are hidden).

Bug: 1017165
Change-Id: I7d9698928ccf42e0b52211f079b82d7ed46805f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1964773
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724363}