bluetooth: Test prompt controller OOB repro case
This change adds a unit test that reproduces the OOB bug that occurred
in WebBluetoothServiceImpl if it was cleared due to an error, but the
BluetoothDeviceScanningPromptController still called
OnBluetoothScanningPromptEvent. The OOB access happened on
WebBluetoothServiceImpl::scanning_clients_ because the method assumed
that it would not be empty at that point.

Bug: 1024116
Change-Id: I1bcd0a286fb0faef7a502aaca2144eed99e3de7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1925883
Commit-Queue: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Reviewed-by: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Auto-Submit: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724375}