[omnibox] Improve multi threading safety for on device head provider.
This includes:
1.DCHECK |client| & |listener| in ::Create(...).
2.Add two sequence checkers for main thread and worker thread
respectively.
3.Pass AutocompleteInput |input| by value during task posting;
4.Explicitly invalidate weak ptr in stop. Note I still keep the request
id here since I'm not sure if this invalidation can prevent the 2nd
callback in base::PostTaskAndReplyWithResult from being executed even
though it's bound to the weak ptr.
5.Check request id at the start of each task posted to main thread;
this will force these tasks to stop after AC::Stop(...) is called, if
the invalidation in 4. above fails to cancel the task. This might be
important as it's missed in current
OnDeviceHeadProvider::StartInternal(...) which can be executed after
AC::Stop() is called.

Tested on Pixel 2 XL / iPhone X.

Bug: 925072
Bug: 1032705
Change-Id: Ibfd60e13931e6dd281baa33915607ff5c9536b2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1960498
Commit-Queue: Ce Chen <cch@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724378}