[Networking] Fixed wifi list scroll issue
When the network list updated, there is a check to see if any of the
network views has IsMouseHovered() true and a programmatic scroll is
done. The problem is that IsMouseHovered() returns true when the mouse
is outside of the bounds of the tray pop-up (below the list). The list
seems to continue down vertically below the cut off of the tray-popup.
The views are "visible" according to Views::GetVisible() but seem to be
masked away by the tray pop-up outline and are not user visible. However
IsMouseHovered does not care and it returns true in this case. The
simple solution done here is to just check if the view with
IsMouseHovered() == true is within the bounds of parent container. If
it is not, we don't do the programmatic scroll.

Fixed: 994674
Change-Id: I0e4c93a7edc0466394873d59b841f8cde002e00d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1966232
Commit-Queue: James Vecore <vecore@google.com>
Reviewed-by: Azeem Arshad <azeemarshad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724395}