Make sure Omnibox is restored for back-forward cache navigators
Currently if we navigate from a page that is not showing the Omnibox we
will not restore it, as the old renderer (which is the one animating the
show) will be replaced instantaneously. More over the restored renderer
might ignore SHOW events as its internal state (blink compositor etc)
"thinks" that it is already showing the Omnibox. The browser process
also caches a lot of related state.

Resetting all the Omnibox associated state "manually" seemed very hard,
so this patch makes sure that the restored renderer and the browser are
in sync again by sending a HIDE event to the cached renderer if needed.

Bug: 100574
Change-Id: Icd5a3ab3e7b66031409e4cb5c80fca2dd25f3dd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1949783
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Carlos Caballero <carlscab@google.com>
Cr-Commit-Position: refs/heads/master@{#724396}