CodeGen: Make blink_type_info work for buffer source types
This is to unblock compilations, and not to work for all the cases
that use buffer source types.

And adds an assertion to show a clear message that an unknown type is
passed in the function. (Otherwise, we may hit a property access error
on None objects.)


Bug: 839389
Change-Id: I98e54356e7dafd98a93023afc220700913b5d6f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1965779
Commit-Queue: Hitoshi Yoshida <peria@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724511}