[code coverage] Enable online profile merging for Android gtests
Reference link:
https://clang.llvm.org/docs/SourceBasedCodeCoverage.html#running-the-instrumented-program
Dumping data to ONLY 1 file may cause warning and data overwrite in browsertests.
Pattern "%2m" is used to expand to 2 raw profiles at runtime.
Example in chromium:
https://cs.chromium.org/search/?q=LLVM_PROFILE_FILE+2m&sq=package:chromium&type=cs

DeviceTempFile cannot be used with the pattern, and NamedDeviceTemporaryDirectory
is not applicable because the directory is not created until using "with" statement.

Bug: 985574
Change-Id: I8c2b9f1d289be1de3bd1b283cc35acc618a7d6a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1884378
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Yun Liu <yliuyliu@google.com>
Cr-Commit-Position: refs/heads/master@{#724532}