[DeviceSync v2] Use SoftwareFeatureManager to call SetFeatureStatus
When v1 and v2 DeviceSync are running in parallel, before v1 DeviceSync
is deprecated, use the SoftwareFeatureManager to indirectly call
CryptAuthFeatureStatusSetter::SetFeatureStatus(). This is to ensure an
ordering of SetSoftwareFeatureState() and SetFeatureStatus() calls.
These functions mutate the same feature states on the CryptAuth backend
and the order of the calls matters; we need to avoid race conditions.

When v1 DeviceSync is deprecated, the CryptAuthFeatureStatusSetter is
used directly.

Bug: 951969, 1019206
Change-Id: I0711b011dc91a0218503a01f3131d4062f51bc36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1944888
Commit-Queue: Josh Nohle <nohle@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724747}