Make ScriptedAnimationController a ContextLifecycleStateObserver
This removes the need for custom pause/unpause logic for
ScriptedAnimationController in Document.

Also, construct ScriptedAnimationController in the Document
constructor, rather than lazily. It will get initialized on the
first frame, so we're not saving much by deferring its initialization.
This simplifies a bunch of logic around its handling in Document, and
guarantees that its lifetime will be predictable for leak detection.

Bug: 1029822
Change-Id: I4cec7985258dd24154032b4cd1812b75cc858757
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1953643
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724751}