Reset app list content when app gets hidden
Resets the app list contents view when the app list view visibility
changes to false - the goal is to ensure that the app list is in
the initial state when it becomes visible again. This is in
particular needed in tablet mode, where AppListPresenterImpl::Show might
not get called when app list becomes visible (so app list contents is
not reset that way).

ResetForShow was also being called when app list was getting focused
while in hidden state, but that might not catch all the transitions
where app list becomes visible (and with this change becomes unnecessary).

Additionally, change when the selected apps page is reset for app list
visibility change - instead of doing it on focus change, the selected
page is now down when the visibility change is starting (i.e. in
OnVisibilityWillChange). This will make page reset happen sooner (before
the app list UI is actually shown), and cover more visibility change
cases in tablet mode.

BUG=None

Change-Id: I6240d7401091a8d778f5a3f8774db919a4cfd6bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1966374
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724752}