[CrOS A11y] Learn More links to be labelled by entire sentence
Currently, when the user tabs to a Learn More link (at the end of a
sentence usually), ChromeVox only verbalizes "Learn More".

The entire sentence that the "Learn More" refers to should be
read out in addition to the "Learn More".

There are two means of how "Learn More"s are presented with preceding
sentences:
1) When the Learn More is already embedded in the settings_strings.grdp
2) When the Learn More is concatenated with preceding sentence in JS

This CL demonstrates the handling of case 1).  Case 2) will be
converted to 1) in a later CL, as there are localization issues
of concatenating "Learn More"s with preceding sentences in JS.

A generalized polymer element that accepts a localized string that
contains a link will be created.  The linked contents will be aria
described by the entire localized string rather than only the
text within the anchor tag itself.  The other parts of the localized
string will be aria-hidden.

If agreed upon, the handling of case 1) will be applied to all
similar strings with "Learn more" embedded in settings_strings.grdp,
and later extended to all cases (once cases of (2) are converted
to (1)).

Bug: 1015924
Change-Id: I222a20b79006dcceedcf3893d5e4f0f77765fc18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1941198
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Hector Carmona <hcarmona@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724763}