Break noDialogs It2Me flag into 2 distinct flags for dialogs and notifications
Using the noDialogs flags is fine for an unattended scenario, however it
there may be cases where we want to provide a notification on the
desktop that a user is connected but still allow the remote user to
connect w/o prompting or requiring periodic reapproval.

This CL splits up the 'noDialogs' flag into 2 flags which control
suppression of the user dialogs (to allow / reapprove the remote
connection) and the desktop notification that is displayed.

Change-Id: Ifd2eb4f9270fadb4ffc2267bd2b1dfbe0ec38ca4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962530
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724766}