[omnibox browser stack] Clear redirect domain when we detect disable
When the DNS-interception-detector is disabled, Chrome no longer sends
detection packets. If the network were to change under this condition
after having learned the redirect domain, the latter would never
change. This CL clears the domain once it detects that the policy was
set to disable the feature.

Bug: 181671
Change-Id: Ie51489272dc3ee92d279e9c3623ba1682ab14330
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962132
Commit-Queue: Kevin Bailey <krb@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724784}