[Chromecast] Fix EOS handling when connection errors occur
Previously, if the caller had sent EOS but not yet received a response,
and the mixer connection was lost, it would get into a state where the
connection was left pending indefinitely (since we would not send EOS on
reconnect, and there was no signal to the caller). Instead, signal EOS
complete to the caller immediately if the connection is lost after EOS
was sent to the mixer.

Change-Id: I50a5adcd35e7c8c200e129076f5d2f274400a856
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967886
Reviewed-by: Yuchen Liu <yucliu@chromium.org>
Commit-Queue: Kenneth MacKay <kmackay@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724806}