Merge LayoutText phases into a single phase
This CL is
  1) merging the work done by
        EnsureLayoutTextUpdated(...) and
        EnsuresLayoutTextAttributeUpdated(...)
  2) simplifying the truncate text logic
        Only truncate text at the end.

EnsureLayoutTextUpdated(...) is copying text_ into layout_text_ and
it is performing codepoints rewriting. An invariant was assumed:
the amount of codepoints need to stay the same.

EnsuresLayoutTextAttributeUpdated(...) is building the layout break
lists. It is making a copy the text break list.

These two phases can be done at once. That way, there is no more needs
to keep the invariant on amount of codepoints. This is preparing the
code to the text eliding refactoring. Text eliding will break this
assumption.

The truncate logic is only used to avoid slow (huge) cases. The only
limit used in chrome is 10000.

The code for truncate was broken because it was possible to break
graphemes. Unittests are added to cover these cases. Also, of the
text is broken at the beginning or the middle, some invariants
are broken.

Example:
  bool RenderText::IsValidLogicalIndex(size_t index) const {
    return index == 0 || index == text().length() ||
        (index < text().length() &&
         (truncate_length_ == 0 || index < truncate_length_) &&
         IsValidCodePointIndex(text(), index));
  }
The truncated text is assumed to be at the end.

Bug: 1025561
Change-Id: Ifdee81c5a4527175258d0c9335cbe798e6d74775
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1948043
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724809}