aw: Always send BeginFrameAck with viz for webview
With viz we must reply to CompositorFrameSinkClient::OnBeginFrame with
DidNotProduceFrame/SubmitCompositorFrame. But as submitting frames for
webview depends on onDraw() and for viz happens only hardware draws
(DemandDrawHw) it wasn't always the case.

This CL ensures that we don't queue up unconfirmed BeginFrames by
sending DidNotProduceFrame for previous frame in onBeginFrame if
necessary.

Bug: 805739
Change-Id: I36c416406447f01914e5ae89fb0eab188a6a9c40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967880
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724824}