Move status area widget to ShelfControlContainer
Moves status area widget and notifications bubble from status container
to shelf control container.

The main motivation is fixing stacking order of hotseat and the
notification bubble - notification should be overlaid over the hotseat
when shown, but hotseat should be overlaid over the notification bubble
while hotseat is being dragged. To enable this behavior, both should be
in the same container. For that move status area to shelf control
container (given that is already visually part of the shelf).
Also, hotseat is stacked at top whenever hotseat drag is initiated.

Renames status container to overview focus container, given that
overview focus widget (which gets focused when overview is active) is now
the only window left in that container.

Also, update heuristic to determine whether a window is a notification
popup in tray_bubble_wrapper (given that there are more items in the
same container with the notification now).

BUG=1028315, 1032694,1032686, 1033342

Change-Id: I28287963c8226eba19c0c82c58fbb41716c02f38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962675
Reviewed-by: Tim Song <tengs@chromium.org>
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724853}