[remoting] Show a Dock icon when the permission wizard is displayed.
This CL displays an icon in the MacOS Dock when the wizard is shown,
and hides it when the window is dismissed.

This also causes an app menu to appear at the top of the screen. This
shows the bundle name by default, so this CL updates the bundle name to
the branded title of the app.

Bug: 1032149
Change-Id: I8e489d74fe1254f452c2d4dab3af919c8932e9ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1968198
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Jamie Walch <jamiewalch@chromium.org>
Auto-Submit: Lambros Lambrou <lambroslambrou@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724865}