Fix incorrect handling of selection in RenderText
A recent refactoring moved the selection color to the colors_ break list.
This caused selectiond to be broken (see crbug/1033194).

Regression in:
  https://chromium.googlesource.com/chromium/src/+/ee7c76abc9016939e4f112615d6eec5ee9b5cbd7

The first issue is that the selection is now an text attribute and needs to
be recomputed. This requires calling:
  OnLayoutTextAttributeChanged(false);

The second issue was an off-by-one when we apply the selection on the layout_colors_.

Bug: 1033194
Change-Id: Ib4287c3b2afa9e58e20c4e4399d413702debd89f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1965956
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724866}