Update visible item indices when ripping off a shelf icon
When a shelf icon is removed by dragging off the shelf, the function
to update |first_visible_index_| and |last_visible_index_| is called
after ShelfView::AnimateToIdealBounds() which relies on the correct
visible item indices. It is why crash happens.

This CL fixes the issue by triggering the update on visible item
indices before ShelfView::AnimateToIdealBounds(). A unit test is
created for this scenario.

Bug: 1032002
Change-Id: I5c51297dd111c180e49c3adbfcd85c55d251eb5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967925
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724867}