AW: ignore user consent for hardcoded URLs
Prior to this change, Safe Browsing CTS tests were untrustworthy because
they would early-return if the user had not consented to Safe Browsing
(worse yet, the CTS test runner tricked GMS into believing the user had
opted out, regardless of the setting value).

This changes WebView's logic to ignore user consent for
chrome://safe-browsing/* URLs. This is OK from a privacy perspective
because we handle these requests entirely in the open source chromium
repo (we hardcode these as "malicious"), and never send the URL to GMS.

This changes the public getSafeBrowsingEnabled() API. This used to fetch
whether or not Safe Browsing was fully enabled, querying for user
consent and exposing this to the app. This was originally done as a
workaround for CTS tests, so we could early-return rather than fail the
tests. Since this early-return logic is no longer necessary, this API
now only reports the app's opt-in/opt-out choice (via
WebSettings#setSafeBrowsingEnabled() or manifest tag).

Bug: 938538
Test: run_webview_instrumentation_test_apk -f SafeBrowsingTest#*
Test: $ out/Default/bin/system_webview_shell_apk launch \
Test: "chrome://safe-browsing/match?type=malware" (site is blocked)
Test: $ out/Default/bin/system_webview_shell_apk launch \
Test: "http://testsafebrowsing.appspot.com/" (site is not blocked)
Change-Id: I41981b0f18d975e3a8f7d5d1eb211daf130c5524
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1966584
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724893}