Fixed software input panel not being raised when switching between tabs.
This issue is actually related to a change that I made to enable
TSF to control software input panel(SIP) through input pane policy.
It regressed because of a bug in TSF which we are planning to service
downlevel. This patch is a temporary workaround until the servicing
fix is reached to earlier releases.

Basically this issue is related to an accessibility focus requirement for TSF
which it uses to process the policy change when focus is switched from one tab
to another. When we tap on the address bar, it notifies an accessibility focus
event by calling ::NotifyWinEvent with MSAA focus event which makes TSF to
evaluate the policy and thus raise the SIP.

Bug: 1022295

Change-Id: I1a0e7e2ebd8c3c3b8f14049039f42cd87eb3bf31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1955085
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Yohei Yukawa <yukawa@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724928}