Add browser type TYPE_APP_POPUP
App popup browser is a new type of browser which behaves like an app browser
(e.g. it should have an AppBrowserController) but looks like a popup (e.g. it
never has a tab strip).

The first user of it is going to be the terminal app, which needs to use an app
popup browser to host its setting page.

Bug: 999950
Change-Id: I2ad76a32fd681f0ee608139367c320bd2ce189c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1947613
Commit-Queue: Jason Lin <lxj@google.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: calamity <calamity@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#724998}