media/gpu: return Optional<Fourcc> from Fourcc::FromVideoPixelFormat()
Invalid Fourccs are dangerous by nature: they can be passed around just
like valid values and create bugs that manifest in a completely
different part of the code. To solve this, we plan to make it impossible
to create an invalid Fourcc instance.

This CL advances towards this goal by making FromVideoPixelFormat()
return Optional<Fourcc>. If a value exists, it is guaranteed to be
valid, and if the conversion failed then nullopt is returned. This
forces the caller to check the returned value and will crash immediately
if the Fourcc is access in an attempt to pass the value around. All
callers of this method are also updated accordingly, with the required
checks added (these checks should have been there anyway).

BUG=1031082
BUG=b:143498574
TEST=VDAtests passing on Hana.

Change-Id: I5916fb4cda423e6ed93f628402ea93c6e62364d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1957197
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725044}