[Payments] Rename VPA to UPI ID
In the Universal Payments Interface, VPA and UPI ID refer to the same thing.
We have chosen to refer to exclusively as UPI ID in all user-facing strings,
since this is the most common way websites refer to it. This CL makes sure
"VPA" does not appear on its own and unexplained in the codebase.

Bug: 986289
Change-Id: Idf225ce3e6cf9835da814cd03d8fe651fbbbd81d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967178
Commit-Queue: Christos Froussios <cfroussios@chromium.org>
Reviewed-by: Jared Saul <jsaul@google.com>
Cr-Commit-Position: refs/heads/master@{#725077}