Propagate may_cause_scrolling_ flag.
This flag indicates that a touch has moved beyond the slop region,
and its descendant is used inside TouchEventManager
DispatchTouchEventFromAccumulatdTouchPoints() to decide whether to
handle the event. It was being set from below the gesture recognizer
on the root_relative_event, but not on the original event.

Bug: b/144801646
Change-Id: I0aa0a966ce678af228aa904598ae471c8dfba06b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967222
Commit-Queue: Alex Sakhartchouk <alexst@chromium.org>
Reviewed-by: Albert Chaulk <achaulk@chromium.org>
Reviewed-by: Alex Sakhartchouk <alexst@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725117}