Always compare bookmark favicons in sync integration tests
The former urls_with_favicons_ optimization is not only unnecessary but
can also hide subtle bugs in the test setup, as was the case until
recent patches associated to the same bug.

Besides, the implementation did not handle well the case where multiple
bookmarks for the same URL were created, a situation that is
consistently hit when dealing with GUIDs (due to local GUID updates
incurring in creations and deletions).

Bug: 783774
Change-Id: Ic51fc2d2a4026650054407310957e3db76b20769
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/810770
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725150}