[PM] Add PerformanceManagerRegistry.
With this CL, PerformanceManagerTabHelper and RenderProcessUserData
must be created through PerformanceManagerRegistry. The
PerformanceManagerRegistry keeps track of objects it creates and
ensures that they are destroyed when it is itself destroyed. This
bookkeeping replaces the linked lists that were kept in global
variables before.

We plan to add a PerformanceManagerRegistry::Observer that will
allow code that maintains a decorator to know when there is a new
node for which the decorator should be maintained.

Bug: 1030399
Change-Id: Iaf1357a99f9d0eba96954d60338fa61c695c20b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1959378
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725164}