[cc/metrics] Change how main-damage is tracked.
It is possible that the args for begin-main-frames can be processed out
of order. For example, if the first BeginFrameArgs emits a
begin-main-frame, then it is possible that the next BeginFrameArgs can
be handled and responded as 'no damage from main' before the main-thread
responds to the begin-main-frame. So update the sequence-trackers to
handle this case correctly.

BUG=1021963

Change-Id: I28a838d7a1eadb0ead93d0ac7fbafc47a62af789
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967952
Reviewed-by: Xida Chen <xidachen@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725194}