Add Assistant integration tests using a fake S3 server.
These tests run as browser tests, use the real libassistant code and use a fake
s3 server to replay previously recorded conversations.

This allows the Assistant to run CQ tests that verify the basic interactions
are not broken.

More specifically, this uses a FakeS3Server library that was previously
created by the LibAssistant team. On the ChromeOS side I added hooks to
tell LibAssistant to use the FakeS3Server, and a test-fixture to start/stop
the fake S3 server during the browsertests.

BUG: b/141264108
TESTS: in target |browser_tests| with names |AssistantBrowserTest.*|

Change-Id: I89985c9e34eaa21a17eef69fa5b472fa39486d30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1894296
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Cr-Commit-Position: refs/heads/master@{#725214}