[HaTS] Fix dereferencing released web dialog delegate
When HaTS web dialog is closed, the widget released the dialog's
delegate before deleting the content view and its children. This means
an ObservableWebView might be alive while the |delegate_| it stores was
already released, which may cause crashes.

This CL changes ObservableWebView to store the address of the web
dialog's delegate which would be nullified at the same time as the
delegate's release.

bug:1033457

Change-Id: I4102032d56192fde64fcd94a02ea78d9014fe36e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967801
Commit-Queue: Wei Li <weili@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725228}