Add support for parsing shortcuts in Web App Manifest
This change is part 1 of 2 for adding App shortcut menu for PWAs.

As part of this change, when a shortcuts section is included in the Web
App Manifest, it gets parsed by the manifest parser.

It is then appropriately formatted to be accessed from the Browser
process using Mojo.

The shortcuts will be used during the PWA installation flow to add in
app shortcuts.( This will be a follow up CL)

Bug: 955497
Change-Id: I913141a0308a279271f862baae20132cae22aa68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1941051
Commit-Queue: Rahul Singh <rahsin@microsoft.com>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Rayan Kanso <rayankans@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725263}