webview: migrate missed profile data.
We missed migrating various data stores in the initial profile directory
migration, because many of them are not created until used. Run the
migration again and move all the missed files as well.

If any new data has been written to the new location already, we rename
it as a backup and then replace it with the old data, as we can't really
merge the data; it's more likely that the old state is useful as the
partial migration only just happened in the M79 rollout for most users.

Bug: 1033655
Change-Id: Ib2b7267454f1d9ae896f13a45e01683871c0b764
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1966383
Reviewed-by: Victor Costan <pwnall@chromium.org>
Commit-Queue: Richard Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725282}