Rewrite certificate tagging tool for the Windows Chrome updater (WIP)
This CL is a mechanical rewrite of the certificate tagging tool at
https://github.com/google/omaha/tree/master/common/certificate_tag

The symbol names and the structure of the code in this CL closely
resembles the original golang code.

The intention is to use boringssl for the x509 and asn1 parts of the code.

Bug: 1031734
Change-Id: Icdf08c078cdcfca30a7d2752556580a0bc559c62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967917
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725284}