[Reland]Update visible item indices when ripping off a shelf icon
The test in my previous CL (https://crrev.com/c/1967925) was flaky.
The reason is that: when explicitly destroying the shelf icon,
the shelf icon may be still in animation progress. In this CL,
stop the animation for the shelf icon to be removed if any before
destructing the shelf icon.

The description of the previous CL is listed as following:

When a shelf icon is removed by dragging off the shelf, the function
to update |first_visible_index_| and |last_visible_index_| is called
after ShelfView::AnimateToIdealBounds() which relies on the correct
visible item indices. It is why crash happens.

This CL fixes the issue by triggering the update on visible item
indices before ShelfView::AnimateToIdealBounds(). A unit test is
created for this scenario.

Bug: 1032002

Change-Id: I3e4d4ef6af3d940c444a301e8ce5a02a73c90aee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970291
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725288}