Only send domains to Safe Browsing if SBER.
Only send saved password domains if the user is SBER and not
incognito. Previously, there was a change to add it if the
saved password experiment is on which should not be the case.
This was where it was erroneously changed:
https://chromium-review.googlesource.com/c/chromium/src/+/1913477

Bug: 1034691
Change-Id: I90da552b323d8f00095a4403ca1eb8c7b3d47f9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970426
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Commit-Queue: Varun Khaneja <vakh@chromium.org>
Auto-Submit: Bettina Dea <bdea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725299}