[Structured metrics] Add an error delegate for reading prefs.
We use a JsonPrefStore to store keys and unsent logs, which needs to do
an initial read of the prefs during initialization. This might fail, so
this CL adds the plumbing to catch the errors.

A follow-up CL will add UMA metrics throughout this class, including in
the delegate.

Bug: 1016655
Change-Id: I121323b8ca57240982a30e8d6192bfa79c88705c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1969184
Commit-Queue: Tony Yeoman <tby@chromium.org>
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725372}