Fix graphemes handling in obscured text
This CL is fixing the behavior of RenderText with
emoji/graphemes when text is obscured.

The previous behavior was to display a replacement
bullet for each codepoint. This is incorrect for
complex unicode sequence.

As an example "e" + "acute" is two different codepoint
but it is displayed as a single glyph. For this case,
only one replacement character should be used. The same
apply for the complex emojis.

Bug: 1025561
Change-Id: I48cb0245aed793e026d353acae4008d00a535103
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967428
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725373}