MediaCapabilities: Use a double for mojo frames_per_sec
We use this mojo struct both in recording performance stats and in
reporting predictions. The former buckets fps on the render-side, so
an int type is acceptable. The latter buckets fps on the browser-side
so the conversion to int caused a loss in precision that could trigger
a DB miss for floating point frame-rates that would otherwise round up.

Bug: 1024399
Change-Id: I639badae4e761b960c0ae77c1947fa815cffb944
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1922914
Commit-Queue: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725378}