Remove unneeded includes from web_input_event.h
We are doing this as part of the removal of individual file dependencies from
chrome/browser/DEPS. The web_input_event.h file has many dependencies that could
prevent it from moving to public/common (as common cannot depend on other parts
of blink). We can remove most of them to the files which actually have
dependencies.

Bug: 1008303
Change-Id: I51839f97473363ca86132d548ad5882eb17e1013
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1969673
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Sergio Villar <svillar@igalia.com>
Cr-Commit-Position: refs/heads/master@{#725458}