skia: Crash when a discardable allocation fails.
A recent change (see crbug.com/983348) made the allocation of discardable memory
return nullptr in some cases, instead of crashing. Skia is not expecting this,
triggering a nullptr dereference.

This goes back to the previous behavior (crash in case of allocation failure),
while the proper fix is to make skia expect a nullptr return value. As a
consequence, this fix is temporary until upstream skia is updated.

Bug: 1034271, 983348
Change-Id: Idc3937898e897a19ca4e6793ab4e79626fe1b319
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1968985
Reviewed-by: Mike Reed <reed@google.com>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725462}