[css-pseudo] Recalculate viewport units in ::marker
If a marker uses viewport units (vw or vh), they should be recalculated
when the size of the viewport changes.

This patch updates node traversal methods to include ::marker, so that
it's taken into account when updating viewport units.

However, this is only fixed for markers with non-normal content, since
otherwise they aren't implemented as real pseudo-elements yet.

BUG=457718

TEST=external/wpt/css/css-pseudo/marker-content-014.html

The test still fails because it includes cases with 'content: normal'.

Change-Id: I7495dd5c4bbd84d6286635a63199bfcc663c62c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970015
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/master@{#725463}