[mediasource] Remove uses of DeprecatedByteLengthAsUnsigned
This CL replaces calls to DeprecatedByteLengthAsUnsigned by calls to
ByteLengthAsSizeT. Unfortunately the content of the ArrayBuffer is copied
into a WTF::Vector, whose maximum size is smaller than the maximum size
of an ArrayBuffer. Therefore I reject the incoming ArrayBuffer with a
kNotSupportedError if its size is too big.

Background: we prepare ArrayBuffers to be bigger than 4GB. Therefore we
changed the size field to size_t. Now we are changing all uses of
ByteLength to be able to deal with size_t, either by accepting a size_t,
or by throwing an exception if the size is too big.

R=chcunningham@chromium.org

Bug: chromium:1008840
Change-Id: I50ffdc4a53613ec121f7e58a6e0dc6874a936f12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1954520
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725487}