[webgl] Change length of TypedFlexibleArrayBufferView to size_t
This CL changes the length field of TypedFlexibleArrayBufferView from
unsigned to size_t. The reason is that internally
TypedFlexibleArrayBufferView already operates on size_t and currently
uses a checked_cast to convert to size_t. To avoid potential renderer
crashes I adjusted now ValidateUniformParameters to check the size of
the incoming TypedFlexibleArrayBufferView and throw an exception if
it is too big.

Background: we prepare ArrayBuffers to be bigger than 4GB. Therefore we
changed the size field to size_t. Now we are changing all uses of
ByteLength to be able to deal with size_t, either by accepting a size_t,
or by throwing an exception if the size is too big.

R=kbr@chromium.org

Bug: chromium:1008840
Change-Id: Iaaacc3672931d11df285876f042cce084c9976c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1963803
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725526}