viz: Simplify OverlayCandidateList by Moving Overlay Promotion Logic
This CL moves the overlay promotion logic that used to be in
OverlayCandidateList onto OverlayProcessorAndroid.

For Android pre-SurfaceControl code path, we keep track of promotion
hints for resources. It was kept on OverlayCandidateList, and it makes
implementation of OverlayCandidateList complicated.

OverlayProcessorAndroid now implements CheckOverlaySupport, which
helps keeping track of the promotion hints.

This is CL 13/? of overlay processor refactor.

R=rjkroege

Bug: 963597
Change-Id: Ia4ff00e1a1f92687505b8ec985ddceec004683b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1961082
Commit-Queue: weiliangc <weiliangc@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725527}