[cascade] Support css-logical properties
Some properties cascade separately, but use the same storage location
on ComputedStyle. For example, -webkit-writing-mode and writing-mode,
or css-logical properties and their physical counterparts. Such
situations can not be resolved cascade-time (at least not always),
hence we need a way to solve this apply-time.

There is currently a mechanism for dealing with writing-mode and zoom:
StyleCascade::Filter. The same mechanism can be used for css-logical
properties, however, since it would mean maintaining a large switch by
hand, this CL generates code for it instead.

Any property which has some other property map to it, is called a
"slot". Slots are marked as such in css_properties.json5 ('is_slot'),
and properties also specify which slots they map to ('slots').

See StyleCascade::Resolver::SetSlot for a more detailed explanation.

Note that this CL adds support for mapping a property to more than one
slot, which is currently unused, but necessary for supporting
-webkit-border-image in the future.

BUG=947004,985043

Change-Id: I22f75e32efa41dafbc2f5aaf4fafa0f7b1a53191
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967321
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725529}