[Autofill Assistant] Re-enable flaky test.
This fixes an issue where the test was failing in some runs due to elements not being visible on the screen. For now, checking the effective visibility seems to work. In integration tests (which this is not), we should instead scroll to the elements instead.

Bug: 1030217
Bug: b/146414854
Change-Id: Ia5222f5b66e30580d69dfa104287e207aa9d3ccc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970176
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/master@{#725532}