[iOS] Fix clipboard implementation not realizing it contains a URL
Usually, if the text in the pasteboard looks like a URL, the |.URL|
property returns non-nil even if the initial copy itself didn't
identify the text as a URL.

Unfortunately, this sometimes doesn't happen. The bug specifically
calls out syncing between macOS and the iOS simulator, but it could
also happen with Universal Clipboard (this was not checked and is pure
speculation).

To fix this, we can just fallback to checking whether the string forms
a URL via NSURL.

Bug: 1033935
Change-Id: I9aa8fe2cf14676158e0e38cfc95947a09dfd7c1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967326
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725546}