Throw RuntimeException when an unregistered SharedPrefs key is used.
Currently, when DCHECKS are enabled, SharedPreferencesManager asserts
when it receives a call passing a key not registered as "in use".

This CL changes that to RuntimeExceptions, which include a message and
are easier to understand.

Asserts were also easily confused with asserts performed by tests.

Change-Id: I9f410bc3d7181cfbd7e51e9101859ce44305b7b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967921
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Natalie Chouinard <chouinard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725568}