[PM] Add PerformanceManagerMainThreadObserver.
The observer interface provides a way to observe when a PageNode is
created for a WebContents on the main thread. This can be used by code
that maintains a PageNode decorator using main thread data, to know
when there there is a new PageNode for which the decorator should be
maintained.

There is no method to observe when a PageNode is destroyed. This is
because code can use WebContentsObserver::WebContentsDestroyed() to
know that a PageNode is destroyed as a result of WebContents
destruction, or use a custom hook prior to PerformanceManagerRegistry
destruction to know that all PageNodes are destroyed as a result of
PerformanceManagerRegistry destruction.

Bug: 1030399
Change-Id: I998de0055ea6d64fdddf4a144673ea5b47dccd9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967434
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725607}