Prepare ProfileBrowserTest for HistoryServiceUsesTaskScheduler activation.
The feature will soon be flipped to ENABLED_BY_DEFAULT. This will make
HistoryServiceBackend's task runner dispatch to the ThreadPool instead
of a dedicated base::Thread. The tests need to be prepared to stay
functional when that happens.

This is done as the last mile of the chrome-wide effort to coalesce as
much work as possible to the ThreadPool.

Bug: 661143
Change-Id: Ia8cbe38a5155a947de5cd04a41fb0d9134379050
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1965151
Auto-Submit: Oliver Li <olivierli@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Oliver Li <olivierli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725614}