cros: Only consider the feature value (!switch) for the hotseat
There are conflicts between the switch and the feature (for example,
kukui has the switch added to its overlay, yet PMs may want to disable
the feature for demos).

Remove the switch for now. It was only used to enable for Kukui, but
this is no longer required because Kukui are dogfooder only, and all
dogfooders have the feature enabled via finch.

The actual switch will be removed in a later CL. This was made to be
merge-friendly.

Bug: 1034626
Change-Id: Ia2a47a3875def2b1c434deabfdbf512ae01baa5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970142
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Vladislav Kaznacheev <kaznacheev@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725639}