[metrics] Support observing many DidFinishNavigation() in FirstWebContentsProfiler.
A fuzzer suggests that it is possible to have the FirstWebContentsProfiler
created after 2 DidStartNavigation() have been dispatched but before
the 2 corresponding DidFinishNavigation() have been dispatched. This
currently results in a DCHECK failure.

With this CL, we invoke DumpWithoutCrashing() and we record
"abandon new navigation" to Startup.FirstWebContents.FinishReason instead.
The goal is to determine whether the problem actually happens in the wild.

Bug: 1027947
Change-Id: I3e651b6f08f9001ba970ccc5c756e45e50a5df6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1947894
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Auto-Submit: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725655}