[Android] Delay deferred startup till after initial navigation
This CL changes CustomTabActivity to run
ChromeActivity#postDeferredStartupIfNeeded() only when the tab
is loaded in the background.

The is OK because a tab cannot be reparented to a
CustomTabActivity from another window.

This CL fixes a bug where ChromeActivity#onDeferredStartup() would run
prior to the initial tab navigation since:
- CustomTabActivityTabController#onFinishNativeInitialization() which
creates the tab is called after ChromeActivity#onStartWithNative()
due to the relative ordering of
NativeInitializationController#onNativeInitializationComplete() and
ActivityLifecycleDispatcherImpl#dispatchNativeInitializationFinished().
- TwaIntentHandlingStrategy#handleIntent() navigates the page
asynchronously.

BUG=1032291
TEST=CustomTabDeferredStartupTest.*

Change-Id: I26d33e0ce6361aceeceecf32a17c48adad265d86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1960477
Commit-Queue: Peter Kotwicz <pkotwicz@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725658}