Increase the Watchdog V2 timeout for Mac and Android
From the watchdog metrics, ~50% of the detected GPU hangs would start to
response if we increase the timeout from 15 seconds to 17 seconds for Mac
and from 10 seconds to 15 seconds for Android.

This CL also changes the starting time tick to the beginning of the
OnWatchdogTimeout() function when we count the extra 60 seconds in
OnWatchdogTimeout().

Bug: 949839
Change-Id: Ia3b87eb0e76d37bc5ab7767f21149bda215d421c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1971051
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725699}