Invert sRGB image detection behavior
All DiscardableImageMap paths should call .Generate(), which fills
in whether images are sRGB or not. However, it has an early exit
if there are no images.

The problem is all_images_are_srgb_ is set to false by default. So
the early exit treats no-image cases as some-images-are-wide-color.

This CL inverts the behavior. It renames all_images_are_srgb_ to
contains_non_srgb_images_, allowing the defaulted-to-false value
to continue to work.

Bug: 955158
Change-Id: I5b1cd49d7ab428b970bf096fdbfb5cbbd41728a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970850
Reviewed-by: Khushal <khushalsagar@chromium.org>
Commit-Queue: Chris Blume <cblume@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725705}