Remove inertial latching override
I tried this out and the TODO is correct, inertial scrolls don't break
the gesture stream so this is a no-op because during a GSB the
CurrentlyScrollingNode will always be nullptr.

This is true of both in renderer scrolling as well as ui::Views.

Bug: 1016229
Change-Id: I7696fcbe3bfb45cd26ec68850570e10f253f821b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1954456
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725712}