Reland "[throughput metrics] Fix handling when no frames are produced."
If a CompositorFrame is not produced in response to a begin-frame, then
it can be either because the compositor is attempting to recover from
high-latency, or because there was no update to the frame (i.e. there
was no damage). The latter cases should not cause the throughput metrics
to go down. This is currently handled for some cases (e.g. when the
main-frame is aborted due to no damage), but there are other cases where
the trackers are not notified, and thus they report lower throughput
metrics. To correct this, use the newly introduced FrameSkippedReason to
notify the trackers when a frame is not produced because there were no
damage.

BUG=1021963

Change-Id: Ibe2f37cae47c652a7a1170c2dec24a01e2e03ae1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962991
Reviewed-by: Xida Chen <xidachen@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725789}