Add active user changed function to ARC windows.
Add ActiveUserChanged to AppServiceAppWindowArcTracker to make sure
that we created items for all apps, not only which have a window, which
is the same implementation as the current code:
https://cs.chromium.org/chromium/src/chrome/browser/ui/ash/launcher/arc_app_window_launcher_controller.cc?q=ActiveUserChanged&l=114&dr=CSs

For those task ids having windows,
AppServiceAppWindowLauncherController::ActiveUserChanged can handle them.

Using OnWindowVisibilityChanged interface, the same as
ArcAppWindowLauncherController, because MultiUserWindowManagerHelper is
used to manage the window when switching between users, and in
OnWindowVisibilityChanging, the window is still invisible, so if in
OnWindowVisibilityChanging interface, after switching back to the
primary user, those windows are lost.

Set previous window full screen mode when the window is inactive,
following the current ARC windows implementation:
https://cs.chromium.org/chromium/src/chrome/browser/ui/ash/launcher/arc_app_window_launcher_controller.cc?q=SetFullscreenMode&l=359&dr=C

BUG=1011235

Change-Id: Ie891e1079807e44b4c6a801ff6dfd568a5f1f3f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1973510
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725871}