Auto-dismiss TouchToFill if UI can't be shown.
Before this CL, TouchToFill assumed that BottomSheetController::requestShowContent always succeeds.
This is not the case if there is a BottomSheetContent currently being shown with higher priority.
In practice, this can happen when autofill_assistant (a high-priority bottom sheet content)
attempts to log-in on a website.

The problem surfaces because TouchToFill prevents the soft keyboard from being shown. Thus, in an
autofill_assistant flow, users would tap the login text input on the website, but the keyboard
would not appear.

After this CL TouchToFill UI is automatically dismissed if
BottomSheetController::requestShowContent fails.

Bug: b/145278496
Change-Id: Ifede88d513fc95692872ad001e538340f367f55f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1968993
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/master@{#725902}