[webgl] Replace uses of deprecatedByteLengthAsUnsigned
This CL replaces calls to deprecatedByteLengthAsUnsigned by calls to
byteLengthAsSizeT. Unfortunately the current implementation cannot deal
with huge ArrayBuffers yet. Therefore I reject the incoming
ArrayBuffers with a SynthesizeGLError if its size is too big.

Background: we prepare ArrayBuffers to be bigger than 4GB. Therefore we
changed the size field to size_t. Now we are changing all uses of
ByteLength to be able to deal with size_t, either by accepting a size_t,
or by throwing an exception if the size is too big.

R=kbr@chromium.org

Bug: chromium:1008840
Change-Id: I96972b25473f03d0af80fa3d73e62bf41a49ecf2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970172
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#725907}