[scroll-snap] Handle fractional offsets for directional snap strategies
When the currently snapped area's offset is a float number it would not
get ignored when exectuing a directional scroll. This happened because
the scrollable area's current scroll offset would be rounded, while the
snap area's position offset was not.

Now when DirectionStrategy compares the snap area's offset to the
the current offset for determining whether to snap, it will round the
difference to an int.

Bug: 959859,989738
Change-Id: I41f79182e923ea7104f678142d9fc4e971a3b881
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1969052
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Yi Gu <yigu@chromium.org>
Commit-Queue: Kaan Alsan <alsan@google.com>
Cr-Commit-Position: refs/heads/master@{#725992}