[iOS] Handle InfoBar replacement.
Updates InfobarOverlayRequestCancelHandler to handle InfoBar
replacements.  When the corresponding request's InfoBar has been
replaced, the cancel handler creates an OverlayRequest for the
replacement and inserts it in the queue after the replaced InfoBar's
request.  Then, the existing request is cancelled, removing it from
the queue.

Bug: 1030357
Change-Id: Idff53a277c4115adb6075017db5aed4c4e3afac7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1960810
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726021}