DevTools: do not retain HostResolver in TCPDeviceProvider
After we switched to using mojo host resolver in
https://chromium-review.googlesource.com/c/chromium/src/+/1199848/
we need to make sure HostResolver is destroyed on the same thread
that it was created on.
Ensuring this with ref-counted DeviceProvider would require
a helper class, so just re-create it each time we need to
resolve a host for simplicity.
The overhead should be small considering to other things we
do (like not caching host resolution lately etc).

Bug: 1028292, 1025369, 921608

Change-Id: I6f8d2e663f0ac872964e1725e261232493bbbfe5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1972781
Commit-Queue: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726022}