[Benchmarking] Filter timing data to only present current stories.
This means that if you abridge a benchmark or remove stories from
a benchmark then that information will be reflected in
http://go/chrome-benchmark-schedules as soon as you check in your
change.

Also, add a --filter-only option so that we can keep this up to
date without needing to run a slow bigquery query each time.
Also, add validation to presubmit so that this stays up to date.
Also, make the output deterministic by sorting it.


Bug: 1030840
Change-Id: I2cda6110a029953559874707852230693d0ac7b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1973232
Reviewed-by: John Chen <johnchen@chromium.org>
Auto-Submit: Caleb Rouleau <crouleau@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726030}