Add manual dependency on Vulkan
libvulkan is dlopen()'ed, so a manual dependency is necessary. It's
available on all RPM-based distros, so a basic dependency is added. But
it isn't available everywhere on Debian-based systems, so it's added as
a recommended dependency instead. If the system doesn't have libvulkan,
OpenGL will be used as a fallback.

R=thestig
BUG=1034606
CC=penghuang

Change-Id: Ibb0a1a94cc163c93a6428b1b408f6b364f9b6bb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1973042
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Auto-Submit: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726036}