[nfc] Replace uses of DeprecatedByteLengthAsUnsigned
This CL replaces calls to DeprecatedByteLengthAsUnsigned by calls to
ByteLengthAsSizeT. Unfortunately the current implementation cannot deal
with huge ArrayBuffers yet. Therefore I reject the incoming
ArrayBuffers with a RangeError if its size is too big.

Background: we prepare ArrayBuffers to be bigger than 4GB. Therefore we
changed the size field to size_t. Now we are changing all uses of
ByteLength to be able to deal with size_t, either by accepting a size_t,
or by throwing an exception if the size is too big.

R=reillyg@chromium.org

Bug: chromium:1008840
Change-Id: Ia0af7fb8c57ec522cf8cd0963d6e0dd3fad9dd92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1973910
Reviewed-by: Rijubrata Bhaumik <rijubrata.bhaumik@intel.com>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726041}