[iOS] Reset OverlayRequestMediator.request when it's cancelled.
A destruction callback is used to reset the reference to a
mediator's OverlayRequest upon destruction.  When a request
is cancelled, it is immediately removed from its queue and
destroyed, but it's possible for the UI to outlive the request
while it's being dismissed.  This CL adds early returns to
prevent accessing the request after destruction.

Bug: none
Change-Id: I256f17e35572c168aecf8c714601ceb17153ceb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1963439
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Auto-Submit: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726063}