Fix active descendant output
An active descendant is always set on the currently focused node. However, in some editables, ChromeVox can have its range on a descendant of the focus. With the new eventing model in AutomationAXTreeWrapper, we never get the active descendant change.

Fix this by intorducing a FocusAutomationHandler which groups all events for which we want to observe on the currently focused node.

Bug: 1035509
Test: manual.

Change-Id: I94566a7ea34df788f40d8baf2789aae8809db888
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1972500
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726066}