Desktop PWAs: ToolbarButtonContainer creates and updates its buttons
WebAppFrameToolbarView::ToolbarButtonContainer now creates the views
that it contains, and listens for updates that affect its members.

The outer class WebAppFrameToolbarView is now much simpler,
as it does little more that create the two containers for the controls
on the left and right of the toolbar, manage their layout and
implement the ToolbarButtonProvider interface.

Design doc:
https://docs.google.com/document/d/1gOU3F8Zk1VGTPZQIr7Bq8gultJ7vg5AZJuujmRuG9To/edit?usp=sharing

TBR=tapted@chromium.org

Bug: 1007151
Change-Id: I0d31dee8a9a19a1e6939c1c2b0f611a377dafc43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1966791
Commit-Queue: Eric Willigers <ericwilligers@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726076}