Make Mac10.X Tests builders into thin testers
This will free up a small amount of Mac hardware which can be
used elsewhere. It also creates a template for other OSes with
builder/tester splits to follow.

We know that this works because the mac-osxbeta-rel CI builder
already does this (as can be seen by the smaller diff in its
generated config).

Bug: 1003062
Change-Id: Iecdb68404dd3ae16eabfc657981f0e82478a040a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1974510
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726077}