Fix altgr for rule-based input methods.
We accidentally broke altgr for rule-based inputs in [1] because we
looked at the "altgr" attribute of the keypress info, which is only
filled by the VK and not the PK.

For the PK, AltRight is always treated as AltGr, so we add that logic
back in.

We add a test to prevent this from happening again.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1837312

Bug: 1035145
Change-Id: I64d3d56b08c8cd2a82c4617f63a3c6f34444f36a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1971638
Reviewed-by: Shu Chen <shuchen@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726113}