[Content] Introduce WebContentsObserver::OnDidAddMessageToConsole()
Introduce WebContentsObserver::OnDidAddMessageToConsole() to (partially)
replace WebContentsDelegate::DidAddMessageToConsole(). This is useful
for consumers that are interested in monitoring the console messages
added, but not changing the behavior of adding the message. It will also
be used in the future to enable easier testing, since it won't require
modifying the WebContents' existing delegate.

Update WebViewGuest<> to use the new method, instead of overriding the
WebContentsDelegate version.

Bug: 1034150
Change-Id: Ia0d3e9aaf72a57c55244d508b4c49d02644970f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1968152
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726121}