rule-based: Use native code for resetting engine.
<This change is hidden behind a flag>

As part of migrating PK rule-based IMEs to native code, this CL migrates
reset to native code path.

We port RuleBasedController#onCompositionCancelled in
chos/pk/rulebasedcontroller.js.

Note that we need onReset to call through to the extension, as it has
side effects that may affect the extension when it is running (e.g.
when VK is up).

Note that the JS code has already been guarded by the runtime flag.

Bug: 1009903
Change-Id: Ic1efc191fb7c963916b58cea4e8055d1d8ae6743
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1968776
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Shu Chen <shuchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726145}