Reland "Make ScriptedAnimationController a ContextLifecycleStateObserver"
This removes the need for custom pause/unpause logic for
ScriptedAnimationController in Document.

Also, construct ScriptedAnimationController in the Document
constructor, rather than lazily. It will get initialized on the
first frame, so we're not saving much by deferring its initialization.
This simplifies a bunch of logic around its handling in Document, and
guarantees that its lifetime will be predictable for leak detection.

TBR=creis@chromium.org

Bug: 1029822
Change-Id: Ia30db6b7dac25c834152ae546364d77ff089fca3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970259
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726206}