Implement Infrastructure for AR and VR permissions
This change begins implementing the infrastructure for AR and VR
permissions so that WebXr can switch to using Permissions instead of a
custom consent flow.  Note that neither permission is intended to be
exposed to blink at this time, and neither permission is expected to
be called, therefore all calls to these permissions are currently
blocked while development is occurring.

Bug: 1033592
Change-Id: Ia51f9df7f66c42d2b5fad17ef9e9206b7bedd799
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1965974
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726239}