[Network Settings] Fix quick settings mobile data toggle animation
This CL fixes issue with "Mobile Data" toggle animating to the on
position when opening the menu instead of appearing in the on state
immediately. The issue was caused because the mobile data header view
always initializes with the toggle off. The correct value is updated
immediately after initialization. But this code path uses the toggle's
AnimateIsOn method that starts an animation. Fixed this by setting
correct initial state for mobile header view.

Fixed: 1020722
Change-Id: I8c17ce28b09b2506221fcbcbf472f80d44f67006
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1975111
Commit-Queue: Azeem Arshad <azeemarshad@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726241}