[bfcache] Fix crash in ServiceWorkerVersion::RestoreControlleeFromBfcacheMap
This CL intends to fix crash happening in ServiceWorkerVersion::
RestoreControlleeFromBackForwardCacheMap.
The crash hit the check !base::Contains(controllee_map_, client_uuid),
so it seems that the client exists in controlle_map_ when it is
supposed to be in bfcached_controllee_map_.

This could happen when a controllee was not controlled by a version
and started to be controlled in bfcache.
This change adds code to make sure that UpdateController checks
bfcache status of the client.

Bug: 1021718, 1034141
Change-Id: I1e054cfb43b4ce0e6df33fbbf0e1ec239faf1803
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1959238
Commit-Queue: Yuzu Saijo <yuzus@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726265}