[NFC] Check if array buffer has 32-bit compatible size
This CL replaces calls to deprecatedByteLengthAsUnsigned by calls to
byteLengthAsSizeT. If the byte length is larger then a RangeError is
thrown to avoid problems in passing a wrongly cast size to the
underlying stack.

Background: we prepare ArrayBuffers to be bigger than 4GB. Therefore we
changed the size field to size_t. Now we are changing all uses of
ByteLength to be able to deal with size_t, either by accepting a size_t,
or by throwing an exception if the size is too big.

R=reillyg@chromium.org

Bug: chromium:1008840
Change-Id: Id713040b55384d2b427d97ffdd66d4963eb614ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1964610
Reviewed-by: Rijubrata Bhaumik <rijubrata.bhaumik@intel.com>
Commit-Queue: Emanuel Ziegler <ecmziegler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726318}