Avoid unnecessary eglGetCurrentSurface call
GLContext::IsCurrent can result in driver calls which have a non-zero
cost, as evidenced by recent sampling profiles. This happens frequently
with virtual contexts on Android in the following circumstance:

- {GL,Skia}Renderer gets a PassthroughImageTransportSurface |wrapper|
  for the display, which is a GLSurfaceAdapter wrapping |surface|

- GLSurfacePresentationHelper::OnMakeCurrent picks up the |surface|
  and subsequently uses it to make a context current in
  GLSurfacePresentationHelper::CheckPendingFrames

As we draw frames, we alternately make current with |wrapper| and
|surface|, leading to unnecessary driver calls in
GLContext::MakeVirtuallyCurrent.

This CLs eliminates that by so that both the |wrapper| and |surface|
are considered current whenever either one of them are.

Change-Id: I153aff272b966c42a5066fb2016c9d353a1b3f87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1974316
Reviewed-by: Peng Huang <penghuang@chromium.org>
Commit-Queue: Jonathan Backer <backer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726378}