Add whole-script-confusable characters for Hebrew
This CL adds a list of characters that look like Latin characters. A
domain consisting only of these characters and not having the .il ccTLD
will be converted to punycode, as is done for Cyrillic Latin-lookalikes.

The impact of this change is minimal on existing domains with <5 domains
being affected.

Bug: 722167
Change-Id: I0fd9c13331b10845958612f58fb6353d5fdb6c88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1975054
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726475}