Internet: VPN Config: Elim policy indicator for provider type
When we switched to using an unmanaged enum for vpn.type (which is an
intrinsic / immutable property of a configured VPN), it broke the
policy indicator for the type selector, which assumes that the
'property' property will be a managed type (e.g. ManagedString).

Since the provider type can not be changed, it is not useful to
show the policy indicator for policy managed networks (the provider
type must be provided by the policy). Therefore we should just leave
'property' undefined for that network-config-select element.

Bug: 1035581
Change-Id: I47704adff49fa1063eab5e93a8654645a2ec1893
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1976492
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726476}