[Benchmarking] Disable dawn_perf_tests on mac.
The upload for these is hanging on mac. My guess is that the root cause
is the same as for windows where it hits a 403. However, the result is
much worse for mac because it waits for over an hour before it hits the
overall timeout. The longterm solution is to convert dawn_perf_tests
to histograms.

Bug: 1035590
Change-Id: I9070d04e527cc4387ba793963de78fa73141a0a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1976020
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Commit-Queue: Austin Eng <enga@chromium.org>
Auto-Submit: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726496}