Write and readback pixels to a texture in OpenXR tests
The old implementation was sort of a hack to be able to figure out
what values to appropriately set based on the configured IPD that
the device set (since one test set specific colors and also had
a different IPD than the rest).The new implementation fix this
by reading back from the texture. So the hack is no longer needed.

These should instead store the
pixels to a texture when the frame is submitted, and then read back
from the texture.
Tests affected as of 8/28/2019:
WebXrVrOpenXrBrowserTest.TestPresentationPixels
WebXrVrOpenXrBrowserTest.TestPresentationPoses
WebXrVrOpenXrBrowserTest.TestLocationIndicatorWhenUserAskedToPrompt
WebXrVrOpenXrBrowserTest,TestMultipleInitialIndicators_OneDeviceAllowed

Fixed: 986621
Change-Id: I18421f1adc136bb4e1799026af486cfe85fbff05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970257
Commit-Queue: Zheng Qin <zheqi@microsoft.com>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726515}