Rename set_glyph_width(...) to SetObscuredGlyphSpacing(...)
The glyph_spacing(...) is only valid for Obscured text.
As an example, this cannot work with arabic characters.
It will break words in piece.

The concept is only used for password. Let rename to
ObscuredGlyphSpacing to make it obvious it is only working
when the text is obscured.

Also removing the DCHECK failing when the text is not
obscured. It is assumed now that the width is not used
for un-obscured text. That DCHECK was often failing while
fuzzing.

TBR=tbarzic@chromium.org

Bug: 1025561
Change-Id: I1700b4047bec9090d5c6719381923f14d9cd8e73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1976013
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726532}