New plumbing of incoming sync trusted vault keys
The API is refactored to isolate the flow for incoming encryption keys.
This better separates the API from other functions that are intended to
be used by actual UI (e.g. user-entered passphrases), and sets the
foundation for triggering fetches initiated by the TrustedVaultClient
implementation itself (useful on Android).

There is a behavioral difference as a result of rewriting the logic
in SyncServiceCrypto, which is arguably desirable: trusted vault keys
are not propagated to the sync engine (and hence the Nigori keybag
unless there is an actual need to do so, which excludes, most
importantly, users that are not in the relevant passphrase type).

Bug: 1012659
Change-Id: I009e02ebd0370eb98e95c50368299b0bf947cfb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1973823
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726535}