Prepare IndependentOTRProfileManagerTest for HistoryServiceUsesTaskScheduler activation.
This test had brittle behavior that was hidden by the way HistoryService used to be coded.

The feature will soon be flipped to ENABLED_BY_DEFAULT. This will make
HistoryServiceBackend's task runner dispatch to the ThreadPool instead
of a dedicated base::Thread. The tests need to be prepared to stay
functional when that happens.

This is done as the last mile of the chrome-wide effort to coalesce as
much work as possible to the ThreadPool.

Bug: 661143
Change-Id: I4f3bae453c93454eb8cecc6cd3046cfd07958bfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1976150
Auto-Submit: Oliver Li <olivierli@chromium.org>
Commit-Queue: Brandon Tolsch <btolsch@chromium.org>
Reviewed-by: Brandon Tolsch <btolsch@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726536}