On Mac, gracefully exit GPU process on GPU disconnect/disconnect request
Using the "SafeEjectGPU" tool, we can see that when the "responsible"
process (the browser process) has a GPUEjectPolicy of 'wait', any
subordinates using the GPU (i.e. the GPU process) get the 'rwait'
GPUEjectPolicy: "Eject actions apply to the responsible process, who in
turn deals with subordinates to eliminate their ejecting eGPU
references." [man 8 SafeEjectGPU]

Empirically, the browser does not relaunch. Once the GPU process exits,
it appears that the browser process is no longer considered to be using
the GPU, so it "succeeds" the "wait".

Documentation:
https://developer.apple.com/documentation/metal/gpu_selection_in_macos/handling_external_gpu_additions_and_removals

Change-Id: Ia44c2ed864d1bdbb52d82ad72e14cbfcb4e32108
Bug: 1035619
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1974856
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726602}