Isolate cert decisions for non-default storage partition
Currently, all cert error decisions (i.e., when a user clicks through
a certificate error) are stored in ContentSettings, associated with a
Profile. This means that if a user clicks through a certificate error
in a <webview> in a Chrome App, that decisions propagates to normal
browsing. Persisting decisions within a <webview> is undesirable on
its own (because there's no UI to remind the user that they've done so
and to allow them to revoke the decision), and it's especially
undesirable for that decision to affect normal browsing. Therefore,
this CL isolates cert error decisions by storage partition. Decisions
made for the default storage partition are remembered in the normal
way (ContentSettings for the Profile); decisions for other storage
partitions are remembered in memory only.

Bug: 639173
Change-Id: If1cca181c80f8d07f5411fbb0d3707cf3755c5a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1974698
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726607}