AW: disable testOnSafeBrowsingHit CTS test
The WebViewClientTest#testOnSafeBrowsingHit() is flaky due to a subtle
timing change when we switched to Network Service. This disables the
test for now, while we look for a fix.

This test was added in OMR1 (which we don't run CTS tests for in
chromium waterfall), so we don't need to do anything for earlier API
levels. The flaky part of the test was split out into
testOnSafeBrowsingHitBackToSafety in Q, which we're already skipping in
the JSON config.

Bug: 1006953
Test: android_webview/tools/run_cts.py (on P device)
Change-Id: I2c3e197484b1678722573ee052aed83150148417
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1975156
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Andrew Luo <aluo@chromium.org>
Reviewed-by: Andrew Luo <aluo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726610}