XPath: Simplify had_type_conversion_error handling
EvaluationContext had |bool had_type_conversion_error| field, and we had to copy
the field manually whenever we cloned an EvaluationContext object to propagate
an error to the top-level EvaluationContext.
This CL changes the field to a mutable reference to bool, and it is shared
between cloned EvaluationContext objects.

Change-Id: I5f7e71a25337cfaf413ca98fccc9d3f16a04dba4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1975472
Commit-Queue: Kent Tamura <tkent@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726634}