Remove animator implementations from UiElementContainerView.
This is part of a multi-CL effort to clean up the Assistant UI element
modeling/rendering pipeline. This CL simply moves the ElementAnimator
implementations that were previously in UiElementContaienrView into
AssistantUiElement. In the future, UiElementContainerView will have no
concept of AssistantTextElementView or AssistantCardElementView, it
will only understand AssistantUiElementView.

More details of the effort in the bug.

Bug: b:145003512
Change-Id: I0228556d5837f5466ca8fd5f9fc59c6ffd1239a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1954666
Commit-Queue: David Black <dmblack@google.com>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726636}