ui: gbm_wrapper: lower severity of log message
On ChromeOS, not being able to create a GBM BO for a buffer prevents us
from using hardware overlays with this buffer, but is not a severe error
in itself. However, currently a warning is unconditionally displayed in
both debug and release builds, which tends to accumulate with time and
suggests that something is seriously going wrong:

gbm_wrapper.cc(286)] Failed to create GBM BO, YV12, 320x240, flags: 0x2021; gbm_device_is_format_supported() = 0

Turn this LOG(WARNING) into a DVLOG(2), so only developers explicitly
interested in understanding what exactly happens at the GBM level are
notified of this failure.

BUG=b:143323148
TEST=VDA unittest builds and runs on Hana/Kukui in Debug mode. Confirmed
     that the log did not appear in regular runs.

Change-Id: Ide75e80455be790aa1bcadaa875945bbc42c35e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1974299
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726737}