Adjust background in themes with a tabstrip stroke.
Due to legacy behavior, themes with background images and a stroke
around the active tab shift their background images down a pixel, which
causes a small visual discrepancy against the background image of the
toolbar and bookmarks bar.

With this change, the background images of the toolbar and bookmarks bar
are adjusted down by stroke width (1 DIP) in cases where there is both a
background image and a stroke, restoring the original behavior in this
case.

Bug: 1035090
Change-Id: Idd028a73d48390604e96f2d64e0d729efab30474
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1975301
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726741}