Post actions from AnimatingLayoutManager
This replaces immediate running with task posting to prevent the tasks
from running from the same stack as ::Layout. This solves a couple of
problems:

* Prevents Layout -> ScheduleLayout due to delayed actions.
* Makes the call stacks similar for actions that run immediately and
  ones that are delayed, so users won't accidentally rely on the action
  being run immediately just because an animation isn't running in most
  cases.
* Allows running views::Widget::Show as a delayed action, which modifies
  the layer stack and therefore cannot be done during Layout as a result
  of damaged layer rects.

The last was discovered as a crash while trying to pop out an Extension
and show a views::Widget once the extension's View is visible. This case
would crash occasionally due to iterator invalidation if the animation
finishes during layout.

Bug: None
Change-Id: I15928f7d53e00d20c25388d574f89d5c72a7ae34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1976292
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726829}