Remove redundant EnsureLayout calls in RenderText tests
Remove the redundant call to EnsureLayout(...) used by
RenderText unittests.

These calls were used to ensures that the layout structures
(internals data structure) are updated. The public API
is taking care of calling EnsureLayout when needed. There
is no need to call it.

For the protected API methods (only for testing), the calls
are lifted into their wrapper function in the tests.

The goal is to ease code reading and have a better code coverage.
This is ensuring the layout is performed on-demand by the
public API.

Bug: 1025561
Change-Id: Ifc8da08fa1bc1b55db33ec500bb4a1bbaae6e53d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1972760
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726846}