Removing ForceRGBA from serializer and canvascolorparams
Serializer: The enums have been renamed, but they have the very same
functionality. Previously the kRGBA8 was effectively handled internally
in CanvasColorParams as if it was Native, and the kForceRGBA8 was being
handled as kRGBA8.

This CL fixes the internal behavior of CanvasColorParams without needing
a forceRGBA.


Bug: 1021986
Change-Id: I32517a0455f41f05113304b95bb8a7d6fd593229
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1947546
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Auto-Submit: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726906}