Fix in app shelf style for Assistant launcher UI
With the Assistant launcher UI and new shelf design feature turned on,
if the Assistant is triggered in tablet mode the shelf style will be
changed back to "home shelf".

Added an additional check in ShelfConfig for the Assistant state to
handle this case.

Bug: 1026727
Test: Manual Test
Change-Id: Ie9ca0980583117e5b17d21fa9257f91f3ebe5647
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1975114
Commit-Queue: Yue Li <updowndota@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726929}