Add AssistantWebViewFactory for use in Ash.
Currently Assistant uses Content Service for embedding WebContents in
Ash. Moving forward, we're migrating back to WebView as it better
handles focus and ChromeVox related corner cases.

To work around dependency restrictions in Ash, this CL adds a new:
- AssistantWebViewFactory, and
- AssistantWebView2

which are implemented in Browser.

AssistantWebView2 is named as it is because there already exists an
AssistantWebView in Ash, but the existing AssistantWebView is soon
to be removed and the name will free up.

AssistantWebView2 abstracts away the concept of WebContents so that
the APIs invoked on AssistantWebView2 are delegated to the underlying
WebView or WebContents as needed.

Bug: b:146351046
Change-Id: Ica75899ca8d5426333b0501a7c4e34698bab4b19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1974460
Commit-Queue: David Black <dmblack@google.com>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727030}