Don't default use_thin_lto to true in official ARC builds.
This will cause it to be false in ARC-bits of chrome-branded linux-chromeos builds.

This should reduce the maintenance burden of Chrome's clang since there's
a non-zero cost of ensuring the official linux-chromeos build config builds
with thin lto.

Disabling it should be safe since linux-chromeos is a developer-only build,
and actual builds of Chrome OS explicitly set it:
https://chromium.googlesource.com/chromiumos/overlays/chromiumos-overlay/+/3c6956bccc6e77a38395f35919d43c502f6c5cdf/chromeos-base/chromeos-chrome/chromeos-chrome-9999.ebuild#271

Bug: 1033839
Change-Id: I8d44482a6cea956afba808480aeed69c64e5c6e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967602
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727093}