plugin_vm: add support for uninstallation
This CL adds a button to the plugin vm subpage of chromeos' settings.
When clicked this button will cause the newly introduced
PluginVmManager::UninstallPluginVm method to run.
This method queries the dispatcher for the vm's state, then tries to
stop the vm if it is running, then finally removes the vm.
It is currently not supported to invoke any of
PluginVmManager::(UninstallPluginVm|LaunchPluginVm|StopVm) at the same
time. Doing so may cause one or all of the running operations to fail.

Bug: 1023256
Change-Id: I3861b14f76e5b6f479ad6dc2eb5e16167e3dc37f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1973511
Commit-Queue: Julian Watson <juwa@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Timothy Loh <timloh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727104}