Use HTTP_NOTSAMESITE_ORIGIN for split cache Web Platform Test
Currently HTTP_REMOTE_ORIGIN is used as the third-party origin, but
this does not work for split cache experiments using eTLD+1 as the
additional key (i.e. when UseRegistrableDomainInNetworkIsolationKey is
enabled). This is due to HTTP_REMOTE_ORIGIN simply prepending 'ww1.' to
the HTTP_ORIGIN, which results in the same registrable domain.

Bug: 1035508
Change-Id: Ib42bb30491bb451a47b80914a48e252a9c93d10f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1978037
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727110}