MacViews: Fix NativeWidgetMac::SetSize
The argument to the function NativeWidgetMac::SetBounds is sometimes
in screen coordinates and sometimes in parent coordinates. The
distinction was made deep in the bowels of NativeWidgetNSWindowBridge.

There was a bug wherein NativeWidgetMac::SetSize was implemented by
calling NativeWidgetMac::SetBounds, and always using bounds in screen
coordinates. Then, deep in the bowels of NativeWidgetNSWindowBridge,
we would re-offset the the bounds.

The result was that the position would shift at every call to SetSize.

Make all structures below NativeWidgetMac only ever use screen
coordinates. Add the function ConvertBoundsToScreenIfNeeded to do (or
not do) the conversion as needed.

Bug: 1036703, 957362
Change-Id: I44ed4f2ed8e0f544d488298ca49135f0c1b363d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1972761
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727113}