Remove unnecessary check from LayerTitleCatch#fetchFaviconForTab.
LayerTitleCatch#fetchFaviconForTab checked if last used profile has
OffTheRecord (OTR) profiles or not and did not proceed if it had.
The check was done to prevent usage of OTR profiles for getting local
favicon images.
This was wrong and not needed because:
 1- Profile.getLastUsedProfile() always returns a regular profile.
 2- hasOffTheRecordProfile() covers OTR profiles and normal profiles
    that have an OTR profile.
 3- getLocalFaviconImageForURL handles OTR profiles as well by
    redirecting the request to a read only copy of the original profile.

The misleading code is removed.

Bug: 1035768
Change-Id: I11f1b629c8cc6d57db873f8beab61795d4942157
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1977998
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727122}