[cc/metrics] Defer resetting state when frames pause.
It is possible that a frame sequence can pause while in the middle
of processing a frame. When this happens, the state-reset should be
deferred until processing the frame ends, so that other notifications
for the frame (e.g. 'no damage') can be handled appropriately.

BUG=1035382, 1021963

Change-Id: Iacc85227bb0bd8171e8f2000459909fde87284c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1979952
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727143}