Prepare to enable Hotseat flag in M-81 - Part 2
When enabling the hotseat, back gesture is enabled as well. Touch event
falling in the area of back gesture is consumed by back gesture handler.
As a result, no gesture event will be generated for it by gesture
provider and it causes the test failure. This CL ensures that touch
location is outside of the back gesture area.

TBR=xiyuan@chromium.org

Bug: 1032702
Change-Id: Ia560025ecf755be0858f0698933a306e5a1ac169
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1981063
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727266}