Remove MockRenderProcessHost::OverrideURLLoaderFactory().
This function was confusing because there is also SetNetworkFactory(),
which overrode it. The only callsites were to give a fake factory in
order to prevent the test from crashing with a null factory. This CL
instead makes all MockRenderProcesHosts have a default factory, and
removes the callsites.

Bug: 1001435
Change-Id: Ib0fd1a3ab8ba7987a1c0b5aac8c5eba9311f7878
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1981237
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727343}