Uses AXVirtualView to make the TreeView accessible
This patch creates an accessible object for every item in the tree and
adds both the permanent accessibility attributes which don't change dynamically,
such as role and text direction, as well as the
dynamic attributes that are refreshed every time they are requested,
such as the item's location and visibility status.

This patch fires appropriate events for all tree changes.

Some accessibility actions are also added but
more work is needed to implement all supported actions on AXVirtualView.

Furthermore, some methods were added to AXPlatformNodeDelegate that allow the accessibility
tree to be dumped to a string for debugging and testing purposes.

R=dmazzoni@chromium.org, aleventhal@chromium.org

Bug: 811277
Change-Id: I73821a425e90415a61328b5fa2a8592a45d1e233
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1917061
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Auto-Submit: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727388}