[Extensions] Mark ResultCatcher::GetNextResult() as WARN_UNUSED_RESULT
Callers of ResultCatcher::GetNextResult() should always, always check
the result. Typically, a return value of `false` means that something
went wrong (because it is returned when a test fails), so it should
be explicitly checked. (Note that we don't just assert in the
ResultCatcher itself, because some tests do expect to fail).

Update the one call site that didn't already check the result.

Bug: None
Change-Id: I690b794b0c1ac07cb0a30b5d3078ffdb258195a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980752
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727473}