Add class for snapshotting the accessibility tree.
For exporting tagged PDFs, we need the ability to
take a snapshot of the accessibility tree for a
RenderFrame, and to have the accessibility context
stay alive long enough that IDs in the snapshot,
and IDs sent to the printing system, are consistent.

We already had code to do AX tree snapshotting,
so this patch is just some refactoring to make it
a class instead, and to have it return a public
type rather than a content-specific type.

This will be used in the follow-up change that
implements support for tagged PDFs:
http://crrev.com/c/1970742

Bug: 607777
Change-Id: I5163561c2cabcb2dd3579ce9d7a2de8ac96106d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970579
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727483}