Changed color id used in bubble border for menu scroll view containers.
Changed menu scroll view container to have its bubble border use the
color id for background color instead of the color id for border color
when determining background color.

Following the change made in patch
  https://chromium-review.googlesource.com/c/chromium/src/+/1949105
the color of context menus in ChromeOS shifted to a darker shade of
grey. This is contrary to how the context menu should be colored
according to UX guidance.

The code path for the bubble border in the menu scroll view container
is only exercised in the following ash components:
  - ash/app_list/views/search_result_tile_item_view.cc
  - ash/app_list/views/app_list_item_view.cc
  - ash/root_window_controller.cc
  - ash/shelf/shelf_view.cc

The scope of this change is thus isolated to ChromeOS, restoring
the context menu color to its intended values on the platform.

Bug: 1035600
Change-Id: I115df0b2781cba696713d96ead77ad50888538a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980768
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727493}