AW: repopulate the flags UI if Activity is recreated
This repopulates the Spinners in the Flags UI when the Activity is
recreated, to ensure the Activity and Service/ContentProvider have
consistent state. Without this change, the flags UI would have an empty
map of flag overrides, and would clear most of the Service's flags when
the user toggles the first Spinner.

This moves some the ContentProvider communication logic to a separate
bridge class to be shared by both the dev UI and the embedded WebView
browser process.

TBR=torne@chromium.org

Fixed: 1036538
Test: Manual - toggle flags in the Activity, swipe away the Activity, \
Test: and relaunch to see the UI is restored correctly
Change-Id: Ife42a0633707c9d59e5678483a95f4267fbbe739
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980857
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Shimi Zhang <ctzsm@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727523}