[Extensions] Clean up and extend PageAction Commands tests
Refactor PageAction commands tests to be less reliant on waiting for
success, and introduce a couple of helper methods. This allows us to
exercise that the page action is *not* triggered in certain cases
(whereas with the way the tests were structured before, there was no
way to do so).

Also add a new test that leverages to test that the page action is not
triggered if the action is not visible on the given tab.

This is also a precursor for getting overflowed page action keyboard
commands to work properly.

Bug: 1037685
Change-Id: I2e5bec90bf3a5f1aeb00e115b94a82863a68e092
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1981192
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727524}