WebLayer: Satisfy Blink's need for InstalledAppProvider impl
As crbug.com/1037884 details, Blink currently implicitly requires that
the embedder of //content supply an InstalledAppProvider implementation
on Android (on all other platforms a default no-op implementation is
provided by //content, but on Android a production implementation is
provided in //chrome).

That bug will hopefully be fixed soon, but in the meantime this CL
provides a stub implementation in WebLayer on Android so that pages
that trigger this flow don't crash in debug mode. When that bug is
fixed, I will remove this workaround.

Bug: 1037884
Change-Id: I7eb17431a8b43704f0ce3056babf4a5f95b6a6b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980743
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727538}