Fix annotated_text from AXPosition<>::ToString
ToString converted the text to UTF8 before wrapping the focused
character with '<' and '>'. This can cause confusion if a developer
uses standard output logging with an AXPosition because |text_offset_|
is relative to the UTF16 character position, not UTF8, which could
result in wrapping the wrong character or wrapping a portion of a
UTF16 character.

Converting the final |annotated_text| to UTF8 instead fixes this.

Bug: 1037921
Change-Id: I1c4ed0f8847197ac2d97d6abbda5899a479ae03c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1981636
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Adam Ettenberger <adettenb@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#727583}