[Extensions] De-flake BrowserActionApiTest.CloseBackgroundPage
De-flake BrowserActionApiTest.CloseBackgroundPage. The test could fail
if the C++ side clicked on the browser action icon before the extension
registered a listener for browserAction.onClicked(). If this happened,
the event would be dropped, and the test would time out waiting for the
extension host to close.

Fix this by waiting for the extension to send a "ready" message when it
has finished setting up listeners.

The test now passes 100 runs on the bots with no flakes (before, it
would flake multiple times on 100 runs).

Bug: 1035075

Change-Id: Ic56ff1026b68db8cd5296098f719b660ad7a4918
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1981835
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727593}