ChromeVox: Add functionality to announce capital letters.
This change adds functionality to announce capital letters. A <select>
has been added to the ChromeVox options page to allow users to
choose how they want capital letters to be signaled. The two options
are increasing pitch and announcing 'cap' before the letter.

This patch does not change the phonetic hint behavior for capital
letters. That means even if ChromeVox announces 'cap <letter>',
it will still follow-up with the phonetic hint for the letter.

Bug: 809094
Change-Id: I05193df0f281b5f1e1fc7ebb8cccc761d61a6ea9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1953645
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727607}