Fix JavaScriptDialogTest.DismissalCausePromptCancelButton
This test fails when DialogDelegate::GetDialogWidgetInitParams is
changed to set views::Widget::InitParams::child (that is, to make the
dialog not be a top-level widget).

The reason this test fails is that, during teardown, focus has not
been removed from the dialog, so the line
  DCHECK(!!focused_now || !new_text_input_client);
in NativeWidgetMacNSWindowHost::OnDidChangeFocus fails (because
the TextInputClient has not been reset, as would have been done
by un-focusing the dialog).

The reason that this didn't fail before the patch is that
Widget::CloseWithReason would un-focus the Widget's views before
deleting destroying them. But only for top-level widgets, for no
clear reason. Other places where similar things should be done
(e.g, Widget::NotifyNativeViewHierarchyWillChange), do this for all
widgets (not just top-level).

Add a helper function Widget::ClearFocusFromWidget and change all
places that are trying to remove focus from the widget to use that
call.

Bug: 957362
Change-Id: I9905055fc12a54441e7f918094ba2ec0a61f6e15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1981265
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727616}