SuperSize: Set no-cache headers for hosted version
Turns out google storage sets a one hour public cache header by default,
which means a bad push results in the page being broken for an hour.

This leaves the default header on for .size files, but sets all other
files (which are tiny) to no-cache.

Note: the .wasm file is big, and should probably be cached, but that is handled
by the service worker already.

Bug: None
Change-Id: Ia3c03e4533c37338b924d5ace3f7b5fd4ac8d6f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1981263
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Eric Stevenson <estevenson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727618}