[CrOS Settings] Remove backspace icon fill transition on pin input.
Currently, there is an unnecessary blink of enabled 'Backspace' button.
This regression occured because a transition time was introduced in
https://chromium-review.googlesource.com/c/chromium/src/+/1898479 to
iron-icon, causing a momentary blink.  This CL removes the transition
time to the fill so there is no blink in the beginning on the backspace
button.  Doing so does not cause any blatant changes to other actions
applied on backspace button (i.e deleting a pin number when the button
is enabled).

Fixed: 1036266
Change-Id: I539ec8b2f775c868d560f8fa2df76299bf8a7a89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1983344
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Reviewed-by: Dan Beam <dbeam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727671}