[Benchmarking] Add gtest perf tests to OBBS.
All Telemetry tests are currently run as part of One Buildbot Step (OBBS):
performance_test_suite. However, gtest perf tests are run separately. This causes
problems since these tests have to finish
before we can start the Telemetry tests. So even thought these tests are
fast, they block a bunch of
other tests.

This changelist adds these test to the shard maps so that
they can be run alongside the Telemetry
tests as part of the same shards. This allows us to
distribute load across all the shards evenly.

As a first test, this changelist only does
the switch for Nexus 5. If this works, I will follow it
up with changes for all the other builders.

Please note that this changelist should not be cleanly
reverted because it changes which shard the various telemetry
tests run on. This change can cause false-positive improvements
and regressions as a result of the bot that the stories are running
on changing. If this change is reverted, then that will cause
another set of these false positives as the stories change back to
running on an earlier bot.

Bug: 921353
Change-Id: I0dabb053ba21b05cc9e97f540e23e59829e3e44c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1983326
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727678}