[CastRunner] Add support for HEADLESS CastRunner components.
Allows for the creation of headless Cast components without Scenic
Views.

The semantics of CreateView() are modified under these circumstances
to handle the view token EventPair as a simple means of scoping the
lifetime of a HEADLESS rendering session.

Refactors the constructor logic for WebContentRunner.

Bug: 1025045,1029204
Change-Id: I348c550d2d9a634112ef2ff3e40b9373012d3e60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1931764
Commit-Queue: Kevin Marshall <kmarshall@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727718}