[WebLayer] Initial integration of captive portal interstitial
This CL provides an initial integration of the captive portal
interstitial within WebLayer. In this initial integration, the captive
portal interstitial is brought up when the user has hit an SSL error
and the underlying Android captive portal detection reports that the
user is behind a captive portal.

The focus of this initial implementation is to make the basic use case
available and develop integration tests of this use case. Functionality
of Chrome's embedding of CaptivePortalBlockingPage beyond this basic use
case is not ported over to WebLayer but instead is targeted for the next
stage of sharing of //chrome's embedding of security interstitials with
WebLayer (crbug.com/1030692). We have highlighted those touchpoints in
this initial bringup, which will help to guide the design of that
sharing.

Bug: 1025059
Change-Id: I10354585f0ca4deff835abce76ed7051faaa853d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980747
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727736}