WebLayer: Hook up autofill detection and communication to browser
This CL brings up renderer-side autofill detection in WebLayer and
communication of that autofill detection to the browser. On the browser
side, we inject an AutofillProvider into the autofill system, to which
the autofill component delegates the browser-side UI flow.
This setup anticipates the usage of an AutofillProvider implementation
on Android that integrates with the system-level autofill support;
bringup of that implementation will occur in a followup CL.

This CL has several moving parts:
- Creation of the renderer-side AutofillAgent that parses documents for
  autofill purposes
- Creation of the browser-side ContentAutofillDriver that drives the
  browser-side flow and communicates with AutofillAgent
- Hookup of the communication between AutofillAgent and
  ContentAutofillDriver
- A minimal AutofillClient implementation that implements only the
  touchpoints used within the WebLayer flow of Autofill and is passed
  to ContentAutofillDriver on creation
- A stub AutofillProvider implementation that enables cross-platform
  browsertests and is passed to ContentAutofillDriver on creation
- Testing infrastructure for such a browsertest and the browsertest
  itself

There as of yet no user-visible behavioral changes, so no manual testing
instructions with this CL.

The followup CL will introduce the Android production implementation via
bringing up the Java-side system-level integration, connecting it to a
C++ AutofillProviderAndroid instance, and passing *that* instance to
ContentAutofillDriver.

Bug: 1025617
Change-Id: I3aa30179c78d23e527fe24b43472f744d4393fae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1965660
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727739}