[Signin][Android] Remove persistent account list from PO2TSDelegate
This CL revises how ProfileOAuth2TokenServiceDelegate is keeping track
of the list of accounts that have been advertised to the native code by
OnRefreshTokenAvailable. Before, the persistent storage on the Java side
was used. Now, a simple instance variable on the C++ side is used
instead.

Bug: 1006772
Change-Id: I4bb7ddec56f387d332abdef6d1b5c43add617207
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1983225
Reviewed-by: Pâris Meuleman <pmeuleman@chromium.org>
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727768}