[Benchmarking] Hotfix for windows failures.
performance test suite is failing here
https://ci.chromium.org/p/chromium/builders/ci/Win7%20Tests%20%28dbg%29%281%29/80918

the issue is that we assume that the file will be named .exe, but it isn't
sometimes. I revert the changes that were made to the shard map in this change
and turn off the unittests that depended on the old version.

R=adithyas@chromium.org

Bug: 1024767
Change-Id: I0a732dd7c3c35a096ea40b65d43f246513a8913a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1984175
Reviewed-by: Adithya Srinivasan <adithyas@chromium.org>
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727772}