[Refactor] Remove ScrollEndImpl
This CL is a first step in making scroll offset animations work sensibly
with the scroll latching logic of the compositor.

In this CL we stop calling ScrollEndImpl from ScrollAnimated,
ScrollAnimatedBegin and ScrollOffsetAnimationFinished so that animations
don't clear the latched scroller (the |CurrentlyScrollingNode|). We
inline the only remaining callsite into ScrollEnd.

As a result of this change, we can now simplify the state of several
scroll element id tracking variables. Since we no longer clear the latch
at the end of a scroll animation, we can set these variables at a scroll
begin and assume they will survive unchanged until a ScrollEnd is
received. These are:

|last_scroller_element_id_| is used to determine which scroller to send
a scrollend event to.

|scroll_animating_latched_element_id_| was previously used to relatch to
the same scroller after an animation ends and another GSU arrives. This
is now likely redundant and will be removed in a future CL.

Similarly, |scroll_animating_overscroll_target_element_id_| was needed
to know which element to send overscroll events to since they're sent
when no scrolling is possible and thus the latch is cleared. Since we
no longer clear the latch this can also probably be removed.

For now, this CL keeps the above variables but simplifies their
lifetimes. We also keep the re-latching logic in ScrollAnimated but
ensure it re-latches to the same node if we're already latched. A future
CL will clean this up so that ScrollAnimated can assume a node is
already latched.

Bug: 996537,940508
Change-Id: I84d5e74ab174043576e829f85b60f269ae21966b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1981066
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727775}