[webauthn] Refactor GetEphemeralKey
Pull GetEphemeralKey down into FidoDeviceAuthenticator and call it from
the methods that require a key agreement instead of exposing it as part
of the public API. This way we avoid repeating the key agreement logic
on fido handlers.

A minor downside of doing this is we now get a new ephemeral key when
using the PIN we just created on MakeCredential calls.

Fixed: 1037617
Change-Id: Ifa2563f1c7ea22c73bed75dc2c407ade95827a21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980957
Auto-Submit: Nina Satragno <nsatragno@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727797}