Add IconLabelBubbleView Delegate method for getting adjacent foreground color.
By default, the ink drop color now maps to this color, which makes sense (most
views want to use the same text/ink drop color).

This doesn't yet help remove duplicated code, since the goal of this CL is to
not change any behavior (yet), and thus the affected bits mostly just move
from one class to another.  As with the parent CL, the real motivation is to
break the fix for bug 1033922 into easily-reviewable pieces, and in addition
to enabling that fix, this will also enable some later cleanup.

Bug: 1033922
Change-Id: Ie68d1cd0ed71f0e52c9b2d290c2ed839d5c2631c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980478
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#727818}