Reset initial virtual time when disabling for testing
There are a few tests which call PageScheduler::EnableVirtualTime at
the start of the test, then PageScheduler::DisableVirtualTimeForTesting
at the end of the test, to reset state. Disabling virtual time currently
does not reset the MainThreadSchedulerImpl's initial_virtual_time*
fields. If a subsequent test then calls EnableVirtualTime, since the
MainThreadScheduler is shared across tests, we can end up in a situation
where the ThreadLoadTracker sees inconsistent timestamps and DCHECKs.

This happens due to the fact that there is at least one task that runs
and records timing between DisableVirtualTimeForTesting and
EnableVirtualTime - the ExitRunLoop task from
blink::test::RunPendingTasks, which is invoked from SimTest::TearDown.
When the next test comes along and does EnableVirtualTime, the
MainThreadSchedulerImpl sees that there was a previous
initial_virtual_time and proceeds to apply the
BaseTimeOverridePolicy::OVERRIDE policy to the
AutoAdvancingVirtualTimeDomain along with a timestamp in the past. When
the next task's timings are recorded by ThreadLoadTracker, it sees that
time has gone backwards and we DCHECK.

This change makes it such that the initial_virtual_time fields are reset
when DisableVirtualTimeForTesting is called. This will cause them to be
re-populated on the next test's call to EnableVirtualTime and time won't
go backwards.

Bug: 845197
Change-Id: I756949db2b4dbbdd1b7b2b35b92fab80a4b20d14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1984875
Reviewed-by: Alex Clarke <alexclarke@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#728001}