Don't run WebLayer browsertests using an incognito profile on Android.
We weren't passing a profile name so it was always in-memory. I'm using this in a follow-up cl to force all downloads in weblayer_browsertests to go to this temp profile directory instead of filling up the default download folder. It also makes sense that WebLayer Android browser tests match desktop, and non-WebLayer as well.

Bug: 1025603
Change-Id: I1f4fe20e77e6405b97365a9dfa30917fd1c5d486
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1984761
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Commit-Queue: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728010}