Communicate color space to ui::Compositor
We don't send the color space to the browser's ui::Compositor on macOS.
This has been broken on macOS for an unknown period of time. It didn't
actually make a difference because
- that only set the raster space for browser UI, which is sRGB anyway
- the renderer got the correct color space for raster via
  RenderWidgetHostImpl::GetScreenInfo
- we would draw using CoreAnimation (not GLRenderer) >99.9% of the
  time, so the ui::Compositor's GLRenderer's color space didn't matter
Because of rounded corner rects, we hit the GLRenderer more often,
causing this bug be visible.

Change RecyclableCompositorMac::UpdateSurface to take a required
color space argument, ensuring that all callers specify a value.

Bug: 1038723
Change-Id: Icde934718371802356c5c26277a628361705009f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1985127
Reviewed-by: Eric Karl <ericrk@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728067}