[iOS] Create OverlayResponseSupport
This CL creates OverlayResponseSupport, which allows objects to specify
support for a subset of OverlayResponse types.  This will allow for the
removal of OverlayDispatchCallbackStorage, which uses templated container
objects to associate callbacks with their required response info types.

Support for a variety of responses can be specified by constructing
OverlayResponseSupports with a vector of support instances to aggregate.
OverlayRequestSupport was also updated to use a similar constructor for
consistency, and AggregateOverlayRequestSupport was deleted.

This CL also adds OverlayResponseInfo, a template to be specialized for
OverlayUserData used to configure OverlayResponses.  This template
declares a static ResponseSupport() that returns a support for responses
created with that info type.

Bug: none
Change-Id: I7d1cd13f9a05511c3dc95739055347284e7f7060
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1984464
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Auto-Submit: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728089}