Ensure GetTextForTooltipAndAccessibleName() doesn't return empty.
This is the cause of the attached bug. Animations in the toolbar mean
that some buttons which are being hidden are visible briefly as they
fade out, which previously created the situation that some tests which
closed the browser window while the fade-out animation was happening
would fail.

This solves the problem by having a default tooltip during fade-out.

Bug: 1038567
Change-Id: Ia24f3493e8b5d062b5ab67e2a8950e2fcf58b49e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1985171
Commit-Queue: Dana Fried <dfried@chromium.org>
Auto-Submit: Dana Fried <dfried@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728093}