Remove GLSurface::ColorSpace
The union of all platforms' capabilities is will eventually be about
as big as gfx::ColorSpace, so we should use that instead.

On Windows this reduces the two functions to produce a DXGI_COLOR_SPACE
from gl::ColorSpaceUtils and gfx::ColorSpaceWin to just the one in
gfx::ColorSpaceWin.

The next patch in this sequence will be to serialize gfx::ColorSpace
instead of sending it as a GLenum in glResizeCHROMIUM.

Bug: 1037954
Change-Id: I7a244543c1eae2826b1a3910b1dc9a860838644b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1984621
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728139}