[css-pseudo] Fix style adjustment for forced-inside ::marker
StyleAdjuster adjusts the style for ::marker depending on whether it's
inside or outside. This is mostly defined by the 'list-style-position'
property, however, Chromium forces some markers to be inside.

This patch is a follow-up of r726310, it fixes a problem with the logic
for detecting these forced-inside markers, it was not working at all.

BUG=457718

TEST=fast/lists/list-marker-computed-display.html

The test is internal since forced-inside markers are a non-standard
quirk.

Change-Id: Ibfda8459defb036eabdb2cd01990fde000bae533
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1985668
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/master@{#728173}