De-flake CrowdDenyPreloadDataTest.*.
The tests used to work with a global singleton CrowdDenyPreloadData
instance. This is wrong semantically, and had also been leading to
use-after-frees because this global instance used to a reference to a
ThreadPool instance that would be created (and destroyed) with the first
ever test case run.

Bug: 1038902
Change-Id: Ie6e560de3a779e7d9621bbe5a53587a76a019826
TBR: andypaicu@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1983176
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Commit-Queue: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728194}