Remove CompositorFrame handling in render_host
This CL starts to remove code in content/browser/renderer_host/ that
existed to receive CompositorFrames and submit them to the display
compositor running on the browser UI thread. The VizDisplayCompositor
feature has launched so this code is no longer used.

This CL is a first pass to delete code that doesn't have require
significant changes to tests.

TBR=sky@chromium.org

Bug: 759653, 812322, 1038683
Change-Id: I7bf29f57e7418eb8d3a16b86169f2b30bb9fbf80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1978997
Reviewed-by: Saman Sami <samans@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728209}