[CrOS Browser Settings] Fix Learn More link in browser language settings
Currently, there is an extra "Learn More" in the browser language
settings.  This error was introduced when replacing "Learn More" links
with settings-localized-link in os settings.  Since browser settings
is not using the settings-localized-link element, but shares the
same strings used by os settings, a new string that does not contain
"Learn More" in the sentence must be introduced.

Fixed: 1038878
Change-Id: Ib450994561e9ea9908543c60377448120361a6c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1984377
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728289}