Add translation context for on/off strings in ash UI and settings
The "desc" attribute is not sufficient to provide context to the
translators because identical strings will be collapsed in the
translation console.

Add additional string constants, eliminate the shared "toggleOn"
and "toggleOff" strings, and provide the "meaning" attribute in
the GRD file to disambiguate.

See bugs for details, or
https://www.chromium.org/developers/design-documents/ui-localization

Bug: 1031145
Bug: b/145301624
Change-Id: I3daf288f5a8a0dbf9ebd65575e7ff77144880ac2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986862
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728293}