Media: plumb latencyHint to audio renderer.
The bulk of the change is to AudioRendererAlgorithm, which has
historically been the decider of when the queue of decoded output
is "full" such that we can begin playback.

This CL adds a new method, IsAdequateForPlayback(), which may return
true well before IsFull() when a latencyHint has been set that lowers
latency from the default of 200ms. In such cases, we will still
allow the decoded output to queue beyond the point of being
"adequate" until we reach "full".

When a latencyHint is set that increases latency vs the default, both
the "adequate" and "full" thresholds will increase to accommodate the
higher value.


Bug: 1018908
Change-Id: I96fd19a7670630c23fb1e05076ac80c6486dba03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1959326
Auto-Submit: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Sergey Volk <servolk@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: Chrome Cunningham <chcunningham@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728311}