Fix mouse access for the android omnibox dropdown.
The issue is that the MotionEvent (ACTION_BUTTON_PRESS) was bleeding through
the view hierarchy and being consumed by the underlying web contents, which
would request focus away from the omnibox triggering an ACTION_CANCEL to
be dispatched.

The solution is to have the suggestion list consume any previously unconsumed
mouse click events to ensure they do not bleed through to the sibling yet
lower in the Z-axis views.

BUG=968414

Change-Id: Ic29114419708a1bf4d7b4ccb03b7cd0144d6bcd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1987073
Reviewed-by: Ender <ender@google.com>
Commit-Queue: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728360}