Expose disable reason custodianApprovalRequired to chrome://extensions
This disable reason will be used in a downstream CL to determine
whether to show a button or a toggle for enabling extensions for child
users. This disable reason will also force-disable the extension
enable toggles when the extension state is REQUIRE_APPROVAL. This
force-disabling will prevent a regression to crbug/1003014.

Bug: 1030958
Change-Id: I63657d01db8c388906579b3cae77d9d78dd487de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1974877
Commit-Queue: Toby Huang <tobyhuang@chromium.org>
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728367}