Componentize //android_webview files that integrate with system autofill
In preparation for implementing integration with system autofill in
WebLayer, this CL componentizes //android_webview's infrastructure that
does the heavy lifting for this integration in WebView. The
infrastructure is self-contained and will be reusable out of the box.

This CL also suitably generalizes the context of the files:
- Renames them appropriately
- Removes/generalizes most references to WebView (a few need to be left
  for legacy compatibility; these are documented).

We note that it would be possible to consider merging
AutofillProviderImpl.java into AutofillProvider.java in a followup; as
far as we know, there are no other implementations of the
AutofillProvider interface in the codebase. We did not do that here as
(a) it would have introduced a large amount of noise into the CL, and
(b) it is debatable whether it is better, as there is some value to
having separation between the tight AutofillProvider.java interface
exposed to Chromium's autofill and the sprawling AutofillProviderImpl
implementation necessary to integrate with the system AutofillService.

This CL also adds michaelbai@ as OWNER of the relevant files for
integration with system autofill on Android under
//components/autofill/android.

TBR=michaelbai@chromium.org

Bug: 1025617
Change-Id: I4ca7bbd03983d8b46558db1cd0f2d632d36f787c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980738
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728479}