Rename MinidumpUploader to MinidumpUploadJob
This paves the way to unfork the weblayer MinidumpUploader, which splits
MinidumpUploadCallable into two classes, one of which is responsible
for uploading a single minidump. Unfortunately that causes a naming
collision; resolve it in favour of renaming the minidump_uploader class,
which is responsible for managing the upload of multiple minidumps as
a JobService job.

Bug: 1029724
Test: run_webview_instrumentation_test_apk -f \
Test: 'org.chromium.android_webview.test.services.MinidumpUploadJobTest#*'
Test: run_chrome_public_test_apk -f \
Test: 'org.chromium.components.minidump_uploader.*#*'
Change-Id: Ib82dd334e77ee0e60a3c35a0d2f8bd38cfb6ed3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1975881
Commit-Queue: Tobias Sargeant <tobiasjs@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Anna Malova <amalova@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728489}