[ios][eg2] Fix flakiness in testRequestMobileSitePropagatesToNextNavigations
This increases the timeout used when waiting for a web state after switching
to desktop or mobile mode from 10 seconds to 15 seconds.

The previous value (10 seconds) had been chosen since that was sufficient to
fix flake locally when this test was converted to EG2.

Bug: 1037663
Change-Id: I37c56cb30ab1158b36c95d88a53b2e45ea009cbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1988488
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728594}