storage: Make IndexedDBDispatcherHost not observe RPH
This code was added as a part of idb mojo-ification in this patch:
https://codereview.chromium.org/2727733004#msg2

From comments in that code from three years ago, this cleanup is only a
potential optimization to close pending transactions that would be
aborted anyway.  It seems fine in the short term to remove this
optimization.

This does bring up some future questions about how the renderer/browser
will communicate with IndexedDBDispatcherHost (which will almost
live in storage service), and in that world it will be easy to hook this
back up if needed by adding a mojo disconnect handler.

Bug: 1015212
Change-Id: Ie02f7c7a925b2296d7403fa542852e912e329246
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986915
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Auto-Submit: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728597}