[iOS] Create SavePasswordInfobarBannerOverlayMediator.
This mediator uses the information exposed in the request config to set
up a banner consumer.  All other functionality is implemented via the
OverlayRequestMediator superclass.

In addition, this CL removes InfobarBannerOverlayRequestConfig, which
was previously added as auxiliary data for the save passwords config.
However, this was implementing UI logic in the model layer, whereas
this should actually live in a mediator.  Now, the banner coordinator
class is shared, but the banner UI can be configured by specialized
mediator subclasses that perform the UI setup logic.

Bug: 1030357
Change-Id: I2ff9aae6f8618ca65774a045631e78402dbcecec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1978558
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728624}