[Paint Preview] Compositing API
Introduces a public API for compositing of paint previews. This should
complete the API surface required for implementers of the base service
to both capture and playback paint previews.

I intend to follow this up with a change that moves the parts that
should be depended upon externally to a separate target in a public/
directory within components/paint_preview/browser/

Bug: 1030269
Change-Id: I9275ffc6e00267e84a2eaf60b1c07d65f6f4b053
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1966073
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Mehran Mahmoudi <mahmoudi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728651}