Allow UIA text navigation in all empty objects
We had this bug with UIA that was preventing us of navigating with word
or character inside of empty objects devoid of a static text child.
e.g.
<div>
    <label for="input1">Some text </label>
    <input id="input1">
</div>

In this case, when reading by word/character using UIA, we would never
jump inside the empty input since it doesn't have a text child in
the AXTree. The same would happen with any object embedded in text
that is devoid of static text child, such as images and text areas.

While I investigated this issue, I stumbled upon the concept of
"embedded object characters" which is used for IA2 and ATK.
I couldn't use it the same way it is used in those other accessibility
APIs, but the idea behind this CL is the same.

This CL introduces the concept of empty object replacement characters.
When an empty object is replaced (only with UIA), the character is
treated the exact same way as an ordinary character in a text node
would be. The expectations are the same as if there really was an
"empty object replacement character" (the character used for empty
object, U+FFFC) inside this node.

For now, all leaf nodes that aren't text are considered as empty
objects expect the document node and ignored nodes. I don't think this
will be causing issues, but I can be wrong. If it is, we'll scope down
the condition inside of IsEmptyObjectReplacedByCharacter().

Known issue: crbug.com/1028843

Bug: 928948
Change-Id: Ic2e7d37495615dce19258630a39aaa2f5b7c88f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1937858
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#728700}