Fix an issue that a new download can be started more than once
For download service, it could issue multiple DownloadUrl() calls to
the download system that have the same GUID at the same time.
This CL fixes the issue by checking the GUIDs that are already pending
or are already persisted in the DB, so that the same GUIDs cannot be
used to create a new download.

BUG=1029746

Change-Id: Ifc17167454b6eee4132fffa08b7d2666571cce35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980758
Commit-Queue: Min Qin <qinmin@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728709}