Prepare SearchResultRankerTest for HistoryServiceUsesTaskScheduler activation.
The feature will soon be flipped to ENABLED_BY_DEFAULT. This will make
HistoryServiceBackend's task runner dispatch to the ThreadPool instead
of a dedicated base::Thread. The tests need to be prepared to stay
functional when that happens. This includes advancing mock time now that
the history backend tasks happen within the mocked time environment.

This is done as the last mile of the chrome-wide effort to coalesce as
much work as possible to the ThreadPool.

Bug: 661143
Change-Id: Id3cd0dbf77bc18bd5a407174af3b7d9b743a5db5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1979153
Auto-Submit: Oliver Li <olivierli@chromium.org>
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Oliver Li <olivierli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728719}