Use synthetic gesture mechanism for TouchpadPinchBrowserTests
The tests which use an extra pinch gesture to test that a previous
gesture did not cause a scale change could themselves not cause a scale
change, which caused the tests to flake. This was likely due to not
properly waiting for the gesture to finish. We now have these tests use
the synthetic gesture mechanism to send events rather than directly
creating and sending events, as the former is more robust about handling
gesture completion.

Also, it looks like a method for synchronizing threads was renamed to
|WaitForJavascriptExecution|. This name is misleading, so we give it a
more descriptive one and comment its use. We also add back a comment
for a test which was deleted when the test was disabled.

Bug: 1033907, 1021541
Change-Id: If05936b4a2077c6b0486afa3922545ddc898585d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1983267
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728733}