OOR-CORS: Remove an unnecessary scheme check
There is a scheme check and it was for the case running without
the network service. Now that the network service is always enabled
and there is no way to disable it, we do not need to have the check.

This check is not covered by any test, and it would be better to
remove it rather than adding tests for the case.

I'd keep a DCHECK here for a while to capture unexpected caller cases.

Bug: 1039174
Change-Id: Ic29db7c5a4ef6da2f24eaae283e8056ad426dce5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986654
Auto-Submit: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728797}