MacViews: Migrate FocusManager listeners
When a non-top-level NativeWidgetMac is re-parented, it needs to
remove its FocusManager listeners and connect them to the new parent.
Add this logic.

Do this by adding a OnNativeViewHierarchyWillChange and
OnNativeViewHierarchyChanged method to NativeWidgetPrivate, so we can
be sure that this will update all descendants.

Bug: 957362
Change-Id: I38eb894fec4b47062e0e4fe0d8f04cd2ac9105d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1983894
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728830}