Add viewport scroll information to PDFAccessibilityTree
This CL adds following information to PDFAccessibilityTree
- ax::mojom::IntAttribute::kScrollXMin
- ax::mojom::IntAttribute::kScrollXMax
- ax::mojom::IntAttribute::kScrollX
- ax::mojom::IntAttribute::kScrollYMin
- ax::mojom::IntAttribute::kScrollYMax
- ax::mojom::IntAttribute::kScrollY

It enables partial implementation of IScrollProvider, following methods
will work with this change
- IScrollProvider::get_HorizontallyScrollable
- IScrollProvider::get_VerticallyScrollable
- IScrollProvider::get_HorizontalScrollPercent
- IScrollProvider::get_VerticalScrollPercent

This information is required by Screen readers to announce the current
scroll position and extent of scroll to the user.

Tests: The best way to test this change will be through browser tests.
However, it requires the second part of the change, detailed in the
associated bug, to be completed.

Bug: 1034521
Change-Id: I629b2773c75f2b6b222d339cd0ae01799d695295
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1985821
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Virender Singh <virens@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#728843}