Two ProfileManager nits for better readability.
Occasionally it is checked if the profile returned from calling
ProfileManager::GetLastUsedProfile is OffTheRecord or not. A comment is
added to note that except on ChromeOS guest sessions, it is never
OffTheRecord.

ProfileManager::IncognitoModeForced() function is renamed to
OffTheRecordModeForced since it indicates being in guest, system, or
incognito mode.

This CL does NOT result in any behavioral change.

Bug: 1035768, 968028
Change-Id: I4459617cfe517777e321974312474ae7875ecc11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980571
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728849}