[iOS] Replace broadcasting by viewVisible
The broadcasting property is actually checking whether the view is
visible or not. This CL changes the name of the property to reflect
that.
It also changes the current checks for the visibility of the view
(i.e. self.view.window checks) by this property.

Bug: 1038034
Change-Id: I8ec8c6a5c0f3a20324552d99a52e1e61c6c2328a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1983229
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728852}