[css-pseudo] Test dynamic 'list-style-position' for nested ::marker
marker-content-013.html was already testing that various kinds of list
markers are properly updated when 'list-style-position' changes.
However, nested ::marker originated by a ::before or ::after were not
included. I'm working on a patch that could have broken that case, so
covering it in the test seems a good idea.

BUG=457718

TEST=external/wpt/css/css-pseudo/marker-content-013.html

Change-Id: I99c72c02298372c2d026adea99fb580a9bc8890a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1985951
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728868}