[Autofill Assistant] Send memory keys of non-empty input fields.
This will let the backend know which memory keys are set in the client,
such that appropriate actions can be taken. A common use case would be
text input fields on a website that are hidden behind a button. Before
this CL, the backend would not know whether the button needs to be
clicked or not.

Note that the values themselves stay on device, only the memory keys are
sent.

Bug: b/146642017
Change-Id: I440083792965ee874b575fcf480c26ba067fcab5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1988545
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#728912}