android: Lower invisible main frame rank
Currently the whole tab is reloaded when user switches to it, if any sub
frame process is killed while it was invisible. This means there is no
point in trying to keep the main invisible frames alive at the cost of
sub frames. So lower the invisible main frame rank to the same as
invisible sub frames.

Bug: 1038686
Change-Id: I282d5f4670397aa3dd9913dfcb04905f2b752c37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986027
Reviewed-by: ssid <ssid@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728935}