Rename "swap out" terminology to unload.
Historically, we have used the terminology of "swap out" to discuss the
change of one RenderFrame to another during cross-process navigations.
It dates back to times before RenderFrameProxy existed, so we had to
navigate the frame itself to the swappedout:// URL.

This has not been the case for a while now and in reality all the code
that uses this terminology actually does unloading. This CL aims to
rename most instances of "swap out" verbage in lieu of unload to make
it more clear to folks who haven't worked in this area for a while.

Bug: 1015882
Change-Id: I60f7fcfd28057a9c189a990df6d8ebb8ea25ac89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1987030
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#728980}