ozone/wayland: window state change must be synchronous.
This CL fixes two issues:

1) After we changed to only create ShellSurfaces after Show calls, starting the
browser with maximized or fullscreen modes broke. That is, in Wayland, it
is required to assign a role to wl_surface and make higher level objects like
xdg_surface and xdg_toplevel that are used by desktop xdg shell.

State manipulation can only be done through those higher level objects and
we must defer setting state to maximized/fullscreen until they are created.

2) Also, this CL changes the intermidiate result of state changes: now, all of
them are considered synchronous and if Wayland decides to change the state
to something else, the state will be overriden and OnWindowStateChange will
be called. Otherwise, if it was the client, who changed the state, no
notification is sent.

Test: StartMaximized, StartWithFullscreen, CompositorSideChanges
Bug: 1033525

Change-Id: I89728297d572d62db59f2d3b1628f2d735dbd4b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1965031
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#728996}