Fix VR browser control handling after recent refactoring.
The http://crrev/723097 changed fullscreen constraints to be push
instead of pull, but the VR constraint was left behind.  This now
makes VR push an update when entering/exiting VR explicitly to
ensure the controls are hidden w/o relying on a transitive update
on the tab itself.

BUG=1034735

Change-Id: I6069e2c6a8f2778c730dedcc95cb3bff928e2e1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1987296
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729011}