[webui-ntp] Set theme colors on NTP main page
This includes:
- The background color,
- Most visited/shortcuts text and icon colors.

Furthermore, this CL makes the ntp-app element the owner of the
current theme on the UI side so that the app can set the theming on
both the customize-dialog and most-visited elements.

The setting of theme colors on the main page has some delay, which will
be addressed in follow-up CLs.

Bug: 1032327
Change-Id: I23e5a4bb9d6ef6f5f6ce774654bcf2e63d1ecf68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1973943
Commit-Queue: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729012}