Ignore the incognito NTP background color for GTK.
This reverts a change I made to respect the GTK color in
https://chromium-review.googlesource.com/c/chromium/src/+/1961012 .  I thought
ignoring the GTK color was an unintentional oversight.  Turns out it wasn't; the
GTK color is effectively the "non-incognito color", and there's no separate
COLOR_NTP_BACKGROUND_INCOGNITO to avoid overriding.

It doesn't help that the incognito NTP content assumes this hardcoded background
color instead of being responsive to the actual background color that's set, but
I'm not going to try to fix that.

Bug: 1038027
Change-Id: Ifeea542cf60d6d1d79910dc1fb43d550bc7c6210
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1987281
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729017}