IDN Display: Add whole-script-confusable characters for Armenian
This CL adds a list of Armenian characters that look like Latin
characters. A domain consisting only of these characters and not
having a .am ccTLD or other Armenian TLDs such as .հայ will be
converted to punycode, as is done for other whole-script-confusables
such as Cyrillic and Hebrew.

This change doesn't affect any real world registrable domains. It
affects a number of subdomains that use "աաա" instead of "www" and
leaves them in punycode. This is a desired outcome of this change.

Bug: 722167
Change-Id: I82c1dc532331f713f9ce937d079d728670fadadc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1987828
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Christopher Thompson <cthomp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729031}