[Omnibox] Makes BitmapFetcherService receive a callback to notify consumers
This is a followup to crrev.com/c/1976297

Makes BitmapFetcherService receive a callback when requested to load an
image rather than an owning observer as well as include its own network
traffic annotations therefore eliminating the need for the
BitmapFetcherServiceObserver interface as well as the BitmapFetcherHelper
class.

Also increases the BitmapFetcherRequest's cache size from 12 to 16 which is
double the default number of maximum Autocomplete suggestions.

Bug: 1035981
Change-Id: Ide6cd953c8fb20af477b7b208dc98071d13df155
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986432
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729037}