Introduce build flag to disable Assistant integration tests.
These integration tests require libassistant.so to support grpc
communication with the S3 server, which increases the library size,
which is why this CL introduces a flag to disable them in the release
builds.

Result:

With the tests enabled:
❯ size out/Default/libassistant.so
   text     data      bss      dec      filename
13107099   365088   112329  13584516   out/Default/libassistant.so

With the tests removed:
❯ size out/Default/libassistant.so
   text     data      bss      dec      filename
11758198   329288    98945  12186431   out/Default/libassistant.so

without this flag (and ensure the Assistant tests work or are disabled).

Bug: b:146653931
Test: Use |size| to test the library size, run browsertests with and
Change-Id: I1dad722b5c95e1dca3cf0d5c30477ac229e7f89a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1981131
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729070}