[Spellcheck] Fix hybrid spell checking
The previous strategy for hybrid spell checking on Windows was unable to
correctly process text with multiple character sets. Due to how the word
iterator works, it was impossible to distinguish between "skipped" text
and text marked as correct. That caused text skipped in Hunspell to be
considered correctly spelled in the native spell checker, so native
spell checker results were always ignored. See crbug.com/1034043 for
more details.

The best way I've found to solve the problem is to change the way
hybrid spell checking works. It now uses the following  strategy (which
is similar to how the spelling service works with Hunspell checks):
  - Perform spell check with the native spell checker first
  - Double-check only the misspellings with Hunspell
  - If the misspelling text was completely skipped by Hunspell,
    don't consider it as correctly spelled

The reason this works is because we know the misspellings found by the
native spell checker are already broken into words, so if the Hunspell
locales completely skipped the word, we should trust the native spell
checker's result.

More context on this design change can be found in the original design
doc (look for "UPDATED DESIGN" sections and deprecation notices):
https://docs.google.com/document/d/1K4blgUnu4jJ9hViqZzK05d6DRiDCv2pk6Ps6HJ-4Rc4

Summary of the changes:
- Removed the RequestPartialTextCheck Mojo API, since it is no longer
  needed. Hybrid spell checks now use the same code path as non-hybrid
  checks when calling into the native spell checker.
    - Tests for RequestPartialTextCheck have been removed.
    - Refactors that had been made to split up partial VS full native
      spell checks have been undone, to bring everything back to a
      single code path.
- A test has been added to get per-language suggestions in the Windows
  platform code (the test was missing).
- A test suite to test the new logic of combining hybrid results has
  been added.
    - The necessary code has been added to mock classes and test
      fixtures.
- In the SpellCheckProvider class, RequestTextChecking and
  OnRespondTextCheck have been adjusted for the new hybrid checking
  strategy. The code path is now very similar to non-hybrid checks.
- A new filter has been created when creating Blink text checking
  results in the case of hybrid spell checking, to perform the logic
  of double-checking native results with Hunspell.


Bug: 1034043, 463364, 978460, 1035450, 1031618
Change-Id: I362e8d8ac567fa37cfc173c95cb2ee99e88f032e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1984672
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729102}