Merge and rename ProfileIOData Handle classes.
Previously, there were separate classes for OTR profiles and
ProfileImpls, but as things have been removed from ProfileIOData, the
two classes are now identical. This CL merges them, and renames the
single class to ProfileIODataHandle. While the class is slated for
removal anyways, this CL will simplify the final CL to it a bit, and
results in a fair bit less code in the meantime.

Bug: 934009
Change-Id: I8ceec3e227c2ead31cd9e36bd1c43b7300c23a1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1989851
Reviewed-by: David Benjamin <davidben@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729125}