Move ChromeContextMenuPopulatorTest to unit_tests
In order to migrate java.net.URI to GURL the junit tests that rely on
java.net.URI need to be moved to a test apk that supports the native
library.

This change moves the ChromeContextMenuPopulatorTest unittests from
junit to unit_tests (Android only). These tests should be just as fast
as they were as junit tests (and much faster than chrome_test_apk unit
tests). This is a new style of test, where only on Android
the native unit test calls up into java to run the java unittest. The
main benefit of this is a java unittest that's able to run native code
and doesn't restart the Activity between tests so it's really fast. Note
that robolectric can't be used for tests of this style, so it can't be
used for everything.

I also fix various erroneous test expectations caused by the
UrlUtilities mock returning the wrong values for various URLs.

Bug: 783819
Change-Id: Id490571f7b7debce7280311f091dc4f2696edab0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1965918
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729134}