viz: Add a Helper Class to Schedule Task On Gpu Main Thread
Add a helper class to schedule a task directly on the gpu thread. This is
to be shared between the output surface and other users (overlay
processor in the future). This class should be a thin wrapper around
|SingleTaskSequence| in the future when we no longer use a command buffer
inside viz.

This is CL 1/? for send the overlay candidate directly.

R=rjkroege
TBR=kylechar@chromium.org, boliu@chromium.org

Bug: 979788
Change-Id: Ifc35b49c7211d366a8786680056c71cdd338997a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1985299
Reviewed-by: weiliangc <weiliangc@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Commit-Queue: weiliangc <weiliangc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729141}