Neuter DefaultDecoderFactory after MediaFactory destruction.
Since we started allowing asynchronous stop of media::PipelineImpl, we
have released the DefaultDecoderFactory on the media thread assuming it
was safe to continue usage after ~MediaFactory. It turns out this is not
the case for the MojoDecoderFactory used by DefaultDecoderFactory.

To fix, we now neuter the DefaultDecoderFactory in ~MediaFactory to
prevent decoders from being created. This is done under a lock since the
decoder creation methods are called from the media thread and
~MediaFactory happens on the render thread at time of ~RenderFrame.

R=xhwang

Fixed: 1029865
Test: Existing tests all pass.
Change-Id: Ie214fb5eee12dd76a1d21e10497d2cb96ed0397a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1990200
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729142}