[CompositeAfterPaint] Fix backdrop filter mask
Same as blend mode, backdrop filter also uses the backdrop as an input,
so when we synthesize a mask clip for a backdrop filter, we also need
to move the backdrop filter up to the outermost mask isolation effect
node. Now use the same code path previously for delegated blend modes
also for backdrop filters.

Different from blend mode, a backdrop filter has geometry which needs
a proper transform space which may be different from the transform
space of the synthesized clip. To deal with this, the synthesized clip
layer needs to transform the clip operations.

Bug: 923429
Change-Id: I6294dc6ec68a6fe0e7966bdae69fd5b5754fb965
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1977892
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729162}