Add COEP value for service worker to the entry stored in the disk
Service worker script will recognize the COEP response header, and it's used for
process selection on starting up a service worker. However, we don't want to
read the header value everytime to launch the service worker. In order to avoid
the disk access, this CL adds the COEP value to the entry of service worker
storage, which is read on scope matching. The actual COEP value will start to be
stored in later patches, and tentatively it's set to NONE_OR_NOT_EXIST.

Bug: 1039613
Change-Id: I0f29d00bb771774860e459d9bfec7203c2be136a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1989296
Commit-Queue: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729208}