End-to-end tests for opening a linearized PDF file.
This CL adds an end-to-end test that makes sure that linearized PDF
documents can be opened in Chrome.  This was broken before this CL in
DCHECK-enabled builds, because of a NOTREACHED assertion failure in
CorsURLLoaderFactory::IsSane() (see https://crbug.com/1027173).

The new test verifies that the opened PDF document can be searched via
find-in-page functionality.  This helps verify that the text of the
document has been successfully read.  Find-in-page doesn't necessarily
verify that the document is properly rendered, but it is sufficient to
catch the problem reported in https://crbug.com/1027173.

Change-Id: I8bd697d2ea7d6f641b5d0042ec29937791472dd8
Bug: 1027173
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1954870
Auto-Submit: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729237}