Reland: [webgl] Introduce ExtractDataLengthIfValid helper function
The original Cl contained an oversight that is fixed in this CL. The
original CL is uploaded in Patchset 1.

Original message:
As discussed in https://crrev.com/c/1970172, this CL introduces a helper
function to check the size of an ArrayBuffer and creates an error if it
exceeds the maximum supported size.

R=kbr@chromium.org

Bug: chromium:1008840
Change-Id: I13f34f5fff785b852dba1ea16ca1a81f4f93beaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1989530
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729268}