[css-pseudo] Support NeedsOccupyWholeLine quirk in ::marker pseudo-el
In quirks mode, if an outside list marker is immediately followed by an
<ol> or <ul> element, or if there is only collapsible whitespace between
them, then the marker should occupy the whole line and force the <ol> or
<ul> to appear in the next line.

This may not happen when the ::marker is a pseudo-element instead of an
anonymous box. The reason is that the marker is technically in-flow, so
an anonymous block box is created around the whitespace between the
marker and the <ol> or <ul>.

This patch fixes that by making that whitespace think that there is no
previous in-flow content.

BUG=457718

TEST=third_party/blink/web_tests/fast/lists/nested-list-quirks-marker.html

Change-Id: Iad209e8325b776c7836c88b113a19977f214e5e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1985950
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/master@{#729309}