[WebLayer] Expose BrowserFragment's Activity via BrowserImpl
There are multiple upcoming WebLayer usecases for which internal
WebLayer browser code requires access to the Activity associated with
the BrowserFragment (e.g., integration with system autofill and
launching intents from the Activity rather than the Application). This
CL makes the Activity accessible via BrowserImpl.java, which is itself
accessible via TabImpl.java's getBrowser() method. Thus, in any (native
or Java) context in which the Tab is available, it will now be possible
to get the Activity.

Change-Id: Ifd6273aa8552dab2822314d9515d42c96f6f02b7
Bug: 1025617, 1031465
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1982553
Reviewed-by: Tobias Sargeant <tobiasjs@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729315}