Fix test flakiness in animation-state-changes-positive-playback-rate.html
Recent changes to animations caused this test to start flaking. The
problem is triggered by floating point precision in the value of current
time. When start time is set to the timeline time, current time ends up
being 0 +/- epsilon. Whether current time is positive or negative
affects the auto-rewind behavior when calling play on a running
animation.

There are no known precision problem with the normal play and pause
of an animation. This is a special edge case in which a play or reverse
call is made while within epsilon of the start boundary on a running
animation. Note that calling reverse on a newly created and play pending
animation is fine as the hold time will be precisely zero in this case.

Bug: 960944
Change-Id: Ie4af4f8894bce9b17e871e1bc9eb678c7e383afb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986908
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729344}