Fix incorrect escaping of LTR ShortFormModifiers text
On Mac, the shortcut text is created with:
  shortcut = ApplyShortFormModifiers(...)
while on other platforms:
  shortcut = ApplyLongFormModifiers(shortcut);

The ShortForm should *NOT* be processed by the RTL hack.


The fix made to shortcut text on MAC triggered the bug.
  https://chromium-review.googlesource.com/c/chromium/src/+/1876468/2/ui/base/accelerators/accelerator.cc#b239
The size of the shortcut is no longer 2, but 1 insteaad.
For details of the bug:
  https://bugs.chromium.org/p/chromium/issues/detail?id=1038321#c3

The bug fixed previously revealed an other bug here. :)

Bug: 1038321
Change-Id: I4762557d1ffdec783fde4f211c9484a23bddfd6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1990127
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729359}