Remove dead code from animation.cc
PlayStateUpdateScope is not needed in ContextDestroyed since the method
does not manipulate any variables affecting play state.

The removal of the last use of PlayStateUpdateScope creates an island of
dead code. In addition to removal of the PlayStateUpdateScope class, the
following methods are removed:
* variant of SetCurrentTimeInternal which takes an update reason.
* UpdateCurrentTimingState
* PlayStateInternal
* CalculateExtendedPlayState
* BeginUpdateState
* EndUpdateState

The patch also removes variables that were required solely for the
deprecated methods.

With the removal of PlayStateUpdateScope, the sole remaining use of
ResetPendingTasks, was refactored to bring it into alignment with the
spec.

Bug: 960944
Change-Id: Ic861c1839115c29393b97bd7a57a4539abd504cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1943616
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729361}