RenderSubtree: Put search tick marks at the locked roots of matches.
This patch ensures that we put the search ticks at the position where
the locked roots are located. This is the only reliable spot that has
to have up to date layout. Any element under such a root may have out
of date (or non-existent) layout information.

Bug: 1030018
R=chrishtr@chromium.org, rakina@chromium.org

Change-Id: I079ef63ecdab41a1434baf9ad200b4b943496423
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1984358
Commit-Queue: vmpstr <vmpstr@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729377}