Add histogram instead of DCHECK to avoid cases when profile data is stale
As explained in the attached bug, the developer was running ChromeOS and
Chrome linux in the same user data directory. This lead to the case
where the primary account was an email even thought the account ID
migration preference was set to "done". This lead to a new DCHECK in the
migration code firing.

Even thought this is not expected to happen in production code, this CL
uses a new entry in the histogram to avoid strange migration scenarios.

Bug: 1034512
Change-Id: I61e0aaadb93a7f7c6c3a09d3a009fa9f2bf2ee6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1971838
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729397}