Bookmark sync: use GUID as temporary server ID during commit
This patch effectively relands the logic reverted in
https://chromium-review.googlesource.com/c/chromium/src/+/1864800 which
caused crashes in the past (CHECK failures).

In this new form, the underlying issue is addressed: for users that
don't use GUID-based merging of bookmarks, local GUIDs are reassigned
randomly to avoid GUID collisions in the local bookmark model.

Bug: 978430,1004205
Change-Id: I812d1cfe1c12b766799e6c5ec238f95fd16139ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1983231
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729405}