Remove unused SandboxType::kProfiling
This supported out of process profiling but this is now back
in the browser, so this sandbox type is not needed.

Also removes |g_use_localtime_override| from libc interceptions
which was the only thing controlled by the kProfiling sandbox type.
This global defaulted to true so it is safe to remove it from the
compound checks that used it.

BUG=772503

Change-Id: Id1639c595d10ef414be36fc1d62fd58864db002c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1988916
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729412}