[ios] Do not add Infobar badge if Infobar's WebState is not active
Adds WebState as a parameter to InfobarBadgeTabHelperDelegate's addInfobarBadge:.
Then BadgeMediator returns early if the WebState passed is not the same as the
WebStateList's active WebState. This prevent a situation where an inactive WebState
gets a new Infobar and the BadgeViewController then adds its badge to the UI
even though there is no such Infobar for the currently active WebState.

Fixed: 1039668
Change-Id: I1eaf5aae4385d372d025dddce72a362e521fece0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1989993
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729435}