[Extensions Cleanup] Remove WaitForExtensionInstallError()
Remove ExtensionBrowserTest::WaitForExtensionInstallError() and
ChromeExtensionTestNotificationObserver::WaitForExtensionInstallError().
The latter was only called by the former, which in turn was only called
in a single place.

Inline the (short) implementation directly into the single test that
called it. In addition to removing cruft from the (very large)
ExtensionBrowserTest interface, this also fixes a potential issue:
the proper pattern for these types of observers is:
<instantiate test observer>
<trigger behavior>
<wait with test observer>

Here, it was:
<trigger behavior>
<instantiate and wait with test observer>

This is can be racy or incorrect depending on the triggering behavior,
since it may already finish before the observer is instantiated.

Bug: None
Change-Id: If35c9faff6ac60e86b476c5064ad0d0d1dd7f739
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1981050
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729471}