Downward shelf drag should not start window drag
When handling shelf drag with extended hotseat, do not start window
drag unless the current drag is moving downwards.

While ShelfLayoutManager::MaybeStartDragWindowFromShelf already checked
whether the scroll direction is not down, it was bypassed when the
method was called from StartShelfDrag (because the scroll value was not
known at this point).

This updates MaybeStartDragWindowFromShelf no to start window drag until
it's clear that the drag is upward drag when hotseat is extended.

Also, removes MaybeStartDragWindowFromShelf call in StartShelfDrag as
it's expected to be no-op with this change.
*   the method is no-op in kShown hotseat state
*   in kHidden hotseat state, drag is started only if drag gets above
    kExtended hotseat bounds (which will not be the case when the drag
    starts)
*   in kExtended hotseat state, the window drag will not start as the
    drag direction is not yet known at this point.

BUG=1039906

Change-Id: Ia8087422d61726316a99cc11cc011b9f90d42d84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1992347
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729519}