[iOS] Create OverlayBrowserAgentBase
OverlayBrowserAgentBase is a superclass for BrowserAgents that use
callbacks installed onto OverlayResponses to make model-layer updates
for user interaction events that occur on overlay UI.

This will be used to handle the shared model layer interaction for
infobar overlays, as the same model will need to be updated for banner,
detail sheet, and modal overlays.  Previously, this was done in
InfobarCoordinator, which remained started and in memory for the
duration of the InfoBar's lifetime.  With OverlayPresenter, however,
the coordinator that handles the UI is only in memory from the first
time an OverlayRequest is presented until its final dismissal either
by user interaction or by cancellation.

Bug: 1030357
Change-Id: I6ceeb1e29523df8d96966ba2800a1562eadd2ee8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986461
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729544}