[Extensions] Store external uninstall info in a separate pref
When a user uninstalls an externally-added (sideloaded) extension, we
store a pref so that Chrome doesn't try to re-add the extension when it
next sees it. Currently, this is stored as a bit in the extension's
entry in the ExtensionPrefs (setting "state" to
"EXTERNAL_EXTENSION_UNINSTALLED").

Unfortunately, this causes a number of interesting side effects, because
it means that, in this specific scenario, the Preferences retain the
Extension's entry even after uninstallation. Then, when re-installation
is attempted, some of those preferences can have an impact. In the past,
this has also led to improper behavior around installation of an
extension (see linked bug).

Fix this by instead storing uninstalled external extension IDs in a
separate preference entry ("extensions.external_uninstalls"). This way,
we can fully remove the Extension's entry in preferences when it's
uninstalled.

Add migration logic to move all externally-uninstalled extensions into
the new array, and remove their entries from prefs. Add unit tests for
the migration logic, external uninstall being set and retained, and
for the scenario of a user uninstalling an external extension and
re-installing it again.

Bug: 795026

Change-Id: I177777ed78cf0a85ad9be4366999a80aecd495ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1982239
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729553}