[FlexNG] Introduce FlexItem::physical_margins.
Previously we'd store intermediate calculations of margins directly on
a FlexItem's box.

This instead stores the intermediate margin calculations on the
FlexItem directly.

This is primarily just a refactor / cleanup. (The existing code was
fine, not causing any bugs/issues), but was somewhat confusing from a
LayoutNG p.o.v. as transient calculations shouldn't be stored directly
on the LayoutBox.

Change-Id: Ia37b3837ecbbfbd3f61c19571536fc8484d85d54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1977196
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: David Grogan <dgrogan@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729572}