[iOS] Add OverlayRequest::GetWebState()
The value is set when an OverlayRequest is added to a queue.
This allows the WebState to be accessed from request callbacks.
Will be used when one request's completion callback needs to
enqueue an additional request with a different ConfigType into
the WebState's queue.

Bug: 1030357
Change-Id: I08fee42f2da58f8451d83d5f0e417374a5b1c5ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1989912
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Auto-Submit: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729603}