Make RecentModel::GetRecentFiles accept file-type filter.
As described in the bug, we are going to filter files returned by
RecentModel::GetRecentFiles() based on media types.
As a first step, I'm  adding FileType parameter on the GetRecentFiles().

Collecting files are implemented in RecentDiskSource, RecentDriveSource,
and RecentArcMediaSource. Actual filtering code will be implemented in
them in follow-up patches based on the passed |file_type|.

Bug: 1040049
Test: Run unit_tests
Change-Id: I48d4a055a74a1b95ad61d75d557ab1dddfb44951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1991070
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Naoki Fukino <fukino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729624}