Enable WCG on Pixel 4 by default
Wide color gamut on Pixel 4 was already enabled by default, but
the rigging was setup as though it wasn't. Part of the feature
roll-out is to mark the feature as enabled-by-default.

This CL mimics that step (but leaves it already enabled-by-default)
by removing the field trial and updating the feature check to
no longer require that field trial.

Bug: 1021181
Change-Id: I556a25b39f667e2e9199fb8d6d9b8f680d97fee0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1992010
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: Chris Blume <cblume@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729638}