Android: Fix preview tab scroll
There is a timing issue that keeps the desired bottom sheet content
height from being set correctly if tab switcher is invoked between
device rotation. The activity tab from which the desired content
height is obtained can be still be null when the height change occurs.
Currently the related logic does nothing in this case, and the following
layout change events cannot update the height either, since they see
no change in the height. Therefore, BottomSheet.mContentDesiredHeight
retains the old value in |ensureContentDesiredHeightIsComputed|, not
being able to get updated with the right value.

This CL fixes it by keeping track of the current content height, and
updating it when a valid height is available.

Bug: 1039204
Change-Id: Id173e90ac35bc3177f647f3c937da61089de7516
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1989305
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729654}