[FragmentItem] Generate NGFragmentItem for culled inline boxes
This patch changes to generate |NGFragmentItem| for culled
inline boxes.

In common cases, inline boxes are not needed for paint
purpose, and that we suppress generating |NGPhysicalFragment|
for them. However, hit-testing, |getClientRects()|, and some
other scenarios require the coordinates of such inline boxes.

Synthesizing it saves memory, but has drawbacks in the
performance, the correctness, and the code complexity. With
this change, we always generate |NGFragmentItem|, while
|NGPhysicalBoxFragment| is still suppressed when possible.

Some related code need to change accordingly; e.g., hit-
testing and |NGInlineCursor|. ~120 new failures are due to
the lack of these changes. They will be fixed in following
patches.

Bug: 982194
Change-Id: I53a336952202087abb0f80214354d092e970a81e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1774425
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729656}