Create mechanism to query when an embargo for a permission started
To support the display of recently changed permissions to the user as
part of the the privacy settings redeisgn, the ability to query when a
permission was embargoed is needed.

This CL introduces a GetEmbargoStartTime function in the permission
auto blocker that facilitates this.

Bug: 1032584
Change-Id: I04da71522175a8e59051b45e7573bd65468e78cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1983216
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729661}