Add Image::Size(RespectImageOrientationEnum) overload
The operation of finding out the "oriented" image size is used in a few
places. Put this in a new overload of Size() in Image which takes and
applies a RespectImageOrientationEnum argument.

This should allow to (finally) eliminate the IntrinsicSize(...) method
from ImageResourceContent.

Bug: 773272
Change-Id: I031cfe735ebf285405c240e294686f755b7e0c44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1991564
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#729680}