Check for image content in ImageDocumentParser::Finish
If the <img> element in the image document has had it's source cleared
out (by having its 'src' changed or by moving it to an inactive document
or similar cases) we may not have any image content associated with it
when we reach ImageDocumentParser::Finish (when the original image data
finishes loading). Check if image content exists before updating the
title of the document et cetera.

Bug: 1036686
Change-Id: I5d0e2e52e9373c541e64d33bbaa6c67dd0535092
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1989755
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#729743}