Revert r702352 to fix issue with visual overflow recalc
The change in r702352 was causing issues in some situations,
when an inline element has some changes that cause an overflow
that should be visible in a scroller.
Problem is that we're no longer calling
MarkSelfPaintingLayerForVisualOverflowRecalc() after r702352
and the visual overflow is never recalculated.

This revert changes the call
in NGBlockNode::CopyFragmentDataToLayoutBox() to the original one,
to SetNeedsOverflowRecalc().
Due to that change we need to disable NeedsLayoutOverflowRecalc
unit test as it's now failing.

We added a new test case to prevent a similar regression in the future.

BUG=1036644,941180
TEST=fast/overflow/visual-overflow-recalc.html

Change-Id: If205d08281ca0186559715a788b132850e3cbdd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1992423
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Manuel Rego <rego@igalia.com>
Cr-Commit-Position: refs/heads/master@{#729750}