[iOS] Use IsURLTrustworthy directly
This CL removes the origin_util file in web and directly uses the shared
IsUrlPotentiallyTrustworthy util in the call sites.

The previous implementation of web::IsOriginSecure() used to consider
data scheme as trustworthy (which is no longer the case since
crrev.com/c/1982540). The explicit checks for the data scheme can be
removed.

Fixed: 939077
Change-Id: I74d4d1972c19dc15691f29d7ecb66b57cbd32ea7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986072
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Maxim Kolosovskiy <kolos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729761}