Enable leaving notification embargo.
There is no notification channel when the notifiction permission for
the origin is merely embargoed, so it is not possible to intent to
notification channel settings.

This CL updates the SingleWebsitePreferences page on Android, so that
when the automatically blocked setting is changed for the first time,
a synthetic CONTENT_SETTING_BLOCKED is directly written just-in-time.
This will trigger lifting embargo, and create an initially disabled
notification channel, and finally we intent into channel settings.

Bug: 996201
Change-Id: Ice838b1ca971c9d8117f1c56ecc107bf41417550
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970473
Reviewed-by: Natalie Chouinard <chouinard@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Commit-Queue: Andy Paicu <andypaicu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729811}