Removed dependency on GetThemeProviderForProfile() from the tabstrip.
Removed the call to GetThemeProviderForProfile() from the webui
tabstrip and replaced it with a call to GetThemeProvider() accessed
through its embedder object.

The intention is to ensure colors are obtained via the ThemeProvider
object accessed via the widget hosting the given ui element. This
paves the way forward for the rolling out of the color pipeline
project.

Bug: None
Change-Id: If298faae91f509c8448ef6b2bdd81b24fa2074c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1990337
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Reviewed-by: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729827}