Cros: Make cloned screen transparent for rotation
During screen rotation we create a copy of the current screen as a layer
and stack it in the LayerTree to ease the animation between the screen
before and after the rotation.
However, some elements in the shelf are obscured by the copied layer.
The new layer is set to be transparent as a workaround so it doesn't
occlude these elements.

Bug: 1025400, 1040279
Change-Id: I003cc149fca54cc821f65bc2f476c57c61d1788b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1992689
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729834}