Register Autofill tab helpers for presentation windows
The Presentation API allows to create full screen windows for a
secondary display. The corresponding WebContents is not associated
with the TabStrip and therefore, TabHelpers::AttachTabHelpers() is not
called on the WebContents and the autofill drivers are not registered.
As per
https://chromium.googlesource.com/chromium/src/+/master/docs/tab_helpers.md,
we must not reuse TabHelpers::AttachTabHelpers() but be selective
about the helpers we attach to each WebContents.

Bug: 1038329
Change-Id: I455a11b997a14ed6e7a7a5dad6a47fdada77cedc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986827
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Reviewed-by: Brandon Tolsch <btolsch@chromium.org>
Commit-Queue: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729842}