Don't apply forced colors when printing
We don't want to print high contrast colors when in forced colors
mode. To avoid this, add a check in Document::InForcedColorsMode()
to see if we are printing. If so, InForcedColorsMode() returns false,
and the printed content is no longer affected by forced colors mode.

Bug: 970285
Change-Id: Ia0338834a6fa945a4446e51377014f951d83e997
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1987985
Commit-Queue: Alison Maher <almaher@microsoft.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729849}