Change never-visible terminology for RenderWidgets to never-composited.
The terminology was defined around never Showing a RenderWidget to the
user and thus not compositing it. However visibility is now split
between Widgets (show for compositing) and Views/Pages (show for blink
timers/js). And since we plumb this bit through RenderView creation and
store it there, it becomes ambiguous. This renames it throughout to be
more clear that it is about compositing (ie RenderWidgets) and not
about page visiblity.

R=avi@chromium.org, dcheng@chromium.org
TBR=sky

Bug: 419087
Change-Id: Ic95d7c0eae473d5a1adfc369fd88272471f6fafb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1992343
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729855}