Android: Clean up feature module JNI registration
DFMs use explicit JNI registration for both Chrome and Monochrome, but
the feature shouldn't care which it's in. Traditionally, JNI registration has
been done based on APK targets, but we can also point generation at an Android
library.

Do that here, and also, tweak the JNI generation template to allow
exclusion of transitive deps. For a DFM, this means we can generate JNI
only for a module's entrypoints, without pulling in methods from the
base module or other unneeded deps. This seems cleaner, and also allows
this change to work in component builds.

Bug: None
Change-Id: I3f3f457815a7e2f06f575f72476cc4c877741c7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1992694
Commit-Queue: Christopher Grant <cjgrant@chromium.org>
Reviewed-by: Eric Stevenson <estevenson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729869}