Add a flag for building HWASAN splits.
The new build_hwasan_splits flag causes HWASAN splits to be created for
the Monochrome6432 and TrichromeChrome6432 bundles, as well as any of its
DFMs that contain native code. The intent is that this flag will eventually
default to true in official builds.

The library is deliberately installed in TrichromeChrome6432 rather than
TrichromeLibrary6432 (where the regular native library lives) because the
Android platform does not support splits for the library containing the
WebView implementation. This means that only Chrome will use the HWASANified
native library, and WebView will use the regular native library.

Bug: 916828
Change-Id: Ie92bb36e83df245ab431b0123a44b0f0a3182100
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1769731
Commit-Queue: Peter Collingbourne <pcc@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729945}