[iOS] Create InfobarOverlayBrowserAgent
InfobarOverlayBrowserAgent is an OverlayBrowserAgent subclass that
handles updating InfoBarDelegate state for user interaction with
infobar UI presented via OverlayPresenter.

For each InfobarType, an InfobarInteractionHandler will be created
that updates the infobar model layer for interactions with the UI.
These interaction handlers are provided to the InfobarBrowserAgent,
which sets up OverlayRequestCallbackInstallers to forward responses
from the UI to the interaction handlers.

InfobarBannerOverlayRequestCallbackInstaller is used to forward user
interaction events for infobar banner UI.  The interaction handlers
and the corresponding callback installers for detail sheet and modal
UI will be implemented in follow-up CLs.

This CL also updates AttachInfobarOverlayBrowserAgent() to install
the browser agent if the feature is enabled.  Subsequent CLs will
be added to provide the browser agent with interaction handlers for
each InfobarType.

Bug: 1030357
Change-Id: I131b03e8dd1e5e4228ade4d9e01730114545fa5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1989855
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729991}