[sms] Allow multiple one-time-code forms on a single page
This changes allows a page to have multiple input forms with
autocomplete='one-time-code'. If subscribe has already been called for
an SmsClient, it will silently ignore. Without this change, the page
will crash upon trying to Subscribe again because each SmsClient is only
allowed to be subscribed once in the SmsQueue[1]. This crash is also
reproducible by reloading the page with the autocomplete input.

[1] https://cs.chromium.org/chromium/src/content/browser/sms/sms_queue.cc?sq=package:chromium&g=0&l=16

Bug: 1037914
Change-Id: I130aeda2b8c9df2328393ec6b42a7cae608ef07f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1981266
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Commit-Queue: Ayu Ishii <ayui@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729999}