Extend sync's Javascript API with key versions
For debugging and future feature development, the web should be allowed
to pass a key version in addition to key material. If multiple keys are
involved, only the version of the newest is relevant, so the API is kept
as simple as possible by plumbing a single integer.

Bug: 1032485
Change-Id: I0e53c54176e92c1dfa6ffcb496611ae53c8c113f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1991432
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730078}