Update focus ring style.
*** SHERIFFS: if this CL triggers a flaky layout test, PLEASE do not
*** revert this CL. Instead, please disable the offending test(s) and
*** add them to the crbug.com/1035582 section of TestExpectations. I
*** will then take care of them as I fix up the remainder of these
*** tests. Feel free to cc me if you create a bug. Thanks.

This CL updates the focus ring to be made of two borders which have a
2:1 ratio. The new width of the focus ring is 3px: 2px nearly black (#101010),
followed by 1px white.

If the element has a border large enough to have an inset outline (width >= 1px)
then we are going to inset the focus ring by 1px as we want it to be partially
drawn over the border to avoid clipping or strange alignment issues caused by
increasing the width to 3px.

For elements that have not been styled and that have an appearance, the border
radius of the focus ring is updated to be the same as the one used for drawing
the element. For all other cases the current behavior is maintained:
border radius = outline width.

Tests are added to validate the new focus ring for multiple scenarios (see
controls-refresh/focus-rect)

Bug: 1026553
Change-Id: Ibe05cf6b7336198bd9513174493f51da5108339a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1975072
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730110}