[Flags] Add flag for WebAssembly tiering
We plan to run a finch trial on mobile chrome to test different
compilation strategies. For this, we need to separate enabling Liftoff
from enabling tiering. The finch config could then enable Liftoff but
disable tiering - a configuration which is not possible currently.

Note that this CL does not change the default configuration. On Desktop,
we enable both Liftoff and tiering. On mobile platforms, Liftoff is
disabled, so the tiering flag has no effect.

R=ahaas@chromium.org, jochen@chromium.org

Bug: 1040030, 1040061
Change-Id: I7cd0a8907d84388d15cb49663d18fe4c5c81dd83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1988360
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730128}