sessions: removes SessionTabHelper::SetTabExtensionAppID
And puts the implementation in the single place that needs it.
SessionTabHelper is really intended to forward WebContentsObserver
related calls to SessionService. SetTabExtensionAppID() isn't that.
As SetTabExtensionAppID() was called in a single place, I moved
the implementation there.

This is part of a larger refactoring I'm doing in hopes of moving
SessionTabHelper to a shared location.

BUG=1033924
TEST=none

Change-Id: Ia70af33908593839d4a22b4ebbc5f9f48c8abc30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994527
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730200}