[discarding] Only log memory details on memory pressure on ChromeOS.
We should minimize the amount of work done when the system is under
memory pressure, to allow it to exit that state.

Since the "Tab Discards Memory details" log is only used on ChromeOS,
this CL removes it on other platforms. Collecting memory details is
the culprit for 3% of IO thread hangs, so we shouldn't do it for no
reason.

Bug: 1040522
Change-Id: Ie4fff140a7024d9372a3888a515466885cce68da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980282
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730202}