Fix render surface occlusion with pixel-moving filters
We have been correctly ignoring occlusion from outside of the subtree
of a pixel-moving filter in the stack of OcclusionTracker (which leads
to correct occlusion inside of the subtree), but when we get the
occlusion of a render surface, previously we returned the current
occlusion of the parent render surface even if the current render
surface has pixel-moving filter, causing the render surface incorrectly
occluded.

Now add a flag in the stack indicating that the render surface doesn't
inherit parent occlusion, and
OcclusionTracker::GetCurrentOcclusionForContributingSurface() should
not return the parent occlusion when the flag is set.

Bug: 1040288
Change-Id: I505b079d62186a404b1be1c83c6cbccc89c10b8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994102
Reviewed-by: weiliangc <weiliangc@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730215}