Adding cache resetting param to NetworkContextParams.
This parameter will be used when the backend changes its
underlying key representation, making previous keys invalid.
Rather than retaining all the old keys in the cache, an empty
cache is created in its place.  Right now this is primarily
used for many of our Finch experiments.  This CL detects and
stores this state in prefs and then passes the appropriate
reset parameter to the cache.

Change-Id: Ifd466dc902e6368df8122223ed34272e2e9db848
Bug: 1002220
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1918392
Commit-Queue: Eric Robinson <ericrobinson@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730219}