CrOS Shelf: Set right margin on status area based on tablet mode
With this change, the padding between the edge of the screen and the
status area widget no longer has anything to do with the hotseat
state. This edge padding will now only ever have two possible values,
one value in tablet mode and another value in clamshell mode.

This is a reland of crrev.com/c/1977025, which was reverted because
of a test failure when the hotseat flag was enabled. The addition of
UpdateIsDense() in OnAppListVisibilityWillChange() fixes the test
failure. UpdateIsDense() has been changed to help minimize the number
of times OnShelfConfigUpdated() will be called.

Bug: 1035716
Change-Id: Iddf1e13928d73f6783d33bddfd6cc1796d0abe31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1990513
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730274}