Alter behavior of local_discovery.enabled pref
A previous CL [1] added the pref and completely disabled access to
the local discovery UI (chrome://devices) if the pref is set to false.
Change the pref to still allow access to the UI, but disable any
capabilities to register or add devices.

Added an early return in HandleRegisterDevice().

[1] crrev.com/726249

Bug: 1011582
Change-Id: I5de45881264043492b9bff4c9596b24b6cc5509c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986907
Commit-Queue: Daniel Hosseinian <dhoss@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730316}