SplitSettings: Start splitting up route.js
This CL:
- Moves the Route and Router class definitions to a separate file, so
that they can more easily be shared by browser and OS settings
- Moves SettingsRoutes to a separate file as this is a shared
dependency of Router and the remaining functions in route.js
- Makes Route class's createChild and createSection methods public
so that they can be called from route.js.

This is in anticipation of forking the remainder of route.js, which
contains different logic for adding OS vs browser settings routes.

Bug: 967861
Change-Id: I6565792c808633f3eb00e1286f66cabb225002d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1991811
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730356}