build: Change definition of is_a_target_toolchain.
https://chromium-review.googlesource.com/1994671 adds a third target
toolchain that will be used to build a HWASANified native library on Android.
This requires adjusting the definition of is_a_target_toolchain. We can
simplify the definition so that it considers all non-host toolchains to be
target toolchains except when targeting the host, as long as current_os
matches target_os, so that for example we avoid considering Android as a
target toolchain when targeting CrOS.

Bug: 916828
Change-Id: Ib688efef7602e961eb53420a9cb89b734c3c76b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995772
Auto-Submit: Peter Collingbourne <pcc@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Peter Collingbourne <pcc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730390}