select: fix border:none issue
The Combobox issue (where the control has unexpected lines around the boundary) can be
reduced to <select> still rendering border when border:none is specified.
The reason is that NativeThemeBase::PaintMenuList is calling NativeThemeBase::PaintTextField
(to draw border and background) and PaintTextField would always draw border even when it's not
supposed to (menu_list.has_border = false). This change addresses this issue by introducing
TextFieldExtraParams.has_border and passing down menu_list.has_border to PaintTextField
so it has the ability to not draw the border.

A new test case is added to select-inpage-appearance-basic.html for <select> with border:none scenario.

Bug: 1023883
Change-Id: I74ef2c8a0227d83b64385b64541cdfb6468cdac8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1989001
Commit-Queue: Alex Keng <shihken@microsoft.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730450}