Perform document order index validation before animation value update
The interval update doesn't not care about the document order of the
timed elements, while the computation of the animation values do. Hence
move the call to UpdateDocumentOrderIndexes() to ApplyTimedEffects() to
get a more logical ordering of the calls.

Also hoist the check for an empty prio queue, since that implies that
there's nothing to do, and UpdateAnimationsAndScheduleFrameIfNeeded can
return early.

Bug: 1017723
Change-Id: If6fac53e01773d084d814c6067863e51812598de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994887
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#730550}