Reformat all gn files in /chrome/browser/net.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=mmenke@chromium.org

Bug: 1041419
Change-Id: I36a069f97ef7b87147c95caeab6ee57fda0f9443
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994075
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730566}