Reformat all gn files in /chrome/browser/resources/settings/chromeos/multidevice_page.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=nohle@chromium.org

Bug: 1041419
Change-Id: Ifc5a6ba0e265ba9e55dbe894c3cd58ab5e8cc3ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1993853
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Josh Nohle <nohle@chromium.org>
Commit-Queue: Josh Nohle <nohle@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730569}