Reformat all gn files in /chrome/browser/ui/messages.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=twellington@chromium.org

Bug: 1041419
Change-Id: Ia86be4784c3abacf4e9f42a7a8b8cb0c15103c88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997919
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730572}