Reformat all gn files in /ui/gfx/range.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=rsesek@chromium.org

Bug: 1041419
Change-Id: If7a4f5f1dd190e901349bf85f669fc8026287d98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997000
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730579}