Add translation meaning for policy on/off strings
The "desc" attribute is not sufficient to provide context to the
translators because identical strings will be collapsed in the
translation console. Add "meaning" to disambiguate these strings
from other uses of "on" and "off". Upload screenshots of the UI
as well.

See bug for details, or
https://www.chromium.org/developers/design-documents/ui-localization

Bug: 1031145
Change-Id: Ic61df8d72593b86e2b59077612d8e0f664357166
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986933
Auto-Submit: James Cook <jamescook@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730580}