Reformat all gn files in /third_party/robolectric.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=yolandyan@chromium.org

Bug: 1041419
Change-Id: If3415d28f4ae5464a00cd86c56707f1367fcf056
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995848
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Commit-Queue: Peter Wen <wnwen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730583}