[Fuchsia] Add AcceleratedWidget to CreateInputMethod in Ozone
Add AcceleratedWidget as a parameter to CreateInputMethod for Ozone
since it is needed by the platform Scenic. The Scenic platform is
only used by Fuchsia, which is the only platform affected by this
change.
The new input service used by Scenic needs a way to associate a
window or view to an IME.

Bug: 988507
Change-Id: I74c1c7eaf5b71b390b8c63e8b232daf6b8290f10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1993679
Commit-Queue: Sharon Yang <yangsharon@chromium.org>
Reviewed-by: Yohei Yukawa <yukawa@chromium.org>
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Michael Spang <spang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730584}