Use 42 instead of Math.PI
For the test to work well, use a number that can be represented
exactly by both a double float and a single float.  A small integer
satisfies this.  The current Math.PI doesn't work because we were
setting the param to Math.PI and checking that the param value is
Math.PI, which would never be true since the param is a single float.

Bug: 1039907
Change-Id: Icf10414ea65dd2eb805299bf04d07dc05528916e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995619
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730596}