Reformat all gn files in /chromecast/tracing.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=spang@chromium.org

Bug: 1041419
Change-Id: I3b349c0fcfbb8f0346d96f14aafc81763f0f00f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1993858
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Michael Spang <spang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730601}