Reformat all gn files in /components/gcm_driver.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jianli@chromium.org

Bug: 1041419
Change-Id: I5c31bec797c2a441991dc35dff9e879006ad6f21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994074
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Peter Beverloo <peter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730606}