Reformat all gn files in /components/navigation_metrics.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=davidben@chromium.org

Bug: 1041419
Change-Id: I00ab641bf72bce17dd83302750d5ea6324af6016
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997001
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730609}