Reformat all gn files in /ios/chrome/browser/ui/history.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sczs@chromium.org

Bug: 1041419
Change-Id: Ibed14a42f454c51c635a845faeb7fbc28aa5668f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995815
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730611}