Add skeleton classes for Isolated Prerender
This CL adds some of the fundamental classes for Isolated Prerenders
and a little bit of the core logic, but mostly it's all boiler plate.

The interceptor will trigger on applicable pages and passes the
IsolatedPrerenderURLLoader on those navigations. For right now, that is
just a simple pass through to the network but it will become more
interesting in future CLs.

Design: http://shortn/_ayPYuHZA4b

Bug: 1023483
Change-Id: I7e4bbe4865bc237507c3c387d6e72696f81debcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980352
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730613}