Reformat all gn files in /components/metal_util.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=ccameron@chromium.org

Bug: 1041419
Change-Id: I6ca1b116f6d23c86853c0cdce207a5e308a60900
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994077
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730614}