Reformat all gn files in /chrome/android/features/test_dummy/internal/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=wnwen@chromium.org

Bug: 1041419
Change-Id: I861c0ae3197a994b69da5cf5c87abe990b5c5b46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1993850
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Commit-Queue: Peter Wen <wnwen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730615}