Reformat all gn files in /content/browser/content_index.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=knollr@chromium.org

Bug: 1041419
Change-Id: Ib9ccf99302675a20cfec621c9b5aeb6ccc11ab43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998069
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Richard Knoll <knollr@chromium.org>
Commit-Queue: Richard Knoll <knollr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730616}