Reformat all gn files in /services/media_session.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=beccahughes@chromium.org

Bug: 1041419
Change-Id: Icb8bb604bb41fe713dab12a57363ebcee751b2bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997758
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Commit-Queue: Becca Hughes <beccahughes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730621}