Reformat all gn files in /chrome/android/modules/extra_icu/provider/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jbudorick@chromium.org

Bug: 1041419
Change-Id: Ib595c782904977a8edb86f84c32706264b9be576
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997793
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: John Budorick <jbudorick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730622}