Reformat all gn files in /chrome/test/data/media/engagement.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=beccahughes@chromium.org

Bug: 1041419
Change-Id: I1e9449a0d6baa472f588f6cd86a8b487fef5007a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995845
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Commit-Queue: Becca Hughes <beccahughes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730626}