Reformat all gn files in /chrome/common/mac.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=rsesek@chromium.org

Bug: 1041419
Change-Id: I5076378956df18ea595ee8fa556ea75e7db83842
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997923
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730628}