Reformat all gn files in /third_party/blink/renderer/modules/battery.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=timvolodine@chromium.org

Bug: 1041419
Change-Id: If653e498a90dd9e2db7f493d7a5de81ba8b98733
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995807
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Tim Volodine <timvolodine@chromium.org>
Commit-Queue: Tim Volodine <timvolodine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730631}