Reformat all gn files in /chrome/android/modules/stack_unwinder/public/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=tiborg@chromium.org

Bug: 1041419
Change-Id: Ia319e28064a679a85da206dd93f55def5ab51ce4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998031
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Mathieu Perreault <mathp@chromium.org>
Commit-Queue: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Mathieu Perreault <mathp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730633}