Reformat all gn files in /third_party/zlib/contrib/tests/fuzzers.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=cblume@chromium.org

Bug: 1041419
Change-Id: I7c5f71a3369485a78c6e7ba9490f893e15b65c9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997933
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Chris Blume <cblume@chromium.org>
Commit-Queue: Chris Blume <cblume@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730634}