Reformat all gn files in /chrome/android/features/dev_ui/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=wnwen@chromium.org

Bug: 1041419
Change-Id: I70252c9919dc07ed01aace76fa6547f042e7d9d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998062
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Commit-Queue: Peter Wen <wnwen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730635}