Reformat all gn files in /build/config/fuchsia.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=wez@chromium.org

Bug: 1041419
Change-Id: I733c080b363275f58bd115ce62ff4f53d64ba53a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997780
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Commit-Queue: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730636}