Reformat all gn files in /components/bookmarks/common/android.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=tedchoc@chromium.org

Bug: 1041419
Change-Id: I786e44bae873aa353f49547af725226d1045de89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997742
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730637}