Reformat all gn files in /third_party/pffft.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=olka@chromium.org

Bug: 1041419
Change-Id: I4b73fd903352e7de1845445841d8189dc23b9302
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997680
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Olga Sharonova <olka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730639}