Reformat all gn files in /chromeos/services/secure_channel.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=nohle@chromium.org

Bug: 1041419
Change-Id: I75fc757ebaf825b6d9faf0638efa5c24d996368e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998050
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Josh Nohle <nohle@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730646}