Reformat all gn files in /third_party/flatbuffers.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=engedy@chromium.org

Bug: 1041419
Change-Id: I7e7e1d81e71f2a8aa132132329ddff072b30efb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997823
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Commit-Queue: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730647}