Reformat all gn files in /chrome/browser/ssl.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=agl@chromium.org

Bug: 1041419
Change-Id: I8de3776cc3d6409c5fbd6ac9d04ecaf1b5ed9388
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998260
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730651}