Reformat all gn files in /remoting/client/jni.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=yuweih@chromium.org

Bug: 1041419
Change-Id: Ic24d85b6ca0c6955a55518b9610ff8fd3cdb8c56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997799
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730659}