Reformat all gn files in /media/muxers.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=mcasas@chromium.org

Bug: 1041419
Change-Id: I28818ce68ed58360c3d783ee7a1c1ae6f5c29243
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997836
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730660}