Reformat all gn files in /chromecast/media/cma/backend/fuchsia.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sergeyu@chromium.org

Bug: 1041419
Change-Id: Ic05ca5e7c4f685d8dfdfc6b3a3b76e3894688ff6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998071
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730663}