Reformat all gn files in /chrome/browser/browser_switcher.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=nicolaso@chromium.org

Bug: 1041419
Change-Id: I13b293a6a76900c8c8d55ea7898d0e0a7e5eb4d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995812
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Commit-Queue: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730664}