Reformat all gn files in /chrome/browser/preferences.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=hnakashima@chromium.org

Bug: 1041419
Change-Id: I3d6e329440eba2bcf597a3c87d25a817be581064
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998249
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730668}