Reformat all gn files in /components/ukm.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=asvitkine@chromium.org

Bug: 1041419
Change-Id: I94706aa025c6b48e7b9b9f1e355382cd8ec9f60b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997924
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730670}