Reformat all gn files in /chrome/browser/resources/chromeos/accessibility/chromevox/run_jsbundler.gni.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=pdr@chromium.org

Bug: 1041419
Change-Id: I549c5714e460510fc249c6fb2a887a5e3d65dfa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997988
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730678}