Reformat all gn files in /ios/chrome/browser/ui/infobars.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sczs@chromium.org

Bug: 1041419
Change-Id: If2ccf7d6589cf1e7fc1d8492de8cf4eccfb34026
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997821
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730680}