Reformat all gn files in /chromeos/dbus/kerberos.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=ljusten@chromium.org

Bug: 1041419
Change-Id: I4afa0dfa554026b374db811152a67a54fb0bbb88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997789
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>
Commit-Queue: Lutz Justen <ljusten@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730681}