Reformat all gn files in /build/linux.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=thomasanderson@chromium.org

Bug: 1041419
Change-Id: Icffbf27d3fa07a38e4b6005c6f84423b5c1940de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997987
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730688}