Reformat all gn files in /content/public/android.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=boliu@chromium.org

Bug: 1041419
Change-Id: I5273bcc7a32b634c296d584f70045db74b47647f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995818
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730690}