Reformat all gn files in /chrome/browser/nacl_host.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=dschuff@chromium.org

Bug: 1041419
Change-Id: Ibd5a9efe48f2c2733923056c5b82919fc53124e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997960
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Derek Schuff <dschuff@chromium.org>
Commit-Queue: Derek Schuff <dschuff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730692}