Reformat all gn files in /third_party/blink/renderer/platform/blob.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=mek@chromium.org

Bug: 1041419
Change-Id: I5bdfbfd31d41aac7cb082c50bac916da9ed7e0a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995515
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730693}