Reformat all gn files in /third_party/openxr.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=alcooper@chromium.org

Bug: 1041419
Change-Id: I78b2bcbe6230791219d6c6b4464f8e1fad250569
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998209
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730694}