Reformat all gn files in /chrome/test/mini_installer.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=grt@chromium.org

Bug: 1041419
Change-Id: Iad33ff4054db54f1cfec89232f38e0ae7e208fc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997913
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730697}