Reformat all gn files in /ios/chrome/browser/ui/authentication.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jlebel@chromium.org

Bug: 1041419
Change-Id: Ic92988ed786e376fe0f80c2c3d8e933ca64b425b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995813
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730704}