Reformat all gn files in /components/dbus/properties.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=thomasanderson@chromium.org

Bug: 1041419
Change-Id: I7433395e8009c1baf9daac14091f5d9236da58c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994514
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730705}