Reformat all gn files in /components/about_ui.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=aee@chromium.org

Bug: 1041419
Change-Id: I8ec8f1f079396b5ffc9f5c9209e1bb62ef478b56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997741
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Commit-Queue: Esmael Elmoslimany <aee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730709}