Reformat all gn files in /device/bluetooth/strings/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=rnk@chromium.org

Bug: 1041419
Change-Id: If1c5861710e73388ba054b4ba6ba58bbaa7ec4f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998339
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Reid Kleckner <rnk@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730714}