Reformat all gn files in /components/page_load_metrics.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jkarlin@chromium.org

Bug: 1041419
Change-Id: I4f13da6b34568bedeef68f047a8ada9f46dd4b4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997753
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Commit-Queue: Josh Karlin <jkarlin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730720}