Reformat all gn files in /components/network_time.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=carlosil@chromium.org

Bug: 1041419
Change-Id: I457a0e6093d2176f18aa5998d5a175afa4f89f33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998068
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Commit-Queue: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730721}