Reformat all gn files in /components/favicon.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sky@chromium.org

Bug: 1041419
Change-Id: Ic76471b6885c3b472e632e1305274ea97e60f654
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997964
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730725}