Reformat all gn files in /chrome/browser/resources/omnibox.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=orinj@chromium.org

Bug: 1041419
Change-Id: I0272eb8e4547f1c7e0cc2a2491d41f892e3bbeee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998043
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730726}