Reformat all gn files in /chrome/browser/metrics/perf.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=gmx@chromium.org

Bug: 1041419
Change-Id: I51432b9b402a20c03d00739e2390f33904643c77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995375
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Gabriel Marin <gmx@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730735}