Reformat all gn files in /components/services/storage.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=rockot@google.com

Bug: 1041419
Change-Id: I402d95b7ef4a446f4f77067171a36f53cd5fe9e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995817
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#730738}