Reformat all gn files in /ios/chrome/browser/ui/table_view.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=rohitrao@chromium.org

Bug: 1041419
Change-Id: I676be0c3e35616e09dd418580b0a0d3811b89538
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998064
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730739}