Reformat all gn files in /crypto.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=davidben@chromium.org

Bug: 1041419
Change-Id: I44e46869f4ee69cd30635c5ac4a0b1c66c467de8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997917
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730742}