Reformat all gn files in /chromeos/network.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=tbarzic@chromium.org

Bug: 1041419
Change-Id: Ieb6b36c3d3cc6d798e8ff119f70dc863a3efa656
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998057
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730747}