Reformat all gn files in /courgette.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=wfh@chromium.org

Bug: 1041419
Change-Id: I206e7e4c8e89f16b2349ed1cd361a63524e688a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1996999
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730753}