Reformat all gn files in /ios/chrome/browser/ui/dialogs.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=kkhorimoto@chromium.org

Bug: 1041419
Change-Id: I8aaa6e7933b5b398d4ae539d72a1037c208a8da6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995855
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730755}