Reformat all gn files in /chrome/browser/resources/offline_pages.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=chili@chromium.org

Bug: 1041419
Change-Id: I04afdc2bee22aa7b471e818140a18838915b7f8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997738
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Cathy Li <chili@chromium.org>
Commit-Queue: Cathy Li <chili@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730756}