Reformat all gn files in /ios/chrome/test.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=eugenebut@chromium.org

Bug: 1041419
Change-Id: Iabf3cc86820f3037cf1f75437c474b959c844d39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998340
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730761}