Reformat all gn files in /content/public/app.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=rockot@google.com

Bug: 1041419
Change-Id: Ibfda06a9304e7ded249c5fe28f337d0ca2b84ec8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994459
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#730772}