Reformat all gn files in /third_party/libpng.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=scroggo@chromium.org

Bug: 1041419
Change-Id: I49f16669fa55af707d6e9f63476835fb82a835b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997903
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Leon Scroggins <scroggo@chromium.org>
Commit-Queue: Leon Scroggins <scroggo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730773}