Reformat all gn files in /components/storage_monitor.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=tommycli@chromium.org

Bug: 1041419
Change-Id: I9013d0b8c6ef76cae32e17b95bb648b70a4a12be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998156
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730775}