Reformat all gn files in /chromeos/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=ellyjones@chromium.org

Bug: 1041419
Change-Id: I4098048975177e4021b0ffd4988ddb0ecd8a33a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998307
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730776}