Reformat all gn files in /ui/accelerated_widget_mac.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sdy@chromium.org

Bug: 1041419
Change-Id: I6480564d70e72089ee693c574b59e1f94b0308fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998201
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Sidney San Martín <sdy@chromium.org>
Commit-Queue: Sidney San Martín <sdy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730778}