Reformat all gn files in /components/services/pdf_compositor.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=weili@chromium.org

Bug: 1041419
Change-Id: I45df05de12463ad06920a470132f04e7c0645b1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994452
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Wei Li <weili@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730779}