Reformat all gn files in /chrome/browser/notifications/scheduler.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=xingliu@chromium.org

Bug: 1041419
Change-Id: I85fe99f3ae0af72e132d30fb10531980d4ec100b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998285
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Commit-Queue: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730780}