Reformat all gn files in /chrome/browser/resources/tab_strip.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=johntlee@chromium.org

Bug: 1041419
Change-Id: I526ef0899c4e707ba32dd78c3d7d7382c8ee6d19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998073
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730782}