Reformat all gn files in /chrome/browser/ui/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=rnk@chromium.org

Bug: 1041419
Change-Id: I36235d519936da73a9effbf83a1a3bc0fb05d01c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994448
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Reid Kleckner <rnk@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730783}