Reformat all gn files in /gpu/command_buffer.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=bajones@chromium.org

Bug: 1041419
Change-Id: I1e0f8766acdd08d3e472dbc846809ca1007eb42c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997746
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730786}