Reformat all gn files in /components/module_installer.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=wnwen@chromium.org

Bug: 1041419
Change-Id: I6adb8d4b11646760fe515a65c1c47d1ad41ba849
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998498
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Commit-Queue: Peter Wen <wnwen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730789}