Reformat all gn files in /chrome/android/features/keyboard_accessory/factory/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=wnwen@chromium.org

Bug: 1041419
Change-Id: I6bc48243420c0b135e3dd91e3d5067e9e73b7e06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998435
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Commit-Queue: Peter Wen <wnwen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730793}