Reformat all gn files in /build/config/freetype.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=bungeman@chromium.org

Bug: 1041419
Change-Id: I33ffff8ff9f431f01eba59d4e857fb05cf657a05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995514
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Ben Wagner <bungeman@chromium.org>
Commit-Queue: Ben Wagner <bungeman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730801}