Reformat all gn files in /chrome/android/modules/chime/internal/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jbudorick@chromium.org

Bug: 1041419
Change-Id: I6d8f90d260e2c85028629600fa8ebee9196493d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997705
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: John Budorick <jbudorick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730807}