Reformat all gn files in /components/upload_list.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=mark@chromium.org

Bug: 1041419
Change-Id: Ib4113905e5dd1ccbbf6d1b46ac04d96b55e09d11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998206
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730809}