Reformat all gn files in /chrome/services/media_gallery_util.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=tommycli@chromium.org

Bug: 1041419
Change-Id: Ie20659a371b62b667f1b01289423eb950c7e7c16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998025
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730812}