Reformat all gn files in /components/component_updater.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sorin@chromium.org

Bug: 1041419
Change-Id: I5ce57fdea4d93ccbaadda76acf903ae8851f15b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997914
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730815}