Streamline and share more jsdialog code.
The interface JavaScriptNativeDialogFactory is replaced by a callback
to reduce verbosity.

For Android, WebContents activation is moved from the factory function
into the ctor of JavascriptAppModalDialogAndroid. This allows it to be
shared. Also, remove ChromeJavascriptAppModalDialogAndroid and just use
JavascriptAppModalDialogAndroid because the former's purpose was to
facilitate UMA, and that UMA was removed in a7d41eae009d1b26d110fb56ec6

Bug: 1025256
Change-Id: Id677208c1ecab28b45d8401ac943b259f1a16180
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1988850
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730816}