Reformat all gn files in /ios/components.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=eugenebut@chromium.org

Bug: 1041419
Change-Id: I7f5bc85c6349a2421d8c4ddcecad509e85bf14dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997837
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730817}