Reformat all gn files in /ios/web/favicon.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=gambard@chromium.org

Bug: 1041419
Change-Id: Ia6414bc1cb147312b8dbedfd8ebde5ff830ae938
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997972
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730820}