Reformat all gn files in /components/cast_certificate.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=spang@chromium.org

Bug: 1041419
Change-Id: Ia6c1e085c63f2dd30b0a664b794131f1b4b6dae0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998423
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Michael Spang <spang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730824}