Reformat all gn files in /chromecast/external_mojo.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=kmackay@chromium.org

Bug: 1041419
Change-Id: I0b718719c39034ee3b4c568a0ffffc4752e95e0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998506
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Kenneth MacKay <kmackay@chromium.org>
Commit-Queue: Kenneth MacKay <kmackay@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730825}