Reformat all gn files in /chrome/browser/resources/safe_browsing.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=asanka@chromium.org

Bug: 1041419
Change-Id: Ie33805911c68e24b2e4d7d05d4a311dee8541b63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997969
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Asanka Herath <asanka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730826}