Reformat all gn files in /components/scheduling_metrics.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=fdoray@chromium.org

Bug: 1041419
Change-Id: I07381c44cb76b93f46e811ba42e39ecd2b002b2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998030
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730827}