Reformat all gn files in /components/printing.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=weili@chromium.org

Bug: 1041419
Change-Id: Id23e45e39b6f8bbbb3eac331a63b4f7f2e915b84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998266
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730830}