Reformat all gn files in /ash/components/fast_ink.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=kaznacheev@chromium.org

Bug: 1041419
Change-Id: Ib2fc5bcc6b423a0683fbfba3b47e672905b2f50d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998631
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Vladislav Kaznacheev <kaznacheev@chromium.org>
Commit-Queue: Vladislav Kaznacheev <kaznacheev@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730842}