Reformat all gn files in /extensions/browser/api/printer_provider_internal.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=tbarzic@chromium.org

Bug: 1041419
Change-Id: I96edaa53571933286a6ed5ceacdb66c29119c8f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998443
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730843}