Reformat all gn files in /chrome/browser/ui/webui/discards.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=fdoray@chromium.org

Bug: 1041419
Change-Id: Ia02279eaacfda3c8acb77a209eefd7d5904afa40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995512
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730844}