Reformat all gn files in /base/util/memory_pressure.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sebmarchand@chromium.org

Bug: 1041419
Change-Id: I48cabac19dd2af422cb87c45b51ef6fdaab3dad5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998286
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Sébastien Marchand <sebmarchand@chromium.org>
Commit-Queue: Sébastien Marchand <sebmarchand@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730847}