Reformat all gn files in /ios/chrome/browser/ssl.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=felt@chromium.org

Bug: 1041419
Change-Id: I35d844ab6c053659ec19f13f522a69cb5eafeda8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997797
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Adrienne Porter Felt <felt@chromium.org>
Commit-Queue: Adrienne Porter Felt <felt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730854}