Reformat all gn files in /build/android/buildhooks.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=estevenson@chromium.org

Bug: 1041419
Change-Id: Ib713adb3d9a56fdba15164c887b72e5c553ca040
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998624
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Eric Stevenson <estevenson@chromium.org>
Commit-Queue: Eric Stevenson <estevenson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730858}