Reformat all gn files in /content/renderer/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=pdr@chromium.org

Bug: 1041419
Change-Id: I2d13bf5c8bb9387c0707d7829fe485b2596f433a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994058
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730865}