Reformat all gn files in /components/nacl.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=bradnelson@chromium.org

Bug: 1041419
Change-Id: I7310bbba217b453a6032d1e31a9b06cf467ee5d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998499
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Brad Nelson <bradnelson@chromium.org>
Commit-Queue: Brad Nelson <bradnelson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730867}