Reformat all gn files in /components/viz.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jonross@chromium.org

Bug: 1041419
Change-Id: Iae1cea16661734334126d0d21781e5bf9526075f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998425
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730872}