Reformat all gn files in /components/download/quarantine.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=asanka@chromium.org

Bug: 1041419
Change-Id: I07f009518b4d547715c8eefbc70772a25e306695
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998326
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Asanka Herath <asanka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730880}