Reformat all gn files in /ui/display/util.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=dcastagna@chromium.org

Bug: 1041419
Change-Id: I71a55972da9b349a1f8c87acf9dc8b426750938c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998504
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730883}