Reformat all gn files in /ui/base/ime/fuchsia/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=kmarshall@chromium.org

Bug: 1041419
Change-Id: I39aea7c2f27c9d6a0120dd89f1735262d16f9e9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998686
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Kevin Marshall <kmarshall@chromium.org>
Commit-Queue: Kevin Marshall <kmarshall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730886}