Reformat all gn files in /chrome/test/media_router.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=mfoltz@chromium.org

Bug: 1041419
Change-Id: I9229b61620dd6dc329f6c05b20e8a44cc6b50f96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997820
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730887}