Reformat all gn files in /chrome/browser/image_editor.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=gayane@chromium.org

Bug: 1041419
Change-Id: I6d0c9ee7c70d53e734d1e1002c3e039bf9e0cd77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998750
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Gayane Petrosyan <gayane@chromium.org>
Commit-Queue: Gayane Petrosyan <gayane@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730889}