Reformat all gn files in /components/image_fetcher.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=wylieb@chromium.org

Bug: 1041419
Change-Id: Ic9677792e928692188accb4070684f5cc35bd8f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997805
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Commit-Queue: Brandon Wylie <wylieb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730890}