Reformat all gn files in /third_party/blink/renderer/modules/payments.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sahel@chromium.org

Bug: 1041419
Change-Id: Ife49d512f0efad68e17bdd7e0833331dd13269e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998438
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Sahel Sharify <sahel@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730897}