Reformat all gn files in /ios/chrome/browser/ui/content_suggestions.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=gambard@chromium.org

Bug: 1041419
Change-Id: I3d9f2d1c960c0a753bfbcf8c24268e699d95bf98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998282
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730898}