Reformat all gn files in /ios/third_party/motion_animator_objc/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=ellyjones@chromium.org

Bug: 1041419
Change-Id: I42c1ecaba4da78b88c762f135b8453fd8c556939
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997791
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730903}