Reformat all gn files in /components/cronet.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=agl@chromium.org

Bug: 1041419
Change-Id: Ie77c19735a2df0e68f7cfd5ab16d4b7cbb3d3f09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995786
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730904}