Reformat all gn files in /components/network_hints.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=mattm@chromium.org

Bug: 1041419
Change-Id: I6fc00babc06e91bff038f861de8851637d510d41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997962
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Matt Mueller <mattm@chromium.org>
Commit-Queue: Matt Mueller <mattm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730915}