Reformat all gn files in /extensions/browser/api/networking_private.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=stevenjb@chromium.org

Bug: 1041419
Change-Id: Iceb59bcb8bcdaf87c0df9b2a70f4b0f02503ac94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998261
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730917}