Reformat all gn files in /components/user_manager.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=xiyuan@chromium.org

Bug: 1041419
Change-Id: I682da943be78f7552e3599552f8dcb10ab73f55b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998501
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730919}