Reformat all gn files in /ios/chrome/browser/webui.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=michaeldo@chromium.org

Bug: 1041419
Change-Id: Id1b9ec569f5ea61bfe99ee97f2e77570f1b1a5ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998033
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730920}