Reformat all gn files in /content/browser/cache_storage.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jsbell@chromium.org

Bug: 1041419
Change-Id: I0d5089e4c5ce6d07cbeade73cae75ff7777e5099
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998641
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730925}