Reformat all gn files in /ui/color.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=pkasting@chromium.org

Bug: 1041419
Change-Id: Ifa634f5939f17473ee3a01f4c4af538c3fa4cfea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994437
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730930}