Reformat all gn files in /extensions/browser/api/hid.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=reillyg@chromium.org

Bug: 1041419
Change-Id: I9b6e47cc5b373c6e2b186a2185bfec96cbf841e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998627
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730932}