Damage Tracker calculates if a surface rect intersects any damage from below during its update
Bug:1035639

The purpose of this changelist is to help improve the performance on the ui when there is background blur. The related bug is https://bugs.chromium.org/p/chromium/issues/detail?id=1035639
The general idea is to avoid recomputation of a background blur if the damaged regions under the blur surface don't intersect the surface, as any damage coming from above the surface doesn't contribution to the re-calculation of backdrop blur.
A flag can_use_cached_backdrop_filtered_result_ is added to RenderSurfaceImpl to indicate if backdrop filter cache from last update can be used.
The major changes here in this changelist are in DamageTracker and they are quite contained. It will be combined with another couple of changelists on GL side to fulfill the end goal.

Description of the approach:

This changelist is based on the presumption that any render surface has to have at least one layer contributing to it. In other words, in the 'tree' where all layers and surfaces pointing to their respective render targets, leaves can only be layers and vice versa, non-leaves can only be surfaces and vice versa.
Even though the changes here are to serve to performance improvement of backdrop blur filters, they are written in a generic way so you can use the functions for other purposes where you see fit.
To achieve the goal requires the damage tracker to be able to tell if any damage coming from below intersects a rect on the surface in question. This is done through the walk over all layers and surfaces to accumulate damage in DamageTracker::UpdateDamageTracking() [1]. As we know, when anytime we visit a surface in [1], all its ancestor render targets on the chain to the root have already accumulated all the damage from layers and surfaces below this surface. Whether this surface has accumulated damage on top itself is not relevant. So at this point, we only have to walk up the ancestor render target chain to root to see if any damage rect intersects the rect in question (backdrop filter surface rect).
However this is not sufficient, as some layers and render surfaces that were in the previous update can be removed and no longer exist in the current update. If any of these layers and surfaces were below the surface in question, they need to be considered as well. This requires us to know the order of layers and surfaces when they are drawn to the same render target. I call these layers and surfaces siblings and added a prev_sibling_ and next_sibling_ pointers in RectMapData so each layer and surface in rect_history_for_layers_ and rect_history_for_surfaces_ points to both its immediate older and younger siblings. These pointers are to help track the draw order of siblings. When you go through the siblings through the prev_sibling_ pointer one by one, you get all surfaces and layers drawn before the current surface. This is how I did the damage intersection check. (There are more details. Please see code.)
When a new layer or surface is added, the prev_sibling_ and next_sibling_ pointers have to be hooked up. This is done when each layer is iterated in [1]. Since we know the layers are visited once and only once in each update, I hook up the sibling pointers for the new layer when it is visited and walk up its ancestor render target chain to hook up the sibling pointers of each target until root or a target that already had the pointers hooked up. This ensures all layers and surfaces are hooked up.

Possible further improvement:
Backdrop filters sometimes have bounds (or clipping?), so maybe we could
use the bounds or clipped region to do the intersection check with
damage below the surface and the backdrop filter cache thus could take less space.

Change-Id: I27e593ffce0046aa60c3eeca76961ce10e9511ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1988155
Commit-Queue: Jun Liu <yjliu@chromium.org>
Auto-Submit: Jun Liu <yjliu@chromium.org>
Reviewed-by: weiliangc <weiliangc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730933}