Reformat all gn files in /remoting.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=yuweih@chromium.org

Bug: 1041419
Change-Id: I0294a2309b33a72deba3f518f6efad470dc10ab0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998786
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730944}