Reformat all gn files in /chrome/browser/resources/usb_internals.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=reillyg@chromium.org

Bug: 1041419
Change-Id: I18ccdffc4fc953b70ed6389c0e3f24809f2301ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998649
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730950}