Reformat all gn files in /chrome/browser/resources/downloads.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=dbeam@chromium.org

Bug: 1041419
Change-Id: I56203e5f4245372031c8d5120d552a5b33caa9a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997764
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Dan Beam <dbeam@chromium.org>
Commit-Queue: Dan Beam <dbeam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730952}