Reformat all gn files in /components/optimization_guide.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=rajendrant@chromium.org

Bug: 1041419
Change-Id: Ic53a705035bee6656c7195b95295fcf69f4f08d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998637
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: rajendrant <rajendrant@chromium.org>
Commit-Queue: rajendrant <rajendrant@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730962}