Reformat all gn files in /third_party/blink/renderer/core/html/parser.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=yoavweiss@chromium.org

Bug: 1041419
Change-Id: I7fff56ed3f06009203fa9cf81fcf4d71577c3b3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998314
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730963}