Reformat all gn files in /services/identity.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=msarda@chromium.org

Bug: 1041419
Change-Id: I8cd0ca11583d6c403b79c1d4b8925d710f402dcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995857
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730967}