Reformat all gn files in /extensions/browser/guest_view.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=ekaramad@chromium.org

Bug: 1041419
Change-Id: I5ecbae9d7f892c8dca80c002daa45c04459614c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998920
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730970}