Reformat all gn files in /third_party/yasm.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=dalecurtis@chromium.org

Bug: 1041419
Change-Id: Ie028bb68b57fb36845f4825459257be95b848685
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998270
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730981}