Reformat all gn files in /ash/app_list.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jennyz@chromium.org

Bug: 1041419
Change-Id: I9ffe3d86934385cec5ba115267c6a064d7cd57b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998978
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Commit-Queue: Jenny Zhang <jennyz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730983}