Reformat all gn files in /third_party/minigbm.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=dnicoara@chromium.org

Bug: 1041419
Change-Id: I38a770cc4fb885eb49e7d6356e99ea4c50fa84da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997713
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Commit-Queue: Daniel Nicoara <dnicoara@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730985}