Reformat all gn files in /ui/webui/resources/cr_components/chromeos/cellular_setup.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=khorimoto@chromium.org

Bug: 1041419
Change-Id: Icb6f30e6c7d6070efa6d732a88dfcc21ae02da55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998712
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730992}