system-web-apps: Add support for chrome-untrusted:// to WebUI's URL loader
This allows us to serve code that parses untrusted content from a
different origin without any of the privileges that chrome:// URLs get.

Note, this only adds support for chrome-untrusted:// DataSources and not
for chrome-untrusted:// WebUIControllers. WebUIControllers are used to
grant extra privileges and APIs to renderers i.e. bindings, Mojo APIs,
etc. and we should be very careful when adding support for them. Since
current clients don't strictly need these capabilities yet, we will add
support for WebUIControllers later.

The plan is to first add support for chrome-untrusted:// to DataSource.
This lets teams that will use the untrusted context migrate over to it
asap and make sure all of their use cases are covered and it shouldn't
introduce any security risks. Then start work on cleaning up
WebUIController so that it can be used for chrome-untrusted://.

Bug: 1012150,1035300
Change-Id: I5ea20090e803734976608f21b774829a24b634e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1898159
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731012}