Reformat all gn files in /third_party/blink/renderer/modules/hid.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=reillyg@chromium.org

Bug: 1041419
Change-Id: I4abdcd91cd82ff2d97f8dc41a4ff12bb186d5cd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997814
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731013}