Reformat all gn files in /components/discardable_memory.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=penghuang@chromium.org

Bug: 1041419
Change-Id: Ibb04065be6174bdc1a3a2d2f4ade9cb7b46f1efa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998303
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Peng Huang <penghuang@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731017}