Reformat all gn files in /media/capture/video/chromeos.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=wtlee@chromium.org

Bug: 1041419
Change-Id: I724e3fc788403459dac134ad589ca1b63c3f648c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997766
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>
Commit-Queue: Shik Chen <shik@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731027}