Reformat all gn files in /media/gpu/vaapi.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=posciak@chromium.org

Bug: 1041419
Change-Id: I9d129dfd4e48234b62ab6705a7c64e55bf8a69e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998044
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731028}