Reformat all gn files in /chrome/browser/resources/chromeos/camera.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jcliang@chromium.org

Bug: 1041419
Change-Id: I26ce808571f6f6469795b55fb37b5a1d82fefc75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997992
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>
Commit-Queue: Shik Chen <shik@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731038}