Reformat all gn files in /third_party/cacheinvalidation.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=nyquist@chromium.org

Bug: 1041419
Change-Id: I4a516e0f2eecd6e4d5f4ca9249ccad6ffcd196be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997930
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Tommy Nyquist <nyquist@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731046}