Reformat all gn files in /chrome/browser/devtools.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=dgozman@chromium.org

Bug: 1041419
Change-Id: Ie424cf8fc39636cc61d494f76a7cfd234f4f1520
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998878
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731049}