Reformat all gn files in /chrome/services/local_search_service.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=thanhdng@chromium.org

Bug: 1041419
Change-Id: I630ae75c8af674f23dd1730c2efd981c6ce0213a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998204
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731051}