Reformat all gn files in /chromeos/disks.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=slangley@chromium.org

Bug: 1041419
Change-Id: Id0ed245af9015f342b2d98305ca43cf7e788f38d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994436
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Stuart Langley <slangley@chromium.org>
Commit-Queue: Stuart Langley <slangley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731070}