Reformat all gn files in /components/dbus.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=hashimoto@chromium.org

Bug: 1041419
Change-Id: I95ddf74755a36c4cfa8b0c8b249f407a311dcea5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997912
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Ryo Hashimoto <hashimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731073}