Reformat all gn files in /chromeos/dbus/cryptohome.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=hashimoto@chromium.org

Bug: 1041419
Change-Id: Ic6a4512dc95f61300c14b1eb9781eb9c654f93c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998729
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Ryo Hashimoto <hashimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731074}