Reformat all gn files in /third_party/pywebsocket.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=yhirano@chromium.org

Bug: 1041419
Change-Id: I14572d278bc0ecf3752a2f1b0c1bcc238e455faf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994446
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731083}