Reformat all gn files in /components/offline_pages.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=dewittj@chromium.org

Bug: 1041419
Change-Id: I382f1b80648b2bce3a3e35b012f53b167d4f70f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997706
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Justin DeWitt <dewittj@chromium.org>
Commit-Queue: Justin DeWitt <dewittj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731085}