Reformat all gn files in /components/dom_distiller.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=wychen@chromium.org

Bug: 1041419
Change-Id: If5f7e519d44cf951607b29c2b731a4b161b689f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995796
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Commit-Queue: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731087}