Reformat all gn files in /third_party/sqlite.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=pwnall@chromium.org

Bug: 1041419
Change-Id: I6172dfb3d12e0259168180b710c8cf9d640a52e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998159
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Commit-Queue: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731090}