Reformat all gn files in /chrome/common/safe_browsing/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=haraken@chromium.org

Bug: 1041419
Change-Id: Ieac1effb679db05a24b2b40c8d674f1be854e635
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997801
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731095}