Reformat all gn files in /ios/chrome/browser/ui/context_menu.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=michaeldo@chromium.org

Bug: 1041419
Change-Id: I128bc74f5f65f6122eaf3cd0357556967d7d702c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997995
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731099}