Reformat all gn files in /extensions/browser/api/file_system.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sammc@chromium.org

Bug: 1041419
Change-Id: I192d719fffdd98ae397a749b65e0d8ff9ad4e2ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997830
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731102}