Reformat all gn files in /media/gpu.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=acourbot@chromium.org

Bug: 1041419
Change-Id: Ida4ed30caf6b2dd5db8282cb6d8da39688dd7b5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998198
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731122}