Fix root window size computation to produce correct size with fractional scale.
* Use gfx::ToEnclosedRectIgnoringError when computing dp bounds of
 the display because it may end up a very tiny amount smaller due to
 the error.

* Do not recompute the root window size using matrix on device.
This was added to support "resizing the host window" for dev purpose
when running on linux desktop, and it doesn't work well with
repeating-decimal scale factors. It can lead to 1px smaller size.
I could have used the same ToEnclosedRectIgnoringError, but since
this isn't unnecessary for device, I just used the original one
when running on the cros device.

Bug: 946509
Test: covered by unittest. also tested on device with other upcoming CLs.
Change-Id: Iee68470c600cc5067a4794bec380a3525bfd0e38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1976721
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731130}