Reformat all gn files in /testing/gtest.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=pwnall@chromium.org

Bug: 1041419
Change-Id: I6598bbc2bf59209c942da418a2db8af51a6776f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997818
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731140}