Reformat all gn files in /third_party/blink/renderer/core/fetch.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=horo@chromium.org

Bug: 1041419
Change-Id: Ic7a24216d1a7c462c92afe0679e14104e8a2e036
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998100
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731142}