Reformat all gn files in /ui/base/ime.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=yhanada@chromium.org

Bug: 1041419
Change-Id: I59006a7aaacad9812afd78061a0522da66fdc3cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998420
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Yuichiro Hanada <yhanada@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731146}