Reformat all gn files in /base/util/values.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jdoerrie@chromium.org

Bug: 1041419
Change-Id: I19f90b2453cf21303876ab8c6ff69b0c65ab8d7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998176
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731172}