[Autofill Assistant] Support launch from normal Chrome Tab
This change is gated behind the same flag as tab switching. The startup
of the AA stack is delayed until navigation has finished, to prevent
timing issues where sometimes scripts were first searched on the SRP
before attempting to start the script on the actual page.

This should be considered a "good enough" solution for M81 / M82, while
we work on a more long-term integration of the feature.

Bug: b/145195894
Bug: b/145106184
Change-Id: I071c947cd90152d3b5c8e2995e11339a5cc33e64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967462
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/master@{#731177}