Reformat all gn files in /content/browser/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=hans@chromium.org

Bug: 1041419
Change-Id: Ie592b54a2efbd0a7c2655df6706b19770fbf3e4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995799
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731181}