Reformat all gn files in /third_party/blink/public/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=hans@chromium.org

Bug: 1041419
Change-Id: Ic416ba8fd7210cb39beaf0e0380cc2261ca37e7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998787
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731183}