Reformat all gn files in /components/services/font.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=drott@chromium.org

Bug: 1041419
Change-Id: I6887b59317f166ea304b6e343ff972824b722874
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998253
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731188}