Reformat all gn files in /components/consent_auditor.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=dullweber@chromium.org

Bug: 1041419
Change-Id: Ib8ad603950701d887e9e4259c499aa5d9ffd1b40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998674
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731193}