Reformat all gn files in /ios/chrome/browser/ui/settings/credit_card_scanner.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=javierrobles@chromium.org

Bug: 1041419
Change-Id: Ia9cc5f6511786dfe123dd46fadee4a589dd55e3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998725
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731194}