Reformat all gn files in /chromeos/ime.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=shuchen@chromium.org

Bug: 1041419
Change-Id: I0ae8cd75de0b025c0d3b622b2336e708579ab873
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998630
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Shu Chen <shuchen@chromium.org>
Commit-Queue: Shu Chen <shuchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731200}