Reformat all gn files in /ios/chrome/common/colors.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=javierrobles@chromium.org

Bug: 1041419
Change-Id: Icc49599dd3bfa5f60571eba35e79d6ba24516a84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994056
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731207}