[FlexNG] When determining the IntrinsicBlockSize make percentages "auto"
This fixes:
 flex-minimum-height-flex-items-012.html
 flex-minimum-height-flex-items-015.html

Previously we were using the same constraint space for determining
flex-basis, and performing layout to determine the intrinsic block-size.

When resolving flex-basis lengths, we should be resolving percentages
(if possible).

However for performing the layout to determine the intrinsic block-size
we should be resolving percentages against an indefinite size.

This separates this space building into two separate methods.

Bug: 845235
Change-Id: I434082e0d36f5706daeaaf1e451528fe8403fd7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1982188
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: David Grogan <dgrogan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731212}