Reformat all gn files in /chrome/browser/resources/browser_switch.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=pastarmovj@chromium.org

Bug: 1041419
Change-Id: I847a5fd4d7a426e03710c534c5f5b849e2d664db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998639
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Julian Pastarmov <pastarmovj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731214}