Reformat all gn files in /media/audio.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=henrika@chromium.org

Bug: 1041419
Change-Id: I8a1a3f2749169511725749b113ebf7a745b2d875
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998039
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Olga Sharonova <olka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731221}