Reformat all gn files in /ios/chrome/browser/context_menu.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=noyau@chromium.org

Bug: 1041419
Change-Id: I05be779515059567459b7b5c23a1b91bedd737b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998780
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Eric Noyau <noyau@chromium.org>
Commit-Queue: Eric Noyau <noyau@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731226}