Reformat all gn files in /services/device/screen_orientation.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=blundell@chromium.org

Bug: 1041419
Change-Id: I7c0d2e5f4881adb04e435b75c641afecfe7c1ea8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997743
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731234}