Reformat all gn files in /third_party/blink/renderer/modules/vibration.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=mvanouwerkerk@chromium.org

Bug: 1041419
Change-Id: I84408fec10e8f1e24e3895194938c3333165d762
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998683
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Commit-Queue: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731235}