Reformat all gn files in /services/tracing.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=skyostil@chromium.org

Bug: 1041419
Change-Id: I2b79b56155a91eb333b227d7d6c8ee1e84e3363b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998575
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Sami Kyöstilä <skyostil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731241}