[fetch] Replace uses of DeprecatedByteLengthAsUnsigned
This CL replaces calls to DeprecatedByteLengthAsUnsigned by calls to
ByteLengthAsUnsigned. By switching to size_t the byte length may exceed
the maximum supported length of FormDataByteConsumer. Therefore I throw
RangeErrors when the ArrayBuffer exceeds the maximum supported size.

Background: we prepare ArrayBuffers to be bigger than 4GB. Therefore we
changed the size field to size_t. Now we are changing all uses of
ByteLength to be able to deal with size_t, either by accepting a size_t,
or by throwing an exception if the size is too big.

R=yhirano@chromium.org

Bug: chromium:1024228, chromium:1008840
Change-Id: Id2fe3259d3c8fa68bfe9e5bd8c531f1d3329e043
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994294
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731258}