Reformat all gn files in /chrome/browser/ui/webui/app_management.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=dominickn@chromium.org

Bug: 1041419
Change-Id: Ia6e24c4f721a39bc43e8cfc321d003c609acaf35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998508
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731263}