Reformat all gn files in /components/arc.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=hidehiko@chromium.org

Bug: 1041419
Change-Id: I9d88478136ff6a5365385522a07ca36ed543de9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997979
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731273}