Reformat all gn files in /chrome/browser/vr.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=vollick@chromium.org

Bug: 1041419
Change-Id: I7470b9bda84fe9f814135fb4596799b9b4ddadd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998290
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731276}