Reformat all gn files in /chrome/browser/notifications.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=yoshiki@chromium.org

Bug: 1041419
Change-Id: I3b44a28686d10b170456b355e781cfc33ede4328
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997811
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Yoshiki Iguchi <yoshiki@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Peter Beverloo <peter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731279}