Reformat all gn files in /third_party/unrar.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=drubery@chromium.org

Bug: 1041419
Change-Id: I7806bac4c1e4d1b3263f03937dafe76a506073d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998749
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731300}