Reformat all gn files in /components/drive.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=hirono@chromium.org

Bug: 1041419
Change-Id: I05ad6ee34f025b1a6c94eff7addf34a541d00b0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998168
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Daichi Hirono <hirono@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731304}