Reformat all gn files in /third_party/breakpad.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=thestig@chromium.org

Bug: 1041419
Change-Id: I8e4c814d593bc3f08b7cf5aa50d8f9d5c3681e1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998939
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731307}