Reformat all gn files in /ios/chrome/search_widget_extension.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=olivierrobin@chromium.org

Bug: 1041419
Change-Id: If6c1eeb0224931d3460af1a7223d42760eb32d6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997751
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731309}