Reformat all gn files in /tools/polymer.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=dpapad@chromium.org

Bug: 1041419
Change-Id: I77c1df24128f8491669467cfdf76727334ad934e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998259
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731312}