Reformat all gn files in /third_party/ots.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=bashi@chromium.org

Bug: 1041419
Change-Id: I3b7cba390b7d10b6476f12d5807c714e32eaefeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998346
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731314}