Reformat all gn files in /ui/android.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=dtrainor@chromium.org

Bug: 1041419
Change-Id: I0db5eff9c826a1e1b5fdecb974341deb788038a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998865
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731319}