Reformat all gn files in /ios/chrome/browser/ui/promos.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=marq@chromium.org

Bug: 1041419
Change-Id: I8f46db2668d1022595cdae274bb64863c4cf1ba8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998942
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731320}