Reformat all gn files in /services/content.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=clamy@chromium.org

Bug: 1041419
Change-Id: I47d1d318ba7b3c1a3e1c6136295742fcccc10397
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994454
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Commit-Queue: Camille Lamy <clamy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731321}