Reformat all gn files in /services/device/public/cpp/power_monitor.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=bajones@chromium.org

Bug: 1041419
Change-Id: I0fb543a75d96ee7134e215cd01ae2bff3591983e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995790
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731322}