Reformat all gn files in /components/paint_preview.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=yfriedman@chromium.org

Bug: 1041419
Change-Id: I3cb222aeff16285032082508b6cc80b9ea5ae198
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998563
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731324}