Reformat all gn files in /chrome/notification_helper.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=finnur@chromium.org

Bug: 1041419
Change-Id: Iacf7411a5ec7d3a8ba6509868b5ee5d9ce9634ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997909
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731335}