Reformat all gn files in /ios/chrome/common.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=olivierrobin@chromium.org

Bug: 1041419
Change-Id: I0e0353bfdafe4ee3949b9424a2cf493205f74f81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998049
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731336}