Reformat all gn files in /components/ui_devtools/viz.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=kylechar@chromium.org

Bug: 1041419
Change-Id: I27ad5c2f99878e9f421f136efdcfd1224068cd50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998048
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731347}