Reformat all gn files in /chrome/browser/resources/chromeos/crostini_upgrader.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=nverne@chromium.org

Bug: 1041419
Change-Id: I223be6c578922e30581064f2cc1dc3c882d1fd10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997983
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731356}