Reformat all gn files in /components/language.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=dvallet@chromium.org

Bug: 1041419
Change-Id: Ie64f6438bb52955a329e7f525952c0966bc43e5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994450
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: David Vallet <dvallet@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731359}