Reformat all gn files in /chrome/browser/resources/conflicts.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=chrisha@chromium.org

Bug: 1041419
Change-Id: Ia04feffd95ffd1ce40cbcc57c2aba1769151f60a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997981
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731362}