Reformat all gn files in /android_webview/tools/webview_log_verbosifier/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=ntfschr@chromium.org

Bug: 1041419
Change-Id: I0a234dd19ffb4582c47709efc40a165f71875e2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998860
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731364}