Reformat all gn files in /services/device/public/cpp/bluetooth.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=ortuno@chromium.org

Bug: 1041419
Change-Id: I8ca57b843f4417b45ee1a027484cb627079e4da8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998431
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731367}