Reformat all gn files in /services/viz.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=danakj@chromium.org

Bug: 1041419
Change-Id: I91cf343e926b04fe8e262c3c012d6cb839b9a59a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998322
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731372}