Reformat all gn files in /ios/chrome/browser/web_state_list.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sdefresne@chromium.org

Bug: 1041419
Change-Id: Ifc94d441111386eea772bf54f8cb146dba77dce6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998940
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731381}