Reformat all gn files in /ui/chromeos.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=oshima@chromium.org

Bug: 1041419
Change-Id: I8660470f0b0d59f793a133632883d1eaef6f9fce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998728
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731393}