Reformat all gn files in /ios/chrome/browser/ui/translate.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=mahmadi@chromium.org

Bug: 1041419
Change-Id: Ibef3bb12e7049acd5bcd588231f1a3674835cc80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998963
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731395}