Reformat all gn files in /chrome/install_static.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=gab@chromium.org

Bug: 1041419
Change-Id: Ib98e577f82d11c993a6226b8c7c9aa0192555812
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995844
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731397}