Reformat all gn files in /components/browser_ui.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=tedchoc@chromium.org

Bug: 1041419
Change-Id: I579aa9d533813b8a4f5dd9ca45e6560f1749a8d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998742
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731399}