Reformat all gn files in /chrome/browser/ui/webui/snippets_internals/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=fgorski@chromium.org

Bug: 1041419
Change-Id: Id68e68d33c1b373850ceb32b4a1606f8c95ff99c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998070
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Filip Gorski <fgorski@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731400}