Reformat all gn files in /ui/events.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sadrul@chromium.org

Bug: 1041419
Change-Id: I5bace1563e3ffdf1dc7000988bede5ab08c40399
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994506
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731405}