Reformat all gn files in /components/bookmarks.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sky@chromium.org

Bug: 1041419
Change-Id: Ia3b916c0b65753996d5e1f64b1bff71e02c898ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998783
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731409}