Reformat all gn files in /content/app.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=rockot@google.com

Bug: 1041419
Change-Id: Ie928d40c07e29fc3ac17eabbc0872074b6704ddd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998622
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731416}