Reformat all gn files in /ios/chrome/app.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=rohitrao@chromium.org

Bug: 1041419
Change-Id: I3ed16ea5a118cc73854205575d0fea107ab41505
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998418
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731418}