Reformat all gn files in /content/browser/background_fetch.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=peter@chromium.org

Bug: 1041419
Change-Id: I12871f7507120ec88249e68c74b8321711d266f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998732
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731419}