Reformat all gn files in /chromeos/dbus/shill.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=stevenjb@chromium.org

Bug: 1041419
Change-Id: I06ebb2384c7af9a1026d08ff06eb61ef2ca01e05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1993854
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731422}