[WebLayer] Integrate with AutofillService on Android
This CL wires up integration with the Android system-level
AutofillService in WebLayer. This integration allows for autofill of
HTML forms that WebLayer renders within a ContentView via glueing
Chromium's detection of those forms and their structure to the
system-level AutofillService's ability to provide autofill data for
those forms given their structure.

Concretely, this integration entails the following:

- Bringing up Chromium's classes that serve as glue between Chromium's
  autofill infrastructure and the underlying system-level
  AutofillManager within the context of WebLayer. This bringup is done
  on a per-Tab basis, as the glue classes are effectively
  per-WebContents and hook into the C++-side infrastructure that is
  already brought up in tab_impl.cc.
- Extending ContentView to listen for relevant callbacks from the
  system-level autofill infrastructure and forward them on appropriately
  to Chromium's above-mentioned glue classes.
- Disabling this system integration in the context of the existing
  cross-platform autofill browsertest, which installs its own stub
  AutofillProvider implementation.

One special complexity here is that unlike in Android WebView, within
WebLayer the Context in which TabImpl is operating can change (due to
the BrowserFragment being retained across an Activity recreation). To
handle this, we tear down the AutofillProviderImpl instance whenever
the Context with which the TabImpl was previously associated goes away
and recreate the AutofillProviderImpl instance whenever the TabImpl is
associated with a new Context. We also extend the native side to be
notified of such teardowns and recreations.

I am investigating testing strategies for this integration as followup.
I currently see two possibilities:

1. Inject a mock wrapper of the system-level autofill infrastructure
   into Chromium's glue classes. This is the approach that Android
   WebView follows. It will require the ability to interact with the
   WebLayer implementation from instrumentation tests, which is not
   yet supported. It has the downside that it will not actually test
   integration with the system-level autofill infrastructure.
2. Write a sample AutofillService implementation and set up tests
   that interact with that implementation. This has the advantage of
   coverage of the actual integration with system-level autofill.
   However, it will require the ability to simulate the user's
   interacting with dialogs presented by the Android system (e.g., to
   set the sample AutofillService as the enabled AutofillService). At
   this time, I do not yet know what support is present in Chromium's
   testing infrastructure for such simulation.

Bug: 1025617
Change-Id: If3c6f9d085e72e62d8879308c108a235a6f035b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980739
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Tobias Sargeant <tobiasjs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731432}