Reformat all gn files in /third_party/blink/renderer/modules/launch.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=raymes@chromium.org

Bug: 1041419
Change-Id: Ic2e6a55467967191b2dffc7250ca3f970387ff9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997944
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731433}