Reformat all gn files in /tools/traffic_annotation/summary.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=msramek@chromium.org

Bug: 1041419
Change-Id: I68c66021c4153517f3f90c5d9be65d390f3654ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998382
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731434}