Reformat all gn files in /chrome/browser/ui/webui/new_tab_page.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=tiborg@chromium.org

Bug: 1041419
Change-Id: I61714cdbb993451260314a611a180293f0ab0b5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998648
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731438}