Reformat all gn files in /chromecast/browser/android/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sanfin@chromium.org

Bug: 1041419
Change-Id: I5a9ece1c068c86532e79f0c398aec308d8e207eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998047
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Simeon Anfinrud <sanfin@chromium.org>
Commit-Queue: Simeon Anfinrud <sanfin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731457}