Make webapps open new tabs in same activity part 1/2
This CL is in preparation for making window.open() open a new
tab in the current activity's tab strip for WebappActivity the way that
CustomTabActivity does today.

This CL adds webapp-only behaviour to:
- TrustedWebActivityBrowserControlsVisibilityManager to show
the browser controls and the close button for tabs opened via
window.open()
- CloseButtonNavigator to make the close button close the tab
if it was opened via window.open() (vs navigating to the most recent
page which is within the WebAPKs scope)

This CL also adds new metrics:
- CustomTabs.CloseButton.ChildTab.ScopeAlgorithm.ClosesOneTab to determine
whether the new close button behaviour differs to the current close
button behaviour for real websites
- WebApk.Navigation.ChildTab.InScope to determine whether window.open()
is used for URLS which are within the WebAPK's scope.

BUG=1038841
TEST=CloseButtonNavigatorTest

Change-Id: I93d7c6e931fb372c470598fce85440175292b5bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986450
Commit-Queue: Peter Kotwicz <pkotwicz@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Glenn Hartmann <hartmanng@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731459}