Reformat all gn files in /components/search_provider_logos.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=justincohen@chromium.org

Bug: 1041419
Change-Id: I1289ae53fb4b753f94447bd98a16336a40e09f00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998782
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731463}