Reformat all gn files in /third_party/wayland.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=oshima@chromium.org

Bug: 1041419
Change-Id: Ia365977ceb2e7bd5f37a4fba1b075eef770edad9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998170
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731467}