Reformat all gn files in /chrome/chrome_elf.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=caitkp@chromium.org

Bug: 1041419
Change-Id: I77578d0ff4b752b7f76edd5a06b739f185a1ba5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998066
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Cait Phillips <caitkp@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731469}