Reformat all gn files in /third_party/blink/renderer/core/probe.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=caseq@chromium.org

Bug: 1041419
Change-Id: I416c0082fddc811674238de2cad6c20d5ae16503
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998734
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731478}