Reformat all gn files in /third_party/tcmalloc.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=primiano@chromium.org

Bug: 1041419
Change-Id: I10f46812f58dbf7c2150cc8a2e76592c55473b3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998598
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Primiano Tucci <primiano@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731484}