Reformat all gn files in /components/grpc_support.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=zhongyi@chromium.org

Bug: 1041419
Change-Id: Iaddb687988f3779b43c816b83f224404d7e2af6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998280
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Zhongyi Shi <zhongyi@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731487}