Reformat all gn files in /components/password_manager.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=kolos@chromium.org

Bug: 1041419
Change-Id: I3fba1be81acec95c7142feaac7e6d62faf5d23a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995858
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Maxim Kolosovskiy <kolos@chromium.org>
Commit-Queue: Maxim Kolosovskiy <kolos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731491}