Reformat all gn files in /components/leveldb_proto.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=nyquist@chromium.org

Bug: 1041419
Change-Id: Iea95170a02dabf08739154fc9f34eec0e0e84623
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998172
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Commit-Queue: Tommy Nyquist <nyquist@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731500}