Reformat all gn files in /components/services/heap_profiling.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=erikchen@chromium.org

Bug: 1041419
Change-Id: Ifa8bf4b865762a9deee417fe3436f7970690667f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998741
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731501}