Reformat all gn files in /components/contextual_search.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=twellington@chromium.org

Bug: 1041419
Change-Id: I878af8db20c7cdf36f5753076ebd835d81081048
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998820
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Donn Denman <donnd@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731503}