Reformat all gn files in /chrome/test/base/js2gtest.gni.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=stevenjb@chromium.org

Bug: 1041419
Change-Id: I61e5049051ec4863dbf8aeb24fcbb92b45293d74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998277
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731504}