Reformat all gn files in /third_party/ced.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jinsukkim@chromium.org

Bug: 1041419
Change-Id: I8f20a0487ace57a103b9583bf9fe65bb59852205
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997756
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731512}