[iOS] Fix first combined character input.
This is a fix for broken input of the first combined character in the
omnibox (for example in Korean).
This introduces a regression on older OS versions when using third party
keyboards; however this regression is *already* happening on the latest
iOS 13 builds.
This is a temporary fix; a proper fix is to change pre-edit behaviour
so that it wouldn't set text or change selection in the omnibox when it
exits pre-edit. Such a fix is a lot more fragile and will be rolled out
behind a flag, but we need to fix the common-case input of combined
character on latest OS now.

(cherry picked from commit 1ed06c8787e7e957c713643a1a014eb6aa3c2cb0)

Bug: 1030476, 988431, 875918, 873544
Change-Id: I33248fdb17d75e215dedea11469de94dd7dc183c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980746
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#728505}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2000618
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/branch-heads/3987@{#524}
Cr-Branched-From: c4e8da9871cc266be74481e212f3a5252972509d-refs/heads/master@{#722274}