Reformat all gn files in /ios/chrome/browser/download.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=eugenebut@chromium.org

Bug: 1041419
Change-Id: I4a7c9d8ee66a4e80184a27b2928e33be35deb10d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995794
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731515}