Reformat all gn files in /testing/perf.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=bsheedy@chromium.org

Bug: 1041419
Change-Id: I9be007cf211e0ed0c8a3d7e005419503d0a3ba8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998569
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731520}