Reformat all gn files in /components/subresource_filter.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jkarlin@chromium.org

Bug: 1041419
Change-Id: If545efc3cf1911956d9bd413860d228ff96719a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998559
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731522}