Move Mojo binding endpoint headers to be owned by SECURITY_OWNERS
Most of the time, the only changes required are in the .cc. On some
occasions though, the header also needs to be updated. It doesn't
really make sense to get two reviewers for the case where the header
needs to be updated too, so just consolidate ownership.

Bug: 1012033
Change-Id: If8179bda928782f182ac086f73ca75415a2641a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1993878
Auto-Submit: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731526}