Reformat all gn files in /third_party/blink/renderer/modules/bluetooth.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=dougt@chromium.org

Bug: 1041419
Change-Id: I5e14d61cfcd41ca650b14bda2c9ddf9cc356ea61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998558
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731528}