Reformat all gn files in /chrome/browser/media/router.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=takumif@chromium.org

Bug: 1041419
Change-Id: Ife4a250bf109ae1137c17113f633be6ae0680ebb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997813
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731529}