Reformat all gn files in /chrome/browser/resources/chromeos/login.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=alemate@chromium.org

Bug: 1041419
Change-Id: Idb09397d39d80887de5a4f6ded65b8919b9a9655
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998065
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Achuith Bhandarkar <achuith@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731537}