sessions: make SessionBackend be deleted on background task runner
SessionBackend is used exclusively on a background task runner (it's
created on the main thread). SessionBackend's destructor does IO.
No provisions were made to ensure SessionBackend was deleted on
the background task runner, which would result in running IO on the
main thread.

This patch ensures SessionBackend is deleted on the background task
runner.

BUG=112512
TEST=none

Change-Id: I47f9c95195dfffdede5e87196c320253de2acef0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2000959
Reviewed-by: François Doray <fdoray@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731542}