Reformat all gn files in /ios/chrome/browser/ui/collection_view.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=rohitrao@chromium.org

Bug: 1041419
Change-Id: I238eb712b932b7e2d408fd8e9b43ebef8af8cd99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998570
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731559}