Reformat all gn files in /ios/chrome/browser/ui/omnibox/popup.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=stkhapugin@chromium.org

Bug: 1041419
Change-Id: If551207bc0b945f223cddd2a6c4ed08fe54b39e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997710
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731566}