Reformat all gn files in /components/search_engines.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=manukh@chromium.org

Bug: 1041419
Change-Id: I77d816e1495a2d404919fb2d0c83162df088af0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998306
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731595}