Reformat all gn files in /third_party/blink/renderer/modules/media_capabilities.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=mlamouri@chromium.org

Bug: 1041419
Change-Id: I37d869acfd366097bf4c1b2c2fc84d49f82c3bfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998349
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Commit-Queue: Mounir Lamouri <mlamouri@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731599}