Reformat all gn files in /ios/chrome/browser/ui/toolbar_container.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=kkhorimoto@chromium.org

Bug: 1041419
Change-Id: I319d7a1fbb3caa71709cce240011a758a598a38f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998798
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731601}