Reformat all gn files in /ash/components/shortcut_viewer.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=wutao@chromium.org

Bug: 1041419
Change-Id: I0ae4ba3bd6daae16e7b81963afa8d5d1726a0d26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998738
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Tao Wu <wutao@chromium.org>
Commit-Queue: Tao Wu <wutao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731602}