Use panel's rotation to rotate cursor.
On ChromeOS, the panel's physical rotation can be different from
logical one. That is, the panel's original orientation can be
portrait, but the default orientation is landscape, which leads
to wrong cursor rotation. This CL fixes the issue by using panel's
orientation to rotate cursor.

For other platforms, they're the same and this doesn't change the
behavior.

Bug: 1027832
Test: covered by unit tests. Also teseted manually. See bug for repro case.
Change-Id: If70076a3a8a1b9b0893a31827bc767f2dd512489
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995060
Reviewed-by: Bret Sepulveda <bsep@chromium.org>
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731603}