Reformat all gn files in /services/proxy_resolver.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=rmcelrath@chromium.org

Bug: 1041419
Change-Id: I3ed2eccb81e26ef4a1b51fe0eb9e0ec584c8012b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998881
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Robbie McElrath <rmcelrath@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731604}