Reformat all gn files in /third_party/blink/renderer/platform/scheduler.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=shaseley@chromium.org

Bug: 1041419
Change-Id: I73d9da6db7cf7347a1f22c13958c5f3723a6c8f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998164
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731621}