Reformat all gn files in /chrome/browser/touch_to_fill/android.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=fhorschig@chromium.org

Bug: 1041419
Change-Id: Icec751dca4a2e208e7b20ddc334d64db92b3958e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998357
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731631}