Reformat all gn files in /ios/chrome/browser/history.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=sczs@chromium.org

Bug: 1041419
Change-Id: I47e645f0f9da05e7ad5daacb64c565fab498585c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998680
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731637}