Hide tab strip when restore completed and overview visible
Currently, we try to show tab strip when tab model restore completed.
However, when return-to-clank is enabled, we are not in tab page when
restore completed, and therefore the strip should not show. This CL
fixes this issue and adds a test.

Bug: 1030606
Change-Id: I01df83b15682b353eb4e5eabada97d457384969e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2000965
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Commit-Queue: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Auto-Submit: Yue Zhang <yuezhanggg@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731642}