fido: filter "no user activity" responses from authenticatorBioEnrollment
As a response to a authenticatorBioEnrollment CTAP command, an authenticator
may send a response with a status of CTAP2_ENROLL_FEEDBACK_NO_USER_ACTIVITY if
it timed out waiting for the user to touch a fingerprint sensor. The
BioEnrollmentHandler reports this to the UI layer by invoking its
SampleCallback and then immediately sends another authenticatorBioEnrollment
command to request the next sample.

Change BioEnrollmentHandler to not invoke the SampleCallback for the timeout
responses so that the UI layer doesn't have to ignore them explicitly. This
fixes a bug where receipt of a timeout response would make the enrollment
progress bar flash.

Change-Id: I3247169f7a21fcb61b0ed9ba727810f7bf6125f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1999087
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#731645}