Add GetMaxInputFramesRequested()
The AudioTrackOpusEncoder runs into an CHECK when |converter_| calls
ProvideInput() twice in a row, and the ATOE underflows. This happens
because the amount of frames converted at a time is bigger than
|resampler_.ChunkSize()|, which means multiple reads can be issued
for a single call to resample.

This issue is similar to the one in recently fixed in
AudioRendererAlgorithm. This CL generalizes that fix by adding
SincResampler::GetMaxInputFramesRequested(), and fixes ATOE by making
sure it has enough data to cover all reads.

In practice, this means that ATOE needs a slightly bigger FIFO, and
carries one more buffer's worth of frames in memory.

Bug: 1025246
Change-Id: I0caae8f3725dbcce92e076ad246f2870db82c84d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1992358
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731651}