Reformat all gn files in /third_party/blink/renderer/modules.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=mlamouri@chromium.org

Bug: 1041419
Change-Id: I7cf01fd713eecdf0509185d87258fbb1d8a04a82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998296
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Commit-Queue: Mounir Lamouri <mlamouri@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731658}