Reformat all gn files in /chromeos/login.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=achuith@chromium.org

Bug: 1041419
Change-Id: I5ab11cc97e64ccf3b92cdca4dc7950d8509c0ef7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998704
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Achuith Bhandarkar <achuith@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731660}