Reformat all gn files in /chrome/android/feed/core/java/src/org/chromium/chrome/browser/feed/library/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=pasko@chromium.org

Bug: 1041419
Change-Id: Ib00515a92a2d56c559e23ae6d806cc38441eb380
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994078
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Egor Pasko <pasko@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731681}