Reformat all gn files in /ui/gl.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=ericrk@chromium.org

Bug: 1041419
Change-Id: I38a8a73da2d421506d7cbeab034ed83850d602f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998513
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Commit-Queue: Eric Karl <ericrk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731683}