Reformat all gn files in /ios/chrome/test/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=rnk@chromium.org

Bug: 1041419
Change-Id: Icc793b1f54f98146eb7d635662d0ae184424f5f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997701
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Reid Kleckner <rnk@chromium.org>
Commit-Queue: Reid Kleckner <rnk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731696}