Reformat all gn files in /ui/events/blink.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=nzolghadr@chromium.org

Bug: 1041419
Change-Id: I86adbcbd780e75ef47856090bca1e42945691977
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995847
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Commit-Queue: Navid Zolghadr <nzolghadr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731701}