Reformat all gn files in /chromeos/timezone.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=alemate@chromium.org

Bug: 1041419
Change-Id: I7317b4bfdc0f3a7092ab59f1e8cc5a7380bf402c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1993849
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Commit-Queue: Alexander Alekseev <alemate@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731706}