Reformat all gn files in /chromeos.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=alemate@chromium.org

Bug: 1041419
Change-Id: I5723ed7e974e973d1e2d66c9ebf60243507e1111
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998797
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Commit-Queue: Alexander Alekseev <alemate@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731710}