Reformat all gn files in /chrome/android/chrome_common_shared_library.gni.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=pasko@chromium.org

Bug: 1041419
Change-Id: Ib7bfeb431e9f0522d2b9b1869bfcabfaf21c34fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998388
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Egor Pasko <pasko@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731711}