[android] Update minimum supported OS version to Lollipop
Chrome currently supports Android KitKat, but as of this CL, a UI will be
presented to users running on Android KitKat. Whether this message is shown
or not is possible override from a server side configuration.

This is in preparation of an upcoming deprecation of Android KitKat support.

Bug: 1041930
Change-Id: I6eef0697887215cc58a1d4e5067ba19ee383c76a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2001241
Commit-Queue: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731745}