Reformat all gn files in /ash/system/machine_learning.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=wrong@chromium.org

Bug: 1041419
Change-Id: I6759f2f2d88ecab9c71e220d6be9093720ba4a1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994079
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Jia Meng <jiameng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731776}