Allow chrome.loginState API for Smart Card Connector app
The chrome.loginState API was only exposed to Chrome Extensions.

This change exposes this API to the Smart Card Connector, which
is a Chrome App. The Connector app needs this API in order to
fix an issue related to user session states (details on the bug).

Bug: 1034927
Change-Id: I17a3d5fb7446ae65b9ef500011615a20b841c4f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970311
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Ben Wells <benwells@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731786}