Reformat all gn files in /components/mirroring.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=miu@chromium.org

Bug: 1041419
Change-Id: Ibdcde4d07f7c038ab1559bdafd86a09d6d3a8574
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995797
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Yuri Wiitala <miu@chromium.org>
Commit-Queue: Yuri Wiitala <miu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731826}