Reformat all gn files in /chrome/services/wilco_dtc_supportd.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=pmarko@chromium.org

Bug: 1041419
Change-Id: I1be9952d103e6f1fbecb58fb15048942dd55c710
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997554
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731852}