CrOS Shelf: Add ink drop padding to first and last app in the hotseat.
This change adds padding to the ink drop for the first and last app
buttons in the scrollable shelf. This padding is added so that when the
ink drop is shown for the left-most or right-most app, the ink drop
appears to cover from the behind the app to the left/right edges of the
hotseat. This means that a "one app hotseat" will be completely filled
by the app's inkdrop.

Bug: 1035263
Change-Id: I91383bf73865f6807d878bde3afeb904ddacb3f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1999767
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731865}