Fix Crash in RenderWidgetHostImpl::GetRootWidgetViewportSize
In current case, if the view is destroyed and fling controller is trying
to generate fling curve, it crashes when it calls
RenderWidgetHostImpl::GetRootWidgetViewportSize to get the size of the
root view.

This CL fixes the issue by performing a null check for the view. If
destroyed, it will not generate fling curve.

Bug: 1030354
Change-Id: I81f7e7170691cbe9c8cd274d71a4825e046b78a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1972609
Commit-Queue: Sarthak Shah <sarsha@microsoft.com>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#731870}