Reformat all gn files in /third_party/blink/renderer/modules/webmidi.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=toyoshim@chromium.org

Bug: 1041419
Change-Id: I8172a49c60fb4fe4787d8fb96a611d8f410e6280
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997959
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731882}