Reformat all gn files in /content/common/fetch.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=hiroshige@chromium.org

Bug: 1041419
Change-Id: I7b00ea2ae90e089a1c6792752cff6456c1ffe5a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998640
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Commit-Queue: Hiroshige Hayashizaki <hiroshige@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731899}