Reformat all gn files in /content/shell.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=mkwst@chromium.org

Bug: 1041419
Change-Id: I7de48deb8977863952c15ac6659dcf1b700efff8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998175
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731901}