Fully sync tabs after batch operations
During a batch operation (which are currently restoration, close all
and undo close all), the internal state of the WebStateList may
not be consistent (in particular, there may not be any active WebState).
Calling sync in the middle of one may cause loss of sync data.

After a batch operation, it is necessary to resync all the tabs as they
may all have changed.

Note: on close all tabs, the local state of sync will be updated but
this will not be reflected on the server as empty windows are not synced.


Bug: 1038926
Change-Id: I7c8ff2d8d335ea20d95b8c3d6a74d04d64ee06f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986084
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731917}