Reformat all gn files in /components/keep_alive_registry.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=benwells@chromium.org

Bug: 1041419
Change-Id: Idd7f6b91eb08ad3f5b3a44323eb03d14443d4ecf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998067
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Ben Wells <benwells@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731929}