Reformat all gn files in /components/crx_file.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=benwells@chromium.org

Bug: 1041419
Change-Id: I8bd35171cb1e83eb86d245fbd7c70c38547fc593
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994513
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Ben Wells <benwells@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731939}