Fix crash due to leaking sync tab pointers after free
SyncedSessionTracker keeps raw pointers to tabs in |synced_tab_map|,
with the ownership either in |unmapped_tabs| or for mapped tabs in
the actual session window (SessionWindow::tabs).

Prior to this patch, the clearing of tabs in
PopulateSyncedSessionWindowFromSpecifics() could lead to dangling
pointers in |unmapped_tabs|, leading to undefined behavior. This
specifically affected the case where the input contains duplicate
window IDs, since any other window should have no tabs as a result
of ResetSessionTracking().

The presence of dangling pointers in |synced_tab_map| led to a CHECK
failure in PutTabInWindow() because of how
SyncedSessionTracker::GetTab() is implemented, which would return
the dangling pointer to freed memory, not linked to any known
window, and detected by a CHECK failure in PutTabInWindow() before the
pointer itself was dereferenced.

After this patch, if a header entity contains duplicate window IDs, it
gets filtered out early. Based on that, the problematic tabs.clear()
call in PopulateSyncedSessionWindowFromSpecifics() is replaced with a
CHECK.

Bug: 803205
Change-Id: Id2f3fd4f0cc5012ed169065b39c5363ee9c23c14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2001303
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731955}