Reformat all gn files in /third_party/blink/renderer/modules/exported.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=kinuko@chromium.org

Bug: 1041419
Change-Id: Iedaf875c56ccf34c6538c46a6a23d238d8dcc944
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998273
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731987}