Reformat all gn files in /components/payments/content/android.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=gogerald@chromium.org

Bug: 1041419
Change-Id: I1db9ad258f3025c25d8c491a665baa616e4eae2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997744
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Ganggui Tang <gogerald@chromium.org>
Commit-Queue: Ganggui Tang <gogerald@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731989}