[iOS] Create BrowserList
A keyed service which can track browsers created or destroyed for a
browser state.

This implementation is intended to be a tool that can be used by
objects that own Browsers (currently in Chrome, that's BrowserViewWrangler).
Browsers themselves are not automatically tracked by the BrowserList;
the code that creates and destroys them must add and remove them.

Change-Id: Ia67c730b7ff0fa6668cd23f7955b78071d21444a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995326
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Mohammad Refaat <mrefaat@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731990}