[ios] Fix tabgrid buttons when using VoiceOver
Previously, the "Close All" and "Done" buttons were not
being updated when using VoiceOver and switching panels
in the tab grid. Additionally, the incognito tabs were
not being hidden in the app switcher (when using VoiceOver).

The problem was that the buttons are being configured in
the scrollview animation completion code. VoiceOver turns
off animation, so the buttons are not being properly
configured when using VoiceOver. Same goes for the incognito
hiding code.

This CL fixes this issue by calling the button configuration
and incognito content visibility update methods from the
VoiceOver codepath.


Bug: 1038010
Change-Id: If76127096638f73283b54a0b1be952b0479d3b3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2001838
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Auto-Submit: edchin <edchin@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731992}