Reformat all gn files in /ios/showcase.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=edchin@chromium.org

Bug: 1041419
Change-Id: Ib28835e0fe99bfe6c4d6226ba60c4ab26f5707a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998748
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732009}