Fix for dropped events on composited scrollbars.
This fixes an issue where mouse event handlers for mousedown/up
were not getting fired when interacting with the scrollbar on the
compositor thread. The fix is to add a modifier to the MouseEvent(s)
and forward it to the main thread and on the main thread side, if
that modifier is found, avoid passing the event to the scrollbar.

Bug: 1038549
Change-Id: Iedebdd98eb3d339bc594d319e6db02b6b6fddde6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986308
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Rahul Arakeri <arakeri@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#732039}