Fix font fallback unittests for Mac
This is the same issue than this CL, but for MAC:
 https://chromium-review.googlesource.com/c/chromium/src/+/1824104

The MAC API is returning ".Noto Sans Universal" which is not
recognized by Skia (i.e. SkTypeface). To work with the API
that converts codepoint to glyphs, the same NativeFont must be kept
and pass to the SkTypeface_Mac constructor. The goal is to avoid
loosing information and keep the exact selected fallback font.

Bug: 1022455
Change-Id: If15d7a1306131fe2469406b45fb614d3d3bcdb39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994158
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732042}