Add a document Type enum to DocumentInit
Currently, we wait to determine what type of Document will be
created until it's time to actually construct the Document. This
CL moves that determination earlier. The Document type is needed
in policy decisions related to ExecutionContext and SecurityContext
initialization. If, as planned, the LocalDOMWindow because the
ExecutionContext, these policy decisions will happen before Document
constructions and will need access to that type.

Bug: 1029822
Change-Id: Ic99ae1d74ef955d3e3abf7df46aa67ce7559a227
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1989065
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732046}