Reformat all gn files in /third_party/re2.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=tfarina@chromium.org

Bug: 1041419
Change-Id: I1bb86a7010450186cd178218f65ed1d5291fcb85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998866
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732049}