desktop-pwas: Omit .exe in Win7 launcher filenames
On Windows, each Progressive Web App (PWA) must have its own launcher
in order for the PWA to be able to register as a handler for file
types. On Windows 7, the filename of this launcher is used as the
PWA's display name in the "Open With..." menu, because Windows 7 lacks
an alternate way to set multiple custom display names for one
executable (and PWA launchers are all identical copies of
chrome_pwa_launcher.exe).

This change omits .exe from the filename of PWA launchers on Windows
7, unless the PWA's name (which will become the launcher filename) is
a reserved filename on Windows.

Bug: 960245
Change-Id: I9713c3151c276bf460b193ce502ecdf233d3aad4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1965560
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Reviewed-by: Jay Harris <harrisjay@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732057}