Reformat all gn files in /ipc.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=tsepez@chromium.org

Bug: 1041419
Change-Id: I0b313bdca77daa75e41535abfe706c7e6652aaa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995853
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732084}