[startup] Do not crash if a visible window has no active WebContents on startup.
It is invalid to have a visible browser window with no active WebContents,
but https://crbug.com/1032348 shows that it happens. This is a fix that
can be merged to M80 to avoid getting crashes on stable.

Instrumentation to better understand the problem is added in
https://chromium-review.googlesource.com/c/chromium/src/+/2001220

Bug: 1032348
Change-Id: Ia20a35fff65532d8f11110a097b94b72fbcf3dd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2001139
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Auto-Submit: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732092}