Reformat all gn files in /ui/compositor.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=flackr@chromium.org

Bug: 1041419
Change-Id: I7ff5060ebb5b369c81513b25dd9da6bb148a6c00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994076
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732105}