Allow popup sliding when the anchor is already occluded
Minor fix to popup sliding. We have a heuristic to prevent sliding the
popup over the anchor rectangle when that would occlude it. There was a
bug where the heuristic would prevent sliding even if the anchor rect
was already occluded, which is unnecessary.

This bug was discovered in the course of fixing crbug.com/992856 and may
contribute to some poor popup placements, just not the ones in that bug.

Bug: 992856
Test: exo_unittests --gtest_filter="*Positioner*Sliding*"
Change-Id: I975e6a15e299b8acf8d096f2b1317e9d4e66a57b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997430
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Cr-Commit-Position: refs/heads/master@{#732148}