[PM] Fix some flakiness in the OT browser tests
These tests are racy, they verify that the OT policy for a page is the
expected one but this can fail if PM hasn't received the OT policy for
all the frames by the time we do this check. This CL simply adds a
counter to ensure that we've received all the messages we need before
verifying the expectations.

Bug: 998071
Change-Id: I7189c42f7ea3e64dfd95b284f5d2ec5789dc2150
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2002938
Reviewed-by: François Doray <fdoray@chromium.org>
Commit-Queue: Sébastien Marchand <sebmarchand@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732150}