Add test automation for manual native file system tests.
This adds two new methods to testRunner:
 - getWritableDirectory, which returns the path to an an
   (empty, per test) temporary directory. And
 - setFilePathForMockFileDialog, which causes all file dialogs that
   are shown by the test to return the provided file path.

This is then used to automate the otherwise manual native file system
web platform tests that try to interact with the actual native native
file system implementation (as opposed to the sandboxed file system).

Also rename the filenames of these manual tests so the test system
actually correctly picks them up. Unfortunately not all of the tests
are currently passing, so this also checks in test expectations for
the failing tests.

Bug: 998917
Change-Id: I0c9100d39387bb211c3108c7566937af0e75db68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1989369
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Robert Ma <robertma@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732156}