Have CastAppController observe WebContents
We need to observe this since this is how the cast apps size is updated
from the renderer. For a slow loading page RenderHostFrame may not exist
when the cast app is being shown so the EXO surface won't be sized when
the WebContents are attached to the EXO surface.
Since it wasn't observing we would never get a resize update.

Bug: b/147376354
Test: None
Change-Id: Idddf3cdadbf8a58b7b2506f3af8c4e5f16c92e6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2003609
Reviewed-by: Albert Chaulk <achaulk@chromium.org>
Commit-Queue: Daniel Nicoara <dnicoara@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732163}