Reformat all gn files in /media.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jrummell@chromium.org

Bug: 1041419
Change-Id: I476454a428344042303d09daaa009e756f2dc4eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995793
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732174}