Reformat all gn files in /components/data_reduction_proxy.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=ryansturm@chromium.org

Bug: 1041419
Change-Id: I5c64a562b5a91207b51204fe58d679b1106d1425
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994071
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Commit-Queue: Ryan Sturm <ryansturm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732181}