Update CompositeCrossOriginIframes feature to promote crossorigin frames
This patch updates the CompositeCrossOriginIframes feature to promote
(composite) cross-origin iframes. The original patch for this feature
(https://crrev.com/708336) had a subtle bug because it did not change
the self-painting status of the iframe's layer. To be composited,
PaintLayers must be marked as self-painting (see:
PaintLayerCompositor::CanBeComposited).

This patch builds on the existing iframe compositing reason by adding a
check for cross-origin frames in
LayoutEmbeddedContent::RequiresAcceleratedCompositing. Because
cross-origin changes now affect compositing, a call to
Element::SetNeedsCompositingUpdate has been added when the cross-origin
status changes in DocumentLoader::InstallNewDocument and
LocalFrameView::CrossOriginStatusChanged.

Bug: 1014273
Change-Id: I096fbad4921345df836b2a61c338cb4f4b5f79d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997639
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732194}