[throughput] Notify trackers when commit happens.
Notify the trackers when a begin-main-frame processing is complete, so
that the tracker can verify (with DCHECKs) that it is notified of the
events in a well-defined and expected sequence. At the moment, there
are synthesized 'begin main frame' notifications to the trackers that
prohibit the DCHECK()s to be added, but we will be turning those on in
follow up CLs.

BUG=1021963

Change-Id: I01e8bb2d869a1707b6a1b412fdab0cadfde12255
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995461
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732238}