Reformat all gn files in /android_webview/tools/automated_ui_tests/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=ntfschr@chromium.org

Bug: 1041419
Change-Id: I24f1a8ce0a32854fad3be2f645dd801f20e583ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998381
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732254}