Reformat all gn files in /content/browser.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=boliu@chromium.org

Bug: 1041419
Change-Id: Ied9d082ee57276d5b82983e6b42ee7914dea9eff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998672
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732255}