Add a Feature, kEnablePalmOnMaxTouchMajor to change max_major behavior.
By default, we mark any touch that has a ABS_MT_TOUCH_MAJOR ==
major_max (as given in device info) to be a palm.

For some experimental situations, we often receive a genuine
TOUCH_MAJOR of this magnitude. We want to be able to treat that as normal.

Adds a Feature, and set it at construction time - this avoids looking
through features on every single touch. Adds a unit test appropriately
to test for this change in behavior (and that current behavior is
maintained).

Bug: 1009290
Change-Id: I4f1c3b2d4733ff00cb13814db4dad834baeb8977
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2003166
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Commit-Queue: Rob Schonberger <robsc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732267}