Reformat all gn files in /chrome/browser/resources/management.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=emaxx@chromium.org

Bug: 1041419
Change-Id: I1bc4cdc82b44fbc1da1cedc429708e0aa58717ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998264
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732268}