infra: Set `enable_ats` value in builders.star
We want to enable ATS for all builders using an RBE backend.

This will be done programmatically in builders.star, so that we don't
have to update builder configs individually.

In this CL, we test out this feature on Windows builders currently using
RBE.

This should be a no-op for generated builder configs. The only changes
in cr-buildbucket.cfg are to these builders' configs:
- win-celab-tester-rel: test-only, unaffected by goma flags
- win-pixel-builder-rel: did not explicitly specify because it is
  assigned a Windows 10 bot elsewhere, added explicit Windows 10 value.

R=agable@chromium.org, gbeaty@chromium.org, tikuta@chromium.org

Bug: 1040754
Change-Id: If7e97e1a444527b5e72413988f02108b06870ac6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2004186
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Simon Que <sque@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732291}