Android: Hook Preview Tab to IPH tracker
Preview Tab is using PreferenceManager to control the visibility of
"new" label on the corresponding context menu items. This CL migrates
it to use the feature engagement system so that it is fully controlled
as IPH.

The context menu has 2 menu items for preview tab feature (open link
and image) but their 'new' labels are handled as one. i.e. they disappear
at the same time if user accesses preview tab via either of the menu items.

The label handling for Google Lens is reverted. It needs to be
taken care of by a separate feature engagement entry. The label remains
visible for now.

Bug: 1027718, 902140
Change-Id: I5d68d4f5e8c71f66c2c3c91a875152d1c4b128dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1985838
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Donn Denman <donnd@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Sinan Sahin <sinansahin@google.com>
Cr-Commit-Position: refs/heads/master@{#732306}