PerUserTopicSubscriptionManager: retry only once on auth errors
On receiving an auth error (HTTP 401), PerUserTopicSubscriptionManager
invalidates the access token and retries, based on the assumption that
the access token was likely expired. That makes sense once, but if the
second request fails with an auth error again, there's really no point
in trying again. In some cases, it might even lead to an infinite loop
of invalidating and re-requesting access tokens.
This CL changes the behavior so that a request is retried exactly once
on an auth error. If it fails again, it is abandoned (which already
happens for all other 4xx errors).

Bug: 1020117
Change-Id: I0c83d3eaa5bafcc8d9f81315921f8375661371af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2000625
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732399}