[Android] Fix password generation integration tests
The manual generation tests were broken because the tab view in the keyboard
accessory could not be found. This CL changes the matcher used.

All tests used to depend on server predictions for filling, which marked
the field as eligible for generation or not. This broke when the predictions
where not as expected.

Bug: 1010540, 1010344

Change-Id: I8ca5392d80b81a67d414eec33cc7a7a2e282d725
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995273
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732423}