Reformat all gn files in /remoting/android.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=lambroslambrou@chromium.org

Bug: 1041419
Change-Id: Ifd8d08b1a26e308bf3d25dfa8cb7a4e8e5853a2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998199
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732433}