Reformat all gn files in /ios.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=pkl@chromium.org

Bug: 1041419
Change-Id: Ic52468b78c9fc7623519e87af64397452228d52b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997973
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732441}