Reformat all gn files in /ios/chrome/browser/ui/activity_services.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=pkl@chromium.org

Bug: 1041419
Change-Id: I4b2a38df9f4ec3c86b5cf57380f44a1bd271a8a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994439
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732450}