Reformat all gn files in /third_party/blink/renderer/modules/gamepad.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=mattreynolds@chromium.org

Bug: 1041419
Change-Id: Ib1ee8d1b1861e40d422f5836045216d36822f451
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997956
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732454}