Reformat all gn files in /components/safe_browsing.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=drubery@chromium.org

Bug: 1041419
Change-Id: Ic101be4bcd915eee3dfdd21fa4b81df1fd63aac7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997810
Commit-Queue: Nico Weber <thakis@chromium.org>
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732464}