Reformat all gn files in /chrome/common/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=msw@chromium.org

Bug: 1041419
Change-Id: Id5fd33ee7d25b36a3197b763f8c08d3f5cdd24c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998032
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732465}