Reformat all gn files in /third_party/blink/renderer/core/xml.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=palmer@chromium.org

Bug: 1041419
Change-Id: I6a009e9f579bd7536f7ac24518c506823d792a5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997993
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Chris Palmer <palmer@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732467}