Reformat all gn files in /components/browsing_data.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=jsbell@chromium.org

Bug: 1041419
Change-Id: I9bd76424e221848d518996431fd4e9fd2381725f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998037
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732468}