Reformat all gn files in /ios/chrome/browser/ui/util.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=marq@chromium.org

Bug: 1041419
Change-Id: Ia5c37b7458d97cc02b04c25a073c8027e37688f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997699
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732469}