Reformat all gn files in /sandbox/win.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=pennymac@chromium.org

Bug: 1041419
Change-Id: I1c20f1c82c2897118b67995697eae165a2f35d29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995846
Auto-Submit: Nico Weber <thakis@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732470}