Reformat all gn files in /third_party/blink/renderer/modules/peerconnection.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=tommi@chromium.org

Bug: 1041419
Change-Id: I8881810cdb7a748263c8cef903207d92f45c2d76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997902
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Tommi <tommi@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Tommi <tommi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732471}