Create content::MessagePort.
This is analogous to the Java org.chromium.content_public.MessagePort
and is the interface that embedders should use for talking to
Javascript. It hides away the fact that the implementation is
currently via Mojo, and allows the implementation to be changed
without changing embedder code.

A further CL will rework the Fuchsia code to use this instead of
raw mojo message pipes, before subsequently flipping over all
clients to use MessagePortDescriptors internally.

BUG=980533

Change-Id: I4a68ed51179c96bae4b12b7001cecb686302973d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1978944
Commit-Queue: Chris Hamilton <chrisha@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732475}