Reformat all gn files in /third_party/blink/renderer/modules/accessibility.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=aleventhal@chromium.org

Bug: 1041419
Change-Id: Iac2cd6da8c44de88fadf500b6cd3f34ba5016a19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1993856
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732479}