Speculative fix for crasher
The presumptive scenario is remote iframe inside a throttled iframe
inside an active top document. The top document runs a lifecycle
update, which succeeds (i.e., the return value from
LocalFrameView::UpdateAllLifecyclePhases is 'true'); but the middle
frame is throttled and doesn't update, so we can't do a hit test on
it. In that case, just give up on trying to compute occlusion
information.

BUG=1041252
R=chrishtr@chromium.org

Change-Id: If7fadc9daf033bceda9f8a01edc6c3cabf71c066
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2003259
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732489}