Avoid updating URL when Tab does not exist
The crash happens when a tab updates after model switched. The updated
tab does not exist in the current TabModel, therefore a null is returned
when we try to locate the selected tab. The simple fix is to null check
before we access the tab.

This CL also hides the onUrlUpdated logic behind the
TabGroupsContinuation flag.

Change-Id: I81090666328158f8a5a5e37459c54e48d0ef6325
Bug: 1035301
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1996130
Commit-Queue: Mei Liang <meiliang@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732493}