Reformat all gn files in /buildtools/third_party/libc++.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=glider@chromium.org

Bug: 1041419
Change-Id: I1eb878817e27ccfb5811a4aacd9b68d374fbe10d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997002
Auto-Submit: Nico Weber <thakis@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732497}