[breakpad] DCHECK that crash key size is below limit.
Currently, Breakpad will silently truncate a crash key name with a
length above the limit of 40 characters. Unfortunately, trying to
retrieve the value of the key with crash_reporter::GetCrashKeyValue()
will fail, as the truncated name is not used for retrieval.

This CL adds a DCHECK for the crash key length, to avoid having
other people chasing the same problem as me when writing a test
for crash keys.

Change-Id: I6f08ddcb5df9589b7604f34fbeb2dbb6573b1b20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1999220
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732520}