Introduce kDisplayChangeModal Flag
This change introduces the ash::kDisplayChangeModal feature and flag
to toggle the use of the display configuration change modal rather than
the notification. In addition, this change parametrizes the tests in
ResolutionNotificationControllerTest so that they are run with the flag
in both positions.

Change-Id: I9c52b618be412b8f3f41afe5bc3e256ac313932d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1979015
Auto-Submit: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Commit-Queue: Bailey Berro <baileyberro@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732525}