Fix issue 1038319: Unable to change chrome language via chrome://settings/languages on Windows-Hebrew machine.
The problem is that the chrome installer uses the old language code for Hebrew, "iw",
instead of the new code "he". This value is saved in the user's local state. The
language model in langauges.js does not recognize this language code, thus the bug.

Fixed: 1038319
Change-Id: Ib227b2972989c2fa87fdc2d4be408040ad137c55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1990476
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Fei Ling <feiling@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732550}