Move command line init to LayerTreeTest constructor
There is a data race in
LayerTreeHostScrollbarsPixelTest.HugeTransformScale caused by
TestGpuServiceHolder being initialized before the command line options
for Vulkan are set in LayerTreeTest::RunTest(). The test was not running
using Vulkan, and was passing as a false positive.

Set |renderer_type_| in the constructor of LayerTreeTest and move the
command line initialization there to make sure that these options are
set before any of the test body code is run. To clean up, remove
RendererType from the LayerTreePixelTest::RunTest() variants, and
replace some direct checks to |renderer_type_|.

Bug: 1040259
Change-Id: If47923b5ba21190733d584b0a709276c46710527
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2003171
Reviewed-by: kylechar <kylechar@chromium.org>
Commit-Queue: Sean Gilhuly <sgilhuly@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732587}