cc: Skip crash dumps for PaintCache failures on SkTextBlobs.
This seems to be a majority of the failures for invalid deserialization
errors with PaintOpBuffers and its unclear whether this is just fallout
from failure to deserialize an inlined SkTextBlob or a caching error.

Skip performing a crash dump for the cache error cases to make this
clearer.

R=ericrk@chromium.org

Bug: 1019634
Change-Id: Ia882b918dc0630a89c7507fb0154f6fb08ee0087
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2001584
Auto-Submit: Khushal <khushalsagar@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Commit-Queue: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732588}