Make the BinaryUploadService treat APP requests as authorized
Currently the BinaryUploadService tracks authorization by checking the DM
token with the server. For Advanced Protection users, we will not have a
DM token, so we need to authorize the upload another way. This CL adds a check
for whether the user is truly enrolled in APP as the authorization.

Bug: 1020418
Change-Id: Ie9d61906aeed1fbb68fb68d455a5c24c9d5d31fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2001107
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Bettina Dea <bdea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732601}