task_manager: Include portal renderer tasks in the task manager UI.
We render them mostly like subframes: with a distinctive prefix, showing the
site URL rather than the title and favicon (which aren't otherwise visible to
the user and so could be misleading). Also like subframes, we mark the host
page as a parent task, so that the parent sorts first in the task manager UI.
Double-clicking on a portal task will activate the tab containing it.

These tags can need to change when portal activation swaps a tab and a portal
within it, so an extra hook after this has been done is added to allow the
delegate in chrome/ to do so.

Finally, tasks are given the opportunity to hold onto the WebContentsTaskManager,
which is useful so that portal tasks can look up their parent. This is somewhat
more dynamic than subframes (whose parent never changes), so looking it up
from the current WebContents relationships seems cleaner than storing a
task_manager::Task* like SubframeTask does.

Browser tests that these update correctly across portal activation, and that
they sort like subframes do are included.

Bug: 932303
Change-Id: I706eb6f8797f395d723b0d4cf0a683cb5665a172
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998999
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Adithya Srinivasan <adithyas@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732614}