Enable support for the xHE-AAC audio codec on Android P+
Android P+ includes support for xHE-AAC decoding. Unfortunately, ffmpeg
does not have support for decoding this content yet. So we can only add
support on Android at this time.

xHE-AAC uses an "audio object type" value of 42, which can't be written
into an ADTS header (only 2 bits in size), which required disabling ADTS
wrapping for src= and MSE when xHE-AAC content is detected.

For some reason our MediaCodec code has always been deconstructing the
ESDS package and only sending bits of it out. Per the MediaCodec docs
we should just send the entire ESDS. Since this may break things, I've
only done this for xHE-AAC profiles in this CL.

No metrics are added in this change, that will be done in a follow up.

Bug: 1019299
Test: Various new unittests; though no bots run P+ unfortunately.
Test: https://www2.iis.fraunhofer.de/AAC/xhe-aac.html on P+ manually.

Change-Id: I4343f202f2444dc0f3f391085da0ecc33a72be72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1992006
Reviewed-by: Yuchen Liu <yucliu@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Simeon Anfinrud <sanfin@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732635}