Print Preview: Fix resizing issue in destinations dialog
https://chromium-review.googlesource.com/c/chromium/src/+/1984717
fixed an issue with the iron-list sometimes setting an infinite
displayed item count by setting the height of iron-list using JS so
that it was always set to something nonzero before the list items were
updated. Unfortunately, this causes the list to never increase in
size if the dialog is resized larger after it has been opened, leaving
an odd empty space between the destinations list and the buttons.

Fix by adding a listener for the resize event that updates the height
of the list.

Change-Id: I7e010293831e5e1459cf1f0f6ba429dd52825b94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2006110
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732688}