Fix context loss when falling back to the software compositor.
We correctly handled context loss when using GPU compositing, but
if we transition to software compositing from GPU compositing, we
were not correctly starting the frame sink.

It's possible we also hit this case outside of context loss if the
OnReceivedContextProvider() is delayed until after EnableSubmission();
I was able to reproduce this by adding a delay to when the context
came back. So likely this fixes some broken video playback for some
users on the software compositor!

R=liberato

Bug: 1027518
Test: Updated unittest.
Change-Id: I738a1a36f4d850248730cd61b9ae4adcb5c1cd06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2006193
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732692}