Filter pinned extensions as actions change
This filters extensions when they are added/removed from the
ToolbarActionsModel. This happens not only on uninstall but also during
reload which is why we need to re-filter on adding extensions.

This leaves a TODO to actually remove pinned-extension IDs when
uninstalling an extension to keep traces of previously-installed
extensions.

For now it'll address a crash where entries in pinned_action_ids() could
refer to extensions that were no longer installed, resulting in a
null-pointer dereference when trying to update their position in the
toolbar.

Bug: chromium:1040856
Change-Id: I3928e931c12d6571c8eb5ad9bd255f0602cfa3d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2005373
Commit-Queue: Peter Boström <pbos@chromium.org>
Reviewed-by: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732706}