[Benchmarking] Increase shard timeout to 6 hours.
With the work to move gtests onto the shard maps, the overhead should
be reduced and we can change this from 5.5 hours to 6 hours.
This will reduce the chance of timeouts of these shards.

This is an adjustment from the earlier change
https://chromium-review.googlesource.com/c/chromium/src/+/1977965

Bug: 1041692, 1041696
Change-Id: I88661366c3f2ed058ed3d3b247c309d0f3e48957
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2006612
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>
Auto-Submit: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732716}