Reformat all gn files in /ios/chrome/browser/itunes_urls.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=mrefaat@chromium.org

Bug: 1041419
Change-Id: If6f5add12f826fc758e76b6540736e540b871b2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997910
Auto-Submit: Nico Weber <thakis@chromium.org>
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Reviewed-by: Mohammad Refaat <mrefaat@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732718}