Remove prefinalizer from ColorChooserPopupUIController
ColorChooserPopupUIController's prefinalizer closing the popup allocates new objects which is not allowed.

ColorChooserPopupUIController's prefinalizer which closes the popup should not exist.
In normal use, the popup will be explicitly closed before ColorChooserPopupUIController is destroyed.
The only exception is the Internals::endColorChooser() which releases the ColorChooser without closing the popup.

This CL fixes Internals::endColorChooser().

Bug: 1040389, 1039092
Change-Id: Ib52c80e5288ac7a115409706c96c3ba226357dd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2004365
Commit-Queue: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732736}