[PasswordGeneration] Conditionally added '0' and '1' to numeric alphabet.
Currently, the '0'/'1' digits are not used in generated password to avoid
visual confusion with letters like 'l','I' and 'O'. With this change, the
digits are added to the numeric alphabet if the generated password does
not contain lower and upper case letters.

Bug: 1038659
Change-Id: I25552b85592b8a3f2e286b9c30652db6664eff00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1985768
Commit-Queue: Matthias Körber <koerber@google.com>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732775}