Introduce sync IsTrustedVaultKeyRequired()
This API is analogous to IsPassphraseRequired() and different from
IsTrustedVaultKeyRequiredForPreferredDataTypes() in the sense that it's
not influenced by the currently-selected sync datatypes.

The rationale for introducing is that, when pending keys exist, certain
operations such as setting up a custom passphrase should be disallowed
by the UI, to prevent data loss. This is because having pending keys
means the locally accessible keybag does not contain all keys, so e.g.
setting up a custom passphrase would produce a new keybag where such
keys may be missing. As a consequence, worst-case, this can effectively
lead to user data loss, since there could be synced passwords encrypted
with the keys that have just been lost.

In this patch, setting pages on various platforms are updated to consume
use the new function instead of
IsTrustedVaultKeyRequiredForPreferredDataTypes().

Bug: 1019687
Change-Id: I6b5c69aac12b05ea34b4edc7dc149c84bd23d5c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1972140
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Maksim Moskvitin <mmoskvitin@google.com>
Cr-Commit-Position: refs/heads/master@{#732818}