Make SelectPopupDialog and SelectPopupDrown reusable.
This change makes above classes independent of SelectPopup, which will
allow reuse of those widgets in other parts of Chrome.

Specifically, they will be used in autofill_assistant for the selection
of time slots and similar information.

This is a refactoring only; there are no user-facing changes.

Bug: b/147577012
Change-Id: Id2d981345640ecf3b201ad4b64e0b5e2c3145426
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2002620
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732833}