[WebLayer] Tie network detection to Browser lifecycle.
Tying to BrowserFragment lifecycle leaves a loophole. If
setRetainInstance() is called on BrowserFragment, it will be detached
and reattached without necessarily being destroyed and re-created.

In the duration the BrowserFragment is alive but not attached, it'll not
receive any network change notifications. Tying registration to the
Browser lifecycle fixes that.

Bug: 1018848
Change-Id: I63be218a0fa2f8422f5cd9cc0d6e2dee2fa2be0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998398
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Reviewed-by: Tobias Sargeant <tobiasjs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732899}