Move ownership of main frame RenderWidget to the main RenderFrameImpl
Previously it was owned by the RenderViewImpl, in order to let its
lifetime extend beyond that of the RenderFrameImpl. Since that is no
longer needed, we can move its ownership over to the RenderFrameImpl
as it is for other local roots. This gives RenderFrameImpl consistent
ownership, creation, initialization, and destruction paths for all
frame-based RenderWidgets.

The GetWidget() accessor is misleading since it can be null and is
dropped. Instead, callers go through GetMainRenderFrame() which is
null for more obvious reasons, and get the RenderWidget from there.

R=avi@chromium.org

Bug: 419087
Change-Id: I268d7ef12e32f5ae131ecc1c7978eb1bc16cadf2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2005389
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732913}