Replace GURL static initializers with functions.
This is a necessary pre-cursor to cleaning up scheme initialization.
No GURLs should be created before schemes are modified/added and GURL
is ready to use. Enforcement of this will come in a followup CL.

See the bug for additional context.

Replacing static initalizers with functions ensures the GURLs are
only initialized when used, which happens after test setup.

Bug: 783819
Change-Id: Ifafe4a45b69eac0df10ac2125a328d24de130bd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1918828
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732950}