Print Preview: Move non-title elements out of dialog <slot='title'>
Remove some noise for screen readers in dialogs by moving non-title
elements out of the [slot='title']. Anything in this slot is read
out when the dialog loads. Move these elements to the body instead.

This change requires updating some CSS styling, in order to keep the
lists the only piece of the dialog that scrolls and to keep the
appearance of the dialogs the same.

Bug: 987324
Change-Id: I0ae6f7b338b7df988c12289f7eca2e876bd99a92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2006216
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732964}