[Cast UI] Specify the correct background color for sink button text
The button text changes its color depending on the background color to
stay legible. This CL specifies the correct background color (that of
ScrollView that the button is contained in) so that the text is rendered
in a legible color.

Bug: 1033670
Change-Id: Ib5c3f15a5e6512a2d259f5fd20dce7fb58ce13a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2008057
Commit-Queue: Takumi Fujimoto <takumif@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732986}