Chrome OS: Add key code for privacy screen toggle
Some upcoming Chromebooks will have an electronic privacy screen which
can be toggled on or off via a hotkey. When this key is pressed linux
will emit a KEY_PRIVACY_SCREEN_TOGGLE keycode 0x279. There is not
currently a USB HID usage for privacy screen control but there is a
request in progress and is expected to be in the 0x300 range of the
consumer page. We use the next available device DomKey 0x60C and the
next unused VKEY 0x9B.

BUG=b:146006851, b:138397511, 952051

Change-Id: Ia519c6227ca1acf3f352ef1d757b0ef3925dca05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1979232
Commit-Queue: Mathew King <mathewk@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Gil Dekel <gildekel@chromium.org>
Reviewed-by: Kevin Schoedel <kpschoedel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733021}