[TTS][Template URL] Add support for exact-search.
Adds support for a new Contextual Search Resolve request parameter:
"ctxsl_exact" boolean used to restrict the server to only resolve
a search for the exact selection.  This means the returned Search
Term is restricted to the exact selection and the selection will
not be expanded.  This is needed for user-adjusted selections
as part of the Long-press triggering experiment.

BUG=956277

Change-Id: I5f1285ce04bd4afff7250fbc76bca79c4da0dbf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994265
Commit-Queue: Donn Denman <donnd@chromium.org>
Auto-Submit: Donn Denman <donnd@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733058}