IDN Display: Add whole-script-confusable characters for Myanmar
This CL adds a list of Myanmar characters that look like Latin
characters. A domain consisting only of these characters and not
having a .mm ccTLD or other Burmese TLDs will be
converted to punycode, as is done for other whole-script-confusables
such as Cyrillic and Hebrew.

No popular domain is affected by this change.

Bug: 722167
Change-Id: I8d099e716f79ef42ab3205fe590e5e9b9a1db7a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2005869
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733059}