overview gesture: Fix the crash when dropping window to overview.
The crash happens if the window is dropped in overview before the
overview start animation is done.

OverviewSession::OnStartingAnimationComplete assumes that
|should_focus_overview| is false only if splitview mode is active, and
tries to activate a window in splitview, but with the cl
https://chromium-review.googlesource.com/c/chromium/src/+/1995948 in
place it's no longer true. To fix the crash, explicitly limit the logic
to when splitview mode is active.

Bug: 1042476
Change-Id: I6667b9d23a6ce25197642f1db0e493e90bc21a56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007515
Commit-Queue: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733071}