Updates when window is snapped during window drag from shelf
Currently, there is a bit of ambiguity about what happens on drag end in
drag_window_from_shelf_controller if overview is not started - the
controller will attempt to scale window down to go home, but may also
attempt to snap the window in overview (if gesture satisfied
preconditions for snapping).

This cl fixes this by allowing window snap only if overview is active
when the drag ends (otherwise the window will be scaled down to go
home).

BUG=None

Change-Id: I7e7fd22bc6e803934779528f2e95cc5589b1bbe0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2004009
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733082}