Merge to M80: ChromeVox should ignore focus events when the event source is touch gesture.
When you use ChromeVox to navigate to something focusable, like a link or
button, ChromeVox focuses it. Suppose you navigate to the next object and
it's not focusable. Focus gets cleared, which is equivalent to moving focus
back to the body element.

In r716808 (https://crrev.com/c/1895791), focus events on the body were
fixed to properly report eventFrom. Normally this is good, it allows
ChromeVox to skip events that were triggered by actions.

This accidentally broke touch navigation, though, because the logic to skip
events from ChromeVox actions explicitly excluded touch gestures. What that
meant is that if you used a touch gesture to move to the next element, and
that moved focus back to the body, ChromeVox would sync back to the document
body and you'd "lose your place".

As a quick fix, adjust the logic so that focus events are still
skipped even if they're from touch gestures - when the source of the
focus event is from a ChromeVox action. As a follow-up we may want to
improve that logic further.

Bug: 1038121
Tbr: akihiroota@chromium.org
Change-Id: Iafd447c69b005aa5bcb030c9ac23c5e4895ad0ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007982
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/branch-heads/3987@{#615}
Cr-Branched-From: c4e8da9871cc266be74481e212f3a5252972509d-refs/heads/master@{#722274}