Add checks to help track down crbug.com/1006814.
Upgrade some DCHECKs to CHECKs. In all cases the pointers are used
unconditionally, so this doesn't cause any new crashes in release,
just makes them a little more explicit. The number of crashes related to
these is quite small.

Add some extra debugging which will catch invalid parent_routing_id on
both the renderer and browser side (DumpWithoutCrashing). It's unclear
that this is actually the problem but this will confirm it.

I've dumped my attempt to understand the most common crash-point into
this doc

https://docs.google.com/document/d/1Ox8b0v273L4-kU0gu4xh61Bsp9ucZP_FncfW_kI9RoY/edit


Bug: 1006814
Change-Id: If85ca2f7fc2ae21b0c3f5030880759cc4de09346
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2006996
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Fergal Daly <fergal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733192}