media/gpu/V4L2VEA: Move SetExtCtrls functionality to the V4L2 device.
This CL moves the SetExtCtrls function to the V4L2 device, so it can be reused
in other places. Additionally this reduces the number of IOCTLS executed
directly inside the VEA, to provide a better abstraction of the V4L2 device.

This CL also introduces a V4L2ExtCtrl wrapper struct, that reduces the amount of
code required to send a request. It also makes sure the v4l2_ext_control
structure is always properly memset before executing a IOCTL. This is currently
not done in many cases, and could potentially cause hard-to-debug issues.

TEST=./video_encode_accelerator_unittest on hana

BUG=None

Change-Id: Ia0a785dc75456e8265b58ceebc292cb028e14d51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1969093
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733211}