media/gpu/V4L2: Use the SetFormat() function of the V4L2 device in the VEA.
Rather than performing a 'VIDIOC_S_FMT' IOCTL directly in the VEA, this CL
changes the VEA code to make use of the SetFormat() function in the V4L2 device.
This improves code reuse and reduces the number of IOCTLS executed outside of
the V4L2 device.

TEST=tast run $HOST video.EncodeAccel.*_360p_i420 on hana

BUG=None

Change-Id: I9d21fcd75fce00d719a79fc9d88137856717b96e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1975495
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733218}