Set the real value to credentials_mode
By this CL(https://chromium-review.googlesource.com/c/chromium/src/+/1994854),
the credentials_mode is given from a renderer process to the browser process.
theHere, the browser process doesn't use the value for shared worker script fetch yet.
In this CL, we use this value instead of the default value which is currently used.
This only changes the behavior when script_type is 'module'.
For classic scripts, we use 'same-origin' credentials mode.
This is the same behavior as before.

This CL doesn't change/add tests because module shared workers is still
not ready to write tests. Later CLs will implement them.

Bug: 824646
Change-Id: I041c88a466c76f7e2c108fa207d458ac1404d8b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994858
Commit-Queue: Eriko Kurimoto <elkurin@google.com>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733232}