[Files ranking] Fix a failed DCHECK in DriveQuickAccessProvider.
The results we get from the Drive API contain *absolute* paths, where
the path root is the the root of user's Drive. We use these as
*relative* paths within the device's DriveFS mount, and so we need
to reparent the paths on top of the mount path.

For example, if we get the result:
  /root/my_drive_file.gdoc
And DriveFS is mounted at:
  /home/chronos/u-.../my/drive/mount/
We want to produce the path:
  /home/chronos/u-.../my/drive/mount/root/my_drive_file.gdoc

But right now we don't account for the absolute-ness of the result
path, so we actually produce:
  /home/chronos/u-.../my/drive/mount//root/my_drive_file.gdoc

These paths are still valid and so don't cause bad behaviour (POSIX
paths ignore duplicate slashes in the middle of a path), but it causes
a DCHECK in the file path code.

This CL just strips the leading separator from the result to fix the
issue.

Bug: 1022150
Change-Id: Ic024af205967000f0d8d8f68304cc131ad6f41b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007798
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733244}