[quickview] Improve error reporting in updateQuickView
Undefined errors in the console are expected when for example, the user
mashes an arrow key. This causes getAvailableTasks_ to reject a promise
with undefined as it tries to throttle access to the metadata system.

Improve this: Only log errors to console that are defined, and add a
stack trace.

Bug: 1038799
Id: I843660a87456a92c3b73548cb5d5dbbde686d6e2
Change-Id: I843660a87456a92c3b73548cb5d5dbbde686d6e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2009602
Commit-Queue: Noel Gordon <noel@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733263}