[iOS] Reload SearchEngine if changed while editing
This CL reloads the SearchEngine TableView if it has changed while the
user was editing it.
This can happen because when the user select a Search Engine as default
it is updated in the TemplateURLService. But as the table view is being
edited, for now  this change is ignored.

With this CL, when the TemplateURLService is updated while the user is
editing the TableView, a comparison with the current TemplateURLService
state and the state of the TableView is done. If it is different, the
TableView is reloaded.

It also re-enable TestChangeProvider which was disabled because of that.

A visible UI change introduced by this CL is that when a custom search
engine is selected, it should be displayed in the first section.
Because we weren't updating the TableView when selecting a search engine,
previously this was only visible if the user left the search engine
screen then re-enter it. Now this is happening instantaneously.

Bug: 1036445
Change-Id: I2a28c030a443663d9ea4dbd5c5dcfe0c5c805dbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2002585
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733302}