[SSL] Move MITM check for enterprise-managed to //chrome-level factory
Currently SSLErrorHandler internally checks for whether the app is
enterprise-managed as part of showing a man-in-the-middle interstitial.
However, this check involves //chrome-level concepts and hence cannot
be brought down into //components when SSLErrorHandler is componentized.

This CL moves the check into ChromeSecurityBlockingPageFactory, which
*is* the //chrome-level configuration for this interstitial. As part of
this change, the infrastructure for overriding the value for tests is
made significantly simpler. We also eliminate unittests of
SSLErrorHandler that overrode the value and checked that the
overridden value was honored; these tests are no longer relevant as
those unittests don't use ChromeSecurityBlockingPageFactory, and the
browsertests that remain provide a superset of that coverage in any
case.

Bug: 1030692
Change-Id: If1c8ed4813d80ba4bde5c83503ebcaa5d57858ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007211
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733318}