Update GpuProcessHostDisableGLBrowserTest.CreateAndDestroy Disabling Hardware
GpuProcessHostDisableGLBrowserTest.CreateAndDestroy wants to use software
compositing, not the hardware path. The vulkan_content_browsertests were
enforcing that software fallback not be allowed. Causing this test to crash.

However BrowserTestBase::UseSoftwareCompositing already is in use by tests which
wish to explicitly be software compositing only. I'm adding this to the setup
of the test to allow it to run in the vulkan_content_browsertests suite.

TEST=vulkan_content_browsertests
GpuProcessHostDisableGLBrowserTest.CreateAndDestroy

Bug: 1009664
Change-Id: I85f00e2e3a4bf710d946ac69d46bd3e0e0879233
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2003251
Reviewed-by: Jonathan Backer <backer@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733397}