No absolute position in smart card SAML PIN dialog
This is a pure cleanup CL. It replaces the use of "position: absolution"
in the smart card SAML PIN dialog (a.k.a. the security_token_pin dialog)
with the normal positioning.

This makes the implementation less fragile against future changes, as it
doesn't rely on the absolute coordinates of the PIN input field and its
sizes anymore.

Bug: 1024251
Test: take screenshot of the SAML smart card PIN dialog before and after the change, check they're equal
Change-Id: Ibce4d99d07016ef9437782ba528c7e3c2744b799
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007647
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733436}