Invoke RenderProcessExited() in MockRenderProcessHost::Cleanup()
This CL makes it so that the behavior of the mock class is closer to
real implementation of RenderProcessHost.

This change will be used in a follow up CL where the lifetime of a
class is tied to the existence of the renderer process.

Bug: 993029, 1035555
Change-Id: I0c933c7c0b3890bc561015be847e145fdd96410a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2006450
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733471}