Handle canceled state in generate{Source,Destination}Message
Sometimes, the feedback panel tries to call generateSourceMessage for a
feedback panel with a cancelled state. This will fail, creating an empty
panel. Handle this case to prevent that from happening. It doesn't
really matter what gets returned in this case because the panel will not
be shown anyway.

Bug: 1015415
Change-Id: I599ec0e348c756d93b1a739dba6a239174eec3b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2009603
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733476}