cc: Disable activate needs draw in webview
cc does not have control over when to produce a frame for webview /
synchronous compositor as it does over chrome. There is a lot of down
side to waiting for draw, when the pathological case for happens
frequently causing both poor performance on the main thread and higher
latency.

Synchronous compositor also had a fallback tick that runs "fake" draws
after invalidate to ensure the compositing pipeline is never stalled.
This is no longer necessary with this change, so remove fallback tick
and some unused code in swap messages.

Bug: 1025695
Change-Id: I240f01662ccdf6f9c35912c421ea1ef0e40b1cc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986934
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733492}