[quickview] Change openInNewButtonTap_ to use executeDefault
If the user opens Quick View after selecting multiple files of
different types, the user is unable to use the "open" button from Quick
View. This is due to the openInNewButtonTap_ (OPEN button) handler using
taskController_.executeDefaultTask, which retrieves the tasks for the
entire mutli-selection, rather than for the item shown in Quick View.

Fix this: create a local variable containing the tasks for the currently
viewed item, and retrieve the tasks from here when the "open" button is
clicked from Quick View.

This could potentially be racy. Will address this in a future patch.

Bug: 1038799
Change-Id: Ie4186f4f4162eff1b89d5334bd5b5bcab9089450
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010236
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733495}