ServiceWorker: Remove controlled frames using add_cleanup() helper
clients-matchall-client-types.https.html manually cleans up controlled frames at
the end of each test case. This doesn't work when the tests fail on the
way, and that unexpectedly prolongs lifetime of a service worker. This is the
reason why the second test failed with the following message.

  "Error: wait_for_state must be passed a ServiceWorker"

This CL fixes it by cleaning up the frames using add_cleanup() helper.

Bug: n/a
Change-Id: I0f0e71a55814d114665eff824c917c918ac993f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010600
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733498}