[quickview] Test the text file is displayed when multiple files are selected
Check-select an image and a text file, show the files in Quick View,
and down-arrow to the text file.

Check the text file is displayed in Quick View (tests a bug where
the text file was not being displayed, and fixed in CL:2007795).

Bug: 1038799
Change-Id: I7ef509de8ffcf418e9476c8216d0ea32d94ca011
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010601
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Commit-Queue: Rachel Sugrono <rsugrono@google.com>
Cr-Commit-Position: refs/heads/master@{#733501}