[ios] Allow use of chrome_browser_provider_forward.h from provider
The provider are not permitted to have dependency on //ios/chrome,
however they were able to forward-declare ChromeBrowserProvider so
they ended up with indirect dependency.

To allow moving the class to the global namespace, allow them to
include the forwarding header since this does not increase the
amount of "bad" dependency.

This is a reland of https://crrev.com/c/2010796 with the correct
path (sorry for the spam).

Bug: 1042208, 1043657
Change-Id: I9abe0f492ab1ff86ca095dc22beb437459e8e239
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010984
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733549}