Added a check against a race condition that would cause a request to...
... be removed before the bubble is shown.

The exact conditions that trigger the race condition are unknown, I've
not been able to reproduce the issue reliably enough to investigate.

It appears that before `ShowBubble` gets to run, something happens that
calls `FinalizeBubble` which removes the request. It happens so rarely
that the DCHECK we have in place has not surfaced the issue.

Bug: 1041222
Change-Id: I0413fa70c77f2a24e17646b21d40ad9fa1cec69f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2000754
Commit-Queue: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733571}