Android: Peak GPU Memory during Tab Change
We previously added measurements for Peak GPU Memory Usage during a tab
change on desktop platforms.

The path that controls this and sets up the compositor is different on
Android. This implements the measurement there, by having
CompositorViewHolder.java notify the C++ Compositor of the tab change.

TEST=CompositorImplBrowserTest.CompositorImplReceivesPresentationTimeCallbacks
Bug: 1022038

Change-Id: Iba8bec10d817042494074871c9f7c7acaa786f7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2001377
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: ssid <ssid@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733633}