Rename components/sqlite_proto's files
This is the first of two CLs renaming files and classes in
//components/sqlite_proto to be more generic (from their
initial loading predictor-specific names). The change is
split into two CLs to make git recognize that a couple
of the larger changes are actually renames, not deletes and
creations.

This CL just renames the files and updates includes; the second
change renames classes implemented in the files and includes
some minor additional cleanup.

R=alexilin

Bug: 1036494
Change-Id: I9e465e3feb2358fe4e0b62371206ee6603a3bb6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2012640
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733646}