Always allocate extra space for alignment of AudioArrays
Instead of trying an allocation to see if it's aligned, just always
allocate extra space for alignment.  We waste a bit of space, but this
should not be huge.  Arrays are typically at least 128 floats or more
(the render size), and we need 16-byte alignment so we'll only waste
3% worst case.  This simplifies the algorithm too.

This means we don't need the static int that leads to data races.

Bug: 1042879
Change-Id: I8ac368737ed7679efceae64e2ca867a4de7b493e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2011132
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733672}