Show "TPM update required" error on login screen
Implement displaying the "TPM firmware update required" error on the
Login Screen, in case this error is received from the cryptohomed
daemon.

This error is currently only generated when a smart card authentication
is attempted to be used on a Chromebook with a TPM chip that is known to
be ROCA-vulnerable (see
https://www.chromium.org/chromium-os/tpm_firmware_update). In order to
fix the error the user or the admin need, as described in the error
message, to install the update of the TPM's firmware according to the
instructions. Before this CL, the only thing the user would see is that
the login fails for some unknown reason.

Bug: 1011327
Test: manual - find a device with a ROCA-vulnerable TPM 1.2 chip, try to add a new user using a smart card, check that the error is displayed
Change-Id: I0039950903196f702994404304c8d89da7e93879
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010985
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733694}