Add meta:timeout=long to FileSystemBaseHandle-postMessage.js tests
Both the sandboxed and native versions of the NativeFileSystem
postMessage() test are flaky due to timeouts in the CI.  Looking at
recent results, these tests take 5 or more seconds to complete
successfully.  This amount of time is dangerously close to the 7 second
timeout, which sometimes occurs, causing the flakiness.

This change adds "meta timeout=long" to both test variants to extend the
7 second timeout to a longer value, which gives these tests enough time
to pass.

Bug: 1043435
Bug: 1043028
Change-Id: I73eab928bce53a36be8c33e4b6e5746e7195f16a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2013580
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Steve Becker <stevebe@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#733757}