Allow copying selected contents from an image document
When copying in an image document, in most cases, there's no selection,
and we just want to copy the image itself. So we have a fast path for
that.

However, Javascript may insert other contents into an image document,
which can be selected and should be copyable. This patch detects this
case, and if so, uses the normal code path instead of the fast path.

Bug: 1042468
Change-Id: I45f5f296dc93040fd770f017b6220c51048b954b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007657
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733819}