Change default behavior to http cache reset to not reset.
In the case where the previous state of the finch trials was
unknown (when the previous status was not saved previously
due to this feature not being in the code), we had been
defaulting to resetting the cache.  This CL changes that
behavior and only resets when there is a known change to the
cache.  It also improves the testing to use PRE_ tests to
more accurately model reloading behavior rather than relying
on explicitly setting variables.

Also fixes flaky behavior introduced due to cross-process
testing.  Note the TODO regarding the bug to clean up the manner
in which we resolve the race condition.

Bug: 1041542,1041810
Change-Id: I9c40c9be9682b5372e1f3fb2851ef6531802f1bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2003249
Commit-Queue: Eric Robinson <ericrobinson@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733822}