If Notifications service isn't registered, check if it could be activated
DBus supports launching services on demand the first time a call is made to them.
In case the service isn't currently running, check whether it would be activated
when we try to call it.
Especially in case of session restoration the browser might be launched before
the shell's notification service is fully up and running.

Change-Id: I6b2905aa1d74fe4c99a10b27637d842f7fb836cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1975711
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733842}