Check the Drive integration service is mounted before use
In guest mode, the Drive integration service exists, but it is not
mounted or enabled, meaning some operations on it will lead to a crash.
Use drive::util::GetIntegrationServiceByProfile instead of going through
the factory, as it also checks if drive is mounted.

Bug: 1029477
Change-Id: I5f3e71ee7e788507613f5fe6f158db4e0c4add87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2011467
Commit-Queue: Austin Tankiang <austinct@chromium.org>
Auto-Submit: Austin Tankiang <austinct@chromium.org>
Reviewed-by: Sergei Datsenko <dats@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733890}