Make Plugin VM installer check for existing VM
The Plugin VM installer currently fails if a VM already exists. This can
happen if the user manually sets up a VM via vmc. This CL updates the
installer so that it first checks whether a VM exists, and proceeds to
the finished state immediately if so.

Bug: 1038816
Change-Id: I19381288c7fd2f2beaa82ffdb72ae5102d1f4c0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2014203
Commit-Queue: Timothy Loh <timloh@chromium.org>
Reviewed-by: Julian Watson <juwa@google.com>
Cr-Commit-Position: refs/heads/master@{#733911}