[aw] Avoid treating dynamite proguarded class names as WebView.
When we determine whether a stack frame refers to WebView we look up the
class by name in the WebView classloader. If dynamite module code used
by the app has been proguarded, then it's possible for there to be a
class name collision, which causes us to treat it as a WebView class.

We can avoid this case specifically by relying on the fact that GMS
proguarding sets the source filename to a predictable value.

Bug: 1043656
Change-Id: Icff4d7eb931d8295902cdda2483da2578f071cd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2014445
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Tobias Sargeant <tobiasjs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733969}