Expect presentation only if swap buffers will happen
There are situations when viz::Display::DrawAndSwap will Draw but will
skip Swap (when resize is happening and we have CopyOutput requests).
For this frame presentation callback will not be called.

This CL moves creation of PresentationGroupTiming for frame only if
|should_swap| is true to avoid waiting for presentation that will
never happen.

Fixed: 1042938
Change-Id: I508679b2e5c7d2dceeba3c915479db7617d432ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2008354
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734004}