CaptivePortalService: Introduce ability to register for callbacks
We are in the process of componentizing CaptivePortalService. As part
of this we need to eliminate its sending of a notification: as there
is no enum of notification types in //components and we don't want to
introduce one when notifications have been deprecated for 7+ years.

This CL starts the process of that elimination by introducing the
ability to register for callbacks when CaptivePortalService completes a
query, notifying the callbacks at the same time as the notification,
and porting a couple of initial consumers to validate the approach.
There remain a handful of consumers to be converted.

Bug: 1030692
Change-Id: I2bc682a092fac0c3ab2ebf4b152d6ee62e80d324
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010775
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734039}