Generalize kAccountStorageOptedInAccounts pref to per-account-settings
Before this CL, the pref kAccountStorageOptedInAccounts was a map from
hash of account ID to bool. Now it is generalized to a map from hash of
account ID to dictionary, and the previous bool is just one entry in
the dictionary.
This will make it easy to add other account-keyed settings later.

Since the previous pref was never launched, there's no need for any
migration logic.

Bug: 1035407
Change-Id: Ib60f30c644bdf457da841b709c433f35572df543
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2014449
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734069}