Rebaseline layout tests with FormControlsRefresh enabled (CL 4-1)
The [1] CL enabled the FormControlsRefresh feature by default for
Windows, Linux, and ChromeOS. In the process, many layout tests were
disabled, because the form controls have a different appearance with
the feature enabled. This is one of several patches rebaselining those
disabled tests.

SHERIFFS: If one of the tests touched in this CL becomes/is flaky,
*please* add it to TestExpectations beneath the section for bug
1035582, and mark it with a *new* bug number. At this point, any
further failing tests are flaky on their own, not because of this
rebaseline. Thanks!

[1] http://crrev.com/c/1967608

Bug: 1035582
Change-Id: Icf39d2aca9cf8c3d12c3f4e0709e3601ef30a014
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2008353
Commit-Queue: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734074}