[iOS] Add text zoom persistence
The text zoom level is stored in a user pref and keyed based off the
current url host and the current iOS content size category. This means
that if the user changes the zoom on one page and then changes the
accessibility text size, the earlier change won't affect the page in
the new text size. Thus, if a page is broken at one accessibility text
size, any user adjustments will not affect other text sizes.

This does not clear the preferences when browsing data is cleared. That
will be added later.

Bug: 1028938
Change-Id: I5a3a71f2bfa3d84750f714d2e2cffa869f40ccc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2002520
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734100}