Use SecurityContextInit to initialize SecurityContexts
This wrapping up of parameters leaves SecurityContext construction
much more readable.

To enable this, SecurityContextInit needs to be usable as a const
reference in the SecurityContext constructor. This doesn't work
currently, because there are two functions that move and return a
std::unique_ptr. This CL therefore changes construction of the
SecurityContextInit to store the state need to later construct
the objects that will be returned wrapped in a std::unique_ptr.

Bug: 1029822
Change-Id: Ic1f17ac22aa38eb77614735d41609941457d6698
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2006074
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734106}