Copy whitelisted enabled features from sign-in screen to guest session
The AutoNightLight feature is enabled only on kohaku via the chromeos
session manager [1]. However, the session manager does not start the
guest session, and therefore, does not setup its command-line flags.

This CL adds the ability to copy a list of whitelisted features to
the guest session command-line if they are enabled in the current
session (the sign-in screen).

[1]: https://cs.corp.google.com/chromeos_public/src/platform2/login_manager/chrome_setup.cc?q=p:chromeos_+AutoNightLight&g=0&l=105

BUG=1044344
TEST=Make sure AutoNightLight is enabled on kohaku in guest sessions.

Change-Id: Ie18e9a0e98d4b54bcd753d3bdbb8ecd422fa16e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2013739
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734114}