Re-Enable SkiaRenderer in blink_web_tests
We removed --fuzzy-diff from the default Linux configuration in:
https://chromium-review.googlesource.com/c/chromium/src/+/1989866

As a part of this we needed to disable SkiaRenderer, as this feature had
been developed with -fuzzy-diff on and we were not certain the impact.

We have since tested re-adding it, and it appears that some rebaselines
are needed. This requires 130 baseline changes in the main tests, along
with similar rebaselines for these tests which have been duplicated in
various VirtualTests.

The vast majority of the differences are small color variations in
anti-aliasing, and some gradient representations.

Some ref_tests cannot be rebaselined, issue 1043675 was filed to track
those affected by SkiaRenderer so that we can triage them accordingly.

TEST=blink_web_tests
Bug: 1039401, 1043675, 1044032

Change-Id: Idfcbdac1416067d6922b20a6cadc9180ecce82fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994088
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Jonathan Backer <backer@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734119}