Implement second FetchKeys() pass for sync trusted vault keys
This addresses a TODO in SyncServiceCrypto about honoring the result of
MarkKeysAsStale(), which is called when a previous call to FetchKeys()
returns keys that are insufficient to resolve the encryption issue.

If MarkKeysAsStale() reports true, it suggests another FetchKeys()
attempt is worth (usually meaning some local cache was invalidated).
This means another FetchKeys() call should be issued, as a second and
last attempt to resolve the encryption issue.

Bug: 1012659
Change-Id: Idd766233d0cacd2fb1030ca107e51efd667ad414
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2003132
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734143}