Add a fake ResourceScheduler for testing.
Another CL I'm working on concurrently involves adding some tests
involving CorsURLLoaderFactories, which are a little laborious to
initialize in tests, in part because a couple of the dependencies
are tricky to mock out. This CL makes it slightly easier by
adding a fake (TestResourceScheduler) for ResourceScheduler
to the network service's test support library.

Change-Id: I758c355ffe0d9241798d119cb12e2f36bb6f3fcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007523
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734154}