Fixed inconsistent behavior of MenuButtonController.
Fixed the trigger check for MenuButtonController so that it always
checks to see if a given menu triggering event is valid during an
OnMousePressed event.

Previously the timing check was made during either OnMousePressed or
OnMouseReleased. If the check was made during OnMouseReleased the
event may or may not have registered as a triggering event based on
the timing of the action (ie long press of the button), resulting in
inconsistent behaviour.

Bug: 1036441
Change-Id: I566fae882b7ef0e7c69a1359237b11f7c4a3d53d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2013747
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734158}