Skip infinite underline clip outs
We calculate clip rects for the underline intersections with glyphs to
clip out the underline. These intercepts are calculated by Skia and can
be infinite which results in an invalid clip-out paint op. This patch
skips these degenerate clips-outs. This is not done for all clips
because the common case is to clip from values from the layout system
which will be finite. This is not done in the intercept calculation
code because infinite values are correct for extreme glyph positions.

Bug: 1043753
Change-Id: I0b63eb892840994537b7fec4fd9b2313f74a7d14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2014310
Auto-Submit: Philip Rogers <pdr@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734213}