Generalize DirectlyCompositedImage support in cc
DirectlyCompositedImages in blink has a dependency on having compositing
information prior to the painting phase. This can't work the same way in
CompositeAfterPaint. We'll be moving support directly to cc, in which
we'll detect the situations under which we would have directly composited
images (i.e. some specific sets of paint ops within a paint chunk) and
adjust the raster scale appropriately.

Blink's usage always matches layer size and image size, due to the usage
of ReplacedContentTransform and directly setting the content layer
bounds. This means we can reuse the code to detect and compute raster
scale changes, where the current code is handling the case where (i.e.
attempt to use raster scale of 1). A future change will add support for
the paint op detection and set image sizes to non-matching values.

This change generalizes the directly composited image code to use the
image size rather than a boolean. Then use the ratio of layer bounds
to image bounds to determine the raster scale.

Bug: 875110
Change-Id: I4eb60f3f9b603b4e553104440695e13fc2fec43e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2009368
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734241}