Reduce size of test rectangles for Pixel_ReflectedDiv.
This CL reduces the sizes of the expected-colors rectangles for the
Pixel_ReflectedDiv test so that it passes on Nexus 5. The result in
that device looks visually fine [1], but pixels along some edges are
slightly different than the expected color:

actual   = rgba(255, 130, 4, 255)
expected = rgba(255, 128, 0, 255)

Therefore, reducing the size of the test rectangles to avoid the edges is
reasonable.

[1] https://chrome-gpu-gold.skia.org/img/images/f357a4dd6c87c632ae787203fcc307fa.png

Bug: 1033279
Test: Same reliability test as in CL:1979059.
Change-Id: I1d7ed61f5d1c7c2cacacb45d2d72b363580cba9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015560
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734248}