Animate the "Start" button for smart card user pod
Implement the animation of the "Start" ("->") button on the Login
Screen for the users that use challenge-response authentication
(a.k.a. smart card login).

The animation is, in accordance with the UI mocks, a continuously
increasing arc around the button.

This change also removes the code that was hiding the "Start"
button - this is replaced by the animation.

Bug: 1016839
Test: log in as a smart card user, log out, click "->" to log in again, verify that the "->" button is animated with a continuously increasing arc
Change-Id: I77e830522c5a0770d583d661122cb3831fbb5390
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2005988
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734251}