Reland blink: Ensure correct shared image lifetime for static bitmap images.
This reverts commit d196259f3397839f8d431c93384bb98cec436757.

The error here was trying to delete a shared image created using a
context before a GPU process crash. The service side considers an
erroneous shared image deletion command as a fatal error which in turn
destroys the new GpuChannel.

This change now depends on a patchset which ensures that shared images
are automatically deleted once a context is deleted avoiding the need
for explicit deletion from the user.

TBR=klausw@chromium.org
R=kbr@chromium.org

Bug: 1042446,1042481,1036142
Change-Id: I886460b8b1a62617ae987a1c96e0626fc3eb6e65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2008076
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734263}