Add a convenience accessor for current_frame_host.
This is accessed in a bunch of places include in the constructor where
it passes through an extra call to render_manager that seems to be
redundant.

This is a no-op refactor that just tidies things a little. I'm open to
better naming and docs.

Change-Id: I954e9ecf46a77de7ea9528eb527c1204b9ada3cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010462
Commit-Queue: Fergal Daly <fergal@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734297}