weblayer: adds support for restoring navigation state
Which includes open tabs.

This should be the core support necessary for the Java side,
but java integration (and tests) will be done on top of this.

The browser_test does not run on Android as it creates C++
Browser objects from the C++ side, which is not how things
are setup to work now.

BUG=1033924
TEST=weblayer_browsertests SessionService*

Change-Id: I3cc4cbdd51d2baa055457a4a58b210d9e3f1d404
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2014051
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734307}