plugin_vm: settings pages respond to uninstallation
Open Plugin VM settings pages should revert back to the top level os
settings when Plugin VM is uninstalled.

In order accomplish this the top level settings cr-link-row needs to be
always visible, as os_settings can't dynamically hide rows. Instead the
contents of the row have become dynamic, and will either say Turn On if
Plugin VM is not yet installed, or show an arrow if it is installed. The
resulting UI is now the same as Google Play Store and Linux (Beta).

Bug: 1023256
Change-Id: I9ade77a4dff3e7799732608430f299b400338ec8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007803
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Timothy Loh <timloh@chromium.org>
Commit-Queue: Julian Watson <juwa@google.com>
Cr-Commit-Position: refs/heads/master@{#734321}