DevTools: Modify Page.getInstallabilityErrors to use structured data
This change modifies the shape of Page.getInstallabilityErrors [1] CDP
event to include the error ids and arguments that are needed to generate
the detail messages that are displayed in the Applications - Manifest
tool. The design is proposed in Localizability of CDP [2].

The strings that are currently present in installable_logging.cc will be
moved to the DevTools frontend. They'll remain in place so we are backwards
compatible. Separate CLs will be made to process the new data in the
DevTools frontend.


[1]: https://chromedevtools.github.io/devtools-protocol/tot/Page#method-getInstallabilityErrors
[2]: https://docs.google.com/document/d/1ZbCiXbVJZJaYqKZkQ_Y5RNpUqQd1JzO17PqFJZk2Ne8/edit?usp=sharing

Bug: 946860
Change-Id: I8af923c1554396c6d79a4233bf342357b3386c7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2008430
Commit-Queue: Mandy Chen <mandy.chen@microsoft.com>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Lorne Mitchell <lomitch@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#734336}