MediaError class
MediaError is meant to be a relatively small (sizeof(void*) bytes)
object that can be returned as an error value from functions or passed
to callbacks that want a report of status. MediaError allows attaching
of arbitrary named data, other MediaErrors as causes, and stack frames,
which can all be logged and reported throughout the media stack. The
error code and message are immutable and can be used to give a stable
numeric ID for any error generated by media code. There is also an OK
state which can't hold any data and is only for successful returns.

doc: go/media-error

Bug: 1043289

Change-Id: I11a7c6d667f9810ba85264e41c8d7ce184858b34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1867991
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: Ted Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734338}