Reformat all gn files in /third_party/chromevox.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=aboxhall@chromium.org

Bug: 1041419
Change-Id: I1e4d95d67075c6855b963f6ed419b352b1102705
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997555
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Alice Boxhall <aboxhall@chromium.org>
Commit-Queue: Alice Boxhall <aboxhall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734348}