Temporarily allow kErrorInvalidDlc in Plugin VM installer
Until the ebuild is actually updated to support DLC, the Plugin VM
installer shouldn't be failing due to DLC errors. This CL partially
reverts the behaviour of crrev.com/c/2001942 so we don't fail.

We can revert this CL once the ebuild is updated.

Bug: b:145814572
Change-Id: I2a9aadef45f5e72d1e7841f0ec3438d4d19f352c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015871
Reviewed-by: Julian Watson <juwa@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734350}