Force-installed extension should not be disabled if was pre-installed
For manually installed extensions, the install location is set to INTERNAL.
When it is added to the force installed list, the install location should be
changed to EXTERNAL_POLICY_DOWNLOAD. This can be done by:

Unloading the extension -> changing the extension install location -> Loading
the extension.

In case the extension is disabled by user, remove the disabled by user
DISABLE_REASON as it is now required by policy.

Bug: 1014123
Change-Id: I1329e62708e1ca083af1becfddf75d602dada58b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1911754
Commit-Queue: Swapnil Gupta <swapnilgupta@google.com>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734404}