Eliminate remaining usage of CAPTIVE_PORTAL_CHECK_RESULT notification
The sending of the CAPTIVE_PORTAL_CHECK_RESULT notification is a
blocker to the ongoing effort to componentize CaptivePortalService. A
previous CL has introduced the ability to register for a callback with
CaptivePortalService as an alternative mechanism of being informed on
the results of captive portal queries. This CL converts the remaining
consumers and eliminates the notification.

Bug: 1030692
Change-Id: I9a2c5f6190922eb6ddb10d572dcbec6b52ffc3ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2012383
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734407}