Componentize CaptivePortalService
All //chrome-level dependencies have been cleaned in preceding CLs, so
the componentization is straightforward. We leave the componentization
of the unittest to a followup as it still has //chrome-level
dependencies to be cleaned. We also defer putting the files in the
captive_portal namespace until all of the related componentizations
have been completed to avoid churn on files that are currently in
//chrome but will shortly be moved to
//components/captive_portal/content.

This componentization is necessary as part of the ongoing effort to
componentize //chrome's SSLErrorHandler.

TBR=jochen@chromium.org

Bug: 1030692
Change-Id: I84bd90cd026cd461a349f8e5a95a6b8965837b05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2011926
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734449}