[WebLayer] Change shell to retain Fragment instance
This CL changes WebLayer Shell to retain the Fragment instance in order
to more accurately model the behavior of WebLayer consumers. As part of
this change, WebLayer Shell no longer handles rotation/resize internally
in its manifest and loads the startup url on weblayer creation only if
there is not already a navigation entry in the current Tab.

Note that to test rotation in an emulator it is necessary to turn on
autorotation, which is off by default in crow.

Change-Id: I8a575e477d1242486f9fbbe72151392b597b86b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007413
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734505}