orderfile: cleanup: actually remove --new-commit-flow
The usage of the flag was removed in crrev.com/719331. I planned to
remove the declaration of the flag separately to make reverts smaller,
but forgot to remove it in the end. Do it now.

Bug: 1028600
Change-Id: Ia2cb8bd18389c25e440c2c95773b4c49ce120d64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2017405
Commit-Queue: Egor Pasko <pasko@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Auto-Submit: Egor Pasko <pasko@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734524}