[FlexNG] Fix percentage-heights-010.html
This fixes the above testcase, marked as failing in:
https://chromium-review.googlesource.com/c/chromium/src/+/2015101

Essentially a flex-item, if it has any %-block-size children, needs to
mark itself as depending on %-block-size, as its definiteness can
change.

See note in ng_physical_container_fragment.cc as this is a *little* too
broad, as we should only check if the definiteness changes, not the
percentage block resolution size.

Bug: 845235
Change-Id: I78120f5dd65186ef9ad597f8a5717923cfa49144
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2016634
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: David Grogan <dgrogan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734553}