Refactoring ClonedInstallDetector so it owns Metrics Id Reset Pref
This is the first step to enabling the ukm service to reset the
ukm client id when a cloned install is detected. By moving the
responsibility of the pref, both services can call in to understand if
we are in a cloned state.

Bug: 981407
Change-Id: I0a27bc0f14f5458cf685ff5abf7ea792e73246d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1992009
Commit-Queue: Joshua Berenhaus <joshber@microsoft.com>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Auto-Submit: Joshua Berenhaus <joshber@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#734573}