Remove root scroll offset code in Android accessibility.
Android accessibility was using a different code path than other platforms
to compute the bounding box of accessibility objects. In particular, we
were ignoring the root scroll offset from the accessibility tree and using
the offsets from RenderCoordinatesImpl instead.

It looks like this is causing problems with ImplicitRootScroller, and
furthermore it looks like this code is no longer necessary. Switching back
to use the scroll offsets from the accessibility tree seems to fix the
problem.

A follow-up change will remove UseRootScrollOffsetsWhenComputingBounds
entirely. This change is minimal in order to be considered for merging.

Bug: 1005121
Change-Id: I82e97f046cee6b8b08238fcd4368228da368b92a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2017662
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734630}