Metal: Test offline-compiled shader link time
Change metal_util::TestShader to test either compile or link, and
change the default behavior to link.

Add inline data for a trivial test shader. Add code to modify the
bitcode, to defeat caching.

Ensure that the Metal calls for the test are invoked on a thread
that has base::TaskPriority::HIGHEST.

TBR=jvanverth

Bug: 974219
Change-Id: I35e688c3758af2e03f893a1ccbef7aef6b778b23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015486
Reviewed-by: ccameron <ccameron@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734645}