[Sampling profiler] Eliminate unneeded Java interfaces for stack_unwinder
Strips the stack_unwinder module interfaces and contents down to the
minimal Java required to get the dynamic feature module to work from
native code. The module will be providing native code interfaces[1],
which will be consumed by native code, so it doesn't make sense to
interpose more Java than necessary in between. It's also the case that
the parts of the DFM Java API that deal with the Play Store can't be
used within the sandboxed renderer process.

This change exposes via JNI the StackUnwinderModuleProvider's functions
to query and request installation. This is the only Java-provided
functionality that will be required for manipulating the module from
native code.

1. Design doc: https://docs.google.com/document/d/1QB5IzWuVKfoPdPuiyRok2MUEta1FmRLIQkq1wDATA4k

Bug: 1027654
Change-Id: Ie69e687335a753c17206fc6c81ae981118c698b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2003298
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Reviewed-by: Fred Mello <fredmello@chromium.org>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Commit-Queue: Mike Wittman <wittman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734653}