Fix SetDrawRectangle() behaviour
There is logic in GLOutputSurface to ensure that SetDrawRectangle() is
only called once per swap. However, it's possible for
Display::DrawAndSwap() to to draw and not swap. GLOutputSurface will
ignore the first SetDrawRectangle() after Reshape() if the last swap was
skipped.

DirectCompositionChildSurfaceWin::Resize() releases the draw texture, so
it should be safe to call call SetDrawRectangle() again after Reshape()
even if SwapBuffer() wasn't called.

This CL also implements equivalent logic in SkiaOutputSurfaceImpl for
SkiaRenderer.

Bug: 1042007
Change-Id: Ib1a8f5092d004172d37fe4c16d8a7144cbc7ce8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2013346
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734656}