Keep Hotseat in extended mode after restoring the active window
Now hotseat is hidden after restoring the active window from overview.
The reason is illustrated in the bug description.

Actually the in-app shelf only exists in tablet mode where no mouse
is available (if there is the mouse connected, device will switch to
clamshell mode). So there is no point to check whether mouse event
is enabled for shelf widget in this scenario. So skip that code block
in tablet mode. Meanwhile, the CL uses the additional attribute to
ensure that hotseat is not hidden by window restoration.

Bug: 1044366
Change-Id: I6cd2568f2eb7a32aabb8a6be373118ec6c501634
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015802
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734659}