Add accessibility call to scroll to text
Screen readers read from the top of the page on a scroll to text
navigation, instead of the targeted text. This patch adds an AX call
that is made when navigating to an anchor, as is done with element
fragment anchors.

Confirmed this works with TalkBack on Android.

Bug: 1044644
Change-Id: If5c905b083189efd7276f64abc4a547f81f6b054
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015132
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Nick Burris <nburris@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734668}