AW docs: clarify commandline flag support
No change to logic, only docs.

This clarifies WebView's flag support. Namely, it explains that we don't
"support" all flags, but that we support toggling them (and the behavior
itself may or may not be stable). This also links to the layer diagram.

This clarifies some of the language around "features," defining
"base::Feature" and updating some references to use either
"base::Feature" or "Feature" rather than "feature" (which could be
mistaken for a general enhancement to the product).

Bug: 918221
Test: tools/md_browser/md_browser.py
Test: Upload to gerrit > open file > click "gitiles"
Change-Id: I570eeed005b8671a003028e0c45b8ac03d37fc71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2018151
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Shimi Zhang <ctzsm@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734717}