chromeos: Fix DCHECK when searching OS settings for not-found string
The C++ OSSyncHandler is only instantiated when the SplitSettingsSync
flag is set. However, the new os-sync-controls HTML/JS can be
instantiated by a search event when the flag is not set. This leads
to a DCHECK for an unhandled message.

Only instantiate the HTML if the flag is set. Do the same for the
old UI.

Fixed: 1044369
Change-Id: If09060e9916405afc524e669f235166cedf8c4eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2017986
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734718}