remoting: Set CHROME_REMOTE_DESKTOP_SESSION in user-session
CHROME_REMOTE_DESKTOP_SESSION is currently only set in the
chrome-remote-desktop Python script. This means that is-remoting-session
returns true for programs started by the script, but returns false for
in the user's .profile, which is necesarilly sourced before running the
script so that settings such as
CHROME_REMOTE_DESKTOP_DEFAULT_DESKTOP_SIZE can be set before the script
runs.

This change causes is-remoting-session to return true in .profile when
running under user-session before the script is launched. This will
allow the user to include things like

    if /opt/google/chrome-remote-desktop/is-remoting-session; then
      export IDE_SESSION_PREFIX=CRD
    fi

in their .profile instead of having to create and customize a
.chrome-remote-desktop-session file.

Change-Id: Ie0f6c4f59ae781360025ff8664145019b15a111d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1999763
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Erik Jensen <rkjnsn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734747}