Fix ignored node computation in ARC a11y.
crrev/c/1961352 changed how ignored states are computed in ARC
accessibility, but it made tast arc.AccessibilityTree fail.

Previously, all the nodes that don't have accessibility importance in
Android are labeled as ignored in Chrome. But the CL changed behavior
so that when a node's descendants have any important property,
unimportant node is recognized as important.
The unit test |AXTreeSourceArcTest:SerializeAndUnserialize| is an
example of how it changed.

This CL fixes the regression to make a node that isn't important should
not be exposed even if it has an important descendant.

Bug: none
Test: tast run DUT arc.Accessibility* (under http://crrev/c/1982005)
Test: unit_tests --gtest_filter="AXTreeSourceArcTest.*"
Change-Id: Icb10919f81a385beb5e706169f23cf755efa51a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2009599
Commit-Queue: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Sara Kato <sarakato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734770}