WebUI Tab Strip: Add ability to drag in and out of groups
The dragover/drop touch targets may need to be reworked, but that can
be done in future CLs. This CL sets up the basic functionality of
being able to drag a tab into a group to add it to the group or to
drag a tab out of the group to remove it from the group.

Bug: 1027373
Change-Id: I52e9e5c0cd4ea24d55714d6baf1c712997a4a558
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015678
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734789}