media/gpu/chromeos/MailboxVFConverter: Set visible size to SharedImage
MailboxVideoFrameConverter sets coded size to SharedImage. It
causes that a green line is shown at the bottom and right edge
when playing a video.

This fixes the issue by setting visible size to SharedImage so
that GPU doesn't access non visible area.
Note that MailboxVideoFrameConverter needs to recreate
SharedImage if the visible rectangle of the current video frame
is changed, which should be rare though.

Bug: 1043582
Test: Play a 1080p video on soraka with --enable-features=ChromeosVideoDecoder
Change-Id: I76878a27fa92d755380d9b92b94d20e765056fac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2011586
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734795}