[Captive Portal] Componentize CaptivePortalTabReloader
This CL componentizes CaptivePortalTabReloader and its unittest as part
of the effort to componentize CaptivePortalTabHelper. The latter is
itself needed by SSLErrorHandler, which we are componentizing for
sharing with WebLayer.

The componentization involves only injecting CaptivePortalService rather
than obtaining it internally via the Profile.

Change-Id: I54dfe0425e817e75cbd82cc5e4acb6a7a9edeea6
Bug: 1030692
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2014907
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734862}