[SSL] Componentize ssl_error_assistant.asciipb and PB generation
This CL componentizes ssl_error_assistant.asciipb and the infrastructure
that is used to generate a binary proto from the ascii file at build
time. The motivation is to support componentization of
ssl_error_assistant.*, which will itself be needed for componentization
and sharing of //chrome's SSLErrorHandler with WebLayer.

As far as I could ascertain from examination of the Python scripts,
nothing on the cloud storage side is impacted by where the binary proto
sits on the client side; thus, this change should be safe to make
without requiring any coordination with the operation of the component
updater for this data.

Bug: 1030692
Change-Id: I474b1a56c91d98e930047e9b122c9a3aa39dd84c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2011808
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734877}