Added integration test for covering bug https://crbug.com/1028577
This adds an integration test that covers a bug that was observed
when default_policy_allowed/blocked_hosts were specified but should
be different per browser profile. It turned out they were global.
This was fixed and the CL:1978680 introduces the test to cover this.

R=marcgrimme

Bug: 1031179
Change-Id: I6a5a0b10109ed4b358bd4344a56eeac705949198
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1978737
Commit-Queue: Marc Grimme <marcgrimme@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734909}