[Autofill Assistant] Merge client memory into user data.
This CL merges client memory into user data; removing parts of the client
memory and parts of the former user data. In a previous CL, user data
ownership was moved to the controller to allow it to be retained across
multiple collect user data actions. With this CL, client memory is
removed/merged into user data so that all the user values retained on the
client will eventually be kept in a single place. Also in this change,
selected_profile, billing_address and shipping_address are no longer
separate/individual fields on the user data, rather they are stored in
the autofill profiles map.
Note that the action delegate can now provide the user data instead of
the client memory. This is plumbed through from the controller via
the script tracker.


Bug: b/148132177
Change-Id: Ife130d7537680fd214459bc2d80610cd8c3dfea3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2014450
Commit-Queue: Marian Fechete <marianfe@google.com>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#734985}