Add the Display Observer for Direct Composition in Windows.
The purpose is to observe the events of monitor plugged-in/unplugged for
Direct Composition overlays so we can update the hardware overlay caps
upon display change. DisplayAdded() and DisplayRemoved() are added to
DirectCompositionSurfaceWin.

The display observer goes through the GPU data manager in the browser
process first. It will then notify the GPU switch observer in the GPU
process through Mojo.

This CL is limited to Windows only. The display observer might work (not
tested yet) for Mac if we need it later. (But no code support for Linux.)


Bug: 1042989
Change-Id: I53dc1689795d569ea60290c13dacfcc7f3ee9152
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2011134
Commit-Queue: Maggie Chen <magchen@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735009}