Add ExtensionRequestObserver for extension request
ExtensionRequestObserver observers the extension policy change. Once an
extension in the request list is put into whitelist, blacklist or
forcelist, pop-up an notification and remove the request from the pending
list once the notification is closed by user.

If there are multiple notifications approved or rejected or installed,
there will be one notifications for each decesion.

If there are new update when the notification is opened, the notification
will be refreshed.

Screenshot:
https://drive.google.com/file/d/1PW5w0grWBKfTWL4HZ7ilimyjlrn7s81w/view?usp=sharing

Bug: 1006899
Change-Id: I4ef0f68f4d171addb2d28de1ebfe6fc2baeb71e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2018149
Commit-Queue: Owen Min <zmin@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735016}