Fix AXPlatformNode::IsUIAControl
One of my previous change introduced a regression in determining if an
element should be considered as a UIA control. The previous CL changed
the default value returned by AXNodeData::IsReadOnlyOrDisabled from true
to false, but AXPlatformNodeWin::IsUIAControl's use of this method
needed to be adjusted to the new implementation.

This CL fixes it.

Bug: 928948
Change-Id: Ia5d2ff96e7200ad8806be19f52a249645037bf6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2018163
Reviewed-by: Adam Ettenberger <adettenb@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#735031}