Remove use of MAC_OS_X_VERSION_10_*
This changes expectations based on the macOS SDK used to build
Chromium.

MAC_OS_X_VERSION_10_12 is true if we use at least the macOS
10.12 SDK, which has been the case for years.

MAC_OS_VERSION_10_13 is a typo and thus is never defined.

Therefore, this always evaluates to true on the Mac.

Bug: 1044838
Change-Id: I3a56d37e7ea9d1acb65d6b8e07996de47da37794
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2020163
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735075}