Cros: Fix icon failing to animate from AppsGridView -> ShelfView
When the drag completes in ScrollableShelfView, there are two
possibilities:
1. The drag was Shelf -> Shelf
2. The drag was AppList -> Shelf

For scenario 1:
 - ScrollableShelfView owns a DragIconProxy |drag_icon_|
 - ShelfView treats ScrollableShelfView as the
   ApplicationDragAndDropHost.
 - ShelfView makes its DragImageView |drag_image_| invisible.
When scenario 1's drag completes, there is no issue, ScrollableShelfView
successfully animates |drag_icon_| to the proper bounds.

For scenario 2:
 - AppsGridView treats ShelfView as the ApplicationDragAndDropHost
 - ShelfView creates a DragIconProxy, |drag_image_|.
 - ScrollableShelfView does not create a DragIconProxy.
 - Animation for the scrollable shelf, implemented in
   ScrollableShelfView fails.

To fix scenario 2, pass ShelfView's DragIconProxy
(ShelfView::drag_image_|) and use it within ScrollableShelfView.

This change is not intended to be a full fix, just a change that we can
successfully land without refactoring.

I am avoiding refactoring because:
 - Branch is soon.
 - Kukui is coming out with this milestone
 - This code is already prone to bugs, I am worried there will be
   negative effects to a refactor

I created a follow up bug for this refactoring:
https://bugs.chromium.org/p/chromium/issues/detail?id=1045186

Also in this change:
 - Use a transform instead of a bounds animation.
 - Simplify DragIconDropAnimationDelegate

Bug: 1041999
Change-Id: I63a428f103a888ef019da919407a60533da68c07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2017923
Commit-Queue: Alex Newcomer <newcomer@chromium.org>
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735082}