Add TODO for missing IA2_EVENT_TEXT_REMOVED in live regions
When a generated live region event occurs, we currently fire
IA2_EVENT_TEXT_INSERTED. However, as recently discovered via
https://chromium-review.googlesource.com/c/chromium/src/+/2015429
we are not firing IA2_EVENT_TEXT_REMOVED in the generated event
case.

This currently adds a todo rather than attempting to correct the logic
for which events fire, as it is not known to cause any actual screen
reader issues.

Bug: None
Change-Id: I4d74acdf47c5909d3efe0dbf2acb9d20ffb98528
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2019410
Reviewed-by: Ian Prest <iapres@microsoft.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735083}