[IndexedDB] Adding last_modified and size to 'file' BlobInfos
The goal of this change is to avoid writing empty blob files in
IndexedDB solely to store the last_modified date.

This changes creates a migration for IndexedDB from v3 to v4. This
migration changes the way BlobInfos are encoded, and adds 'size' and
'last modified' to the 'File' BlobInfo.

Now that the metadata is changed, this change also avoids writing
'empty' files to disk. Previously this is where the last_modified time
was stored, but since that is now in the metadata these no longer need
to be written.

Bug: 1024966,1015214
Change-Id: I1309561373e0d917119fece4c552edd6ff6f5c18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2014060
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: enne <enne@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735149}