Updating expiry date for system-keyboard-lock flag
This flag is used for two purposes:
1.) BrowserTests which would fail if system-keyboard-lock were enabled
    (this is because the test infra installs its own keyboard hook
     which prevents the test from running correctly)
2.) Chrome users who want to capture all browser key combos in
    fullscreen mode but do not want local OS combos to be captured.

I'm extending the flag expiry out by several milestones as we are
working on an update to the KeyboardLock API which would allow the
caller to specify whether to install the system keyboard hook or
not.  That would mean that we can satisfy the two use cases above
via the javascript API and the flag can be removed.  We are working
on the spec updates now so hopefully we'll have the feature
implemented early this year

Change-Id: Ie2d6915eeb6351c96dd67bcd06f34ca07317a6cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2016349
Commit-Queue: Jamie Walch <jamiewalch@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Auto-Submit: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735152}