Wire up conversation starters fetching in Ash.
In conversation starters V2, conversation starters are fetched from
the server using the dedicated ConversationStartersClient.

When the server call fails to result in a set of conversation starters,
we fallback to locally provided conversation starters.

Note that this CL fetches as the UI is becoming visible to maximize
freshness. This logic may need to be changed once we assess latency.

Bug: b:148246719
Change-Id: Icc978d95588cb2088fd7aa7df83be69ded9cb22a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2017916
Commit-Queue: David Black <dmblack@google.com>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735173}