Refactor Android Consent Prompt lifetimes
In preparation for splitting out the DFM installation logic which lives
inside of the consent prompts, this change refactors the lifetime of
those objects to no longer be static. This will pave the way to
managing the installation logic in a much cleaner way.

Bug:1045130

Change-Id: Ib5150f690d9d5beafcf40e98afc7fd01bd6f5c1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015611
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735211}