blink: Replace Members with raw pointers for stack allocated classes
Since in incremental/concurrent marking the stack is considered white,
there is no need to issue write barriers for stack pointers. Therefore,
stack allocated classes don't need to wrap Members and can simply use
raw pointers or references to heap objects. This is consistent with
regular uses of raw pointers on stack.

This CL updates all classes marked as STACK_ALLOCATED to use raw
pointers instead of members. The main reason is to greatly reduce the
number of write barriers.

Since default initialization of raw pointers doesn't imply
zero-initialization, it must be ensured that pointers are always
initialized with a value or zero. As suggested in Chromium C++ Dos and
Don'ts, a member is not initialized with in-class initializer in case
all the constructors explicitly initialize it.

Bug: 1044611
Change-Id: I15587eeb407a624d499978fd3dfb7b8034d02056
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2014940
Commit-Queue: Anton Bikineev <bikineev@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735242}