Bake SDR white level into display::Display's color space
By doing this, we do not need to adjust the brightness of SDR content.
The only time that we need to adjust content is when displaying HDR
video that uses the PQ transfer function.

Adjust the default SDR white level to be 100, so that it will behave
correctly on macOS, now that Windows is no longer relying on the value.

This fixed the bug where PDFs were double-darkened, and may also fix
other filtering bugs, though that will need to be verified.

Bug: 1038770
Change-Id: I22dc9d50bd65897234905390e2c2e2f8416a41eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2018339
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735263}