Add wrapper of ServiceWorkerStorage::GetUserData{,ByKeyPrefix}() to ServiceWorkerRegistry
Currently ServiceWorkeStorage depends on ServiceWorkerContextCore
to schedule database error recovery (see the original design doc [1]).

This dependency needs to be removed so that ServiceWorkerStorage can be
moved to the Storage Service. In other words, we need to move
ScheduleDeleteAndStartOver() to ServiceWorkerRegistry (see the related
section of the refactoring plan [2]).

This CL is the first CL to move scheduling DeleteAndStartOver() to
ServiceWorkerRegistry. Specifically, this CL modifies
GetUserData{,ByKeyPrefix}() so that we can call
ScheduleDeleteAndStartOver() from ServiceWorkerRegistry.

Subsequent CLs will do the same for other methods in ServiceWorkerStorage.

[1] https://docs.google.com/document/d/1eXdgnAOZC4dDDybmRpXT0t0lOYGhFf7xajOSanEb91Y/edit#heading=h.kboaqpwv2bvf
[2] https://docs.google.com/document/d/1PXg_8SKtXqfwcwTfJaPHMyUeZ4el92nU76jDbEMwkt4/edit#heading=h.hrlvq4dn719m

Bug: 1039200
Change-Id: Ic91013a244538160a2600af6d9fe201358896003
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2018625
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735374}