Portals: Add portal AX role and omit contents from tree
This CL adds a new accessibility role (kPortals) for portals, which is
eventually implemented as a button on platform implementations (with
the exception of ChromeVox, impl will be in a follow up). This CL also
omits a portal's AX tree from the root AX tree.

Change-Id: I2b6b27d7c2d04116e7e4cfa23792bbd1188d7628
Bug: 1045608
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980898
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735396}