Implement sqlite storage layer for conversion measurement API
This change implements a storage layer in sqlite for persisting
impressions and conversions from the API to disk.

The storage for the API consists of two tables, an impression table and
a conversion table. Every row in the conversion table is associated with
an impression in the impression table to reuse common data such as
conversion and reporting origins.

The storage layer does not implement specific functionality like last
clicked attribution, bit limits, or reporting delays. Instead it abstracts
that functionality to an injectable delegate class.

This change also includes some common structs that are used by the API,
such as a new mojo struct blink::mojom::Conversion.

The functionality of the API is described on
https://github.com/csharrison/conversion-measurement-api

Reference prototype change:
https://chromium-review.googlesource.com/c/chromium/src/+/1967220

Bug: 1014604
Change-Id: Icff2d720fe7c595398835fcfd026f6d321cb9c99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1965450
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735418}