[Spellcheck] Move WindowsSpellChecker to its own file
This is the first step towards making the native spell checker class
(WindowsSpellChecker) more testable. Right now, it can't be easily
tested because its life cycle is tied to a base::NoDestructor pointer,
but it holds references to thread pools. So, between tests, the
WindowsSpellChecker is kept alive, but its references to the thread
pools are invalidated.

The next step will be to move the WindowsSpellChecker from a static
base::NoDestructor pointer to the BrowserProcessImpl object, which is
destroyed and recreated for every test.

In addition to the above, the spellcheck_platform_win.cc file was
getting very large, so that's another reason to split it up.

There are no functional change as part of this; only code being
moved around to new files.


Bug: 463364, 1035044
Change-Id: I0b621ad2beba6ed57856dd3cff1910aebebbb710
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2008031
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#735425}