Make tests navigate to about:blank instead of the NTP.
These tests open a new tab and immediately navigate to a different URL.
The problem is that chrome::AddTabAt by default opens the new tab to the
NTP, which takes a considerable amount of time to load. By specifying
that the new tab should open to the about:blank page instead of the NTP
I observed a decrease in execution time of up to 4 seconds for some of
these tests.

Change-Id: Ife413d2d102610012ff509a712fbfb39314d0f68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2019805
Reviewed-by: Patrik Höglund <phoglund@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Austin Orion <auorion@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#735512}