Fix HttpServerTest.Request.
The test was waiting for the client end of a socket to be connected,
but not the server end. In general, the server side is connected first,
but there's an intervening Mojo pipe in this case wrapping both client
and server connections, and it's also not clear if platform APIs are
guaranteed to notify the server of the connection first, either.

This CL makes both the Mojo and non-Mojo HTTP server tests not rely on
that ordering.

Bug: 1032620
Change-Id: I925b9220d39a8e2f54b224e46c1c1853d2fdd067
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023005
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735569}