Update WebXr on Android exit behavior
Currently when a user is in WebXr on a Daydream device, when they press
the "App" button to end the session they are taken to the VR Browser,
even if that is not how their session was started. After this change,
when the user presses the "App" button they will be taken to Daydream
home to continue a VR experience, unless they initiated the Session from
the VR Browser (in which case they will be returned to the VR Browser).
If the user simply takes the device out of the Daydream View headset and
closes the window, they will be returned to 2D chrome.

As a result of this change, a few tests needed to be updated.

Fixed:1046056

Change-Id: I8d0e8466e09a13aecccb331b0bf9fc2f2ebbf177
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2020905
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735587}