Retry event upload batches that contain individual event failures.
If a reporting UploadEventsRequest contains too many events, the
reporting server will only upload the first N events, and the
UploadEventsResponse will contain a record of which events were accepted
and which were rejected. The retry logic ensures that in the case where
there were some temporarily rejected events, that job is retried until
all of the events are either accepted or permanently rejected.

This involved some refactoring of the
DeviceManagementService::JobConfiguration so that a job can take part in
the decision of whether it needs to be retried, and to allow a job to
make changes to itself before it is retried.

R=poromov@chromium.org
BUG=1039739

Change-Id: Ic947f678d63fbe644e6af3575c7bea1dd970caba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2005898
Commit-Queue: Brian Malcolm <bmalcolm@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735590}