Reland: viz: OverlayProcessor Sends Candidates Directly for Android Pre-SC
For Andorid Pre-SurfaceControl, overlay processor can send the
overlay candidates directly.

This is a reland. The original CL was reverted due to clusterfuzz bug
and causing crash. The problem is the destruction order of objects owned
by the display. The OverlayProcessor class needs to be destructed before
the DisplayResourceProvider is destructed.

The crash is at crbug.com/1045574.

This is CL 6/? for sending the overlay candidates directly to gpu thread.

R=rjkroege

Bug: 979788
Change-Id: I8b7057977daa7900d415e3720e2a194efbcc0816
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023188
Commit-Queue: weiliangc <weiliangc@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735614}