Reporters are not advanced to Active before Submit
Reporters should not be advanced to Active stage prior to Submit.
Currently the reporter in BeginImpl and BeginMain will be move to
Activate stage in case of ImplFinish and BeginMainAbort.
This might result in losing the previous reporter in Active stage which
might be needed for submit.

Bug: none
Change-Id: I90b7161f8daebb8884ce23b4fc4cc9d593dea09d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1993743
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Behdad Bakhshinategh <behdadb@chromium.org>
Auto-Submit: Behdad Bakhshinategh <behdadb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735629}