[TTS] Add a Feature for Translation in the Bar.
Adds a new Feature ContextualSearchTranslations that does inline
partial translation using CoCa to get a translation during the
Resolve request.

This adds a flag to chrome://flags and will send an integration
code to the server to allow Translate card responses.

The server is all set up for this already, except that it switches the
positions of the translation and the original text from what we're used
to on our team.

TODO: unit tests.

BUG=952401

Change-Id: Ia22b5bba4a6c98a3d8717decf99a20c9698a3038
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2018485
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Auto-Submit: Donn Denman <donnd@chromium.org>
Commit-Queue: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735639}