[Files ranking] Delete expensive UMA metric in UI.
In crrev.com/c/1911065 we introduced a ground-truth metric for zero
state files appearing in the UI, that was useful but costly to
compute. We should take this out before M80 rolls out to stable.

This CL deletes the function that computes the metric, to make the
change we need to merge back into the M80 branch as small as possible.
The rest of the infrastructure for the metric will be deleted in a
follow-up CL.

Bug: 1011221
Change-Id: I88a610a30347ade7d1a7fdfe3db341adb629b076
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2016813
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735670}