Remove flag for BLE authenticators in WebAuthn
The Chrome WebAuthn team has decided not to ship support for BLE
authenticators in Web Authentication due to reliability issues.

This CL removes the flag that allows it to be enabled. BLE
authenticators can still be used with Chrome running on versions of
Windows 10 that support the Hello API since that does not depend on this
flag or Chrome's implementation.

BUG=1046131

Change-Id: I2992fec02b95677d84a0b4fc65a5432a325e6c23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023671
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735677}