Use ERR_ABORTED when SB committed interstitial is not shown
With Committed interstitials, ERR_BLOCKED_BY_CLIENT is used instead of
ERR_ABORTED when cancelling a safe browsing flagged navigation. This
causes an issue when an interstitial is not shown, triggering a net
error page. This changes the code back to ERR_ABORTED when no
interstitial is shown. The plumbing to the URLLoaderThrottles was
also setting showed_interstitial to true in some cases where it should
be false. This CL also fixes that.

Bug: 1022477, 1045549
Test: This fixes a CTS test that was broken with committed interstitials
(android.webkit.cts.WebViewClientTest#testOnSafeBrowsingHit)

Change-Id: I8f10bea767749919e4a5ef6b460c9b96d6cbb75f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2008310
Commit-Queue: Carlos IL <carlosil@chromium.org>
Reviewed-by: Nathan Parker <nparker@chromium.org>
Reviewed-by: Tao Bai <michaelbai@chromium.org>
Reviewed-by: Tim Volodine <timvolodine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735690}