Set initial pointer shape on entering pointer lock.
When pointer lock is active on the client, the host composites the
mouse cursor into the desktop image. When pointer lock is deactivated,
it sets a blank cursor to disable compositing, but this resulted in
no compositing after pointer lock was reactivated until the cursor
shape changed.

This CL inserts ClientSession between the MouseShapePump and the
DesktopAndCursorComposer, allowing cursor shape changes that occur
while pointer lock is inactive to be saved so that compositing can
resume when pointer lock is reactived.

Bug: 1043325
Change-Id: I70edc0ae3d1a45595fe5a7d9298c4f02b921d696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023519
Auto-Submit: Jamie Walch <jamiewalch@chromium.org>
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Commit-Queue: Jamie Walch <jamiewalch@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735693}