Use mojom::ScrollIntoViewParams::Behavior type in Blink
This is a follow-up CL of https://crrev.com/c/1993076 and this CL
is a part of the effort to use the mojo type in Blink without
converting between Blink and Non-Blink.

This CL uses mojom::ScrollIntoViewParams::Behavior and clean up
ScrollBehavior enum in Blink.

Bug: 1039285
Change-Id: Id9a6a03c9ff64b8d36bc76511ad5ac784f660576
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2021567
Commit-Queue: Miyoung Shin <myid.shin@igalia.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735711}