Allow the IncreasedContrastThemeSupplier to Opt-Out of Most Incognito Colors
Previously, the IncreasedContrastThemeSupplier bypassed most of the
incognito colors. This worked fine as the colors it did specify worked
with non-incognito colors in incognito.

http://crrev.com/719713 removed this bypass, resulting in incognito
colors applied to the IncreasedContrastThemeSupplier and white-on-white
text as well as light-on-white controls.

This change reintroduces this opt-out as a workaround.

BUG=1045630

Change-Id: Ib83172fbd90559bc72e94e850591ee66cd4ad4e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024468
Commit-Queue: Robert Liao <robliao@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Auto-Submit: Robert Liao <robliao@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735723}