[Structured metrics] Add metric collection and upload logic.
This is the final CL for the core of the structured metrics system. It
ties together the KeyData class (for managing keys and computing hashes)
with the metrics provider, and implements logic to:

 1. Hash and store incoming structured metrics events.
 2. Provide stored metrics events for upload.

It also adds integration tests for the system, checking that an event
created through the public API is correctly prepared for upload.

Other misc changes:

 - Some minor refactoring of event_base.{h,cc} and equivalent changes
   to events_template.py, to fix some bad style.
 - Added a missing RemoveObserver call to the metrics provider dtor.

Bug: 1016655
Change-Id: I3daae2ba7927742d26db8d5750538584bf13b94f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1975386
Commit-Queue: Tony Yeoman <tby@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735748}