arc: net: print NetworkConfiguration service name
It is sometimes necessary to correlate Android logs with host logs for a
specific network service for debugging some connectivity issues. This is
difficult to do with anonymized logs from feedback reports which do not
have GUIDs or SSIDs.

This patch adds to ARC++'s net.mojom NetworkConfiguration objects the
DBUS path of the associated shill service. This will allow ARC++ to
print it in logcat and dumpsys without leaking any PII. The service path
is essentially a monotonically growing integer that does not persist
after reboot.

dumpsys wifi.

Bug: 147270970
Test: flashed ARC and Chrome on eve, manually checked logcat and
Change-Id: I61c1a7b3b5476975a72e31347b6af485f9e0961e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2021786
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Cr-Commit-Position: refs/heads/master@{#735750}