[settings] Remove unnecessary diceEnabled checks in people_page.html
Dice is now always enabled, these checks are unnecessary around
#edit-profile.

The sync-overview section can be completely removed, because it is only
shown when dice is disabled and signin is allowed, which is no longer a
possible state since dice was enabled by default.

The test suite DiceUITest can now be removed and merged with the general
test suite.

Bug: 891781
Change-Id: Ida75b66481d8a29ed983864aa5da24f0e1d835a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2019447
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735787}