[Captive Portal] Move opening of login tab out of CaptivePortalTabHelper
This CL starts the componentization of CaptivePortalTabHelper, itself a
prerequisite to the componentization of SSLErrorHandler for sharing
with WebLayer. In this CL, we move the functionality of opening a new
tab via interaction with //chrome's Browser object out of
CaptivePortalTabHelper into ChromeSecurityBlockingPageFactory. The
latter is a good landing place for this functionality as it in fact
already needed to invoke this functionality as part of the configuration
of CaptivePortalBlockingPage.

Bug: 1030692
Change-Id: I3fc447a1fe74c8cc7c8172f6bd8433e68d4c854f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2017488
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735791}