Use mojom::ScrollIntoViewParams::Type type in Blink
This is a follow-up CL of https://crrev.com/c/1993076 and this CL
is a part of the effort to use the mojo type in Blink without
converting between Blink and Non-Blink.

This CL uses mojom::ScrollIntoViewParams::Type and clean up
ScrollType enum in Blink.

Bug: 1039285
Change-Id: Ib0e6db68d371807892831ffabce5cd3d04d4be8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024628
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Miyoung Shin <myid.shin@igalia.com>
Cr-Commit-Position: refs/heads/master@{#735799}