Use mojom::blink::ScrollAlignment::Behavior type in Blink
This is a follow-up CL of https://crrev.com/c/1993076 and this CL
is a part of the effort to use the mojo type in Blink without
converting between Blink and Non-Blink.

This CL uses mojom::blink::ScrollAlignment::Behavior and clean up
ScrollAlignmentBehavior enum in Blink.

Bug: 1039285
Change-Id: I43743284c1ed33f5f720b4973dccfd72de4801ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023711
Commit-Queue: Miyoung Shin <myid.shin@igalia.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735801}