[Captive Portal] Componentize CaptivePortalTabHelper
This CL cleans CaptivePortalTabHelper of its remaining //chrome
dependency by passing in the CaptivePortalService from the embedder
rather than obtaining it internally via the Profile. It also
componentizes the class. Componentization of its unittest will be done
in a followup.

The motivation for this componentization is to enable the
componentization of SSLErrorHandler for sharing with WebLayer.

Bug: 1030692
Change-Id: Ibca11f2f3d603a411cfe03f00c26ca5be71ada33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2019102
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735808}