Add reason for hiding autofill popups
This CL should have no visible effect (yet).

By adding a reason whenever the autofill popup is asked to be hidden,
the controller can decide whether it should close the popup immediately.

This comes in handy when a popup is closed only to be reopened with
fresh suggestions (see linked bug). With the reason, the popup can be
kept open and directly updated which prevents flickering.

TBR=vasilii@chromium.org
TBR Reason: API change for autofill client.

Bug: 1043963
Change-Id: I77cb4b920114c37012b9383508d04fc4250422c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015055
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Christos Froussios <cfroussios@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735821}