Hang with UIA while editing certain text inputs
This change fixes a hang encountered when deleting text from an input
at the end of a document. Since AtEndOfAnchor checks for strict
equality with MaxTextOffset, a position where the offset is beyond
MaxTextOffset will return false for AtEndOfAnchor. This results in an
infinite loop in places such as CreateNextCharacterPosition, where
AtEndOfAnchor is used to break out of a loop.

I considered changing AtEndOfAnchor to be >= rather than strictly ==,
as well as snapping positions at the end of the document to be within
the document, but both of those approaches seemed to violate a
fundamental property of text positions - that a position is simply a
node and an offset, and if that node or that offset are invalid, the
position itself is no longer valid. See IsValid() for this property,
which was in fact false for the cases that hang.

So with the concept of position validity being a responsibility of the
owner of a position in mind, this change adds a new method for the
owner of a position to snap that position to MaxTextOffset if desired.
This method (SnapToMaxTextOffsetIfBeyond) is called upon the call of
any UIA TextRange method, to ensure that our text range doesn't
become invalid due to the underlying document changing. For performance
reasons (there is a call to MaxTextOffset), this is only called on the
last node in the document. Other nodes are far less catastrophic because
other methods will snap positions beyond MaxTextOffset to the following
node (which is why positions at the end of the document are not able
to do this).

Unit tests are added for all scenarios described and new methods added.

Bug: 1045761
Change-Id: I304a657bdddb7cf2f45e64148b4d82670ff85286
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2021844
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735887}