Autofill popup controller checks for updates before hiding
Currently, a AutofillPopupControllerImpl::Hide call will always hide the
AutofillPopupView it manages.

If the delegate that provides data for view is preparing to push updated
data to the controller, hiding might cause flickering or might close the
updated pop-up. Therefore, the controller now can enter a waiting state
in which it only proceeds to hide, if the reason is urgent (e.g. if the
tab is gone or the view destroyed) but not if the user might actually
wait for the data update.

Bug: 1043963
Change-Id: I6b766d2b92f55711a551f21954b40f02cfc21dfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2019025
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735895}