[iOS] Fix the identity disk position
This CL fixes the position of the identity disk (by fixing the position
of the invisible toolbar). The constraint was updated when the phone was
rotating, but it was too soon to take into account the safe area.

This CL removes this updates and instead relies on the update from
-updateConstraints as it is called by the CollectionViewController
when its view safe area is updated.

It also removes the implementation of -viewSafeAreaInsetsDidChange
as it is not reliably called.

Fixed: 1041831
Change-Id: I5a8c2681cf91f440ffd50b74dbb537143dcd4495
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023889
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735907}