[Audio Process][Sandbox] Add  libnss_compat to the whitelist.
After manual testing on a Ubuntu distribution (more precisely a XUbuntu
18.04) I was able to determine that libnss_compat is needed by ALSA to
communicate with libnss and be able to read the groups available on the
system.

This CL adds the affected library to the loaded list of libraries in the
audio sandbox hook for linux.

Bug: 1032949
Change-Id: I779158f6cf4eb48ef6b29469ee52181f0c69a70e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023893
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Armando Miraglia <armax@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735928}