[Reland#3] Make programmatic scrolls respect smooth scroll flag
Programmatic smooth scrolls should also be instant if the smooth scroll
flag is explicitly disabled.

Previously landed in
https://crrev.com/c/1536759
https://crrev.com/c/2009842
https://crrev.com/c/2022356

To reland, this CL force enables the scroll animator in Blink web
tests. Without this, the usual flow for determining if the animator is
enabled is followed, which leads to inconsistencies. On Mac, the
animator is enabled in Chrome but not in content shell. This leads to
inconsistencies with other platforms (Mac tests ran without smooth
scroll while Linux/Windows ran with) and ensuring we test what we ship.

This change required some minor fixes to tests that assumed instant
scroll on Mac (and uncovered one real bug).

Second reland fixes a few more tests on Mac. See diff from original
patchset for changes.

Third fix addresses flakiness in some of the Mac tests. Original
patchset has reverted CL.

TBR=flackr@chromium.org,nasko@chromium.org

Bug: 944583
Change-Id: I6b1ccb719b6bbed2b6d7495efa4d7846424796d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024163
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735931}