In Dark Mode, list item markers should be treated as text.
In Dark Mode, we use DarkModeFilter::ElementRole to distinguish between
text, foreground, and background objects being drawn. List item markers
are functionally playing the role of text, but the actual drawing commands
like DrawOval are assumed to be background by default.

Introduce a mechanism to allow ListMarkerPainter to override the
DarkModeFilter::ElementRole for the subsequent drawing operation. That
approach should be flexible enough that we could use it in other places
and doesn't require adding a bunch of additional parameters to
GraphicsContext APIs.

Bug: 1030457
Change-Id: Ic147feababeea5cd210896497fd49087a5bcc11f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1948449
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Aran Gilman <gilmanmh@google.com>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#735983}