[android] do not filter locales in resource packing
This CL makes it so that all locales are used during resource_packaging

It looks like previously resource_packaging was only called for apk
builds.  But now that it is also used for bundle builds the locales
for Android bundles should be included as well.

Before this change all language splits for the 27 + 9 new languages
recently added to Clank were about half the size as older languages in
monochrome_public_bundle.apks

After this change the new languages and old languages are about the same
size in monochrome_public_bundle.apks.

Example Before (af,as, az, be, bn, and bs are new languages):

   158921  2001-01-01 00:00   splits/base-af.apk
   424765  2001-01-01 00:00   splits/base-am.apk
   432527  2001-01-01 00:00   splits/base-ar.apk
   226815  2001-01-01 00:00   splits/base-as.apk
   165763  2001-01-01 00:00   splits/base-az.apk
   203683  2001-01-01 00:00   splits/base-be.apk
   473236  2001-01-01 00:00   splits/base-bg.apk
   234710  2001-01-01 00:00   splits/base-bn.apk
   163986  2001-01-01 00:00   splits/base-bs.apk

Example After:

   305787  2001-01-01 00:00   splits/base-af.apk
   424765  2001-01-01 00:00   splits/base-am.apk
   432527  2001-01-01 00:00   splits/base-ar.apk
   569512  2001-01-01 00:00   splits/base-as.apk
   325961  2001-01-01 00:00   splits/base-az.apk
   468522  2001-01-01 00:00   splits/base-be.apk
   473236  2001-01-01 00:00   splits/base-bg.apk
   577862  2001-01-01 00:00   splits/base-bn.apk
   324504  2001-01-01 00:00   splits/base-bs.apk



Bug: 1035665
Change-Id: I31d3bddba38148931212e6bccb63155376b82bda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024159
Reviewed-by: Megan Jablonski <megjablon@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Trevor  Perrier <perrier@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736004}