Add Windows high contrast support to the ink drop.
This change begins adding high contrast support to the ink drop. This
work is behind the EnablePlatformHighContrastInkDrop flag, which is
disabled by default.

The InkDropHighlight will now use 1.0 opacity (completely opaque) when
in platform high contrast mode. This will ensure that the ink drop color
is fully visible in high contrast.

To start, the ink drop color is set to the correct high contrast color
for the toolbar buttons. Future work will set the ink drop color
correctly in more areas, and fix element layering to ensure the
foreground is always visible on top of the ink drop background.

Bug: 776871
Change-Id: Iecdbedc7f4d735096e650d2b65a53d3de64643f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023362
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736050}