[Android] don't map zh-hk to zh-tw in bundle builds.
To give Clank language parity with Android 27 new locales are being
added.  26 of those locales have been added in two previous CLs:

   - https://crrev.com/c/1941058
   - https://crrev.com/c/1952288

zh-hk was skipped because currently resources for it are mapped to
zh-tw.  This CL skips that mapping for bundle builds where zh-hk is
now included. zh-hk is still mapped to zh-tw in apk builds.

There is an increase in size even when language splits are used in
bundles.  A bug has been created to look into this:
- APK Bundles wastes a lot of space in resources.arsc (400kb)
- https://crbug.com/1031645

Bug: 1034774
Binary-Size: Size increase is unavoidable (see above).
Change-Id: I4edb9d5213e8dd3443522deee5395e500af34371
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970258
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: (gone) Eric Stevenson <estevenson@chromium.org>
Commit-Queue: Trevor  Perrier <perrier@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736071}