Clean up in GpuDataManagerImplPrivate
Simplify GpuProcessStartAllowed(), currently it returns false on
Windows with GPU and swiftshader disabled, or on Chromecast with the
flag --disable-gpu. This matches exactly when |gpu_mode_| is
GpuMode::DISABLED.

The test GpuStartsWithGpuDisabled has a special case for Chromecast
audio-only builds, but it isn't run on Android Chromecast. Separate out
the Chromecast-only test.

Also fix a couple of typos.

Change-Id: I166bf1c7f482e7b97c31ed607949044f3ef7ac6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023949
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Sean Gilhuly <sgilhuly@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736101}