[omnibox] Relocate On-Focus config in fieldtrial_testing_config.json
Currently, the on-focus configuration in fieldtrial_testing_config.json
is in the "NTPZeroPrefixSuggestions" trial.

The problem is that about_flags.cc configures on-focus suggestions with
the "OmniboxBundledExperimentV1" field trial. This mismatch between
fieldtrial_testing_config.json and about_flags.cc causes a crash.

It's most proper to match what's in fieldtrial_testing_config.json
and about_flags.cc.

We won't be able to match the server-side config all the time, since
Dev/Canary uses the "unbundled" experiment format, and Beta and Stable
use the "bundled" experiment format.

But so long as fieldtrial_testing_config.json and about_flags.cc match,
the client-side code is correct.

Bug: 1015174
Change-Id: I93fd6cf6392f3de2a0e326bbdfe736bb9a10acc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2019329
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736116}