M80: Never show warning for sideloading at login screen
The warning should never show on M80 since the entry to turn on the
feature is not enabled (but the code path to show the warning is).
However, unexpected timeout in this code path triggers the warning.
As we're promoting M80 to Stable soon, this change simply force to
not display the warning.

Bug: 1046443
Test: build
Change-Id: Id53ef199653bed616f6416aa963733e6ccc55f94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2025970
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Victor Hsieh <victorhsieh@chromium.org>
Cr-Commit-Position: refs/branch-heads/3987@{#741}
Cr-Branched-From: c4e8da9871cc266be74481e212f3a5252972509d-refs/heads/master@{#722274}