Convert ScrollState's delta_granularity to enum
This is used internally as ScrollGranularity enum value, so convert from
being a double to the real type. Note that we're currently keeping the
double version for script bindings of Blink's ScrollState. There is a
behavioral change where the ScrollStateInit deltaGranularity is no
longer round-tripped correctly but the ScrollCustomization feature was
never shipped and no longer being pursued.

R=bokan@chromium.org

Bug: 1045306
Change-Id: I8092ebe8ee4833f2566b234660b8c4b5880e51f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026468
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#736212}