cc: Fix handling of mips for images on the CPU with OOPR.
If an image is greater than max texture size it can not be uploaded to
the GPU and we keep a copy on the CPU instead for rasterization. While
for GPU backed images adding mips requires re-allocating a new mipped
texture, for these images it should be a no-op since its handled by skia
internally. The existing code was erroneously trying to upload them.

R=cblume@chromium.org

Bug: 1046414
Change-Id: If88f76578dae99470991f7e4fb136107b94e23a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026227
Commit-Queue: Khushal <khushalsagar@chromium.org>
Auto-Submit: Khushal <khushalsagar@chromium.org>
Reviewed-by: Chris Blume <cblume@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736239}