Rename to |Offset/RectInContainerBlock|
|NGPaintFragment| has two offsets; one relative to its parent
and the other relative to the container block.
|NGFragmentItem| has only one, so it was named as |Offset()|,
but r736251 <crrev.com/c/2024133> revealed that this can be
confusing and mis-used, even by myself.

This patch adds |InContaininerBlock| suffixes to all functions
that return the offset relative to the container block.

This patch has no behavior changes.

Bug: 982194
Change-Id: I64f5ae133c11b4d21e455b7c13c2a8884273b6f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024190
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736265}