media/gpu/v4l2(s)vda: stop accessing visible size from wrong thread
GL or EGL image creation happens on the child thread, which should not
access any of the members that belong to the decoder thread. Yet this is
precisely what we were doing by reading the visible size of the buffer
to be imported, without any guarantee that the decoder thread will not
have changed it following a resolution change event.

Fix this by making the visible size a parameter of the bound method run
on the child thread. That way, it is read from the decoder thread and is
guaranteed to be consistent.

BUG=b:148421581
TEST=video.Seek.switch_vp8 passing on Kevin.
TEST=video.Seek.switch_vp8 passing on Hana.
TEST=video.Seek.switch_h264 passing on Krane.

Change-Id: Id8bc87ac50d6ebd2d74b2f6c25d42da83c933605
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026852
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736274}