[SyncSetup]: Allow querying sWAA independently from sync.
As part of friendly settings project, we communicate the sWAA in sync
setup page. This CL enables to query the sWAA bit independently from
sync or history sync status. This is needed for the 2 following
scenarios during which the |WebHistoryService| is null:
1- The advanced sync setup flow (click on settings from the turn syn on
confirmation dialog).
2- Switching sync history from 'Off' to 'On'. This change does not take
effect until we close the |syncSetup| settings page.

Bug: 1035421
Change-Id: Id1651bd0d9508643e6218c6cef0c6f310109bef8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2012950
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736295}