Limit interval search space after discarding an interval
When the timing update needs to process a large "jump" in time, after
for instance the page having been suspended for a significant amount of
time, it is possible to get into a state where a lot of new intervals
are generated (essentially all the intervals within the elapsed time
interval). New intervals in turn can generate additional, future
(relative to the current time in the elapsed time interval), intervals.
Intervals that are not yet started can be discarded and resolved again
when new instance times arrive. In this case we would start the interval
resolution from the earliest possible time, meaning that any such
(re)resolve would devolve into a linear search through all the previous
intervals. For animations with short cycles (and cyclic dependencies)
this could turn into a lot of work.

When we've resolved the/a first interval, and it's not the actual first
interval we're (re)resolving, we can use the end of the  previous
interval as the basis for computing the "begin after" ("start of
search") time passed to ResolveInterval(), since this should provide an
accurate lower bound for the new interval. This limits the search space
greatly thus reducing the amount of time spent resolving intervals in
the "catch-up" scenarios described above.

Bug: 1039243
Change-Id: Ibd5586336b56b0ccb3d61a92f1a4f09a7af5eaec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026977
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#736381}