UIA: Fix heading navigation hang caused by empty <div>
When using Narrator's heading navigation, there's an issue when the next
heading we want to read is followed by an empty <div> element. It hangs
for 30 to 45 seconds (on a debug build) trying to get to that heading.

This CL fixes the issue. It was caused by the replacement character
exposed to the text pattern provider to represent an empty object,
introduced in one of my previous change, CL:1937858.

I had to modify AbortMoveAtFormatBoundary in ax_position.h. The issue
was caused by the wrong AXBoundaryType returned by
GetFormatStartBoundaryType and GetFormatEndBoundaryType when in
presence of an empty object. When we are navigating by headings and we
encounter an empty object, the AXBoundaryType returned should be a
AXBoundaryType::kUnitBoundary and not a AXBoundaryType::kNone.
In other words, we want to break when the anchor is an empty object,
since the format is necessarily different.

Bug: 1046112
Change-Id: Ia83e31df5a242908186c03a2ea110a36bccb365b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023392
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#736402}