Create mapping in FeatureUtilities between feature flags and prefs.
The map contains the equivalence between a feature flag from
ChromeFeatureList and the constant in ChromePreferenceKeys with the
SharedPrefs key that caches that flag.

Use the equivalence to replace:
- #cacheFlag() with #cacheFeature() and inline
- #isFlagEnabled() with #isEnabled()
- #sFlags.put() with #setForTesting()

This way, the features are references by the ChromeFeatureList constant
rather than by its SharedPrefs key, which is an implementation detail.

Note there are tons of special cases in FeatureUtilities. This CL
addresses only the regular cases. The special cases will be handled in
future CLs.

Bug: 1012975
Change-Id: I0e02a52cbb841b551cf5920a13cb007c8aacecd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023199
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736408}