Add policy to disable Advanced Protection scanning
The Advanced Protection deep scanning feature will prompt users enrolled
in the Advanced Protection program to send their files to Google for
malware scanning. While the user needs to choose which files to upload,
admins may be concerned about the potential sharing of data. So this CL
adds a policy which can be used to prevent this prompt from appearing.

Bug: 1020418
Change-Id: I1f3bf1952aa9774bd503256c943179e6fdc81cbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024081
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736441}