Accessibility: Implement IsStructure() ax_role_properties method.
This change implements the IsStructure() method, which returns true
if a node's role is a subclass of the Structure abstract role.
Please see the aria spec for more details:

https://www.w3.org/TR/wai-aria-1.1/#structure

To keep the code clean and organized, this change also implements
a few helper methods for a couple other aria abstract roles, such as
Landmark and Select.

Change-Id: If8ff7d6d59f217af85c9b1bd5e853b00b4052e33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023194
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736461}