DevTools: Modified LayerTree.compositingReasons to include a new property named compositing_reason_ids.
This change modifies the LayerTree.compositingReasons method [1] to include an additional return property named compositing_reason_ids. This new property provides an ID that can be used by CDP clients (such as the DevTools) to localize the compositing reasons.

The design was reviewed here in [2].

A separate CL will follow for the DevTools frontend to consume compositing_reason_ids and map them to strings that are localized in the frontend.

[1] https://chromedevtools.github.io/devtools-protocol/tot/LayerTree#method-compositingReasons
[2] https://docs.google.com/document/d/1ZbCiXbVJZJaYqKZkQ_Y5RNpUqQd1JzO17PqFJZk2Ne8

Bug: 946860
Change-Id: Ie924a8dc65fe224e418e4d8b75c07433d2050645
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015551
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Mandy Chen <mandy.chen@microsoft.com>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Lorne Mitchell <lomitch@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#736492}