ui/base/clipboard: Make ClipboardFormatType a class.
It's possible this should have been a struct when it was created, but that time
is long gone. ClipboardFormatType is now fairly complex and can hold multiple
private members. Therefore, as per the Style Guide[1], this should now be a
class instead of a struct.

[1]: https://google.github.io/styleguide/cppguide.html#Structs_vs._Classes

Change-Id: I3eab2fc75d9660fbc5aa05322df7d12034d113fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2020477
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Jarryd Goodman <jarrydg@chromium.org>
Commit-Queue: Darwin Huang <huangdarwin@chromium.org>
Auto-Submit: Darwin Huang <huangdarwin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736544}