AppCache: replace HasSecurityState() with ChildProcessSecurityPolicyImpl::Handle().
HasSecurityState() was a temporary solution to deal with security
checks on processes that have been destroyed, but which still have
messages in flight.  ChildProcessSecurityPolicy security checks on
those messages should still be able to reference the old security
state.  Previously, HasSecurityState() was used to detect that the
process was gone and just fail open.  Now, we have a better solution
in place in the form of ChildProcessSecurityPolicyImpl::Handle, which
extends the lifetime of a process's security state even after the
RenderProcessHost has gone away.  Switch AppCache to use this approach
instead, keeping the security state around while the corresponding
AppCacheHost is alive.

Bug: 943887
Change-Id: I7ebe7e127dec1c301ed61d811f0298c38778fd31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026314
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736562}