DNR: Fix "Compile called before Add" error message
RE2 was printing an error whenever a ruleset with no regex rules was
initialized. It wasn't causing any problems, but it was bugging me. Fix
it by skipping the initialization of an empty RegexRulesMatcher, because
it wasn't actually doing anything anyway (other than print an error).

R=karandeepb@chromium.org

Bug: 974391
Change-Id: Ia381bca2b9d60ac2407fd9121657fccc4b05e81c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026592
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736586}