[omnibox] Conditionally keep pedal on triggering match
This CL partially revives the logic from crrev.com/c/1471251 :
[omnibox] Eliminate side-button Pedal suggestion mode

Note that only the match handling part is restored and the code for
handling the Pedals button is ignored. The is due to the conditional
nature of application of this logic. When the button row feature is
enabled, we will use the button row on the triggering suggestion to
show the pedal on the same suggestion. When disabled, the existing
dedicated Pedal suggestion logic will be kept.

For now, the button row work is in progress, and until the row is
actually presented, enabling both Pedals and the row will
effectively hide the pedal in the triggering suggestion.

Bug: 1046523
Change-Id: Iacb1e02158ee054b877dd0095f3ddd3a0b533b1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2027900
Commit-Queue: Orin Jaworski <orinj@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736625}