BitmapFetcher: Let BitmapFetcher handle data: URLs correctly
The omnibox uses BitmapFetcher to get Entity Suggestion images.

For on-focus suggestions, sometimes the images are included inline with
the Suggest results as data: URLs. This currently fails to render,
because BitmapFetcher cannot handle them.

This CL makes BitmapFetcher handle them and includes a test.

I implemented this functionality as part of BitmapFetcher instead of
further upstream, because there are multiple callers to BitmapFetcher
that may also run into this gotcha in the future, and it would be
unfortunate if they each had to write their own code to handle data:
URLs.

Bug: 1015174
Change-Id: Ibb68f0d0efdcaa8107d486dd88e3e3ccb50e2b84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2017962
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Rachel Blum <groby@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736635}