Run WebProtect Deep Scanning even when Safe Browsing is disabled
If deep scanning is enabled by policy, it should still occur even when
Safe Browsing is enabled. This CL ensures that will happen. It
implements the following flow for deep scanning:
- If SB is disabled, immediately begin deep scanning
- If SB is enabled, begin the consumer SB check
  - If this identifies malware, skip deep scanning and warn the user
  - Otherwise, deep scan the file

To accomplish this, we restructure the interaction between the
ChromeDownloadManagerDelegate and DownloadProtectionService so that
the DownloadProtectionService is now responsible for determining whether
scanning (either consumer or enterprise) should be performed.

Bug: 1041578, 1020215
Change-Id: Ia681b2cfee9cf2adecbba2aa62e83b530faaedeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2020408
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Nathan Parker <nparker@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736687}