weblayer: adds ability to save/restore the minimal state of a Browser
This is intended for scenarios where the full state is not necessary.
The minimal state attempts to save all the tabs and at least the
current navigation entries of each tab. This saves as much as will fit
in 512k. 512k is roughly half the max size allowed by most android
versions for saving application state. If application state is too big,
the OS throws an exception.

This code path will be used when full persistence is not desirable and
the system asks WebLayer to save state. Ideally when the system asks
WebLayer to save state and no id was given, we would use the same code
path as when an id was given (perhaps by generating one off ids and
saving that), but there is no clear way to know when Android no longer
needs the data, which means it's nearly impossible to automatically
delete the data. Trying to use heuristics will be error prone, for this
reason I'm adding the minimal code path.

I plan on moving session_service into weblayer/browser/persistence
after this.

BUG=1033924
TEST=component_unittests CommandStorageBackend*

Change-Id: Ie0cce1f4123a54d816d11f51a6ee6546310d3b02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2027129
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736694}