WebUI Tab Strip: Update drag image when dragging groups
This CL updates 2 UIs: the drag image that is used as the ghost image
following the finger during a drag session, and the placeholder of the
tab group as the user continues to drag.

Tab groups are resized to the size of a normal ungrouped tab to allow
for easier dragging, as groups can otherwise extend the entire view.
Tab groups are only draggable by their chip (color dot or title) to
prevent tabs from shifting beneathe the pointer as the tab group is
resized.

Bug: 1027373
Change-Id: Ib4692389390b8ffea8545760b8308c0d3d20000e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024124
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736722}