QuicTransport: Implement createSendStream()
Implement the SendStream interface and OutgoingStream mixin from the
WebTransport specification. Implement the QuicTransport
createSendStream() method to create a SendStream() object.

SendStream is trivial as all interfaces are inherited from
OutgoingStream.

A new interface, WebTransportCloseProxy acts a bridge between
QuicTransport and OutgoingStream. It abstracts away that a single
stream_id can be shared between an IncomingStream and OutgoingStream in
the case of a bidirectional stream.

Also add a browser test for the new API.

Bug: 1011392
Change-Id: I68731e606527d5dd94d54f9bfed05a22fde4d59e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024393
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736817}