ConvolverNode channel count and mode can be changed
The channelCount and channelCountMode for a convolver node is no
longer fixed.  The count can be 1 or 2 and the mode can be set to any
valid value other than "max".  This is basically the same constraints
as for a StereoPannerNode.

Also see https://github.com/WebAudio/web-audio-api/issues/2050
Feature: https://www.chromestatus.com/feature/6248507407073280

Bug: 1021552
Change-Id: I1cd814376d88ff4a525bae1f5e608ca4b989ac07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1988714
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736899}