Correct Japanese language keys for Mac.
The Apple Japanese keyboard has two language keys: kana to the right of
the spacebar, and eisu to the left. These use the USB codes for Lang1
and Lang2, respectively.

The uievents-code spec mentions this in the (non-normative) notes for
Lang1 and Lang2:
https://www.w3.org/TR/uievents-code/#table-key-code-alphanumeric-functional-2

This can also be seen in the USB-to-ADB mapping table used by the HID
driver source code published by Apple:
https://opensource.apple.com/source/IOHIDFamily/IOHIDFamily-1090.260.23/IOHIDFamily/Cosmo_USB2ADB.c.auto.html

Change-Id: Ib856fbba2d0916c6016625abe226d394edcb5fa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2029173
Reviewed-by: Wez <wez@chromium.org>
Commit-Queue: Erik Jensen <rkjnsn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736936}