[throughput] Handle off screen main damages
We could have off screen damages that is currently counted as
frame expected. For example, there is a main thread animation outside
of the current viewport, then for each begin main-frame, we
neither report is "no-damage", nor report "submit-frame". In
this case, we treat that main-frame as it caused no damage.

Bug: 1021963
Change-Id: I3d9c37e625b78d6cf275208d77676ffe17b47611
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1987603
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736964}