Fix a DCHECK of the first predicted event timestamp
The DCHECK on prediction metrics is to assert predicted event is later
than the real event. However, because we are using frame_time - 5ms as
the predicted time, it is possible that the predicted event is actually
prior to the first real event when the input frequency is high.
It's not likely to happen in real device as the input frequency is
120Hz, but it may happen on emulator.
Therefore, we should return instead of DCHECK this case.

Bug: 1040654
Change-Id: I61990cea0f7e7f7679e7f5b28db22b37a78865ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2031550
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Commit-Queue: Ella Ge <eirage@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736978}