Increase timeout of SimpleURLLoader tests.
Some of the tests started flakily timing out in December, with no change
in the SimpleURLLoader code.  Given that Chrome still seems to be
working and updating, and there's no obvious problem in the wait loop
of the code, my guess is the tests are just running slower than they
used to.  This CL is an attempt to test that theory.

Bug: 1035127, 1046745
Change-Id: I46bec60e001a21afd76ea66b92f864f8ffbddece
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2031544
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736985}