[Instant Start] Use tab ID to get JPEG thumbnail.
This is a precursor CL for https://crrev.com/c/1992341 which creates
the TabContentManager pre-native, and splits its creation into
creation + post-native initialization.

In this CL, we refactor TabContentManager::getTabThumbnailWithCallback
to use a tab ID instead of Tab which can't be created in pre-native.

Bug: 1041865
Change-Id: I8a0c82e40b5e0f05974a66bceb7cc96d2e4bcf5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2003433
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Commit-Queue: Xi Han <hanxi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736990}