[Translate] Add a histogram for the hrefTranslate hint
For Google navigations, the hrefTranslate hint may trigger a translation
automatically. Add a histogram to track if the hint is present on a
Google navigation, record whether or not automatic translation happened.
If automatic translation was already going to happen, then we use the
auto translate target language. Also record if this caused the target
language to be different than the hrefTranslate target language.

Bug: 900586
Change-Id: Idaf09ba10ad980a0cfdeca23e21a247548dcbeb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2029290
Commit-Queue: Megan Jablonski <megjablon@chromium.org>
Reviewed-by: anthonyvd <anthonyvd@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737005}