Fix synchronization of animation starts.
When there are composited animations in the process of being started,
we defer the start of main thread animations in order to synchronize
the start times. This process can lead to main thread animations getting
stranded in the pending state if composited animations are being
continuously generated. With this patch, main thread animations queued
up in a previous frame will no longer get blocked waiting to synchronize
with a fresh batch of composited animations.

Bug: 666710
Change-Id: I21eb1184bbf636230991e09691498a190ccf8f86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024476
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737015}