[Extensions] Enable content verification for new policy non-webstore installs
After this commit:
 * For non-webstore policy force-installed extensions, write
 computed_hashes.json on installation.
 * If (according to computed_hashes.json) extension's resource is
 detected corrupted, this extension will be disabled and  reinstalled.
 This applies only for policy force-installed extensions, and for
 webstore extensions it worked before this commit.

This commit will NOT change behavior for already-installed extension.
So, if some (non-webstore) extension was already corrupted, it will not
be detected. The reason for this is that there is no source of truth for
resources of already-installed non-webstore extension. It's possible to
reinstall all previously-installed non-webstore extensions, but it's out
of scope of this commit. See the bug for more details.

Bug: 958794
Change-Id: Ie9c014e7fd998f4eae6e852a59bd692d57ff7d8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2012964
Commit-Queue: Oleg Davydov <burunduk@chromium.org>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737034}