Android A11y: fix class name for static text.
On Android, the virtual views in web content mimic their corresponding
Android classes, and that includes returning the name of an Android
class name as a string.

This change fixes the class name for a static text node - it should
correspond to an android.widget.TextView. That will enable the Accessibility
Scanner to detect color contrast issues.

Bug: 1046651
Change-Id: Ia579b83dafa0ce814be059282c594136a37bf33d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026670
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737044}