Introduce ResolutionChangeDialog
This change introduces a system modal dialog that is displayed when a
user changes the resolution of an external display. The majority of the
logic is encapsulated in a more generic DisplayChangeDialog so that
behavior can be reused for a future RefreshRateChangeDialog.

Mocks: go/cros-displays-split
Test: ash_unittests
Change-Id: Iaca64bbabbe4f831a3578012aac148fee95f3066
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1979137
Commit-Queue: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737053}