Fix crash in handling AssistantWebViewImpl suppressed navigation.
It is potentially unsafe to destroy AssistantWebViewImpl during
handling of suppressed navigation events. This was not previously an
issue for Assistant as the Content Service backed implementation
notified handlers of this event over mojo, so destruction was performed
after the suppression sequence had completed.

Note that this is currently only affecting proactive suggestions as only
proactive suggestions currently uses target="_blank" for its links in
conjunction w/ the destruction sequence which triggers code path [1] as
opposed to code path [2].

[1] http://shortn/_H18kMl1jSt
[2] http://shortn/_J5ss5Vjirn

Bug: b:148484855
Change-Id: I9e42c371a3f53e16c6112bd4f7506ea2de749f2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026689
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: David Black <dmblack@google.com>
Cr-Commit-Position: refs/heads/master@{#737062}