viz: Handle empty swaps in GLOutputSurfaceBufferQueue.
This CL modifies GLOutputSurfaceBufferQueue to handle the case where
GetOverlayTextureId() and SwapBuffers() are called without first calling
BindFramebuffer(): this indicates that the client didn't draw but still
wants to swap. Prior to CL:1977969, on such case, GetOverlayTextureId()
would return the texture that was obtained in the last call to
BindFramebuffer().

This CL is intended to do the same: if GetOverlayTextureId() is called,
and we know that the client is not currently drawing, we still return
the last texture that was attached to the framebuffer.

Additionally, this CL ensures that we end access to a shared image prior
to emitting the command buffer command to swap the buffers to avoid
synchronization issues.

Android Q no longer produces flickering. Inspecting a page using
added a unit test for empty swaps.

Bug: 1046055
Test: Visiting https://www.funwithmama.com/bento-boxes-for-kids/ on
chrome: //inspect on Android Q no longer causes a DCHECK crash. Also
Change-Id: I86dc72f920a97b598028023f203aa5be4da2a840
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024177
Reviewed-by: Khushal <khushalsagar@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737067}