Re-land: Fire live region events when a node is removed.
Adds an AXLiveRegionTracker class to keep track of live regions
in an AXTree. Uses it to fix a bug where we weren't firing the
LIVE_REGION_CHANGED event on the live root when a node was removed,
only when a node was added or changed.

I'm going to follow this up with code that optionally computes
the text of a live region change, that we can use on Android,
Chrome OS, and some older versions of macOS. So AXLiveRegionTracker
is simple now, but it will be a convenient place to put that
logic.

Originally landed as https://crrev.com/c/1464325, but reverted
due to a UAF. The issue was that in LiveRegionTracker we were
keeping track of a mapping from node to live region root, and
when a node was deleted from the tree we were deleting it from
the key of that map, but we also needed to delete it from the
value. We do that now by temporarily keeping track of a set of
deleted node IDs.

Bug: 560599, 930763
Change-Id: I2005656fb50a208b00f744def828ce8db2134be1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1988742
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#737100}