[Extensions] Check for initiator origin in ExtensionNavigationThrottle.
Changes one of the checks in ExtensionNavigationThrottle to check if the
initiator origin of a navigation is empty, to more correctly handle
history.back() being used to navigate a window. Adds tests to cover this
case.

Also adds a test for a similar case which navigates a local frame, which
results in the navigation being blocked at the renderer level.

Bug: 1043965
Change-Id: I63e7e6775dbc56afdf3cd96452bf59939202370e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2021194
Commit-Queue: Tim Judkins <tjudkins@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737162}