Implement live-update and double-escape for month input popup
This change implements live-update and double-escape behavior for the
month input popup.  When arrow keys are used to change the
selection, the changes are immediately reflected on the in-page control
(and input events are fired).  When the user hits escape, if they have
made changes these are reverted without closing the popup.  A second
escape (or the first one, if no changes were made) closes the popup.

Behavior when clicking on a month is not changed; the popup is still
submitted immediately (as opposed to other types like datetime-local
where the user may want to go on to select other values after clicking
the date).

The selection/highlight model is changed so that using the arrow keys
moves the selection instead of the highlight, invalid dates cannot
receive selection, and the highlight is only used as a hover style.

Initialization of the control is changed such that if there is any
valid month, one will be selected, avoiding scenarios where we don't
paint a focus ring because there is no starting selected month.

Fixed a bug where months invalid due to the step attribute were not
greyed out.

The above changes required a lot of the same valid-date-range-seeking
functionality already implemented in CalendarPicker, so I removed the
side-effects from these functions and moved them into a mixin consumed
by both CalendarPicker and YearListView.

All of above behavior changes apply only to the standalone month picker;
the month selector embedded in date/week/datetime-local popups is
unchanged.  I'm not sure that live-update and double-escape makes as
much sense in those contexts, although we could consider a later
follow-up to change the selection behavior to match the new standalone
picker's behavior.

Bug: 1045195
Change-Id: I264a5285ec7df8bca4b0ad6389886b9580c12ae2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2029243
Reviewed-by: Ionel Popescu <iopopesc@microsoft.com>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#737203}