[LayoutNG] Associate all fragments with their layout box.
We'll now put *all* fragments generated from a node into a vector in
LayoutBox, instead of pointing to (at most) one fragment (was only
stored if the fragment was cacheable).

When a fragment is regenerated (after a relayout), we'll overwrite
entries in the vector. Each fragment now has a "sequence number", which
is stored in the break tokens. We need this to tell the fragments apart.
Block fragmentation may cause a node to generate multiple fragments, and
multicol balancing [1] and rebalancing [2] may go back and relayout and
overwrite fragments for the current column row (until everything fits
nicely). Then we need to know at which fragment indices to store the new
fragments (and nuke the old ones following, if any).

The sequence number will also be used for other purposes in an upcoming
CL, for example to cache the drawing of each fragment; see
ScopedDisplayItemFragment. At least we'll need this as long as we don't
create a unique DisplayItemClient for each fragment.

This is in preparation for NG fragment traversal (for painting and
hit-testing), which is needed by block fragmentation. This CL, however,
should not affect behavior in any way (but is filed as a separate CL
nevertheless, due to its size) - hence the unit tests. We already have a
bunch of block fragmentation tests in ng_column_layout_algorithm_test,
but I put the new tests in a new file, since they aren't really about
testing the column layout algorithm, but rather block fragmentation in
general (and the same can be said for many of the tests in
ng_column_layout_algorithm_test, BTW).

In an upcoming CL, PaintLayer will walk through these fragments and
paint them separately. The list of fragments will also be used to map to
FragmentData entries. In the future, we might also consider making
FragmentData an (optional) member of NGPhysicalFragment, since there is
a 1:1 relationship between NGPhysicalFragment and FragmentData.

[1] See NGColumnLayoutAlgorithm::CalculateBalancedColumnBlockSize()
[2] See the exit condition for the outer loop in LayoutRow().

Bug: 829028
Change-Id: I2e0b678c378327ea21d60a5c306845ce7b6e9e21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2030949
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737236}