[LayoutNG] Paint by traversing the fragment tree.
This introduces a feature LayoutNGFragmentTraversal, which will make us
perform layout read operations (painting, hit-testing) by traversing the
NG physical fragment tree instead of the LayoutObject tree.

This CL only adds basic support for painting, and only for nodes that
generate just one fragment. Painting is still driven from PaintLayer
(which is LayoutObject-centric -- we probably want to fix that in the
future), and the near-term plan there is to collect each physical
fragment and pass to NGBoxFragmentPainter. For now we just pick the one
and only fragment, and enter fragment traversal for painting, where
possible.

Note that this feature is still actively disabled inside multicol. We
need to get rid of the "artificial" paint layers for the flow thread and
multicol container (and still calculate FragmentData correctly) before
we can turn it on.

Added a virtual testsuite. The tests in fast/block/ will mainly verify
that the new code works, and the ones in fast/table/ will mainly verify
that we fall back to legacy painting correctly.

What remains (roughly) after this CL: Hit-testing, support for multiple
fragments, overflow calculation for fragments, and actual support for
block fragmentation (multicol).

Bug: 1043787
Change-Id: I1f61bb15458e7162ad61960c317e50891bd13e69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2002879
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737259}