[mathml] Treat unknown MathML elements as <mrow> elements.
The MathML CG decided that unknown elements in the MathML namespace with
a math display should behave like an <mrow> element [1] [2]. This CL
ensures that a MathMLRowElement is created for unknown MathML elements
in order to make that possible. It does not seem necessary to create a
specific MathMLUnknownElement C++ class at that point. Whether a
specific MathMLUnknownElement IDL is needed is still open [1][3].

[1] https://github.com/mathml-refresh/mathml/issues/139
[2] https://mathml-refresh.github.io/mathml-core/#new-display-math-value
[3] https://bugs.chromium.org/p/chromium/issues/detail?id=1021837

Bug: 6606
Change-Id: Ia493392ee8bfe9a7073e2ea6f1ce65927ec7fcae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007738
Commit-Queue: Frédéric Wang <fwang@igalia.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737263}