Apply no-var to ChromeVox (manually)
Apply
s/var /let /g

This is more dangerous than the no-var rule applied by Eslinter which tries to be conservative/safe.

For example,

try {
  let foo = 1;
}

if (foo == 1) {}

is an error since `let` scopes to local block while var scopes globally (or to local function). Eslinter would not replace in this case. Fix by declaring foo above the try.

The other variation of issues that cropped up was:

switch (...) {
  case 1: let foo = 1; break;
    case 2: let foo = 2; break;
}

causes an error because of the re-definition of `foo`. Explicitly scope them in a block (i.e. { let foo = 1; }).

Test: browser_tests pass.

Change-Id: Ib28758d6b723d43077410f60e112d21ecb526dde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2031529
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737349}