Fix identityDisc tapping in landscape mode.
The bug appeared after crrev.com/1986242 because identityDisk is now
placed on toolbar. Toolbar sits underneath the doodle and doodle simply
covers the toolbar and steals the gesture.

This CL swaps z-index for toolbar and doodle and disables user
interactions for toolbar in landscape. This way the user can still tap
on that toolbar in portrait and tap on doodle in landscape.

Bug: 1046303
Change-Id: I76d49f9c440c1f4cf6be8ee24513ec78848b2b26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2028741
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737379}