Keeps auto-hide shelf showing after opening the system tray
Due to the side effect from https://crrev.com/c/1988855, auto-hide shelf
is hidden after opening the system tray. Update on the shelf's
visibility state is triggered by the visibility change of system tray.
However, calculation on the auto-hide state depends on the state of
tray bubbles which are created/destructed after the visibility change.
This CL delays the update on the shelf's state when the visibility of
system tray changes. In addition, a test case is added for this
scenario.

Bug: 1046464
Change-Id: I72d48b5872ca9efc8160c9fd6bba6bd55addfa43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2028766
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737437}