[WebLayer] Add UrlBarController, createUrlBarView.
Create the UrlBarController interface and a simple version of the method
createUrlBarView() without UrlOptions.

Use it in WebLayerShellActivity to show a URL bar.

UrlBarController updates url text whenever the visible URL changes. This
means that the embedder doesn't have to listen to onVisibleUriChanged and
create a new URL Bar view.

Bug: 1025607, 1044571
Change-Id: I83bff6f82354ab970dbd84d3d24e5fac2cf66db7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2017408
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Tobias Sargeant <tobiasjs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737442}