[Signin] Dummy Reauth dialog
This CL adds an implementation of a dummy Reauth dialog to
SigninViewController. This dialog is needed only for developer needs
and should never be shown in production.

The dialog is hidden behind "SigninReauthPrompt" feature flag.

This CL also adds minor details to the comment of
SigninViewController::ShowReauthPrompt() method.

Bug: 1045515
Change-Id: I74b066157ddfa3d697f12e45511c9cbdceaf1c21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2030766
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737445}