[3P blocking] Fix tests accessing the ModuleDatabase on the UI thread
The ModuleDatabase was moved to its own sequence but both the
IncompatibleApplicationsBrowserTest and ThirdPartyBlockingBrowserTest
test fixtures were not updated.

This CL moves the accesses to the ModuleDatabase instance to the right
sequence.

Also added an override for the OutOfProcessQuarantine feature to allow
local testing. This was not an issue on the bots because they respect
the fieldtrial_testing_config.json file.

Bug: 850517
Change-Id: I2e8955b33cd7ffc75f74dd56b01df9dec96fc604
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2031742
Commit-Queue: Chris Hamilton <chrisha@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737448}