Hide DeviceSamlLoginAuthenticationType policy
Mark the DeviceSamlLoginAuthenticationType policy as "future", in order
to hide it from the policy documentation.

This policy is unused. It was originally intended for configuring the
Chrome OS smart card user login feature for all users on the given
device, but the design was later changed to support users of both types
(smart card and non smart card) on the same device, by determining the
case dynamically.

BYPASS_POLICY_COMPATIBILITY_CHECK=policy was never used

Bug: 826417
Change-Id: If76b55076d09d5c09e985a3730bf954c72a9f963
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2030898
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737467}