[ios] Prevents InfobarModal from being configured twice.
If an there's an existent modal it means one is currently presented, or in
the process of being presented. Configuring a new one would only cause
InfobarCoordinator to hold pointers to a Modal that won't be presented.

This CL prevents that by stopping any new configuration if a Modal already
exists.

It also removes modalNavigationController since we can get the
NavigationController from the modal.

Bug: 1040066
Change-Id: Ia2da22d67804e39780b19b78af2dcba58fa8a580
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2032565
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737468}