Relax request_initiator_site_lock checks for PDF plugin.
In some scenarios (e.g. http range requests for linearized PDFs) the PDF
plugin may need to initiate requests on behalf of the PDF's origin.
Such requests are currently proxied by the renderer process hosting the
plugin (i.e. the renderer hosting the PDF extension -
mhjfbmdgcfjbbpaeojofohoefgiehjai) and therefore go through
URLLoaderFactory that is locked via request_initiator_site_lock to a
different origin (i.e. the origin of the PDF extension, rather than the
origin of the website that hosts the PDF file).

Before this CL, such requests would be classified by
CorsURLLoaderFactory::IsSane and
VerifyRequestInitiatorLockWithPluginCheck as incorrect and possibly
malicious.  After this CL, the request_initiator checks are relaxed by
PluginServiceImpl::FindOrStartPpapiPluginProcess using a new mechanism
similar to AddCorbExceptionForPlugin exemption available for Flash.

We don't want to relax request_initiator_site_lock checks for renderer
processes hosting arbitrary origins from the web, so as a
defense-in-depth the CL adds a CHECK to FindOrStartPpapiPluginProcess
that verifies the origin lock of the renderer process that is granted
the extra capabilities.

Relevant tests:
- ChromeFindRequestManagerTest.FindInChunkedPDF (tests that range
  request do not result in a request initiator lock mismatch)
- PDFExtensionTest.EnsureInternalPluginDisabled (tests that the PDF
  plugin may only be embedded in trustworthy contexts)
- PDFExtensionTest.RedirectsFailInPlugin (the 'noRedirectsSucceed' testcase
  tests non-range requests initiated by the PDF plugin)

Bug: 1027173
Change-Id: Ib5111fd7a5d5b31727a9754c46c316e75def5143
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986923
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737503}