[CompositeAfterPaint] Update impl-side scrollbar geometries on contents size change
In CompositeAfterPaint, a scroll node is associated with the scroll
hit test layer (which is in scroll container geometry) instead of the
scrolling contents layer. This is because in CAP we don't always have
a single scrolling contents layer.

We need to update impl-side scrollbar geometries when the contents size
changes. In pre-CAP, that was triggered by the change of bounds of the
scrolling contents layer.

Bug: 931486
Change-Id: I5aec888180b3b43416e43a17b634d41f5da650c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024156
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737582}