[CompositeAfterPaint] Store cc scrollbar layer in ScrollbarDisplayItem
This simplifies and improves performance of scrollbar layer management
in PaintArtifactCompositor, and ensures that we will create a new
cc scrollbar layer for a new ScrollbarDisplayItem. Previously we failed
to re-create a scrollbar layer when a scrollbar was removed and added
again (scrollbars/reconstruct-coordinated-scrollbar.htm).

This CL will cause more re-creation of scrollbar layers, until we fix
over-paint-invalidation of scrollbars in composited scrolling
(crbug.com/1046544).

Bug: 931486
Change-Id: Ic0538d3063cab91877fd96749bceec871f6973bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024075
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737586}