blink/bindings: Report the size of ParkableStrings in ScriptResource
Parkable strings which are large enough to be parkable don't see their
memory footprint reported as part of the owning ScriptResource. This
shows as an underestimate in the web_cache category in memory-infra.

Report the size accurately for both cases, marking ParkableString
instances as either children of PartitionAlloc for the ones not tracked
by ParkableStringManager (short strings), or as children of
"parkable_strings" for the large ones.

Note for sheriffs: If there are metrics change in range: this is likely
responsible for an increase in reported "web_cache" memory footprint,
and a decrease in "parkable_strings" "effective_size". However it should
be neutral for parkable_strings *size*, and PartitionAlloc's size.

Bug: 1043679, 924164
Change-Id: I991b7d0064358808fd36008f239b0bed30a3cbea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033254
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737604}