Add use counter to feature policy to gauge web-compatibility of change.
A proposed change to the semantics of the Feature Policy header could
change the availability of features in some frames, depending on how the
Feature-Policy header and allow attribute have been specified. This CL
adds enough of the proposed logic to be able to tell whether the output
would change for any given call to IsFeatureEnabled, and counts how
often that situation occurs.

Bug: 937131
Change-Id: I2b587546f8d08a0c9d5e0c2bab60b77d771751c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2028630
Commit-Queue: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737617}