[system-web-apps] Add ability to grant requestable schemes in a WebUI
Currently chrome:// URLs are not allowed to embed chrome-untrusted://
URLs because we never grant request access to chrome-untrusted:// URLs.

To allow for this, this CL makes a couple of changes (Note that this is
only fixes the issue on the browser side. The renderer still has some
checks that stop us from loading chrome-untrusted:// iframes.):

1. Make the set of requestable schemes part of content::WebUI. This way
 clients of WebUI can control which additional schemes they want their
WebUI to be able to request. The function to add a scheme is added as a
webUI API because it is expected to be used by chrome/ WebUIControllers
that use chrome-untrusted://.

2. When creating a WebUI in RenderFrameHostImpl::CreateWebUI(), grant
request access to the schemes the WebUI specified.

3. Decouple granting access to chrome:// and file:// from
ChildProcessSecurityPolicy::GrantWebUIBindings(). Instead add this to
the list of schemes in content::WebUIImpl. Long term, we should be able
 to remove these schemes from content::WebUIImpl and instead have each
client specify the schemes they want access to, similarly to how
clients will call EnableMojoJsBindings()[1].

[1] https://cs.chromium.org/chromium/src/content/browser/process_internals/process_internals_ui.cc?l=53&rcl=0ab24b983b19f971f40b2f1e3f5a5b95c7a1ad9a

Bug: 1043548
Change-Id: I71d31f86e7c21da253df4f32634d945b45d071cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2009589
Commit-Queue: Jessica Huang <jessing@google.com>
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737655}