Make ServiceWorkerRegistry::GetOrCreateRegistration() private
In production code, this is called only from ServiceWorkerRegistry
now.

ServiceWorkerResourceStorageTest used the method but we can
replace it with a combination of CreateNewRegistration() and
CreateNewVersion(). This replacement also removes
NewRegistrationId() and NewVersionId() calls. This is a positive
side effect as these calls are going to be async (see issue
1046335).

Bug: 1039200, 1046335
Change-Id: Icef9951785b86b5321659ec4569da94ad593c2b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2032456
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737671}