Files app: Files list type icons for FilesNG
Resize the icons in the file list from 16x16 to 20x20 and overwrite all
icons for new style.

Check mark and folder icons use mask-image to follow the font color:
Folder has to be the same grey, instead of its black color and check
mark has to switch from blue to grey when it doesn't have the focus.

When the file list is in the multi-select mode (body.check-select)
background and image mask are reset so they aren't visible behind the
check mark icon.

Bug: 992821
Change-Id: I9314e4b0f93778e046c0fa746a55bfcb8e516340
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2029558
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737693}