Remove ServiceWorkerStorage::ScheduleDeleteAndStartOver()
There were three ScheduleDeleteAndStartOver() calls in
ServiceWorkerStorage which we couldn't simply move to
ServiceWorkerRegistry as these are callbacks of internal operations.

This CL replaces these ScheduleDeleteAndStartOver() with Disable() and
makes ServiceWorkerRegistry schedule a recovery process when the next
public operation is performed (the next operation will fail because
the storage is disabled, then the registry tries to recover).
This slightly changes the existing behavior but I don't think this
change has a real world impact given that the low error rates of storage.

Bug: 1039200
Change-Id: Iec112cf6690de4d68c3498282736a5f564924d39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2034385
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737697}