Block the XML parser on parser blocking stylesheets.
The html spec talks about script blocking stylesheets[1], but we do
script blocking in body by pausing the html parser. The APIs for doing
that for the XML parser had empty implementations which meant that
querying style from a script after an external stylesheet resource would
not wait for the stylesheet to load.

[1] https://html.spec.whatwg.org/multipage/semantics.html#interactions-of-styling-and-scripting

Bug: 1043944, 917116
Change-Id: I08422fa0608db8972154a8b19854906e154dfdd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2012021
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737721}