Migrate tests from "Mojo" bot to new "Network Service" bot.
"Mojo Linux" was really about running browser tests with the Network
Service in different configurations: in-process and web request proxy.
We don't want these to regress and have been watching the FYI bot, but
this isn't sustainable. We considered putting the tests on the CQ
but these tests are expensive so it would significantly increase CQ time.

Move the tests to a new waterfall bot.

This is step 2) and 3) from comment #5 on the bug:
2) Add an entry for the new builder to mb_config.pyl and to various
   files in buildbot/
3) Add an entry for the new builder to ci.star

(Doing both steps in the same CL seemed needed to pass presubmit.)

Bug: 1040384
Change-Id: Ic77927dcca52a556729e1d6b1174fc87c77ed656
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2029395
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737762}