Rename |getMainProfile| to |getLastUsedRegularProfile|.
|getMainProfile| function is renamed to have the similar name
in Java and C++, and to reduce the usage complexity of the function
on matching platforms. This change only contains Java side changes,
but there will be a following change for C++ side.
Note: This change is only code clean-up, does not change any behavior.

Bug: 1041781
Change-Id: I3d4aa6bfe4739c440ea3b7f40c2deb070561e865
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2027811
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Commit-Queue: Side YILMAZ <sideyilmaz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737773}