Animate the SAML smart card PIN dialog
Implement the "progress bar" animation in the smart card PIN dialog
shown during the SAML login (the <security-token-pin> Polymer element).

The animation is shown while the PIN is being verified by the smart card
middleware, and the PIN pad is hidden during that time. Once the
verification completes, the animation is stopped and either the dialog
is closed or the error message and the PIN pad are displayed.

Bug: 1016839
Test: start smart card SAML login, enter PIN, check that the progress bar is shown while the PIN is verified
Change-Id: I25e0c7412640f85221e0a73b3a64763c0f61ca18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2030411
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737794}