[WebLayer] Introduce SecurityBlockingPageFactory implementation
This CL introduces an implementation of the
//components/security_interstitials SecurityBlockingPageFactory
interface in //weblayer in preparation for //weblayer using the
component's SSLErrorHandler. In this CL we simply move the creation of
interstitial pages that is currently done in //weblayer's
ssl_error_handler.cc into the factory. We do not yet introduce
creation of interstitial pages that WebLayer does not currently create;
that will occur in a followup.

Bug: 1030692
Change-Id: I8fd776e29e502fab99c5b4966ed3ca5bebd2e5dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2028129
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737839}