Workaround PDF viewer brokenness when created in portal contents
A portal WebContents has TabHelpers even when embedded. When the helpers
are added, the portal WebContents is assigned a distinct tab ID. This is
an issue when attempting to load the PDF viewer as
ExtensionTabUtil::GetTabById cannot find the specified tab ID in the
TabStripModel.

Until we have a more thorough plan to deal with issue 1042323, we have
the PDF viewer load without a tab id when created in a portal.

Bug: 932691
Change-Id: Idbde43efcc1c0453abcea9b507cf01e9ef81f5f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033325
Reviewed-by: Lucas Gadani <lfg@chromium.org>
Reviewed-by: Zachary Kuznia <zork@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737854}