overview: Add metrics for enter/exit when all windows minimized in tablet.
This is a common case in tablet because users frequently press the home
button which minimizes all windows. It also is slightly different from
other overview animations - it animates the wallpaper blur and opacity
while fading in the mirror views and slides up the hotseat. Also, UX
wants to remove delay on wallpaper blur when exited, so these metrics
will catch a regression. A tast test will be added shortly.

Test: ash_unittests *Overview*
Bug: 1045185
Change-Id: I3ae72d876857415b11d1a3a5870b404b3e997660
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2018448
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737857}