Fix app list button on secondary display
Clicking app list button while app list is shown should hide the app
list. The click is handled by AppListPresenterDelegate first, and the
delegate dismisses the app list if the event is outside certain parts of
UI. The logic is supposed to be no-op if the event is targeting home
button, but it does not work for app list button on the secondary screen
(as it compares the display root window coordinates with screen
coordinates), so when clicking app list button on secondary display,
AppListPresenterDelegateImpl dismisses the app list before the event is
handled by home button, at which point the app list is shown again (in
peeking state).

This cl fixes the logic to use correct coordinates - additionally, uses
View::HitTestPoint to test whether the event should be handled by home
button (the home button uses custom event targeter, and HitTestPoint
takes this into account).

BUG=1047769

Change-Id: I18fd36bfff1583ab4cf1e78cd5732cb8525ef174
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033071
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737884}