macOS HDR: Add HDRCopierLayer
It used to be that setting the contents of a CALayer to an IOSurface
that has an HDR color space, and setting that the CALayer wants
extended range content was sufficient to display HDR content. This was
an undocumented behavior, and macOS has recently fixed it (made it not
work).

The documented way to get HDR is to use a CAMetalLayer. Unfortunately,
one cannot call setContents (or provide a pre-existing IOSurface) to
this API. As a consequence, we need to perform a copy from the
IOSurface to the CAMetalLayer. Add HDRCopierLayer subclass of
CAMetalLayer, which overrides setContents to blit the specified
contents to the CAMetalLayer.

Add tests for this. It appears that the test suite is not running,
because most of the tests are broken. Fix all the tests (next step
there is to add them to bots, but that will get a separate patch).

Bug: 976426
Change-Id: I8115f2ad19f9a9fd3d1018d2bde7132f7bcfc6e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033906
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737885}