Combine filter list rules which only differ by domain during generation
Currently, whitelist rules in the easylist are not always deduplicated
by the rule. It is possible for there to be two separate lines with the
same rule but different domains.

This CL adds an awk command during filter list generation which combines
these lines.

This shrinks the filter list size by 1.7% and reduces the number of
overall rules by 4.7% on the current list.

Bug: 1048224
Change-Id: Ib3e71e9b371605380cfc5600fe8a92f5898f5b6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033987
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737886}