Fix browser frame view not getting a relayout after a state change
views::NonClientView has 2 things: a views::NonClientFrameView and a
views::ClientView. We were previously only invalidating the layout on
the ClientView after a state change.  This was causing the browser
frame to paint as if it were still maximized after restoring from
a maximized state on Linux.  Invalidating the layout of the frame view
fixes the issue.

BUG=1046122
R=sky

Change-Id: I7da525efe1f436564ffffb410afe294e901e5d89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033759
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737890}