cast: webview: Use WebContents window size as the EXO surface size
Simplify the resizing logic. We're only supposed to be resized by the
embedder, so only resize when we're told to.

This way we no longer need to worry about the case when a WebContents is
embedded inside another WebContents as there's no size updates on the main
one.

Bug: b/147376354 b/148398442
Test: Verified resizes are applied to the EXO surface.
Change-Id: I70d2578bfe91018680b052af3dccf698c5303c72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2032201
Reviewed-by: Albert Chaulk <achaulk@chromium.org>
Commit-Queue: Daniel Nicoara <dnicoara@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737920}