[WebLayer] Use //components/security_interstitials' SSLErrorHandler
This CL changes WebLayer to use the //components/security_interstitials
SSLErrorHandler that is used by //chrome rather than the custom fork
that it has been using to date. As part of this change,
WebLayerSecurityBlockingPageFactory is augmented to create the blocking
page types that were not necessary before (as they were never created
in the fork).

With this change, WebLayer now shares the core logic for determining
which interstitial to show on SSL errors with Chrome.

Bug: 1030692
Change-Id: Ibd1dadb9d03937cf578fd785ae4b32c943283749
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2028068
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737925}