A11y:Refactor AXTree::PopulateOrderedSetItems
Refactored AXTree::PopulateOrderedSetItems, which is used for computing
set size and position in set, to make it more straight forward.
I made no change in the actual algorithm.

This change is needed to prepare for various tasks for exposing PosInSet
and SetSize for various elements

Change-Id: I2d4883bd5e3132ebbed721abb547bc24f5be0f10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2020544
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Victor Fei <vicfei@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#737931}