[NativeFS] Fix issue when a path changes from being a file to a directory.
(or vice versa.) We don't want to re-use the existing permission grant
especially in the file->directory case to avoid giving a website access
to an entire directory while we only told the user about a single file.
However in our logic to guard against this we left the old permission grant
dangling, which would result in a DCHECK when it was eventually destroyed.

This CL changes the logic in these cases to deny the old permission grants,
and updates the DCHECK to make sure it only triggers for non-DENIED grants.

Bug: 1047588
Change-Id: I8275d7b3b66ad9b687ca16e4129dba2323f5645e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2032236
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737959}