Improve trybot ForTest check to also track inlined symbols
android-binary-size's 'ForTest' check which looks for symbols in the
production apk with 'ForTest' in the name is currently lacking in that
it does not track inlined symbols. This change fixes that by having the
script parse the proguard mapping file and use that instead of using
just the final list of symbols.

Bug: 1012373
Change-Id: Ic2752ddc792150f426fb6beac8c9d6553799f6c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1993623
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737968}