Checks if model exists before validating the prediction model.
This change adds an earlier check for whether a model is included
in the PredictionModel proto before validating it. This should
resolve the prediction model validation histogram being recording
as invalid when the model is not provided by the remote
optimization guide as there is not a newer one.

Bug: 1046848
Change-Id: I0a974b45883f5940a1a63f77876c615ae420d8dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2035050
Auto-Submit: Michael Crouse <mcrouse@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Commit-Queue: Michael Crouse <mcrouse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737970}