[TTS][Overlay][Dark] Correct mdpi top_round asset.
Fixes the asset for the rounded corners on the top of the Bar of
the Overlay panel when in Dark Mode on an mdpi device.

The wrong asset was originally uploaded -- white instead of dark.

This asset was crushed with optimize-png-files.sh but still ended up
substantially larger maybe due to the non-white color that's dominant.

BUG=1000792

Change-Id: I60faa1f33c019d2a9f473444a5be11d3eb09fae2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036732
Auto-Submit: Donn Denman <donnd@chromium.org>
Commit-Queue: Theresa  <twellington@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737986}