Improve relative mouse support on Windows.
The DesktopCapturer implementation used by ClientSession on Windows
uses read-only shared memory, and so the mouse cursor composition needs
to be done by the DesktopSessionAgent instead. This CL refactors that
logic so that it can be used in both places, cleans up some raw pointer
ugliness, and adds support for compositing the mouse cursor on Windows
when pointer lock is active.

Bug: 1047908
Change-Id: Ic21a750379a8f144227d662cce200cdc95ce3967
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2034104
Commit-Queue: Jamie Walch <jamiewalch@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738003}