plugin_vm: allow uninstallation when policy is disabled
If the policy is disabled, the user still needs a way to cleanup the vm
from their disk. This change allows them access to the plugin vm subpage
as long as either the policy is enabled, or there is a vm on disk.

Note: It is not enough to simply delete the vm for the user, as the
policy change could be accidental or temporary.

Bug: 1023256
Change-Id: Id1cfae86061c8f7e1a0b771165833e412009edbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026511
Commit-Queue: Julian Watson <juwa@google.com>
Auto-Submit: Julian Watson <juwa@google.com>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Timothy Loh <timloh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738029}