Skip clipboard CSS sanitizer when there's no style sheets to sanitize
The clipboard CSS sanitizer has a side effect that, it may add
additional inline styles to elements to match their computed style in
the dummy document.

This isn't necessary when the markup doesn't have any style sheet to
sanitize, in which case the side effects are undesirable and causes
some regressions. Hence, this patch skips the sanitization when there
are no style sheets.

Bug: 1047806
Change-Id: I62e20fb5fe29d56bda9d86ca12499b7ffb1497e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036731
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738083}