Prepare breakpoint web test for source map asyncification (1/2)
This CL removes a race in a web test where breakpoints get removed and
then the current breakpoint information is dumped. Instead of racing,
we properly wait for the debugger to confirm that the breakpoint got
removed before dumping the information.

Drive-by: Previously {removeBreakpoint} was also marked as async.
This is not needed and this CL removes async/await for
{removeBreakpoint}.

R=petermarshall@chromium.org

Bug: chromium:1032016
Change-Id: I836387085b4b3b8b85b2d035e8031efd4b54e0b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033074
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738102}