Clean up unsized-media feature policy
Since the 'intrinsicsize' feature was removed [1], there's no reason to
have an IntSize (that is always set to either 0x0 or 300x150) in the
relevant elements. Remove the IntSize and use the various bools to deduce
the size instead. Also fold and remove the IsMediaElement(...) helper
since it's easier to just use knowledge at the callsites (in the various
element constructors).

[1] d1a6a7200ef9195b6cd701758eddb7a660790942
    https://chromium-review.googlesource.com/c/chromium/src/+/1782660

Change-Id: Ie10bcd3725ab9993df274f5ed87f950284bc0609
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2029516
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#738116}