Disable ctrl+alt+h shortcut after the enrollment started
The moment the enrollment process started, switching to hangouts meet
or back is disallowed, to avoid getting wrong policies. In order to
enroll the device in hangouts for meet, the ctrl+alt+h shortcut should
be pressed after ctrl+alt+e in the enrollment process.

BUG=chromium:1040523
TEST=Manual tests on the device

Change-Id: I4a5f589ab278bf2c6f47c8640d833b02e785735f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2014406
Commit-Queue: Igor <igorcov@chromium.org>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738140}