Reject invalid UTF-8 when deserializing strings from DOM objects.
These are guaranteed to be valid, non-null strings. Getting a null
string here indicates that UTF-8 decode failed, in which case
propagating a null string is not the right thing to do.

Test included. This test would trigger the DCHECK in the linked bug
without this fix.

Bug: 1047753
Change-Id: I9e31d50176563d339b07883bbc2e5b52c4ff7641
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033568
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738200}