Should not create GpuDataManagerImpl in the IO thread in the content tests
Some content unit tests create GpuDataManagerImpl in the IO thread.
PeakGpuMemoryTrackerImpl needs GpuDataManagerImpl and GpuDataManagerImpl
is created it when PeakGpuMemoryTrackerImpl is running on the IO thread.

This will cause a crash (sequence checker) if the DisplayChange observer is
added. This CL makes sure GpuDataManagerImpl is created first in the test
setup so GpuDataManagerImpl will run on the main thread later.

Bug:1042989

Change-Id: I825937f800dfeeddf2d25b03e5e50bf418e6942e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2019882
Commit-Queue: Maggie Chen <magchen@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738217}