Migrate NTP from URI to GURL
Replaces usages of java.net.URI with GURL.

This change is slightly complicated as GURL and URI handle about: URLs
differently, in that GURL considers everything after the about: scheme
as a path instead of a host/path/etc.

This is usually handled in the native code by running FixupUrl on URLs
with the about: scheme to convert them into the chrome: scheme (unless
they're about:blank or about:srcdoc), and this is how navigation to
about:newtab works (by converting it to chrome://newtab).

So in order to know if something will resolve to the NTP, we need to
first fixup the URL, and then check its host/validity.

Bug: 783819
Change-Id: Idf6ad46fa6ef9793a969f3a8ee721c9bb05b97a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2031710
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738273}