tracing: Attempt to make startup file writing more reliable in tests
Some of our startup tracing tests timeout flakily, and it seems that the
flakiness stems from the file not being written quickly enough on the
background task runner.

This patch overrides the task priority for the task runner for these
tests to USER_BLOCKING, which should make file writing more likely to
finish in time.

Also reenables one of the tests on windows, where it was previously
timing out flakily.

Bug: 1046447
Change-Id: Ibbe2ef61f81567a364f4d3396162e72be22b4457
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2032109
Reviewed-by: oysteine <oysteine@chromium.org>
Commit-Queue: Eric Seckler <eseckler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738293}