[ios] Translate Messages EGTest
This creates an EG2 test suite for Translate for the new Messages UI.
Many tests are either direct copies or similar to those found in
TranslateTestCase. Some of them were also not moved over, because they
do not apply to the new Messages design.

Bug: 1014959
Change-Id: I4df264e6dcfabcdc7ad3933aeb43f2bb255c3d0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036371
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738305}