Improve GetRegistryLength() performance by reverse lookup order
The make_dafsa script is extended with a switch that when applied will
reverse the order of all input strings, so that end characters appears
first in the generated graphs. Together with a new lookup function host
names can be looked up from right to left to find the longest suffix
matching entire components separated by dots.

GetRegistryLengthImpl() is modified to use the new lookup function,
replacing the loop with repeated lookups by a single lookup.

Tests on a larger set of host names from an external source showed a
reduction in the total number of lookups by 50% and total time spent in
GetRegistryLengthImpl() was down by 40% in average. All hosts used in
the tests were valid and resulted in positive registry lengths. Similar
performance gain was obtained also after the set of host names was
modified to produce negative lookup results.

BUG=678334

Change-Id: Ie6da70ef6ee33194da2988078e78c5eca6495453
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033512
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Commit-Queue: Olle Liljenzin <ollel@opera.com>
Cr-Commit-Position: refs/heads/master@{#738313}