Convert supervised user auth token fetch to unconsented primary account
SplitSettingsSync will allow the user to opt-out of browser sync.
However, IdentityAccessor::GetPrimaryAccountInfo() assumes the user has
consented to browser sync. The supervised user feature isn't coupled
to browser sync consent.

Switch to using the "unconsented" primary account. On Chrome OS this
account always exists for the logged-in user account, whether or not
the user consented to browser sync. On Android, it is always equivalent
to the primary account.

See go/cros-sync-mock and go/cros-primary-account for details.

Bug: 1042400
Test: updated existing unit tests
Test: add user and signin works with both parent and child accounts
  webui for parent controls works, webui for child account setup
  works

Change-Id: Ic45dd82bb3d5e3cbd53d32bc5b68abec677ba636
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026176
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738328}