Enable cursor compositing for uncurtained Mac.
I previously thought that the cursor was always part of the captured
screen on Mac, but it turns out that's only true if curtain mode is
active. This CL makes creating of a composing capturer conditional
on the curtain mode setting, which should give the correct beheviour
for both remote support and uncurtained remote access.

Bug: 1048774
Change-Id: I28cc01b0b615794e06cd31308c10a40af550e0e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036546
Commit-Queue: Jamie Walch <jamiewalch@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738347}