Fix ReadyToCommit->DidCommit VS request_initiator_site_lock race.
This CL makes sure that a correct request_initiator_site_lock is used
when a URLLoaderFactory is pushed to a renderer process
during a ReadyToCommit->DidCommit transition window.  In particular,
during that window it is incorrect to use |last_committed_origin_|
(see the bug for details).

Bug: 1047436
Change-Id: If696a326f162585f42ab0a0ecbe8f8f42f07f5a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026168
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738362}