Wrap |kAutofillAssistantDirectActionsName| in the right conditional.
The |kAutofillAssistantDirectActionsName| constant is only declared for
OS_ANDROID in flag_descriptions.h, but is defined for all platforms in
flag_descriptions.cc. This did not trigger a compiler warning because
the definition has an erroneous extern keyword. Remove the extern
keyword and move the constant to the right spot. Do the same for
|kAutofillAssistantDirectActionsDescription|.

Change-Id: I0a8b174a2a27a360a97d0ffa7f5e69c110d2f970
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2025248
Reviewed-by: Stephane Zermatten <szermatt@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738367}