Rename FeatureUtilities to CachedFeatureFlags.
FeatureUtilities is not accurate anymore, as this shouldn't contain
code for specific features anymore, and it is (and has been) a
stateful class.

The responsibility of this class is to cache feature flags for use
without native being loaded.

This is CL 3/5 for renaming FeatureUtilities.

Bug: 1012975
Change-Id: I8b4c72a65361379ec34272bdff55edf416d35b55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2028111
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738380}