Fix issue getting hypertext for omnibox
When calling AXPlatformNodeBase::FindTextBoundary, it would attempt to
get the hypertext of the node, but this was returning an empty string
for UI controls causing ui::FindAccessibleTextBoundary to fail.

This CL merges AXPlatformNodeWin::TextForIAccessibleText into
AXPlatformNodeWin::GetHypertext so that it can handle UI controls.

Bug: 1047906
Change-Id: I84f0f5d1241e9b634b481dbbc49fdefc7c706d70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036363
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Adam Ettenberger <adettenb@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#738386}