Fix a crash in pop-up tab strip
In the popup TabStrip used by Duet-TabStrip integration, there is a
crash in RecycleView when user very intensively clicks on the TabStrip
favicon to close multiple tabs. The stack indicates that the
recyclerView is trying to remove the same ViewHolder in two consecutive
remove operations, and the second call would crash. The fix introduced
in this CL is to early end the animation if we know for sure that this
item is going to be removed. This should help with better identifying
the ViewHolder to remove.

This fix is inspired by the discussion here:
https://b.corp.google.com/issues/34184109
and the CL here:
https://critique.corp.google.com/#review/188055979/depot/google3/java/com/google/android/apps/play/music/app/ui/common/MediaListRecyclerFragment.java
Even though we early end the animation, there should be no visible
animation change from the UI perspective.

Bug: 1045944
Change-Id: I0d8a7b098e17cdc86c4ae3227cd9dc1ba37b2adc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036874
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738392}