[OnionSoup] Move resource_load_info.mojom into Blink
The current mojom struct ResourceLoadInfo defined in //content is used
to notify the status of a resource loading, which is unavailable
inside Blink.

Along with the implementation of OnionSoup of loaders, we are going to
notify the loading status from Blink directly. So we have to move this
mojom struct into Blink in order to leverage it there.

This CL makes the changes below:
- Remove the ResourceType in //content and its typemap and replace with
  the mojom one.
- Move the resource_load_info.mojom into Blink.

Change-Id: I11ab3b7107bd4e06ee3ec737f825202c533a466c
Bug: 860403
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986640
Commit-Queue: Minggang Wang <minggang.wang@intel.com>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738417}