[throughput] Add new story for on-screen no damage animations
This CL adds a few perf tests for the on-screen no damage cases.
To be more specific, we create animations that causes no damage, for
example animation from translateX(0) to translateX(0).

At this moment, the on-screen no damage cases should not report
any histogram, because the frames_expected would be 0.

Bug: 1048163
Change-Id: Id315e98b32d5a9e55b85a96d1ed2d5d59003a983
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2035858
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738418}