viz: Remove loop in RemoveOverdrawQuads
Originally RemoveOverdrawQuads looped trough all the renderpasses
and did some computation for the non-root ones, and then executed
occlusion logic on the root one.

A CL that removed the computation for non-root renderpasses landed,
making the loop always skip his body for all the renderpasses
but the root one.

This CL removes the loop and executes the occlusion logic only for
the root renderpass.

Test: RemoveOverdrawQuadPerfTest.*

Change-Id: Ie92deae2b5fb65e1f5be605f79bc47bfc9de36bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033368
Reviewed-by: Sasha McIntosh <sashamcintosh@chromium.org>
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738425}