Stop using WebURLRequest in the subresource loading path
We're planning to introduce a non-copyable request body to
blink::ResourceRequest and blink::WebURLRequest in order to support
streaming upload.

WebURLLoader::LoadAsynchronously and WebURLLoader::LoadSynchronously
take a const reference to WebURLRequest, but we'll need to modify the
request when a non-copyable request body is introduced.

Here we stop using WebURLRequest and pass
std::unique_ptr<network::ResourceRequest> instead. This is better than
using a mutable reference to WebURLRequest because it is hard to
distinguish whether an argument is const reference or mutable reference.
Moreover, WebURLRequest is planned to be deprecated and we are planning
to use network::ResourceRequest or network::mojom::URLRequest for the
"request" struct for IPC. This change is also a step of that larger
refactoring effort.

third_party/blink/PRESUBMIT.py is updated to reflect the discussion at
https://groups.google.com/a/chromium.org/forum/#!topic/platform-architecture-dev/ntSzerId4gw .

Bug: 860403, 688906
Change-Id: Ieb42432e30495d03f8cfb924a41a1327070c7aa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2032688
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Yoichi Osato <yoichio@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738470}