Files app: File list FilesNG
Table header move padding to the label so clicking in the padding area
activates the click event in the label to sort the column.

Change the cursor to pointer in the column header to indicate that it
has an action (sorting the column).

Reset the table header font-weight because it's already using Robot
Medium which is already the right weight.

Fix the border radius for the file row thumbnail so it shows perfectly
rounded.

Reset the padding top in the file row because in the current spec the
text doesn't need the additional padding to align with the icon [1].

[1] -  https://cs.chromium.org/chromium/src/ui/file_manager/file_manager/foreground/css/common.css?l=343&rcl=80e9555e489ff4a6743f2d8f99dafdbaccc7397f

Bug: 992821
Change-Id: I06b1b6e5c842162f6b1a3ea027932869704995a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033059
Commit-Queue: Noel Gordon <noel@chromium.org>
Auto-Submit: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738513}