[Android] Pass fully initialized view to accessory adapter or exit
As described in the linked bug, the adapter assumed WindowAndroid to
never be null which causes a crash.
With this CL, the view will be initialized early on and if WindowAndroid
is nullptr, the creation of the view (and the adapter) will be aborted.

To do that, the unused constructor parameters for label width and length
were removed.

Bug: 1048849, 1049090
Change-Id: I60151a560907a984ba4e45ee949624400b8f86da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2037672
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738564}