Add basic structure for secure DNS setting.
This change introduces a toggle and radio buttons that are shown when
the toggle is enabled. Future cls will introduce a dropdown menu and
text field alongside the second and third radio buttons, respectively.
The visibility of the setting is controlled by a new param, ShowUi,
associated with the DnsOverHttps feature. The new param currently
defaults to false.

Change-Id: I9d9322cbdf64f9d66cd744de01829300446b8896
Bug: 1040145
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1980480
Commit-Queue: Katharine Daly <dalyk@google.com>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738588}