Convert AppDataMigratorTest.FileSystemMigration to RunLoop
This test has been disabled due to 747589, and so it was using
deprecated RunLoop constructs that DCHECK when running.

For now, this test is still deprecated because it might be flaky.
I will followup and turn it back on separately in case it gets
reverted again.

I am trying to get this test running again to test changes in:
https://chromium-review.googlesource.com/c/chromium/src/+/2020533

Most of the changes are fairly straightforward to use RunLoop
Quit() as callbacks to quit running instead of RunUntilIdle.

The tricky part here is that this test now crashes because QuotaManager
kicks off a task in NotifyStorageModified that runs after the test has
been torn down:

[204729:204736:0204/170817.027368:1900997872433:ERROR:database.cc(1584)] Quota sqlite error 1032, errno 0: attempt to write a readonly database, sql: INSERT INTO OriginInfoTable (used_count, last_access_time, origin, type, last_modified_time) VALUES (?, ?, ?, ?, ?)
[204729:204736:0204/170817.027526:1900997872589:FATAL:database.cc(1597)] attempt to write a readonly database

To avoid this, a new function to turn off the QuotaManager database is
added for testing.

Bug: 747589, 1015214
Change-Id: I98d07ffc62e0ff8047da3e994fbf40728f0972bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2038133
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Jarryd Goodman <jarrydg@chromium.org>
Auto-Submit: enne <enne@chromium.org>
Commit-Queue: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738607}