Add Omnibox Suggestions histograms to monitor performance.
This change adds metrics to measure:
- Time spent creating individual suggestions
- Time spent measuring and positioning suggestions
- Efficiency of suggestions re-use mechanism.

To enable the code to perform view re-use and create time the CL
extracts two methods from the ModelListAdapter#getView():
- tryConvertView() - attempts to re-use supplied view for a new
  suggestion, and
- createView() that constructs a new view of a specified type.

We do not yet know what range of values to expect, so we assume 10ms
to be the upper limit. This will be revisited.

We do not yet know how well the Debug.threadCpuTimeNanos will work as
it *might* return -1. this will also be a benchmark for us regarding
reliability of this method. The alternative
SystemClock.currentThreadTimeMillis() has much worse resolution.

We hope to narrow this time down to - say - 5ms with 50us resolution
if that proves to be reasonable.

See https://crrev.com/c/2007943 for more context.

Bug: 982818
Change-Id: I1cc83b8f5d45ae06e5ee6aa58d2cffa78d5b7ab2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023195
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Commit-Queue: Ender <ender@google.com>
Cr-Commit-Position: refs/heads/master@{#738615}