Refactors PolicyConversions to remove dependencies on //content.
Creates a PolicyConversionsClient abstract class with a concrete
ChromePolicyConversionsClient subclass.  The code remaining in PolicyConversions
and PolicyConversionsClient does not have any dependencies on //content or
//chrome and can be moved to //components/policy in a future CL.  Does not yet
create an iOS implementation of PolicyConversionsClient, but that will also come
in a future CL.

PolicyConversionsClient is an abstract class which contains methods that
individual embedders need to implement.  For example, //content-specific
dependencies are hidden behind these methods.  PolicyConversionsClient
also contains some helper methods that need to be shared by all
embedders.

ChromePolicyConversionsClient is the //content-specific implementation
of PolicyConversionsClient.  Its code is largely moved from
PolicyConversions, with some small refactorings to call methods on the
superclass instead of directly accessing private ivars.

BUG=1027249

Change-Id: I64e01ff487df9f96ebb657e4cde6651490833f20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2025927
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738622}