Implement Omaha 3.1 Protocol Upgrade.
This implements the section related to 'manifest objects' in the Omaha
protocol:
https://cs.chromium.org/chromium/src/docs/updater/protocol_3_1.md

```
A manifest object contains details about how to fetch and apply an
update.
*`arguments`: A string, indicating command-line arguments that should
be passed to the binary specified in `run`. Default: "".
*`run`: A path within the CRX archive to an executable to run as part
of the update. The executable is typically an application installer.
If unsent or the empty string, no particular update-delivered
installer needs to be run. Default: "" (empty string).
```

Such values are being parsed and passed over as an optional parameter
to the CrxInstaller::Install call. An std::unique_ptr is used to
avoid forcing all callers of CrxInstaller::Install to use the
base::Optional type.

The //chrome/updater code picks up this parameter and, on Windows,
it runs the program specified by the update response. The other
platforms don't implement this call yet but code will land soon
for macOS.

This change is transparent for the component and extensions updaters
since their installers ignore these optional parameters.
A compliant server would not respond with {run, arguments} for
these update_client requests.

Bug: 1042218
Change-Id: I00368182d0aa1189893808c1b6972284c5b3407e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036287
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738633}