[Paint Preview] Add performance UKM
Adds UKM support for reporting the capture time of the main frame (in
ms) for a paint preview. Support for memory overhead and on-disk size
may be added to this dataset in future, but is out-of-scope for this CL
as getting reliable memory metrics is non-trivial and on-disk size is
handled on a per-implementation basis.

Bug: 1038390
Change-Id: Iaaf77b3722633773455babd3e7e51cd95f713577
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1988222
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738638}