Add Windows high contrast support to the toolbar button icons
This change adds different foreground icon colors to the toolbar buttons
for the hovered and pressed states when the high contrast ink drop
feature is enabled. This is to support the desired platform high contrast
colors. Future work will add foreground color support to other buttons.

New theme properties were added to support the additional _HOVERED and
_PRESSED state high contrast colors. When not in high contrast, these
are the same as the normal icon color. In high contrast (+ flag) they
return the selected text HC color. These new colors are used to create
images in the toolbar buttons for the various button states.

New helper functions we added to ToolbarButton to simplify setting the
images for the different button states, including one that just uses
the standard colors returned by |GetForegroundColor|.

In the future, we might want to do work to allow icons to be separated
from colors to allow for easier recoloring.

Bug: 776871
Change-Id: Id72d8c60a23889f018044a5fcce7abd9359ad61a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2029492
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738660}