trybot_commit_size_checker.py now outputs size improvements
trybot_commit_size_checker.py now specifies if a metric delta is a large
improvment (greater than the limit, in the opposite direction). A
boolean is set on the binary size gerrit plugin json object for each
metric if it is a large improvement. This allows the plugin to showcase
those differently on the CL page.

Bug: 921090
Change-Id: I0b809835289f8af9046d477826abdd7d7d9a7387
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2037958
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738679}