Invalidate throttling on main-frame cross-origin changes
Invalidations are needed for throttling when a frame's
cross-origin-to-main status changes (see https://crrev.com/451041 and

be set from javascript by modifying document.domain. We issue throttling
invalidations on the frame that changes origin, but missed the case when
a main frame origin change causes subframes to become
non-cross-origin-to-main. This patch invalidates all descendant frames
when the main frame's origin changes.

LocalFrameView: :UpdateRenderThrottlingStatus). The origin of a frame can
Change-Id: Ibd6b99f6666deb4f583f37cad9cfd2a823ea749e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036269
Reviewed-by: Stefan Zager <szager@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738687}