Fix a crash in ProfileDestroyer due to delayed RenderProcessHost destruction
This CL addresses a crash (CHECK_EQ being hit) on desktop when a
RenderProcessHost instance is still alive while the associated
ProfileDestroyer is on its way to destruction. This started happening
after https://crrev.com/c/1971031 landed to fix other crash on Android
by delaying the destruction the off-the-record profile, which would
have destroyed the RPH before hitting the CHECK_EQ.

This CL moves the off-the-record profile destruction task where it was
but uses |DestroyOffTheRecordProfileNow| instead. This change takes
care of the desktop crash but also makes sure the clean up task
required for the Android issue comes before the profile destruction.

Bug: 1040839
Change-Id: Iabab0a304a802d1bd64adec255e88d81b6987cde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2032385
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738746}