Re-enable telemetry_perf_unittest's BenchmarkSmokeTests on Chrome OS.
According to the comment in the code, disabling this test disables all
benchmark smoke checks in telemetry_perf_unittest. That seems like a
critical set of tests to lose coverage for. So this re-enables them.

Also increases the shard count of telemetry_perf_unittest's on the
chromebook tester since this increases its runtime significantly.

Bug: 1041328

Change-Id: I849a04c57c8f9ac4eb687e93b85322b67cb47cfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2032325
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738751}