viz: Ensure correct alpha info before buffer allocation.
Buffer allocation for the display compositor is currently done before
overlay processing. This order is incorrect since we need to know
whether buffers used by the compositor need to support alpha and this
is not known until overlay processing. If a quad is promoted to an
underlay then the display compositor buffers will need to support alpha.

R=dcastagna@chromium.org,ccameron@chromium.org

Bug: 946360
Change-Id: Ibb854cdb7764d4f6f6f41e490012416f9f0497ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026968
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738755}