Force VoiceOver to use Chrome's computed text for links.
VoiceOver was incorrectly reading links that contained
absolute-positioned text inside. As a workaround, expose
the text of a link in AXDescription instead of AXTitle.
That forces VoiceOver to treat it similarly to a link that
has an aria-label on it, so it favor's Chrome's accessible
text, rather than treating it as a link containing visible
text and using its own (broken) heuristics to describe
what's inside.

It's possible this could have a side effect, if VoiceOver
was announcing more detailed information inside a link.
But it'd still be possible to access the contents of the
link, so this is probably okay.

Bug: 1048782

Change-Id: I2764afe80af1f6afc4c191cf7f98bbcb6b3c025f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2037710
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738791}