[UI-PrintPreview] Align the margin control with the contents
In right-to-left direction, the scrollbar of the print preview area may
be on the left, and we should align the margin controls with the preview
contents.

The bug was a "regression" caused by crrev.com/588201 which corrected
position of contents in a scroller with a scrollbar on the left, because
before the CL the preview content was not at the correct location and
the margin controls happened to be at the same location.

Bug: 1049038
Change-Id: Ida362fbfc1622998a1ad928217ce8abf7ff5c34c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2040438
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738838}