Extend AXLiveRegionTracker to generate live region change text.
When a live region changes, optionally provide a way to generate the
text associated with that change. On some platforms we might expose
that text directly so that AT doesn't need to try to compute it.

The computation is somewhat subtle because it involves worrying about
deleted nodes and handling aria-atomic, aria-relevant, and aria-busy.

Bug: 1042893
Change-Id: Ia79f0c976bef73ea25a8821dbb9fc7f22fde0e3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2006131
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#738876}