Add initial support for AutomationPosition
This patch introduces AutomationPosition. It acts as a wrapper of an
AXNodePosition and utilizes gin::Wrappable to ease the burden of surfacing
positions in javascript. gin::Wrappable handles the details of object lifetime
between the object within V8 and the native C++ object.

This change has gone to great lengths to make it as easy as possible to define methods and properties on this wrapper object so that they are self-contained and trivial to add, remove, and edit. This will facilitate future changes in the underlying AXPosition class and make it straightforward for anyone changing the api.

Bug: 1048367
Test: browser_tests
Change-Id: I9c881002ffcb991e4f8a1db3c269e5be0e9144da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1814650
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738880}