[Code Health] Reformat throughput_test_case.py
The format in this file is in-consistent, sometimes there is one
blank line after a test case, and sometimes it is two. I looked at
other test cases and it seems two blank lines is the right format.
This CL does the re-format, no beavior change.

TBR=sadrul@chromium.org

Bug: None
Change-Id: I6e6246eb9278ed5bd5acd3d9bdd86d0468cc78f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041798
Reviewed-by: Xida Chen <xidachen@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738983}