Add a TabModelObserver to control show/hide of new tab variation
Currently, we check the number of incognito tabs in
onIncognitoStateChanged in TabSwitcherModeTTPhone to decide whether we
show new tab variation or not. We assumed that there will definitely be
an incognito state change when we close all incognito tabs. However,
this is not true in a corner case where user closes all incognito tabs
from Android notification while staying in normal tab model. This CL
covers this corner case by introducing a TabModelObserver to control
the show/hide of new tab variation.

Bug: 1028497
Change-Id: Ia71d829a9ea5a87e50c5930e60f2165759fbbb06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015833
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#738993}