[CrOS Settings] Remove SettingsSectionName enum
After design discussions with jamescook@, we've decided to remove this
enum, since it's not strictly needed for the metrics we're trying to
compute (number of navigations until a setting has changed). This
complexity is unnecessary at this time.

Bug: 1047062
Change-Id: I6280115e051f9bbd67434e5a5faf25c1bc03854d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2040913
Auto-Submit: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739028}