[Extensions Menu] Properly clean up the views in BrowserActionApiTest
Properly dispose of the custom views created in BrowserActionApiTest
during the tear down phase. This prevents a UAF when the view tries to
clean up after the profile is destroyed.

This was just an issue in the test, not in the wild (where views
associated with a browser will never outlive the profile).

Bug: 984654

Change-Id: Ied7dc0bd9c471fa2445b92096550adf09575043a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042390
Reviewed-by: Peter Boström <pbos@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739077}