Remove watcher process and move exit code watcher to crashpad process
This change removes the watcher process.  The watcher process is
responsible for waiting for the browser process to exit and logs the exit
code to a known registry location. Subsequent launches read from the
registry location and log the value to a histogram.  The watcher process
was originally created prior to the creation of crashpad.  Since all the
watcher process is doing is waiting for the browser process to exit it
makes sense to move this functionality to the crashpad process and get
rid of the watcher process and code entirely.  Use of the registry is
now removed and the histogram is logged directly on shutdown.
This change removes the chrome_watcher.dll which gives back ~ 1MB of
disk space.

Bug: 1043844
Change-Id: I666819e64817197e8950a9c5749c921fcebdb964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2035054
Commit-Queue: Chris Davis <chrdavis@microsoft.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739109}