Support result_selection with Skia readback
Add support for CopyOutputRequest::result_selection when using
SkiaRenderer with the Skia readback API. Previously the field was
ignored in SkiaOutputSurfaceImplOnGpu. The skia readback API doesn't
support specifying a clip rect in the destination pixel space, so this
is a bit of a workaround with the potential for some rounding errors.

This CL also adds SkiaReadbackPixelTest. It draw a frame containing a
source image with a CopyOutputRequest. The pixels that are read back are
compared against reference PNGs. The test is based on
GLRendererCopierPixelTest and uses the same source and result PNGs.

Bug: 1047217
Change-Id: I9f97cacafff15833d6ce97d36c69c70b55d09479
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033624
Commit-Queue: kylechar <kylechar@chromium.org>
Reviewed-by: Jonathan Backer <backer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739127}