[SSL] Clean ChromeMetricsHelper of //chrome deps
This CL changes
ChromeMetricsHelper::StartRecordingCaptivePortalMetrics() to take in the
CaptivePortalService instance rather than obtaining it internally from
the Profile and similarly changes ChromeMetricsHelper to take in the
HistoryService instance in its constructor. The motivation is to enable
this class to be componentized in a followup in order to make it
available to WebLayer.

Bug: 1030692
Change-Id: I6447dcc21e317d7ed4ba2b01b821575028066d4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2040412
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739172}