Fix intent picker automatically dismiss issue.
Click to call dialogs needs to be closed if the tab navigates to a
different origin. In ChromeOS, click to call uses the same dialog as
intent picker. Intent picker dialog will show when tab navigates to a
different origin and there are Android apps that can handle this origin.

The click to call dialog closes cause the dialog to automatically closed
after briefly show up. This CL change the code to only close the dialog
for click to call. This CL also fix a issue the intent picker icon don't
show up for auto dialog pop up.

BUG=1045036

Bug: b/146227259
Change-Id: Iadb1db30c7e24108bee933eb93d57d73a206d3d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2035821
Reviewed-by: Richard Knoll <knollr@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739217}