[quickview] Add |canDelete| property to |files-quick-view|
In preparation for adding a delete button to the Quick View toolbar add
property |canDelete| to |files-quick-view|, which indicates whether the
file shown can be deleted or not.

In a future CL, |shouldShowDeleteButton_| will then toggle the display
of the delete button in Quick View.

Test: Tests coming in future CL.
Bug: 803259
Change-Id: I880fd7f2065c11435a16ec923abca12277b0c824
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2040502
Commit-Queue: Noel Gordon <noel@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739221}