[Structured metrics] Re-enable launcher collection.
This re-enables the launcher collection that we disabled while moving
from the hashed logging prototype to the structured metrics framework.

All logging is the same, with one exception: we previously had two
metrics aimed at recording which provider a search result comes from,
with each determining about half of the providers. I've simplified this
by combining them into one metric, which records the value of
RankingItemType, the enum we use for type identification for all
on-device models.

To do this I've cleaned up the RankingItemType enum, removing item
types that we no longer use and adding a "don't change this" header.

Bug: 1016655
Change-Id: Id04156a4818f842af6489aea23d4048d7aeb4824
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2038292
Reviewed-by: Charles . <charleszhao@chromium.org>
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739241}