Use HbScoped in Blink platform
Now that we have scope helpers defined in a shared location for typical
HarfBuzz types, use those in Blink. In order to be able to do that for
HarfBuzzFontCache, FontGlobalContext needs to forward-declare and
isolate HarfBuzzFontCache from leaking outside the header, otherwise all
clients need to be made aware of HarfBuzz types.

Bug: 1049527
Change-Id: Ic2e6a3fec01a4a6f22eb879d77615a986dc4b987
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041471
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739318}