Store line/column on pending callbacks for the XML parser.
When the XML parser is paused, we temporarily store pending callbacks to
call the XML parser hooks when the parser resumes. For error reporting
we used the line/column number for the current position of the parser
input which is not correct when handling errors for the invoked
callbacks. This CL adds these source positions to the callbacks to have
correct positions for error reporting.

The test below will fail without this CL when the
BlockHTMLParserOnStyleSheets flag is enabled.

TEST=svg/hixie/error/013.xml

Bug: 891767
Change-Id: Ie3e6c608253dfa355b9947433c79ef21d912cbc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041443
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739338}