Remove LayoutMenuList::options_width_
options_width_ is updated at the beginning of
ComputeIntrinsicLogicalWidths(), and referred only in it. It doesn't
need to be a data member of LayoutMenuList.

This CL renames UpdateOptionsWidth() to MeasureOptionsWidth(), and it
returns the measured value, and removes LayoutMenuList::options_width_.

This CL has no behavior changes.

Bug: 1040828
Change-Id: Ic33faffa74ede63444e150914541ed4985f70243
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2043340
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739343}