Guard for null typefaces in Mac system font instantiation
The previous change [1] to always apply opsz to Mac system fonts did not
handle a situation in which loading a system font comes back with a null
result. Cover this case for the base typeface as well as for the case
where instantiating the variable font fails.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2020767

Bug: 1049391
Change-Id: I2aad15c31f7ad793abc389f66afab3fbdf963df7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041667
Reviewed-by: Ben Wagner <bungeman@chromium.org>
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739391}