[SSL Interstitials] Componentize ChromeMetricsHelper
This CL componentizes ChromeMetricsHelper and renames it to
ContentMetricsHelper, as it is now specific only to //content embedders
that use //components/security_interstitials, rather than being specific
to //chrome as it was previously. The motivation is to reuse this class
in WebLayer (which will happen in an upcoming CL).

Bug: 812808, 1030692
Change-Id: I48e533fe7d56bba84bf16c877148aa98a5779ec5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041625
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739412}