aw: Not use WeakPtr in TaskForwardingSequence
TaskForwardingSequence created and destroyed on DisplayCompositor
thread. Destruction will cause weak pointers to invalidate. The weak
pointers are checked when RunTask happens and it's on Android Render
Thread. Weak pointers are not thread safe, so this is incorrect.

This CL fixes the issue by not using WeakPtr and binding all necessary
data to task.

Change-Id: I89ef10ff332bd806d660fa4d7b6d9ec6417fcc1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042384
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739448}