viz: Ensure correct pairing of Begin/EndPaint in SoftwareRenderer.
Prior to [1], Begin/EndPaint on the SoftwareOutputDevice always happened
in pairs since it was done in Begin/EndDrawingFrame. But now BeginPaint
is done only if we draw to the root render pass, which can be skipped in
case of an empty swap. Ensure that we only call EndPaint if a BeginPaint
was needed.

[1]: https://chromium-review.googlesource.com/c/chromium/src/+/2026968

R=ccameron@chromium.org

Bug: 1049545
Change-Id: Ib0b27e62051107670852ede5d3ce73371cd3e51e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042514
Auto-Submit: Khushal <khushalsagar@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739455}