[LayoutNG] Make hit-testing work inside an anonymous block.
When self-hit-testing an anonymous block (because the point is within
the bounds of the anonymous block but not within any of its lines), we
need to UpdateHitTestResult(), so that we provide the right node in the
hit testing result. Otherwise we'll fall back to the node of the
enclosing PaintLayer.

Bug: 1043787
Change-Id: I735e50a27e19d178b36f5e4049f91eea109412f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044092
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739481}