ash/shell: Ingest privacy-screen key press
Certain chromebooks now come with panels that have integrated privacy
screens on them. This CL adds an accelerator that toggles the privacy
screen when the privacy screen keyboard key is pressed.

In addition, this CL relaxes the restriction in ash/components to depend
on //ash as per https://crbug.com/1046624

Bug: 1046624, b:147451643
Test: tested locally on a test device
Change-Id: I2bfec87a258d89222826e292d704ac5a30949957
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2021014
Commit-Queue: Gil Dekel <gildekel@chromium.org>
Reviewed-by: Gil Dekel <gildekel@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Tao Wu <wutao@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739489}