Don't let devtools override reserved hotkeys
Based on https://chromium-review.googlesource.com/c/chromium/src/+/1741434,
(and specifically the comment on deprecated_is_app()) I'm going to conjecture
that this was unintentional in the first place. In any case, none of these
particular commands are anything DevTools should be overriding, and this
causes unfortunate inconsistencies in Mac key command handling (see c#11 and
c#12 on the bug).

Bug: 1047046
Change-Id: I74c22fb510a7831667037e4fc3ae72d4873afa64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044140
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739557}