CrOS Shelf: Fix spacing between navigation button and left screen edge.
This change fixes spacing between the screen edge and shelf buttons
on the edge of the shelf along the primary axis. For example, with
bottom shelf this fix changes the spacing between the left and right
most buttons and the edge of the screen.

Also in this change, some code which unnecessarily enlarged the size
of the navigation widget's ideal size after the size was already
enlarged has been removed.

Bug: 1046115
Change-Id: I0fa2ffea8f23260f3aa122238bf789da80604a1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2031862
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739575}