[UKM] Record the reasons for a BeginMainFrame
The CC frame sequence metrics have knowledge of why we need to be
performing frame updates. Pipe that through to the renderer to
annotate the Blink.UpdateTime UKM event. This will allow us to
identify common performance patterns.

Note that this information is already available in UKM via the
Compositor Throughput UKM event, so does not represent new
information. But it cannot be correlated with a Blink.UpdateTime
event for performance analysis of the renderer.

Bug:1040551

Change-Id: I3159b4363d86448df200c6ce8978f4aa6f293830
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2040181
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739576}