Add an expires_after tag for 2067 histograms that do not have one.
This is required because a new presubmit check will soon require all
<histogram> entries to have an expires_after attribute.

The expiration of "M85" was chosen to be about six months away as that
is the threshold for auto-delay.  Auto-delay is a weekly job that pushes
out the expiry date of histograms that have been

  a) accessed (even infrequently) on the UMA dashboard, or
  b) the cause of a Chirp alert.

If no auto-delay occurs then crbugs will be filed as the expiration date
approaches providing plenty of time for owners to manually push out the
date.  Significant effort was expended last year to update the owners of
all histograms so these bugs should not go unnoticed.

This will undoubtedly cause some extra work for some people but it is
important that these histograms are not forgotten about.  Unnecessary
histograms bloat Chrome's code size, waste user bandwidth, and cost
significant money to store and process.

While expiring histograms will not help with code size, it will reduce
the others.  Please, if a histogram is no longer used, remove it from
the code base and add an <obsolete> tag to its description.

Change-Id: I09726a63efae5a21e6b6a8ee993659f4990912f6
Bug: 1030096
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044139
Commit-Queue: Brian White <bcwhite@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739639}