WebUI Tab Strip: Make tab and group element placement methods public
This CL updates the method to insert or move TabElements to not rely
on TabElements always having a |tab| object, and instead moves the
variables needed to place them to params instead. This also moves
the method to move TabGroupElements out to its own method.

This allows for both the DragManager to eventually place elements
externally without calling private methods, and allows for placing
elements that do not have data yet in the cases where a tab is being
dragged from another window.

Bug: 1048894
Change-Id: Ic5d4cbf6bcf196dbec956ad85c5703362d38cf2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044878
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739655}