AppCache: Fix 304 resource response handling
AppCache previously never implemented 304 response handling for
resources at all.  The correct handling sequence causes the cached
item on disk to be updated based on the 304 response's results,
including its request and response times.

When corrupted cache entries were created due to another bug (issue
968179), this missing 304 handling resulted in those corrupted cache
entries persisting indefinitely.

This commit changes the 304 handling sequence for resources
so any cached entry's info will get updated based on the incoming
request's info and the entry's data will get copied to that new
entry.  One implementation detail: it was easier to create a
new cached entry based on the old entry rather than updating
the old entry directly, so the change is written to copy cached
entries over, updating them in-flight based on the response.

As part of this change, this commit adds an UpdateJob unit test to
verify the cached data is correctly updated by first writing a cached
entry to disk with broken request and response times, kicking off the
update, letting it complete, and then reading the newly cached data
to verify that the request and response times are correctly updated.

This commit doesn't address 2 cases:

* It doesn't address 304 fixes for non-resource handling which will
remain as it was since the AppCache feature is close to removal.

* It doesn't handle the case where a cached entry has a Vary header
and the incoming response does not.  In those cases, we'll continue
to be non-compliant with the spec.  We believe this is acceptable
since support requires a lot of extra work and we expect it to be a
rare case.  See the CL for more details.

Bug: 989611
Change-Id: Ifcb1c853da49d8f3a26104edad96d23386715f1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2018442
Auto-Submit: Chase Phillips <cmp@chromium.org>
Commit-Queue: Victor Costan <pwnall@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739668}