Use Browser in CrashRestoreHelper instead of TabModel (v2)
V1: https://chromium-review.googlesource.com/c/chromium/src/+/2024028
TabModel is provided to the CrashRestoreHelper as a SessionWindowRestore
protocol delegate. and it's used to restore sessions from there. Replace
that with browser and use sessionRestorationBrowserAgent from the
CrashRestoreHelper instead.

V2: This  CL:
Also i changed member method: moveAsideSessionInformation to be
a class method moveAsideSessionInformationForBrowserState, the method
is stateless and should be called before the creation of the browser. so
it makes sense for it to be a  class method which returns true if the
operation was a success.


Bug: 1010164
Change-Id: Ib3bd85423552134a3f358cdeab31335d1a25bf02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2039348
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739731}