[cascade] Add CascadeExpansion
CascadeExpansion is an iterator that can traverse a declaration block,
and expand 'all', '-internal-visited', and '-internal-ua' properties.
In other words, when e.g. 'all' is physically encountered in the
declaration block, CascadeExpansion will generate and emit the
declarations 'all' represents, such that the caller doesn't have to
handle the 'all' case with special code.

This will simplify StyleCascade later, but is unused for now (except
for tests).

Bug: 947004
Change-Id: I11c924e8469a66d13e0ef759941dbe5832454b47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041654
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739864}