grit: gather: fix up encoding handling
The image data is read as bytes and then converted to base64.  The
base64 module reads bytes (which is what we're giving it) and returns
bytes which we then treat as a string.  Convert it to a string so we
can merge it with the other string here.  It's guaranteed to be ASCII
so this should always be safe.

For the other html files, we've been reading them as binaries, but
then running string operations over them.  Lets convert reading of
these files over to UTF-8 and see if it matters to anyone.  If so,
we can reevaluate by trying to convert all the rewriting logic to
only operate on bytes.

These require fixing the GetData encoding handling in chrome_html.
We also tweak the other modules to make sure the encoding requested
always matches the code behavior.  If that changes, we'll know.

Also change GetDataPackValue to operate on bytes.

This func already uses bytes in a few places and relies on the fact
that Python 2 allows for arbitrary binary data in strings.  Clean it
up to use bytes all the time to work with Python 3 too.

Bug: 983071
Test: `./grit/test_suite_all.py` passes
Change-Id: I7be2246652ca5744227bdd205141f371c5306fb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2045475
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739972}