pinpoint_cli: Stop erroneously modifying non-tryjob configs
Logic in pinpoint_cli/commands.py is intended to alter outgoing
tryjob configs to swap out the 'start_git_hash' arg, if present, for
'base_git_hash', but it does not explicitly check that the config at
hand is a tryjob (as opposed to a bisect)---and removing the
'start_git_hash' arg from a bisect config can lead to server-side
errors failing the Pinpoint API request.

This CL fixes the bug by conditioning the request modification on
whether the config is for a tryjob or a bisect (and no longer removing
the argument in the latter case).

Bug: 1050173
Change-Id: I17f0874665ec02a0d2cd5b2fde2b268f58fb3457
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044723
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#739981}