cc: Remove raster color space id plumbing.
This was added to allow color space to be cached by the transfer cache,
when it was being sent to the GPU process for each tile raster. This is
no longer the case since the color space is now stored on the shared
image.

R=ccameron@chromium.org

Change-Id: Ie93e6d0425e4af812d9b2a65c1e1595f8a8a7c9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2047407
Commit-Queue: Khushal <khushalsagar@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Auto-Submit: Khushal <khushalsagar@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740007}