Disable optimization for objects painting nothing
The optimization is to avoid paint invalidation for objects that will
nothing. It was important before we have separate raster invalidation.
Now it might be less important because these paint invalidation won't
cause raster invalidation which is based on painted result.

Will revert this CL if the regression is not acceptable. Otherwise
will remove the dead code for these optimizations.

Cluster-telemetry [1] shows that this CL will not affect record time in
the benchmark because this CL only affects paint invalidation which is
not measured by the benchmark. Will see how this CL affect real-world
PrePaint and Paint performance after it is landed.

[1] https://ct.skia.org/results/cluster-telemetry/tasks/chromium_perf_runs/wangxianzhu-ChromiumPerf-4638/html/index.html

Bug: 1049440
Change-Id: I79481036dc3f4ada31093aa1f525606ce17bb56d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042262
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>