Try to lookup DeviceInfo by guid for WebRTC Sharing
When we setup a WebRTC connection for Sharing messages we might not have
the remote DeviceInfo available yet. Instead of always assume that's the
case this CL tries to lookup the DeviceInfo first in case we do have it
already. This is better as we can then log all metrics in the
SharingMessageSender with the correct values.
This also fixes a possible null dereference in
SharingServiceProxyAndroid and adds logging for when the DeviceInfo was
not available.

(cherry picked from commit 0e5eeedae1e491e4218deb2a4e8fda0aa7c820fb)

Bug: 1049134
Change-Id: I509010108f2f685987a1b48f757d3ad0f2aefbec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2039611
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Commit-Queue: Richard Knoll <knollr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#739001}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044276
Reviewed-by: Richard Knoll <knollr@chromium.org>
Cr-Commit-Position: refs/branch-heads/4044@{#159}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}