Fix Creating a app window from the teleport window crash issue.
When the Chrome app window is teleported from the inactive user, the
item controller tries to get the app window from the current user's
profile, so the app window is null, and it crashes when calling a
function on a null pointer.

Modify the item controller's function GetAppMenuItems and
OnWindowTitleChanged to check all profiles saved in
AppServiceAppWindowLauncherItemController, to get the correct app
window pointer.

AppServiceInstanceRegistryHelper::IsOpenedInBrowser is updated to check
the proxy for all profiles, because the window could be teleported from
the inactive user.

BUG=909925

Change-Id: I041f4b1f58292f3a8c01ef7957d3f395e998f4bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2043330
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740061}