Add a histogram for downloads initiated by the tab's window.opener
This CL adds a histogram to record when a download is initiated on a
tab/window whose window.opener is non-null. The goal is to
understand the prevalence of the scenario demonstrated in crbug/121259.

Example: If a.com calls window.open("http://b.com"), it can now navigate
the newly opened window. This navigation can result in a download. In
that case the download shelf will display the downloaded item over
b.com's window. Since the download shelf doesn't display the origin
of the download, this can be used to trick users into thinking that
the download was initiated by b.com.

Bug: 121259
Change-Id: If6743ae7a21795efa40ce7ea3ef81f30b8d414a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042149
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Christopher Thompson <cthomp@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740062}