Skip virtual/off-main-thread-css-paint/* on Win10
Tests under virtual/off-main-thread-css-paint flakily crash/fail on
Win10. Because of crbug.com/1050838, run_web_tests.py crashes when
any of the tests fails, and because of crbug.com/1050834, the crash is
not caught by the try bot, so we have to skip these tests on Win10 for
now to ensure coverage of other tests.

TBR=robertma@chromium.org,xidachen@chromium.org

Bug: 1050834,1050838
Change-Id: I175567308cf89f721f2f156c9200f413df226a46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2048995
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740138}