Add APIs for performing WebAuthn origin security checks
This is the first of three patches to allow Web Authentication in Clank
to perform origin checks in the browser process.

This CL refactors the security checks into a separate class, and
adds content APIs that allow them to be invoked by WebAuthn request
handlers outside content. The refactor also adds a central place that
hopefully makes it easier to make changes to the security policies in
specific situations, in future.

The second CL will have the WebAuthn request handler code in Clank
internal to call these APIs.

The third will remove the security checks currently performed in the
renderer process.

Bug: 827542
Change-Id: Icf5cc4df2e02c5dee8c2211adac34e90832cae90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2031548
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740183}