Change RenderPass to use ContentColorUsage instead of ColorSpace
Prior to this change, SurfaceAggregator would assign a ColorSpace to
each RenderPass, based on the alpha and content needs of that
RenderPass.

This starts falling apart because SurfaceAggregator doesn't know the
final answer of "what alpha is needed by this RenderPass" and "what
color spaces are present in this RenderPass". This is because the
DirectRenderer's OverlayProcessor can add the need for alpha (because
of hole punching) or remove the need for HDR (by taking an HDR video
and promoting it to an overlay).

Change the RenderPass to store the ContentColorUsage, from which the
DirectRenderer computes the required color space.

Change the way that SurfaceAggregator decides if an HDR pass needs to
be added by calling ColorSpace::IsSuitableForBlending on what is the
expected output color space. Note that this decision is not correct
(or was it before), because it assumes that the OverlayProcessor will
not change the content or alpha needs of the RenderPass. A more
comprehensive fix would be a more substantial refactor, moving the
OverlayProcessor to be more closely tied to SurfaceAggregator.

Bug: 1049334
Change-Id: I0d21a291cf20231946e3e8fdda77742deedd5ba4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2047453
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740194}