[SyncSetup]: Add histogram to record sWAA state.
This CL adds a histogram to record the state of web and app activity.
The histogram is recorded on the completion of the request. The main
goal is to track how many users have sWAA On/Off and to monitor the rate
of failure to fetch this value.

Bug: 1050070
Change-Id: Ie4b0b4020b097a29850bb4f38b34e66ea62d084b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044214
Commit-Queue: Monica Basta <msalama@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740208}