[iOS] Use the actual error page in new workflow
This CL updates the new error page workflow, making sure that real error
page HTML is used when the error is done on provisional navigation.

For that, it is making sure that the context of the failed navigation
is passed to the navigation to the error page file.
It is also injecting the HTML of the error page into the loaded file
when the navigation finishes.

Bug: 991608
Change-Id: I382d9172bc4cdbd624d8629e6099c366f67d68a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024895
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740256}