Add tests for web documents navigating to WebUI ones.
We are currently lacking explicit coverage for web documents starting
a navigation to WebUI document. This is currently handled by FilterURL,
which disallows the navigation and rewrites the URL to
about:blank#blocked. The CL adds explicit tests for this behavior and
those should be changed as we move towards blocking the navigation
instead of rewriting the URL.

Bug: 1002276
Change-Id: Ie61558b55e4bb3c887844b53b91f731b1e7029e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044335
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740469}