SharedWorker: Use CSP headers to outside settings
This CL passes real ContentSecurityPolicy and ContentSecurityPolicyType
values to creation_params which are used for SharedWorker scripts fetching.

By this change, outsideSettings CSP is applied to scripts statically
imported from module SharedWorker top-level scripts.
This behavior is tested in the static import section in the added test.
(Previously the CSP was not applied at all for such scripts,
but it's OK because module SharedWorker is not yet shipped.)

This CL doesn't affect the behavior of classic SharedWorker and
dynamic imports (either from module or classic SharedWorker).

Web-platform-tests to check its behavior on Module SharedWorker are added.
WPTs for Classic SharedWorker will be added in the following CLs.

Bug: 1049011, 824646
Change-Id: I6fa79b0394426f09b19da1abbdf9ecf483d558ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2038297
Commit-Queue: Eriko Kurimoto <elkurin@google.com>
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740538}