Fix bindings for WebUI error pages that reload successfully.
WebUI bindings aren't granted to error pages, and we weren't granting
them when reloading a NavigationEntry that had earlier been an error
page.

To fix this, reset all sensitive state (including bindings) on the
NavigationRequest when the destination SiteInstance changes from what
the NavigationRequest expected it to be.

Bug: 1046159
Change-Id: I2b004423ad2f1f7aecd86220be75d32f26404040
Fixed: 1046159
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2045143
Commit-Queue: Charlie Reis <creis@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740557}