Fix IPsec configuration update using UI
Currently on network configuration updates through UI, there are some
empty fields populated and not omitted before being validated by the ONC
validator. One of these fields is the client certificate type. This CL
ignores this field if it is empty and adds tests with client certificate
type set to empty to make sure that the ONC validator will not complain
and that the VPN configurations will get updated as expected.

Bug: chromium:1043126
Test: chromeos_unittests -gtest_filter=ManagedNetworkConfigurationHandlerTest.*
Change-Id: Ic84e4250e2c01744c6d4632108427d2478f64e16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2035974
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Omar Morsi <omorsi@google.com>
Cr-Commit-Position: refs/heads/master@{#740593}