Move out enum types from ScrollIntoViewParams to global types
This is a follow-up CL of https://crrev.com/c/1993076 and this CL
is a part of the effort to use the mojo type in Blink without
converting between Blink and Non-Blink.

This CL moves out enum types from ScrollIntoViewParams to
scroll_enums.mojom as global types. These enums are used by lots of
different Blink scroll methods, not just ScrollIntoView.
ScrollIntoView just happened to be special because it is called
cross process and so required the Blink-internal enum to be cloned
into the mojom. Then the internal types were replaced by the mojom.
This CL makes these two types be more general than ScrollIntoView.

Bug: 1039285
Change-Id: Ie3387e6f0de757336dfd2917ee9a5bf91fac117b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026855
Commit-Queue: Miyoung Shin <myid.shin@igalia.com>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740595}