BackForwardCache: Record UKM only for history navigations
In BackForwardCacheMetrics::DidCommitNavigation,
|last_committed_main_frame_navigation_id_| can be not -1 not only on
history navigation but also on reloading.

This CL changes the conditions to record UKM only on history navigation.
We plan to add other UKMs for back-forward cache, and they should not
be recorded when reloading. For example, we are trying to record the
reasons why the page was not in the cache, but recording them when
reloading does not make sense. Then, we need more strict conditions.

Bug: 1033410
Change-Id: I933e7093805f643e60f9059c753a1f5ec2cacd9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1988093
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740610}