Remove scroll_into_view_params_type_converters.{cc, h}
This is a follow-up CL of https://crrev.com/c/1993076 and this CL
is a part of the effort to use the mojo type in Blink without
converting between Blink and Non-Blink.

This CL removes scroll_into_view_params_type_converters.{cc, h} and
uses mojom::blink::ScrollAlignment mojo type instead of
ScrollAlignment class in Blink, and we keep ScrollAlignment class
as a helper class.

Bug: 1039285
Change-Id: I8cc47a8188b15a5e73f2d27783ef382a015f9720
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024638
Commit-Queue: Miyoung Shin <myid.shin@igalia.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740618}