[SSL] Componentize launching of date and time settings
This CL componentizes the implementation of launching of date and time
settings found in //chrome's SSLErrorControllerClient in order to reuse
this functionality in WebLayer (that reuse will happen in a followup).
The code has no //chrome dependencies. We placed it in the
security_interstitials component as that is the only consumer of this
functionality at this time; if need be it can be moved into a more
generalized location (e.g., a standalone component) in the future.

Tested manually that visiting the bad clock interstitial and clicking
"update date and time" still works as expected on Android (note that on
Linux, the other platform where I can locally test, this already
doesn't work as expected; filed as crbug.com/1051427).

Change-Id: I062dbf61392c329f11d9a3f8af748a4ec2086d05
Bug: 1030692
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1988353
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740629}