[LayoutNG] Clean up ChildrenInline()ness.
Introduce the term "in inline formatting context" and "inline formatting
context root" for box fragments and nodes, which hopefully avoids some
confusion.

On one hand, we can ask a LayoutObject whether its children are
inline. This isn't exactly the same as asking an NGBlockNode whether its
children are inline. And none of those two is the same as asking an
NGPhysicalBoxFragment whether its children are inline.

Also remove the recently introduced IsBlockLevel(), which was almost
(but not quite) the opposite of ChildrenInline().

Change-Id: Ic000ba3a882a1edd2a2e0c9aa3e70e88419a1104
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2047243
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740669}