[iOS] Create user sign-in flow for sign-in architecture migration.
This patch embeds the UserConsentViewController into the sign-in view
host. Interactivity and consent buttons will be added separately.

This CL is part of a series of refactors to move the existing sign-in
architecture to the Coordinator-Mediator design paradigm.

Fore more information, see go/chrome-ios-signin-migration.

Bug: 971989
Change-Id: Ie0e87842036c68a72aed2a404dfe6a98a1d8cae5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041599
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740732}