Fix Crash in XrRuntimeManager::InitializeProviders
All Providers are created in XrRuntimeManager::GetOrCreateInstance, of
these providers all are created with make_unique except for the
ArCoreDeviceProvider, which is created via a factory method. If this
factory hasn't been installed (which happens via a static method from
ArCoreInstallUtils.java via the ArDelegate init method), then the
provider may be null. The presence of the ArCoreDelegateImpl (which
actually installs the ArCoreDeviceProviderFactory) is gated on the same
buildflag (enable_arcore) as adding the provider to the providers list.
This _should_ mean that this provider shouldn't be null. However, it
appears as though there may be some race condition or other issue that
causes this provider to be null.

To fix the crash, both the processing of the providers and the creation
of the ArCoreDeviceProvider are wrapped with null checks. This will stop
the crash, but will result in end-users believing that their device does
not support AR. In order to help diagnose those users (and to help track
down a root cause for this crash), Release logging is added as a result
of both null checks, and to indicate that this race likely occurred.

(cherry picked from commit c054e3c1cfee15825cf6693752a110c83033c0fc)

Bug: 1050470
Change-Id: Ib809253093260da396cf82324456c2ff87ae238b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2047362
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#740015}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050672
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/branch-heads/4044@{#228}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}