[Sampling profiler] Use first sample time as profile start time
Now that per-sample timestamps are provided to the ProfileBuilder we
can set the actual profile start time -- the time the first sample
is taken -- rather than using the prior approximation of when the
CallStackProfileBuilder class was created.

This will be used when applying retrospective metadata, to ignore requests
on time ranges prior to the profile start time.

Bug: 1034758
Change-Id: Idf51c4e40dd853408d5f7baee6e9f657d18aa061
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015724
Commit-Queue: Mike Wittman <wittman@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740806}