Moved certain enums to Mojo and added some new ones for use when implementing event intents
This is the first patch that starts the process of implementing event intents for accessibility.
An event intent describes the cause of an accessibility event.
While implementing these intents in
https::/crrev.com/+/2031047/
it became clear that some enums that are used in AXPosition, e.g. AXTextBoundary and AXBoundaryDirection,
would also be useful when defining an event intent.
This patch moves all such enums to Mojo and defines some more that would exclusively be used by
the above patch for implementing intents.
We could always revisit the list of intents, but we'd always need a foundation to start with.

R=dmazzoni@chromium.org

Bug: 989156
Change-Id: If91bafb5b8d237a7f3357ab527eb57275f22a62d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050572
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740935}