Crostini Port Forwarding: Preference tracking (activate/deactivate)
As part of the port forwarding work, when user's interact with the UI
the profile preferences for port forwarding should be updated.

This CL includes the changes to update the preferences when a user
activates or deactivates an existing port as well as unit tests to
support.

Bug: 848127
Test: crostini_port_forwarder_unittest.cc
Change-Id: I42e563056dccd486929f76e757e87429ee8b4847
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2049149
Commit-Queue: Matthew Chen <matterchen@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Cr-Commit-Position: refs/heads/master@{#740990}