[throughput] Change debug info
Right now at ReportFrameEnd, we output a debug info that looks like:
e(3,2). But the unit testing isn't consistent with that. Specifically,
GenerateSequence is doing e(3), because the second parameter is known
from the previous ReportFrameProcessed. This CL changes the unit tests
to be consistent with the code, which takes e(3,2)

Bug: None
Change-Id: I0e1d9f5614c35d3097b4d5eb9bba251aee7b779a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050605
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741026}