[Autofill Assistant] Add List popup section
Adds a new type of additional section which prompts a list
of items for the user to chose from.
If it's a required field and no item is selected an error
message is displayed and the Continue button is disabled.
We technically support multiple selection if specified (although
not needed yet it was not much more complicated to support it rather
than not). However at the moment we only display the first of the
selected items in the ui when the section is collapsed.

Moreover, the results of TextInputSection can also be sent to the
backend now, if specified in the proto.

Screenshots
Collapsed:
https://screenshot.googleplex.com/C9B0DLtT8nU.png
Popup:
https://screenshot.googleplex.com/3OzWxsbDrgY.png

(cherry picked from commit 685e460834294313f209ac6e7b65b109b8aeffe7)

Bug: 1049568
Bug: b/147653652
Change-Id: I6bad58d1a6b5b79b2c2ca0bab5a396dca7b8f313
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2014915
Commit-Queue: Luca Hunkeler <hluca@google.com>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#738617}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2054543
Cr-Commit-Position: refs/branch-heads/4044@{#258}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}