CSP |directives|: convert to map.
It used to be a vector of pair (directive_name, source_list) with no
duplicated |directive_name|. This is equivalent to a map<directive_name,
source_list>.

The main reason it wasn't a map was the ContentSecurityPolicy used to be
transmitted using the old Chrome IPC system. It is now using mojo. Mojo
support map<K,V>.

Using a map slightly simplify the logic, there is no more need to loop
through the list of directives for instance. std::map::find is used
instead.

Description of "future" patches relying on this:
------------------------------------------------

[1] https://crbug.com/888079  Browser process (rather than renderer)
                              should calculate the origin to commit.

[2] https://crbug.com/1041376 Compute the sandbox policy of a document
                              before loading it.

[1], which is blocked by [2]. [2] requires the CSP sandbox directive to
be parsed in the network service before loading the document. The
sandbox directive doesn't contain a CSPSourceList, but a SandboxFlag.
It means we need to store different kind of "Value" for a directive.
This can be achieved by declaring:

union DirectiveValue {
  CSPSourceList source_list;
  SandboxFlag sandbox_flag
};

It could potentially be extended for storing "empty" values for
UpgradeInsecureRequest or TreatAsPublicAddress.
The map<CSPDirectiveName, CSPSource> will become
    map<CSPDirectiveName, CSPDirectiveValue>.

Then we could get access to the Directive value using:
- csp->directives[kSandbox].get_sandbox_flag();
- csp->directives[FrameAncestors].get_source_list();

------------------------------------------------

Bug 1041376

Change-Id: I63979802b15c55452cc8faf13fdeac4c7c29de86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050278
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741117}