NetworkStateHandler: Avoid Matches call to uninitialized networks
Briefly: When a ManagedState entry is added to a list, no initial
properties have been received, so these entries are *mostly* ignored.

In NetworkTypePattern::Matches(), we have a DCHECK to help ensure that
uninitialized states are not used. That means that when we are
iterating through a list of networks, we need to avoid calling
Matches on ManagedState.type_.

This adds a convenience wrapper for places where we need this extra
test, and uses it in UpdateManagedList which is currently triggering
DCHECKs.

TBR=vecore@chromium.org

Bug: 1049808
Change-Id: I1a159622adb24b221849e1fa1a98ce89a3357744
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050759
Auto-Submit: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741121}