Update xr_browser_tests for permissions prompt
Updates xr_browser_test infrastructure to appropriately grant (or deny)
permissions, as well as providing tests with a way to override that.
This ensures that tests will keep running when the permissions prompt is
enabled by default and can be cleaned up as time permits.

Additionally, refactors consent flow tests to a base class and two sub-
classes. These sub classes can be used to explicitly test the custom
consent flow or permissions prompt behavior. A (local) helper macro was
created to indicate which of these tests should be run for both versions
versus which tests were specific only to one or the other. To simplify
this change (and because the permissions flow is runtime agnostic),
consent flow tests are now run against only one runtime.

Bug: 1043241
Change-Id: Ifb9040c46993f8b0375383426a4cafa7e342f6da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2055304
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741172}