CrOS Shelf: Add an 'UpdateLayout' call that was left out previously
The idea behind some of the recent refactorings is to split the logic
for calculating target bounds and upating the layout to each shelf
component instead of doing everything in the shelf layout manager.

Previously, some (limited, soon more) work was delegated to the
shelf widget's |UpdateLayout| method, but calling that method from
the "central" place in the shelf layout manager was omitted. This change
reinstates that call.

Bug: 1051824, 1050383
Change-Id: Ibcf3bfd26210278ec464f186ae6a0a958f03d644
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2055392
Commit-Queue: Manu Cornet <manucornet@chromium.org>
Auto-Submit: Manu Cornet <manucornet@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741176}