Reland "Track DoH probe results per-context"
Original: crrev.com/c/2001144
Revert: crrev.com/c/2050749

Was reverted due to crashes where
|ResolveContext::doh_server_availability_| size was not matching up with
the number of DoH servers in the config in |current_session_|. Because
the config is const in the session and |doh_server_availability_| is
only and always set when setting |current_session_|, the most plausible
theory I have is that the session pointer comparisons were invalid due
to memory reuse (new session but incorrectly recognized as the old
session because it has the same pointer).

Protected the |current_session_| pointer from reuse via WeakPtr. Not
ideal, but DnsSession ownership is already very loose with both ref
counting and other WeakPtr instances. Also added a bunch of CHECK
validations to help spot similar issues if this continues to or later
crashes similarly.

Bug: 1022059
Change-Id: Ibf030fa06882f02830bbaa13e45d9669815e9c5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050758
Commit-Queue: Eric Orth <ericorth@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741180}