chromeos: Convert feedback to use consent aware identity API
SplitSettingsSync will allow the user to opt-out of browser sync.
However, IdentityAccessor::GetPrimaryAccountInfo() defaults to
ConsentLevel::kSync, so it returns an empty struct if the user has not
consented to the browser sync feature.

Feedback isn't tied to browser sync. It has its own consent text in
the feedback dialog.

Switch to using the "unconsented" primary account. On Chrome OS this
account always exists for the logged-in user account, whether or not
the user consented to browser sync.

go/cros-primary-account and go/consent-aware-api-dd

Bug: 1042400, 1046746
Test: Send feedback, report shows up in dashboard
Change-Id: I3f515046994dbb13e826b4c0d62e765ac7025857
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2053128
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741204}