viz: Remove DCHECK for occlusion in content spcae
Initially this DCHECK was placed to prevent an increase in case #3,
checking quads against the target space occluding region. This case is
not a particularly high cost computation and is still relatively
uncommon.

The occluder in conentspace region may be empty if the target space
occluding rect is reduced to 0 after mapping into content space.

Bug: 1051922
Bug: 1051566
Change-Id: Iffd3433a7bf38be9951479c0d8d4c0d5f2b93dde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2053410
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741218}