storage: Migrate IndexedDBContextImpl::Observer to mojo
Devtools was using this interface, and when IndexedDB moves to the
storage service, it will need to access it via mojo instead.
This creates a function for observing IndexedDB changes on the
IndexedDBControl interface and replaces users and tests.

The IndexedDBDispatcherHostTest tests are disabled and were both broken
(identically without this change).  They are updated to at least pass,
although enabling them will be done in a followup to avoid getting this
reverted.

Bug: 1015214
Change-Id: I455d06545e9bbc780173244846b7e801f6bae2cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2052768
Commit-Queue: enne <enne@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741220}