Polish icon drag from applistview to shelf
There are two issues in the current code:
(1) ScrollableShelfView::GetTargetScreenBoundsOfItemIcon is used to
calculate the target bounds of the animation. However, that function
assumes that shelf is horizontally aligned.

(2) ScrollableShelfView::drag_icon_ is created when moving an icon to a
different shelf page. There is an edge case that the user drags an icon
from AppListView and meanwhile drags it to a different page. In this
case, |drag_icon_| is created but never destructed.

This CL fixes the above two issues.

Bug:1052123

Change-Id: I463bb5a672ed309ee674618c36ae743711927739
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2053150
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741246}