ozone/x11: Fix broken software compositing.
We used to be requesting a SkCanvas from the bitmap presenter and
creating own SkSurface for that canvas despite the fact that
X11BitmapSoftwarePresenter has already stored a SkSurface for
that canvas. This resulted in a weird behaviour - frames
were not presented.

So, instead of doing that roundtrip and looking for solutions for
creating a surface storage for that SkCanvas, change the
SurfaceOzoneCanvas to return SkCanvas instead of SkSurface.

This does not imply any behavioural changes.

Bug: 1045347
Change-Id: I09442a6019b1c073f4b1b3407c529792cf74c984
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2017432
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741369}