[NativeFS] Revoke permissions when last top-level tab is closed.
When using the new origin scoped permission model this makes sure to revoke
permissions if no top-level tabs remain with that origin open. Every time
a tab/window is navigated away from an origin we start a timer to check if
that was the last tab/window for that origin, and if so revoke all its
permissions.

Bug: 984769
Change-Id: I3840f1885f3a8794baa09316ab1930068e808baf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2028179
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741406}