Store whether origin is blacklisted in CredentialCache
There are 2 pieces of information that are stored:
1. Whether the origin was blacklisted when the cache was populated (upon
navigating to the page).
2. Whether the origin is currently blacklisted.

The difference between the two is that the user can unblacklist the
origin while being on the page. Thus, the first will be used to determine whether a toggle to re-enable saving for the origin should be shown to the user, while the second will be used to determine whether the toggle is on or off and whether Chrome should show the save password infobar.

Bug: 1044930
Change-Id: Ic5de843ab843337c52502ec718916d39fd42186f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2043841
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741426}