[throughput] Avoid reporting underflow cases
Right now we report throughput to UMA whenever the number of expected
frames are >= 100.

This CL added a condition to prevent underflow. That is, the frames
expected could be decremented to be < 0, but given that the type of
frames_expected is uint64_t, we will get a very large number instead
of < 0.

Bug: 1021963
Change-Id: I2ec718545fa4913e941d2492c622df8144477512
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010183
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741431}