Add presubmit to force existence of expires_after histogram attribute.
A change to histograms.xml was needed to validate the presubmit so
<obsolete> messages had "deprecated" changed to "removed".
"Deprecated" means it still exists (but is in disfavor) when in fact
the histograms have been completely removed.

Removal of an expires_after triggered the presubmit error but the file
was otherwise okay.

Bug: 1030096
Change-Id: I07422aabc6b1fdcf9e65ce1abc792239c865e64c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2052623
Commit-Queue: Brian White <bcwhite@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741432}