Onion soup client-side of prerendering system.
This eliminates Blink public WebPrerender and WebPrerenderingSupport
interfaces in favor of having core/loader talk directly via mojo to
the browser-side PrerenderProcessor. This deletes a bunch of code :)

Move corresponding interfaces to blink/public/mojom/prerender/.

PrerenderCanceler remains defined in chrome/common/ as it is used by
the renderer process that is created to load a prerendered / prefetched
web page. Onion soup of this part could be a future CL.

Change-Id: If7ec0f417930f1a72a515a7fb7ae288a2ba843cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042139
Commit-Queue: Darin Fisher <darin@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741456}