Do not show username in android page info
PageInfo in Android currently shows the full url used to navigate,
including username@ if the navigation included it. This CL changes it
so it omits it. This does not affect the copy function which will
still copy the URL including the username@ part. This matches the
omnibox functionality on desktop.

Bug: 1050011
Change-Id: I3b6794f94774a7300f5afe43262b130c0d534949
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2055806
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Commit-Queue: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741463}