Narrator does not announce deleted characters in some scenarios
When TextRange positions are at MaxTextOffset, NormalizeTextRange will
move them to an equivalent position at the beginning of the next node
via AsTextPositionBeforeCharacter. This can cause issues with
Narrator's logic that determines which character was deleted.

The fix for this is to exit NormalizeTextRange if a caret or selection
is visible. Note that this doesn't fully fix the issue - instead of not
announcing anything, after this change it will announce one character
off from the deleted character. That will require a follow-up change
once we are clear what the expecations from Narrator are.

This change also fixes an issue with HasVisibleCaretOrSelection. It
was not properly accounting for sub-objects within text fields. This
change updates it to use PlatformGetClosestPlatformObject instead of
comparing against the "this" pointer.

Bug: 1049682
Change-Id: Ia46fe47bde33e81f4c615609b90c71783c6b2af3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042214
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#741475}