chromedriver: makes chrome driver write out exe name for weblayer
Prior to this the executable name was an empty string. This is problematic
as the executable name is written to a file followed by the command line
args. If the executable name is empty, it means the first argument is
treated as the executable name, which effectively means what should be
the first argument is not seen as an argument to weblayer.

BUG=1052173
TEST=covered by wpt

Change-Id: Icac6504e7174582a1744d23079b193b6da6a40e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2057085
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741481}