[Signin]: Optimize fetching Image in the Account Fetcher Service.
Add |last_downloaded_image_url_with_size| to |AccountInfo| as well as
|kLastDownloadedImageURLWithSize| to |AccountTrackerService|. The goal is to
have an identifier to the last downloaded image. |AccountInfo.picture_url| can
not be used since it points to the picture related to the account but it does
not necessarily map to the current image, that can happen while fetching a new
image or failure to fetch. The |AccountInfo.picture_url| is set on
|UserInfoFetch| while the |last_downloaded_image_url_with_size| is strictly set
when the image is set.

This is useful to avoid re-fetching the image every refresh period (currently
24 hours) while the user has not changed their account GAIA image.

Change-Id: Ic60c20a9caaf13415c045d11c34ecd9a60fc525d
Bug: 1050643, 1050375
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1834190
Commit-Queue: Monica Basta <msalama@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741489}