Skip popup Tabstrip update when opening tabs from background
Currently, we use AnchoredPopupWindow#onRectChanged to force update
popup strip layout when new tabs are added. This could lead to a race
condition where RecyclerView tries to recycle a view that has been
temporarily detached due to the re-layouting process. This is only
reproducible when there is recycling happening, and when tab is added
from background.

This CL bypasses this crash by disabling the force update for adding
tabs from background. Note that it is already a known issue that strip
is not properly updated when tab is added from background, so this
CL doesn't introduce any UI difference before/after.

Bug: 1045944, 1050401
Change-Id: I4d7ba97f53d41042867bea9562ce2e76aade118b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2048988
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741526}