Remove workaround for Web Components v0 trial for headless clients
In [1], we implemented a workaround to re-enable WebComponents V0 in
headless, regardless if any valid origin trial tokens are present. That
workaround has been merged to M80. In [2], we implemented a fix so that
origin trials are properly supported in headless (that is intended for
merge to M81).

This CL removes the workaround, as it is no longer needed. It also
implements a basic test to verify that trials are working in headless,
and specifically the WebComponents V0 trial.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2039601
[2] https://chromium-review.googlesource.com/c/chromium/src/+/2042382

Bug: 1049126
Change-Id: Ie118f2d53a535ff8cb8f9f1644b866298e57f036
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042689
Reviewed-by: Jason Chase <chasej@chromium.org>
Reviewed-by: Johannes Henkel <johannes@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Jason Chase <chasej@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741632}