Fix committed interstitial auth login for non-default StoragePartitions
Make sure that the credentials returned from the login delegate are
stored in the StoragePartition that the WebContents that requested it is
actually using.

This also adds a browsertest that verifies this for the Chrome OS
gaia sign-in webview, which uses a different StoragePartition from the
default one.

Bug: 1049349, 1051053
Test: browser_tests --gtest_filter=SamlTest.IdpRequestHttpAuth
Change-Id: Ie0a60eef561fe736ffa84ea2edd2b1c649815908
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2056083
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741646}