overview: Revise comment
The present CL revises a comment about why after the split view drag
indicators are updated, the bounds of the virtual desks bar may need to
also be updated. It involves "the top splitview guidance indicator,"
which exists only in tablet mode, because clamshell split view is always
horizontal. It is important to understand that the code only matters in
tablet mode. That is why the code calls MaybeUpdateDesksWidgetBounds
only on the overview grid with the dragged item.

Bug: None
Change-Id: I3f0db698c631021e37a740f4927768d253d55ded
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2056567
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741672}