[iOS] Remove incorrect DCHECK in MainContentUIState.
This DCHECK assumes that every |-scrollViewDidEndDraggingWithGesture:
targetContentOffset:| callback is received following a corresponding
|-scrollViewWillBeginDraggingWithGesture:|.  This is not always the
case though, since sometimes the begin dragging signals is not sent
if the drag begins at the same time as WKWebView events like page
loads or navigations.

This CL removes the DCHECK and updates FullscreenModel to send the
FullscreenModelObserver::FullscreenModelScrollEventStarted() callback
when |scrolling_| is set to true, which occurs when the scroll view is
either dragging or decelerating.  This allows for this callback to be
handled even if the drag begin callback was skipped.

Bug: 1001872, 1048010
Change-Id: I6dec6280ba738ecdd4eb035e1133bae7fe7868ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2057809
Auto-Submit: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741691}