Polish the mouse wheel handling of shelf
There is the legacy code in ScrollableShelfView::HandleMouseWheelEvent:
if the shelf is horizontally aligned, ScrollableShelfView checks whether
it should handle the mouse wheel event and if not it will redirect the
event to Shelf (to affect the launcher visibility). In fact, event
redirection is unnecessary in the current code.

On one hand, for the touchpad scroll (which may be also received by
View::OnMouseEvent), ScrollableShelfView::OnScrollEvent will check
whether the scroll event should be handled here or by Shelf. In
addition, OnScrollEvent is hit before OnMouseEvent.

On the other hand, for the real mouse wheel event, from the perspective
of the expected UI behavior, it does not make sense that
ScrollableShelfView redirects the event to Shelf. Because when
scrollableShelfView receives the mouse wheel event, it implies that the
mouse cursor is within the bounds of ScrollableShelfView.

In this CL, this part of legacy code is removed. Besides test cases
related with the touchpad event and mouse wheel event are created.

Bug: 1052191
Change-Id: I8b42545ad89edba23ebec2b4ae19f4c5dbf2505c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2057095
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741695}