[blink/bindings] Include more details in HighEntropy API list.
When generating the high entropy API list, instead of only showing the
high entropy ones, include all of them with proper annotations so we
know which ones are already marked as high entropy.

The output is also explicitly a CSV file that is now located at the root
of the build tree where it is easily accessible.

To generate the list of Blink JS APIs, use the following ninja
incantation:

    ninja -C out/Debug generate_high_entropy_list

Now look for the file named high_entropy_list.csv in the build
directory.

R=yukishiino@chromium.org

Bug: 973801
Change-Id: I1bf7ee7a48d0b8db0cf2dd284e3557a4fa73a7e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2055685
Commit-Queue: Asanka Herath <asanka@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741760}