Propagate invalidations when FontFaceCache is rebuilt
There are some cases where we need to reappend all stylesheets (e.g.,
when 'media' attribute on a style element is modified), which rebuilds
FontFaceCache. This patch ensures a full style recalc in this
case, so that the invalidations are propagated to all Font objects.

Bug: 1050564
Change-Id: I51c8e000198c1285a29349d278f14e054584878a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2057593
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741803}