[settings logging] Add a timer to prevent brightness event spam.
When the user drags the brightness slider, it triggers many brightness changes at small increments of time. We are more interested in the final change by the user than in each individual brightness change, so this CL introduces a timer as a way to approximate whether or not the user has finished changing the brightnes.

Bug: 1014839
Change-Id: I9486843a941c90abd97ec3e15fa36eb82b7fdf9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2056188
Commit-Queue: Rachel Wong <wrong@chromium.org>
Reviewed-by: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741804}