chrome://settings/passwords: Show badges for local vs account passwords
This CL adds badges to each password and exception item in
chrome://settings/passwords that identify the storage location of the
item (on this device, or in the Google Account).
The badges only show up if the feature "EnablePasswordsAccountStorage"
is enabled (aka chrome://flags/#passwords-account-storage), and the user
is opted in to the account-scoped password storage.

The badge icons and labels are not final.

Screenshots:

Feature disabled or user not opted in (not changed by this CL):
https://screenshot.googleplex.com/jzCLUwcxu3J

Feature enabled and user opted in, BEFORE:
https://screenshot.googleplex.com/B3D7vq5BrgG

Feature enabled and user opted in, AFTER:
https://screenshot.googleplex.com/rY9xGnVf3Xs

Bug: 1049141
Change-Id: Iee3a36651474df9d310e1f801e2ae42284958667
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050381
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741873}