WebSocketStream: Use a hwm of 1 for the readable
Initially WebSocketStream used a hwm of 0 for the ReadableStream that is
returned from the `connected` promise, but that resulted in poor
throughput when the application wasn't continuously reading.

Change the hwm of 1 so that up to one message will still be transferred
to the render process even when it is not inside a read() call.
Backpressure will still be applied if more than 1 message is in transit.

BUG=983030

Change-Id: I883d9e62b8465d45bc9b383db8f639fbf791c78d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2054605
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742029}