media/gpu/test: Improve loading of test images.
When loading test images, we try to resolve the file to an absolute
path. Unfortunately when the path doesn't exist this will result in an
empty path, causing error messages to be printed without the file name.

The CL makes sure the file name is properly printed in error messages.
This CL also improves the logic when resolving file paths, to check
whether the path is absolute, relative to the current path, or relative
to the default test data path.

TEST=./image_processor_test on hana

BUG=None

Change-Id: I8765174e87b25664abc56a2ec3f4ddf4141e60a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2035496
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742066}