[fuchsia] Fix NetInterfaceToNetworkInterfaces() for lack of IPv4 address.
It is possible for a network interface to have only IPv6 addresses, but
NetInterfaceToNetworkInterfaces() was assuming that the non-optional
|addr| field in the fuchsia.netstack.NetInterface would always contain a
valid address, resulting in a net::NetworkInterface with INADDR_ANY
being returned in the interfaces list.

Fix the call to omit the IPv4 NetworkInterface if the |address| IsZero().

Bug: 1052410
Change-Id: I66dc50fcac4911416399210558484130fbfe571c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2060072
Auto-Submit: Wez <wez@chromium.org>
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742067}