Remove blink::LayoutListBox
LayoutListBox provided
 * Intrinsic height, and
 * Adjustment of minimum width.

For the former, this CL folds it into LayoutBlock and NGBlockNode
because it's small.

For the latter, this CL moves it to LayoutBox, and
ClampIntrinsicBlockSize(), NGLayoutInputNode::ComputeMinMaxSize(), and
LayoutBox::ComputeLogicalHeight() call it because the code size is not
small.

This CL has no behavior changes.

Bug: 1040828, 1040826
Change-Id: I9dd95c0971928a50895c32fa977668421078e780
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2053648
Commit-Queue: Kent Tamura <tkent@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742068}