Fix flags of login-screen policy extensions
This removes the flags that were incorrectly specified for the
force-installed login-screen extensions/apps (installed via the
DeviceLoginScreenExtensions policy) :

* The FROM_WEBSTORE flag wasn't necessarily true, since the policy can
  specify any update URL - so it can also point to a self-hosted CRX
  file. (In particular, this is used in a few browser tests.)
* The WAS_INSTALLED_BY_DEFAULT flag seemed to be not suitable as well,
  since the policy-installed extensions aren't "default" (unlike, for
  example, the "offline Docs" extension).

This CL is expected to have no functional effect for the existing
scenarios for the login-screen policy apps/extensions (e.g., smart card
login or Imprivata login). However, the CL makes the flags more precise,
and prevents problems with future code if it tries to make some
decisions based on the presence of these flags.

Bug: 888172, 1051513
Change-Id: Ibdb70ff664b462edf660a21db0451f22d4632803
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2049855
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742098}