[Autofill Assistant] Introduce a BROWSE state.
This CL introduces a new AutofillAssistantState called BROWSE, which
allows users to browse around on a website while the AutofillAssistant
waits in a prompt-like state. Currently, this in implemented as a
browse mode (bool) on PromptAction, but it could be refactored to its
own action in the future. Gestures such as GoBack do not terminate
the BROWSE state like they do for PROMPT. However, navigating to a
different domain will terminate the AutofillAssistant.
Preconditions are still checked, but there is no overlay or touchable
areas in browse mode. Also, navigation and user gesture filtering as
applied when in PROMPT is not applied for BROWSE so that autofill
assistant does not terminate in those cases.
Also adding unit tests for both PROMPT state and BROWSE state for
the controller, script executor, and prompt action.

This is based on a prototype by Mathias Carlen <mcarlen@>.

Bug: b/147412447
Change-Id: Ieeee6207ab72bec38363ddc9dee031602ea78e6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2056726
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Marian Fechete <marianfe@google.com>
Cr-Commit-Position: refs/heads/master@{#742116}