Increase initialization screen contrast on Kiosk app
The contrast ratio should be at least 4.5 on all text. On the initialization screen of the Kiosk app it was too low on one of the strings. The color of one string was changed from gray to #666 (same as the other strings on that screen) which increases the ratio from 3.94 to 5.74.

Bug: 1015832
Change-Id: I53f14c5f7831b9a7c54d817e2b28d1a277c59954
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036054
Commit-Queue: Anina Koehler <aninak@google.com>
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742137}