[iOS] Create separate find in page command protocol
This separate protocol is in preparation for moving the find bar
coordinator usage code from BVC to BrowserCoordinator. Once there's a
separate protocol, the BrowserCoordinator can implement it instead of
the BVC.

From this, most of the code here is passing the new protocol through
various levels of command handler.

Bug: 1052818
Change-Id: I0d1d3be56632d21851545a47ffb9e6550729d237
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2060495
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742202}