[ExpandablePaymentHandler] Set full height to be 90%
Context:
The full state of the Expandable Payment Handler should leave some room
above the sheet, in order for users to see the main tab's toolbar.

Before change:
The height of the full state of the PH bottomsheet covers the whole
tab.

After change:
The height of the full state of the PH bottomsheet occupies 90% of
the tab height.


Bug: 1052158

Change-Id: Id683da48a10119e97a935aabf7dc06ed5a6b69bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2058909
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742211}