Fix feedback for pinning app under RTL
Basically two elements are required by View::ScrollRectToVisible:
the rectangle to show and the bounds of the visible space. However,
in ScrollableShelfView::ScrollRectToVisible, the rectangle to show
is not mirrored under RTL while the visible space is mirrored.

This CL mirrors the visible space from RTL to LTR when RTL is enabled.
A test case is also created for this scenario.

Bug: 1052553
Change-Id: I3a051c40a87f93f612edf558944479aad61d6ca2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2056938
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742221}