Update WebXr tests for permissions prompt
Updates the Android WebXr test infrastructure to allow accepting or
rejecting permissions prompts in addition to the consent dialog, based
on whether or not the feature is enabled. This ensures that tests will
keep running when the permissions prompt is enabled by default, and can
be cleaned up as time permits.

Additionally, consent prompt tests that still make sense to run when
the permissions prompt is enabled were pulled into an impl method and
a variation of the test was created which explicitly sets the feature
to the desired state. The old tests can thus be more easily removed when
the feature flag is removed. This ensures both some coverage of the
permissions prompt flow now, and continued coverage of the consent flow
after it is disabled, until such time as it can be removed.

Some tests were determined to be testing logic specific only to the
current consent flow, and as such those tests forcibly ensure that the
permissions prompt is disabled.

Bug: 1043241
Change-Id: Ib61d81a3f3aa90b1869d13ad53112ecd5606b627
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2051367
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742234}