Origin policy: add disconnect handler
This CL adds a URLLoaderFactory mojo disconnect handler for
OriginPolicyManager to recover its URLLoaderFactory, after factories are reset.

This CL also adds a browser test that exercises the disconnect handler,
demonstrating that it plays a critical role when e.g., an extension with
the declarative web request permission is loaded in the browser, and the
NetworkContext's URLLoaderFactories are reset.

Co-authored-by: Domenic Denicola <domenic@chromium.org>
Bug: 1047275
Change-Id: Iaa978348d069dfd61bc157ddbe159b1766a0e7ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2053228
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Domenic Denicola <domenic@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Domenic Denicola <domenic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742253}