viz: Refactor BeginFrameArgs creation in BeginFrameSource.
The sequence numbers in BeginFrameArgs are expected to be incremented
based on vsync intervals. Refactor the code for this from
DelayBasedBeginFrameSource and ExternalBeginFrameSourceAndroid into a
BeginFrameArgsGenerator.

This is a simple refactor of the code, and does not change behaviour.

BUG=none

Change-Id: Idb51e1306dc6e6c64acbe2027cc28df0ba440055
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2058032
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Auto-Submit: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742256}