Prepare to migrate Tab#getUrl users to GURL (3 of 3)
Going through each of the usages will take some care, so I want to
essentially fix the API in a no-op pass and properly migrate each of
the callsites independently (or in batches).

I'm still working on migrating users of WebContents#getVisibleUrl, but
some to the APIs those usages call into are shared by users of
Tab#getUrl, so exposing a GURL Tab#getUrl early makes this less painful
and requires less duplication and converting.

This will be a 3-sided patch:
1. Add the getUrlString() method and migrate upstream usage.
2. Migrate downstream usage to getUrlString().
3. Change getUrl() to return a GURL (this CL).

Bug: 783819
Change-Id: If841d509bff8ea9d2da048021da6ba8635a8cd1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044379
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742257}