views: ax-ignore MenuScrollViewContainer
Right now, MenuScrollViewContainer claims to be an AXMenuBar, which
produces the following puzzling AX tree:

  AXGroup (menu root)
    AXMenuBar (scroll container)
      AXMenu (actual menu view)
        AXMenuItem ...

An AXMenuBar is something like the top system menu bar, at least on
macOS, not the "contents of a menu" or anything similar to that. After
this change, the AX tree is instead:

  AXGroup (menu root)
    AXMenu (actual menu view)
      AXMenuItem ...

which makes much more sense.

Bug: 954701
Change-Id: I00215874d75d9ee4870acd0e2101ef280dc2a199
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962587
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742311}