Clear the subresource redirect hints when navigation is started
RenderFrames could be reused for same-origin navigations. This causes
the hints from previous navigations to be used in next navigation
until the new hints fetch finishes and updates the hints. This CL clears
the hints.

Unfortunately this causes race conditions between the image fetch and
the hints update, which causes images to not redirected to compressed
versions. So the image fetches are delayed a bit to circumvent this.

Bug: 1051283
Change-Id: Ie9e5351e68c0081ab9dbfe23faed2c3c5e8e4d42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2056207
Commit-Queue: rajendrant <rajendrant@chromium.org>
Reviewed-by: Michael Crouse <mcrouse@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742326}