viz: Add a flag to do double-buffering compositing
Chrome on ChromeOS uses triple buffering by default, keeping up to 2
buffers in flight while one is displayed.

This CL adds a flag in chrome://flags that limits to 1 the number of
pending frames, effectively reducing compositing to double-buffering.

This can be useful for experimenting on devices where we want to
minimize latency at the cost of more dropped frames.

Bug: 1010457
Change-Id: I0cddb4c1831d5b909ee3c0193d42f3903e10d03a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2059709
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742351}