Refactored HoverButton to use layout rules.
Updated HoverButton code such that it will responsively update its
own and its children's layouts. This is in contrast to its original
behaviour in which layout was set based on the preferred sizes of
it's children at construction time.

This helps towards having HoverButtons with icons that update to
match the color of the Widget's current theme provider. Such
changes could incur changes in the icon's preferred size and
could result in incorrect behavior in the old code.

Also cleaned up unnecessary and unused code and fixed broken logic.

Bug: None
Change-Id: I502c32ea673d8f680f38a45ce2940e155872dd40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2045145
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742352}