Adjust width of action bar search box for files-ng
Changes the width of the search box to expand the containing search
wrapper to 288px.

The value used in the CSS calc property 60px is the sum of 48px for the
search button plus 12px margin to the right of the search input field
cancel button.

Note, the width of the search button margins may be reduced in a future
CL subject to visual evaluation.

Bug: 992823
Change-Id: I372df330a5bb446c3430dcd85c0582bd3a3d9b01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2060401
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742358}