extensions: Remove AppDataMigrator
The unit test is not testing anything (it both early outs on copying
files and also uses the same source and destination directory when
checking if files have been copied).  This is also very deprecated.
Rather than keeping this working[*] and updating the test to not
be broken, instead remove this functionality that is incredibly
deprecated.

A followup will be to remove the functions in IndexedDBContext that
this is the only user of.

[*] alternate approach:
https://chromium-review.googlesource.com/c/chromium/src/+/2020533

Bug: 747589, 1015214, 302577
Change-Id: I7ac05f966c02be39fcd05788b1e562efae214825
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2048032
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Ben Wells <benwells@chromium.org>
Commit-Queue: enne <enne@chromium.org>
Auto-Submit: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742367}