Add missing DesktopCapturer methods.
DesktopCapturer is not a pure virtual interface, meaning that wrapping
implementations such as DesktopAndCursorConditionalComposer need to be
updated whenever a new method is added--the compiler will not give an
error in this case. This was the cause of the multi-monitor breakage
for relative pointer mode.

Change-Id: I5a37a0e1fc47bf72cf9ccc7810c9e17d0a418237
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2063412
Commit-Queue: Jamie Walch <jamiewalch@chromium.org>
Commit-Queue: Gary Kacmarcik <garykac@chromium.org>
Auto-Submit: Jamie Walch <jamiewalch@chromium.org>
Reviewed-by: Gary Kacmarcik <garykac@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742384}