Cleanup WebXr flags to match support
Remove flags for shipped/shipping features:
* WebXr Sandbox
* WebXr Device API
* WebXr AR Module
* WebXr AR Hit Test

Remove webxr-runtime choices for runtimes that we do not intend to
support natively given that OpenXr will allow for support:
* OpenVr
* Oculus

Removes "Orientation Sensors" choice from webxr-runtime as it is inline
only and doesn't really make sense in the list for android, and will
not be supported on Desktop.

webxr-runtime is kept to ensure that tests can be run on both Windows
Mixed Reality (WMR) and OpenXr. As usage of WMR is expected to decline
in favor of OpenXr, this flag will be removed when WMR can be removed.

Bump expiry of webxr-incubations flag to M85 to reflect ongoing usage.

Note that this does not remove any features, so --enable-features,
--disable-features, and --force-webxr-runtime= flags will still allow
these features/options to be set.

Fixed: 1049797
Change-Id: I54eaff1cfcf6ecebc574113759dc06c806bcd0f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2063535
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Mounir Lamouri <mlamouri@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742407}