Crostini Port Forwarding: Remove active variable in pref
Due to the spec change of not forwarding "active" ports on container
start up, we have removed the active state in prefs as it is not
required anymore.

This change removes the active pref, resulting in changes to fit the
new context such as: on deactivate all ports, we just need to remove
ports from our lifeline fd map.

Bug: 848127
Test: crostini_port_forwarder_unittest.cc
Change-Id: I19ad693682122f1edd47c8a21fff5b69780fb4c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2063191
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Matthew Chen <matterchen@google.com>
Cr-Commit-Position: refs/heads/master@{#742483}