Refactor LayoutSVGResourcePattern::PreparePaintServer/BuildPatternData
Change the structure of these pair of functions such that the former
handles the cache lookup and calls the latter to mint a new PatternData.

Remove the redundant check for an empty 'viewBox'.

LayoutSVGResourcePattern::PatternForClient is removed, because the issue
mentioned in the FIXME should no longer be present.

Bug: 109212
Change-Id: I5514dce39c20e07adbf45ea46c2dd472a4c71303
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2061752
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742562}