Add support to override ProfileSyncService in SyncTestRule
SyncTestRule doesn't provide a way to override the instance of
ProfileSyncService it holds. For this reason some tests might override
the ProfileSyncService with a subclass but SyncTestRule will hold a copy
of the old ProfileSyncService instance. This cl adds a method in
SyncTestRule to override ProfileSyncService and refactors some tests
that need to use it.


Bug: 1030725
Change-Id: I9fc1b770df4c153e726a4e071e432fc585561889
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2012380
Commit-Queue: Tanmoy Mollik <triploblastic@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742596}