[Android] Photo Picker: Remove deficiencies when flinging.
Removing an image decode from the queue should not also delete info
about which decode is currently in progress, since that causes the
queuing mechanism to think new requests should be sent for decoding.
This adversely affects flinging, because requests that had been queued
up (but are about to be deleted because the user has already flung past
them) would be sent for decoding when new entries are added.

This change also removes the Map of current requests, which may give
off the impression that we're switching from parallel to serial
decoding. However, that is not the case since it was serial to begin
with -- we currently only fire off one request at a time, so a Map is
not required to track what is in progress.

Bug: 1051925, 895776, 656015
Change-Id: Ieaef6d905148eb67d813de397d1582ca151aca79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2054249
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742626}