[message_events] Notify GC of ArrayBuffer memory owned by MessageEvent
With the recent ArrayBuffer refactoring, the backing store memory is
announced to the GC when it gets wrapped by a JSArrayBuffer object. The
assumption is that backing store memory is always wrapped by a
JSArrayBuffer object. However, the bug report shows one example where
this is not the case. When an ArrayBuffer is post-messsaged to a Worker
but the Worker never takes the ArrayBuffer out of the message, then
the backing store memory becomes dead before it ever gets wrapped by a
JSArrayBuffer object. Therefore the GC does not know about the memory
used by the backing store and does not trigger garbage collection.

With this CL the MessageEvent now announces the backing store memory to
the GC. With that, garbage collection is triggered, and the backing
store memory is collected as expected.

R=haraken@chromium.org

Bug: chromium:1049005
Change-Id: I517f65a4cb1546ae469be29514183314f62e3d42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044272
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742675}