extensions: Make destination for base::PostDelayedTask() explicit
When posting tasks with base::PostDelayedTask(), explicitly specify whether the
tasks should run on the thread pool on a specific named thread. This is in
preparation for making it always mandatory to specify the destination for posted
tasks.

This patch has no functional changes.

This CL was uploaded by git cl split.

Bug: 968047
Change-Id: I9595134c52d067fc61f8d3c705fa7f21db18b10e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2063168
Commit-Queue: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742708}