Make TabListMediatorUnitTest independent
http://crrev.com/c/2036513 introduced a test dependency. It requires
TabAttributeCacheUnitTest to run before TabListMediatorUnitTest, or
TabListMediatorUnitTest would fail.

This was due to different handling of shared preference in these two
tests. TabAttributeCacheUnitTest treats shared preference as a real object,
while TabListMediatorUnitTest uses a mocked one. In this CL,
the shared preference in TabListMediatorUnitTest is converted to a real
object.

After this fix, when running in a per-class way as described in
https://crbug.com/1032883#c3, there are no failures.

Bug: 1032883, 1048255
Change-Id: I73d7bf723b3d039e3b494abb5d5bd0013d710ef0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2060909
Reviewed-by: Yue Zhang <yuezhanggg@chromium.org>
Commit-Queue: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742740}