Only show paused dialog for active apps that reached daily limit
We were showing paused dialog for every app that was stopped when
time limit was reached. This was resulting in bad user experience when
Chrome was reaching the limit, because all installed web apps are paused
then and user was seeing multiple dialogs (one per every web app).
Now we show the dialog only for active app. When web app is active
there will be one dialog displayed. When Chrome is active there will be
no dialog, but the interstitial page. This should not change anything
for ARC++ apps.

Bug: 1052040
Change-Id: If4b4827ffb6dd7d022170c6503bf8e109c4bfb94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2063219
Commit-Queue: Aga Wronska <agawronska@chromium.org>
Reviewed-by: Yilkal Abe <yilkal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742745}