Refactor WptMetadataBuilder to loop over tests to lookup statuses.
Previously we'd ask typ for all "Skip" expectations and then loop over
all tests to look up their baselines. In a future CL we'll need to look
at other (non-Skip) expectations for all tests as well, and since we're
already looping over tests to find baselines we might as well query the
expectations at the same time.

This also changes the return type of "get_tests_needing_metadata" into a
dict (rather than list of pairs), which makes it easier to attach new
more statuses to a test as we process more input.

This change should functionally be a no-op.

Tested: built metadata with and without this change and recursively
diffed. Validated that changes make sense.

Bug: 937369
Change-Id: I80c68c6326f20c864c4b1c6334116c5d2c0498ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024457
Commit-Queue: Luke Z <lpz@chromium.org>
Reviewed-by: Robert Ma <robertma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742749}