Move PermissionRequestImpl to //components/permissions
This required moving all the strings and icons that are referenced in
PermissionRequestImpl as well. kProductIcon which is used for the
PROTECTED_MEDIA_IDENTIFIER on ChromeOS was left in //chrome since
kProductIcon is different depending on chrome branding.

The IDR_* resources used on Android were moved in a way that still works
with chrome's ResourceMapper. WebLayer will need a similar system to
map the native IDR_* constants to Android drawable resources.

Bug: 1025609
Change-Id: I21724ce14a006b89e30f027971165a40fce85177
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2055888
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742750}