cbuiv: fix DesktopMedia tab list crash
When the tab list in the desktop media picker has keyboard focus, it
interprets a Return key press as meaning "accept selected source", but
table views don't always have a selection. This can cause a crash in
DesktopMediaPickerViews itself, which assumes that its Accept method can
only be called with a source selected.

To fix that, have DesktopMediaListController reject calls to
AcceptSource when there is not actually a selected source. Also, add a
unit test for this case.

Bug: 1042976
Change-Id: I840cf4353f59f33a6950b580196f34a60d04c448
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2063044
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742832}