bluetooth: Intercept navigation in browser test
The web_bluetooth_browsertest.cc test cases are flaky because the
EmbeddedTestServer is initialized with a hardcoded port, and other tests
in the test suite can potentially use the same port number. This causes
the test cases to fail.

The port was hardcoded because the test cases that check Web Bluetooth
permission persistence require that the origin of the test page be the
same across browser restarts.


This change updates WebBluetoothTest to intercept navigation requests
with URLLoaderInterceptor to dynamically serve the local test page
without the use of EmbeddedTestServer. This enables the test to maintain
the same origin across test cases while still navigating to the test
page.

Bug: 1053464
Change-Id: I75f50c909030b59a9bbcda76038ac664abe8b513
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062899
Commit-Queue: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742844}