Create a Zygote which starts unsandboxed for utility processes.
This ensures that utility processes don't end up starting with a
different on disk binary than is currently running.

This works by adding a new "--no-zygote-sandbox" command line flag
which is set in ZygoteCommunication based on a construction flag. It's
used to launch the zygote without the setuid and namespace sandboxes;
which sounds scary, but is already the case today for utility processes
with specialized sandboxes (net, ime, audio, soda).

Within the ZygoteMain this functions as if "--no-sandbox" was specified
to the Zygote and skips its sandbox setup. The actual sandbox is later
applied by UtilityMain().

Bug: 22703,1049234
Change-Id: Ib61e7c71c6ee43e2ab518fa0ed109bea36677bff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2047932
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: John Abd-El-Malek <jam@chromium.org>
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742875}