Fix wide color gamut on Pixel 4
The Pixel 4 has wide color gamut enabled. However, a CL [1] accidentally
caused the Pixel 4 to look muddy. This was because it no longer set the
color space on the texture.

For Android, when using the scanout usage for shared images, we always
use the AHardwareBuffer backing. The GL backing was setting the color
space, but the AHardwareBuffer backing was not. And in our wide color
situation, those shared images were using scanout.

This CL sets the color space of a shared image with AHardwareBuffer
backing.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1977969

Bug: 955158
Change-Id: Id811457e6198ba2b9cbe9667817edc385aa5ef4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2065430
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: vikas soni <vikassoni@chromium.org>
Commit-Queue: Chris Blume <cblume@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742897}