[throughput] Fix a DCHECK crash
Currently we hit a DCHECK in report histogram. The problem is that
for certain type of throughput data (such as CompositorAnimation +
main thread), we do not report it. In this case, we do not reset the
throughput data and it kept accumulated and eventually the number
of expected frames reaches a threshold and caused crash.

This CL fixes the crash and added test.

Bug: 1052848
Change-Id: Ib150872f950f0901a6bce5f10e683e08644c5809
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2061398
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742947}