[system-web-apps] Allow chrome-untrusted schemes to be embedded in WebUI pages.
We want to embed an iframe with a URL of scheme 'chrome-untrusted://'.
Currently, embedding chrome-untrusted:// iframes is blocked in multiple
areas: the WebUI navigation throttle in the browser side and in the
renderer side which currently treats chrome-untrusted as a display
isolated scheme. This means it expects only the same scheme can load the
same scheme e.g. chrome-untrusted:// can only load chrome-untrusted://.
Since we want chrome:// to load chrome-untrusted://, it makes sense to
no longer treat chrome-untrusted as a display isolated scheme.

This cl unblocks those 2 areas (though there are more blocks we need to
remove) by deregistering chrome-untrusted as a display
isolated scheme and we allow chrome-untrusted schemes to proceed in the
WebUI navigation throttle.

Bug: 1048951
Change-Id: I9bfce2909f6a25c535866d1b8d57507393429bb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2038192
Commit-Queue: Jessica Huang <jessing@google.com>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: calamity <calamity@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742982}