[system-web-apps] Add chrome-untrusted iframe to sample system web app
An example of how to embed a chrome-untrusted iframe into a system web
app.

To do so, we need to add abstract methods in WebUI despite WebUIImpl
being the only class to actually implement it. This is so the
WebUIController can add requestable schemes so the WebUI can request
the chrome-untrusted resource.

Bug: 1048951
Change-Id: Ia3152e21412845cf421df74bb00a16e0a8f264b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2037186
Commit-Queue: Jessica Huang <jessing@google.com>
Reviewed-by: calamity <calamity@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Auto-Submit: Jessica Huang <jessing@google.com>
Cr-Commit-Position: refs/heads/master@{#742999}