Files app: Change directory to update expand icon instead of full rescan
Change TreeItem classes to avoid doing a full rescan just to work out
if we should show the expand icon or not. Instead, only scan down to
the first sub-directory using the new method updateExpandIcon().

Replace calls to updateSubDirectories() with updateExpandIcon() where
the content of sub-directories aren't needed. The most important call
replaced is in SubDirectoryItem constructor, because this caused
scanning of an additional level down from the folder that user was
interacting with.

This significantly improves the performance of navigating on folders
that have large sub-folders. For example, the time to run integration
test directoryTreeExpandFolder() cut down from ~40 secs to ~7 secs,
where the most of time is spent expanding a sub-folder that contains
1k children.

Test: browser_tests --gtest_filter="DirectoryTree/FilesAppBrowserTest.Test/directoryTreeExpandFolder"
Bug: 1039820
Change-Id: Ib629601f2c795a6dedd80a04f85827fd33e96af0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2059729
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Auto-Submit: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743011}