Block VMware Virtual Webcam from MediaFoundation. Take 2.
With MediaFoundation, VMware exposes a camera even if no physical camera
exists. The behavior of this camera is problematic and causes getUserMedia()
calls to hang.

A previous CL (crrev.com/741120) tried to block this camera, but it had
no effect because the blocking mechanism worked only on USB cameras.
This CL introduces a blocking mechanism that works on the display name for
the camera.

Local testing has shown that under DirectShow, the virtual camera is not
exposed if no physical camera exists. It works well when a physical camera
exists too.

Bug: 1044974
Change-Id: I544e33ea4a5717dd91d09784a1bc32e3e268cf0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2064996
Reviewed-by: Armando Miraglia <armax@chromium.org>
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743037}