Passwords account storage: Don't use or offer for Sync users
If Sync-the-feature is enabled, there is only a single, combined storage
for passwords. So we should not offer the account storage in this case,
and if the opt-in already exists it should be ignored.
This CL adds the necessary "!IsSyncFeatureEnabled()" checks plus
corresponding tests.

Bug: 1024332
Change-Id: I496ecb902297983d51526a74211ff1b754376e00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062981
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743058}