Fix input autofocus in smart card SAML PIN dialog
This fixes the regression that the input field wasn't auto-focused when
the smart card PIN dialog (<security-token-pin>) is shown during the
SAML login.

The regression was introduced in https://crrev.com/c/1968989, which
erroneously changed the focus to be set synchronously. This didn't work,
because in practice this happens too early, while the dialog still has
the "hidden" attribute - because Polymer doesn't guarantee the order of
updates triggered by the changes in |pinDialogParameters_|.

The fix is to schedule the focus assignment to be done asynchronously
(soon), so that this happens after Polymer updates all properties
synchronously.

Bug: TODO
Test: enroll Chromebook under domain with smart card settings, start SAML login via the smart card, wait until the PIN dialog appears, verify that the PIN input field is auto-focused
Change-Id: I2f1b4961867f735ac00290cd0cb62cb10fc3d33d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2047167
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743168}