(Hopefully) Fix flaky ChromeNavigationBrowserTest.RedirectErrorPageReloadToAboutBlank.
This CL tries to deflake
ChromeNavigationBrowserTest.RedirectErrorPageReloadToAboutBlank. The
theory is that the extension web request listeners aren't getting set up
by the time the page is reloaded, hence the redirect is not happening.
Fix by waiting for an ack from the extension background page.

BUG=981295

Change-Id: If1171578dcbf248b910cbb6f774491ccb348f2ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2056942
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743241}