Clipboard: Check permission and feature flag in browser process (M81)
Adds a check in the browser process for Clipboard permission and feature
flags being on.

NOTE THAT THIS IS NOT A CHERRY-PICK. This is a smaller diff, for the
sake of reducing complexity for backporting. It will have slightly
different behavior from the equivalent change on master, (only 1 mojo
interface instead of 2, frame-scoping in renderer is unchanged, and
failed calls will silently fail), but it will fix 1034023 in a smaller,
easier-to-review CL and fixes the bug in question correctly.

Bug: 1034023
Change-Id: I259a9fc3338b9b7ff920f7e7ba3fa430a3bd841f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067446
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Darwin Huang <huangdarwin@chromium.org>
Cr-Commit-Position: refs/branch-heads/4044@{#389}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}