Eliminate task executor duplicate code.
For UpdateApps we did not need handling of native UI messages
since the execution flow is silent. I also fixed the IWYU headers.

For InstallApp, there were two SingleThreadTaskExecutor objects
being instantiated, which is confusing. One of them was eliminated and
replaced with a SequencedTaskRunner. The thread checker was
replaced with a sequence checker.

The UI code has thread affinity and continues to use a thread checker.

Change-Id: I43ef545a7b86ff7f8cedea7b922cd84b0c7bbb99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067686
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743360}