Merge to 81 for month live update and month switching selection update
This is a combined merge of two changes to 81, with image baseline merge
conflicts resolved.  Initial change descriptions follow:

Change #1:
Implement live-update and double-escape for month input popup

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2029243

This change implements live-update and double-escape behavior for the
month input popup.  When arrow keys are used to change the
selection, the changes are immediately reflected on the in-page control
(and input events are fired).  When the user hits escape, if they have
made changes these are reverted without closing the popup.  A second
escape (or the first one, if no changes were made) closes the popup.

Behavior when clicking on a month is not changed; the popup is still
submitted immediately (as opposed to other types like datetime-local
where the user may want to go on to select other values after clicking
the date).

The selection/highlight model is changed so that using the arrow keys
moves the selection instead of the highlight, invalid dates cannot
receive selection, and the highlight is only used as a hover style.

Initialization of the control is changed such that if there is any
valid month, one will be selected, avoiding scenarios where we don't
paint a focus ring because there is no starting selected month.

Fixed a bug where months invalid due to the step attribute were not
greyed out.

The above changes required a lot of the same valid-date-range-seeking
functionality already implemented in CalendarPicker, so I removed the
side-effects from these functions and moved them into a mixin consumed
by both CalendarPicker and YearListView.

All of above behavior changes apply only to the standalone month picker;
the month selector embedded in date/week/datetime-local popups is
unchanged.  I'm not sure that live-update and double-escape makes as
much sense in those contexts, although we could consider a later
follow-up to change the selection behavior to match the new standalone
picker's behavior.


Change #2:

Update selection when changing month in datetime-local/date/week popups

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2040550

Prior to this CL, when changing the displayed month in
datetime-local/date/week via the month switcher menu or the prev/next
month buttons, the selection did not update.  This caused an issue for
keyboard selection: if the user changes the month via keyboard and
then tries to adjust the selected day, the calendar will jump back to
the old month because the selection is still in the old month.

This change fixes the issue by moving the selection to the new month
whenever the month is changed.

This change also fixes the 'm'/'y'/'d' hotkeys, which were not pushing
updates to the in-page control.y

The (benign) accessibility tree changes are due to the addition of the
click handler to the popup document's body element.

(cherry picked from commit e243a8618c1e79439b1a58c2c5e1f112a477f354)

Bug: 1045195,1047439
Change-Id: I024d3cbe899dee555eca42200e9d594ce994cd3c
Cr-Original-Commit-Position: refs/heads/master@{#737203}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067262
Reviewed-by: Ionel Popescu <iopopesc@microsoft.com>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/branch-heads/4044@{#396}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}