[video-raf] Use latest frame metadata
Currently, video frame metadata is reported back up from the compositor
to the blink layer as soon as a new frame is presented. Since video-rAF
callbacks have their execution delayed until the next rendering steps,
by the time the callbacks are run, there is a chance for a newer frame
to have been presented. This means that the metadata we have would be
referring to the frame that is now being displayed on screen, instead
of current frame.

This CL removes the pushing of the metadata, and instead pulls it at the
last moment, right before the callbacks are run. This means we are more
likely to get metadata that is relevant for the upcoming v-sync, instead
of the one that just happened. However, we might miss more frames, as
measured by non consecutive "presented frames".

Bug: 1049761
Change-Id: I544d190a477f79c9d917a65b781ae68cd099dc37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042143
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743400}