[Suggested Files] Ranking model
Added a new ChipRanker type, which will house a new Dolphin model.
This model will rank apps alongside both local and drive files, which
will allow us to compare the different types of items.
These ranks will be used to adjust the scores of the files to be
in line with the app scores, while still preserving the original
ordering of both groups.
This ensures the files will display appropriately in the suggestion
chips, but that their ranking elsewhere will be unaffected.

Bug: 1034842
Change-Id: I618e2694b74c6c5917eef66a900ef822f5422b92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1989742
Reviewed-by: Tony Yeoman <tby@chromium.org>
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Eleni Dimitriadis <edimitriadis@google.com>
Cr-Commit-Position: refs/heads/master@{#743401}