Allow OffscreenCanvas to access the high-performance GPU.
Instead of only checking whether the containing frame is same-origin
with respect to the top-level document during
transferControlToOffscreen, check in the OffscreenCanvas constructor,
both for the main thread as well as dedicated workers. Dedicated workers
are same-origin per specification, so always grant access to the
discrete GPU in this scenario.

Add new pixel tests that verify WebGL-rendered OffscreenCanvas into an
ImageBitmapRenderingContext (i.e., where the OffscreenCanvas was not
originally tied to a canvas element). On dual-GPU macOS devices, also
verify the active state of the low-power and high-performance GPUs.

Simplify some of the tests most recently written for low/high-power GPU
switching.

Bug: 1048892
Change-Id: I85addc2b53a58997499e444b357d433d7082e262
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2043076
Commit-Queue: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: James Darpinian <jdarpinian@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743419}