Add "Move to another window" tab context menu command
Support moving tabs between different open windows via the tab context
menu, allowing this to be done using only the keyboard. The option
replaces the "Move tab to new window" entry if there is at least one
other window that can be moved to. Inside the submenu, there is an
option to create a new window or to choose any existing windows of the
same profile.

A new class, ExistingWindowSubMenuModel, represents the submenu. It
requests the list of existing windows from the Browser, which it
receives as a list of titles. Once selected, the index of the chosen
browser is passed back to Browser, which references a stored list of
WeakPtrs.

The windows enumerated in the submenu are each calculated titles of the
form "PAGE_TITLE and x more pages" that fit within an allotted pixel
width. First the length of the "x more tabs" extension is calculated,
and the title is elided based on how much width is left from the default
400 pixels. In addition, the title is always guaranteed at least a few
characters if the extension is extremely long.

Bug: 1040582
Change-Id: I9e5c2ba1a1e7f850126df019a9bc2a5b1de842e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026265
Commit-Queue: Justin Gallagher <jugallag@microsoft.com>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743436}