CrOS Gallery: Workaround new treatment of EXIF orientation in m81.
The Chrome OS Gallery is affected badly by Blink's new treatment of
EXIF orientation for <img> and <canvas> elements. With crrev.com/741635
(and crrev.com/725393), it's now necessary to add the "new Image()", and
the <canvas> element it is drawn into, to the page temporarily.
Without this, <img> and <canvas> elements attempt to respect any EXIF
orientation in the source bitmap. This conflicts with logic in
Gallery to support the earlier behavior of these elements.

This CL sets attributes to disable the new treatment, and performs the
necessary DOM machinations to ensure these attributes take effect.

The approach should be robust enough to function as expected whether
or not the new EXIF orientation treatment is actually taking effect.

Skip presubmit: A linter complains about changes outside the scope
of this CL. (OWNERS approval still sought).

No-Presubmit: true
Bug: 1043561
Change-Id: Ib58c8bed9dd5589e3908a8336b4446eeb523170a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2065688
Commit-Queue: Trent Apted <tapted@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743438}