Clear-Site-Data must uncontrol affected service worker clients
Adds an immediate flag to unregister jobs for service worker
registrations.  When set, the immediate flag enables unregister jobs to
run the "Clear Registration" algorithm immediately without waiting for
the active worker's controllees to unload.  After an immediate clear,
the registration's controllees become uncontrolled.  The controllees
receive the 'controllerchange' event when the
navigator.serviceWorker.controller is set to null.

This change updates ServiceWorkerContextCore::DeleteForOrigin() to set
the immediate unregister job flag to true.  All other callers set the
immediate flag to false.  The storage quota manager uses
ServiceWorkerContextCore::DeleteForOrigin() when clearing data for the
origin.  In addition to evicting origins to free up storage quota, this
code path is used by scenarios like the Clear-Site-Data network
header and the "clear browsing data" privacy UI setting.

The change adds a new function to the ServiceWorkerRegistration class,
DeleteAndClearImmediately(), which is used by ServiceWorkerUnregisterJob
when the immediate flag is true.  DeleteAndClearImmediately() is very
similar to the existing function, ClearWhenReady(), but it does not wait
to clear the registration when the active worker has controllees.

ServiceWorkerContextCore::DeleteForOrigin() is also updated to call
DeleteAndClearImmediately() on the origin's uninstalling registrations.
Registrations stuck in the uninstalling state are waiting for the active
worker's controllees to unload.
ServiceWorkerContextCore::DeleteForOrigin()  must stop waiting and clear
the registration immediately.  GetUninstallingRegistrationsForOrigin()
is added to the ServiceWorkerRegistry to enable
ServiceWorkerContextCore::DeleteForOrigin() to find the uninstalling
registrations it needs to clear.

For testing, the change adds a new unit test that sets the immediate flag
to true in ServiceWorkerJobTest.  The change also adds a new service
worker WPT test, unregister-immediately.https.html.  The new WPT test
uses the Clear-Site-Data network header to unregister service workers
with the immediate flag set to true.  The WPT test adds a test case for
each service worker state.  For example, one test case uses
Clear-Site-Data to clear a registration with a service worker in the
'parsing' state.  Other test cases cover the other service worker states
like 'installing', 'installed', 'activating' and 'activated'.
Additional test cases verify controllee state, which includes verifying
what happens to pending fetch events when the controller clears.

Bug: 1014114
Change-Id: I3baee1117e6dbd349ec4c98c796d4e1b2bddde96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2040533
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Commit-Queue: Steve Becker <stevebe@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#743477}