Use 'foreground duration' to measure when the peak CPU usage occurred
Before this CL, the time when the peak CPU usage occurred was measured
as the time delta between the navigation start and the time when that
happens. This didn't consider the background state and back-forward
cache. If the tab was in background for a long time, the measure time
was longer than we expected.

This CL fixes this by recording the CPU usages only when the tab was not
backgrounded in the window.

Bug: 1014174
Change-Id: Ib15d8f189b1ce9a0329498da9883d54d54ac8bc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050295
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743481}