Rework the SVG resource cycle-checker
This moves the traversal code from SVGResourcesCycleSolver to
LayoutSVGResourceContainer. The reason for this is to be able to perform
slightly more heterogeneous traversal - i.e not only using SVGResources.
The primary use case this is to get rid of the "linked" resource from
SVGResources (currently only used by <pattern>s). This also the reason
for the split into FindCycleFromSelf() and FindCycleInDescendants() -
the former is expected to be overridden by a resource that needs to take
special care of some specific reference. (This mechanism could hopefully
also be used/extended to handled cycles via <feImage> in filters.)

Bug: 1028063
Change-Id: I34c80037516e0059e8b4cb4113fa0f4147d7ec31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2066722
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#743488}