Update PermissionPromptBubbleView to better handle small prompts
Recently PermissionPromptBubbleView::CalculatePreferredSize was added
to ensure permission prompts do not expand too large when large
prompt text is present. While this helped ensure the prompt respected
the same max bubble width used elsewhere in Chromium it also caused
small prompts to become larger to fill the space.

This change updates the CalculatePreferredSized impl to take into
account the preferred size of the context along with the max size and
use the smaller of the two.


Bug: 1054699
Change-Id: Ie3237d50bafd1b19dae9928b8d2c834dc37c9ba5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067081
Commit-Queue: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743492}