Remove existing_user_controller as policy observer
Remove existing_user_controller as an observer of MinimumVersionPolicyHandler. This is because in case of OOBE flow after enterprise enrollment, the policy is fetched and if the minimum version requirement is not satisfied, then the update required screen is forced to be shown from existing user controller.
This abruptly interrupts the flow and causes the DCHECK() in the
EnterpriseEnrollmentHelperImpl::~EnterpriseEnrollmentHelperImpl() to fail.
Currently, not sure what would be the correct place to have the observer for the policy handler, but removing existing_user_controller for the time being to fix crash.

Bug: 1048607
Change-Id: I84c7cb9de91c2b3de26345119519e7ad348c12ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050486
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743516}