Allow WaitForDom to handle multiple cases combined with and/or.
This change introduces the concept of "element condition", which are
arbitrary combination of element check that support and, or and not, and
is expressed in the proto as "all_of", or as "any_of and not as "none_of"

In proto, the concept is represented by ElementConditionProto. In C++,
the existing ElementPrecondition.

This change also switches all "element conditions" to use this concept,
that is to decide:
 - whether script preconditions match
 - whether wait for dom can continue
 - whether a prompt choice should be shown
 - whether a prompt choice should be auto-selected

The idea is to have only one piece of code that does element condition
that can be reused in multiple places that is powerful enough not to
require constant extensions.

This change is not backward-compatible with previous proto definition.
The idea is to have the server generate both the old proto and the new
conditions before submitting this change.

Bug: b/147865211
Change-Id: I74925e95a7c4898baaec3a64746c8c25e988ceb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2020951
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743529}