extensions: don't crash in ExtensionInstalledWaiter destructor
As part of unit-testing this class, I added a hook to it named
g_giving_up_callback. The destructor looks like this:

  ExtensionInstalledWaiter::~ExtensionInstalledWaiter() {
    if (done_callback_)
      g_giving_up_callback->Run();
  }

The hook allows unit tests to plug into the "browser destroyed while
ExtensionInstalledWaiter is live" behavior, which improves test
coverage. However, in production, g_giving_up_callback is null, so the
effect of this test hook was that if this case (ExtensionInstalledWaiter
being destroyed by browser exit) ever happened *in production*, the
browser would crash :(

Bug: 1049190
Change-Id: I434613d292ffd115dde9b6d8bc1658a14cd5dd0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067465
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743538}