aw: Unsubscribe from BeginFrameSource
During destruction of WebView it's possible that not every observer has
unsubscribed from BeginFrameSource in RootFrameSink (the latest it will
happen in dtor RootFrameSink). This leads to situation when
RootFrameSinkProxy is still observing BrowserViewRenderer's
BeginFrameSource in dtor which leads to DCHECKs.

This CL adds explicit RemoveObserver in RootFrameSinkProxy dtor to
handle sutiation correctly.

Bug: 1054777
Change-Id: I00811c79ae430872964075c83a308cad17a8694a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067087
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743567}