Add origin trial support for CSS media features/values
This CL adds plumbing of an ExecutionContext into the creation of
MediaQueryExp objects. This will allow RuntimeEnabledFeatures that are
slated for origin trial experimentation to pass the ExecutionContext to
the REF::*Enabled check so that origin trials can appropriately enable
new media features based on experimentation status per origin.

Callsites that perform media query parsing now pass an ExecutionContext,
except for tests, media_feature_overrides, and html_preload_scanner. The
overrides don't need an ExecutionContext since they are used to produce
emulated values - parsing/setting the new features/values would still
need to be enabled for this to have any effect. Not having the
ExecutionContext for the preload scanner means that if experimental
media features are gating resources, they may not be preloaded. However,
given that these features are experimental, will eventually get loaded,
and the fact that the amount of extra plumbing required is non-trivial,
I've decided not to add the plumbing here.

This CL also adds a test-only media feature that is enabled by the
basic sample API origin trial feature in order to test various aspects
of the plumbing.

Bug: 1051382
Change-Id: Ifaafa2816f65a04f9cdd87bab8b1f216b54a348c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062699
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#743570}