Android WebView save/restoreState with TransientEntry.
TransientEntry will chang to a general entry when we restore it
from the state which was saved before. while TransientEntry
usually used for InterstitialPage(eg. blocking pages) then an
TransientEntry will be created again when we restore and load
with the entry.

Use the  LastCommittedEntryIndex as selected entry index same
with chrome.

BUG: None
Change-Id: I56146535799a6cbaa367b9f9de9485fa870b4640
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068218
Commit-Queue: Bo <boliu@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743578}