[WebLayer] Integrate HostContentSettingsMap
HostContentSettingsMap will shortly be required in WebLayer for reuse
of //chrome's SSLHostStateDelegate implementation. This CL enables that
by creating a BrowserContextKeyedServiceFactory implementation for
HostContentSettingsMap in //weblayer. The implementation is modeled
after that in //chrome but can be significantly simpler. Two relevant
differences (see discussion in comments on the CL for more details):

1. In WebLayer, we do not reply on the //chrome-level feature to
   determine whether to set
   migrate_requesting_and_top_level_origin_settings to true but simply
   set it to true. The feature has been enabled by default for 2+
   years and is shortly set to be cleaned up, so it is not worth
   componentizing.

2. We did not bring over the Android-specific code from //chrome's
   factory (which would need componentizing). The features in question
   are not currently supported in WebLayer.

HostContentSettingsMap requires a PrefRegistrySyncable in order to
register its prefs. To facilitate this, this CL changes //weblayer's
BrowserContextImpl to use a PrefRegistrySyncable (as //chrome's
ProfileImpl does), rather than its previous usage of a
PrefRegistrySimple. This change should have no meaningful behavioral
impact as WebLayer does not use sync.

Bug: 1030692
Change-Id: I86053d749141b7e223812b042949aaf7ff5ecee4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062342
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743599}