android: Return global default display
Android will soon only allow retrieving the display from a "display"
Context. In practice for chromium, this means retrieving the
WindowManager or Display is only allowed on Activity (or a
ContextWrapper wrapping an Activity). In particular, it will cause
problems to use ContextUtils.getApplicationContext() to retrieve the
Display, which unfortunately happens in a lot of places.

This CL converts ui/display to store a default display, and return that
if any code tries to use an illegal Context to get the display. This is
really going against the intentions of the Android change and will keep
allowing chromium to do the wrong thing. However the ui/display library
(in both java and native) are already set up to force code to do things
the right way already, which is the correct long term solution.

Bug: 1042063
Change-Id: Iea12bcea9d0b7b99bd35bdf2a4de7d6f08272d71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2064472
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743601}