[M81] Paper-over the crash in UpdateSubresourceLoaderFactories.
This CL adds a temporary workaround to avoid the crash in
content::RenderFrameImpl::UpdateSubresourceLoaderFactories.
This is not a real fix, but only an attempt to suppress the crash spike
on M81 branch.

This CL should not be needed once:
1) A real fix is found for https://crbug.com/1013254.
or
2) CorbAllowlistAlsoAppliesToOorCors ships
   (see https://crbug.com/1013254#c19)

Bug: 1013254
Change-Id: I3e918bd2cdf569e65431c8b62e618fd540394afd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068480
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/branch-heads/4044@{#408}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}