Grant file access to controller SW when accessing subresources.
When controller Service Worker is running in a different process than
the controllee page, it is unable to access files uploaded by in
subresource requests. This change makes it so we explicitly grant the
file access to the Service Worker before proceeding with dispatching
Fetch event.

The proper long-term solution here would be to put a Mojo remote in
network.mojom.URLRequestBody / .FetchAPIRequestBody , which grants
access to the file, and the renderer could clone it and pass it to other
renderers as necessary. This would let us avoid using FilePaths which
had been a source of security attacks in the past.

Bug: 1028866
Change-Id: Id437f2041315c74b204827c633ff7a0c5970e37a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042647
Commit-Queue: Kasjan Siwek <kasiwek@microsoft.com>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743630}