webauthn: allow phone pairing via Windows UI.
When caBLE v2 is enabled, if no phones are paired
and no caBLE extension was received then we'll jump
straight to the Windows-native UI for a WebAuthn request.
However, this means that the user has no opportunity to
pair a phone.

This change causes the transport selection dialog (which
includes a button to pair a phone) to be shown if the user
cancels the native UI. Since that dialog also includes a
button to trigger the native UI, the user can decide to go
back if they wish.

BUG=1002262

Change-Id: I559f3586772bfc7aa138c55aa21877e9bd3ff898
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2032014
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Auto-Submit: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743644}