Only snap background src rects that are close to integral
Code to snap the src_rect for background image painting was snapping
the rect to an integer under the assumption that properties modifying
the source size were used overwhelmingly by sprite maps. This caused
severe jitter when animating the background properties of a small
image scaled up to something large.

Change the snapping to only snap when the rectangle is already close
to integral. This requires adding a method to FloatRect.

The test is in WPT form but not the WPT directory because I can
see no way at all to make a ref test for this case given it relies
on image resize filtering that is highly implementation dependent.
In particular any code that snaps will have integral pixels with no
filtering. Hence to show sub-pixel src rect we need to have something
filtered thus introducing the platform-dependent filtering issue.

Bug: 925196
Change-Id: I1fcf16be6b453a5b411d59c91251fac06a153721
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2063916
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#743647}