Paper-over the crash in UpdateSubresourceLoaderFactories.
This CL adds a temporary workaround to avoid the crash in
content::RenderFrameImpl::UpdateSubresourceLoaderFactories.
This is not a real fix, but only an attempt to suppress
the recent crash spike.

This CL should not be needed once:
1) A real fix is found for https://crbug.com/1013254.
or
2) CorbAllowlistAlsoAppliesToOorCors ships
   (~M83; see https://crbug.com/1013254#c19)

Bug: 1013254
Change-Id: I3e918bd2cdf569e65431c8b62e618fd540394afd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068638
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Auto-Submit: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743675}