Domain Reliability: Refactor configs and contexts
This CL does some refactoring around DomainReliabilityConfig and
DomainReliabilityContext. The main change is to construct the
Google configs lazily to avoid needlessly consuming memory (with
configs that are largely the same) for domains that may never
be visited.

We also now require collector URLs to be https. This was already
the case except in tests.

Bug: 1047437
Change-Id: I6abc43e8123ae98daac417cbd3d9955acb41132f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2055548
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743681}