Deflake attempt for SettingsSiteSettingsPageBrowserTest.labels test.
This test was flaky most likely because it was previously a
SettingsPageBrowserTest which is known to be flaky. Converted to a
more targeted CrSettingsBrowserTest instead.

Also stop referring to the private |defaultSettingLabel_()| method
from tests. Instead properly export it with cr.define().

This is in preparation of upcoming changes in
settings-site-settings-page.

Bug: 709442,1032584
Change-Id: I4435e4e03a3c7e0f303bd70f39f2fd8cb99199bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067547
Auto-Submit: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743699}