[throughput] Remove some checks.
We already have checks in place to prevent underflows. So remove the
explicit check that was introduced to check other underflows, but it
is tripping tests for valid sequences. So remove the DCHECK.

We also report the length of the sequences in UMA, and according to
the data, the lengths are:
 50%ile: < 20
 99%ile: < 500
 99.5%ile: < 800

So underflows are not a concern at the moment.

BUG=1052790

Change-Id: Ia87b3fae331b01c483468887894f16364938a624
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068894
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Auto-Submit: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743769}