Use flinged bounds for snap fraction calculation
We currently save snap fraction in CompleteDrag, but if the PIP
window is flinged, the bounds of the PIP window is not the flinged
bounds at this point as it just creates a bounds event and sends
the client a bounds change request. So we need to pass the flinged
bounds to SaveSnapFraction.

Before introducing snap fraction, we used restore bounds and saved
the flinged bounds properly, so this CL basically does the same
thing with snap fraction.

BUG=b:149875362
TEST=Manually confirmed flinging PIP and rotating the screen works.

Change-Id: I7b109eaf7786aee5d3c70da167c27ff5c1a69e90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2065510
Commit-Queue: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743799}