AppCache: Remove OnceCompletionCallback alias.
This CL removes the content::OnceCompletionCallback alias, and replaces
it with net::CompletionOnceCallback. Explicitly referencing the net::
namespace is intended to be a readability improvement, because it makes
it clear that the callback's int argument follows the //net convention,
where negative values are errors, and non-negative values are byte
counts.

Bug: 1055133
Change-Id: Ie5057f0abe27f97620ce1897ac839e9e1d53bf82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2069507
Auto-Submit: Victor Costan <pwnall@chromium.org>
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743850}