[Spellcheck] Use PostTaskAndReplyWithResult in WindowsSpellChecker
Makes the implementation of the background helper in the
WindowsSpellChecker class cleaner by using PostTaskAndReplyWithResult.

Currently, a reference to the main thread task runner is passed to the
background helper at construction, and the background helper uses it in
several methods to post callback tasks on the main thread. With this
change, the background helper is simplified and just needs to return
results normally.

Note that with this change, we lose a little bit of visibility on the
fact that the callbacks are run on the main thread. This isn't really
an issue because the callbacks don't need to be run on the main thread.
The important thing is that COM objects are interacted with on the
background thread, and the background helper already enforces that.

Change-Id: I86afa6daf6d4e038bd4bdce9a058a9dd6417eed8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067263
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#743944}