Fix fuzzer crash with CSSParserImpl's usage of context Document pointer
In fuzzer runs, the CSS parser is invoked with a CSSParserContext object
that does not have a Document. There are a few places where the CSS
parser uses the context's Document pointer in order to get an
ExecutionContext to enable origin trial controlling the exposure of
RuntimeEnabledFeatures.

This change adds an ExecutionContext accessor to CSSParserContext that
will return null when the CSSParserContext doesn't have a Document.

R=futhark@chromium.org

Bug: 1055119
Change-Id: I394976bd1d4548fd13db1845f0b84b7612076eca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2069851
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#743948}