[iOS] Allow height of FindInPage bar to change with toolbar height
On iPhone, the height of the toolbar changes with dyanmic type size.
Before, the height of the FindInPage bar was fixed. According to the
design, the actual content of FindInPage should not change (as it is
semantically a toolbar). The height should change, however.

Now, if the toolbar grows, the FindInPage bar stays centered in the
background view beneath the safe area.

One small thing to note is that the toolbar layout guide changes when
the toolbar shrinks into fullscreen. In this case, the FindInPage bar
will change height along with the toolbar until a minimum height is
reached. At that point, it will stop shrinking.

Fixed: 1053976
Change-Id: I2172b37dc039239b13b44e42caebb03176e6100a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2066751
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#743954}