Create Recent Site Permissions element
As part of the privacy settings redesign to the capability to surface
recently changed permissions is required. A previous CL introduced the
capability to retrieve these changes and this CL introduces a Polymer
element to display them.

This CL also places that element in the site_settings_page hidden behind
the privacySettingsRedesignEnabled flag.

Bug: 1032584
Change-Id: I744f2a57de65200c2c0684b4777bd77549496e6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2030525
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743964}