[Paint Preview] Fix capture timing
Use onPageLoadFinished() rather than onDidFinishNavigation() for capture
experiments. This ensures the page is loaded and most of the content is
renderered before we attempt a capture. At present this is happening as
a navigation occurs which is too soon as there is often no content.

Known Issue: some sites may trigger too soon as they lazily load
resources. It would be better to trigger on navigate away, but AFAICT
such a trigger/handler doesn't exist as we normally avoid doing work
when unloading the page (requires further investigation).

Bug: 1055353
Change-Id: I33b99835f96543c476a858dd629ce1b41c130b78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2070104
Reviewed-by: Mehran Mahmoudi <mahmoudi@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743975}