Make sure to init Mac's VideoToolBox before GL
Otherwise this leads to crashes in VTDecompressionSessionCreate.

The CL http://crrev.com/c/2013806 changed the order but
the intention was to fix an issue on Linux/Wayland. So
restore the previous behavior on Mac. Also restore the
previous behavior on other operating systems to make the
change safer. And guard the new order with OS_LINUX.

Bug: 1047643
Change-Id: I9637a684996ee095d32a336610cb65a7d0d20370
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068207
Reviewed-by: Maggie Chen <magchen@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Julien Isorce <julien.isorce@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744001}