device/fido: clean up OWNERS.
Kim is no longer active in this area but has done a review recently and
we want to make it easy for her to return :)

Other emeriti are not, in practice, reviewing in this area. Remove them
so that tools don't auto-suggest them for CLs that they'll only have to
bounce to someone else.

Change-Id: I5d46659e1621e526912b3b48323b2027c8878593
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2070563
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744025}