app_list: Hide drag item title during a drop to a folder
When dragged app item view is dropped into a folder at the end of the
drag, hide its title, and keep it hidden. Generally, the title will be
hidden at the end of the drag while the item is animating to the new
target bounds, but the item view is removed from the views model before
the bounds animations run, so AnimateToIdealBounds() will not update
the dragged view title.

BUG=1053935

Change-Id: Ief279b9f11f7c6bf7d562e4e4d9a690e89acf229
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068279
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744031}