[Switch Access] Improve interactions with Virtual Keyboard.
Previously, Switch Access would get stuck in the virtual keyboard when
it was hidden, and would not jump to it when it was activated by any
method other than the Switch Access menu.

This change listens to events corresponding to the keyboard being shown
and hidden, and reacts appropriately.

Bug: 996899
Change-Id: Ic6c58c8f66afdbf7422b74a8dc86b0d64c700607
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062584
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744039}