Use -debug-info-kind=constructor on ChromeOS
This uses the new debug info size reduction feature introduced in
https://github.com/llvm/llvm-project/commit/651128f5). This was
previously enabled for other platforms (crrev.com/c/2040441). For
ChromeOS, we needed to upgrade the compiler to get the new
feature. Now that that is done, we can enable the feature for
ChromeOS, too.

Bug: 1050118
Change-Id: Ide359fcb8d1e748bb321927055802bf6ebd7fad2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067741
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744100}