Stop discarding events to disabled views
Currently, events that would be dispatched to a disabled view are
discarded instead. This seems to be a holdover from old event handling
code and not necessary anymore.

To allow pre-target handlers to still see these events, this CL changes
these methods to dispatch the events to the closest enabled ancestor
view.

Bug: 1043374, 1011082
Change-Id: I82bca238e98b9b820c0c7619a5880fbf3497d886
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2029087
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744102}