Make popped out extensions stay popped out if they have a context menu shown in the new extensions toolbar container.
Previously we would only check if a context menu was running when trying to hide a popped out extension. In the case where the context menu was not yet running when this happened then the popped out extension would be hidden when it shouldn't be. This change lets the ExtensionsToolbarContainer keep track of an extension with an open context menu and adjust visibility accordingly.

Bug: 943702
Change-Id: I547b16eb324852c9559238c646d189c39892fb55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2065009
Commit-Queue: Caroline Rising <corising@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744119}