system-web-apps: Implement show_in_launcher and additional_search_terms
There are two changes in this CL needed for show_in_launcher and
additional_search_terms.

1. Re-publish apps once SystemWebAppManager syncs them. When
ExtensionApps and WebApps get notified of a new app installation
SystemWebAppManager is still unaware of them. This is an issue because
we store information relevant to the app in SystemWebAppManager e.g.
additional_keywords and show_in_launcher. To fix this, we change
ExtensionsApps (and in a follow up WebApps) to listen for
SystemWebAppManager::on_apps_synchronized() and refresh the info for
SystemWebApps.

2. Retrieve additional_search_terms and show_in_launcher from
SystemWebAppManager in ExtensionApps when on_apps_synchronized() gets
called.

Bug: 1054195
Change-Id: I5af5e24bd4d61d524e3ad383093c1d00767b4106
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2065194
Commit-Queue: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744152}