Save last reset time in AppTimeController.
This cl saves the last reset time in AppTimeController in
user pref. The value is saved every time the
|AppTimeController:: reset_timer_| fires; the timer fires once
every 24 hours.

Last reset time value is read from the PrefService when
AppTimeController is constructed. It is used to check if the reset time
has been crossed while the device was off.

Bug: 1050876
Change-Id: I11d189256a403b3c6f452905e2ccfc50f1660873
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067242
Commit-Queue: Yilkal Abe <yilkal@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744162}