Allow videos to be fullscreened in chrome://media-app.
The actual fix is just to add `allow="fullscreen"` to the guest
<iframe>, but we want to ensure we're setup for testing things like
this.

The Mock app_main.js is updated to support mock loading of videos as
well as images, and the test message API is updated with a mechanism
to pass a command (i.e. to invoke <element>.requestFullscreen()`).

language_in=ECMASCRIPT_2018 is added to typechecking flags to permit use
of the {...spread} operator.

The test case tickles a DCHECK in Blink's layout logic
(crbug.com/1052791). A neat enough workaround for that is to remove the
`overflow: hidden` style from the main frame before attempting to
fullscreen a video element. The DCHECK is hit in manual interaction as
well, but seems to be innocuous when DCHECKs are disabled.

Bug: chromium:1052791, b/146459531
Change-Id: I234a82b2c33d238252815d80da14b7cc1e106a49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2055951
Commit-Queue: Trent Apted <tapted@chromium.org>
Reviewed-by: Bugs Nash <bugsnash@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744231}