ui: Fix smoothness report when attaching to compositor late
Smoothness is not calculated correctly when underlying Layer
is attached to a Compositor after the animation starts. One
such case is the "Show" animation of WindowCycleView. It is
always report as 100% incorrectly. Smoothness metric depends
on frame number from Compositor. |start_frame_number_| is
captured when the animation starts. However, if Layer is not
attached, the frame number could be be obtained and stays as 0.
As a result, smoothness is always reported as 100% when the
animation finishes.

Bug: 1053732
Change-Id: I60c750afaf5b4820ad17aebfa9809178399c50f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067674
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744284}