[FlexNG] Layout overflow:auto items before calculating intrinsic widths
This brings back the behavior from
https://chromium-review.googlesource.com/c/chromium/src/+/1978103 that
was lost in
https://chromium-review.googlesource.com/c/chromium/src/+/2027075 . Now,
if a flex item with overflow: auto gets a scrollbar, we add the width of
the scrollbar to the intrinsic size. Blocks has similar behavior but
gets there differently. This could cause performance degradation.

This fixes the "inner dimensions" test from
css3/flexbox/overflow-auto-resizes-correctly.html .


I removed the intrinsicHeightBox part of the test. It asked that we give
the item 100px height in this case:

<flexbox column height:auto>
  <item style="min-height: min-content; flex: 0 1 0;">
    <div style="min-height: 100px">

This requires laying out the item to determine what the min-content is.
Legacy does this, Firefox doesn't. NG doesn't do it for blocks, e.g. the
outer div gets 0 height here:

<div style="height:0px; min-height: min-content">
  <div style="height: 100px;">

Bug: 845235
Change-Id: I35ba43d63676798f7a632c33ead4e147ea36f34f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2058066
Commit-Queue: David Grogan <dgrogan@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744318}