Introduce canonical paths as StrongType: CanonicalRelativePath
Currently these paths in content verification is
base::FilePath::StringType, which can be unintentionally converted
from any base::FilePath. This CL make the concept strong typed,
using util::StrongType. This CL also names it with a hint that
this is a relative path and not a full path.

No behavior changes are expected with this CL.

Bug: 1051401
Change-Id: Ib4117e4d2f83e58ae3b294a22b6d7e2720ed5fb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2059572
Commit-Queue: Istiaque Ahmed <lazyboy@chromium.org>
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744334}