Remove nonexisting feature flag from MediaRecorder test
This test is flaking sometimes with an error message about
a nonexisting flag. This removes the attempt to use that flag.
This is not the reason the test is flaking, but the message is
misleading.

Bug: 986021
Change-Id: I847ddee6b63a52b25040c49567de29cad91e2560
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2071018
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Auto-Submit: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744335}