Update Structured Headers implementation to support V15 lists.
The format for lists has changed since Draft 13, to allow for both
inner and outer list members to carry parameters. (This is directly
supporting the needs of the Reporting API).

Items now intrinsically have parameters, which are ordered. These will
be serialized if present.

The allowed token characters are also changed with this CL, the
characters "!", "#", "$", "&", "'", "+", "^", "`", "|", and "~" are now
allowed in tokens.

This CL updates the parser and serializer to support the new format.

Bug: 1048752
Change-Id: Ib6a6ffdb497d64cf15601fe4a4fa6601809d4ed6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2037344
Commit-Queue: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744338}