Portals: Handle edge case with connecting child frames
We shouldn't set the portal element's content frame if
SubframeLoadingDisabler::CanLoadFrame() returns false, this can lead to
discrepancies in the connected subframe count and result in frames not
being disconnected.

In the test case in this CL, we try inserting a portal into a subtree
that is being disconnected.

Bug: 1041406
Change-Id: I7e0377392457f5b0ca6537c27cb6a5835b8eedd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2002859
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744350}