Fix leaking OffscreenSurface objects
SkiaRenderer was failing to delete OffscreenSurface objects
corresponding to RenderPasses that are drawn to texture because of a
CopyOutputRequest but not embedded in another RenderPass. No
ImageContextImpl would be created but an OffscreenSurface would be. In
SkiaOutputSurfaceImpl::RemoveRenderPassResources() it ignores
RenderPassIds with no corresponding ImageContextImpl.

This caused a memory leak for the lifetime of the SkiaOutputSurfaceImpl
unless another RenderPass with the same ID was created later. It also
likely caused some size mismatches if another RenderPass with the same
ID but a different size was created later.

TBR=vasilyt@chromium.org

(cherry picked from commit 375bf8cd0bda18363d7455ecd87859206294476c)

Bug: 1043676
Change-Id: I3dea76bc20171146814bf47b2fc3d3ded40f8fd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067332
Commit-Queue: kylechar <kylechar@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#743350}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2070557
Reviewed-by: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/branch-heads/4044@{#472}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}