Refactor Editor Bubble to take TabStripController (Editor Bubble refactor Part 0)
This doesn't get us to the final API that we want (ideally it would take a model rather than a controller), but it's a step in that direction. It also easily allows us to delete a large part of the TabController API that was used exclusively by the Editor Bubble.

Bug: 1055537
Change-Id: I0914175aade82ebf874f85d979977372281dec40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2070959
Commit-Queue: Connie Wan <connily@chromium.org>
Reviewed-by: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744477}