[IntentHandling] Send replaced preferred ARC app to ARC.
This CL collects the replaced preferred app when setting a preferred app
for an intnet filter. This information will be send to ARC to sync with
ARC. In addition, a set preferred app to ARC intent helper package will
be send to ARC as well.

BUG=853604

Change-Id: I1b715498cd2edca58319f70847a2919d351c0603
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2043332
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Reviewed-by: David Jacobo <djacobo@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744487}