[IntentHandling]Fix common apps navigation throttle dead loop.
Currently when we defer navigation, it will try to find apps, and if
there is no preferred app, it will try to show the bubble and resume the
navigation. However when there is no app found, the navigation will be
trapped in start navigation, defer navigation and resume navigation all the
time. This CL fixes this issue by adding checks when there is no app found.

BUG=853604

Change-Id: Id4215e6b0885da7ab37da9b81f75409ad915af7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2072481
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744561}