Replace deprecated getLastUsedProfile function in DataReductionPreferenceFragment.
|Profile#getLastUsedProfile| is deprecated and replaced with
|Profile#getLastUsedRegularProfile|. This CL updates
DataReductionPreferenceFragment class.

Since settings is considered as always running on regular mode,
it is safe to use regular profile even called in incognito mode.
For this reason, there is no need to make change on code behavior to
act differently for each profile. Therefore, it is safe to leave
behavior as-is today.

Note: This change is only code clean-up, does not change any behavior.

Bug: 1041781
Change-Id: I0585683c1d6d844b2749e36095d941f02f19d237
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2066908
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Commit-Queue: Side YILMAZ <sideyilmaz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744585}