[iOS] Add cancel sign-out operation needed for architecture migration.
This patch adds the authentication flow required to cancel a sign-in
operation following a click on the 'Cancel' button in the
'Turn on Sync?' page.

This CL is part of a series of refactors to move the existing sign-in
architecture to the Coordinator-Mediator design paradigm.

Fore more information, see go/chrome-ios-signin-migration.

Bug: 971989
Change-Id: I72e8bf3c8fc15338d419dc2d3a888eecc430b807
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067076
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744607}