Load Mac system colors for both light and dark scheme.
System colors used in CSS should per spec depend on the used
color-scheme for that element. Since the used color-scheme may not be
the same as the preferred color-scheme, we need to provide the system
colors for both light and dark at the same time. This CL temporarily
sets the appearance to light and dark aqua to retrieve those colors.

TEST=StyleEngineTest.InitialColorChange

Bug: 929098
Change-Id: I4bb9553cf1b2fc4e98cd1c7c13041adce3e50353
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1860023
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744630}