cbuiv payments: remove two GetDialogButtons overrides
This makes it hopefully more clear which events can cause the dialog
buttons to change visibility. For CardUnmaskPromptViews, whether the Ok
button is visible or not depends on whether the autofill client is in a
persistent error state. For LocalCardMigrationDialogView, the existing
code never calls DialogModelChanged, so presumably the value of
`controller_->AllCardsInvalid()` is constant across the lifetime of that
dialog.

Bug: 1011446
Change-Id: I5921f086b3078e68aa417945560ad73a357fe190
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067874
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744659}