metrics: Clear the global memory dump once data has been logged.
This is a partial reland of
https://chromium-review.googlesource.com/c/chromium/src/+/2056820, which
was reverted due to crbug.com/1055093. When a process memory dump fails,
this could lead to a DCHECK() in debug builds, or a crash in release
ones.

Compared to the original CL:
- Don't crash when the global dump failed
- Add a comment and a test explaining the reason.

Compared to the current state of the code, this fixes a bug, where
previous values can be re-used when the new memory dump fails. This
should be rare enough to not skew metrics in practice though (according
to the low frequency of crashes), but is still incorrect.

Also fix a few include-what-you-use "git cl lint" warnings.

Bug: 1054360
Change-Id: Ia3ed513db4be5bd323787c8b2914db345cfeae4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2074699
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744690}