heap: Fix data race in visiting TraceWrapperV8Reference.
Visiting a TraceWrapperV8Reference checked whether the reference is
empty or not. If that check happens on a concurrent thread, it could
race with v8 setting that reference.
The fix is to check only on main thread. For concurrent threads, the
reference is put into the worklist regardless of whether it is empty
or not. The main thread then checks if its empty when popping from
the worklist before pushing that reference to v8's worklist.

Bug: 986235, 1054852
Change-Id: I877844c72994b0b6cb97559d25d225a78a2b3649
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2075278
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744743}