Disable tabbing through Ash's PIN keypad buttons
This makes the buttons of the LoginPinView view
untabbable. It affects the quick_unlock and smart card PIN
keypad UI in the user pod in Ash, and also the Parent
Access Code dialog.

This should NOT be an accessibility regression, since these
buttons don't provide any extra functionality: i.e., the PIN
can still be directly entered into the input field via other
methods. The change was also approved by the a11y team (see
the attached bug).

This also makes it consistent with the HTML/Polymer-based
PIN keyboard, which was recently changed to not support
tabbing through the buttons (see the attached bug). The main
motivation is to avoid struggling with the a11y behavior of
these buttons (specifically, issues of keeping the focus and
of pronouncing the "bullet" word of the screen reader).

Bug: 1043194
Test: log in via a smart card, log out, start a login again, in the PIN dialog press Tab a few times, verify that the buttons of the PIN keyboard ("1", "2", etc.) don't get selected
Change-Id: I650878e2fcecb2fd5e35d9fe847664fa8e4a6d87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2074441
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744749}