[ios] Fix NTP snapshot edge insets
There were two changes to NTP frame and inset calculation
that broke NTP snapshot insets and resulted in a snapshot
with a smaller frame than needed.

https://chromium-review.googlesource.com/c/chromium/src/+/1392956
https://chromium-review.googlesource.com/c/chromium/src/+/2019244

The NTP's fullscreen behavior is different from a regular
webpage in that the bottom toolbar stays fixed (in compact width
size class). Since the NTP's frame is always above the bottom
toolbar, there is no need to further inset the snapshot frame
on the bottom.

This CL fixes the snapshot inset.

Fixed: 1040101
Change-Id: I0f5c0289b873a031ac66550d1c4aca1f5a3e73cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2073011
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744758}