WebLayer: De-flake and re-enable FindInPageTest#testHideOnNewTab
The flake seems to occur when the touch is handled by selecting text on
the page, so switch to a page that doesn't have text in the way of the
touch.

This also puts a timeout on each CountDownLatch since an unbounded
await() leads to very unhelpful (i.e. non-existent) error output.

Bug: 1052802
Change-Id: I44393c4a821413e626eb3a5ef4c2e16609022040
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2074812
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744793}