Advertise a VideoFrame's visible size as the coded size.
This CL fixes the VaapiVideoDecodeAccelerator to advertise the visible
size as the coded size for decoded video frames. Here's the background:

CL:2011583 changed the VA-API decoding pipeline so that the EGLImage
backing a VideoFrame's service-side texture is created using the visible
size instead of the coded size. This is correct because the graphics API
does not care about the non-visible area. This was done to avoid a green
line error: making the size of the EGLImage equal to the visible size
allows the texture sampler to not sample beyond the visible area, so the
green line (part of the non-visible area) is avoided.

The problem is that the coded size of the VideoFrame sent to the client
still contained the non-visible area. This is arguably strange because
as far as the client is concerned, this non-visible area does not exist.
This was also causing problems when PaintCanvasVideoRenderer requested a
copy from the VideoFrame's texture: the coded size was bigger than the
visible size used to create the EGLImage. This caused black images to be
displayed some times (e.g., thumbnails of videos in the Camera Capture
App) because the GL commands sent from the client didn't pass validation
in the service side (e.g., "GL ERROR :GL_INVALID_VALUE :
glCopySubTextureCHROMIUM: source texture bad dimensions").

This CL changes VaapiVideoDecodeAccelerator so that the coded size of
the VideoFrame only includes visible data.

Additionally, it is possible for a decoded frame to end up as an overlay
plane. In that case, the size of the DRM framebuffer should match the
visible area of the frame. This is important so that the non-visible
area doesn't appear in the overlay path. Therefore, this CL plumbs the
desired framebuffer size all the way through Ozone.

There is a caveat, however: there is an unusual corner case where an
H.264 video includes information about a visible rectangle whose origin
is not at (0, 0). Therefore some of the top and left parts of the video
should not be displayed. However, Chrome does not currently deal well
with this case, and this CL doesn't change that (i.e., those non-visible
areas are still displayed).

Note: follow-up CLs will deal with V4L2 and the VideoDecoder interface
variations.

Bug: 1045597
Test: Record video in nocturne's CCA and inspect thumbnail.
Change-Id: I7076b66bfa7cb2284f6a5431968e98053f9b8103
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2056285
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744799}