back gesture: Show hotseat for ARC full screen apps on swipe.
The back gesture will now have special behaviour for ARC apps.
The first swipe will animate in the shelf and hotseat, and
the second swipe will send a back event.

I changed the container the back affordance is in because:
  1) AlwaysOnTop container is watched by the workspace, which
means adding/removing/hiding/etc the affordance window will
trigger backdrop/shelf/etc updates.
  2) AlwaysOnTop container is underneath the shelf container,
which would make it tricky the part with the second swipe
without the shelf pre event handler eating the event and hiding
the shelf.
   3) AlwaysOnTop container is still underneath a lot of things.
I spoke with Ben and he seems to want sometime that is even
above the power menu button. I think maybe we'll have our own
GesturesContainer but for now the OverlayContainer (autoclick,
laser) should work fine.

The affordance was also creating a renderpass at the end which
I removed. It was small, but should help out with the large
minimize animation that commonly follows.

Test: manual
Bug: 1031268
Change-Id: I4f3be651604ac91cf0d2a07a24aa0f51efd44023
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2073199
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744803}