Install xcode via run.py instead of the test runner constructor
We don't guarantee that xcode installed and set before we use otool,
so we want to ensure that it's installed first. This CL pulls this out
of the constructor and into run.py, such that we don't need to pass
xcode_build_version, xcode_path and mac_toolchain into the constructor
of every single runner class.

Change-Id: Id8f315645c22c07448ad6d0211511bc83a9af46e
Bug: 1055330
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044305
Commit-Queue: Jeff Yoon <jeffyoon@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744810}