Desks: Fix MRU and stacking order of windows of removed desks
The windows of removed desks should be demoted to be LRU across
all desks. That's why they're appended at the end of the overview
grid list. However, their relative order should be maintained.

This CL makes sure the MRU order is updated to match the above
expectation, as well as the windows' stacking order (that is
windows of the removed desk should be stacked at the bottom
of the target desk's container's children).

BUG=1048271
TEST=Expanded existing tests.

Change-Id: I06bc69e068e3ffc29bccf7a5983ed6d56fc04830
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2073366
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744856}