[fuchsia] Allow the process' ComponentContext to be overridden by tests.
ComponentContextForCurrentProcess() returns a sys::ComponentContext
connected to the process' incoming service directory, outgoing directory
for services & debugging APIs, etc.

Provide a TestComponentContextForProcess, which overrides the process'
ComponentContext with one which by default has no services available,
and which allows services to be provided by the test itself, and for
services available in the test process environment to be exposed to the
test.

Bug: 1038786
Change-Id: I0d754a1629ab25527dce377d6ffaa039791ef3e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2064908
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Auto-Submit: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744858}