Describe mode override cases in secure DNS setting subtitle.
In cases where the secure DNS setting is forcefully overridden by Chrome
(e.g. detection of parental controls), this cl replaces the setting
subtitle with an explanation. In these cases as well as when the
underlying prefs are managed, the setting is locked and the radio
buttons are not shown.

Change-Id: I27e5b580e09f43bd36a4a924ce4795c06c947c44
Bug: 1040145
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1984660
Commit-Queue: Katharine Daly <dalyk@google.com>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744885}