Make tab strip auto scroll to position of the selected tab
Currently, when a tab in group is selected or a new tab is created
within group, the tab strip doesn't scroll to the position of the
currently selected tab. This could lead to a state where user needs to
scroll by themselves to find their selected tab in strip. This CL fixes
this issue by making following changes:
  * Always scroll the strip to the last position when new tab is added
    in group.
  * When selected a tab in group (through tab grid dialog etc), try to
    scroll to a state where the selected tab is in the middle of the
    strip, if possible,

Bug: 1054079
Change-Id: Ic6921225d744a7b5df1853aa78e14a8f9249f2c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068888
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744892}