Refactor storage handler
The objective is the keep the StorageHandler simple, getting requests
from the UI side and returning the requested numbers.

The logic for the calculation of the storage items is moved outside of
the StorageHandler. The StorageHandler is observing storage size
changes and sending the numbers back to the UI side.

For each storage item, a test api is defined as a friend class, allowing
access to private methods in tests. size_calcualtor_test_api.h is only
linked in the unit_tests target.

Bug: 1051364
Change-Id: I1eacf84f80b649fed1e63c92fca8e03b821c2f4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2051785
Commit-Queue: Jeremie Boulic <jboulic@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744898}