a11y: Enable accessibility on auralinux when API calls are made
This change introduces a mechanism to enable accessibility (on
auralinux) by creating the kApplication node and waiting until there
is a API query to enable accessibility fully. This is done to ensure
chromium will avoid performance overhead for users who don't need
accessibility, by not creating ATKObjects when they are not being
used.

Accessibility is enabled when one of two conditions are met:

1. A client queries an ATKObject API.
2. An AXMode is set on AccessibilityNotificationWaiter (used during
   testing).

This is implemented by using AXMode set on the AXPlatformNode, when
this is set it will allow for all ATK objects to be created. This
change does not remove the need to specify environmental variables and
the commandline flag.

The original author of this patch is Jessica Tallon<jtallon@igalia.com>

Bug: 977112
Change-Id: Ie5e46347296507d214df345320d3d58c9921c84d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2060359
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Julie Kim <jkim@igalia.com>
Cr-Commit-Position: refs/heads/master@{#744905}