[Android] Componentize //chrome's UrlConstants.java
This CL componentizes //chrome's UrlConstants.java for reuse by
WebLayer. Specifically, this file is used by //chrome's external
navigation launching code, which we are in the process of componentizing
for reuse by WebLayer.

As this file references concepts that are conceptually above the
//content level (e.g., browser-level features), it makes sense for it to
live in //components. Rather than create a small component just for
this we placed it in //components/embedder_support. It will shortly be
joined by UrlUtilities.java.

Bug: 1031465
Change-Id: I2f24f625d9dc1b5e977ecb5dcbe3dc98a26e67ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2073938
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745031}