Wire up smart cards with new PIN UI.
The CL introduces a new security token request controller similar to the
parent access controller, which was the only user of the PIN UI until
now.

This CL also adds small bits of functionality to the UI that is required
for the new use-case: Disabling input and clearing the input field.

The new UI applies to login and lock screen, but not in-session or for
OOBE, which use different UIs.

Bug: 1001288
Change-Id: I8c69064a0d944bb6b5a7dac13a2ec2bd968deece
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2020968
Commit-Queue: Fabian Sommer <fabiansommer@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745037}