Make NodeData::is_rare_data_ and NodeData::is_element_rare_data_ const
These flags are used when creating NodeRareData. Making them const
serves to indicate that they are never changed and clarifies the
process of choosing which NodeRareData type to create.
(this was suggested by mlippautz@ on issue 1055194)

This also requires refactoring the initialization a bit.

Change-Id: Ie17684e3050e59c1f6bf16c1e37832e1064be5b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2074881
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745040}