Migrate from identity::ScopeSet to signin::ScopeSet
The mojo identity service (//services/identity) is not used in
production. However, a variety of chrome code uses the typedef
identity::ScopeSet when talking to the C++ IdentityManager and
related auth-token code.

Introduce signin::ScopeSet for these use cases.  It's just a
copy of identity::ScopeSet, which is a std::set<std::string>.
I didn't eliminate identity::ScopeSet in this CL due to existing
usage in //services/identity as well as typemapping code.

I'll delete the identity::ScopeSet version when I delete the
rest of the identity service.

TBR=rockot@chromium.org
TBR=droger@chromium.org

Bug: none

Change-Id: Ib938f08763eda81de22a1a79b460011e012a81ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2076426
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745049}