[iOS] Add authentication operations needed for sign-in migration.
This patch adds the authentication flow required to sign-in or sign-out
a user account following a click on one of the "Turn on Sync?" buttons.

This CL is part of a series of refactors to move the existing sign-in
architecture to the Coordinator-Mediator design paradigm.

Fore more information, see go/chrome-ios-signin-migration.

Bug: 971989
Change-Id: Id9e38721b4343929d23f7861b5b5199a5dfb6a41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2059414
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745077}