Fix in-session password change success detection for Ping IdP
In-session password change success detection mechanism is based on
IdP-dependent heuristics. Small differences are possible even within a
single IdP which may lead to improper detection of successful password
change and left it unsynced between a chromebook and IdP.

This CL extends success scenarios for Ping IdP. The problem was detected
in customer installation and verified with a test account.


Bug: 1056602
Change-Id: I6a83c7e3e9eb070cd1677afe181db395b4593420
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2078451
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Maciek Slusarczyk <mslus@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745084}