Load DnsSession into ResolveContext at registration
In HostResolverManager::RegisterResolveContext(), perform an initial
cache "invalidation" (in this case, more of an "initialization") to load
the current DnsSession (if there is one) into the ResolveContext. This
will allow DoH requests and other transaction operations requiring
per-context state to work if the context is created after the last
config load or connection change.

Bug: 1022059
Change-Id: I6c6eddd9ae992430b5c72f00caf03e9694be56a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068833
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Eric Orth <ericorth@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745243}