Fixed color API bug in ExtraInsetsBorder.
Fixed a bug in ExtraInsetsBorder where the ExtraInsetsBorder was
composing the original border but neglecting to set the color in
its own Border superclass.
This resulted in the border painting correctly but returning the
incorrect placeholder color when asked for its color through the
Border's API.
Added unit test to guard against regression.

Bug: None
Change-Id: I71613ed468384dc3780e2acb76a43351e96d3bac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2069498
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745246}