DNR: Add a max memory limit for individual regular expression filters.
This CL adds a maximum memory limit for individual regular expression
filters using the re2 |max_mem| option. The limit is initially set to 2
Kb.

Manifest regex rules exceeding the limit are ignored and will cause an
install warning instead of a hard error. To prevent a lot of warnings,
we limit the number of such warnings to 10. In case of more than 10 such
rules,  we'll show a single warning listing all the affected rule IDs.

New dynamic rules added using the
chrome.declarativeNetRequest.updateDynamicRules call and exceeding the
limit will cause an error.

Earlier indexed rules (static or dynamic) which now exceed the memory
limit will be ignored silently. This will be modified subsequently to
raise an error in the extension console.

BUG=974391

Change-Id: I759d651b851200613c8d3b2e4c7146fdd4749a6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2032028
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745317}