Fire UIA_SelectionItem_ElementRemovedFromSelectionEvent when unselected
Currently, when we select an item through SelectionItemProvider, Windows
Narrator announces "selected" but when we remove the selection, we do
not hear "non-selected". This is due to not firing
UIA_SelectionItem_ElementRemovedFromSelectionEvent when we unselect.

This CL adds firing UIA_SelectionItem_ElementRemovedFromSelectionEvent
for non-multiselected selection item, so we can properly notify Windows
Narrator of the unselect action.

Bug: 928949
Change-Id: Idbce6a2ecaaf3ff90231cf1a07d7a3bd06a23475
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2066591
Reviewed-by: Ian Prest <iapres@microsoft.com>
Commit-Queue: Victor Fei <vicfei@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#745319}