media/gpu/vaapiVEA: Manage VASurfaceID by ScopedID
VaapiVEA manages VA Surfaces with the ids. They need to be destroyed
by vaDestroySurface(). The ids are pooled for recycling. Some
ids has not returned to the pool, so that they are never
destroyed. This CL fixes the leakage by managing VASurfaceID
by ScopedID.

Bug: b:143957628
Test: video.EncodeAccel.*
Test: Open & Record & Close CCA multiple times
Change-Id: I86743fe80973fc3e850d543e38dcfe2764b2a710
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2060388
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745351}