Centralize invalidation for style attribute changes.
Style attribute changes could affect style matching of other elements
than the one that gets its style attribute modified because of selectors
which have a style attribute selector in the non-right-most compound.
For instance:

  [style] + div

In this case the div will be invalidated by an invalidation set for the
style attribute scheduled in StyleEngine::AttributeChangedForElement().
This method was not invoked for all variants of style attribute
modifications. In particular setting the text for Element.style.

This CL centralizes the invalidation handling to cover all cases.

Bug: 1056151
Change-Id: Iceb70c284e2bb6b209a4243d61f941670eab323e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2078532
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745382}