Implemented some functions for NewLinkedHashSet
Added size, Capacity, isEmpty, front, back, insert and erase for
NewLinkedHashSet.
NewLinkedHashSet is a new implementation of LinkedHashSet.
LinkedHashSet maintains a linked list
to iterate all entries in the order in which they are inserted.
We replace it with NewLinkedHashSet for the following purposes:
To simplify moving of the backing store.
To implement LinkedHashSet on top of existing WTF primitives.
In NewLinkedHashSet, the linked list is implementing in a vector
(with links begin indices instead of pointers).

Design Doc for NewLinkedHashSet is here:
https://docs.google.com/document/d/1ts2L8XJcGQYlxdVerVBp8_625h6ky7RNGw21AqMJ9Z0/edit?usp=sharing

Change-Id: I327143051ff1f2a1cd36cc9f3ff26923267c5694
Bug-Id: 992506
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2065815
Commit-Queue: Kei Nakashima <keinakashima@google.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Auto-Submit: Kei Nakashima <keinakashima@google.com>
Cr-Commit-Position: refs/heads/master@{#745383}