Move code to send the variations header to the renderer into content
This allows it to be shared by Chrome, WebLayer and WebView. For Chrome
this functionality is tested by VariationsHttpHeadersBrowserTest.

To support this a VariationsClient has been added, with an accessor on
BrowserContext, which allows the variations code to tell if the user is
signed in (some of the variations header logic needs to know this).

Bug: 1025612
Change-Id: Ibb5a0dd54e23a3536de4c0fb84624173a6974262
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967245
Commit-Queue: Alex Clarke <alexclarke@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745404}