[CoreAudio] Set stats for CoreAudio workaround to never expiring.
As long as the work-around to handle serialization of CoreAudio suspend
and resume tasks is in place (see crbug.com/772884), we will need these
stats. When the issue is resolved by Apple and the work-around can be
removed, we will also remove the stats along with it.

Fixed: 1055397
Change-Id: I867bb80891c462e05a6c8c97834f54612d47ccde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2071269
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Oskar Sundbom <ossu@chromium.org>
Commit-Queue: Armando Miraglia <armax@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745408}