Fixes to wpt/svg/styling/presentation-attributes-* tests
Missing negation in the "d presentation attribute not supported on other
elements" subtest in ...-special-cases.html test.

Set 'irrelevantElement' to null for 'd', 'height', 'r', 'rx', 'ry',
'transform', 'width', 'x' and 'y' since they are all mentioned in the
table in the spec[1] as being special. They are also covered by the
...-special-cases.html test.
The definition of 'Applies to' for the 'transform' attribute varies
slightly between the SVG specification and CSS transforms
specifications, so this applies essentially the intersection of those
two.

[1] https://svgwg.org/svg2-draft/styling.html#PresentationAttributes

Bug: 1056055
Change-Id: Id7a6e604d4f48abbd601bdc09ec9f7ff795b86ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2078581
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#745452}