[DeviceSync v2] Cache encrypted local device metadata
When a successful SyncMetadata response is received, cache the encrypted
and unencrypted local device metadata, along with the encrypting group
public key, that was sent in the SyncMetadata request. During the next
SyncMetadata request, if the local device metadata and the encrypting
group public key have not changed since they were last cached, reuse the
cached encrypted local device metadata.

Why are we doing this? Because the ECIES encryptor uses a different
session key for each encryption, the blob could change even if the
underlying metadata and group public key have not changed. We do not
want the CryptAuth server to act as though device metadata has changed
if the underlying data and encrypting key remain the same.

Fixed: 1054654
Change-Id: Ie94247295c9d76d984a861ed934219a5496bf268
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2079437
Commit-Queue: Josh Nohle <nohle@chromium.org>
Reviewed-by: James Vecore <vecore@google.com>
Cr-Commit-Position: refs/heads/master@{#745471}