[Passwords] Generalize SortKeyIdGenerator
This change generalizes the SortKeyIdGenerator to take arbitrary types
as keys, and allows customization of the id type and key comparator.
This will be useful when we need to generate ids for compromised
credentials as well. To make the name more generic the class is renamed
to IdGenerator.

TBR=schwering@google.com

Bug: 1047726
Change-Id: I1b9e25aa4ccd0a732c810bde20163d46d5a46d67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080430
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745485}