Allow Editor Bubble to take an anchor Rect (Editor Bubble refactor Part 2)
WebUI doesn't use TabGroupHeaders for tab groups, so Mohnstrudel needs another way to position the Editor Bubble. BubbleDialogDelegateView allows for an alternative way to set an anchor, which is to give it the anchor bounds as a Rect. This CL exposes that method of anchoring the bubble while still defaulting to using the TabGroupHeader in the desktop UI.

Bug: 1055537
Change-Id: I9d38971f867a362007abd626aa425515bef16af3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2071911
Commit-Queue: Connie Wan <connily@chromium.org>
Reviewed-by: Charlene Yan <cyan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745542}