Improve DoH server availability tracking
Now that server availability and stats are tracked per-context, DoH
servers only need to be disabled for consecutive failures. And now that
availability behaves much more similarly to normal server failure
tracking, the availability has been mostly merged into the normal
tracking instead of using a separate flag.  A DoH server is now
considered "available" (eligible for use in automatic mode) if
|ServerStats::last_failure_count| is less than
|kAutomaticModeFailureLimit| and there has been at least one success
recorded on the current connection.

Net effect is that DoH servers can be enabled or failure-tracking reset
for any successful request or probe to that server on the same context,
even if a failure disable has taken effect (eg a transaction started
before the server was disabled, or a new probe when we restart probes in
a future CL).

Also, previously the DoH probe logic reset availability when restarted.
Because the probes were generally only restarted for connection change,
this had the effect of resetting availability for connection change.
Now, the ResolveContext handles that itself on connection change by
resetting the flag that it has seen at least one success on the current
connection.

Bug: 1022059
Change-Id: I872057cbb9f7a0a41f580dff57d906b1e4e5d33c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062676
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Eric Orth <ericorth@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745551}