MSE: Rename and clarify the MediaSource attachment methods
AttachToElement() name was unclear: while it technically attaches the
MediaSource to the Element, the underlying MediaSource cannot be used
for appending media, etc. by the app until it is opened with a call to
SetWebMediaSourceAndOpen().

This change renames these two methods, respectively, to
{Start,Complete}AttachingToMediaElement(), along with some comment
updates. Actual function remains unchanged.

This clarification CL is expected to improve readability and review,
especially of the in-progress MSE-in-Workers CLs.

BUG=878133

Change-Id: Ic726cbee71d92e82a5a8c7d3be7d8d9185d1fb76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2079384
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Commit-Queue: Matthew Wolenetz <wolenetz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745588}