Settings: Mark all entries in SettingsRoutes typedef as always defined.
Previously every member of that typedef was marked as potentially
undefined. This required a type-casting whenever a !settings.Route
was expected, which was tedious without producing any type-coverage
benefits.

As a result of this change a lot of typecasts like the following

/** @type {!settings.Route} */ (settings.routes.FONTS)

are no longer necessary and are removed.

Bug: 1026426
Change-Id: I0becc11cc8a5479c4390d4499e4956f73126c8ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2078822
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Auto-Submit: dpapad <dpapad@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745625}