Fix url load params for StartupTabPreloader
Really, the code in StartupTabPreloader should be reusing the Tab
Creator code, but that's a bigger refactor and this needs to be merged
back (and also I don't want to do it :) ).

The initial preloader implementation forgot to set the transition type
for the LoadUrlParams, which happened to default to a value that mostly
works, but failed to record that we were coming from an intent, and so
it broke intent redirects.

The test for this case didn't catch it because the test extraneously
started Chrome to a blank page before running the test, as that's what
the test suite did.

(cherry picked from commit a941f0fe4272646e9a34c6cc3ba96488e12d2835)

Bug: 1039612
Change-Id: I38e918f6e6576a37f5b1c0c893497f3f1b488560
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067246
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#744532}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080812
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/branch-heads/4044@{#548}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}