Dynamically get span count for message card in GTS
http://crrev.com/c/2023197 makes message card adapt to the span count
when it changes. This is achieved through updating the span size
based on spanCount that we calculate when orientation change is
detected. However, there is a corner case where the span size lookup
happens but the orientation change is not detected and therefore
the spanCount is not updated. This CL fixes this issue by letting
SpanSizeLookup() to dynamically get the current span count from
GridLayoutManager instead of replying on a calculation when
orientation change is detected.

Also, this CL adds some render tests for the IPH message card so that we
have some coverage on the orientation change related logic.

Bug: 1050494
Change-Id: Ie86a8583db926c41a46451a1c5eff3c6a4a5ec54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2076550
Commit-Queue: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Auto-Submit: Yue Zhang <yuezhanggg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745683}