media/gpu/vaapi: Create VAImage with va surface size on UploadVideoFrameToSurface()
This is a workaround of the failure at 1080p video encoding
test on grunt.

Originally vaCreateImage is created with a coded frame size.
However, it causes SEGV_MAPERR on vaPutImage, which copies
vaImage to vaSurface.

As a result of experiments, the vaImage size needs to be the
same as va surface size. Since this is not stated in VA-API
document, so this might be a bug of AMD gallium driver.

Besides of the workaround, this CL copies visible size area
of VideoFrame to the VAImage.

(cherry picked from commit 95ad9e5a05acab4b62cc1d273aec114a7312640c)

Bug: 1048908
Bug: 1050377
Bug: b:148744040
Test: video.EncodeAcel.* on grunt, eve and soraka
Test: camera.* on grunt, eve and soraka
Change-Id: I1ee853964387846d2452e72cfbd4c939f0a18f62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041382
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#744953}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2082215
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/branch-heads/4044@{#561}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}