Replace deprecated getLastUsedProfile function in Sync Settings.
|Profile#getLastUsedProfile| is deprecated and replaced with
|Profile#getLastUsedRegularProfile|. This CL updates
Sync Settings classes.

Since settings is considered as always running on regular mode,
it is safe to use regular profile even called in incognito mode for
initialization of HelpAndFeedback activity and checking whether profile
is child profile. Therefore, it is safe to leave behavior as-is today
and no need to make change on code behavior to act differently for each
profile.

Note: This change is only code clean-up, does not change any behavior.

Change-Id: I130ae3c5daedaf29ef412234c1d04594d1ea9093
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2072179
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Commit-Queue: Side YILMAZ <sideyilmaz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745911}