Update PageInfoController to use incognito profile when called incognito.
This CL update the profile passing to TrackerFactory from
PageInfoController. Currently, it is always regular profile, however it
is not safe because of possible data leakage from incognito to regular
profile.

Therefore, this CL uses the profile (i.e., regular or off-the-record
profile) received from webcontent in PageInfoController, then passes it
to TrackerFactory.

Bug: 1041781
Change-Id: Icbfa68762c0fede68de217621e054472fbdcfb61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080254
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Commit-Queue: Side YILMAZ <sideyilmaz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745961}