[SyncSetup]: Record a histogram for sWAA displayed to the user.
This CL adds a histogram to record the displayed sWAA value in
|chrome://settings/syncSetup| for sync users.  We neglect the short
period of time while sWAA is not fetched or is being fetched. In case of
failure to fetch sWAA, we display nothing.  In case of success, we
either show on or off with an explanation of the reason for which sWAA
is off. The histogram is recorded every time sWAA is
calculated/refreshed.

Bug: 1050070
Change-Id: I6ad75626edc590707c2594d86db4cb61a2069196
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062421
Commit-Queue: Monica Basta <msalama@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745969}