[Passwords] Add Infos about Compromised Credentials to Extension API
This change adds obtaining information about compromised credentials to
the passwordsPrivate extension API. This is done via a getter, as well
as an event that is fired when the credentials change. Furthermore, this
change introduces a new PasswordCheckDelegate class that handles the
implementation of these APIs.

Bug: 1047726
Change-Id: I77d06a030cc823e665d5c49dec0eaed192104a65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2072742
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746021}