Don't create subsequence for the document element
The document element is always contained in LayoutView. The subsequence
for LayoutView and the early return in full-cached situation in
GraphicsLayer::PaintWithoutCommit() (for pre-CompositeAfterPaint) make
the subsequence for the document element useless.

This CL avoids paint chunks that were forced for the document element
by the subsequence. These paint chunks are useless for layerization in
CompositeAfterPaint because they should be always merged into the
scrolling contents layer of the LayoutView, but had overall drawback on
raster invalidation.

Many paint invalidation web tests have raster invalidation clients and
reasons. Some tests progress with less raster invalidation because of
stabler paint chunk ids.

paint/invalidation/css-grid-layout/grid-item-z-index-change-repaint.html
regresses with more raster invalidation when the document element's
negative-z-index children becomes positive-z-index, which is rare and
acceptable.

This is also a preparation of crrev.com/c/2070688 which will fix
transparent background hit test in CompositeAfterPaint, to avoid
extra paint chunks that would otherwise be created for the document
element in that CL.

Change-Id: Ic72c4ad4c1e0ce6cd967973b1158965f549580e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083380
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746070}