Promote Mac10.14 Tests to main waterfall
Mac10.14 Tests is now stable and green, and is ready to be promoted
to the main waterfall alongside its prior-release siblings. This CL
makes a few changes to accomplish that:
1) Updates the OS of the existing main waterfall Mac Builder to 10.14
2) Moves Mac 10.14 Tests to be triggered by Mac Builder
3) Moves it from just master (like all FYI bots) to exist in branch CI
   as well
4) Moves it from FYI to CI consoles

This change leaves Mac Builder Next intact, so that it can be updated
and used for 10.15 soon.

Bug: 1018852
Change-Id: I3d5723ab7c565c37681696acb2f2de79ddcac1f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036769
Commit-Queue: Aaron Gable <agable@chromium.org>
Auto-Submit: Aaron Gable <agable@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746074}