[WebNFC] Re-check permissions in the browser process
This change adds an NFC permission check in NfcHost and ensures that the
Mojo connection passed to the renderer is scoped to the lifetime of the
RenderFrameHost which brokered it.

This ensures that if the renderer navigates or the tab is closed the NFC
connection will be broken immediately, even if for some reason the
NfcProxy in the render process is not garbage collected. The connection
is also closed immediately if the user revokes NFC permissions.

(cherry picked from commit afbe2244300a07412f1735a764be7342b2ad71ba)

Bug: 1035048, 1047838
Change-Id: I73d05c112674a82dd15146403e4ad03765ee1fbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2051384
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#743254}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2081520
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/branch-heads/4044@{#573}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}