Create Service with new name if creation with original name fails.
We create the service with a new name if the creation with the original
name failed. I think this should take care of failures, and may perhaps
fix issue 1023080 as well. The idea here is that in some cases, a
deletion of the original service may not be complete and there may still
has some stale state left over in the SCM database. In such cases, we
may be unable to open it or recreate it. In these cases, we now will
create a new name and install a new service.

Bug: 1055470
Change-Id: I7e0f5e76892fe29fa597c77cd4b64702a3f717c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2071387
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: S. Ganesh <ganesh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746101}