[m-81] Show back button when IME is showing
This was broken a while ago but went unnoticed.

We should show the back button, pointing downwards, so users can get out
of the ime VK.

Another change will be landed for M-82 (ToT as of this writing). The
M-82 change will be different because we have to account for removing
the navigation buttons, which is enabled in default in M-82.

Bug: 1056334
Change-Id: I170f1c24fc405d7ff322ada5a9816cde4e198078
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2079266
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Commit-Queue: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/branch-heads/4044@{#580}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}