Break connections before removing from active_source_handlers_.
In DeferredTaskHandler::BreakConnections, we want to remove finished
handlers and break the connection.  when a finished handler is removed
from active_source_handlers_, it might be deleted, but we were still
using that to create the connection.  Instead, break the connection
first and then remove it.

Manually ran test from the bug and it passes with this change.  Without
this, it failed right away.

Bug: 1057593
Change-Id: I3c9346a6842f412100d608876adb268befb80470
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083436
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746142}