Fix ASTC HDR WebGL extension.
Tested on Pixel 3 XL (with HDR support) (validating) and Nexus 5X (without HDR support)(validating/passthrough) with up-to-date WebGL ASTC test.
Not tested with passthrough angle on Pixel 3 XL due to the build config and cmd line settings for passthrough angle
not work properly locally (reported to Geoff). But given that angle exposes the
extension with the same name as OpenGL it should be fine.

Bug: 1049880
Change-Id: Iadc1a78591068927bbed9523131924187f53fe3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080823
Commit-Queue: Shrek Shao <shrekshao@google.com>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746146}