Set unrecognized values of "scrolling" attribute to kAuto
Prior to this CL, unrecognized values of the "scrolling" attribute
would leave the prior value unchanged. That behavior is counter
to the spec [1], and is also confusing to developers, as it
introduces state to the "scrolling" attribute. This CL fixes
that.

[1] https://html.spec.whatwg.org/multipage/rendering.html#the-page

Fixed: 1057454
Change-Id: Ib6b4cd4c6952fe580003dc2577c72db2fe622fae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083595
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746151}