WebUI NTP: move scroll border code into utils
Moving the scroll border code out of the customize dialog makes it
easier to test. If there is a flake, it will also be easier to diagnose.

This CL enables the ntp-customize-dialog unit tests which was previously
disabled due to flakiness (due to the scroll borders tests).

Bug: 1056688
Change-Id: Ic74f4b79c33c578f9a9d759d80ed569c93f12550
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2082901
Commit-Queue: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Auto-Submit: Esmael Elmoslimany <aee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746221}