Fix test flake for Crostini settings page.
The flake was in the setting page's existence test for the
"Upgrade Available" section. The existence is determined by a value that
is initialized from loadTimeData and updated from a webUIListener.

There was a race between loadTimeData's default "true" value and the
webUIListenerCallback's replacement "false" which would change right
when the page was loading, so the settings test would see the
occasional "false" instead of expected "true" and fail.

Bug: 1057823
Change-Id: I3822eda06c5631c2328f80891f9f5e5de4821172
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083773
Commit-Queue: Nicholas Verne <nverne@chromium.org>
Auto-Submit: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746236}