[iOS] Refactor WebStateListWebUsageEnabler to be a Browser Agent
This CL refactors the BroswerState keyed service that handles enabling
and disabling web states (WebStateListWebUsageEnabler) to be a Browser
Agent. The API is unchanged, aside from accessing the agent via
FromBrowser() instead of FromBrowserState().

The URLLoadingService unittest needed to be heavily updated for this
change:
   - The unused TabModel mocks were removed.
   - The test was updated to do work in the constructor and destructor,
     rather than SetUp and TearDown. Thins includes moving most of the
     member init to an init list.
   - The test now uses no-parameter TestBrowser constructor which
     creates and owns its browser state and web state list.

The accessory classes for WebStateListWebUsageEnabler (the factory and
abstract base class) are also removed.

Bug: 1032556
Change-Id: Ib973d31253c5c92586184089db31d780083a8fb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062970
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Mohammad Refaat <mrefaat@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746311}