[Invalidations] Use Topic API in InvalidationLogger
This CL is a part of ObjectID -> Topic API migration. It replaces
ObjectID with Topic in InvalidationLogger and its dependencies.

There should be only one behavioral change: chrome://invalidations page
would expose 0 as an ObjectID::source for any subscription. It's not a
real change, because |source| was already deprecated and the same value
used. In the future chrome://invalidations should be updated to use
Topic instead of ObjectID.

chrome.invalidations.logInvalidations now receives
TopicInvalidationMap::ToValue() instead of
ObjectIDInalidationMap::ToValue(), but their ToValue() implementations
are the same, so there should be no changes.

Bug: 1029698
Change-Id: I17f36f7c43b664185bf87da4edb8d7d99a0afec9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2074877
Commit-Queue: Maksim Moskvitin <mmoskvitin@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746338}