[Passwords] Save manager should check for opt-in rather than sync state
Whether saving in the account store should be guarded by the current
settings not the current state. This is relevant for when the user
just opted in butter, it may take couple of seconds till the passwords
get synced. We shouldn't however block storing in the account store
in the meanwhile.

Bug: 1012203
Change-Id: Ieac2799b88ab8d8d64dee84bbd51450af82790ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083453
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746355}