Make Tasks executable outside of TaskQueue
Task now exposes Execute(OnceClosure) instead of Run(), and can be
called outside of a task queue, or from within another task.
This change was made so that tasks to be composed of multiple
smaller, nested tasks.

We plan to use this feature in components/feed/v2. An 'upload' task
can be run on its own, or as an intermediate step in a 'load' task.

Change-Id: Icbe64021b8fecd8abbc09c64d2d15dfab2e2708c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2078786
Commit-Queue: Dan H <harringtond@chromium.org>
Reviewed-by: Justin DeWitt <dewittj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746395}