Avoid Network Error Screen on "Back" in SAML PIN
This fixes the UX glitch that pressing the "Back" button
in the SAML smart card PIN dialog (<security-token-pin>
Polymer element) was triggering the "Network Error" screen.

The underlying issue is that cancelling the PIN dialog
indeed aborts the TLS handshake with an error, but it
doesn't make sense to show this error to the user.

The proposed fix is to remember whether the PIN dialog was
canceled during the current authentication attempt, and
suppress the network error in that case.

Bug: 1052409
Test: start SAML login via a smart card, proceed to PIN dialog, press "Back", verify that no Network Error screen is shown
Change-Id: If54dbd067da763afca7ceec4957bebf0d748da9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2069542
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746435}