Stop writing in dtor in SkiaRenderer BufferQueue.
After BeginWrite was moved out of critical path
SkiaOutputDeviceBufferQueue::Image is often in the middle of write.
When resize or reallocation of display happen (e.g rotate screen on
Android) we need to call EndWrite on the image before destruction to
ensure all necessary semaphores are submitted.

Bug: 1056221
Change-Id: Id55c25f43cbb4bf082ad3e3ef6cd1b59e164937d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083835
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Peng Huang <penghuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746438}