WebUI Tab Strip: Update swipe animation to only happen after tab closes
Previously, the swipe animation was mapped to where the user's pointer
would be and would make the tab element disappear before the close event
was confirmed. This CL sets it up such that the animation does not
progress and only detects when a swipe happens.

This is to address a bug where the front-end is not able to confirm
whether the tab was successfully closed, due to an unload dialog.

Bug: 1050747
Change-Id: Ia0a4658260ebd73a05bd616b4a8df5d4c4e2525e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083403
Reviewed-by: dpapad <dpapad@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746452}