[Payment Handler] CanMakePaymentEvent.respondWithMinimalUI() #7
Patch #7: On Android, pass currency to the CanMakePaymentEvent when
minimal UI feature is enabled, so the updated WPTs can pass.

Before this patch, the "currency" field of the CanMakePaymentEvent was
set only on desktop.

After this patch, the "currency" field of the CanMakePaymentEvent is set
on both desktop and Android when the minimal UI feature is enabled,
e.g., through chrome://flags/#enable-web-payments-minimal-ui flag, and
the relevant WPT passes on both desktop Android when opened in a full
browser.

Bug: 1005076
Change-Id: I9d24b97730c10bc4e6f4ce38ced89a3c33bb1a14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2079686
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Danyao Wang <danyao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746477}