[remoting] Create encoder before notifying of captured frame.
This ensures the encoder is created before passing the captured frame
to the scheduler. This is to allow the scheduler (in future CLs) to
examine properties of the encoder before sending the frame to be
encoded. Such properties could include:
* Preferred quantizer values for normal and big frames.
* Whether big-frame-detection is handled by the encoder.

Bug: 891571
Change-Id: I28c32a1944913ca6dec240ab483b9a65d38f9576
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2085120
Auto-Submit: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746495}