gpu: Plumb CreateSharedImage to use SurfaceHandle
This CL begins the work of adding the ability to pass a SurfaceHandle to
the SharedImageInterface. In some circumstances, we previously got
around this by first creating a gpuMemoryBuffer, and using the
CreateSharedImage that uses the gpuMemoryBuffer. This is problematic
because the BufferQueue needed to keep track of the gpu memory buffer,
instead of the gpu process doing it for us.

This CL modifies four interfaces: SharedImageInterface,
SharedImageFactory, SharedImageBackingFactory, and ImageFactory. They
have all been modified to accept and pass through a SurfaceHandle. This
allows us to remove the |buffer| from BufferQueue::AllocatedSurface, as
well as the |gpu_memory_buffer_manager_| from BufferQueue.

This is a mechanical change that is not intended to change behavior. I
followed these rules when modifying the interfaces:

1. Any previous call to CreateSharedImage or CreateAnonymousImage was
passed a SurfaceHandle if applicable, gpu::kNullSurfaceHandle if not.

2. The implementations of SharedImageBackingFactory were modified to
accept a SurfaceHandle. Only SharedImageBackingFactoryGLTexture passes
the SurfaceHandle to ImageFactory::CreateAnonymousBuffer

BUG=958670
TEST=autoninja -C out/Default components/viz:viz_unittests

Change-Id: I5f1263a6421761c8a2805d4ab087c35f848cfe17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2070820
Commit-Queue: Paulo Warren <pwarren@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746529}