Added the chooseFileSystemEntries externs
This is an upstream of web_app_file_handling.externs.js to cl/294762966.

Additional context:

chooseFileSystemEntries is still in the draft stage and as such is not
known to closure. This cl adds the function to the native file
handling externs file and gives it a type definition matching the wicg
spec.

https://wicg.github.io/native-file-system/#api-choosefilesystementries

Bug: 1040328,996088
Change-Id: I6f3d2efacc092d36149866f05fccac9cb6e98ada
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2084039
Commit-Queue: Zain Afzal <zafzal@google.com>
Reviewed-by: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746673}