Migrate AccessibilityMsg_FatalError from the legacy IPC to mojo
Based on crrev.com/c/1847304, by Colin Blundell <blundell@chromium.org>

This CL starts converts the message that kills the renderer, on
top of the infrastructure already laid out by crrev.com/c/1991446.

In addition to that, this CL implements the required changes to move
to a model where the mojo pipe for RenderAccessibility is only valid
while a kWebContents-compatible accessibility mode is enabled, in order
to address some security-related concerns. And it also adds a new unit
test to check that this new logic works as expected (i.e. that the mode
changes in the renderer when the pipe is closed from the browser side).

Bug: 1010831
Change-Id: Ic71c33f654af9fbc1c453d2aa04c858e8285c1a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015045
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Cr-Commit-Position: refs/heads/master@{#746702}