Return all windows in GetAppWindowList()
Currently the following windows are not included as we use the same
logic as building a list for the cycler:
- PIP
- Non-activatable windows
- Windows with HIDE_IN_OVERVIEW key.

In MAPIP, the PIP window is not included in the list returned by
BuildWindowForCycleWithPipList as we don't want to show both the pip
and original window in the cycler.
The actual issue we are facing is, currently we use this list to
query Chrome windows from Tast, but as PIP is not included in this
list, Tast can't find the PIP window.

BUG=b:145552924
BUG=b:141899229
TEST=tast --verbose run dut arc.PIP

Change-Id: I16d67f14d8f3bf666e8c2b9a7c2ee99b6740a509
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026911
Commit-Queue: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746737}