Update ImageDownloadCallback comment for empty |bitmaps| parameter
This CL updates the comment for WebContents::ImageDownloadCallback to
specify that if the image fails to download the callback will receive
an empty |bitmaps| vector.

This detail is necessary for:
https://chromium-review.googlesource.com/c/chromium/src/+/2072344/

At the same time this CL reformats the parameter comment to match the
general style used in Chromium.

Bug: 926083
Change-Id: Ide1d38ff21be467085373e366423a8a27390cec4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2086363
Auto-Submit: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Commit-Queue: Jochen Eisinger <jochen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746777}