build: Fail early if NaCL is enabled on macOS 10.15+ hosts.
The NaCL SDK is 32-bit only. This situation is unlikely to change,
because NaCL is deprecated and slated for removal from the codebase.

macOS 10.15 removed support for running 32-bit binaries. Since the NaCL
compiler cannot run on macOS 10.15 or above, attempting to build
Chromium with enabled_nacl=true on this OS currently fails with obscure
errors.

This CL adds a GN assert with a clear error message that includes an
easily actionable step for developers encountering this situation.

Bug: 1049832
Change-Id: I2d4bccc7110cdb1f54cf302327cce0858e01bbb4
Tested: Checked out the error message on macOS 10.15.3.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2087430
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746905}