Filter ignorable codepoints from missing glyphs
Default ignorable codepoints should not be considered for
choosing a fallback font.

see:
 http://www.unicode.org/L2/L2002/02368-default-ignorable.pdf
"""
  Default-ignorable codepoints are those that should be ignored
  by default in rendering (unless explicitly supported). They have
  no visible glyph or advance width in and of themselves, although
  they may affect the display, positioning, or adornment of
  adjacent or surrounding characters.
"""

The API used to determine fallback fonts on Mac is
CTFontCreateForString(...). The API is not able to match default
ignorable codepoints like ZWJ and NZWJ. GetFallbackFont(...) was
failing when an ignorable codepoints is present and the expensive
GetFallbackFonts(...) is used instead.

Bug: 1036652
Change-Id: If44c5b8887a7186863d8e40bde5d4ba4f94ed0f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2076439
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746906}