Read basic printer info differently for macOS in CUPS IPP code
CUPS destination names and info are organized differently of macOS and
other CUPS platforms. On macOS, the destination name is stored in the
"printer-info" option, while the description (usually the make and
model) is stored in the "printer-make-and-model" option. In contrast,
on other platforms, the name is reliably stored in the |name|
attribute of the cups_dest_t struct, and the description is stored in
the "printer-info" option.

Additionally, populate the |device_name| attribute of |printer_info|
in CupsPrinter::ToPrinterInfo(), so the name gets plumbed up to the
print preview UI.

Meanwhile, consolidate some common CUPS consts into
print_backend_consts.h.

Bug: 226176
Change-Id: Idb71bf7ad7602dcf185d4f49d81294b13d70365d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2066297
Commit-Queue: Daniel Hosseinian <dhoss@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746974}