[remoting host] Check channel last active time
This CL makes the host log the last time it receives an FTL pong before
sending out a heartbeat to help us debug and see if the unresponsive
host is caused by a dead FTL connection. It also CHECKs that to make
sure the channel isn't inactive for too long.

Bug: 992722
Change-Id: I2c93266ee6cc7586e9192ca8fde98e1ded37dd67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2086322
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#746999}