gfx::BufferFormat:: s/BGRX_1010102/BGRA_1010102/
gfx::BufferFormat::BGRX_1010102 was landed as a Chromium-name for
XR30 to support GMBuffer allocation of GL_RGB10_A2, particularly
on Mac. This was a bit of a misunderstanding on my side: we should
operate using an alpha format and only use the opaque format when
trying to import a buffer of this type via EGL on CrOs Intel,
similar to e.g A/XB24 etc (See GetFourCCFormatForOpaqueFramebuffer()).

This CL does **only**:

- sed -i -e 's/BGRX_1010102/BGRA_1010102/g' `grep -rn BGRX_1010102 cc components gpu media ui | cut -d: -f 1 | uniq`
- sed -i -e 's/image_xr30/image_ar30/g' `grep -rn image_xr30 cc components gpu media ui content | cut -d: -f 1 | uniq`

Leaving fixing resource_format_utils.cc and gl_image_io_surface.mm
for the new format, and the DRM related changes to the follow up
CLs.

TEST=display and gl_unittests passing and
http://crosvideo.appspot.com/?codec=vp9.2_30&loop=true&mute=true
reproducing correctly on Mac.

Bug: 949260
Change-Id: I9ccb3d39d4c4b6ac4465fa575cb6e5001100b040
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068722
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747000}