Reland "Add a testing config for ReducedReferrerGranularity"
Landing a field trial testing config to change the default referrer
policy led to a number of pings from perf "regressions". Erik Chen
suggested that this might be due to Web Page Replay not acting well when
page JS starts following new code paths (which latter might very
plausibly have happened because the change in question sometimes alters
the return value of a commonly-used JS API). The plan is to reland the
field trial testing config---which re-enables the change in tests---and
re-baseline the perf stories that recorded large (positive or negative)
metric changes: see crbug.com/1056329.

== Begin original commit message

We're going to 50% beta on this feature
(changing the default referrer policy to
strict-origin-when-cross-origin) so it's time to add a field trial
testing config.

Also updates the following tests, which were failing with
the testing config enabled:
- ContextMenuBrowserTest.OpenInNewTabReferrer
- DownloadReferrerPolicyTest.DownloadCrossDomainReferrerPolicy

Bug: 1014207, 1056329
Change-Id: I6fb0006e1f89fd7f2943f102f72caf54fb389716
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044713
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747024}