Move PermissionPromptAndroid to //components/permissions
This required moving a few of the other permissions java files as well.
A PermissionsClient.java class was added to call into embedder specific
logic in AndroidPermissionRequester.

We may not need to support infobars in WebLayer, so the bulk of the
infobar logic has been kept in //chrome.

The xtb files for the android strings were generated from the same
script that was used in http://crrev.com/c/1966700.

Bug: 1025609
Change-Id: I6ea6f7380cee4de6590b9a906852fa4b2ce8f89f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083714
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747032}