Add browser_test to test allow sign-in behavior during sync setup
During sync setup the toggle which controls chrome sign in should be
disabled in an on state, signifying to the user that this control cannot
be changed during sync setup.

Previously this behaviour was tested manually. This CL extends the
existing browser_test to confirm this behaviour.

Bug: 1043948
Change-Id: Idd54b3ad0fd83e70b7ba356aa11dae66fdb27e52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2088372
Reviewed-by: dpapad <dpapad@chromium.org>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/master@{#747064}