Create dialog that shows a Parent Permissions prompt
This dialog's purpose is to prompt a parent of a child-type user to
enter their credentials to approve an action.  This dialog will also
allow you to pass in an extension::Extension, whose details (i.e.
permissions, title) will be displayed above the credential entry form.

This CL is inspired by the functionality of
ExtensionInstallPrompt, but simplifies the approach used by
that code.

Bug: 957832

Change-Id: I3a9395f56505b43cd5c8f8c269b0b19c99ce6379
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1972417
Commit-Queue: Dan S <danan@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747083}