Redesign Privacy Settings Card
As part of the Privacy Settings Redesign, controls which are currently
present on the main privacy card have been moved to subpages. This CL
adjusts the main privacy card to contain links to these subpages and
remove any now duplicated controls.

All of the changes in this CL are hidden behind the
privacySettingsRedesign flag. Test have been added to ensure the
existing behaviour is preserved when the flag is off.

Bug: 1032584
Change-Id: I174cd6b404c9d704165e3a8d7991e8429088dbe7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2088233
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747092}