Allow ManifestIconDownloader to download images from a specified frame.
Before this CL, ManifestIconDownloader always downloads icons from the
main frame. This creates a buggy interaction with Content Security
Policy, when the manifest icon download is initiated from inside a
frame.

After this CL, the callers of ManifestIconDownloader can specify an
initiator frame (as a GlobalFrameRoutingId) in which to download the
image. The frame routing ID is passed to WebContents to find the correct
frame that exposes the mojo.ImageDownloader interface.

This CL also exposes the mojo.ImageDownloader interface in all frames
instead of just the main frame.

A subsequent CL will update the callers of ManifestIconDownloader to
pass in the correct frame routing ID.

Bug: 1055360
Change-Id: I31fd58453392ec4e1165ac20dc3a1284dc05f6f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2085752
Commit-Queue: Danyao Wang <danyao@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747105}