Add a11y function/event to terminalPrivate extension API
Crostini terminal needs to know when a11y spoken feedback is on so that
it can also turn on its a11y features. However, the SWA (system web app)
terminal cannot access the "accessibilityFeatures" extension API because
it does not support SWA. To fix this problem, we add a new function and
an event to terminalPrivate API to allow the terminal to get and
subscribe to a11y status.

Bug: 1019034
Change-Id: Ief4a5edceed470eac8d5b85d72f4ae581ce2dcb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2086357
Commit-Queue: Jason Lin <lxj@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Ben Wells <benwells@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747130}