Add DCHECKs to make sure ServiceWorkerRegistry runs on correct threads
This is a preparation for moving SpecialStoragePolicy out of
ServiceWorkerStorage.

We are planning to take an approach similar to crrev.com/c/1945940.
ServiceWorkerRegistry will own SpecialStoragePolicy and observe policy
changes. We will need to observe policy changes on the IO thread so we
will add a method to ServiceWorkerRegistry which runs on the IO thread.

For now, all methods in ServiceWorkerRegistry run on the core thread.
This CL adds DCHECKs to annotate the fact explicitly.

Bug: 1055677
Change-Id: I767296f045a6d3d454c25cdaf18587073a111209
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2087320
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747134}