Remove use of generic SyncableService::GetAllSyncData in unit test
This is a preparation for the removal of said method. Although these
unit tests seem to use an arbitrary SyncableService at first glance,
inspecting SyncValueStoreCache shows the object is actually always one
specific implementation of the interface (SyncStorageBackend). This CL
downcasts the return value of SyncValueStoreCache::GetSyncableService
so that the tests don't depend on the SyncableService interface having
the method.

Bug: 1057577
Change-Id: I8b6e674d28d4b28829b519caf1d930c89552f5c7
Fixed: 1057577
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2084991
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#747163}