UIA: Hide an embedded object's children when all ignored
An issue came up with CreateAdjacentLeafTextPosition(), used to create
boundaries start/end positions in ax_position.h and help with text
navigation. It was caused by the UIA-only concept of embedded object
replaced by the empty object character when all children of a node are
ignored.

One of the conditions for a node to be considered as an embedded object
was to look at its child count: if it had at least one child, it could
not be an "empty object". However, a corner case where a button (or any
other control) contains an empty div forces us to modify this
condition.

Consider the following example:
****************************************
<button>                               |
  <div></div>                          |
</button>                              |
                                       |
would result in the following AXTree:  |
                                       |
++kButton                              |
++++kGenericContainer ignored          |
                                       |
****************************************

Because the button node contains one child, it is not considered as
an empty object replaced by character. However, because this button has
no text content and its child is ignored, nothing will be exposed to
text range provider.
Furthermore, CreateAdjacentLeafTextPosition() will not be able to
create a position in the button even though it should. This will result
in the button being either skip (thus inaccessible) or causing
Narrator to get stuck on this node when in Scan Mode.

This CL fixes the issue with a two steps solution:
1. If all children of a node are ignored, we might consider it to be an
   empty object replaced by character even though it has a child or
   more. Remember that this logic is only activated by UIA.
2. I introduced AXPosition::IsLeaf, which returns true when a node has
   no child or is an empty object replaced by character.

Now, when we call CreateNextLeafTextPosition (or any other similar
method) right before a button that contains an empty div, the
position created will be inside the button node and not inside of its
leaf ignored node.

Final note: it is important from now on to call IsLeaf() instead of
AnchorChildCount() in ax_position.h to determine if a node is a leaf.
We shouldn't notice a perf difference, since
IsEmptyObjectReplacedByCharacter is of constant complexity (+ it has
a shortcut that returns false when called by something else than our
UIA implementation).


Bug: 928948
Change-Id: If95281745d1c8dda3e646dba00923791cc1af7e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2079619
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Ethan Jimenez <ethavar@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#747168}