[Android] Move WebappScopePolicy out of ExternalNavigationHandler.java
This CL is another step toward the componentization of
ExternalNavigationHandler.java for sharing with WebLayer. It removes
the dependence that ExternalNavigationHandler.java currently has on
//chrome's WebappScopePolicy.java. This dependence is to assist in
determining whether the user is navigating within the context of an
installed PWA within which the intent should stay. At this point in
time, this is //chrome-level functionality. Hence, we abstract this
check through ExternalNavigationDelegate.

Bug: 1031465
Change-Id: I65869090553f2cf337db39cfbb9aaac6d341f7e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2087678
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747175}