[Android] Photo Picker: Add a simple RenderTest.
The test selects the first photo, scrolls to the bottom and back up
to the top to make sure the selection border is active and of correct
size (old bug I encountered once).

This should also function as a regression test for bugs like the one
seen in issue 1051523 (where a white circle got overlaid on top of
all photos).

Bug: 1051523, 656015
Change-Id: Ifb119c5e9450be5ef06b303eb3549315c0e4a623
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083178
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747188}