Convert ViewMsg_EnablePreferredSizeChangedMode to mojo
This CL ends up also converting ViewHostMsg_DidContentsPreferredSizeChange
to mojo as well. The reason is because it happens that
blink.mojom.LocalMainFrame.EnablePreferredSizeChangedMode() can trigger the
invocation of blink.mojom.LocalMainFrameHost.ContentsPreferredSizeChanged(),
but the later can also be invoked independently of the former, hence it is
not a "completion callback" like mojo message.

Target mojo interfaces are LocalMainFrame/LocalMainFrameHost, respectively.

Some of the methods and class variable in RenderViewImpl
got moved as-is to WebViewImpl, including:

- RenderViewImpl::OnEnablePreferredSizeChangedMode()
- RenderViewImpl::UpdatePreferredSize()
- RenderViewImpl::send_preferred_size_changes_
- RenderViewImpl::needs_preferred_size_update_
- RenderViewImpl::preferred_size_

The call flow now for mojo.EnablePreferredSizeChangedMode() is

browser -> mojo EnablePreferredSizeChangedMode -> blink::LocalFrame ->
blink::ChromeClient -> blink::WebViewImpl.

BUG=1051155
R=dtapuska@chromium.org

Change-Id: I3c6fdc73f9d41b30aeee1f6214ee8a0db6abc76e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068699
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Antonio Gomes <tonikitoo@igalia.com>
Cr-Commit-Position: refs/heads/master@{#747190}