Autoupgrade redirects to HTTPS for mixed content
When the mixed content autoupgrade feature is enabled, content that
would have been autoupgraded but is already HTTPS does not get the
UpgradeIfInsecure flag set, which means if it is eventually redirected
to HTTP, that mixed content is not upgraded. This CL fixes that and
adds a regression test.

Bug: 1058077
Change-Id: I359d720ea5caf0833fabd810de7876e32f17b099
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2086258
Auto-Submit: Carlos IL <carlosil@chromium.org>
Commit-Queue: Mike West <mkwst@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Daniel Vogelheim <vogelheim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747318}