MacPWAs BMO: Remove more uses of extensions::Extension
Remove ExtensionAppShimHandlerMac::Delegate::
MaybeGetAppExtension. That's the goal of this patch.

Removing it means no longer playing with extensions::
Extensions directly in ExtensionAppShimHandlerMac. To
accomplish what we were doing by those queries, add
the following new Delegate functions
- AppIsInstalled
- AppCanCreateHost (rename of AllowShimToConnect)
- AppUsesRemoteCocoa
- AppIsMultiProfile

Also rename GetHostForBrowser to
GetHostForRemoteCocoaBrowser, to clarify that it is
only for use with PWAs.

After this, we'll get rid of the
ExtensionAppShimHandlerMac static functions
MaybeGetAppExtension and MaybeGetAppForBrowser,
and move ExtensionAppShimHandlerMac to a separate
place where it can select between extension and
app functions based on id.

R=loyso
TBR=ellyjones (chrome/browser/ui/views)

Bug: 1057195
Change-Id: Ia29810633ec91b7e74a156902fa7f4c952c16eb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2088711
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747321}