Stop distracting firewall error messages
The "Failed to add Windows firewall rules" message shows up when running
tests as non-admin. This error is usually benign but looks like a
failure so it causes confusion when investigating test failures. The
message can be avoided by not trying to add firewall rules when this is
guaranteed to fail (when running as non-admin).

Change-Id: I0980529d3b289f99444d7b2f7ca99440432cefaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2088533
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747359}