Ensure that startup time baseline is zero
DO_NOTHING is supposed to output an empty result,
but it currently reports 100-500 ms.

It seems that setContentView() and Activity seem to cause additional
delays. Posting the measurement task with a delay seems to fix this.

Also, this adds useful comments such as example run and sample results.

Bug: 1055970
Change-Id: I0e458746b1d4051adef9662bddaff2d368493d9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080586
Commit-Queue: Changwan Ryu <changwan@chromium.org>
Reviewed-by: Tao Bai <michaelbai@chromium.org>
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747379}