Fix Shadow DOM v0 crash in style resolution.
StyleResolverState::DistributedToV0InsertionPoint() doesn't mean that
it is a V0 host child. There is no use in trying to make sense of that
now as Shadow DOM v0 is going away, but the fact that it can actually
be a shadow root child means parentElement() might be null.

Bug: 1058605
Change-Id: Ia791f2caa1a4c7c19935135611a1fbdcfb07dc47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089753
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747416}