[BrowserSwitcher] Launch IE asynchronously
Previously, IE would be launched directly from the UI thread. This was
not only causing jank, but also browser freezes when `DdeConnect()'
would take a long time to complete (>1min in some cases).

This CL makes `DdeConnect()' and `base::LaunchProcess()' run on a worker
thread with blocking allowed, so the browser is still responsive in the
meantime.

Bug: 1054895
Change-Id: I305063c12ebce67a9feffecfeb9f1d4c75211cf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2082994
Commit-Queue: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747424}