Delete NetworkThrottleManagerImpl.
This is dead code, no being built, and would fail to compile if it were,
due to the OnceCallback work.  It was added when we were planning to
experiment with throttling requests below the disk cache layer, but that
plan ended up not coming to fruition.  When the other code for the
project was removed, this was accidentally left in place.

Bug: None
Change-Id: I3b124af1b0872a178651cd8cfe4ddb2a37678a0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2088017
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747452}