Add learn more to show shelf navigation setting sub label
Adds support for learn-more-link attribute to settings toggle button -
the attribute adds "Learn more" anchor element to the setting sublabel,
and triggers an event when the link is clicked (used to record the
user action for metrics).

Uses the new attribute in manage_a11y_page for show shelf navigation
buttons setting - "learn more" will link to a help page article, but the
URL has not yet been finalized, so keeping it blank for now.
Also, adds a user action for clicking the link to metrics.

BUG=1050544, 1055688

Change-Id: I4126a6a85bf54c0039d2a14063f83f1e08db0d55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2069847
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747476}