blink: makes block-flow handle more line-clamp
This makes block flow's line-clamp handling correctly add ellipsis
when if clamping happens on a line where are the text fits, but
there are subsequent lines. Common case for this is when using
'white-space:pre'.

This does not give the exact same behavior as old. In particular,
external/wpt/css/css-overflow/webkit-line-clamp-008.html is an example
of a case that now renders differently (nested divs). See
https://test-results.appspot.com/data/layout_results/mac-rel/302753/webkit_layout_tests%20%28with%20patch%29/layout-test-results/results.html
for the set of failures.


BUG=none

Change-Id: Ie188d5b3efc1817377f7ba5523815ba6bdc4a986
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050093
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747530}