Roll protobuf javalite runtime + fix gn package name
This rolls in the new version 3.11.4 of the protobuf javalite runtime.
The package name changed so this also fixes all references in gn files.
A similar internal commit exists at crrev.com/i/1676922

This also changes how protoc is called to ask for the java lite runtime.
The protoc javalite code gen plugin is no longer needed and will be
removed in a followup cl.

             but reduces them per generated proto.

Binary-Size: new runtime increases the number of methods in the runtime
Bug: 988512
Change-Id: I8dd2ba97badea0b7dbaa8282afeb1da40636aee5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1762684
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747540}