Creating a variants.pyl file for reference from test_suites.pyl
There are a lot of variants that are repeatedly defined. To
minimize duplication in test_suites.pyl, and to keep it readable,
I've extracted the definitions into a variants.pyl.

All variants defined as a string will look for a reference in
variants.pyl.

Change-Id: I5bfc206ead0fd10c3d497d6557761c2795ba76f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090871
Commit-Queue: Jeff Yoon <jeffyoon@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747553}