[Merge M81][WebLayer] Allow running WebLayer and WebView in the same process
The basic idea here is to create a new ClassLoader for WebLayer to get
a new Java address space, and load separate native libraries for
WebLayer and WebView to get a new native address space. On Android N+
we are able to symlink the library directories, and the convinces the
system it is a separate library. On <=M, we need to copy the libraries
AND rename each library to convince the system it is a different lib.

This has been tested manually on L, M, P, and Q in standalone,
monochrome, and trichrome.

TBR=torne@chromium.org

(cherry picked from commit d3020c5b09587f5cdd714dd6518b8ead2e7be635)

Bug: 1051358
Change-Id: I3207ae3de5e64612e67a0114294bd5dae0d3eee1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2040629
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#746961}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090921
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/branch-heads/4044@{#674}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}