DNR: Remove AllowAllRequests_SrcDoc browser test.
r744448 added logic to ensure allowAllRequests were always applied to
srcdoc frames. That said, the added test is flaky. It seems the fetch
can be aborted when the frame unloads causing flakiness. However,
waiting for onload event before the fetch causes the request to arrive
in the browser after the corresponding navigation commit, defeating the
purpose of the test. Since there doesn't seem to be a reliable way to
reproduce requests from a frame arriving before a navigation commit,
simply delete the test.

BUG=1056031

Change-Id: I23fc2fdfe5719a0f10bdce025cf3556840f0c3a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2088346
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747580}