[iOS] Restore items state when restoring session
When a session is restored, the restored NavigationItems have all the
information of the restored items, including scroll offset, user
agent...
But the items were only used for their URL and all the other
information were lost. This CL fixes that by storing those items
until the restore is complete. At this point their information can be
restored in the items associated with the WKBackForwardItem.

For now those information aren't used with kEmbedderBlockRestoreUrl
disabled.

Bug: 771200
Change-Id: Ic04f3eb0a893899acfcfdd46b33c950b2a8e71e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2061151
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747630}