[signin] Bump Signin.IsShared histograms expiration date
This histogram is going to be useful in the short term as we are planning
ont improving the multiprofile experience. This histogram is an indicator
about how often multiple users share a single Chrome profile, which is a
scenario we want to discourage more in the future.

Fixed: 1037291
Change-Id: I8d535b28ac8e8354a1c58c6248177721c70a09d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089760
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747660}