[css-grid] Fix referencing implicit grid line after auto repeat()
The indices of both implicit grid lines defined with grid-template-areas
and explicit ones defined with grid-template-rows/columns used to be
stored together in NamedGridColumnLines and NamedGridRowLines.

However, this was problematic, because the former indices already refer
to the final explicit grid so they don't have to be increased when
expanding an auto repeat(), but the latter ones should.

Therefore, this patch stores the indices in separate fields and uses the
correct logic for each one.

BUG=966090

TEST=external/wpt/css/css-grid/placement/grid-placement-using-named-grid-lines-005.html
TEST=external/wpt/css/css-grid/placement/grid-placement-using-named-grid-lines-006.html

Change-Id: I6d423148af0e4dd865f130742f7a927a325cef90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2081823
Reviewed-by: Javier Fernandez <jfernandez@igalia.com>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/master@{#747669}