Gpu Peak Memory Attribution
We already have measurements in place for the Gpu's peak memory allocation over
various user interactions. However when we see spikes there is no way to
attribute them.

This change builds out attribution for the source of the allocations being done.
It also includes this in the existing trace events.

Bug: 1001168
Change-Id: Ibf9256eae51e076457218ee558aa25b5caa94c68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2024920
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Jonathan Backer <backer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747683}