Hoist empty-list special case out of SVGSMILElement::NextAfter
SVGSMILElement::ResolveActiveEnd() is the only caller of
NextAfter(kEnd, ...), and it already handles the case of an empty list,
so hoisting this check makes it more obvious that 'indefinite' is the
initial value, and the failure to find a proper instance time means
failure.
This makes NextAfter() less dependent on what list it is searching,
which will help split it out (and optimize certain operations on it).

Bug: 1039243
Change-Id: I8fbc8b86521df53b32ea67bbd5e3281e261246cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089867
Auto-Submit: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747696}