DownloadURLParameters: Remove NetworkIsolationKey parameter.
Downloads are currently treated as main frame navigations, with respect
to same-site-cookies, but as subresources, with respect to
NetworkIsolationKey. This means we don't gain any cross-site tracking
protections as a result of the NIK, and there are often mismatches
between NIKs and their site-for-cookies values.

This CL makes downloads act like navigations with respect to
NetworkIsolationKey, to mirror the site-for-cookies behavior.  This
does mean that downloads leak data across sites, and we may want to
address this in the future, but for now, we should at least be
consistent, since NIK and site-for-cookies are similar concepts, and
should have complementary behavior.

This is one of several CLs aimed at letting us DCHECK that NIK and
site-for-cookies are consistent.

Change-Id: I6e35d189c87dc6b04b35ca4ef16170e65a07f990
Fixed: 1046453
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050987
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747743}