[tbmv3] Remove json config and use proto annotations
Now that trace processor can read and emit proto annotations, we no
longer need a separate json config.

I turned the description into comments, and it will likely stay that way
because descriptions are rarely used. We eventually want the
requiredCategories to be a file/message option instead of the comment,
but right now trace processor only supports field options and not
file/message options, so we'll have to wait until that's implemented.

Bug: 1012687
Change-Id: I4ef713ebef151e959b9c373a220f9f2630415ed5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090452
Reviewed-by: Deep Roy <dproy@chromium.org>
Reviewed-by: Mikhail Khokhlov <khokhlov@google.com>
Commit-Queue: Deep Roy <dproy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747777}