[Switch Access] Persist back button focus ring
A bug was causing the focus ring around the back button in the Switch
Access menu to disappear after about a second. This happened because
the FocusRingManager clears the focus ring around the back button when
it's not the primary focus.

This change fixes that bug by moving all direct calls to the menu panel
into a single class (the MenuManager), so it can maintain the focus
ring when the back button is an auxiliary focus.

Bug: 1056785
Change-Id: I2e3173ba819b4b1b270b40e1f935d70d028b3a20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2079141
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747831}