[Build] Fix OS check for generate_linker_map.
The build flag |generate_linker_map| is useful for SuperSize. There is
an assert in toolchain.gni to limit it to {android, linux}, but it's
unclear whether this should be for current_os or target_os. Attempts to
change this led to some past churn, e.g.:

  https://chromium-review.googlesource.com/c/chromium/src/+/1038430
  https://chromium-review.googlesource.com/c/chromium/src/+/1715024

This CL relaxes the check to allow all combos of
  {current_os, target_os} * {android, linux},
in hope to put the problem to rest.

Fixed: 986842
Change-Id: Id44bd76a900f0f55e10efc8bc25d75ac52e92f44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089787
Commit-Queue: Samuel Huang <huangs@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747848}