Prevent fullscreen while dialogs are up (1/2)
Chromium already drops fullscreen when a dialog is first displayed.
Extend that behavior so that a WebContents may not enter fullscreen
for the duration of a dialog's display.

This is part 1: Extend the fullscreen IPC so that the browser can
decline a renderer's request to enter fullscreen.

Bug: 1042210, 1020026, 1037730
Change-Id: Iafba087b22c51bf3c8fb6f9a4ce02921d51f0044
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2041871
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747870}