[Tab Groups] Add new tabs in other browsers to the end of the tabstrip.
Currently, the tab_strip_index is not set allowing the
tab_strip_model_order_controller.cc to decide on where to insert the tab
(which is currently based on the last active index). This can insert
a tab in the middle of a tab group breaking group continuity.

Bug: 1049195
Change-Id: I9a7f7cacdcb01845cd736818bc07d0efc9a594e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080079
Reviewed-by: Charlene Yan <cyan@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Connie Wan <connily@chromium.org>
Commit-Queue: Charlene Yan <cyan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747890}