AW: treat asset, resource, and chrome:// URIs as internal
WebView permits apps to load their own Android assets and resources with
file:///android_asset/<name of asset> and file:///android_res/<name of
resource> respectively. Because these are meant to access the app's own
resources, we should treat them as internal to WebView and not search
for apps which expose intent filters matching this pattern.

This also avoids a crash. If another app targets N+ and exposes such an
intent filter, the embedding app will crash with
FileUriExposedException.

This does the same with chrome:// URLs, which was previously overlooked
because this can only happen if clicked from a frame loaded on the
chrome:// scheme (which can happen if the app uses loadDataWithBaseUrl).

This implements the fix for the default shouldOverrideUrlLoading()
implementation (in WebView) and WebView shell's
shouldOverrideUrlLoading() (which is meant as a best practice for apps
to follow).

Bug: 1059193
Test: Manual - built a test app to let me click these URLs, verified
Test: default shouldOverrideUrlLoading works as intended
Change-Id: I797c4b45cf15bf6c519bf55a848ed646456a6fd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2092416
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747945}