[Switch Access] Make panel window transparent
The specification for the Switch Access menu (see bug for full details)
specifies that the menu should be translucent, with rounded corners.
To accomplish this, we need to make the underlying view transparent.

Previously, the menu showed with a white background filling the
bounding box.

Bug: 973719
Change-Id: I101bb6203fd8c972f7b235821e0f0a51cafa25e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2085596
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/master@{#747952}