Add enterprise policy for ScrollToTextFragment
This change adds a policy for ScrollToTextFragment, and plumbs a
content switch down to blink, where it is used to override the state
of blink::features::kTextFragmentAnchor.

This policy is not marked as dynamic_refresh, since the values are set
early on in renderer startup, with no ability to change dynamically.

Bug: 1055513
Change-Id: I28aae1eb036e94d68974423c871b1d806275d03b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083383
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#747987}