Support typemaps inlined as GN targets
This introduces support for typemaps specified as GN targets, with new
cpp_typemaps and blink_cpp_typemaps variables that mojom() targets can
used to reference their typemap rules.

The underlying work to use the typemap configuration is shared with the
existing typemap infrastructure, but the net result is that we no longer
need separate .typemap files or global "bindings configurations" once
everything is converted to this approach.

Bug: 1059389
Change-Id: Id2e5fe765d3c7600a3f50e337fb693f1b3a3cc0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090716
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747990}