Add a Trust Tokens request signing helper.
This CL implements the request signing operation of the Trust Tokens
protocol by adding a signing helper to //services/network. Request
signing involves the following steps:
1. Retrieve a Signed Redemption Record (SRR) from storage and attach
it as a request header.
2. Optionally, add a timestamp header.
3. Construct a canonical representation of the request, including a
collection of the request headers specified by the caller, and compute a
signature over this canonical representation using a stored public key
associated with the SRR.
4. Attach this signature---but _not_ the request's canonical
representation, which server-side consumers will be able to
reconstruct---as a request header.

The Trust Tokens design doc [*] contains the normative description
of how to construct this canonical signing data.

[*]: https://docs.google.com/document/d/1TNnya6B8pyomDK2F1R9CL3dY10OAmqWlnCxsWyOBDVQ/edit#heading=h.6a92f2gfl9le

Bug: 1042962
Change-Id: I0acf0a7ba29d193013411db911f612d694667a17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050756
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Reviewed-by: Chris Palmer <palmer@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748051}