media/gpu/chromeos: Add VdVideoDecodeAccelerator.
Currently GpuArcVideoDecodeAccelerator bridges the video codec from
ARC++ to a VDA instance. We plan to deprecate to VDA implementation at
chromium by new VD implementation. So we temporary need an adapter to
convert between these two interfaces, until the
GpuArcVideoDecodeAccelerator is ported to the new VD interface.

This CL implements VdVideoDecodeAccelerator as the adapter. With this
class, GpuArcVideoDecodeAccelerator could bridge the video codec to a
VD instance via the adapter.

Bug: b:136716838
Test: Pass video_decode_accelerator_tests with following CLs
      on Kevin and eve

Change-Id: I17f8607cc422c682eaad7760f5680a425de90095
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1913228
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748086}