Mark the key repeat events as handled in exo::Keyboard::OnKeyEvent().
Clients of exo expects to handle the key repeat in the client side
according to crrev.com/c/1079768.
Therefore, exo::Keyboard silently ignores the key repeat events.
However, it doesn't mark the key repeat events as handled and
they are used to invoke the accelerators.
This CL changes exo::Keyboard to mark the key repeat events as handled,
then they are not used to invoke the accelerators anymore.

Bug: b/149477038
Test: Hold Ctrl and W on an ARC++ app which eats Ctrl-W key event, and
check that the app doesn't close.

Change-Id: Iab480b94b98ea8d3e12ce67b810162f1f671e3b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2091328
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Yuichiro Hanada <yhanada@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748134}