Clang packaging script: don't build lld with thinlto
This shaves a few minutes off the clang packaging time, simplifies our
build script, and brings the build config for lld closer to that for
clang so that we can potentially build them in one go in the future.

Building lld without thinlto does not appear to affect its performance.

Bug: 884608
Change-Id: I4c3c39aa935fdb0a0977b1dd46400eee02d76dde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089835
Commit-Queue: Hans Wennborg <hans@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748150}