Implement PWA.Visit.WebAPKableSiteVisit metric.
See http://go/webapk-retention-metrics,
http://go/webapk-retention-metrics-impl for details.

The eventual hope is to augment this metric so that it records
page-loads to installed experiences as well, and record a boolean value
along with the metric so we can distinguish between the scenarios. This
will give us a more direct way to see exactly what proportion of
installable page-loads per app are taking advantage of the benefits of
installation.

It seems like that might require some refactoring to do properly, so
I'm going to do the easy part first so we can start getting at least
some numbers ASAP, then we can look at improving the situation later.

Bug: 1044709
Change-Id: I472012cffe4d0860457a586b2b01a3e7c573a342
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2057108
Commit-Queue: Glenn Hartmann <hartmanng@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748204}