blink/bindings: Extend and align expiration of ParkableString histograms.
The "Memory.ParkableString.*.5min" histograms are useful to get data
about how the ecosystem is changing, as they mostly reflect the behavior
of sites. In addition, they do not take a lot of space either on the
client or on the server as they are recorded once per renderer for
renderers surviving more than 5 minutes.

As a consequence, extend these, and align their expiration milestone.

Bug: 1025871, 1025856, 1025856, 1025844
Change-Id: I1d9a613b0c458c491dde795fe51e837ae2abc725
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2094131
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748210}