[Kiosk] Refactor AppLaunchSplashScreenHandler so it works more reliably
This handler was little bit overcomplicated. This led to a bug in web
kiosk mode when we update network state, but do not update the handler.

Because of this convolution, in web kiosks, the network configure
screen did not update itself when connected to network(which is what
client will expect from the handler which observes network state).
Now it behaves more logically and predictably.

Bug: 1015383
Change-Id: I94eb7d8c1ff60f0f5b4a5843f6e3b2885c63b776
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083018
Commit-Queue: Anatoliy Potapchuk <apotapchuk@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748240}