User data snapshot: Delete snapshot when user deletes browsing data
When the user deletes browsing data, snapshot files are deleted if they
are in relation to the deleted data. Prefs are deleted if one of the
deleted data also affects the prefs. Snapshotted data created before the
start of the time range where browsing data is deleted will not be
affected.

Bug: 1052788
Change-Id: I560a844183905a58cdab2c2266869dd91c5b0f63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2066564
Commit-Queue: Yann Dago <ydago@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748254}