[TTS] Add a chrome://flags for client/server debug
Adds a new Feature ContextualSearchDebug that enables internal debugging
for the client and the server.  This should help us debug issues with
the Definitions feature and upcoming server-supported features.

This adds a flag to chrome://flags and will send an integration
code to the server to enable debugging for the backend requests.

We already have a mechanism to handle integration codes like this
but the server will need to be updated to handle this code for
debugging.

Hopefully the experimentation framework can send a bit automatically
to the server to indicate this debug feature is required for the
server request.  If that works we can simplify this code a bit.

Currently there is no client-side debugging enabled when this
flag is flipped, but we'll probably add some eventually.

BUG=1059004

Change-Id: Ib39911e410d36aa2ffc3874722885ba3a5cd5917
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2026421
Commit-Queue: Donn Denman <donnd@chromium.org>
Auto-Submit: Donn Denman <donnd@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748300}