Adds metrics for policy global size.
The backing store for low level rules uses a fixed allocation which must
be satisfied for all process launches on Windows. This adds plumbing
and a metric to track how much of that allocation is really necessary.

The plumbing can be removed once the metric is no longer needed.

Bug: 1059129
Change-Id: Id02a1002ef006fc14ff9197aed91465baedd08ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090850
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748312}