Cros: Use floating values for obtaining icon bounds
Because we are scaling the icon bounds to animate long press gestures,
some values might be incorrectly rounded.
Change the method that calculates the bounds with scale to use floating
values and then Round the resulting rect to adapt to grid layer.
Also, include the icon shadows for scaling.

Bug: 1050730
Change-Id: Ief19fcf0d2894f570a77047f8209e53eeb461b73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2086317
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748337}