Update maps_pixel_test docs
The maps_pixel_test makes use of Skia Gold Triage links, but not only as an
artifact. Not actual baselines.

For those unfamiliar this can be confusing, as the lack of positive image in
Skia, is juxtaposed with there only being a json of expectations in code search.

To clarify this I'm adding some docs to the maps_integration_test.py to have a
note that the actual pixel failures will be reported in the logs, and that the
Skia Gold Triage is a supplimentary artifact.

TEST=maps_pixel_test

Change-Id: I1c802ba2d4d021890333cbe9c857e54824a0476a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095207
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748391}