Enable GDI patching only for processes which require it.
GDI patching (inconsistently) results in delayload crashes due to the
sandbox restrictions applied to various processes. In an effort to catch
all processes which need the patching, we applied it to all utility
processes. However, this has proven to be too inclusive, as many of
those processes have sandboxes which prevent this code from working
properly. This change detects the sandbox type and only executes the
patching if the sandbox type is ppapi, PrintCompositor, or PdfConversion.

Bug: 1028670
Change-Id: I651f50d073a94bf3a7c1cfa9f01aa21761c68132
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083384
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748400}