assistant: Defer screenshot request
When showing Assistant UI, we used to request the screenshot. This will
significantly reduce the smoothness of Launcher animation.
This cl defers the screenshot request to the time when users click the
"What's on my screen?" suggestion chip.
We still request and cache the Assistant structure of current active
window before showing the Launcher, which has not shown noticeable
effect on the smoothness.

Bug: b/145218971
Test: Added new AssistantScreenContextControllerTest
Change-Id: I1bb6c7597414d4a981cf0ce8df50eb5e5b981427
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2013498
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Jeroen Dhollander <jeroendh@google.com>
Commit-Queue: Tao Wu <wutao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748466}