Make search box clear icon focus visible (files-ng)
Icons for the action bar buttons are controlled using -webkit-mask-image
applied to a child of the buttons. The clear button in the search box
had the mask applied directly to the button which caused any border
applied by focussing to be masked out.

Add a child <div> to the clear button to be consistent with all the
other action bar buttons.

Applies the mask to the child div instead of the containing clear
button.

Tbr: noel
Bug: 1059012
Change-Id: Ie0273eb97b06fb4b588ce523a06673d5606fe056
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2094427
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748477}