[webauthn] Fix GetAssertionRequestHandler DCHECK
When handling a response on GetAssertionRequestHandler, it's possible
the device may be able to use PIN for fallback and not wait for a second
touch if it supports internal UV but not UV token. Fix a DCHECK to avoid
crashing in that case.

This patch also improves the internal user verification tests to verify
behaviour when PIN is supported.

Bug: 1059914
Change-Id: I5be2e4c85ee97f6298bc5f34ec111a757a027cc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095151
Commit-Queue: Nina Satragno <nsatragno@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#748511}