[Modularization] Remove InfoBar dependency on SnackbarManager
InfoBar provides an accessor method to get to SnackbarManager.
TranslateCompactInfoBar is the only infobar that uses this method. Having this
method means that InfoBar needs to know how to get to SnackbarManager instance.
Modularized InfoBar shouldn't have access to ChromeActivity.

This CL removes InfoBar dependency on SnackbarManager and makes obtaining
reference to SnackbarManager a responsibility of the calling component (in this
case TranslateCompactInfoBar).

BUG=1056346
R=mdjones@chromium.org

Change-Id: I33f366ed49dfd6ad51c406a7a6d0e8ffea056d03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095708
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Pavel Yatsuk <pavely@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748557}