oobe: Use show() to show pages in gesture navigation screen
Oobe dialog elements have show() method that focuses the default element
in the dialog, and updates the dialog scroll. Use it in gesture
navigation dialog to show pages (which are oobe dialogs).

This fixes the issue where the scroll background is show on the intro
page (as the scroll is calculated using obsolete bounds).
Upadating scroll also made the animation pages scrollable, so this tweaks
the animation positioning in the screen.

While here, mark next buttons with focus-on-show class, so they get
focused when a screen page is shown.

BUG=1058885

Change-Id: I23c44cb03afd713c35aa0a83f9fc59081a8a7b41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2092948
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Matthew Mourgos <mmourgos@chromium.org>
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748561}