Add user-consent view for Quick Answers
Add a user-consent view informing the users about the Quick-Answers
functionality. The quick-answers will only be enabled if the user
grants her consent by acknowledging this view. The view will be shown to
the users a maximum of 3 times or until an accumulated of 8 seconds it
is impressed.

Bug: b:148956341
Test: Tested locally on atlas device
Change-Id: I46d0cfd5d82ced03ef610ab8624b909fa1403223
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2077589
Commit-Queue: Abhijeet Singh <siabhijeet@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748567}