[iOS] Reland "Fix bookkeeping for RecentTabs presentation"
This CL is a Reland of crrev.com/c/2071938
This CL makes sure that the animations of the RecentTabs are disabled
when the recent tabs aren't presented when BVC is contained.

The original CL wasn't working because the updates weren't prevented
when the class was instantiated. If the user is signed in, the first
presentation is trying to update the TableView even if it wasn't shown
yet.
Fix it by preventing updates in -init.

Bug: 1038034
Change-Id: Id2ec3e63dd78a014a87abf91ec99fe001b27df4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095116
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748601}