Add SMILInstanceTimeList abstraction/wrapper
Add a simple wrapper class for the Vector<SMILTimeWithOrigin>
representation of an instance list, and move all list mutating
operations to it.
This will allow easier tracking of which "origins" are present in the
list, and thus avoiding certain operations in some cases. This will be
added in a follow-up CL.

Bug: 1039243
Change-Id: Iac126f0f6628696fea436514dcab5d824c1dc588
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089765
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#748605}