Avoid redundant scans of the instance lists
When the instance lists grow to a sizable length, checking for and
removing instance time with a kScript origin can take significant time
in some cases - especially when there are no such instance times present
at all.
Add a bitmask encoding which origins are present in the list, and then
use that to avoid redundant scans (in RemoveWithOrigin) when possible.

Bug: 1039243
Change-Id: I7008927338508f08e8b416822fbe52073a86da1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089706
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748627}