[ios] Convert WebStateListObserver::ChangeReason to a real enum
WebStateListObserver::ChangeReason was sometime used as an enum and
sometime used as a list of flags (as reflected by the definition of
the value that are power of two).

As there is no reason to use it as flags (since the values are not
compatible), convert it to a proper enum (without specifying the
individual value of the enumerators). Also convert it to an "enum
class" so that it is not convertible to int types implicitly.

Remove CHANGE_REASON_NONE value since it can never been passed in
a meaningful way (it made some sense when used as a flag since it
could be combined with other) and convert the uses of the constant
to the proper value.

Bug: none
Change-Id: I1a955fd558208cc3d6e108781b407d1a23fa8676
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089891
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748647}