Make LayoutSVGResourceFilterPrimitive inherit directly from LayoutObject
This is one of those "we're only in it for the ComputedStyle (and
notifications)" LayoutObject, so inheriting via LayoutSVGHiddenContainer
only adds overhead. Inherit directly from LayoutObject instead. While
doing this also rename to LayoutSVGFilterPrimitive to avoid any
confusion with LayoutObjects for actual resources.

Change-Id: Icfc293d3ece94d56cc2136975f49bd21e7e8d32f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2094304
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748686}