gpu: Reset share group in LoseAllContexts
Follow up to
https://chromium-review.googlesource.com/c/chromium/src/+/1960465

With change above and with virtual context enabled, it is fairly likely
that when a context is lost, the new context is created before the old
context is destroyed. Since GpuChannelManager keeps using the same
GLShareGroup, the new context will end up in the same share group as the
old context.

Avoid this by creating a new GLShareGroup in LoseAllContexts.

Bug: 1059286,1052485
Change-Id: I1d60578fa6359329406d634fe04c7f242314839a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090137
Commit-Queue: Bo <boliu@chromium.org>
Auto-Submit: Bo <boliu@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748708}