Use GetPaintRecord in SVGFilterRecordingContext
This cuts out one of the middle steps as well as the additional clip to
|bounds|. We will be clipping the results of the filter anyway, so
hopefully this extra clip step doesn't matter too much. When the full
'filter' syntax is supported on SVG content we won't be able to use the
filter region clip in this way anyway.

Bug: 109224

Change-Id: I8cf8799c8c0992b763291f57759ad7d83f98d258
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2096448
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#748716}