Extend/remove histograms emitted from //net/extras/sqlite
This change extends expiry for a number of histograms emitted from
the Cookie and Reporting/NEL databases in //net/extras/sqlite,
and also removes a couple of them.

They are useful as a debugging aid for users reporting problems
loading/saving cookies.

The one that is removed is {Cookie,ReportingAndNEL}.DBSizeInKB.

Fixed: 1053115, 1052954, 1037004
Change-Id: I6eb0229758d7791e97dbff1e9b69195e14e760dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095438
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748717}