WebGPU: Make DissociateMailbox take device client ID
Previously it didn't take a device client id, which meant the map
associating shared images to texture reservations in the wire was global
to the WebGPUDecoderImpl. This is incorrect because there is one wire
per device.

This CL adds the missing argument to DissociateMailbox and makes the
associated shared image map a per-device thing. It also adds a
regression test.

Bug: chromium:1059777
Change-Id: Id4db2556dc19f90a13e1af6381f855f7d633e113
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095123
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748731}