[Web Payment] More private methods.
Before this patch, the RespondWithCallbacks class in
payment_app_provider_impl.cc contained a number of public methods that
were either not used outside of this class or not used at all.

This patch moves a number of RespondWithCallbacks methods from public to
private and removes one method that was never used.

After this patch, it is more clear which part of RespondWithCallbacks is
used outside of the class.

Bug: 1005076
Change-Id: I1058137899177454cb786a91e16fe7f88dc8eaee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089409
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Danyao Wang <danyao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748740}