[Live Caption] Add entrypoint in settings page.
Move the entrypoint to the live caption feature to accessibility
settings, which is the location specified by the PRD. Include the
description text, specified by the PRD. The entrypoint allows
live caption to be enabled on Linux and Windows only. Screenshots
are included and uploaded to Google cloud for translation purposes.

Bug: 1055150
Change-Id: I0c645e9273db95a2e5ec929960bc842b6b3e8cae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090100
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Cr-Commit-Position: refs/heads/master@{#748807}