Prevent fullscreen while dialogs are up (2/2)
Chromium already drops fullscreen when a dialog is first displayed.
Extend that behavior so that a WebContents may not enter fullscreen
for the duration of a dialog's display.

This is part 2: Modify WebContents::ForSecurityDropFullscreen() to
support a span of time that fullscreen is prohibited and modify all
callers to correctly request that span.

Bug: 1042210, 1020026, 1037730
Change-Id: I9d2ccc1e459cf37bfbf3499063d87d93ef9910e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044658
Commit-Queue: Avi Drissman <avi@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Christopher Thompson <cthomp@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748808}