Update ToolbarButtonInProductHelpController to use incognito profile when called incognito.
This CL update the profile passing to TrackerFactory from
ToolbarButtonInProductHelpController. Currently, it is always regular
profile, however it is not safe because of possible data leakage from
incognito to regular profile.

Therefore, this CL uses the profile (i.e., regular or off-the-record
profile) belong to active tab in ToolbarButtonInProductHelpController,
then passes it to TrackerFactory.

Bug: 1041781
Change-Id: Ia140d6a63b63325515af9164c2cc89b7fe559ad8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2078477
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Side YILMAZ <sideyilmaz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748880}