Simplified logic for GetFrameColor() in BrowserNonClientView
Simplified GetFrameColor() logic to a single GetColor() call to the
browser frame's ThemeProvider.

GetFrameThemeProvider() will return the app's ThemeProvider if
the AppBrowserController is present. The AppBrowserController's
ThemeProvider will query the AppBrowserController's ThemeSupplier
first as part of its GetColor() method before falling back to the
default.

The effect of this is that the desired GetColor() functionality can
be simplified down to the single call.

Bug: 1056916
Change-Id: I44cfc977c4fb651732a02f5badc688c621889da7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2091940
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748883}