AW: use distinct task affinities for each channel
Previously all WebView providers on the device would use the same
(global) taskAffinity for developer UI. This means launching Dev UI from
one channel would obliterate the previous Dev UI session, re-using the
original channel's icon for the task (which may no longer be correct).

The simplest solution to ensure we always have the right icon is to let
different channels launch in different tasks. We achieve this by using
the package name as part of the taskAffinity. This uses a "." instead of
a ":" delimiter because ":" is apparently an illegal character for
taskAffinity (can't compile).

Fixed: 1060013
Test: Manual - build 2 local builds and verify I can launch each
Test: side-by-side
Change-Id: If8b1def9a4d8af9187a117f39816dcce2c77913b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095930
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748885}