bookmarks: Simplify visibility of permanent nodes
This removes BookmarkPermanentNode::set_visible(), which was misnamed
in the sense that it didn't directly determine what
BookmarkNode::IsVisible() returns, but instead the returned value for
the case where the node has no children.

As it turns out, none of the callers need to change this attribute
(i.e. the visibility when empty) dynamically. For the managed node, it
needs to always be false, and the remaining permanent nodes take this
value from BookmarkClient upon construction.

BookmarkClient gets a minor refactoring to adopt more accurate names
and to allow evaluating the visibility of permanent nodes prior to their
construction.

Change-Id: I4435140e80216800ce1dcee5604e8370bcc2cedc
Bug: 1060311
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095125
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748891}