Construct FeedStream in FeedService
This is glue code to connect FeedStream to its dependencies.
The idea is to glue up everything we can in components/,
leaving the minimal set of dependencies for gluing up in
chrome/browser/android.

Not all dependencies are ready, so this is just a start.

Bug: 1044139
Change-Id: If88f5814df55737a39626454e5a25e7bdf65e7e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2081188
Commit-Queue: Dan H <harringtond@chromium.org>
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748970}