[ios] Creates InfobarBanerContainer to inform of Banner dismissals.
- If the Coordinator that owns the InfobarBannerVC is destroyed before
the InfobarBannerVC finish disappearing then infobarBannerWasDismissed
won't be called. This means that the InfobarContainerCoordinator in
charge of queuing the InfobarCoordinators/Banners won't know its time to
present a new Banner. In order to achieve this an InfobarBannerContainer
delegate is set directly on the Banner.

- If the Coordinator that owns the InfobarBannerVC still exists once the
InfobarBannerVC has dissapeared then we should let it handle the next
banner presentation (e.g. Translate might present another Banner)

- When InfobarCoordinator is presenting a Banner, we currently do not
hold a strong reference to it in the completion block, so if the
InfobarCoordinator (along with the transition/presentation objects)
gets destroyed then the presentation looks like Bug 1057125. We also
do not do some cleanup tasks like
configureAccessibilityForBannerInViewController, which are needed.
Since stop is called whenever the Coordinator is destroyed then
badgeDelegate should be nil and we shouldn't be trying to update the
badge then.

Bug: 1057125
Change-Id: I12f3437a2bf23d2e54ff05a146d2bd4bd0459e55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090711
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749017}