Fires a focus event when navigation completes or is cancelled
In BrowserAccessibilityManager, we were relying on the presence of other accessibility events, such as kChildrenChanged,
in order to tricker the firing of a focus event.
However, when page navigation is cancelled, or is simply completed after the page contents have been loaded
in the accessibility tree, there are no other events that would tricker the focus event. This creates a situation where
the screen reader does not move its virtual cursor to the web contents. As a result, focus appears to be nowhere, and the user
develops the mistaken belief that the virtual buffer in screen readers such as NVDA is also blank.

R=dmazzoni@chromium.org, kbabbitt@microsoft.com

Bug: 1058959
Change-Id: I29150977d032152d5337889b1ae0c2e196fda065
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095758
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Ian Prest <iapres@microsoft.com>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749022}