MixedContent: Remove cross-origin workers WPTs
This CL removes mixed-content wpts for cross-origin workers.
Cross-origin workers are not allowed regardless of its security policy,
so there is no need to test it.

The auto-generator was unintentionally changed to be able to delete
cross-origin workers wpts in the previous CL:
https://chromium-review.googlesource.com/c/chromium/src/+/2049953
but, the previous CL didn't appropriately delete these tests.
This CL actually deletes such tests. (generator.py is not changed.)

It is discussed here:
https://chromium-review.googlesource.com/c/chromium/src/+/2077319/8/third_party/blink/web_tests/external/wpt/mixed-content/gen/top.meta/opt-in/sharedworker-module/cross-http.no-redirect.https.html#b1

Bug: 906850
Change-Id: I50570a1d41437e7212655252c2f3eae4042dce3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2097682
Commit-Queue: Eriko Kurimoto <elkurin@google.com>
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749081}