[SyncSetup]: Force feature flag value in tests.
We have two tests for sync controls one with sync setup friendly
settings on and another off.  This CL enforces the value of the flag for
each of the tests, it was assumed that the flag is 'disabled' by
default. This will no longer be true after enabling field trial testing
for sync setup friendly settings.

Bug: 1043128
Change-Id: I9f40497ee7eab256bf53e4518c115d6356f3d53f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2096739
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Commit-Queue: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749092}