[Android] Alter ExternalNavigationHandlerTest's overriding of feature
Currently ExternalNavigationHandlerTest overrides the value of a
feature in ChromeFeatureList.java via the functionality that the latter
supports for enabling features in test contexts. However, we will be
porting the feature in question to be defined via a //components-level
ExternalIntentsFeatureList.java that will be as minimal as possible
(see crbug.com/1060097 for context); in particular, we do not want to
have to port over the functionality that supports enabling the feature
in a test. Thus, this CL changes the test to enable the production
functionality in question via an explicit testing API on
ExternalNavigationHandler.java.

Note that all of this will go away if/once the functionality is enabled
by default and the feature is removed altogether, which is work under
development (see crbug.com/839751).

Bug: 1031465
Change-Id: Ibcdf9e0e72b7c7564ee067625af7e0f4bb663d69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2093607
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749097}