Clean up remaining pref_mapping entries in policy_test_cases.json
This converts the remaining "pref_mapping" entries in
policy_test_cases.json (which are currently just ignored) to
"policy_pref_mapping_tests".

As one of the pref mappings is only available if USE_CUPS is defined,
this also adds a way to specify that some tests should only run if a
preprocessor macro is defined.

Bug: 1017632
Test: browser_tests --gtest_filter=PolicyPrefsTest.PolicyToPrefsMapping
Change-Id: Ia8aac403eaf561af0d657e557b36b231bd1c2576
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2055104
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Yann Dago <ydago@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749117}