[COEP] Add CacheStorage > ServiceWorker support.
For every CacheStorage request, bind it with the correct COEP value.

A difficulty: the browser process do not always have the correct value
at startup for new ServiceWorker. It would requires PlzServiceWorker to
be implemented first. Instead, this keeps track of the CacheStorage
requests and bind them later when the COEP value will be known.

Bug: 1058787, 1031542
Change-Id: I7abcadf67d9627c82c295d5a05430ee9c6bca109
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095060
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749128}