[COEP] Move CacheStorage binders toward the actual class.
(pure refactor)

Before this patch, the BrowsingInterfaceBinder collected all the
attributes needed to bind the CacheStorage:
 - process,
 - origin
 - COEP
 - COEP_reporter (in the next patch)
and binded it.

After this patch, this work is done by the classes representing the
context on which the CacheStorage is requested:
 - RenderFrameHostImpl,
 - DedicatedWorker,
 - SharedWorker,
 - ServiceWorker

The code is slightly shorter and simpler. It will allow all theses
classes to provide their own COEP and COEP_reporter. This will be used
by the following patch in the branch.

Bug: 1059727, 1031542
Change-Id: Ic142da4f023b20c3ae25e98a179203c8d2bdfc6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2096453
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749160}