Remove fallback to FaviconCache from HistoryUiFaviconRequestHandler
Previous to this CL, if a) HistoryUiFaviconRequestHandler could not
find an icon in the local favicon database and b) Chrome was not allowed
to query the Google Favicon Server, then the layer would make a fallback
request to FaviconCache, the in-memory cache for synced icons.

Favicon sync has been stopped (https://crrev.com/c/1917173), which means
a) FaviconCache is no longer populated with remote data.
b) It is still populated with local data, however that data never gets
synced up.

This CL removes the fallback code to FaviconCache, under the assumption
that, except for rare cases, an icon would not be found in the in-memory
cache that was not found in the local database.

Bug: 978775
Change-Id: I9d4f92b19eb6a31ebd06b3579ffe1794b9737b00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2085375
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749170}