[Android] Abstract ExternalNavHandler setting of user gesture in intent
ExternalNavigationHandler currently directly calls into
IntentWithGesturesHandler when launching an external intent with a user
gesture to record that fact in case the intent ends up coming back into
Chrome. This is //chrome-level behavior, as it is relevant only in the
case where the app handles incoming intents. In addition,
IntentWithGesturesHandler would not be straightforword to componentize.

This CL abstracts this dependence through ExternalNavigationDelegate,
similar to how we have done for other such dependences.

Bug: 1031465
Change-Id: I2138400f7a62d1911073b1fd7b31ef484e7fbc17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095130
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749191}