android: Remove the linker tests.
The linker tests don't run on bots (and likely not locally either) and
do not support the latest library loading patterns. This is the first
step to remove them, as they add non-trivial complexity in the codebase.

More specifically, this removes the build rules and the test APK.

Bug: 1059707
Change-Id: I12102d59116ad1f0a152d3110cf3346471c8f390
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2093606
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Ross McIlroy <rmcilroy@chromium.org>
Reviewed-by: Egor Pasko <pasko@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749238}