Accelerate long press gesture with stylus button
ChromeOS devices currently do not have styluses with buttons,
and there is no support for handling stylus button events.

In the long term, we will want to make stylus button clicks act
as a right mouse click button (open context menu).  In the short
term, we want to use the stylus button to accelerate the long press
gesture.

If the user has the stylus button pressed when touching the screen,
or presses it before the long press timeout, generate a long press
gesture.

Bug: b:139781900
Change-Id: I3869989f3f731de18ecec480577eb209c08b9c6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2076663
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Sean O'Brien <seobrien@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749240}