Change Privacy UMA Logging Metrics to focus on Elements instead of pages
Given that individual navigations to various routes with settings are
already logged via UMA, the additional UMA logging provided by the
SettingsPageInteractions enum/histogram was of little value.

Additionally supporting this logging required restrictive code on the
site_settings_page.

This CL deprecates the existing enum/historgram and replaces it with
one focusing on interactions with elements that do not trigger a route
change.

Bug: 1059098
Change-Id: I9130836da821eae1aa0ac40ae0e96dcb4f398054
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090795
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749244}