viz: SkiaOutputSurface Checks OutputDevice for SchedulePrimaryPlane
SkiaOutputSUrface checks the OutputDevice implementation for whether it
supports scheduling the primary plane as overlay. This is to match
GLOutputSurface's behavior, where only the BufferQueue implementation
would return true for IsDisplayedAsOverlay(). This also helps avoid
the case where a Surfaceless GLSurface is created for
SkiaOutputDeviceGL.

R=kylechar

Change-Id: Iadfce3551f90e2049db73ed33dfff6ad042c4ccf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095142
Reviewed-by: kylechar <kylechar@chromium.org>
Commit-Queue: weiliangc <weiliangc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749269}