Move PermissionManager to //components/permissions
The PermissionManagerFactory is kept in //chrome, and an equivalent will
be added in //weblayer to initialize permission contexts. The map of
permission contexts are now injected into the constructor of
PermissionManager.

Some unit tests were kept in //chrome which relate to chrome-specific
logic in GetCanonicalOrigin().

Bug: 1025609
Change-Id: Id56c6233a65714cb5c68a13223bfc5020d9ac98a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095870
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749274}