Show appropriate warnings for AR permissions in site settings
https://crrev.com/c/2085847 began showing an appropriate string to users
when they denied camera access after granting AR permissions; but missed
a few other places to indicate to the user that additional permissions
are missing. This change ensures that these warnings appear, similar to
the behavior with the camera permission.

Note that crrev.com/c/2087704 handles ensuring that the user is prompted
for the OS permission when attempting to use the site permission.

Bug: 1058055
Change-Id: I26f427c874100b4bfbe52925c3bcded2729591bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090499
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Kamila Hasanbega <hkamila@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749275}