Ensure that AR prompts for system permissions
The underlying ARCore runtime requires camera access in order to
function. https://crrev.com/c/2072667 caused AR to prompt for OS/App
level camera access; however, this only happens if the user is shown the
permission prompt and grants the site access to the AR permission. If
the camera permission is denied either as a result of this prompt, or
revoked through the settings UI later, it is not checked again.

This adds a step to the site permissions request flow that ensures that
appropriate OS permissions are set. This mimics the flow that
MediaStreamDevicesController does, but puts the logic into the
permissions context class like GeolocationPermissionContextAndroid. If
indicated (via the persist bool), whatever permission is determined for
AR for the requesting origin permission will be saved, but the end
result bubbled back to whoever requested the permission will indicate
that the permission was granted iff both the site was granted permission
for AR and Chrome itself has permission to use the camera from the OS.

Bug: 1058058
Change-Id: I6f2d74d146de03998b59992a70c94d6c04dabbe6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2087704
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749317}