Fix to handle a SourceDevice switch during a gesture based scroll.
This CL fixes an issue where source device changes mid gesture scroll.
For example, when user initiates a mousewheel scroll, a GSB and a GSU
are dispatched. Before the GSE for mousewheel is dispatched, lets say
the user initiates a scrollbar scroll. At this point, a new set of
GSB and GSU are dispatched. While the scrollbar scroll is in progress,
the deferred GSE from the mousewheel will arrive and end up clearing
the scroll and current device (i.e scrollbar) would no longer be able
to scroll. To fix this, we maintain state in the InputHandlerProxy
that tells which source device is currently in charge of scrolling. If
a GSE for a device, other than the one in charge, is dispatched, it is
dropped. This state is maintained in currently_active_gesture_device_.
It is set on DID_HANDLE or DID_HANDLE_SHOULD_BUBBLE (for cases where a
scroll doesn't occur and the event needs to be bubbled) for GSB/GPB and
cleared on GPE/GSE.

Bug: 1055237, 1060157
Change-Id: Ib42637d33efbdc0cef5e957e095b3015eb5ec2c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068743
Commit-Queue: Rahul Arakeri <arakeri@microsoft.com>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749336}