vk: Add swipe up gesture to VK to access shelf gestures.
Right now, when the VK is up, the user cannot perform shelf gestures
to access the hotseat or overview mode. This patch adds the ability
for users to swipe up from the bottom of the VK to perform the normal
shelf gestures.

It does this by detecting SCROLL_BEGIN gesture events and forwarding
them directly to the shelf, so that the rest of the gestures act
directly on the shelf.

Change-Id: I8d409f8242b4b95f061afb86254c99103de38798
Bug: 1030814
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2071678
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749397}