Stop using HttpResponseInfo in ServiceWorkerVersion (1 of 2)
This CL replaces |main_script_http_info_| of which type is
HttpResponseInfo with |main_script_response_| of which type is
a new struct called ServiceWorkerVersion::MainScriptResponse.
MainScriptResponse is a subset of HttpResponseInfo and only
contains information needed by ServiceWorkerVersion.

The motivation of this change is to remove HttpResponseInfo dependency
from ServiceWorkerVersion so that we can eliminate URLResponseHead ->
HttpResponseInfo conversion in service worker code.

This CL also removes OnMainScriptHttpResponseInfoSet() from
ServiceWorkerVersion::Observer. ServiceWorkerContextCore is the only
class that implements the method and ServiceWorkerVersion has a
reference to the instance so this indirection isn't needed.

This CL doesn't modify SetMainScriptHttpResponseInfo(). A subsequent
CL will change it to remove HttpResponseInfo use in ServiceWorkerVersion.

Bug: 1060076
Change-Id: I5c632fb7e9b88ce826018c2175506a8602d3b9cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2093971
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749426}