cc: Clip render surface damage to its bounds
This patch clips the accumulated damage from a render surface to its
bounds. This is needed for scrollable layers within the render surface
that can have bounds much larger than their render target thus
indirectly contributing to the damage rect of their render target's
render target.

It also removes the expansion of the damage rect based on background
blur filter. This is okay because the background blur is computed within
the bounds of the render surface.

Test=Manually ran chrome os to check for ui glitches.

Bug: 1056014
Change-Id: I942aae3452250540fa283403894078f68630dff0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2075962
Reviewed-by: weiliangc <weiliangc@chromium.org>
Commit-Queue: Malay Keshav <malaykeshav@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749438}