WebUI Tab Strip: Update drag image scale and positioning
The drag image was previously misaligned because of added padding and
because ChromeOS automatically scales and offsets drag images. This CL
fixes the alignment issue and forces the drag image to scale to 110%
instead as per UX request.

The offsets are calculated by determining the pixels the drag started
from the center of the element, and then mapping them onto the bounds
of the drag image.

https://i.imgur.com/yLSJQ47.png

Change-Id: Ifa17ef20d9968cb7b9139bca7391c2c5971c0059
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2097187
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749477}