Disable HC controls painting when forced colors feature is turned off.
In order for high contrast to work correctly it needs the ForcedColors
feature to be enabled.

This CL disables the HC controls painting when the feature is turned off.

Also since content_features is not a separate component and adding
"//content/public/common:common" as a dependency for ui/native_theme
would create a circular dependency, the kForcedColors feature flag
is moved to ui_base_features.h

Bug: 1056398
Change-Id: I7f60c55a9c37b26ddedd7bcf2ec5fc92cf3909a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2097357
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749484}