Dynamically update cookies link row label based on cookie setting.
Previously link rows for content settings displayed one of two labels
describing their current setting. This CL extends the capability of
the cookies link rows specifically to express one of the four possible
effective cookie settings.
    -Allow
    -Block Third Party only in Incognito
    -Block Third Party
    -Block

To ensure that this label remains correct across both the preference
and content settings changes that affect the effective cookie setting,
an additional WebUI listener is created which pushes updates to JS
for immediate display.

This updated label is applied to both cookies link row locations
when the privacySettingsRedesign flag is enabled, and applied to
neither when it is disabled.

Bug: 1032584
Change-Id: I11a984bce60878b6224506bfc1cd1a5cf3a2e63e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095925
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: dpapad <dpapad@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749510}