Plumb COEP reporter to a service worker
CORP check happens before a service worker sends the response back to
the client. This CL plumbs a reporter to notify a server of failures of
the check. Note that the error is going to the COEP reporter for the
client, not for the service worker because the check in
FetchRespondWithObserver is considered for the network request is
initiated by the client.

Bug: 1056122
Change-Id: Id8b83f8c506a78e15653a993e9471d9eab0532b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2094073
Commit-Queue: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749513}