[directorytree] Avoid accessing SMB shares until the VolumeItem is activated
When an SMB share is mounted that requires credentials (eg. a
password-protected share added by enterprise policy), the user will be
prompted to enter those credentials when the share is accessed. Ideally,
this occurs when the user explicitly accesses the share, such as when
the user browses the share. However, the directory tree attempts to
access the share on startup in order to build the tree and render expand
icons. This results in the user being shown a credentials dialog for the
SMB share every time they open the Files App.

Preventing the directory tree from accessing the share before the user
has navigated to it, prevents the credentials dialog from showing until
the user actually tries to access the share.

Bug: 1060061

Change-Id: I5a0743cd2608b07455d145c2a1ec5a446737b40d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2094849
Reviewed-by: Noel Gordon <noel@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Anand Mistry <amistry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749551}