Worker: Create inside settings security origin from constructor origin
This CL changes how to create security origin in inside settings.
Previously, the security origin was created from request script url.
After this CL, it will be created from constructor origin if script url
is not data url, and set to an unique opaque origin if it is data url.

This CL focus on the behavior clearly defined in HTML spec. As for
chrome-extension who is not included in the spec, we still need further
discussion before deciding its behavior, so we keep the previous
behavior in this CL.

See the doc for more information:
https://docs.google.com/document/d/1Vfrvow1PO2OBr_3yI08aHYBl_YW1fQ87enqiHrdlqeA/edit?usp=sharing

Bug: 1053390
Change-Id: I64a2b63979e3ba2bb59673018b7cb89c0ef498f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2060224
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749704}