Android: close tab modal dialogs on navigate.
This doesn't affect most tab modals, such as permissions and js dialogs,
because those have associated C++ objects which react to navigations and
close the dialogs.

This affects/fixes the form re-post confirmation dialog, which until
now would not dismiss on a navigation. It also removes code to
explicitly close that dialog when the tab is destroyed, as it duplicated
functionality from TabModalLifetimeHandler.

Bug: 1059853
Change-Id: I9d6c3c1843c1b39faf5836adf535f108bfa00eaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095137
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749750}