Switch tests to non-declarative webRequest API.
Using declarativeWebRequest API fails on some bots, because this API is
not available in the Stable release channel.  The failure happens even
in presence of
extensions::ScopedCurrentChannel(version_info::Channel::BETA).  This
issue is tracked in https://crbug.com/1061019.  Because of the above,
this CL switches two tests to the webRequest API.

Additionally, https://crbug.com/1061019 shows that the test is timing
out in some cases.  The CL speculatively adds
ExtensionTestMessageListener to see if it helps.

Bug: 1057599, 1061019
Change-Id: If79ca92f22b7a87e8d1a315fbef685e72207d05a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090022
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Auto-Submit: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749802}