[DevTools] Throttle new page navigation before sending request
New page navigation (e.g. clicking on a link with target=_blank)
starts in a new tab and is initiated in the browser process similar
to the user initiated address bar navigations. In order to apply overrides
such as user agent that affect main resource request we throttle
the request before it's sent to let the client configure emulation
agent (as well as others) in the browser before it's called for the
request. In most of the other cases (OOPIFs, cross-process navigation)
the request is baked in the opener page so it was not necessary.

Design document: https://docs.google.com/document/d/1hbTowhyfrf6CaYuNEHjY_OVx2KgZmnbp8kg_wUdGFiA/edit#

Bug: 1051687
Change-Id: Iddcddf1d42878719cbc700f68eda5faffe13f18e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095373
Commit-Queue: Yury Semikhatsky <yurys@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749806}