Add IUpdater::UpdateAll function stub.
This is a mechanical change which is not doing much at all other than
landing the changes for generated IDL files under
win_build_output\midl\chrome\updater\ to get that out of the way.

The goal here is to connect --ua with --server and complete an RPC
out of process update all applications. This CL moves us a bit
in that direction.

Bug: 1053729
Change-Id: Idc3e66d7516143e5e748d110210f29a7d3107385
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101291
Reviewed-by: S. Ganesh <ganesh@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749809}