Connect new cookies controls to preferences and content settings
A previous CL introduced a new set of Cookie controls on the new
Cookies page. This CL introduces the logic to connect these controls
to the appropriate preferences and content settings.

As the effective cookie state is a combination of multiple preferences
and a tri state content setting, all preferences and settings are
manually configured depending on the combines state of the quad
state radio button and clear on exit toggle.

Bug: 1032584
Change-Id: I7ac499384a81e04d4e111e3114e59cc6979e43d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2097470
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: dpapad <dpapad@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749860}