[ToolbarMVC] Call refreshSelectedTab on TabModel switch
This updates the incognito state when switching tab models without a
corresponding tab selection, e.g. when toggling incognito from the start
surface.

The ChromeTabbedActivity/IncognitoUtils change is to fix a related bug
where we close the newly-created incognito TabModel immediately upon
launch. The closure of the TabModel was added to prevent a crash caused
by deleting the profile, but the deletion is the real problem; we should
avoid it when resuming with a selected, incognito TabModel.

Bug: 1059769
Change-Id: Ie4840b94ca3bc458940e707d162f3a978aef14fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2097829
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Patrick Noland <pnoland@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749876}