[discarding] Ensure that tab is reloaded when window is unminimized.
Currently, when a window is minimized, TabLifecycleUnitSource does
not update the "focused tab" unless another window gains focus,
because it considers that the "focused tab" is the active tab
of the *last* active window. When the window is unminimized, the
focused tab does not change, so no reload is triggered.

With this CL, the "focused tab" is the active tab in a window
that is *currently* active. That means that when a window that
has focus is minimized, the "focused tab" is updated. When the
window is unminimized, its active tab becomes the "focused tab"
and it is reloaded if discarded.

Bug: 871834
Change-Id: Icc657abcd70f0b899ded4302c95ed91e2616f167
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2008230
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749889}