DNR: Change the schema for the "rule_resources" manifest key.
Now instead of taking a list of strings, it takes a list of dictionary
of type Ruleset. This is in preparation of introducing support for
multiple static rulesets.

Note that this is a breaking change to the API surface, but it's ok
since the API is on beta.

BUG=754526
NOPRESUBMIT=true

Change-Id: I6a41f80dd675c9c614eca95cf1da2f3dce634e40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2092427
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#749999}