Ensure the gfx::BufferFormat of ScreenAsh be valid
When asked for the Display for a display that does not
exist, ensure that ScreenAsh and DisplayManager return
something with a usable gfx::BufferFormat.

To do this, update ScreenAsh::GetPrimaryDisplay to call
a new function, DisplayManager::GetFakePrimaryDisplay,
which will return a display that has a valid
gfx::BufferFormat. A similar workaround was previously
needed to ensure that a non-empty display was returned
from this function.

Bug: 1057501
Change-Id: I9e05c1f81da5057876b38c6ea65a06d48566b87a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090429
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750053}