[iOS] Stopped using unique names set by Chrome for calculating form signatures.
Currently a unique identifier is set on iOS for every form or form
field as a form name. Later this name is used for calculating the form
signature, which makes signatures of nameless forms on iOS differ
from the same forms' signatures on other platforms. As crowdsourcing
votes are not sent on iOS at the moment, no server predictions are
available for such forms.
This CL fixes it by ignoring names set by Chrome.

Bug: 986631
Change-Id: I020703e8b4e74e55cfdd62be0d681c4a4aef1c6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2100809
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Commit-Queue: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/master@{#750123}