Delayed AccessibilityEvent creation for throttled events on Android
This CL continues the Android performance improvements for users that
have accessibility features enabled.

In this CL we delay the creation of AccessibilityEvents until we are
ready/certain they will be sent. In previous work we throttled the
amount of events being sent to the system, but we did not modify the
creation of those events. This meant we were creating events that
were sometimes discarded, or we were creating events before they
needed to exist, and both of these negatively impacted performance.

We now wait to create an event until immediately before it will be
sent to the system.

Change-Id: I35bc7f5b21990caa23bc80677823d3a46f9736a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2100669
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750185}