[Code Health] Convert Closure to OnceClosure
Change GetClosure() in test_helpers.h to return a OnceClosure instead
of Closure.  This remaining changes are knock-on effects of the
GetClosure() change.  Basically need to call RunOnceClosure() instead
of RunClosure() and need to update some class members to be
OnceClosure instead of Closure.

Bug: 1007799
Change-Id: I53024573f6e3cca94787a8213b00c069cb7c3aa4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095858
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750199}