Prepare infra/config for Branch CI/CQ Phase D.
Phase D migrates all of the try builders that are not includable-only.

Migrate the try builders and the builders that they mirror to the
versioned subdirectory. bucketed_triggers is set to False for the ci
builders since there are not yet multiple builders with the same name to
result in a different scheduler job name. They will be removed on branch
day.

Bug: 1024644
Change-Id: I68f05c82b631b0ae56200f62599a440ecb02d70b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2098274
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750204}