[webauthn] Fallback to PIN for non UV authenticators
Some older authenticators support internal user verification but not UV
token or getting UV retries. Before this patch, if internal user
verification failed for one of these authenticators we would either drop
the request or show an error to the user. With this change, when an
authenticator supports falling back to PIN, chrome will show the PIN
dialog and alert the user.

Some authenticators will immediately fail when sent a uv = required
request if their internal uv is locked. To mitigate this behaviour, a
touch is requested from devices that return a response too quickly.

Bug: 999153
Change-Id: Icffccf208951cd33227fd532b6a569dcc8d93c99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2097121
Commit-Queue: Nina Satragno <nsatragno@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#750215}