Start blocking downloads when the file type is unsupported
Enterprise deep scanning for DLP violations has a specific list of
supported file types and the policy BlockUnsupportedFiletypes should
determine whether to blocked files not on that list. This CL introduces
a new download danger type for files blocked for that reason so that
we can, in future CLs, add UX specific to those files. It also generates
this danger type when appropriate, but uses the generic blocking UX.

Bug: 1060853
Change-Id: I6a670664d766bed860625cabcbd0d0e9a142e38e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101294
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Dominique Fauteux-Chapleau <domfc@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750276}