Allow non-elements to use layout tree builder traversal
Any non-elements can use the standard layout tree builder
traversal to find accessible children. Non-elements include
text nodes, and most importantly, the #document node, which
is the kRootWebArea.

This change only affects <dialog>, but it's an improvement.

Previously, a <dialog> would be added as a child in two places:
- As a child of the root node, because of layout traversal
- As a child within its actual DOM structure

Fortunately, it does not show up in the Blink tree twice, because
BlinkAXTreeSource has a !IsParentUnignoredOf() which ends up
not adding the DOM structure <dialog> child.

However, this double adding of children blocks landing of a chain
of other fixes necessary to avoid accidental resets of the
accessible tree during serialization.

NOTRY=true

Bug: None
Change-Id: I1ebb0ee4d7cdaea285773d391b3e73c0f1ecdb6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101325
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750289}