[Safety Tips] Auto-'ignore' safety tips when UI not shown.
Before this change, we could not disambiguate when users saw the UI (in
the treatment group) from the equivalent "first visit" in the control
group. This is because subsequent page visits in the control group were
still flagged with the original "bad reputation" or "lookalike" flag.

This CL makes it easier to compare between control and treatment groups
by automatically setting the 'ignored' flag for a given URL after the
first visit if the UI is not enabled. Previously, this state was never
used in the control group.

This CL is not user visible, since it only impacts the control group,
and the control group sees no UI.

Bug: 1061254
Change-Id: I2e4bd626d1b7b81d675259d7555f15604fda43e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101932
Commit-Queue: Joe DeBlasio <jdeblasio@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750308}