CrOS Shelf: Fix a bug where the status area wouldn't follow swipes
This bug happened when in tablet mode, and with an auto-hidden shelf.
When swiping the shelf up or down, the status area (which is visually
part of the shelf) would be expected to follow the shelf's movement,
but this was not the case. This change fixes that behaviour.

Add a test that would have caught the problem.

Bug: 1058997
Change-Id: I8c60d872875d47aba6b35a08b8e9ea4ff13a8e2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2102580
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Manu Cornet <manucornet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750313}