Desks: Implement multi-display drag and drop on desk in overview
This CL makes it possible to drag and drop a window on a desk
mini view on a different display than that of the window.
This CL is part of the work done for the feature
"MultiDisplayOverviewAndSplitView".

Demo: https://bugs.chromium.org/p/chromium/issues/detail?id=1059387#c2

BUG=1059387
TEST=Manually, added a new test.

Change-Id: I683a7d776981e5d71a996ed8e50381ac430aeac1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101726
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750322}