Added ACTION_IME_ENTER for Android on R SDK or higher
This CL continues our previous work on adding the new
accessibility action "ACTION_IME_ENTER" as an exposed
action for Android users on R SDK or higher.

We add the action during populate AccessibilityNodeInfo if
the given node is a descendant of a form (defined
as any kForm node).

The aciton leverages the existing ImeAdapterImpl to perform
an editor action based on the current TextInputAction of
the form field, allowing the ime button to perform next,
previous, done, etc based on the context of the click.

Bug: 1057254
Change-Id: I06171a2888d3bd1849f46edb1193ce7b79fe0365
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2096904
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/master@{#750326}