CrOS Shelf: Fix a bug where the status area wouldn't follow swipes
This bug happened when in tablet mode, and with an auto-hidden shelf.
When swiping the shelf up or down, the status area (which is visually
part of the shelf) would be expected to follow the shelf's movement,
but this was not the case. This change fixes that behaviour.

Add a test that would have caught the problem.

(cherry picked from commit e1e3056209ae6f87e2dd09e6ecb5ee6f4526a898)

Bug: 1058997
Change-Id: I8c60d872875d47aba6b35a08b8e9ea4ff13a8e2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2102580
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Manu Cornet <manucornet@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#750313}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2102537
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Cr-Commit-Position: refs/branch-heads/4085@{#8}
Cr-Branched-From: 938fda43077d4622c5b88984608d6becd5ebbb82-refs/heads/master@{#749737}