Enable web content accessibility when any UI Automation fragment root is requested
One of our detection points for enabling web content accessibility was
when a UI Automation element was requested from the legacy window. On
non-desktop Windows devices such as Hololens where HWND support is
limited, the OS may not talk to the legacy window, and it won't hit any
of the other existing detection points either, since they're tied to
MSAA/IA2 and UI Automation is the only supported API on these devices.

This CL moves the legacy window point of detection to where we obtain
a platform object for any fragment root, including either the top-level
window or the legacy window.

Bug: 928811
Change-Id: I8867258f4d1e2109c27066f978e9f5cf70766137
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101257
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Ian Prest <iapres@microsoft.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#750346}