Create extension blocked by policy dialog
Blocked by policy dialog is shown when an extension can be installed due
to policy. This also includes extension request rejection. But it will not
be used if the extension is blocked in child account.

The dialog is for information only and user can only close the dialog. When
the dialog is closed, the webstorePrivate.beginInstallWithManifest3
function will return user cancel to stop extension instllation.

This dialog replaces the old "Oops" dialog in CWS. The dialog UI has been
approved by UX team.

The dialog is by design very similar to ExtensionInstallDialogView
including Layout, title, icon and a ScrollbarView for additional message.

Screenshots:
https://drive.google.com/file/d/1qj3wx8wMv3F5B6PZAAGcShViHZO-VYF0/view?usp=sharing
https://drive.google.com/file/d/1KgQPU-yTgiiYPJvOxsZNpT-w7WJnY1VL/view?usp=sharing
https://drive.google.com/file/d/19W6W4K3bH4OGXByi-mM1o87mwU7BQQ1b/view?usp=sharing

Bug: 1006899, 1061559
Change-Id: I3887c2ff3247d4df6401ec9829dce046461aa075
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068680
Commit-Queue: Owen Min <zmin@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750355}