[iOS] Implement custom infobar banner presentation via OverlayPresenter.
This CL updates InfobarBannerPresentationController to subclass
OverlayPresentationController so that it can specify that it resizes
its presentation container view.  This allows the overlay presentation
context view to be resized to match the container view so that touches
that fall outside the banner can pass through to the underlying browser
UI.  When the infobar overlay flag is disabled, subclassing this
presentation controller does not change any behavior.

Additionally, this CL updates InfobarBannerOverlayCoordinator to use
the InfobarBannerTransitionDriver.  In order to improve upon the
InfobarBannerPositioner interface, this CL also updates
|-bannerYPosition| to specify that the returned position is in window
coordinates.  This disambiguates the coordinate system in which the
returned position is calculated, allowing a shared point of reference
for both the overlay and legacy infobar implementations.

Bug: 1030357
Change-Id: I6425a81dbed40715c2cebd53f1459d3f924b44b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2040748
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750368}