[Paint Preview] Add Paint Preview demo mode
This CL implements Paint Preview demo mode. When enabled, the current
tab will be captured and displayed as a paint preview. Note that this is
not hooked up in this CL. A follow-up CL will add a menu item for
activating the demo mode.

Binary-Size: Size increase is due to adding the paint preview player.
Bug: 1058405
Change-Id: I07b5afe418e1b957f830e01a63d009d30e54c7c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2087993
Commit-Queue: Mehran Mahmoudi <mahmoudi@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750380}