Enable setting focus on anchored bubbles via hotkeys.
The F6 (rotate pane focus) and Alt+Shift+A (focus dialog) hotkeys
ignored the existence of the "restore pages" dialog and probably
also similar ones.

This dialog is outside the views hierarchy but it's anchored to the
AppMenuButton, so we can get it through the kAnchoredDialogKey
property. This will also work with other kinds of dialogs that may be
anchored there.

BrowserView::ActivateFirstInactiveBubbleForAccessibility() has been
created so the callbacks for both hotkeys can use it.

Bug: 1042010
Change-Id: I9befe07d92847279726d8f51d5f5fa9e0a6e5cf6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2055624
Commit-Queue: Jacobo Aragunde Pérez <jaragunde@igalia.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750402}