[Benchmarking] Reshard Nexus 5 Perf to use half the devices.
Some of the reimaging to android M didn't work.

Do not revert this change if it appears to
cause a perf regression. That regression
was likely caused by upgrading the devices from
Android version K to Android version M. DO
NOT REVERT this change.

TBR=johnchen@chromium.org

Bug: 1041944
Change-Id: Idbcc6de3bd69dfe2724e146dfee1b25e32518f0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2104453
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750446}