Optimize the buffer usage for 180 degrees rotation
When the system is extremely slow, the buffer allocation of the output
buffer can fail and upper layer would get the raw un-rotated buffer.

This CL mitigates the issue by reusing the original buffer for 180
degrees rotate, which is the default orientation of the user facing
camera on Nocturne.

Bug: 1061346
Test: Run Hangouts Meet with screen rotated with 0, 90, 180, 270
degrees.

Change-Id: I7051fa9e0c3857345ff3ad1916fd3d91d6a59f71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101324
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Shik Chen <shik@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750487}