heap: Make visitation const
Tracing on-heap objects should treat everything as const and not change
anything. This CL makes all visitor tracing/visitation methods and
TraceTraits expect const objects.

The guideline this CL tries to establish is that from the perspective
of tracing everything is const, including data that will later be used
for compaction. Everything the visitors records is const.
If and when needed, const_cast is used to  get mutable versions (e.g.
in weak callbacks and compaction).

This CL also updates the collection Trace methods to const as these
are highly intertwined with the TraceTraits.

Bug: 1058387
Change-Id: Icd26693e029d5c1d03ba50e54390b1d9aa350620
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095720
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750530}