ash: fix feature disable dialog close behavior
As part of the DialogDelegate close refactor I changed the default
behavior of dialogs that have both cancel and accept callbacks to
require opting in to receiving a callback on close as well. This dialog
is one of the few that were both depending on this behavior and had no
test coverage for that dependency.

This change:
1) Opts that dialog into receiving a callback on close, using the
   pattern called "pattern 2" from r743269
2) Adds test coverage for the behavior here

Bug: 1056116
Change-Id: I3b4a30c6015bf76534b6b083586a13fdc55bac2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2100727
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750563}