[iOS] Add workaround for Apple UIPasteboard.changeCount bug
For unknown reasons, on iOS 13, whenever the user focuses a UITextField,
the pasteboard's changeCount is incremented by 2, although the contents
don't change. To fix this, if the changeCount is incremented by 2, we
fallback and check MD5, because there's no way to know if the difference
is because of this bug or because two actual changes occurred.

It is possible to focus multiple text fields in between calls to this
method (so properly, we should fallback to MD5 any time the change count
difference is even), but that occurs much less frequently. The most
common time this method is called is on tapping the omnibox, which will
trigger the one-focus case.

Bug: 1058487
Change-Id: I4d2b831840f98c071cc2dd76af10bd5d4731b0bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2096562
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#750583}