Disables SaveCard expiration date check on iOS
When attempting to upload saved a card, there's a check where it
assumes that the user providing a date expiration value is because
the date was expired. This ends up calling
LogSaveCardRequestExpirationDateReasonMetric which DCHECKS if the
provided date is not expired.

After enabling the editing support on the Save Card modal on iOS, users
should not be able to send an expired date, but they will provide a date
when attempting the upload (even if it is the original the date).


Bug: 1029067
Change-Id: I874c8af5f24159e00a6cb3c50d46d988d3f391cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2099481
Auto-Submit: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Parastoo Geranmayeh <parastoog@google.com>
Reviewed-by: Jared Saul <jsaul@google.com>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750590}