[iOS] Disable Text Zoom option when it is already showing
Now, the tab helper holds whether the ui is active or not. Besides
disabling the Text Zoom menu option when the UI is active, this also
allows for better state logic to re-present the bar after a temporary
hide.

For instance, it is hidden when you begin a side swipe-to-navigate.
Before, it wouldn't come back if you interrupted the swipe, but now
it does.

Fixed: 1056814, 1056821
Change-Id: I59d0b885bb4143c63563569837d1558af0b58024
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080413
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#750594}