Adjust battery level threshold for animation disabling experiment
I looked at existing Study and the number of entries is too low for
statistically significant experiment. We should get more events by
disabling the animations even if battery life is higher than 20%,
so we can determine if animations affect PLT.

This CL raises threshold to 50%, which is arbitrary chosen number.
The experiment is only running on dev and canary, and not intended
for beta or stable.

Bug: 1006431
Change-Id: I8b6f0cb1ba876c7db11ec23169136e41c21b058e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2096170
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750595}