cros: Fix components_unittests when SplitSettingsSync is enabled
Fix the tests in anticipation of enabling the feature by default.
The tricky bit is that UserSelectableType::kApps isn't used on Chrome OS
when the feature is enabled, but it is still part of
UserSelectableTypeSet::All(). Update the tests to use the registered
selectable types, which is closer to what the production code does.

Test: components_unittests --enable-features=SplitSettingsSync
Bug: 1060289
Change-Id: I16d4f119c1daa1cb51f2b9a35ea247beec75ece9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101722
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Auto-Submit: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750599}