Reduce temporary Font creation in canvas font resolution
Font = FontDescription + FontFallbackList

In canvas font resolution, the result is just a FontDescription.
However, we currently cache the resolved fonts as Font objects with a
null FontFallbackList, and pass Font objects to
CanvasRenderingContext2DState::SetFont(), which is unnecessary.

Hence, this patch changes the code to use FontDescription instead.

This patch also makes the code easier to analyze, so that we can
eventually eliminate Font::Update(font_selector).

Bug: 1049295
Change-Id: I1e3c5a34347355aad2a24f26d35685e9125f7c5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103136
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750628}