Erase ChromePreferenceKeys.START_SURFACE_SINGLE_PANE_ENABLED_KEY
This pref is only needed while transitioning clients to use the dynamic
pref. It is there to honor a previously cached value, but after the
dynamic pref is written, keeping this old pref could wrongly override
the desired value in dynamic pref. For example, if this old pref is
true, isStartSurfaceSinglePaneEnabled() would be stuck to true when
Start surface is enabled. This makes it impossible to switch from
single-pane to other Start surface variations. This CL fixes this bug,
which was introduced in
https://chromium-review.googlesource.com/c/chromium/src/+/2070967

Bug: 1062034, 1062013
Change-Id: If836d22e97ad3b08e3ce6974ad752d530693e925
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2098268
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Ganggui Tang <gogerald@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750673}