Add timeouts for EventWatchers on CSS animation tests
Previously the test harness could timeout while waiting on the event
watcher. In this patch, a timeout was added to the watchers in cases
where the events should fire on the next animation frame. With the
change, we still see timeouts on individual tests, but no longer see
the test harness timing out.

Next steps:

1. Add animationcancel. For some reason, this event got overlooked
   previously and was never implemented. Adding this event resolves
   the majority of timeouts (confirmed in an experimental patch).

2. Reconnect and event delegate after calls to animation.effect.
   Currently, the lack of a delegate on the new effect halts the
   dispatching of subsequent animation events.

Bug: 1059968
Change-Id: Ib5bf46762e316a5f3d5f33aa95bf74edaf42e773
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2105424
Reviewed-by: Yi Gu <yigu@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750745}