Remove workaround for absent SupervisedUserNavigationObserver.
This cl removes the workaround for the absence of
SupervisedUserNavigationObserver even when the profile
of the WebContents is supervised. The workaround is removed
so that we can test the fix that landed in crrev/c/2102764
actually fixed the underlying issue.

Bug: 1061427
Change-Id: I3801c87779563fa7f76a38bb9e4403fa4c639ca8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2105623
Commit-Queue: Yilkal Abe <yilkal@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750769}