Implement time picker infinite scroll without duplicating or reordering cells
The purpose of this change is to work around an issue encountered when
adding AX roles to the time picker popup to implement proper screen
reader accessibility.  The trick used to create hour/day/minute columns
that loop infinitely when navigated by arrow key involves duplicating
all the cells and placing them out-of-order in the DOM.  This confuses
screen readers so that they read out confusing things like "minutes
column, 25, item 70 out of 120 selected", even with the right ARIA roles.
I looked at two alternative fixes for this issue:
1) Applying aria-posinset and aria-setsize to the TimeCells.  This seems
like it should work, but these stop functioning when the number of items
in the DOM is greater than aria-setsize or when the aria-posinset values
are not increasing with respect to DOM order of the cells.
2) Marking cells aria-hidden and controlling screen reader announcements
with an aria-live region.  This could mostly work, but Windows Narrator
doesn't draw proper focus rects for aria-hidden items, and this behavior
is explicitly warned against in the spec per
https://www.w3.org/TR/using-aria/#fourth so there could be
interoperability issues.

So as a last resort, this change re-implements the infinite scrolling
behavior by dynamically manipulating the layout of the TimeCells so
that they can be present in the DOM in an order and number that makes
sense for a11y.

This is achieved by a scroll event handler and ArrowUp/Down keydown
event handlers that rotate the cells' layouts to ensure that more cells
are always visible in the direction that the user is navigating.  These
rotations handle outside of the visible portion of the scroller so they
are not visible to the user, who just sees an infinitely looping column
of cells.

One advantage to this change aside from the a11y issue is that mouse
scrolling no longer bottoms out after 1-2 trips through the values; it
now appears really 'infinite'.  The obvious disadvantage is that this
approach is more complicated and likely has more performance overhead.

A subsequent (much simpler!) change will add the ARIA roles to finally
make the control accessible.

Bug: 1026544
Change-Id: If4eeb2e84731d3107b1f516544761b908cc62e4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2097250
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Ionel Popescu <iopopesc@microsoft.com>
Commit-Queue: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#750784}