Windows: Fix focus on restored window minimized w/o focus
If a browser window is minimimized w/o focus in the location bar,
and Chrome is launched from a shortcut or the command line, the window
is restored, but doesn't get focus, which breaks keyboard accessibility.

This CL adds a HandleWindowMinimizedOrRestored notification to
HwndMessageHandler::PostProcessActivateMessage, when restoring a
minimized top level window from ::Activate(). This causes the root
window to get shown, which allows the view to subsequently get focus.

Bug: 1048195
Change-Id: I147ffccc2a6d9b20623f1989a8ea4395ad56334d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2088293
Commit-Queue: David Bienvenu <davidbienvenu@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750786}