[ntp][omnibox] Add OmniboxTabHelper to mediate omnibox -> NTP calls
This CL is preparing to add the fakebox to the WebUI NTP. The new
OmniboxTabHelper lets the WebUI NTP register an observer for omnibox
events such as focus changes. At the same time, it also lets the local
NTP register observers to support existing behavior.

Calling the WebUI NTP directly from the omnibox, or letting the WebUI
NTP register an observer at the omnibox was considered. However, there
is no convenient handle pointing to the WebUI NTP the omnibox has
access to. Furthermore, the omnibox responsible for the WebUI NTP may
change when moving tabs to other windows. The OmniboxTabHelper makes
this loose connection more mangeable.

Bug: 1061073
Change-Id: I7d0edee57176cb051ee9d77a2bb9300c24f54b06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101700
Commit-Queue: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750798}