Change Crostini's "unclean startup" check.
Successful calls to ListVmDisks just show that vm_concierge is still
running after a Chrome crash. As long as no vm is running, this can be
fine because users may have shut down their vm manually before the
crash, in which case the startup is "clean".
In this CL we changed to GetVmInfo which shows that vm_concierge is
running, and a vm is running too.

Bug: 1060813
Change-Id: Id6d6644c8a9bb88c4e2fec3803f6290f39e4e9de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2102030
Commit-Queue: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Cr-Commit-Position: refs/heads/master@{#750814}