Update user-consent logic for Quick Answers
This CL updates the logic for showing Quick-Answers to the users when
they have explicitly granted the consent by acknowledging the
User-Consent view or implicitly done so when the user-consent view has
been sufficiently impressed.

The major change will be that once impression caps have been reached,
the Quick Answers will begin to be shown in addition to the consent-view
not being shown anymore. This is because the requirements updated to
treat the consent-view as more of a notice.

Bug: b:151363933
Test: Tested on Chrome OS VM.
Change-Id: Ic7868ed208e0fe5c05da57509459d7888665d3e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101725
Commit-Queue: Abhijeet Singh <siabhijeet@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750822}