WebApps: Change parameters to RegisterFileHandlersWithOs.
The parameters to RegisterFileHandlersWithOs currently include separate
sets of MIME types and file extensions corresponding to the file
handlers for the web app being installed. This CL replaces those with a
single vector of apps::FileHandlers. There are a couple of reasons for
doing this. First of all, the Linux implementation of this method will
require an explicit mapping of MIME types to file extensions, which is
needed to register new MIME types with the OS, and which is lost when
extracting the separate sets. Second, the work of extracting those sets
is better done by the callers to this function anyway; for example, in
the Windows implementation, the MIME types set isn't used at all,
meaning that unnecessary work is being done to produce it in the first
place.

Bug: 938103
Change-Id: Ie0a6e7196e7f71338602d9c3e2d0d593283be8e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101393
Commit-Queue: Robert Woods <robertwoods@google.com>
Reviewed-by: David Bienvenu <davidbienvenu@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750828}