Create public interfaces for BrowserXrRuntime
Creates interfaces for BrowserXrRuntime to be consumed by classes that
will remain in chrome/. This updates existing usages that will no longer
be able to access the BrowserXRRuntimeImpl class to use the public
interface (primarily this is everything outside of vr/service).

Note that although right now BrowserXRRuntimeImpl lives in chrome/, it
will soon be moving to content/, once the remaining dependencies for the
rest of vr/service have been broken (so as to avoid over-exposing
methods needed by these classes to talk to each other). Currently the
only way to get a BrowserXRRuntime is via the XRRuntimeManager. While
some of the XRRuntimeManager methods will be moved to a public interface
in future changes, it will still only be possible to get an
XRRuntimeManager (and therefore a BrowserXRRuntimeImpl) in chrome/ until
the move actually happens.

Bug: 1031622
Change-Id: I42e1cfd18842166ce1f8eeaa6003d7f84e52aef6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103305
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750845}