Add PrintBackend::CreateInstanceForCloudPrint() to use CUPS PPD backend
Change the interface of PrintBackend::CreateInstanceImpl() to include a
flag indicating that the backend is to be used for Cloud Print. The new
interface PrintBackend::CreateInstanceForCloudPrint() passes |true| for
the flag, but usage of the flag will follow in a future CL.

The flag is necessary for runtime switching of the CUPS backend on
macOS. Because cloud print is still configured to rely on PPD
attributes, we want to still use the CUPS PPD backend over the CUPS IPP
one. CreateInstanceForCloudPrint(), which will only be built and called
on CUPS platforms, would ensure that the CUPS PPD backend is used.

Bug: 226176
Change-Id: Ic9c376af9992bb70516d1de39da3c9666cff145e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089987
Commit-Queue: Daniel Hosseinian <dhoss@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750854}