ReferrerPolicy: Fix http test with shared worker
This CL fixes the test regarding referrer policy with shared workers.
worker.php contains evaluation error because postMessage() is not
defined in SharedWorkerGlobalScope, but as postMessage runs after
onconnect handler is registered, the evaluation error is ignored and
had seemed to be running correctly.

Bug: n/a
Change-Id: I895bdb54b97122dd6d9828096fc2154fbb6fb2d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106043
Commit-Queue: Eriko Kurimoto <elkurin@google.com>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750888}