Make SVGFilterPainter::PrepareEffect return FilterData
This makes it possible to discern "no valid filter" from "contents does
not need recording". In turn this makes it possible to skip the
corresponding FinishEffect() call when there's no valid filter.
This allows simplifying FinishEffect() and drop the Abort() method on
SVGFilterRecordingContext.

The call to SVGFilterRecordingContext::BeginContent() is sunk out of
PrepareEffect() and likewise the clearing of the invalidation mask is
hoisted.

Bug: 109224
Change-Id: I440a0d85f93ec0bd0d043a4f3545d793364b9f1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2105374
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#750892}