[ios] Add lock preventing nest batch operation on WebStateList
WebStateList is not re-entrant (already asserted by locked_)
so there is no point nesting batched operations. Add a bool
to ensure this is not the case.

Add a method to check whether a batched operation is in
progress (could be used internally to also omit sending
notification during a batched operation).

Bug: 1014526
Change-Id: I6b6dc3278312895a32898d898245f538f9a5264d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089675
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750916}