Update login shelf button on locale change
The LoginShelfButtons (Shut down, Browse as Guest and Add Person)
are not updated on a locale change issued through a policy change.
Both the text language and the alignment of the buttons (in case of
RTL) are not updated.
LoginShelfView does have an OnLocaleChanged function and adds an
observer that are meant to update the buttons, but the function is
not triggered when the locale policy changes.

This CL adds a function, NotifyLocaleChange, which is called after
the language switch (OnLanguageSwitchedCallback) handled in
login_display_host_webui.cc. NotifyLocaleChange uses
LocaleUpdateController to notify all OnLocaleChanged observers.
Both the text and the alignment issues are resolved after adding the
proposed changes.
Browser tests are also added for checking the Shut down button text,
alignment, and the text of an UnifiedSystemTray element.

Bug: 1051945, 1054417
Change-Id: Ida6ea782622dacdf118666c64bb908929b4cdb89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2081820
Commit-Queue: Maria Petrisor <mpetrisor@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Reviewed-by: Jit Yao Yap <jityao@google.com>
Cr-Commit-Position: refs/heads/master@{#750917}