[webtransport] Avoid using deprecated ArrayBuffer
This CL removed a use of the deprecated ArrayBuffer class. In the
|OutgoingStream| class, ArrayBuffer was used to own a byte array and
its length. This Cl replaces ArrayBuffer with a small inner class that
does exactly that, it stores a byte array and its length. The
constructor uses the wtf::Vector allocator to allocate the byte array.

R=ricea@chromium.org
CC=​haraken@chromium.org

Bug: chromium:1008840
Change-Id: I21286265eeaabe359ea7709b8d05042c854b7c25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2102585
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750924}