Move Vulkan EndSwap semaphore to EndPaint semaphores
This CL is follow-up to refactoring that allowed SkiaOutputDevice to
provide additional semaphores to be submitted at flush time.

Before this CL SkiaOutputDeviceVulkan had special pass to submit
additional semaphore that can be used for presentation. This CL moves
it to GetEndPaintSemaphores to unify code paths.

Bug: 1043114
Change-Id: I88092bb0d61ceec9108ab4e9275569af44fa80d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101305
Reviewed-by: Michael Spang <spang@chromium.org>
Reviewed-by: Peng Huang <penghuang@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750954}