Ad tagging: Add functionality to tag elements as ads.
When an HTMLImageElement is created by an ad script, if
async ad tagging is enabled, the HTMLImageElement will be
tagged as ad-related.  When the image finishes loading,
the HTMLImageLoader will check the ResourceRequest via
the ImageResourceContent and ImageResourceInfo to see
whether the resource was tagged as an ad resource.

Eventually we would like to expand the ad-tagging
functionality to some other types of elements, notably
HTMLDivElements.

Tests: blink_unittests --gtest_filter=*ImageLoaded*Async*

Change-Id: I5c6115f0b0be180a4647389096983378191cd0be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2031086
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750956}