Regenerate test certs with shorter validities
macOS 10.15 limits the maximum validity of locally trusted certificates
to 825 days, documented at https://support.apple.com/en-us/HT210176

This updates the test certs that interact with the macOS APIs to meet
that requirement.

In doing so, the CRLSet logic is also updated to be able to create
fake CRLSets during testing for serials with the high bit (0x80) set
in the first byte of the serial. These have a leading 0x00 in the
actual encoded form, to ensure they're positive INTEGERs, while the
CRLSet logic expects that leading byte to be stripped for the internal
representation.

Bug: 1061024
Change-Id: Ie225545b493550cac7e007c0978e843af3b52a37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103370
Reviewed-by: Matt Mueller <mattm@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750972}