Delete XR.RuntimeAvailable metric
This metric is only used to record the availability of the OpenVr
runtime. Rather than expand it to cover the other supported runtimes,
it is being removed since this information would not be used to drive
any decisions.

As a bonus, removing it simplifies refactors to XRRuntimeManager for
it to move to content.

Bug: 1031622
Change-Id: Id0fbaa57abd4c2115921f3f3c23781aa099989ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106456
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750978}