heap: Use assignment in ConstructTrait for member
When calling ConstructAndNotifyElement, we assume the element already
exists and can be concurrently traced. In such cases updating the
member should use atomics. Using operator= in this case takes care
of both atomics and the write barrier.

Construct still calls the member ctor since this is assumed to be a
new buffer that cannot be traced yet.

Bug: 986235
Change-Id: I6c7053e8d084e9c2852bee1e3192196483bfb727
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2105373
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751023}