[iOS] Add workaround for Apple UIPasteboard.changeCount bug
For unknown reasons, on iOS 13, whenever the user focuses a UITextField,
the pasteboard's changeCount is incremented by 2, although the contents
don't change. To fix this, if the changeCount is incremented by 2, we
fallback and check MD5, because there's no way to know if the difference
is because of this bug or because two actual changes occurred.

It is possible to focus multiple text fields in between calls to this
method (so properly, we should fallback to MD5 any time the change count
difference is even), but that occurs much less frequently. The most
common time this method is called is on tapping the omnibox, which will
trigger the one-focus case.

(cherry picked from commit 0189c6d16be199350fcce30c54ced424f5d0442b)

Bug: 1058487
Change-Id: I4d2b831840f98c071cc2dd76af10bd5d4731b0bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2096562
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#750583}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107542
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/branch-heads/4085@{#73}
Cr-Branched-From: 938fda43077d4622c5b88984608d6becd5ebbb82-refs/heads/master@{#749737}