Merge the getters/setters for force dark mode.
The hand-coded methods to set and get the force dark
mode setting are being used alongside the auto-generated
methods based on the settings in settings.json5.

The hand-coded methods are merged with the auto-generated
methods as they are meant to do the same thing.

The hand-coded methods invalidate the color scheme while
the auto-generated methods invalidate paint. A change in
the force dark settings should ideally invalidate both.

So a new change type is introduced which invalidates both
color scheme and paint.

Bug: None
Change-Id: Ice0c4b97a30d5d7f7d91b7ea5a30a67b5b19d7e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2081727
Reviewed-by: Aran Gilman <gilmanmh@google.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Prashant Nevase <prashant.n@samsung.com>
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Varun Chowdhary Paturi <v.paturi@samsung.com>
Cr-Commit-Position: refs/heads/master@{#751030}