[webui][ntp] Add fakebox
The fakebox is an NTP search input field that forwards focus and pastes
to the omnibox. It's merely an affordance on the NTP to redirect the
user to the omnibox, which implements the actual search and suggestions
UX.

This CL adds a fakebox to the WebUI NTP similar to the existing
fakebox on the local NTP. See crbug/1061073 for how the WebUI NTP
fakebox looks like.

Bug: 1061073
Change-Id: Ic84bbe45867baeb2040b4981ca2d5be4ee1709d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101299
Commit-Queue: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751050}