Add toggle to the set of elements that can be displayed in an  accessory sheet
This change adds a custom toggle to the accessory sheet to be displayed by the
passwords accessory sheet when password saving is disabled for the current website.

A custom toggle is needed because the view needs to display a title as well as
a subtitle that indicates whether the toggle is on or off and the toggle needs to
be vertically centered in the view.

The SwitchView is set to be unclickable so that the parent view handles the click
and the tap animates properly. The background is set to null so that there is no
ripple effect on the thumb of toggle. This is consistent to the appearance of
ChromeSwitchPreference.

Bug: 1044930
Change-Id: I6ba5fdf8b70c117729fc0163f0ca62408a6e59d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095119
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751057}