[Fuchsia] Register fuchsia-dir URL scheme in runners
Previously fuchsia-dir was registered as a standard URL scheme only in
WebEngine, but not in runners. As result fuchsia-dir:// URLs were not
always handled correctly, particularly because GURL::GetOrigin()
returns an empty string when called for an URL with a non-standard
scheme. When trying to start an app from a custom content directory
cast_runner was getting empty string from GetOrigin() and passing that
empty string to Frame::SetPermissionState(). As result WebEngine would
terminate the Frame.

Added RegisterFuchsiaDirScheme() helper that's now used in runners to
register fuchsia-dir scheme. This ensures that GURL::GetOrigin() works
properly for fuchsia-dir URLs.

Bug: 1062120
Change-Id: Ibe28f07e4a6b08cf62526c5f7ee81e2783c6d667
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106748
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: David Dorwin <ddorwin@chromium.org>
Reviewed-by: Kevin Marshall <kmarshall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751096}