[Android] Use a new flag to control real time URL check on all devices.
In crrev.com/c/2092697, a memory upper threshold was introduced
to control real time URL check. However, it turns out the change
won't work because high-end and low-end devices are still being
controlled by the same flag "kRealTimeUrlLookupEnabled". Once
high-end devices are launched, finch will warn us with "both
high-end and low-end studies apply to the same feature flag and
affects the same population." See the warning in cl/300834155.

To avoid the conflict, it might be easier if we revert the change
in crrev.com/c/2092697, and add a new flag named
"kRealTimeUrlLookupEnabledForAllAndroidDevices".
The new flag will be independent of the old flag, so it won't be
affected by the high-end devices launch.

Bug: 1014202
Change-Id: Id7c0ef84a0ba8f45ea5628fcec472a678450445b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2104869
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751168}