[throughput] Turn a if into a DCHECK
Right now in the ReportMainFrameCausedNoDamage, we set the
begin_main_frame_data_.previous_sequence to 0 only if the frame
args is the same as the current main-frame args.

This CL turns the if statement into a DCHECK. It basically means:
if the no-damage report comes, it must come before the next
ReportBeginMainFrame.

Bug: None
Change-Id: I6c02855c5fac4721a430cd03610c8ead17d8a411
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2105400
Auto-Submit: Xida Chen <xidachen@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751176}