Add targeting embedding heuristic for lookalike warnings.
Checks to see if a URL contains a target embedded inside it in a way
that needs to trigger a safety tip. This function will return a "safe
url" which is the embedded target. Currently we consider the following
as Target Embedding: sth-google.com-sth.com, sth.google.com-sth.com,
sth-google-com-sth.com, sth.google.com.sth.com, sth-googlé.com-sth.com
where the embedded target is google.com. In addition to these if a TLD
is used that is considered to be an important TLD but that domain does
not actually exist we consider that a target embedding as well. (e.g.
google.edu instead of google.com in the above examples.)

A more detailed explanation could be found in the project document:
https://docs.google.com/document/d/18YQjj17a820SLVWuIF7cOay_Y8Cq7nt721m67WqCtW0/

Change-Id: Id8ced3db6dc74326dddbe45c0a6c6e8cf53f1c96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2096026
Commit-Queue: Behnood Momenzadeh <behnoodm@google.com>
Auto-Submit: Behnood Momenzadeh <behnoodm@google.com>
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751193}