[SuperSize] Ensure BulkForkAndCall() cleans up |pool| on exception.
For BulkForkAndCall(), the first exception encountered can get
re-raised. However, this bypasses code to clean |pool|. For the
upcoming work to convert SuperSize to Python 3, this causes

  parallel_test.py: testBulkForkAndCall_exception()

to hang, if another test is also run, and vpython3 (instead of python3)
is used.

This CL prevents the problem by ensuring that BulkForkAndCall() cleans
up |pool| when an exception is raised.

Also fix missing parameter to _ForkTestHelper() in
testBulkForkAndCall_exception().

Bug: 1054018
Change-Id: I2942438130700fa7ccf804064ae317159c2aa895
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107121
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Samuel Huang <huangs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751228}