Revert static member LayoutThemeDefault cleanup.
Making the selection colors and caret interval members non-static
members of LayoutThemeDefault in [1] caused regressions for the devtools
inspector because GetTheme() returns different objects depending on
whether we emulate a mobile device or not, and these values are only
updated on the current object returned by LayoutTheme::GetTheme().

Also added a test to avoid this regression in the future.

[1] https://crrev.com/67575409937d6cc6d2226f7f5bcea667075e5d7d

Bug: 1062214
Change-Id: I06144fb4ff74e204432591498e7cd2e934018819
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106070
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751273}