[cascade] Respect filter when analyzing interpolations
The filter was accidentally unused when analyzing interpolations.
Everything still worked, because Apply-time filtering was honored. It
might be that we actually don't need Analyze-time filtering at all,
but we shouldn't remove it until the performance implication of such a
change has been investigated.

In the meantime, the filter should be respected.

Change-Id: Ib04e044ab4725940c17312e80206a75630bb8d80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106590
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751277}