Remove more base::Bind() calls in chrome/renderer
In this CL, we replace base::Bind() with BindOnce/BindRepeating.
BindOnce is preferred, so here is the rationale for the BindRepeating
cases:
* In SearchBouncer, the base::Bind is used as a parameter of
AddInterface(), which requires a repeating callback. It also makes sense
to allow for multiple receivers.
* In TranslateAgentBrowsertest, it is used as a parameter to
SetBinderForTesting, which also requires a repeating callback.

Both Repeating cases are associated to methods in Blink which are
already marked as receiving repeating callbacks.

Bug: 1007641
Change-Id: I03c1b25669a63e58ce5b384f301d2c5b1399e0bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107716
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751333}