[aw-dark] Update prefered-color-scheme when page is updated.
We rely on preferred-color-scheme is set to no-preferences
if UA darkening is turned on, so we do not accidentally darken twice.
In prefer-web-theme-over-UA-darkening mode (apply web theme if
available, otherwise apply UA darkening) we need to ensure that
preferred-color-scheme is correctly set for each loaded webpage.

Change-Id: Id3246269f5fb9336ff5a924a95e4768420aefe8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2105376
Commit-Queue: Anna Malova <amalova@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Tobias Sargeant <tobiasjs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751352}