WebUI NTP: load one-google-bar
The goal of this CL is to make the OneGoogleBar as functional as
possible and allow collaboration with the OneGoogle team to make the
OneGoogleBar completely functional as well as provide the ability to
test end-to-end.

The OneGoogleBar is hosted by
chrome-untrusted://new-tab-page/one-google-bar. The WebUI NTP embeds
this using an iframe.

In order for OneGoogleBar overlays to work, this CL uses mouse event
forwarding. While this may not be the approach used for launch, it will
allow the NTP and OneGoogle teams to test the OneGoogleBar and iterate
towards a better approach.

The Gmail and Images links are also updated so they open in the top
frame. This is temporary until the OneGoogleBar updates those links to
use <a target="_top">.

Bug: 1039913
Change-Id: Ic265d5bfd5c8728c6ff3b6556cede51c92571863
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106739
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Esmael Elmoslimany <aee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751407}