Make sure jobs for CI builders work with or without bucket prefix.
When the number of builders with the same name goes from 1 to >1 or
vice-versa, transient infra failures occur because the scheduler
configuration will be updated while there are already builds running
that will attempt to trigger the old job name. This change removes that
as a problem.

Change-Id: I1547659f61912ef758edb53f86f2092e723a7c54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107805
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751409}