[PM] Fix invariant tracking
This CL modifies PerformanceManagerImpl::DeleteNode() and
PerformanceManagerImpl::BatchDeleteNodes() so that the nodes are no
longer deleted if the task posted to the PM sequence is never run.

This is necessary to ensure the invariant of the graph are still held
during shutdown. At destruction, some nodes DCHECKs that their
properties are valid according to the invariant of the graph. If the
PM task runner has stopped executing its tasks, then those DCHECKs may
spuriously fail because a dependent task was not run.

This CL also fixes a lint warning by upgrading a loop to a range-based
for loop.

Bug: 1058705
Change-Id: I835c5aac5cfc9a12d52818e806ec5ba978af261c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106567
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751421}