[omnibox] [rich-autocompletion] Add no-op flag and params.
*Rich autocompletion background*
Currently, autocompletion is limited to when the user types a prefix of
the URL; if the user omits the beginning of the URL or includes parts
from the title, then the suggestion won’t be autocompleted. This is
problematic because hindering the user when they recall some part other
than the the exact URL prefix conflicts with the omnibox’s goal of
helping users navigate to desired destinations. This feature will allow
autocompletion even when the input is from the suggestion title or not a
hostname prefix.
Design doc with further details: go/richomniboxautocompletion

*This CL*
Adds a flag & 4 params which are unused at the moment, but will be used
by follow up CL's. The work is being split up for easier review. The
intention of the params are:

1. Autocomplete titles.
Currently, suggestions can only be autocompleted if the inputs match the
URLs.
With this param, suggestions will also be autocompleted if their title
matches input. In this case, the suggestion URL will also be shown in
order to differentiate them from search suggestions. Typically, the URL
will shown surrounded with `()` following the title; i.e. `title (URL)`.
If the 'two line omnibox' param is also enabled, the URL will be
displayed below the title.

2. Autocomplete non-prefixes.
Currently, suggestions can only be autocompleted if the input is a
prefix of the URL.
With this param, suggestions will be autocmopleted if the input matches
anywhere in the URL (or title if 'autocomplete titles' is also enabled).

3. Show titles.
Currently, navigation suggestions display only their URLs in the
omnibox.
With this param, the titles will also be displayed. This helps add
parity to the UI regardless of whether the input matches a title or URL
since, otherwise, a URL matching input would display only the URL, while
a title matching input would display both.
For URL matching inputs, titles will be displayed surrounded with `()`
following the URL; i.e. `URL (title)`. If the 'two line omnibox' param
is also enabled, the title will be displayed below the URL.

4. Two line omnibox.
Currently, the omnibox displays 1 line; this is sufficient for
displaying the URL.
With this param, the omnibox incresaes its height to fit 2 lines. This
helps when displaying both the title and URL of a suggestion. If the
'show titles' param is disabled and the input does not match a title,
the 2nd line will be unused since only the URL will be displayed. I plan
to avoid this combination in finch experiments.


Change-Id: If3b0acd403acd72986a929506ba5b7253a991396
Bug: 1062446
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107796
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751441}