Refactor Assistant alarms/timers.
Previously the AssistantAlarmTimerController/Model only handled fired
timers. Due to new UX requirements, we are going to need to handle
scheduled and paused timers as well.

This CL makes the necessary changes to the AssistantAlarmTimerController
and model to support scheduled, paused, and firing timers. Note that
though this CL contains sufficient code to support them, it stops short
of actually handling them as this will be done in another set of CLs.

That being the case, this CL does not have a functional impact on the
existing experience.

Also note that this CL removes some "alarm" related nomenclature. Alarms
may/may not be implemented in the near future and (if we do) the code
will be friendlier to treat alarms and timers as distinct entities.

Bug: b:149570650
Change-Id: I6025471b80def7610625dcfa3c66ce17ba7e2817
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095939
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: David Black <dmblack@google.com>
Cr-Commit-Position: refs/heads/master@{#751512}