Make ComputedStyle::SetFontDescription() preserve FontSelector value
ComputedStyle::SetFontDescription() clears the FontFallbackList (and
hence, FontSelector) on the Font object when successful. As a result,
callers, need to store the FontSelector value before calling, and use
Update() to restore it after calling.

This patch makes SetFontDescription() preserve the FontSelector value,
so that the pattern can be eliminated, and we have fewer call sites of
the deprecated function Font::Update().

Bug: 1049295
Change-Id: I6f4c17d3d0cfe79d2f32ca610ad39ecbf75e2d54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2108867
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751545}