[WebSocket] Add support to consume Datapipe in network service
This commit adds support to network service to consume websocket frame
data from mojo datapipe produced from blink. To achieve this, this
commit adds a new message SendMessage() in mojo WebSocket interface.
It also adds SimpleWatcher to watch the consumer and be notified when
needed, and add callback function OnReadable() to support this.

This commit also adds ReadAndSendFrameFromDataPipe() to actually
consume data from the datapipe.

The producer side of the datapipe and the caller of SendMessage() are
not implemented yet for security review. WIP CLs are uploaded here:
https://chromium-review.googlesource.com/c/chromium/src/+/2082869
https://chromium-review.googlesource.com/c/chromium/src/+/2083777

Design Doc:
https://docs.google.com/document/d/1YWj1z9r8wxemGdod6S2tkchudhp6PvNaH3qSO0oucfY/

Bug: 1056030
Change-Id: I580e3776bf0f1a27e8d83e5c92087f5542b19cb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2071189
Commit-Queue: Keita Suzuki <suzukikeita@google.com>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Yoichi Osato <yoichio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751557}