New terminal works better with upgrade dialog.
Several tweaks made to upgrade flow to make it play better with the
new terminal. Since the terminal launches its window prior to
starting crostini, we move the upgrade availability check to an earlier
phase of LaunchCrostiniApp. Also, when the user decides not to upgrade,
we don't restart crostini twice. (Previously, this could happen because
the upgrade process needs to shut down crostini first).

Bug: 1024693
Change-Id: I46ee09877a7092d91894b80bc798ed99e4f855ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2108300
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Commit-Queue: Nicholas Verne <nverne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751581}