bindings: Make NotShared/MaybeShared smart pointers
NotShared/MaybeShared templates are on-stack pointer wrappers to ensure
the pointed ArrayBufferView is shareable between threads or not.

We could not store these pointers as a part of objects because they are
on-stack objects, and had to convert to a pointer with |.View()|.

  class KeepNotSharedABV : public GarbageCollected {
    using T = DOMUint8Array;
    void setFoo(const NotShared<T>& v) { m_ = v.View(); }
    NotShared<T> getFoo() const { return NotShread<T>(m_); }
    Member<T> m_;
  };

Because of it, it was not obvious if an object holds a shareable view
or not.

This CL makes these templates like smart pointers, and be able to be a
part of GCed objects.  So now we can make it like

  class KeepNotSharedABV : public GarbageCollected {
    using T = DOMUint8Array;
    void setFoo(const NotShared<T>& v) { m_ = v; }
    NotShared<T> getFoo() const { return m_; }
    NotShared<T> m_;
  };


Bug: 839389
Change-Id: Id3d47f1d6a01953b8c82316ebc0d2af68d1685b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2104787
Commit-Queue: Hitoshi Yoshida <peria@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751585}