Avoid generating full document metafiles in PrintRenderHelper
The print compositor is now able to concurrently generate the print
document while compositing the individual pages for preview.  When this
is happening we can improve the renderer by not collecting and
capturing a full document object.

Testing:  On Windows platform watching the "Peak Private Bytes" in
Process Explorer, observed an increase of renderer process memory of
only 59MB, roughly about the size of one page's worth of metadata.  This
is much less than the hundreds of megabytes that it used to grow by when
the renderer would still create the full document metafile.  This was
observed for the testcase [1] (taken from prior issue 872935).

[1] https://www.gmbinder.com/share/-LFhhtssjN-teWCOregn

Bug: 1026968
Change-Id: I697ebd4129722c63e4847d54a65d75a4b3e9f1af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068936
Commit-Queue: Alan Screen <awscreen@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751614}