Reenable test sandboxed-iframe-plugins.html
[Note to sheriff]
Feel free to immediately revert this CL if this test is failing
in the next 3 days.

This test was marked flaky 2 years ago, but nobody was assigned or CC.
As a result, it hasn't been properly investigated. This patch enables it again.
If it works, great. If it doesn't work we will get useful data about it.

R=mkwst@chromium.org

Bug: 745887
Change-Id: I0c50d61992dd56a13aee891e4eedfd14c43bc9df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2108144
Auto-Submit: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751620}