Clean-up: remove redundant End-of-file sentinel from AX test files.
A line containing the string <-- End-of-file --> is included
in the output of DumpAccessibilityTree tests, but it's not
supposed to be checked in.

This change removes those lines. A follow-up will fix the diffing
algorithm and make it so that rebasing tests won't keep adding
those lines to expectations files.

Bug: None
Change-Id: I7bd3186793651e7c4c8361882444b6da9c3d35dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2109176
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Auto-Submit: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751692}