Add option to autofill menu to both enroll in b4p and generate password
In case the user is not opted in to "passwords account storage", the
button in the autofill dropdown menu used to generate passwords will now
have a different UI and prompt a reauth flow to opt in the user to the
feature.

This new button can be shown simultaneously with the pre-existing button
for opting in then displaying password suggestions.

Bug: 1060131
Change-Id: Ibcd85f831f3b81ab28fcb69d4ca0ff02bf4d9c8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107620
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#751714}