Eliminate deprecated function Font::Update()
This patch eliminates the deprecated function, together with its last
callers. Now we can ensure that we never switch FontSelector for any
Font and FontFallbackList objects.

For the last callers in FontBuilder, since it needs to setup
FontSelector as part of the font-related style computation, we expose
SetFontInternal() from ComputedStyle to it.

Bug: 1049295
Change-Id: Iea40e7150389bf0db1d10d0e0269d579bd2948fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2109259
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751715}