Reland "Reland "Ensure UI Automation GetFocus reaches focused node in web content""
This is a reland of a471ff9bebd798c4ac27a32ac6f49280a038083a

The failures that led to the second revert had the same root cause as
the first revert, in two additional tests, with the added feature of
being flaky. The affected tests already had a history of flakiness, but
it seems this change exacerbated the problem.

As near as I can tell it's a timing issue. I verified with VLOG() calls
that we're firing exactly two focus events: one on the document root at
or near startup, and one on the select control when focus() is called
from script. The OS is generating additional focus events for clients.
Some of these appear to be sourced from Win32 focus movements; others
may be coming from the MSAA-to-UIA bridge, or from UIA itself, as a
feature, when selection changes.

AccessibilityEventsMenuListCollapse was introduced in crrev.com/396653
and AccessibilityEventsMenuListCollapseNext was introduced in
crrev.com/583367. From looking at those CLs, it appears they intend to
verify that focus events are *not* fired in the test scenarios for some
platforms, so I don't want to suppress the events from the logs.
Instead, this CL seeks to stabilize the output by moving the initial
focus() from global scope to a go() pass before the existing test body.
By doing so, the focus() call - and the accessibility event it triggers
- is guaranteed to come after kLoadComplete is fired.

Original change's description:
> Reland "Ensure UI Automation GetFocus reaches focused node in web content"
>
> This is a reland of 756820dd8e7bd8e08c5dfb1d05fa570a0846abb5
>
> The original CL was reverted due to test failures on Win7.
>
> On both Win10 and Win7, UI Automation was generating an initial focus
> event when the test listener first connects. On Win10, this initial
> focus event comes from a window managed by the operating system
> (ClassName=ApplicationManager_ImmersiveShellWindow). Since that element
> is outside the scope of the test app, the test listener throws it out.
> On Win7, the initial focus event comes from an element in our web
> content, so the test listener logs it.
>
> My suspicion is that the difference in behavior is due to mitigations
> for focus stealing introduced between Win7 and Win10. Regardless, the
> fix is to generate separate expectation files for Win7.
>
> Original change's description:
> > Ensure UI Automation GetFocus reaches focused node in web content
> >
> > AXPlatformNodeDelegate::GetFocus() returns the node within the called
> > node's subtree that currently has focus. When an element in web content
> > has focus, and a delegate in Views is asked for focus, that delegate can
> > only reach as far as the web content root. Calling GetFocus() again on
> > the web content root ensures that the platform API returns the currently
> > focused node.
> >
> > Bug: 928811
> > Change-Id: Ice47eed0a8ea4281beb48624942ac2428fb1950f
> > Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2099162
> > Reviewed-by: Ian Prest <iapres@microsoft.com>
> > Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
> > Cr-Commit-Position: refs/heads/master@{#749386}
>
> Bug: 928811
> Change-Id: I32e59febfb6dbb7b5386c155c252c880a0f2ebc5
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101930
> Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
> Reviewed-by: Ian Prest <iapres@microsoft.com>
> Cr-Commit-Position: refs/heads/master@{#750781}

Bug: 928811
Change-Id: I6a1095335547078db2427599ef5d8c4d30438f34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106490
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Ian Prest <iapres@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#751732}