Fix invisible AXVirtualViews in TableView by recalculating their relative bounds.
The rows and cells of TableView are given a state of kInvisible when
they should actually be visible. This is because their relative bounds
are calculated before the TableView's size is set up, so the width of
the row or cell is determined to be 0 and thus does not intersect the
bounds of the containing TableView.

The fix is to recalculate the relative bounds of the AXVirtualView rows
and cells whenever the bounds of the TableView are updated. Then the
populate data callback for each AXVirtualView will correctly determine
if the row/cell intersects the TableView and should be invisible or not.

The reason we don't do this recalculation in the populate data callback
is that it is called from the const function AXVirtualView::GetData, but
we want to save the result in the member variable custom_data_ (to
return from AXVirtualView::GetBoundsRect) which we can't do in a const
function.

Change-Id: I5ed6b3dd03ceec8e02bfd7c2491a764598e616b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083397
Commit-Queue: Anna Mace <anmace@microsoft.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751733}