Fix accessibility dump tree tests to not need an end-of-file sentinel.
The DumpAccessibilityTree tests generate an output text file and diff
the results against an expected file. Previously, the way that the
algorithm handled files of different lengths was by adding an
end-of-file sentinel to the end. This resulted in that sentinel
getting checked in, which was just confusing.

Fix this by modifying the DiffLines helper function to properly compare
files that don't have the same number of lines. Add full unit tests for
DiffLines so we can be confident it works.

Continue to LOG the end-of-file sentinel, because it's needed by
running rebase_dump_accessibility_tree_test.py to parse actual output
files from remote logs.

Bug: none
Change-Id: If1bcef063822fd5786acaeff9133c05d5fc5b065
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2109458
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Auto-Submit: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751791}