[CrOS settings] Fix issue where os settings opens in browser
Before this CL, os settings could be opened in the browser. This
CL removes a check that prevented os settings from only being
opened in PWA (settings app).

Removed Disposition_OSSettings_UseNonIncognitoWindow because this test
checks if os-settings is opened in a browser but not opened when browser
is in incognito mode. As os-settings no longer opens in browser this
fails every time.

Fixed: 871144
Change-Id: Idc142e47e7c058597afd6bc7d5402a511119d696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2086316
Commit-Queue: Nnamdi Theodore Johnson-kanu <tjohnsonkanu@google.com>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751819}