[LayoutNG] Ensure fieldset content applies padding in all cases
When the fieldset is expanded to encompass the legend, the fieldset
content no longer applies the fieldset padding. This change
ensures that in this case, the fieldset content is aware of the change
in the border box size so that it will properly take care of fieldset
padding.

Bug: 875235
Change-Id: I2ee67221f195bdbcb06122ee4f01a1024e49ceb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2109128
Commit-Queue: Alison Maher <almaher@microsoft.com>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751829}