[iOS] Only GoBack if CanGoBack on committed interstitials
Check whether it's possible to go back before trying to perform a
back navigation on a committed interstitial when the "Back to safety"
button is pressed. This CL fixes the stability issue--a follow-up CL
will fix the actual functionality of the button on e.g. interstitial
opened in new tab.

Bug: 1059142
Change-Id: Icc03af381ca376731b451238ccb671a216764bfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2111411
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Livvie Lin <livvielin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751881}