Workaround to turn off DXVA video decoder on some AMD devices.
This CL adds |disable_dxva_video_decoder|, which turns off the
DXVA VDA.  In addition to refusing to instantiate one if the
workaround is enabled, it also removes it from SupportedConfigs
so that the renderer shouldn't bother to try.

This CL also makes the D3D11VideoDecoder SupportedConfigs only
show up if the decoder is enabled, so that the renderer doesn't
try to instaniate one that will fail anyway.

Bug: 625785
Change-Id: Ie9323049ce4263ea450127557e7ce20fec8e5625
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106892
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751900}