Explicitly use GPU series type for Intel only
The data type GpuSeriesType seems like it apply to all GPU vendors, but
currently it represents Intel's GPU only. Since this name is ambiguous
and there is no intention to use it for other GPU vendors, this patch
renames it to IntelGpuSeriesType to make explicit.

Bug: 896932
Change-Id: I5b2e92ac622d254015d3669ab9e234dbe8c5a551
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015423
Commit-Queue: Lin Sun <lin.sun@intel.com>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751954}