Enable Wunreachable-code for chromium_code on Linux.
(But not in use_libfuzzer builds.)

Suppress the last few issues. This doesn't always go for the nicest
fix for now, but instead for the simplest behavior-preserving change.

The hope is to get this in while the codebase builds with
the warning on, to make sure that no new regressions land. I'll then
clean things up a bit on linux (e.g. make nice conversion functions for
skia to libyuv and libpng color formats), and gradually enable on
androidchromeosfuchsiaiosmacwindows in follow-ups.

Bug: 346399
Change-Id: Ic920d3b63ac83aa59c2bf862f60662196d0b28e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2093537
Auto-Submit: Nico Weber <thakis@chromium.org>
Commit-Queue: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#751978}