Fix AX hit testing not to send a message from an untrusted renderer
Create a dedicated method in content::mojom::RenderAccessibility to
perform hit testing over an accessibility object that expects a reply
that might include the required information to perform another hit
testing, if a child frame was found to exist at the especified point.

By doing this we can remove the HandleAXChildFrameHitTestResult()
method, previously added to content::mojom::RenderAccessibilityHost,
and fix the current situation where a message from an untrusted
renderer would be sent to the browser when a child frame is found,
since now such information would be shared as a reply to the original
HitTest() method invoked from the browser process.

Bug: 1010831, 1060223
Change-Id: I5b60d8a5e4b9f7fb2fc658c4f9156d7721b38db9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2105217
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Cr-Commit-Position: refs/heads/master@{#752001}