Adjust snapping for background image offsets
When a background image is positioned with respect to the bottom edge
of the background positioning area it may be incorrectly snapped. This
is most obvious for very narrow images intended to attach to the edge
of the positioning area. Fix such cases by accounting for the existing
snapping changes in the positioning area when computing the offset.

Bug: 994096
Change-Id: I24f971082a35c4bc9b3e85632f51511cdbf86c2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2108342
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#752013}