[iOS] Refactor method to display first-run sign-in view controller.
As a part of the sign-in architecture migration clarify the difference
between presenting the sign-in flow for first-run using a navigation
controller and presenting the sign-in flow for upgrade and normal
sign-in using a view controller.

This is a follow-up to patch 2107866.

Bug: 971989
Change-Id: Ia31a47b86956e9db701b8cc1ce6679e71c293278
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095069
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Auto-Submit: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752023}