[ios] Fix crash in tab grid animator
The tab grid animator relies on the condition that
BVCContainerViewController only has one subview which has
the named guide it needs. However, crbug.com/1053452 shows
this is condition is not always met.

Tab grid animator should not be the unwitting enforcer that
BVCContainerViewController does not contain superfluous subviews.
Tab grid animator should just find the named guide it needs.
If BVCContainerViewController having other subviews is a problem,
that invariant should be enforced somewhere other than in
this animator.

Bug: 1053452, 860234
Change-Id: I4a622920301998e8d3211039c6594a637ad54bf5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2112139
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752032}