make translate and scale interpolations return z-preserving OperationType
Set the OperationType in TransformOperation::Blend() according so that
it is a 3D operation type iff either of the blended values are
(previous behavior was to use the second value and ignore the first).
This makes Is3DOperation() work for interpolated operations stops
z values getting dropped during serialization.

Bug: 1057307
Change-Id: Ibcbb3125660ab172a0f65f08b8ccd67534cdd665
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2102640
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Commit-Queue: George Steel <gtsteel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752071}