Add logging to test to diagnose flakiness.
The test ChromeFindRequestManagerTest.FindInPDF seems to have two
failure modes, one where expectations fail, and one where it crashes.
Neither of these seem to reproduce locally, so this CL adds some logging
to determine the state/progress of the test when it fails on the bots.

Bug: 992419
Change-Id: Ia60c26f50318a77299a4433e7ebca96c928a3400
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2112575
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752106}