Desks: Fix a crash when a highlighted desk_name_view is removed.
Tabbing again after removing the desk that corresponds to a
highlighted desk_name_view used to cause a crash.

This CL fixes this crash.
It also fixes an issue when multiple highlightable views are
removed from the same call, where the |delete_index_| (which is
used to determine where to continue traversing after views are
removed) is updated properly.

BUG=1063324
TEST=Added a new test.

Change-Id: I77f21dcec16610ca6d2bc06176753f14f86cdc31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2112817
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752171}