Origin policy: reject unparsable headers with an interstitial
This adds a new message for the origin policy interstitial page which
is specific to the case of an error parsing the Origin-Policy header.

We then hook up the recently-introduced origin policy header parser,
using it in the most basic way: if the header is unparsable, the result
is a network error, showing the new interstitial message.

Fixed: 1042036
Change-Id: Ib169e294cfc66f68b87791aeed96cf92d386f469
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106474
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Christopher Thompson <cthomp@chromium.org>
Commit-Queue: Domenic Denicola <domenic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752175}