Remove UsbChooserContext dependency from website_preference_bridge.cc
The PermissionsClient is used to retrieve the chooser context based on
the ContentSettingsType. WebLayer will not support USB for the time
being, so it returns null for every type. Chrome's implementation only
works for USB_CHOOSER_DATA but can easily be expanded to other
*_CHOOSER_DATA types as needed.

FooChooserContext::GetObjectName is also made a virtual non-static,
renamed to GetObjectDisplayName with a return type of string16, which
is what user-facing strings should always use.

Bug: 1058600
Change-Id: Ida9c37aaa7ca84374640444c846afb08a3e286c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107801
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752190}