Trust tokens: Fix issuance to correctly fetch keys for the issuer
There was a bug in the initial TrustTokenRequestIssuanceHelper
implementation where the issuance helper would attempt to fetch keys
scoped to the top-level origin, not the issuer origin. (Thanks for
catching this, Charlie!) This CL fixes the bug and expands test
coverage by making the FixedKeyCommitmentGetter mock verify that the
origin it's called with matches an expectation provided at construction.

R=csharrison

Bug: 1042962
Change-Id: I03711639945fdc49596d748a5b7aeea76347792f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2113075
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752232}