Move GeolocationPermissionContext to //components/permissions
This breaks out embedder-specific logic into a Delegate class which is
implemented in //chrome. The Delegate can then be passed to the
GeolocationPermissionContext in the constructor.

This is the first permission context moved to //components/permissions.
I created a new subdirectory there (contexts/), so we can separate DEPS
that are only needed for permission contexts from DEPS needed for the
core permissions code.

All the unit tests were moved except for a couple that were specific
to the //chrome version of the delegate.

Bug: 1025609, 1025625
Change-Id: I98cf85e4b7b51d421d0101d27a6ae1bf60594b27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2109247
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752458}