webview: Try deflaking AndroidViewIntegrationTest.
Various tests in AndroidViewIntegrationTest have been flaky for some
time; analyzing failing test logs shows that the WebView renderer is
being shut down while the test is waiting for a callback to occur, many
seconds before the callback timeout is reached. This suggests the
AwContents is becoming unreachable earlier than expected and being GCed.

The Java language spec explicitly permits optimisations which shorten
the lifetime of stack variables - since in many of these tests the only
reference to the AwTestContainerView or AwContents is a stack reference
which is never accessed after loading the initial content, it's entirely
possible for this class of optimisation to break the test.

Store the AwTestContainerView in an instance variable on the test class
instead to prevent this, and ensure its lifetime matches the test class
- this optimisation is not permitted for heap references.

Bug: 949391
Bug: 1058304
Change-Id: Ie6865f6dd7219f5a2d75c914e805d3989f930573
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2113411
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Richard Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752462}