Hook up mayLaunchURL to navigation predictor keyed service
External Android apps may use mayLaunchURL() API to notify
the default browser of the next likely navigation. This API
may be used by the external app to notify browser of the set
of URLs that user may navigate to next.

This CL plumbs that information to the navigation predictor keyed
service. Note that this information is currently not used
in any way (except for recording a local histogram).

Change-Id: I93bfe53a0f15116563021fa45768f9ca459f06d2
Bug: 1014210
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1928035
Commit-Queue: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Reviewed-by: Egor Pasko <pasko@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752466}