[WebLayer] Part 2: Refactor PageInfo.
Currently, PageInfo embeds function pointers in a ChooserUIInfo struct
that map from a Profile to the appropriate ChooserContextBase instance
that is associated with the Profile. However, this organization blocks
componentization of PageInfo.

This CL instead has PageInfo obtain the ChooserContextBase instance to
use for a given ChooserUiInfo via the delegate; the Chrome delegate
impl has access to the Profile and can key off the ContentSettingsType
param in the ChooserUiInfo object to determine the appropriate
ChooserContextBase instance to fetch. External clients of the prior
ChooserUIInfo functionality all have access to a PageInfo instance,
hence, they are ported to this model as well via PageInfo exposing the
necessary API publicly.

Bug: 1052375
Change-Id: Ic6750e5504fb7b2ea23dc6b3cff64542d44b6084
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2111171
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752490}