Move test_runner/test_common.h to content/shell/common/web_test/
This file does not access content internals and is not tied to the
files in test_runner/ that do. But it is also used by
content/shell/*/web_test also, so we move it out to content/shell/ for
them both to use it there.

R=nasko@chromium.org

Bug: 866140
Change-Id: I16722f3b06e0c6487afd546395c1277d20133d12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2113430
Commit-Queue: danakj <danakj@chromium.org>
Auto-Submit: danakj <danakj@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752513}