[throughput] Track last frame
It could happen that after a BeginImpl(Main)Frame is reported, a tracker
is scheduled to terminate. In this case, any subsequent handling of
that impl(main)-frame will be lost. In other words, we completely lost
track of the last frame.

This CL fixes the impl frame, the fix for main frame comes later because
this CL is already big.

Bug: 1055844
Change-Id: I1cc65e9a5d973ecb67c83105be03f7bc08139dae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2079134
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752529}