Support zeroconf discovery of "socket://" printers
Chrome OS supports socket:// and lpd:// printers when specified
manually. Both these backends have corresponding mDNS service types
and can be automatically discovered using almost all of the existing
infrastructure.

https://www.cups.org/doc/network.html says: "Due to limitations in
the LPD protocol, we do not recommend using it if the printer or
server supports any of the other protocols." For this reason, and
because the LPD backend is more complicated, this change only adds
support for socket printers.

Another good reason to not support LPD is that its zeroconf
implementation specifies the use of multiple TXT records, which is
confusing and hard to debug (avahi-browse does not show all 4 records
for my printer, though they are visible with tcpdump). This misfeature
is explicitly called out in the DNS-SD RFC:
https://tools.ietf.org/html/rfc6763#section-6.8.

This enables my HP LaserJet 1320 to be discovered and configured
with the correct PPD.

Bug: 1063535
Change-Id: I9b84b4e6b0e726d5410af31e71013c569d61ef8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2113674
Reviewed-by: Sean Kau <skau@chromium.org>
Commit-Queue: Adam Goode <agoode@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752552}