[omnibox] Adds a MatchStability metric to measure if ANY match changed
Currently we make a histogram log for each match position that changed
asynchronously. This is a valuable metric, because it's proportional to
the number of pixels that are changed that the user can see.

But  it also has the weakness of allowing a single match to displace
multiple matches and generate multiple log counts. In other words, more
bins are logged if more matches exist.

This CL adds a analogous boolean histogram that logs True if ANY match
positions change, and False if NONE of the match positions change:
Omnibox.MatchStability.AsyncMatchChangedInAnyPosition

I suspect both types of logging with be complementary.

This new one should also be normalized by:
Omnibox.Start.WantAsyncMatches::True

Bug: 1062706
Change-Id: Ia325c15f3dfe67631613a136ede7bde51363ad39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2108531
Commit-Queue: Tommy Li <tommycli@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752569}