Measure U2F API web usage in a UseCounter
The existing kU2FCryptotoken{Register,Sign} UseCounters are currently
counted by Blink after the cryptotoken extension issues a WebAuthn
request. Blink counts them to the ExecutionContext from which the
WebAuthn request originates, which belongs to the cryptotoken Chrome
extension.

Instead, increment those UseCounters on the cryptotoken side, using the
MainFrame of the tab associated with the U2F request, so that the
recorded events are attributed to the site that actually made the U2F
request.

Change-Id: I672df8d2abdd0a32af887f2cb37fdb3e11c3338e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107708
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752572}