sync: Allow closing tab when trusted vault key retrieval completes
When the user goes through a reauth flow to retrieve trusted vault keys,
the page should be closed upon completion.

In order to achieve this, NavigateParams.created_with_opener is set to
true, which allows Javascript on the page to close itself. In addition,
the reuse of the NTP is disallowed because, in many relevant scenarios
such as FRE, the NTP is the only existing tab and cannot be closed.

As a fallback mechanism, a URL parameter is added to the reauth URL, in
case the page decides to redirect instead of closing (or if closing the
tab was rejected).

Change-Id: I124ddd299c47a7c5705b410829a38130ea78e208
Bug: 1062977
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2108731
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752585}