Rename some fields in InputMethodSettings.
VK and PK are abbreviations which are ambiguous to readers outside our
team, so rename them with full words.

Rename zhuyinSelectionKeys to zhuyinSelectKeys to match the code in
google3.

Rearrange the sequence of properties, move all non-language-specific
options to top and sort them alphabetically.

InputMethodSettings have not been used in production code, so it is safe
to rename these properties.

Bug: 1060906
Change-Id: Idcf8eedcb1f7c3ee7d97f894599753a35fcd9179
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2114515
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Jing Wang <jiwan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752611}