[AF] Enable AutofillEnableAccountWalletStorage by default
Also cleans up the fieldtrial_testing_config entry for the
SyncButterWallet launch now that it's enabled ToT (and 100% launched).

Per request, not going to clean up the code usage of the flag quite yet
because Butter for Mobile may be coming in the near future.

Bug: 983940
Change-Id: Icfb628cb509c77960d7686c3ad8c5a6161ca7547
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2092064
Commit-Queue: Jared Saul <jsaul@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752621}