[version 2] Hide virtual keyboard when Applist shows
My previous CL (https://crrev.com/c/2106958) does not cover all of
scenarios. For example, show the home launcher by back gesture in tablet
mode or by toggling applist button in clamshell mode.

In this CL, moving the code which hides the virtual keyboard to the
function [1] which handles the visibility change of applist. [1] is
called before the search box view within the applist is focused (which
shows the virtual keyboard through transient-blur feature). In addition,
in [2], this CL ensures that the last blur time recorded in virtual
keyboard controller is reset by [3].

A ash unit test case is added to verify this scenario.

[1] AppListControllerImpl::OnVisibilityChanged
[2] KeyboardUIController::HideKeyboard
[3] KeyboardUIController::HideKeyboardExplicitlyBySystem

Change-Id: I9464e21854c9dfc90e974c2ea98c183419581ab5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2111877
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752686}