Add KeepScriptResourceAlive feature and flag
This CL puts the KeepScriptResourceAlive base::Feature under the
blink::features namespace. It also adds a corresponding about:flags
flag, and RuntimeEnabledFeature.

ScriptLoader's behavior depends on the RuntimeEnabledFeature, which is
enabled when the base::Feature is enabled. The base::Feature is enabled
when the corresponding about:flags flag is enabled.

These are the first steps to running experiments with the
KeepScriptResourceAlive feature. A subsequent CL will add to
fieldtrial_testing_conflig.json, so we can control the base::Feature via
Finch.

R=kinuko@chromium.org

Bug: 1043679
Change-Id: Ib7cd83346c6634dad547311100a614cc43e0fce6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2112172
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752772}