Files app: Fix multi-menu to follow its button
Use ResizeObserver to observe the container parent of toolbar buttons,
when it resizes, means that the buttons have moved, in this case
re-position the dropdown menu to align with its button.

This fixes the issue when the user clicks from the search box in a
button with dropdown and the search collapses and moves the dropdown
button.

Bug: 1063337
Change-Id: Ia8ae426778a25ebf472631b1903b884454d8f33f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2114316
Commit-Queue: Noel Gordon <noel@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Auto-Submit: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752777}