[Passwords] Always show dropdown to unlock account store
With this CL, the passwords dropdown is shown on password fields even if
no password is available. This allows to show the unlock button for
account-stored credentials as introduced with Butter for Passwords.

Restrictions: although the renderer will (with the flag enabled) always
send the signal to show the dropdown, the UI suppresses it in the last
minute if its contents aren't sufficient (i.e. only "Manage Passwords"
would be there).

Therefore, the dropdown is only visible to users who either:
a) have passwords or
b) are logged-in and an opt-in for generation or
c) unlocking remotely stored passwords are visible.

Bug: 1052783
Change-Id: If19199b2774452a6a5a1ef403613a890f55cb554
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062128
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Maxim Kolosovskiy <kolos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752846}