Reland "Enable -Wunreachable-code in Chrome OS builds."
This is a reland of 2ff5fdbc7071dbdd8bca67f3cb72cdcb636e543b with
a build fix in code that's only built on some internal bot
(it changes a FAIL() to an ADD_FAILURE() there, but the test still
passes with that). Also enable on arm builds, since the fix for
that has rolled in since the first attempt.

Original change's description:
> Enable -Wunreachable-code in (non-arm) Chrome OS builds.
>
> Bug: 1063529,346399
> Change-Id: I313ebf4cc16885d1f9bdc5a2ed8db524467735f2
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116590
> Commit-Queue: Ben Pastene <bpastene@chromium.org>
> Reviewed-by: Ben Pastene <bpastene@chromium.org>
> Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
> Auto-Submit: Nico Weber <thakis@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#752618}

TBR=bpastene

Bug: 1063529, 346399
Cq-Include-Trybots: chrome/try:linux-chromeos-chrome
Change-Id: I1dd3cefcaa229b6e2655cee3cd20cf98f32be04b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116912
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752856}