[omnibox] Enable chrome://omnibox logging of NTP realbox and Android
OmniboxControllerEmitter is effectively an observer of
AutocompleteController. This CL makes it the case.

This CL also hooks it up to a lot more AutocompleteController instances
that were previously unlogged (like NTP and Android).

See the bug for a complete plan. This should finish out the bug.

Bug: 1058486
Change-Id: I24a5a23e1c4390d9bced1681d1a233f1b507bdee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2111510
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752866}