Allow same language translation on Android
When the user attempts a same language translation we should just try to
do the translation rather than dismiss the UI or not change the source
and target language.

Under the hood this falls back to allowing the server to decide the
source language.

Bug: 1059573
Change-Id: Ifa52a54dc17481237f8fa29fd71e4e14b433bca8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2097130
Reviewed-by: Doug Arnett <dougarnett@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Megan Jablonski <megjablon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752871}