Hopefully deflake IgnoreKeyWithExpiryInThePast
TrustTokenKeyCommitmentParsing.IgnoreKeyWithExpiryInThePast was reported
as flaky. From inspecting the code, it looks like the test can fail when
the mock timer is within a minute of the Unix epoch. I don't know for
sure if this is the cause of the flakes, but, if it is, this patch
should resolve the flakes: it ensures that the mock timer is at least a
minute past the epoch.

R=csharrison

Bug: 1064069
Change-Id: Ib61285431a7fc05de57e8a6596f387c784816fa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116914
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752878}