Adding priority for OOBE Screens.
This CL adds priority for all OOBE screens. Some OOBE screens have an
explicitly defined priority while others have it as DEFAULT. In the
next CL that would follow, the wizard controller would use this
priority to check if it is allowed to transition from the current
screen to the requested screen.

Bug: 1064271
Change-Id: I392ece10cfdf434fdf5a4027fd4ba0f8604bfe1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116224
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752910}