Desks: Fix backdrop stacking after dragging and dropping a window to another desk
In overview, the backdrop is hidden, but even though we don't show it
we still have to update its layout and stacking order, without waiting
for any on-going animation of the window having the backdrop.

Updating the backdrop may lead to its removal or its creation. This
should not result in a double update of the desks mini views' layers.

BUG=1055732
TEST=Manually, added a new test.

Change-Id: I69f3c9577f2b7967e5738ed8a766933edc8f6f73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2117164
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#752914}