Check autoscroll conditions on mousemove to determine when to stop
Currently, autoscroll continues even when the page is resized to the
point that scrolling is no longer available, including when the page is
maximized. To fix this, autoscroll conditions (if a page is scrollable)
will be checked each time HandleMouseMoveForMiddleClickAutoscroll is
called, and if the conditions no longer exist, autoscrolling will be
stopped.

Bug: 1059721
Change-Id: Ieb89b86a73370d61380fb51fd59405b147e34a30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2109286
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Mario Bianucci <mabian@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#752927}