X11: EGL: Check if eglGetMscRateCHROMIUM is loaded.
If Linux/X11 is used with system libegl/libgles2 (currently, it
uses angle by default) aka --use-gl=egl, it crashes because
eglGetMscRateCHROMIUM is not loaded. That API is not part of
drivers, and particularly not part of Mesa drivers.

Thus, check if it is loaded. If not, return early.

Test: Run Linux/X11 with --use-gl=egl
Bug: 1042393
Change-Id: I93436e163ffed6cde41c1faa9088a43e12579422
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2115473
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#752962}