Fix ChromeWebContentsViewDelegateHandleOnPerformDrop TSAN failure
The issue was introduced by https://crrev.com/c/2092062 when it changed

UploadFileForDeepScanning(data_.paths[index], std::move(request));

to

request->GetRequestData(base::BindOnce(OnGotFileInfo, ...));

in DeepScanningDialogDelegate. That change is correct in context, but
it makes the response asynchronous since the file is now opened in
tests. This introduces a race condition with this test since it's not
using ScopedFeatureList correctly. The fix is to only set the feature
in the test ctor (as recommended by base/test/scoped_feature_list.h)
and to never re-set it in the tests to avoid races.

Bug: 1063080
Change-Id: Ie965c158f39e0cca697e59899f2b15232a1cba2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2111905
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753013}