Fix highlighting bug for extensions toolbar button.
Both the BubbleDialogDelegateView and the MenuButtonController were
handling the ink drop highlight for the ExtensionsToolbarButton which
caused the highlight to not appear when
MenuButtonController::Activate() is used to show a popup rather than a
menu. This change lets the ExtensionsToolbarButton observe the widget
and take the lock. It also prevents ExtensionsMenuView from updating
the ink drop a second time. This bug is specific to showing a popup,
instead of a menu, anchored to the button.

Bug: 1046422
Change-Id: I6f35d1415e66b603d8498c5950f15acc03cde8b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083673
Commit-Queue: Caroline Rising <corising@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753044}