Change the type of surrounding text to string16.
In the series of SetSurroundingText and OnSurroundingTextChange calls,
the current type of surrounding text is string, but the cursor index
is calculated based on string16. So if the surrounding text contains
emoji, the text before cursor calculation could be wrong. So change type
of all surrounding text to be string16.

Test: tested with Linux emulator
Bug: 1042084
Change-Id: I17df2d0a8a73b8db9d756c782aff8caae0ecb554
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2105264
Commit-Queue: Jing Wang <jiwan@chromium.org>
Reviewed-by: Shu Chen <shuchen@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753057}