[ios] Adds a fixture to support unittesting of individual policies.
The PolicyTest fixture supplies PrefServices that are managed by policy.
It also supplies a MockConfigurationPolicyProvider that can be updated
with test-specific policy values.  Individual tests (to be written in
the future) can use this infrastructure to verify that preferences are
properly set based on policy data.

BUG=1024115

Change-Id: I5ad4e9ae0089eb55253701806e371680f608d81f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2117317
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753083}