WebLayer: enable DRM video
This consists of enabling the protected media id permission and hooking
up MediaDrmStorageImpl.

Chrome's ProtectedMediaIdentifierPermissionContext is not reused because
the majority of it is Chrome OS specific and the rest of it is
functionality WebLayer doesn't need. The default for Android M+ is
allow, so no prompt is shown. See
GetInitialDefaultContentSettingForProtectedMediaIdentifier()

Test:
https://shaka-player-demo.appspot.com/demo/ > Sintel video should
play (widevine)

Bug: 1062474
Change-Id: I0e68cb78b9d88ad97d272253b62c3722fd7e3f2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116550
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Auto-Submit: Evan Stade <estade@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753087}