screen_win.cc cleanup part 8
More simplification.

GetScreenWinDisplayNearestDIPRect() calls
SquaredDistanceBetweenRects() up to twice as often as before, but
since the number of displays is small and this function is cheap,
that won't have any meaningful performance effect, and this was the
simplest, clearest way to write the code.  (The previous patch set
used a longer form that involved transform() to avoid calling this
function more times than necessary, but if we cared about perf, that
would probably be worse on net, given the cost to allocate/deallocate
an extra vector and read/write different objects in the loop.)

There is one behavioral change here: previously,
IsWindowUnderCursor(nullptr) would return true if GetCursorPos()
failed.  This seems incorrect (if unlikely).  The current version
returns false in this case.  (If WindowFromPoint() returns nullptr,
IsWindowUnderCursor(nullptr) will still be true.)

Bug: none
Change-Id: I0f051a6f0da23eff37ecfb24f1057f4b1fb4e3bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2104369
Commit-Queue: Robert Liao <robliao@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753094}