chrome: Add workaround for samsung dex display with SurfaceControl.
SurfaceControl relies on transaction acks from the framework to
manage pipelining/backpressure for frames sent by display compositor
to SurfaceFlinger. These are not getting dispatched when Chrome is
rendering on an external samsung dex display due to a platform bug
which stalls the rendering pipeline.

This change adds a workaround to disable SurfaceControl while the
Chrome activity is associated with a samsung dex display.

R=boliu@chromium.org

Bug: 1042581
Change-Id: I38c067f6f7f48cf04c192cc59efe95a4c21fa4ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2099602
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753118}