Add chrome://settings buttons to choose opt-in to account passwords storage
Feature kEnablePasswordsAccountStorage disabled:
https://screenshot.googleplex.com/AaNqS7OxpNw

Feature enabled, user not opted in:
https://screenshot.googleplex.com/RBR6E3yzmkk

Feature enabled, user opted in:
https://screenshot.googleplex.com/4GbzCGPeddp

On click of the opt-in button, a reauth flow will be triggered. On
click of the opt-out button, no reauth is triggered. The UI is only
temporary and will be updated once mocks are available.

Bug: 1049141
Change-Id: Iceb293f3809e8995056a7c36989a6abeb462db62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2108030
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#753165}