Add tests for process reuse and COOP+COEP
This CL adds two tests that exercise the interaction between COOP+COEP
and the process reuse mechanism. Currently, the process used to commit a
COOP+COEP page is included in the process reuse mechanism. We'd like to
take it out of the mechanism, to ensure that no document are ever placed
in a COOP+COEP process.

Bug: 922191
Change-Id: I2055b20676350cf13c24328ac2bc019a6adcdcbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2063005
Commit-Queue: Camille Lamy <clamy@chromium.org>
Reviewed-by: Pâris Meuleman <pmeuleman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753171}