Reland "WebGPU: Add early-returns in the request of adapter and device when context is lost"
Reland with the fixes on the missing check in the test bodies so that they can early
return when the platform does not support WebGPU.

Original change's description:
WebGPU: Add early-returns in the request of adapter and device when context is lost

This patch adds early-returns in the request of GPUAdapter and GPUDevice
when the WebGPU context is lost.

In the following patches we will implement more clean-ups when the
WebGPU context is lost.

TBR=kbr@chromium.org

Change-Id: I64f6b4c88d868f0e83c131875bfacd465279948a
Bug: chromium:996713
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2119350
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753173}