[WebApk] Fix logic that checks if a PWA is installed.
We were checking that the url of the installed PWA is a prefix of its
scope. However, the url we have available from the PermissionContext
is only the requesting origin, and hence this check fails if the scope
is longer than the origin.

This CL reverses the check to verify that the url is a prefix of the
scope. A unit test has also been updated to test for this case.

Bug: 1062177
Change-Id: Ic81f9b5b10fcb378cc3d65dadee947f53e8755b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107599
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753220}