clang build.py: change LLVM_ENABLE_ZLIB from ON to FORCE_ON
This uses the FORCE_ON option to enable zlib to ensure that clang
actually finds and uses zlib. Also moved this setting out of
base_cmake_args so that it is only used in the main clang build,
since there is another 32-bit build that can't use this zlib.

Also changed to remove and re-download the zlib source on
every build.

Change-Id: Ifceabc2ba06e4e4a56464f18de80610a0f043641
Bug: 1061189
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2118674
Commit-Queue: Amy Huang <akhuang@google.com>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753301}