Modify Tabs Extension API to handle "unloaded" pages.
Since M69 it has been possible for tabs to restore and not be
loaded for an indefinite period of time. This new tab state was
not reflected in the tabs extension API.

Concretely, this adds a third "unloaded" state to the "status"
enum, which previously was one of "loading" or "complete".

BUG=1061209

Change-Id: Icadb41fe033ae47630affeabffb38fa5b3e9d79e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2105564
Commit-Queue: Chris Hamilton <chrisha@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753345}