Allow single-axis flex rules.
FlexRule by default doesn't know the orientation of the layout it's in,
so the existing pre-baked rules you get from calling:
FlexSpecification(MinimumFlexSizeRule, MaximumFlexSizeRule, bool)
apply the same logic on both axes.

This is typically not the desired behavior, however. For example, if you
have a FlexLayout with cross axis alignment = center and the default
flex rule uses kPreferredSnapToZero, then if a child view is slightly
too large on the cross axis, instead of being clipped at the top and
bottom, it drops out entirely.

This resulted in several kinds of yucky bugs (see below for examples).

What we really want is the ability to easily provide a flex rule that
applies only on the main axis; possibly with a different cross-axis rule
that would default to using the view's preferred size.

This CL adds a constructor of the form:
FlexSpecification(
    orientation,
    main_axis_min_size,
    main_axis_max_size = preferred,
    use_height_for_width = false,
    cross_axis_min_size = preferred)

The order of arguments is chosen so that those with defaults least
likely to be overridden are last.

This new constructor would be used in 95% of the cases where we
currently call the three-argument FlexSpecification constructor
mentioned above, and can still replace the other constructor in the
other 5% of cases at the cost of a little verbosity.

The old constructor is not removed but a note is placed on it
suggesting the new constructor be used in most cases.

Bug: 1061869, 1063275, 1063455
Change-Id: Id8a85f9d80bfe393344ce25f3ed885f18b48d29f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2113690
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753365}