Fixed Bubbles not tracking the NativeTheme of their anchoring Widget
Currently child windows do not track the themes of their anchoring
widget. This can lead to theme mismatches in spawned bubbles (in
particular icons relying on native theme colors in the profile menu
for linux themes).

This change follows on from a previous change that fixes the issue
for ThemeProviders in Bubbles. Follow up CLs should address the
issue for the more general case of widgets tracking the themes of
the Widgets that spawned them (see https://crbug.com/948859).

Bug: 1057962
Change-Id: Ie3a71208cb920e8bced6a8bb2f42968bad46f403
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2121075
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753398}