[CompositeAfterPaint] Do the same for paint property compositing reasons as pre-CAP
When we added the conditions we expected that crbug.com/900241 would
be fixed soon for CompositeAfterPaint, which hasn't happened yet.
Removing these conditions to reduce difference between CAP and pre-CAP.

Another reason of removing the conditions is to let the effect/filter
node know some of the compositing reasons of the transform/effect node,
so that when an object is forced compositing with
will-change:transform/opacity, we won't decomposite that effect to avoid
raster invalidation on change of the effect.

Bug: 765003
Change-Id: I4f2deda6b08c1efb93841a9a804c22e99dd42f2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2118390
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753401}