[media benchmarking] Replace myself with John and Frank.
It's best to have multiple owners in the event of OOO,
so I have nominated John as an additional owner of
load_library_perf_tests and Frank as an additional owner
of media_perftests.

Owners are responsible for understanding how these tests
are important for the product. They do not need to
understand how the tests are run or how the infrastructure
works.

I am leaving Google.

TBR=johnchen@chromium.org

Change-Id: Id217cf0364285f6d09a4e8eba25ea61789ab08fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2121034
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Auto-Submit: Caleb Rouleau <crouleau@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753440}