Pass in a populated file list to media-app.
At the moment media-app receives a file list with only 1 item in it.
This CL updates the launch code to send over all files in the current
directory if the app was launched via the files app.

This is part of a larger set of CLs to expose more directory methods and
data to the guest frame.

Bug: 996088
Change-Id: I485bbe143fcc953c8cd189ded30d0cf23f337fbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2108453
Reviewed-by: Trent Apted <tapted@chromium.org>
Commit-Queue: Zain Afzal <zafzal@google.com>
Cr-Commit-Position: refs/heads/master@{#753464}