Platform agnostically rename PrinterStatus and PrinterQueryResult files
crrev.com/752905 moved the definition of PrinterStatus out of
cups_jobs.h, and named the new file with the assumption that it will
only be built on ChromeOS. Given that CUPS IPP will be supported on
macOS in the near future, this naming was not appropriate. Do the same
thing with PrinterQueryResult.

Bug: 226176
Change-Id: I54e99b63f40221a832ab95196c8c9c14ba7f2011
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2121081
Commit-Queue: Daniel Hosseinian <dhoss@chromium.org>
Reviewed-by: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753472}