MacPWAs BMO: Access AppShimRegistry from web_applications
This access was previously living in apps/platform_apps,
and using extension mechanisms.

In a previous patch, I had moved AppShimRegistry into
apps/app_shim. This was a mistake -- AppShimRegistry is
only for web apps (not for all app shims). Move it to
be in web_applications/components, and update it
in web_app::AppShortcutManager.

R=loyso
TBR=avi (chrome/)

Bug: 1057195
Change-Id: I284a56b472fde55f33f30a84c5959a68e4f0a04a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116564
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753483}