Use OOP-R APIs in CanvasResourceProvider behind flag
This change sets up a code path to render a simple canvas with OOP
rasterization. It covers the most basic Canvas2D cases (ie. draw
operations with stroke()).

This CL also introduces diverging code paths based on the
CanvasOopRasterization flag. OOP-R canvas functionality will remain
divergent behind this feature flag as we work towards parity with the
current code path, at which point the current code path will be removed.

In this change, the flag is often used to bypass uses of GrContext,
SkSurface and SkImage (all of which are unavailable with OOP-R).
References to these objects will be replaced in future changes as we
flesh out the OOP code path.

This is being done as part of the OOPR-Canvas2D project. For more info
about the project see the tracking bug here: crbug.com/1018894

Bug: 1023277
Change-Id: Iffd279418f7b104a84513f14495983fc63e623c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2076555
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753520}