[Autofill] Use country code from Variations for autofill addresses.
Currently, either the country field in an address form is used to
determine the address requirements for autofill profile creation, or the
the app_locale is used as a fallback option.

This change adds the country code supplied from Chrome Variations as a
primary fallback before the app_locale is used to deduce the country of
a user. Also the country code is used to as a source for the default
country for address creation.

Change-Id: Ia7d28c66a31775b26d7d33153670b1db6eade74c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107981
Commit-Queue: Matthias Körber <koerber@google.com>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753543}