[WebLayer] Check URL change behavior in external navigation tests
This CL adds verification of the behavior of the URL display to the
tests of WebLayer's external intent launching:

- When a normal navigation occurs, the displayed URL should be updated
  accordingly.
- When an intent is blocked from launching, the displayed URL should
  remain that of the previous navigation.
- When an intent is launched, the displayed URL should remain that of
  the previous navigation.

This CL also changes the tests so that the navigations to intents are
not the initial navigations in the Tab. The reason is that intent
launching on initial navigations is a special case; we will add
specific tests for that case in a followup. As preparation for those
tests, this CL also moves the launching of the shell into each test
so that the tests explicitly specify whether the shell should be
launched with an initial URL (and hence navigation).

Bug: 1029710, 1031465
Change-Id: I7a5fe59dd3010f3f1d7eceaa7294e121d88cd99e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120302
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Mugdha Lakhani <nator@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753618}