[ScrollTimeline] Consolidate flaky scroll-linked worklet animation tests
This patch puts all related flaky tests together for better tracking.

The flaky tests are not related to worklet animations because it's the
reference file that is not painted correctly.

Mac scrollbars have their own painting path and they paint greedily.
There might be a race condition on the order of painting that caused the
scrollbar being completely black. crbug.com/682209 tracks the effort of
unifying the scrollbar painting logic.

Bug: 946534
Change-Id: I9347887addbaf974e2502ed7c5def3a8145e7d91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120733
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Commit-Queue: Yi Gu <yigu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753642}