[UI] Check any extension action for an icon to ues in the zoom bubble
The zoom bubble will look for an extension icon to show if the zoom
change was caused by an extension; preference is given first to the
extension's icons specified in the "icons" dictionary, and then to
the action icons. Instead of only looking at browser action icons, look
at any action the extension specifies.

Bug: 893373
Change-Id: I424484202be48655de61e98b9b317d9001d214cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080822
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753705}