gpu: extract common VA-API components into a new target
This CL moves vaapi_wrapper into its own target so that it can be used
in the SharedImage interface. This involves removing the dependency on
platform_video_frame_utils.h so that
//gpu/command_buffer/service:gles2_sources can eventually depend on
VaapiWrapper without causing a circular dependency. I moved all
call sites of VaapiWrapper::CreateVASurfaceForVideoFrame() to use
VaapiWrapper::CreateVASurfaceForPixmap() instead to remove the
dependency on platform_video_utils.h.

This CL is a prequel to CL:2083662

BUG=1058103
TEST=on atlas: tast run video.DecodeAccel.*

Change-Id: I10e0a28cf9cc38bef43839dad326b7e2babc6387
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106485
Commit-Queue: Paulo Warren <pwarren@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753715}