Ensure that translate bubble always shows within the window bounds.
Translate bubble was getting confused in some cases because of the way
the translate button is placed in PWA mode, especially during immersive
mode (see attached bug).

This change causes the translate bubble to prefer to be on the screen
rather than switching its mirroring from the button (which may not even
be visible).

Bug: 1063745
Change-Id: I8308d84f130b1ea59b2ba3db82179561bc035213
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2122387
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753731}