Add OnBeforeConfigurationRemoved to NetworkConfigurationObserver.
Wi-Fi sync needs to know the ssid and security type of a network
in order to remove it from sync.  Once a network has been removed
this information is no longer available.  This observer method
allows us to cache the required metadata for use once we know
that the removal was successful.

Bug: 966270
Change-Id: Ic850ae8fc75fe600ce70ca2a161bb4281be4119e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2110392
Commit-Queue: Jon Mann <jonmann@chromium.org>
Reviewed-by: Azeem Arshad <azeemarshad@chromium.org>
Reviewed-by: James Vecore <vecore@google.com>
Cr-Commit-Position: refs/heads/master@{#753777}