[Bookmarks] Remove conditional hiding of default bookmark commands
Extensions used to be able to override the bookmark star view in order
to provide a custom bookmark manager (this was used for the Stars
extension). The extension has been deprecated and removed, and the API
never launched. Cleaning up the remnants.

Remove the logic to hide the default bookmarks commands conditionally if
an overriding extension is installed.

Bug: 894447
Change-Id: Iddbdb96171fb567ee9adf06390cfa4668996d58f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2119805
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Mike Wittman <wittman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753794}