[Extensions] Remove Bookmark Overrides capabilities from CommandService
Extensions used to be able to override the bookmark star view in order
to provide a custom bookmark manager (this was used for the Stars
extension). The extension has been deprecated and removed, and the API
never launched. Clean up the remnants.

Remove the logic in CommandService to check if an extension overrode the
bookmark commands.

Bug: 894447
Change-Id: I2ed7aa58acd7796155248649b1a0fad32f932f8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2119389
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Mike Wittman <wittman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753834}