fido: add googleAndroidClientData CTAP extension
This adds support for an extension that allows passing of unhashed
inputs to the clientData struct to a CTAP2 authenticator.

An authenticator implementing this extension signals support via its
GetInfo response. If supported, Chrome encodes the
AndroidClientDataExtensionInput structure as a CBOR map into the
`googleAndroidClientData` extension in any makeCredential or
getAssertion request sent to the authenticator.

An authenticator receiving the extension MUST use the origin and
challenge values when making requests to the Android FIDO API because
that API requires those values in lieu of the CTAP2 clientDataHash. The
authenticator MUST include the clientDataJSON from the Android FIDO API
response in its CTAP response map under key 0xF0 as a CBOR bytestring
value.

This extension is a temporary hack to work around a limitation in the
GMSCore FIDO API for the caBLE v2 authenticator. Support for the
extension is therefore gated behind the WebAuthenticationPhoneSupport
flag.

Bug: 1002262
Change-Id: Id16534f05f5dbfa4fb2a6adb7c77cb19f63e9e05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2119266
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753841}