Refactor assistive suggester
As there are many assistive experiments, refactor assistive_suggester to
allow for more suggesters to be added.

Coming use case would be emoji suggester in crbug/1049826.

PersonalInfoSuggester is almost a complete fork of AssistiveSuggester,
except for HandleKeyEvent and Suggest function. It should probably be an
observer of AssistiveSuggester, but will do that in another CL.

Bug: 1060965
Change-Id: I639593cc81822ca99c8b9a34b62693cd5ec7b26a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120116
Commit-Queue: My Nguyen <myy@chromium.org>
Reviewed-by: Keith Lee <keithlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753851}