[serial] Reduce the number of GC iterations to avoid timeout
Dashboards are showing that serial/serialPort_readable_gc.html is timing
out flakily. I cannot reproduce this locally but a simple fix is to
reduce the number of garbage collection iterations performed in the
test by calling collectAll(), which uses only enough to force a full GC.
The downside of this is that it may reduce the ability of the test to
detect regressions.

Bug: 1057938
Change-Id: Ieca2c93ce2e67cf60999ac8a74b1ed9da97c8e9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2123510
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Kouhei Ueno <kouhei@chromium.org>
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753860}