Have DefaultScreenPositionClient() keep the Aura root window updated.
It sets itself as the root window's client on construction and nulls
that out on destruction.  DesktopScreenPositionClient already did this,
this just hoists that to the base class.

This makes it harder to introduce bugs like "constructed the client
but forgot to set it on the window" or the more likely (and insidious)
"failed to null the client on shutdown, leaking state into another test
and triggering test flakiness".

This relands r753540, with bugs fixed.

Bug: none
Change-Id: I35c8062c27683c9d04281df3af37fe438f4c64f0
Tbr: shend, yucliu, rdevlin.cronin, sadrul
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2123027
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Yuchen Liu <yucliu@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753864}