Implement delayed warnings for SafeBrowsing
This experiment delays SafeBrowsing warnings for phishing warnings
until the user interacts with the page by typing into it. The goal is
to measure the effectiveness of site identity related changes in the
UI.

Implementation:
The CL adds a UserInteractionObserver class which is a
WebContentsObserver. SafeBrowsing URL checker creates an instance of
it when it finds a phishing verdict. Upon creation, the observer
stores the unsafe resource and starts observing keyboard events with a
callback.

The observer listens keyboard events by passing a keyboard handler
callback to the current RenderWidgetHost instead of implementing
WebContentsObserver::DidGetUserInteraction(). This is to intercept
input events early so that they don't leak to the renderer (also,
DidGetUserInteraction() doesn't allow cancelling events).

Once a keyboard event is received, the observer calls into UIManager
to display the interstitial and deletes itself.

There will be follow up CLs which will show the delayed
interstitial at other security moments such as popups, downloads and
permission notifications.

Delayed warnings will only be implemented for the browser and not
for Android WebView or WebLayer.

Bug: 1057157
Change-Id: I7c12d102d3be2b4fe2e4ede22223749091f7034e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050018
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753871}