WebIDL: Make Union definition object as old generator does
Current code generator generates two union types (A or B) and (B or A)
as different types.
In the new code generator, which is enabled behind a flag, we plan to
merge these union types.

This CL makes the new generator work with the old Union's style during
the migration.


Bug: 839389
Change-Id: Icd24043994134f48e323f2011115d9f834c9a42c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2121947
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Commit-Queue: Hitoshi Yoshida <peria@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753933}