Add API to notify pwd manager UI of deletion of credentials on sign out
The CL adds a new method to the API of PasswordStoreSigninNotifier,
which will be used in future CLs to plumb to a bubble UI any credentials
that were unsynced at the time of sign out. The method is responsible
for choosing the tab where the UI will be futurely displayed and passing
the credentials to the ManagePasswordsUIController attached to it. The
latter is done via a new method in the UI controller, left with
NOIMPLEMENTED for the moment.

Bug: 1060132
Change-Id: I52e7cce0adce4d550f02a4d61a01d04dd19c0fff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120474
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#753943}