gl: x11: rename gl_surface_egl_x11 to gl_surfae_egl_x11_gles
Currently, NativeViewGLSurfaceEGL has a bunch of code guarded by
USE_X11. This should be unified and placed under X11 interfaces
managed in compile time. What is more, this code must be shared
with ozone/x11 as well.

Thus, to avoid confusions in names of the classes/files, rename
gl_surface_egl_x11 to gl_surface_egl_x11_gless and rename the
class NativeViewGLSurfaceEGLX11 to NativeViewGLSurfaceEGLX11GLES2
as that is always used when --use-gl=egl is passed.

In the next CL, I will extract all the x11 code from
NativeViewGLSurfaceEGL and place it to NativeViewGLSurfaceEGLX11.
The reason why this is going to be managed in such a way is that
currently when Angle is used with X11, NativeViewGLSurfaceEGL is
used instead of NativeViewGLSurfaceEGLX11. And, thus, all the
code that is guarded by USE_X11 is compiled and executed.

Thus, when --use-gl=angle is passed (this is default for x11/linux
now), NativeViewGLSurfaceEGLX11 will be used. And when
--use-gl=egl is passed, NativeViewGLSurfaceEGLX11GLES2 will be used
(it initializes native platform window as well, which is not
required in case of angle).

Bug: 1036285
Change-Id: I9bfe89a24c36ce80061137eae05f3c4dc00aa1c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116629
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Jonathan Backer <backer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753983}