Reset high-priority bits in CascadeMap
Not resetting this didn't lead to any bugs, because it's just a
performance optimization. Even if a given bit is set, nothing will
happen if the corresponding property is missing from the cascade.
Also, we only really Reset for DCHECK_IS_ON.

Still it's proper to actually reset the value.

Change-Id: I57b023316c46c8f8def9c634319116d63c4503f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124250
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753995}