Reland "Mac: clean up FullscreenToolbarController"
The original code retained the window's content view in the
controller because it "might be needed in dealloc". Turns out,
this was sort of true. In some circumstance (can't repro),
if fullscreen is ended after the content view has dealloced,
we blow up trying to remove the mouse tracker's tracking rect
from the view.

Since we can't test a more targeted fix due to lack of repro,
Patchset 3 adds a view retain directly in the mouse tracker with
a less vague comment.

This is a reland of 5407de10518bb232852b50a71d2674a092fdbd09

Original change's description:
> Mac: clean up FullscreenToolbarController
>
> - Remove some structure that was previously necessary for Cocoa browser
> - the visibility lock controller is vestigial: we release the lock,
> but never acquire it
>
> Bug: 1063417
> Change-Id: I55ccc6b14b0ffed7d81beb640a54ec8b1110e831
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2117321
> Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
> Commit-Queue: Leonard Grey <lgrey@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#753211}

Bug: 1063417
Change-Id: I2d88c1bec50cc2ff116ff66372e51fa05cc10361
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2122389
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754025}