Implemented a repeating callback in UpdateService::UpdateAll.
This callback is going to be needed shortly when --ua implementation
is going to use COM RPC and UpdateServiceOutOfProcess.

A couple of smaller fixes:
* Added missing callback on success in
UpdateServiceOutOfProcess::UpdateAllOnSTA
* renamed "done" to "callback" to match the existing naming style in
this code and update_client.
* the actual implementation of the repeating callback is mechanical but
provided in the existing code to show the state mapping and ensure the
code is not crashing when a repeating callback is invoked.
* similar work is needed on macOS.

Bug: 1058858
Change-Id: I509b01ae4970478ecb617b6361715f62706cde45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2123318
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754027}