Remove dependency on /extensions in ManagedBookmarksPolicyHandlerTest
The dependency on /extensions was solely due to the use of a list builder for testing. This CL removes this unnecessary dependency.

Also, the dependency on /extensions prevented these tests from being run on Android, and would have prevented them from being run on iOS in the future.

Bug: 1065187
Change-Id: Ida63651f4c67f08ba3981f5d277afcdbf438cb3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2122820
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754071}