Fix back button in Ash PIN request dialog
This CL fixes the regression from https://crrev.com/c/2089684 that the
back ("x") button in the Ash PIN request dialog wasn't clickable via
mouse or touch screen.

The underlying problem was that the button's view was overridden by
another invisible view, so that one can see the button and even tab into
it but not click it via mouse. The fix is to make sure that the button's
view is on top of other views, i.e. it's added last into FillLayout.

This CL also improves the existing unit tests to become a regression
test for this bug. Now the tests will simulate mouse clicks in a way
that is closer to the real-world flow (piping them through the whole
views hierarchy, rather than directly firing the button press event).

Bug: 1064368
Test: log into Chromebook using a smart card, lock screen, start unlock, click "x" in PIN dialog, check that dialog closes
Change-Id: I9b20896a83f5fa251d3b05c5a8bc97ace9ebcd7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120295
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754089}