DumpWithoutCrashing when trust token params deserialization fails
crbug.com/1062637 reported URLLoaderFactory
VALIDATION_ERROR_DESERIALIZATION_FAILED crashes starting shortly after
adding a new member to ResourceRequest. The new member, of type
OptionalTrustTokenParams, is only used for functionality currently
behind a flag (and not enabled anywhere in production).

To help diagnose the cause of the crash (and, in particular, whether
it has to do with the trust_token_params field) this change separates
the deserialization of this field from the large "if (!Read(A) ||
!Read(B) || ...)" conditional in url_request_mojom_traits.cc and
specifically dumps on failure to deserialize the field.

R=yhirano

Bug: 1062637
Change-Id: Icc199fe29bcf1179d762971fb383ffee95b31ee9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2119764
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754090}