AW: float non-default flags to the top of the list
If no flags are toggled, flags are shown in the same order (the order
hardcoded in ProductionSupportedFlagList). If the user has already
toggled some flags (enabled/disabled), these flags float to the top of
the list when the user revisits the UI.

The flag order only moves when we recreate the flag UI. So flags will
stay in the same place as the user interacts with the UI, and only move
as a convenience if they return to the UI and need to go back to
default. This is mostly consistent with chrome://flags (the difference
is that we won't reset the order when the user clicks "reset all to
default," for simplicity and convenience if the user changes their mind
to re-enable some of those flags).

Fixed: 1059478
Test: Manual - verify flags float to the top when I revisit the flag UI
Change-Id: I01cac5e7c38b35be765c3e111f4b0d5ea4e3941f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124532
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754165}