[CompositeAfterPaint] Allow non-atomic effects in multicol
This is not fully correct, but we have to accept it in our current
multicol painting based on FragmentClips.

Actually we have already allowed non-atomic effects in multicol within
composited layers since SlimmingPaintV175. A fragmented element with
effect applies the effect on each fragment separately. We have to do
this because each fragment has its own fragment clip, and its own paint
properties because they may depend on the fragment clip.

This CL enables non-atomic composited effects for CompositeAfterPaint.
This is not an issue in pre-CompositeAfterPaint which doesn't allow
fragmentation of composited layers (which is incorrect).

The root issue will be fully fixed by LayoutNG block fragmentation
which doesn't need FragmentClips.

Bug: 1064341
Change-Id: Id6ae4c263381f24d037f3c2bc493d73cd050598e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124340
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754166}