Stop relying on the value attribute for boundary and hypertext calculations
Blink doesn't always remove all deleted whitespace at the end of a
textarea even though it will have adjusted its value attribute, because
any extra layout objects that might happen to exist in its layout tree are invisible.
Similarly, the value attribute is not updated when an autocomplete
popup is visible in a native text field.
Therefore, We stop relying on the value attribute for hypertext and text boundary calculations
when a text field originates from Blink and use its inner text instead.
We detect this by checking if the text field has descendants in
the internal accessibility tree.

For finding the next / previous text boundary, e.g. the next line start,
we need to access the inline text boxes that represent each text span.
Similarly, for exposing the text on a node on Windows IA2 and
Linux ATK, including the value of an autocomplete popup
we rely on hypertext calculation.
Before this patch and in the case of a web-based native text field, the
text in the field was retrieved from the HTML value attribute.
This could cause a mismatch, if the text that is
produced by concadenating all the inline text boxes was different
from the contents of the value attribute.

With this patch, we switch to BrowserAccessibility/AXPlatformNodeBase::GetInnerText
to compute the inner text of a subtree.
To do this, we need to correct GetInnerText to collect text from all
leaf nodes, and not only text-only objects.
For example, <button aria-label="button"></button>
will have no text-only children, but will still need to expose the
text of its label in AtkText and IAccessibleText.

This patch surfaced an issue with placeholder text in <textarea>
and <input> fields which was not previously exposed in hypertext.

Split off from:
https://chromium-review.googlesource.com/c/chromium/src/+/2110471/18

when there is a blank line at the end of a textarea.
R=dmazzoni@chromium.org, aleventhal@chromium.org

Relnotes: Screen reader correctly announces "blank", instead of a random character,
Bug: 1044410
Change-Id: I64ad2502c729076858fe52fc64d1ed1fda3cff1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2122718
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754209}