Use gpuBenchmarking in wheel event paging vertical scrolling tests
After we finish the scroll unification, the scrolls happens mainly on
the compositor thread, and the scroll code in the main thread will be
removed. eventSender sends the scroll events to main thread, so it
would not work after the scroll unification. We should replace
eventSender.continuousMouseScrollBy with gpuBenchmarking.smoothScrollBy
in the wheel event paging vertical scrolling tests. To simulate paging
scrolling, we can set scroll_by_page to be true in smoothScroll
function.


Bug: 1047176
Change-Id: Id9964bc2fcec4ce88cbe2d09fa562b4a163af7dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2111072
Commit-Queue: Lan Wei <lanwei@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754221}