Do not leave AssistantUI when pressing TAB
Previously the focus was incorrectly passed to the launcher search box
when pressing tab while the focus was on the last suggestion chip in the
Assistant UI.

This caused 2 issues:
   * The focus seemingly disappeared, forcing the user to press TAB
   again.
   * Pressing <enter> at this time would open the search box behind the
   Assistant UI, completely messing up the launcher state.

The root cause for this is that the AppListView has a special controller
to pass the focus to the search box widget. This is required because
otherwise the focus would never go in the search box widget, even when
it is what we want.
I updated this controller to only pass the focus on if the search box
widget is visible.

Bug: b/151800188
Change-Id: Iac88f8aa5baf1b65c62f06d795acfb44e3a635ff
Tests: new ash_unittests AssistantPageViewTest.FocusShouldRemainInAssistantViewWhenPressingTab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2121445
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Tao Wu <wutao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754233}