Add missing include for mojo binding.
The file mojo/public/cpp/bindings/array_traits_stl.h is using the
definition of std::unordered_set, but it doesn't include it.

array_traits_slt.h was able to compile, because it was always included
in files where <unordered_set> was included before (directly or
transitively). It is not able to self-compile. This patch fixes the
issue.

Bug: None
Change-Id: I06c028b4842ae449cd3d2bda535c7139405cc145
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124788
Auto-Submit: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754257}