Cleanup: On Mac, checks the bounds of a range before returning its value
Various cleanups and a small fix:
attributedValueForRange on Mac was not checking the bounds of the NSRange it received against
the length of the node's value or inner text.
Also took the opportunity to cleanup the use of GetValue in tests and
drop the use of "auto" where not necessary.
Split off from  https://crrev.com/+/2110471/18

R=dmazzoni@chromium.org, aleventhal@chromium.org

RelNotes: N/A
Change-Id: I3e6bed5fd1aab3ea9a66d48f012eb6b5805d5e75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2119804
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754277}