Strengthen WTF::LinkedHashSet robustness with const.
No behavior change, merely compiler-enforced modification protection.

Modifying the value in the set is dangerous, because the underlying hash
mapping isn't modified. This CL bring the behavior closer to std::set
and std::unordered_set, whose |iterator| is in fact a const iterator
under the covers.
https://cplusplus.github.io/LWG/issue103 describes how it wasn't always
the case and why it was changed.

Change-Id: Ief7116b0c6e84c458c6085494b6199fd1e67d1ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2123504
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Bartek Nowierski <bartekn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754303}