desktop-pwas: Move DCHECK for kDesktopPWAsTabStrip
This CL updates a DCHECK for kDesktopPWAsTabStrip being enabled to only
occur if we're updating an app to use the experimental tabbed mode.
Previously it was possible to hit this DCHECK when the experiment is
disabled and the user is setting the window mode to something other than
tabbed mode.

Bug: 897314
Change-Id: Ie1bc99cf9967d45c4f9f3048ff8bf88025ed34d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2126334
Auto-Submit: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Eric Willigers <ericwilligers@chromium.org>
Commit-Queue: Eric Willigers <ericwilligers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754394}