Longhand/Shorthand should use CSSProperty::Flags instead of uint16_t
We are approaching our current 16-bit limit for flags, and last time
I exceeded this limit locally (and hence redefined CSSProperty::Flags
to uint32_t), it took a while to understand why things didn't work.
The reason was that the Longhand/Shorthand constructors truncated the
integer to uint16_t.

No new flags are planned at the moment, but I still want to eliminate
this trap for when we _do_ add a new flag.

Change-Id: I4527b7ed16b9f289135a3a37665cbcf319d8bb7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124252
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754432}