[WebLayer] Move PageInfo and PageInfoUI to components.
PageInfo and PageInfoUI have already been refactored into parts
that can be shared among embedders, and chrome-specific parts.

This CL simply moves the former to components.

In addition, this moves some  vector icons to components,and creates a
a new build target in components/resources to house all resources for
components resources that are mapped to an Android resource.
Permissions code is updated to depend on this new target.

Bug: 1052375
Change-Id: Icd80244bf64b52794412b6ece47fe2d52937b3aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2091496
Reviewed-by: Peter Conn <peconn@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754435}