[iOS] Move some dispatcher setup to BrowserViewWrangler.
When the BVC owned the CommandDispatcher instance, it made sense for it
to be responsible for setting up application-layer dispatch. Now that
Browser owns the command dispatcher instances, it's possible to remove
this responsibility from the tab-scoped objects (BVC and BVCoordinator).

This CL instead has the BrowserViewWrangler set up dispatch to the
application and browsing data removing objects; sine this is where
these objects are injected into the tab-scoped object graph, this
removes two layers of object forwarding.

(In the near future, when SceneController and BVWrangler are unified and
cleaned up, another layer of object passing will disappear).

Change-Id: I37747e349aed98501d133310ae50662aab2577df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124381
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754469}