Start migrating to always defining WebView upstream.
Downstream can now set `define_upstream_webview_targets` to cause the
upstream version of the WebView targets to always be defined, even when
`public_android_sdk` is false.

To make the target continue to have the same dependencies as when it was
defined downstream, depend on the downstream code required to support
the R SDK if the build is targeting R and the SDK is not yet public.

Bug: 760187
Change-Id: Iadb62e4da95f8760e420bc264c780d27ba70d35a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2125068
Commit-Queue: Richard Coles <torne@chromium.org>
Reviewed-by: Tobias Sargeant <tobiasjs@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754477}