viz: SoftwareOutputDeviceOzoneTest: use test impl of ozone canvas.
The SoftwareOutputDeviceOzoneTest.CheckCorrectResizeBehavior should
only test SoftwareOutputDevice and it doesn't matter if ozone
platforms support that or not. Thus, add a test impl of
ozone canvas so that we don't have to check what platform the
test is run on and we don't have to expose the platform's name.

Change-Id: I3263a0267580df125c0b6ea8348cad6ef3a0461a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124374
Reviewed-by: kylechar <kylechar@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#754479}