gpu: Add a new vaapi shared image representation
This CL creates a new shared image representation as a VASurface. This
representation will be used in an Ozone context to represent a video
frame that was produced by the video acceleration api.

This representation will allow us to pass a NativePixmap to the VA-API
for the purposes of hardware accelerated video decoding.

We currently follow a VideoFrame centered design where the decoded
buffer is written onto from the hardware decoder as a VASurfaceID. The
decoded graphics buffer is wrapped in a VideoFrame, and passed to the
graphics pipeline as a Mailbox. This change eliminates the need to
convert buffers into the various required forms, and instead centralize
the lifetime as a single entity.

BUG=1058103
TEST=None

Change-Id: I482f3aa6d29a7093e8016ab4e74762e61c600865
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083662
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Commit-Queue: Paulo Warren <pwarren@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754533}