[usb] Read BOS descriptors from the hub driver on Windows
An advantage of the new Windows USB backend is that it can use the
descriptors that are cached by the USB hub driver. In this patch I split
the ReadWebUsbDescriptors() function in two so that the Binary Object
Store descriptor (which is standard) can be read from this cache and the
device only needs to be opened to read the custom WebUSB landing page
descriptor.

This will avoid opening all USB 2.1 and higher devices and instead only
open those that advertise support for a WebUSB landing page.

Bug: 637404
Change-Id: I58cedea166da3cf719c471873cb1fc4faa1549bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2125530
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754575}