Store unsynced credentials in ManagePasswordsState
If ManagePasswordsUIController gets notified of the deletion of
unsynced credentials, it will now store them in ManagePasswordsState.
A corresponding method is added to retrieve those credentials, to be
used by UI in future CLs.

However, in this CL, the notification method is still not called.

Bug: 1060132
Change-Id: I08d502af520f0432184087b70aeb81600c3dd99a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2111159
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754593}