[Extensions] Don't check for extension overrides in bookmarking a tab
Extensions used to be able to override the bookmark star view in order
to provide a custom bookmark manager (this was used for the Stars
extension). The extension has been deprecated and removed, and the API
never launched. Clean up the remnants.

No longer check for extension overrides when bookmarking a tab, and
simplify the methods to simply be BookmarkCurrentTab().

In addition to removing what was thought to be dead code, this also
fixes a bug that exists today that allowed non-Stars extensions to
implicitly override the "Bookmark this tab..." behavior, independent of
whether they were actually a bookmark manager. (This was caused because
we didn't check the command was active before using it in
browser_commands.cc.)

Bug: 894447, 1065167
Change-Id: I975cac58e72e113c351fe63735c4be5c3f08aa4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2123315
Reviewed-by: Mike Wittman <wittman@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754621}