Allow profile avatars to take default icons for empty images
Added the ability for callers of SetIdentityInfo() to specify which
vector icon to fall back to in the case that an empty |image| is
supplied.

This allows the incognito profile menus to specify the incognito
icon and color required.

This is necessary as previously the incognito image was generated
prior to having access to the Widget, resulting in incorrect colors.

Bug: 1065294
Change-Id: I4bc78e692abb7354e6cd5fcdeccd5483dc2c964a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2127410
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754642}