Rename PLM Observer for Isolated Prerender
This PLM Observer is the one that logs the PrefetchProxy UKM event
which I need to add to for IsolatedPrerender. For now the subresource
loading metrics are left alone, those are still nice to have until we
get to the NSP impl.

No functional changes, just a rename.

Bug: 1064726
Change-Id: Ic1dbbe22d8cfe5e6a343b9a5eb4be09590f54424
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2125565
Auto-Submit: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754688}