Deflake the SubframeShowsDialogWhenMainFrameNavigates test.
Currently, the test sends two beforeunload IPCs for main frame and
subframe, then waits for the subframe's beforeunload to trigger a
dialog, and then assumes that the main frame's beforeunload has
completed by that point when it checks the number of outstanding
beforeunload requests.  This need not be true -- even though the main
frame does not have a beforeunload handler, we still (currently) send
the beforeunload IPC to it, and it could still reply slower than the
subframe's IPC to put up a dialog.

This CL fixes the test to instead check the number of outstanding
beforeunload requests on the same event loop iteration, before we wait
for a dialog, since at that point we know we couldn't have processed
any beforeunload ACKs yet.  Then, after the beforeunload dialog is
shown, we only check that the subframe has an outstanding beforeunload
request, and don't check anything about the main frame beforeunload
(which may or may not have completed at that point).

Bug: 1044599
Change-Id: I3eb53b3a390a1945cc5e40ca388cb985a60b6d36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2128587
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754712}