grit: support building/installing as a normal package
This uses standard python setuptools so we can build dedicated source
and wheel distributions, and install them outside in other systems.
We want this for CrOS for example, but it'd also be helpful with some
other projects I've been experimenting with.

The grit_runner change is to support setuptools' console wrapping
logic.  It will invoke our main func, but not pass in argv, so we
need to pull it out of sys.argv ourselves.

The unit change is needed since we don't install the test modules in
the system.  The code still works locally, but no longer blows up the
use of `grit` when installed.

Test: `./setup.py sdist` works, and building a new wheel with that installs/runs fine
Change-Id: Ie929d9370af142bf29e246d543162d974159429b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2127847
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754715}