Long main thread work is not blocking Impl work being reported
During submit, if the frame is not activated yet, a copy of reporter at
Begin_Impl stage will be created and reported. This reporter is the
Begin_Impl portion of the existing reporter which is in Begin_main or
Commit stage.

So in the case of:
BI_1 BM_1 S_1 P_1 BI_2 C_1 A_1 S_2 P_2

In addition to:
BI_1 -> BM_1 -> C_1 A_1 S_2 P_2

This set will also be reported:
BI_1 -> S_1 -> P_1

Bug: chromium:1059282
Change-Id: I581088a41cd8ee1247fa18e28e4a24f5d4870103
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095350
Commit-Queue: Behdad Bakhshinategh <behdadb@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754719}