Move web test setup from ShellMainDelegate to WebTestBrowserMain.
This code is for web tests only so does not need to live inside
content shell.

We still read the kRunWebTests switch in ShellMainDelegate, which is
actually a content shell switch (it modifies content shell to inject
web test code).

This puts code to set command lines that was previously running in
every process into just the browser process, and relies on the allow-
lists of switches for each process to pass the appropriate flags.

This uncovered 2 interesting points along the way:

The kRegisterFontFiles switch is only for windows. It is used as a
global variable in the browser process to set the list of font files,
and then to open a hole in the renderer sandbox to read those files.
Then it is passed to the renderer which reads those files. I've made
this more clear with code, ifdefs and comments.

Web tests were setting --disable-gpu-rasterization in the gpu process
unconditionally, since every process sets command line flags but the
--force-gpu-rasterization is not passed to the gpu process. This doesn't
seem to have an impact on results except for on mac. By removing this
--disable-gpu-rasterization from the gpu process, when gpu raster is
being used in virtual/gpu-rasterization/ tests, the results change.
This CL rebaselines them.

R=nasko@chromium.org

Bug: 866140
Change-Id: Ia49983517ff0b9c91b1ae7549a47fd16a45d3497
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116535
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Florin Malita <fmalita@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754724}