Enable WebNavigationApiTest.Failures test.
This CL enables the WebNavigationApiTest.Failures, which was flaky on
various platforms. After tracing the source of flakiness, it boiled down
to internals of Blink and behavior difference when there is a
subresource that is still being loaded. If the JS "cancel()" method is
called and such a subresource is still in the process of loading, the
load is cancelled and the DidFailLoad callback for the document does
not get dispatched due to how internal state has changed. In the
success cases, the subresource is not in loading state, which allows
the expected dispatch of DidFailLoad callback for the document.

As such, a simple fix for the flakiness is to remove the image tag
from the document, allowing it to be deterministic in the cancellation
behavior.

Bug: 477840, 746407, 868880
Change-Id: I8a23575a9c31f0670315fd6ef55b65874e9f1700
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2128571
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754729}