[Extensions UI] Simplify accelerator priority logic
Extensions used to be able to override the bookmark star view in order
to provide a custom bookmark manager (this was used for the Stars
extension). The extension has been deprecated and removed, and the API
never launched. Clean up the remnants.

When an extension overrode the bookmark shortcuts, the accelerators were
given normal priority, whereas other extension accelerators are given
high priority (in order to override the default shortcut in Chrome).
Now, this can be simplified to have all extension shortcuts be high
priority. Replace the GetAcceleratorPriority() logic with a constant,
and update call sites appropriately.

Bug: 894447
Change-Id: I7a87f323f4c91f8d13bed60c4f845dbe7de45ecf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120291
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Mike Wittman <wittman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754756}