Fix ChromeTabbedActivity clobbering when recreating in multi-window
The current implementation is looking for a currently running tabbed
activity to handle the incoming intent. This is the correct thing to do
when handling outside intends, such as an open request from a messaging
app.  When reparenting tabs in multi-window mode, this clobbers a
ChromeTabbedActivity instance which is being recreated as a result of a
theme change.

Bug: 1055228
Change-Id: I722365c43ec97ab12120e9139eb43a781de8090a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2117062
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Brandon Wylie <wylieb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754780}