Work toward synchronizing feed store and model
- Added LoadStreamFromStoreTask to read model data from the store.
- Update LoadStoreTask to try loading from the store first, via the new
  task.
- Save operations executed to persistent storage.
- Added some conversions to textproto for tests.

My primary goal when making this change was to save operations to
persistent storage as they are executed. Pulling this thread led to
several changes.

We were saving a bit too much data in the single 'stream_data' record,
which meant we would have to re-serialize the whole model to save a
single piece of information. I went ahead and split 'stream structures'
out into another record. All stream structures for the stream now need
to be read in at load time, and they need to be sequenced appropriately.

In making that change, I wanted to verify we could actually load the
data back in. So I implemented loading a stream from storage as well.

Bug: 1044139
Change-Id: I98767281b4a5d1ae401dba5db5bff966f3473452
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2123249
Commit-Queue: Dan H <harringtond@chromium.org>
Reviewed-by: Ian Wells <iwells@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754782}