Android: Delete TabObserver.notifyContextualActionBarVisibilityChanged
ChromeActionModeCallback is a WebContents object that handles
events for action mode. This CL introduces a new class
ChromeActionModeHandler that manages the callbacks in response to
the lifecycle of the active Tab's WebContents. It is owned by
RootUiCoordinator which hooks up the handler with ToolbarManager.
It replaces TabObserver.notify..Changed with the observer list
managed by the handler.

Bug: 995903
Change-Id: Ib9f3ec4c16f71141d339ff8a807ae2167e6e8d83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1912100
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754816}