[Android] Fix crash uploading via Photo Picker.
Apparently, as seen in the crash report from the field,
the ContentResolver can be in a weird state when asking
for an AssetFileDescriptor for an image. The content
URI looks a bit weird, so hopefully this is a one-off
problem, and not something systemic for the whole image
collection.

Bug: 1063578, 656015
Change-Id: I00a46e26f634a080b4d10110ef4d545804bad1f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2127787
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754847}