PasswordManagerSyncTest: Try to fix flakiness
The Update*InBothStores tests are flaky on TSan, failing on
IsUpdatePromptShownAutomatically().

This CL tries to fix this by enabling the feature kFillOnAccountSelect,
which effectively disables autofilling on page load (which is known to
cause issues in tests, see
PasswordManagerBrowserTestWithAutofillDisabled).

In case that doesn't fix the flakiness, this CL also adds logging for
the state of the ManagePasswordsUIController, which might help diagnose
what's going on: Is it showing a Save bubble instead (indicating that
the fetch of existing credentials hasn't finished yet), or is it showing
no bubble at all (indicating that something went wrong with the form
submission), or something else.

Bug: 1058339
Change-Id: I89647573899fa6d895e3265df5892952d02fb8e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2128487
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754869}