[iOS] Add flag for new architecture in first-run sign-in.
Because first run does not have access to experimental flags on startup
this flag will not be used as a part of the migration experiment. It
is simply an additional tool to more easily test that first run is
working as we expect before migrating the code.

Refactors user sign-in constants into a global sign-in constants class
in order to hide the sign-in implementation details from outside
clients (e.g. WelcomeToChromeViewController).

Bug: 971989
Change-Id: I97316c952f4b215c33e62547112dde9f0d60efa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2121261
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754885}