Hide system row in guest mode
A user in guest mode only has access to temporary storage, the
downloaded files and other data coming from a guest user are stored
in virtual memory.

This is inconsistent with the way the "System" row is calculated, which
takes into account data stored in a persistent storage space. We
therefore hide the system row when the user is in guest mode.

browser_tests --gtest_filter="*OSSettingsDevicePageTest.StorageTest"

Test: 
Bug: 1061401
Change-Id: I5b7caafc9698f2853aacae3b043ea7ca6294387e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124276
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Jeremie Boulic <jboulic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754912}