[Media History] Only record audible videos
For Media History we have a sessions table that
records media sessions that we can resume. At
the moment we are only interested in video with
audio.

This adds a MediaAudioVideoState enum that contains
data about whether a media session is audio only
or has both audio and video. For web sessions, this
will follow the routed frame since that is where
the metadata we are using for media history is
coming from.

BUG=1041662

Change-Id: I75d93e339de8e221d8798ee2508aacda0f120b11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036772
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Commit-Queue: Becca Hughes <beccahughes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754944}