Work toward synchronizing feed store and model 2
This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/2123249

- Added LoadStreamFromStoreTask to read model data from the store.
- Update LoadStoreTask to try loading from the store first, via the new task.
- Save operations executed to persistent storage.

My primary goal when making this change was to save operations to persistent
storage as they are executed. Pulling this thread led to several changes.

We were saving a bit too much data in the single 'stream_data' record, which
meant we would have to re-serialize the whole model to save a single piece of
information. I went ahead and split 'stream structures' out into another record.
All stream structures for the stream now need to be read in at load time,
and they need to be sequenced appropriately.

In making that change, I wanted to verify we could actually load the data
back in. So I implemented loading a stream from storage as well.

Bug: 1044139
Tbr: iwells@chromium.org
Change-Id: Icf01f301488bc977f6f4ee2bf0f1c3332037176b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2129775
Reviewed-by: Dan H <harringtond@chromium.org>
Commit-Queue: Dan H <harringtond@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754975}