Handle change to pending state when resolving the ready promise
The ready promise depends on the pending state, which in turn can be
impacted by a pending style change. Checking the pending state prior
to checking the ready promise fixes the problem.

The affected test still produces an unhandled rejection error presumably
due to missing the step of marking the promise as handled when resetting
pending promises.

Bug: 1064640
Change-Id: I25191dd26206368e497638ce9f73984b2e61cad4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2118183
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#754988}