ChromeVox: Ensure locale switching works for char and word navigation.
Before this patch, moving by word or character would not trigger a
locale switch, even if the node was labeled with a language. This is
because movement by word and character goes through Output.subNode_,
which was not hooked up to locale switching logic.
This patch fixes that issue.

This patch also removes the experimental sub-node language detection
and switching logic, since it is still unreliable and not crucial
for launch.

Finally, this change removes the chromevox sub-node language switching
flag, since the logic is being removed.

Bug: 1020439
Relnotes: N/A
Change-Id: I5abfa9d5e493f82682fe18e19e87992d693f993c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2091721
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755009}