Make all LocationBarBubbleDelegateViews adjust on-screen if possible.
Thanks to corising@chromium.org we have previously made special cases
for a number of bubbles of this type to prefer offsetting rather than
mirroring, as mirroring can't fix certain cases where the anchor is
off-screen (e.g. in immersive mode, or when the browser window is
partially off-screen).

This makes the fix universal for almost all bubbles anchored to the
toolbar and location bar.

Bug: 1065327
Change-Id: I62e7b060ed20fb6bfa49ec38a2f4b2b1803bedd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2125468
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755042}