[Clank SSM]: Inject Unwinder in StackSampler.
Add Unwinder as optional param in StackSampler that's used if present.
This is necessary for follow up NativeUnwinderAndroid that doesn't have
a default constructor.

Late creation of StackSampler in StackSamplingProfiler is also removed
because this won't be possible with NativeUnwinderAndroid. Instead, StackSampler
is always created in constructor or injected.

Bug: 989102
Change-Id: Ib90416659f28d7139baed8d7adba62836337b13e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095372
Commit-Queue: Etienne Pierre-Doray <etiennep@chromium.org>
Reviewed-by: oysteine <oysteine@chromium.org>
Reviewed-by: Mike Wittman <wittman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755046}