Make assert_reports_js use assert_throws_js
This was originally left out of https://crrev.com/d255670 due to an
error when trying to use assert_throws_js. After investigation, it turns
out just to be that the function using assert_reports_js was not passing
a TypeError from the right global.

Bug: None
Change-Id: I8067949ddf6e9108bd862287cb49eef7aa788eff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2128243
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755070}