X11: Avoid printing error message when no visual exists
When running under Xvfb, no visual will be set.  We allow this case
because it is useful for testing.  However, an error message still gets
printed out that says to report the issue to a specific bug on the
tracker.  This CL avoids printing that error message.

BUG=1025266
TBR=msisov

Change-Id: I1e692225f84a87bd6d9ab218e1b04102b4971ee6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2127907
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Auto-Submit: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755134}