[ios] Modifies PolicyLoaderIOS to match MDM formatting expectations.
Previous versions of Chrome (M48 and earlier) looked for policy data
under the "ChromePolicy" key, but current MDM vendors do not support
placing policy data in a child dictionary.  This CL updates
PolicyLoaderIOS to read policy data from the top-level dictionary.  In
order to prevent existing configurations from taking effect as we enable
policy support, this CL also adds the requirement that an
"EnableTrustedTesterSupport" key be present and set to true, or else
PolicyLoaderIOS will not load any policy data, even if data is present
in the configuration dictionary.

This CL adds EG tests to verify that policy data is ignored when
the required key is not set.  It also splits the existing tests into
two files and simplifies how policy data is passed on the command line.

BUG=1065261

Change-Id: I4434b863dc9c52e3312c91365487a1fb05dc028e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2123408
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#755137}