Support enums in generated schema.org code and extractor
Generate code for types inheriting from Enumeration and use this code in
the extractor and the validator. The extractor will accept any strings
that start with "http[s]://schema.org", while the validator will ensure
the strings match one of the enum options for one of the enum types in
the property config.

Previously these were treated as a nested entity type, but we will only
ever find enum as string values.

Bug: 1065511
Change-Id: I53b7d3cf290c94f44ede7776eb2cb287c21e513c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2128527
Commit-Queue: Sam Bowen <sgbowen@google.com>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755189}