[Suggested files] Change ranking algorithm.
During test review we've found a significant issue with the current
re-ranking algorithm for suggested files, which makes it easy for the
ranking to get into a permanently bad state and not show file results.

For now, I'm changing the feature back to a simpler version of the
ranking algorithm, entirely based on group scores.

I've had to delete quite a few unit tests for this change, which I'll
re-implement in a follow-up CL. I've tested the behaviour manually and
want to submit the core algorithm on its own to unblock the test
reviewer.

Bug: 1034842
Change-Id: I0f59241af6e5977f899f2be867d7ae0e4b9606ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2128032
Commit-Queue: Tony Yeoman <tby@chromium.org>
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755235}