[Android] Unflake automatic password generation integration tests
This CL disables waiting for server predictions in tests and makes
sure to wait for the text field to be filled before comparing the
generated password to the text in the field.

A previous attempt at unflaking the entire test suite was made at
crrev.com/c/1995273, but the fix uncovered other flakes for the
manual generation tests. This is why this CL is only scoped to the
automatic generation cases, which are simpler.

Bug: 1064960
Change-Id: I3c4c60cebbeef0ce73f4e3dddbd4747b89e5f61f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132248
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755351}