X11: X11Window must use bounds in px from display when fullscreen.
When X11Window goes to fullscreen, it takes bounds of the display
and provides XServer with updated size and origin of itself.

XServer expects the bounds to be in pixels, but X11Window got the
bounds in dip.

This CL fixes the issue by taking display's size in pixels instead.
Also, a test was written to verify this functionality.

Bug: 1055057
Change-Id: If964856386e700a4a651303fe0e4ca81c769e851
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124710
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#755375}