Set PulseAudio context name based on product branding
Currently, no matter what the branding of the browser is, the name of
the PulseAudio contexts are always "Chromium" and "Chrome Input". This
is shown in e.g., GNOME Settings. Instead, use current branding to
determine the name to use: "Chromium" and "Chromium Input", or "Google
Chrome" and "Google Chrome Input".

Change-Id: Ibff97112b990cb43eb568be11310d85604642b69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2130766
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Timothy Gu <timothygu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755378}