[iOS] Update the ClearSyncedData sign-out feature to use new UI mocks.
Previously the ClearSyncedData sign-out feature in iOS mimicked the
current design in Android by adding a third option to signout for
clearing Chrome data. This caused a lot of confusion to the dialog
and was not iOS idiomatic.

The new design adds two signout options to the accounts table screen
when signed into a non-managed account and only the signout and
clear all data option when using a managed account.

Bug: 1043949
Change-Id: Idee2869fe4687da57d6388bf65f122d5b570b7ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124454
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755485}