[CrOS settings] Updated OS route path (RESET) string
Before this CL some os and browser routes had the same name. This was
a bit confusing. If a some c++ code called '/reset', it is not clear if
this url is a chrome or browser subpage. A distinction needed to be made.
This CL renames OS specific route RESET, which is shared with browser
routes to a more distinguishable name OS_RESET. Allowing this bug(990169)
to be unblocked.

Bug: 990169
Change-Id: Ia805b6889bf3b6afe0c22dfc1b11b3ea93fdb2a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2108873
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Nnamdi Theodore Johnson-kanu <tjohnsonkanu@google.com>
Cr-Commit-Position: refs/heads/master@{#755496}