Disallow QuickLook for popups (color picker, date picker, etc.)
Prior to this CL, using the QuickLook feature (hard-press on the
touchpad of a Mac) on a popup like the color picker or date picker
would cause a crash. TextInputClientMac::GetStringAtPoint()
requires a RenderFrame attached to the RenderWidget on the
renderer side to do anything, which only is true for main frame
and iframes.

Prior to [1], this would not do anything either, but after [1]
this became a crash. This CL returns the behavior to a no-op for
popups.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2006209

Bug: 1065328
Change-Id: I6a1a1cf12898c615ba756d32dd75bb5f87b4cc86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132963
Commit-Queue: Mason Freed <masonfreed@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755514}