Switch Skia Gold tests to use mixin
Switches all tests that use Skia Gold to use a mixin instead of
individually specifying arguments in each test definition. As a
prerequisite, also ensures that all test types that use Skia Gold use
the same naming for Gold-related arguments.

This is requisite work for adding a universal Gold kill switch in the
unlikely event of an outage that can't be fixed quickly, as the switch
can just be added to the mixin.

Bug: 1057848
Change-Id: Ib57480f374eab088ac3a389637bfbff1ee41e8b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2121419
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Sven Zheng <svenzheng@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755521}