[ios/infra] conform iOS test output to std json format
The iOS output sets flaky tests to 'actual' = 'FAIL', even though
it passes. This is confusing, because the Swarming task passes,
but it's reported as a fail. Aborted tests are also simply marked as
failures, but should be reported as TIMEOUT.

This conforms the output.json to std json format by properly
handling flaky scenarios via appending PASS to 'FAIL', and setting
is_flaky=True for that test, allowing the isolated merge script
to properly handle the result. Aborted tests are marked TIMEOUT as well.

Bug: 1064766
Change-Id: I95df0a64efd54923f0ed37dc67e807c08ad43919
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2121338
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Jeff Yoon <jeffyoon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755545}