Disabling accessibility event throttling and node caching on Android
This CL disables AccessibiliEvent throttling, and disables the
AccessibilityNodeInfo caching on Android.

This CL leaves all the code in place but does not leverage it so that
we know M83 will be cut without regressions. We can then revert this
change and land them in M84 or later.

Change-Id: I405e712431a4b114ab4bebe7924bb7e75d98e989
RelNotes: N/A
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132927
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755558}