[WebLayer] Create SiteSettingsFeatureList
This FeatureList contains the APP_NOTIFICATION_STATUS_MESSAGING feature,
which is used by the Site Settings and page info UI.

The new files live in a new //components/browser_ui/site_settings folder,
which will eventually contain all of the existing site_settings code.

This is part of an effort to componentize the Site Settings UI so it can
be reused by WebLayer.

Bug: 1058597
Change-Id: Ia50f920fcdc2c57719a8fe8f58d03ea63621749b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124603
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755595}