Removed isForm requirement for ACTION_IME_ENTER on Android
This CL loosens the restrictions on when the ACTION_IME_ENTER action
can be added to an AccessibilityNodeInfo object for a given node.

We now allow it to be set on any textfield, regardless of whether
or not the field is a descendant of a form.

Change-Id: Ieccc23ba0f3edc9adaaac69607a104e14164ffa8
Bug: 1057254
RelNotes: N/A
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132833
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755607}