remoting: Inject mouse events on UI thread on Mac
We recently encountered a crash on Mac when keys were injected due to
CGEventCreateKeyboardEvent asserting it was running on the main UI
thread. While the mouse events don't currently trigger an assert, it's
probably safer to move them to the UI thread as well.

Change-Id: I33e3d213135ade27c34a9dc1af81c6d708fb1a91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2133222
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Commit-Queue: Erik Jensen <rkjnsn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755663}