[cr-action-menu] Add tab key closed custom message
When cr-action-menu is closed using the keyboard, clients have no idea
if the close was due to a Tab key. Clients might use that to decide to
shift focus left or right of the menu anchor element.

Fire a custom tabkeyclosed event if the menu is closing due to Tab key
event. Include the shift key state in that event. Emit the event after
this.close() because that call can shift focus too. Add a test.

Test: interactive_ui_tests --gtest_filter="*CrElementsActionMenuTest*"
Bug: 1035691
Change-Id: I279bc9fefdbc8c5549cb4dcd92484b23515f8ceb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2126336
Reviewed-by: calamity <calamity@chromium.org>
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755728}