Code to help investigate problematic tree resets
Add a DCHECK that will be triggered only on address sanitizer builds,
but importantly, will trigger on clusterfuzz builds.

The hope is that we clusterfuzz will file bugs on us with test pages
that will reproduce the problem.

If this technique works, we can look at adding it in other helpful
places, such as for some of the DCHECKs in ax_selection.

Bug: None
Change-Id: I643ec0fb6c40f4063d437f62384211483bb50868
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2133027
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755839}