[Sync] Reupload new title field for pre-existing bookmarks.
This change initiates bookmark reuploading for pre-existing bookmarks
which are stored locally. A new field is introduced to BookmarksMetadata
to determine if local bookmarks were reuploaded.

To prevent the case when the bookmark specifics is committed without
favicon the entity is marked as unsynced after it was committed. In this
case the entity will be committed again when the favicon is loaded.

Bug: 1061411
Change-Id: I4a26a2babceb99a0f6532d2357884c75aff9fc05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2127032
Commit-Queue: Rushan Suleymanov <rushans@google.com>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755858}