Add UKM for IsolatedPrerender via PrefetchProxy
Most of the metrics are coming from the TabHelper's Metrics struct,
but the prefetch usage enum comes from the interceptor, by way of the
TabHelper.

prefetch_usage is read on commit since that value is set between
navigation start and commit and does not change afterwards.

prefetch_metrics is trickier. It may update numerous times while a user
is sitting on a Google SRP page and won't stop getting written to until
the following commit (or tab destruction). Since PLM's OnComplete is
called after the following commit (not helpful) or at tab destruction
(helpful), the PLM observer can't read prefetch_metrics during
OnComplete because the metrics would be wrong whenever another
navigation happened in the same tab.

The implemented solution to this timing mess is to make the prefetch
metrics struct into a RefCounted class. Then the PLM Observer will get
an additional ref (only by being a friend) and hang on to it until all
needed metrics have been recorded, which happens after the tab helper
has moved on to the next page.

Triggering the PLM Observer to get a ref to the metrics class is done
with a PLM Event. This is done for two reasons, 1) because we can't
rely on the order in which |DidFinishNavigation| is called for WCO's and
2) because using a PostTask off of OnCommit was still too early in my
testing.
Furthermore, using a PLM event allows the observer to only take the
metrics ref when it will actually be used. This is done by only sending
the event when an eligible navigation prediction is made and is being
acted on.

Bug: 1064726
Change-Id: I5de582650e2d29f7d62f2871cd1b7db5a72144fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2129231
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755891}