Update date picker ARIA attributes to follow selection instead of highlight
The pre-FormControlsRefresh date picker considered the cell with
'highlight' to be the active cell with respect to ARIA, and it was the
highlighted cell that moved around by arrow-key navigation.  In the
refreshed design, highlight is no longer managed in JS, and the cell
marked as 'selected' in JS is the one that is moved around by arrow-key
navigation and is therefore the one that should be considered focused
by accessibility.  However the management of aria-activedescendant and
aria-selected was never migrated from highlight to selection. This
change does this migration.

For test coverage, the change re-enables
DumpAccessibilityTreeTest.AccessibilityInputDateWithPopupOpen and
modifies it to dump the full popup tree instead of stopping at the
'Today' button.  The column header values are skipped so that the
/blink subtest doesn't produce different results on different
platforms (Windows uses Mo/Tu/We etc instead of M/T/W/etc for the
day  abbreviation localizations).

date/week/month-picker-ax.html are updated to account for the extra
MarkDirty events that are fired to to the fact that focus/selection
is now being managed again for the calendar cells.  month-picker-ax.html
is moved to promise_test to match the format of the other two tests and
changed to assert instead of timing out on failure.

Bug: 1022920,1054901
Change-Id: Iaa98605582a12fcf44d87748ca3bc572518528fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2111750
Commit-Queue: Dan Clark <daniec@microsoft.com>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Ionel Popescu <iopopesc@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#755938}