Fixing shadow and filter issue with scaling
https://html.spec.whatwg.org/multipage/canvas.html#dom-context-2d-shadowoffsetx
The standard in the above link says that shadow offset should not be
modified by the transform matrix.

Adding a test to verify that, and fixing it so the offset does not
get modified by the transform matrix.

I tried to add a html reference test, simulating a shadow but there is
some slightly aliasing errors due to the change in the way the
compositor handles the shadow and the blur.
Adding a png reference test to validate that it will still work in the
future.

I'm also removing an unnecessary c-style cast in:
  composite_flags.setBlendMode((SkBlendMode)GetState().GlobalComposite());

Bug: 1044803
Change-Id: Ib98a35d9b9fdcb71c0cccca114099e985b569fb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2128766
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#755980}