Add new util in TouchCommon for drag
This patch adds a new method for generating a complete drag event in
TouchCommon: "performDrag". The difference between this method and the
existing ones is that it is a single call for the down/move/up stream
and most importantly uses deterministic timing. The existing methods
are inherently flaky because they use the system time when generating
sequential events. This will confuse gesture detectors and cause them
to fire flings or empty events when they otherwise wouldn't. The new
method specifies a duration that can be used to simulate a fast or
slow drag.

Bug: 813584, 882003, 947694
Change-Id: I42ab8e8803c8b4e3c81a2876240c365ffc39f4f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134200
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756000}