[Extensions] Split up the Service Worker-based webrequest tests.
One of the tests in this suite is flaky, so this CL splits them
up into two different C++ test fixtures, in case the flaky one
needs to be disabled. It's currently flaking at about 0.1% in
the worst case, so hopefully, this won't affect too many builds.

Bug: 1056853
Change-Id: Id7bb0342cbb0b67b61f3bea01fc38601f80ec3b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2128466
Reviewed-by: Archana Simha <archanasimha@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756030}