Prep to disable GLInterface on shared main thread context
To support RasterImplementation on the shared main thread context we need
to remove all references to ContextGL on that context. This CL prepares
to fully disable ContextGL() on the shared context. A follow up CL
will change |support_gles2_interface| = false which will actually
disable it. New requisite API's should be added to RasterInterface
instead of GLES2Interface. During the transition to OOPR canvas we will
support RasterImplementationGLES but that class will be deprecated once
we fully transition to RasterImplementation at which point ContextGL and
GRContext will no longer be accessible on the shared main thread context.

Bug: 1018898
Change-Id: I85b2bc50c937d715a248a0c451278624f8d3e7b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116782
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Nathan Zabriskie <nazabris@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#756048}