Wire up toggle for (dis)allowing printer access from Plugin VM
This CL wires up the toggle in settings for controlling printer access
from Plugin VM to a new pref plugin_vm.printers_allowed. When this or the
PrintingEnabled policy is set to false, the CUPS proxy will return a 401
forbidden error to every request.

Bug: 950431
Change-Id: I14c395b53ea11bcb9c438266e36aa21e6fcd85b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120110
Commit-Queue: Timothy Loh <timloh@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756051}