Don't crash when animating with CSSCascade disabled
Running without the CSSCascade flag doesn't work anymore, since we
hit a DCHECK. We're not supposed to calculate the animation update more
than once, and with CSSCascade disabled it can potentially already have
happened once we reach ApplyAnimatedStandardProperties. This CL re-
introduces an if-check that was removed in [1] to fix the issue.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2064248

Change-Id: I340f31543f63ed05a6a033ece0779ffaa8d57e28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132400
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756062}