[CrOS settings] Updated OS route path (LANGUAGES) string
Before this CL some os and browser routes had the same name. This was
a bit confusing. If some c++ code called the route '/languages', it is
not clear if this url is a chrome or browser subpage. A distinction
needed to be made. This CL renames OS specific route LANGUAGES, which
is shared with browser routes to a more distinguishable name
OS_LANGUAGES. Allowing this bug(990169) to be unblocked.

Bug: 990169
Change-Id: Ic2d6dc3120d505ef2a864023c6d89a57301acdd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2123311
Commit-Queue: Nnamdi Theodore Johnson-kanu <tjohnsonkanu@google.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756092}