Add OTRProfileID to profile class.
For proper support of multiple off-the-record profiles for one regular
profile, an OTRProfileID is added to profile class.

This CL is based on the following design doc:
https://docs.google.com/document/d/1xj21DPfwo1gZ8kagj6i1lSmKwG1dW-4KPt3yvVcsvOg

This is the first step of implementation of this design doc, only adding
OTRProfileID to C++ classes and adding/updating some tests.

Bug: 1033903
Change-Id: Iaed207994c01372f3349ecce1333ee55fea44d2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080253
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Brandon Tolsch <btolsch@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756161}