gl: x11: Extract x11 bits guarded by USE_X11.
This is a clean up of ui/gl, which goal is to get
rid of USE_X11 guards and place the code to corresponding
x11 source/header files so that ozone/x11, which is not
guarded by USE_X11, can utilize the same classes.

In this CL, I moved all X11 code to gl_surface_gl_x11,
and also created a utility that each platform can
implement to provide custom values for display attributes,
alpha and buffer size.

Also, I had to add a singleton with an optional interface that is
set in ui/gl/init. The reason why I decided to do it like that
is that we need X11 to set some x11 specific egl attributes.
While it could have been not a problem to have just utility methods
inside ui/gl for x11 and stub for others, it's a problem for ozone
as we can't depend on ui/ozone from ui/gl. Thus, we have a singleton
interface in ui/gl, which is then set in ui/gl/init right before
display initialization.

R=backer@chromium.org

Bug: 1036285
Change-Id: I3117e65b8ba240ad16e87ec6c1489ccd6d71934f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116443
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Jonathan Backer <backer@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#756165}