ozone/wayland: allow to run Wayland without mojo.
Even though we set require_mojo to true in the PlatformProperties,
in certain scenarios Ozone might be initialized without mojo, but in
a single process mode (for example, unittests, ozone_demo).

In this case, we should instantiate a direct connection between
buffer managers without mojo. I could use base::TaskEnvironment, but
it will require too much changes to how different tests are initialized.

For example, binding an interface and sending a mojo call requires a
a task runner handle (installed in WaylandBufferManagerGpu). And
that fails with compositor_unittests - some tests create
TaskEnvironment, and some don't.

Initializing the TaskEnvironment in CompositorTestSuite won't help,
because (as I said) some tests do that. This results in crashing when
tls delegate is set.

Bug: 1067156
Change-Id: If3236f9604c692b4ed6efc5dbaf9cc67ce36d62e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2128111
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#756173}