[WebLayer] Add test of link click to page that launches intent
This CL adds another test of WebLayer's external intent launching:

When the user clicks a link to navigate to a page that then launches an
intent without a user gesture (e.g., via onLoad()), the intent is
launched. This case differs from that of the user navigating to said
page directly via typing; once Weblayer incorporates //chrome's
TabRedirectHandler the latter will be blocked by policy, but the former
(the behavior being tested in this CL) is supported by Chrome and should
be supported by WebLayer as well.

Bug: 1031465
Change-Id: I58e1fc7bbde9fc6235606523f06797a7877bb108
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132385
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Mugdha Lakhani <nator@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756183}