[Autofill Assistant] Share backend url with extension api code.
Before this patch, the backend url logic was in the android client and
not directly available to the extension code. Since switching backends
is useful during development and testing scenarios, this patch moves the
command line switches to a shared file. The default URL can be in the
service impl and does not need to be part of the client interface. The
patch also cleans up the use of the api keys, which don't need to be
handled in chrome, only the installation channel needs to be part of the
Client interface.

Bug: b/143736397
Change-Id: I30caeffc226e6936d43e0000d897f9bfbf53d9c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2129549
Commit-Queue: Mathias Carlen <mcarlen@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/master@{#756192}