blink/loader: Discard the encoded fonts after decoding.
FontResource holds onto the encoded and decoded versions of the
font. However the encoded version is never accessed after it has been
decoded, and the decoded version is never discarded. This means that we
are holding on the SharedBuffer without a need for it.

Fonts are almost always compressed, with a ~3x compression ratio
typical. From local benchmarking, encoded fonts are commonly ~500kiB on
popular news websites.

This CL discards the fonts after they've been decoded.

Change-Id: Ic72009431084cdd1e30fcd69ba1ec59788c0cede
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134291
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756195}