[Web Payments] Add a PaymentAppServiceBridge JNI interface.
The bridge implements PaymentAppFactoryInterface on the Java side and
acts as a PaymentAppFactory::Delegate on the C++ side to allow Java to
invoke C++ PaymentAppService.

This patch sets up the basic scaffolding. A subsequent CL will fill in
the actual onPaymentAppCreated() callback, which is the most substantial
part of the bridge.

Bug: 1063118
Change-Id: Ie3b93b719712b9b95865ee769415c4cd91d35392
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2133120
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Danyao Wang <danyao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756257}