[WebLayer] Add animation to show and hide security icon.
When there's no security status icon to show, currently WebLayer
keeps the blank space for it. A more elegant UI is to start the URL bar
text at the leftmost edge in this case, without leaving any gap for the icon,
and animate to the right once the icon is ready to be shown.

This CL reuses code from Clank for the security button animation.

GIFs before and after the change can be seen here:
https://drive.google.com/drive/folders/1qyRRQzysAkFdKgwiKBk_aEqrvCB48M8i?usp=sharing

Bug: 1025607
Change-Id: I831ae02f1fdf096170a8296b3d8b8916c505aa70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132333
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Auto-Submit: Mugdha Lakhani <nator@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756270}