Desks: Fix a11y announcement of desk in overview
DeskMiniView was changed from a Button to a generic View, but
the actual button now is DeskPreviewView. The a11y node data should
be initialized first by the button, before we add the extra attributes.

Also announce the desk name if the miniview is selected.

BUG=1067392
TEST=Manually

Change-Id: I1646e0aeff95f089a9f2252532b09f2c7acdcb3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134920
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756312}