[Offline indicator v2] Fix hiding animation for Android L
An integration test was failing on L bots because the hiding animation
in the compositor wouldn't complete and the indicator would get stuck
shown. Upon investigating, I discovered that this issue was not limited
to tests and was reproducible on an emulator.

The issue was that #onLayoutChanged would be called during the hide
animation causing a wrong height value to be sent to the compositor as
if it's being shown.

A simple fix is to add a boolean to ignore #onLayoutChange calls when
we no longer need them, and we're basically done once we get a height to
run the show animation.

Bug: 1059297
Change-Id: I943300ebc1e93172eec9311ec28c44172464ba4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2130781
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756352}