[SB Delayed Warnings] Handle mouse clicks
This CL introduces the ability to observe mouse clicks when a page with
a delayed SafeBrowsing warning is loaded. If the feature parameter
"mouse" is enabled, the warning is immediately shown on mouse click.
Otherwise, a metric is recorded the first time user interacts with the
page.

Bug: 1057157
Change-Id: If81e9c5867b7e72703a284b8226e2987f5edf3d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120428
Auto-Submit: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756418}