Add a mechanism to enable defining consoles in distributed fashion.
Consoles defined using ci.console_view instead of luci.console_view can
specify a dictionary that controls how entries in that console are
sorted. Calls to ci.builder can then specify the console_view_entry
attribute with the category and/or short name for the builder and an
entry for the builder will be added to the console corresponding to the builder's mastername.

Bug: 1067057
Change-Id: Icbbb310bbf4434aeb2434ab930e8fb64fff85a52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2136107
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756440}