Make test independent of wizard controller state.
The test was based on the assumption that the wizard controller is
recreated in the login display host and hence, had the expectations
according to that. If instead the wizard controller object is reused,
this test fails the expectations set in the test class at the start.
Hence, calling AdvanceToScreen(...) directly instead of going through
login display host.

Bug: 1064271
Change-Id: I12b11f0a08269d91d776ae0e438b7b0b8ce7262d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132382
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Cr-Commit-Position: refs/heads/master@{#756444}