[Tab Groups] Streamline tab context menu patterns.
Updating the context menu for Tab Groups to show a submenu with the
entry "Add to Group" if another group exists and the submenu to had
"New group" as the first item followed by a separator and then the
other existing groups.
If there are no other groups, a submenu will not appear and the menu
item will simply be "Add to new group".

Bug: 1064674
Change-Id: I88b470171d4f44e6969201fea7e991d5cd2905c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2133269
Commit-Queue: Charlene Yan <cyan@chromium.org>
Reviewed-by: Connie Wan <connily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756446}