WebXR AR: disable shared buffer for frame transfer
This is a workaround for an issue where the WebXR framebuffer content is
displayed out of order, leading to jumpy animations and other glitches.

Until the root cause is identified, disable shared buffer frame
transfer, and instead use the Mailbox+Surface path. This was originally
intended for use on Android N which doesn't support AHardwareBuffer,
and is less efficient due to needing an extra pixel copy, but the APIs
are still supported on current Android versions.

Bug: 1067795
Change-Id: I256f878836eeac201934392ec6ce79f9f2fd7c53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2135957
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Commit-Queue: Klaus Weidner <klausw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756492}