Use ChangeMouseLock function from CrossProcessFrameConnector
The FrameConnectorDelegate is supposed to be an interface which
is implemented by an object supplying platform semantics for a child
frame. The recent change(https://crrev.com/c/2071788) failed to
implement the ChangeMouseLock in CrossProcessFrameConnector and
therefore these calls were not getting to the correct place.

I also added the error handler for the rarely hit
PointerLockResult::kUserRejected error.

Bug: 1062702
Change-Id: I9c6aad9bac455b578b39e2aa086732f69f09fe61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116327
Commit-Queue: James Hollyer <jameshollyer@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756517}