Extract out HID test for Chromebook.
The test was based on the assumption that the wizard controller is
recreated in the login display host and hence, had the expectations
according to that. If instead the wizard controller object is reused,
this test fails the expectations.
We were set the device type in the class constructor and invoking the
login wizard in SetUpOnMainThread(). Due to this, when the device type
is changed to Chromebook in the test, the test assumed that invoking
login wizard would recreate wizard controller with no HID screen.
Hence, extracting out the test to a separate class and setting the device
type to Chromebook in it's constructor.

Bug: 1064271
Change-Id: I23e57dcedf3e1ea6d8e37bf8064784d8e27a5d9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2135615
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756541}