[CrOS settings] Updated OS route path (PEOPLE) string
Before this CL some os and browser routes had the same name. This was
a bit confusing. If a some c++ code called '/people', it is not clear if
this url is a chrome or browser subpage. A distinction needed to be made.
This CL renames OS specific route PEOPLE, which is shared with browser
routes to a more distinguishable name OS_PEOPLE. Allowing this
bug(990169) to be unblocked.

Bug: 990169
Change-Id: If1528c70ed31429711044985e9589a6025568dfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2117936
Commit-Queue: Nnamdi Theodore Johnson-kanu <tjohnsonkanu@google.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756571}