Move NotifyUISignoutWillDeleteCredentials out of signin notifier
This CL reverts the changes to PasswordStoreSigninNotifier and moves the
API to new a dedicated class, owned by the PasswordStore. The class is
lazily initialized via a setter method (as it is done for the notifier),
and only for the account store, since that's the only place where it
should be exercised.

Bug: 1060132
Change-Id: I6d78fc5e2c061acb474fec4cf7be7418368fd8a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2135750
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756607}