[iOS][Credential-Provider] Credential from Form
Create credential provider extension objects from password forms.

|recordIdentifier| is built in a similar way to how the logins DB
identifies credentials. This way Chrome can remove and update entries in
the credential provider store. This identifier is also used to identify
indexed credentials in the system for password retrieval.

The keychain identifier will come from the login database in a follow
up CL. This is the attribute used to retrieve the unencrypted password
from the keychain.

Favicon and validationID are not related to the password store, and
will also be implemented in a follow up CL.

Bug: 1066803
Change-Id: I20d573301cc468e9984f9524239fd47a7826aa85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2133982
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756645}