Take only account store conflicts into account for generation in b4p
MultiStorePasswordSaveManager now passes only matches in the account
store to the PasswordGenerationManager API, causing matches in the
profile store not to be taken into account, as specified in the b4p
requirements.

A pre-existing unit test checked that only the account form saver was
exercised when triggering generation with the account store enabled.
That test is extended to verify this is still true in the multiple
possible conflict cases that can exist in each store.

Bug: 1060131
Change-Id: I19092cf11603c1308172362b181dc505d91986cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132456
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756658}