Update MostVisitedListCoordinator to use incognito profile in incognito.
This CL update the profile passing to OfflinePageBridge and
SuggestionsDependencyFactory from MostVisitedListCoordinator.
Currently, it is always regular profile even when called in incognito
mode.

This CL uses tab model selector of the chrome activity to
determine whether it is regular mode or incognito, and passes the proper
profile (regular or off-the-record).

Bug: 1041781
Change-Id: If3928158e93961d72e06dc5711653bbb7cb1c6b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134230
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Commit-Queue: Side YILMAZ <sideyilmaz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756672}