Get the correct app_id from tab for the guest user.
For the guest user, GetExtensionForTab returns null, so AppService
Instance can't get the correct app_id for OS setting, and no instance
record for OS setting when OS setting is launched. So the close menu
items is not added when OS setting is launched.

Modify GetAppIdForTab to get the app_id from browser directly, to return
the OS setting app_id from tab, so that we can have the correct instance
and add the close menu item for the guest user.

BUG=1067198

Change-Id: I759228624fac3c69166a0bb2d1c9a199afe323dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2136007
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756692}