gpu: Add a new vaapi shared image representation
This CL creates a new shared image representation as a VASurface. This
representation will be used in an Ozone context to represent a video
frame that was produced by the video acceleration api.

This representation will allow us to pass a NativePixmap to the VA-API
for the purposes of hardware accelerated video decoding.

We currently follow a VideoFrame centered design where the decoded
buffer is written onto from the hardware decoder as a VASurfaceID. The
decoded graphics buffer is wrapped in a VideoFrame, and passed to the
graphics pipeline as a Mailbox. This change eliminates the need to
convert buffers into the various required forms, and instead centralize
the lifetime as a single entity.

This CL also creates two interfaces: VaapiDependencies and
VaapiDependenciesFactory.

This interface exists to allow the SharedImage system to have access to
the VA-API without directly including any VA-API headers.
VaapiDependencies and VaapiDependenciesFactory will be implemented in
//media/gpu/vaapi and will call the VA-API. The implementation of
VaapiDependencies is responsible for creating a VASurface that can be
accessed in the SharedImageSystem via SharedImageRepresentationVaapi in
a follow up CL.

BUG=1058103,1066160
TEST=None

Change-Id: I00bd3d89c7a5bdaab5544c4816baac9a4216428d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132528
Commit-Queue: Paulo Warren <pwarren@chromium.org>
Auto-Submit: Paulo Warren <pwarren@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756693}