Minor clean up for .pkg installation
These are several cleanups inspired by a reading of the excellent book
"Packaging for Apple Administrators".

https://books.apple.com/us/book/id1173928620

1. The package now requires that no copies of the app be running while
being installed. This is more of a completeness thing than any real
protection.

2. This changes the source comments from "product archive" to
"distribution package" which is a broader term, and more technically
correct, as we're not actually trying to make a product archive (which
is a specific thing).

3. This adds an identifier/version pair to the distribution package
in addition to having it on the component package. This technically
makes the distribution package a product archive :) but may be useful
in some situations such as NetInstalls and system install setups (see
https://managingosx.wordpress.com/2017/11/17/customized-high-sierra-install-issues-and-workarounds/
for more information).

Bug: none
Change-Id: Iefde2aae6a3b22721fc16876e6dbce505c3c9601
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134405
Commit-Queue: Avi Drissman <avi@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756697}