Make IDB RenderThreadShutdownTest non-single-process.
The IndexedDBBrowserTestSingleProcess.RenderThreadShutdownTest
was already disabled on Android and ASAN. The test has flakes
on MacOS, Linux and TSAN as well. Rather than just disabling,
convert it to being a non-single-process test as that's not
a realistic configuration. The test is kept around although
likely the render process is simply terminated so it's not
testing much.

Bug: 368901, 788788, 1065523
Change-Id: If31017da91f50f91f1ff677fcdd78d08d075a3eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134930
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Joshua Bell <jsbell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756727}