[css-grid] Allow huge grid to be inspected without freezing the page
Since r629436, trying to inspect a 1000x1000 grid in the devtools would
freeze the page. That's because, when determining the areas to be
highlighted, the code would try to encode each cell independently.
But there can be way too many cells.

This patch changes it to encode the rows and the columns, making the
algorithm linear instead of quadratic.

The `highlights` object returned by `Overlay.getHighlightObjectForTest`
now no longer has a `cells` property, and instead has separate `rows`
and `columns` properties.

I have ensured that with this patch it will be possible to inspect
1000x1000 grid (the current limit) reasonably fast. If I raise the limit
to 10000x10000, then it still works well in release builds, but in debug
ones it takes almost 2 minutes.

BUG=1067748

TEST=http/tests/devtools/elements/highlight/highlight-css-grid.js
TEST=http/tests/devtools/elements/highlight/highlight-css-grid-huge.js

Change-Id: Ic285a07550b4799d3c22026bdcd387b207f23512
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2135618
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/master@{#756735}