TestVDAVD: Use GetRectSizeFromOrigin() for ALLOCATE mode.
CL:2134417 and predecessors fixed ProvidePictureBuffersWithVisibleRect()
so that GetRectSizeFromOrigin() is used to calculate the texture
dimensions and the coded size of a VideoFrame. However, I forgot to
update the test implementation of VideoDecoder (TestVDAVideoDecoder).
This test does just that to be consistent with CL: 2134417.

Bug: b:153085451
Test: ran video.DecodeAccel.* on kevin, kukui, and veyron_minnie.
Change-Id: I647bed4271da186e0658fb5ba03e5c991f27d6d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2136846
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756812}