Retain Owner from HeapMojoReceiver::Wrapper
There is a chance that the HeapMojoReceiver::Wrapper may outlive the Owner in case of a ConservativeGC.
The stack scanning may create a root to HeapMojoReceiver::Wrapper, prolonging its life.
To make sure they die together in the same GC cycle, we will make the Wrapper retain the Owner.
So in case HeapMojoReceiver::Wrapper is kept alive, the Owner is guaranteed to be alive too.

Bug: 1067476
Change-Id: I6456c368dc2ad446a627d1b75102bd42a5770c0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134093
Commit-Queue: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756856}