vulkan: fix gpu hang or crash during destroy vulkan surface.
The hang is because fence helper is used for destroying vulkan
surface asynchronously. However the platform window (X11, android
or Win32 window) could be destroyed before destroying the vulkan
surface which is created for the native window. It causes hang or
crash. Fix the problem by calling vkWaitQueueIdle() and destroying
the vulkan surface synchronously.

Bug: 1062727,1062703
Change-Id: I73715fe0548221dc33cd5ffb2c266d53795dbb13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2138253
Commit-Queue: Peng Huang <penghuang@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756865}