Simplify thumb drag.
This CL simplifies the implementation of thumb drag for the Compositor
threaded scrollbar scrolling feature. It also aligns the implementation
to main thread thumb drag more than before. The aim of doing this is to
reduce the possibility of bugs arising from the disparity between main
and cc manipulated scrollbars.

Bug: 1067518
Change-Id: I501c68e9993afcb4b1e097d0aafce7670d8bec4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2136412
Commit-Queue: Rahul Arakeri <arakeri@microsoft.com>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756892}