[Extensions] Remove Command::Type
Command::Type specified a type for a command of page action, browser
action, or "named" (i.e., any other command). This was posing a bit of a
problem for the generic "action" key, because we'd need to add a new
entry for it.

As it turns out, Command::Type was never really used. It was only ever
passed as a hint of which command to look up for a helper function, and
we can use ActionInfo::Type directly there.

Remove Command::Type completely.

Bug: 893373
Change-Id: Ie26b7ffbc50c541700947988bcef94dc5f4083a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132809
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756895}