[selectionbar] Change 'cancel selection' to icon
Removes the 'Cancel selection' text from the selection bar cancel
button.

Changes the icon for selection bar to match the file list selection
icon.

Positions the icon and number of files selected message to align with
the icons and file names in the file list.

Removes the resize listener on the toolbar controller which was used
in the old layout to size the 'Cancel selection' button to match the
position of the splitter bar between the tree and file list views. The
listener function itself remains, with a comment added so that it can
be removed after files-ng work is complete.

Bug: 1068079, 1065837
Change-Id: I857f037995a5bc0ba631fa41f62af8d72ff99dd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2137185
Commit-Queue: Noel Gordon <noel@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756933}