Remove NetworkContext::SetFailingHttpTransactionForTesting method
FailingHttpNetworkTransaction was test-only code but called by
NetworkContext::SetFailingHttpTransactionForTesting, so it was linked
in the production code.

Since it was only called by a test to force a network error, this patch
removed SetFailingHttpTransactionForTesting method and used an
alternative way to do that.

Bug: 1012116
Change-Id: I918e04a3c143b80292f66d3c0e47651b84d6580e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2129791
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Commit-Queue: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756939}