[FragmentItem] Make delta_to_next_for_same_layout_object_ mutable
This member needs to be set in |LayoutBox::AddLayoutResult|,
after |NGFragmentItems| was built and set to the cached list.

Since this member is used as a cached index, allow changing
by marking it as |mutable|.

This change is needed to properly mark |NGFragmentItem| as
|const|.

This patch has no behavior changes.

Bug: 982194
Change-Id: I83131b744cd73994b861be9eb374e5e5e47fd5c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2138998
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#756990}