[fuchsia] Fix WebContentRunner to handle Context failures.
WebContentRunner used to exit() if its Context terminated unexpectedly.
This was removed when support was added for a single Runner service to
run WebComponents in their own isolated Contexts. Launch of the Context
requires move()ing the contents of the CreateContextParams, resulting in
some of its fields (in particular the service-directory) still being
set, but to nulled values.

Failure of the default Context would therefore leave the
WebContentRunner active but unable to launch any new non-isolated
WebComponents. This affected both the Cast and web Runners.

WebContentRunner & CastRunner are fixed to consume a callback which they
can Run() to get a fresh fuchsia.web.CreateContextParams table from the
caller.

Bug: 1066826, b/152696080
Change-Id: I3610765751c402cd0c3d2acca13c3f0e164f85d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134231
Commit-Queue: Wez <wez@chromium.org>
Reviewed-by: David Dorwin <ddorwin@chromium.org>
Auto-Submit: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757011}