[Android] Prepare TabRedirectHandler.java for componentization
We had initially thought that it would make sense to fork //chrome's
TabRedirectHandler.java for usage in //weblayer. However, on
investigation its //chrome dependencies are essentially layered on top
of its core logic, the latter of which can be shared directly with
WebLayer. This CL thus prepares TabRedirectHandler for componentization
by moving its association with Tab to a RedirectHandlerTabHelper thin
wrapper object. Future CLs will componentize TabRedirectHandler as
RedirectHandlerImpl and then merge it back into the RedirectHandler
interface (eliminating the latter as an interface).

A side benefit of this change is that it removes the remaining
//chrome dependencies from ExternalNavigationHandlerTest.java, as this
test now explicitly passes in the values for the flows that it wants
to test to TabRedirectHandler#updateIntent() rather than needing to
configure the latter's behavior via mutating //chrome-level state.

Finally, as part of this change various CustomTab-related unittests now
need to mock out an access to ChromeFeatureList. This is because with
this refactoring that access now happens before an early return in
TabRedirectHandler#updateIntent(), whereas before the access was after
that early return.

Bug: 1031465
Change-Id: Idabfe225fb77cf3bfe454d7ff1d63c9e801b8fcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2127090
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757012}