[PlayBilling] Support app-store billing in AndroidPaymentAppFinder
After:
When a merchant page runs in a Trusted Web Activity that's installed
from an app store (e.g., Google Play) and the PaymentRequest is not
requesting shipping or payer contact, if it satisfies the following
conditions, the TWA itself would be included:
- the PaymentRequest supports the (TWA installer) app store's
billing method in the payment request.
- the TWA can handle pay intents.
- the TWA can handle the app store billing method.

Before:
AndroidPaymentAppFinder could not include a TWA for an app store
billing method.

Change:
* In AndroidPaymentAppFinder#findAndroidPaymentApps, check if the
merchant page is a TWA installed from app store and if the
PaymentRequest is requesting shipping or payer contact. If it is,
add the TWA itself as a payment app if it's eligible.
* remove mIgnoredMethods.

Note:
* Counterintuitively, the merchant page would send the pay intent to
the TWA instead of the Play Store, because it would be the TWA who
is responsible to interact with the app stores.

Bug: 1064740

Change-Id: I0a2f6baebae422aaeab574e3c39a10cd61fafb4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2135152
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757128}