[WebLayer] Update UrlBarView when active tab changes.
Before this change, when the active tab was changed but
it's webcontents not updated (we switched to a quiescent tab),
the UrlBar wasn't getting updated.

After this change, it is always updated when active tab changes.
Tested manually, and added a browsertest to verify behavior.

Bug: 1025607

Change-Id: I085b4748f25d3f2f483a5d9964bb767ed178f1de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2139707
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757134}