Do not prefer raster in sRGB in WebView
Previously, the raster color space was set to prefer raster in sRGB
for Android. However, this is not desirable for WebView, which should
prefer raster in whatever color space the app is using. That was
already being done and was overwritten by the preference of sRGB.

This CL prevents preferring raster in sRGB when using WebView.

Bug: 955158
Change-Id: I316e646bc7ebac309fa609c2c30a5efa8e6c3650
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116560
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Chris Blume <cblume@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757136}