[LayoutNG] Fieldset border block-start
This change ensures that the intrinsic_block_size_ for fieldsets
includes the amount of the border block-start that has not already
been consumed in previous fragments, rather than simply setting
this to 0 when the fieldset breaks.

This change also alters the design for large legends/small borders.
Instead of adjusting the fieldset border block-start offset when
the legend doesn't break, this change updates the design so that
we adjust the offset when the border block-start doesn't break.

Doing so ensures that we know the size of the border consumed in
previous fragments without having to store its offset for later
fragments to reference.

Bug: 875235
Change-Id: If29c0166453c95c43d9cc08197095b2221e1b309
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2137525
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#757154}