Instantiate PolicyValueStoreTest/ValueStoreTest.
This wasn't instantiated because the source is in
value_store_unittest.cc, which is a bunch of TEST_P()s that are
instantiated in several other files in different test suites.  However
the file was in .gn as part of just the extensions tests, instead of
being part of test_support, so it wasn't linked into unit_tests, so the
attempted instantiation for PolicyValueStoreTest didn't work.

Bug: none
Change-Id: I5089eeb39a92b1f5e72c5a27c7ad3a2b283bf60b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2139080
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757173}