Remove the test_runner namespace.
The files are part of //content/shell (for web tests) and are moved into
content:: with the rest of the web test code.

In the process, we have to rename some method collisions. The
GetWebWidgetTestProxy() method in web_test_support_renderer.h can
be removed, and WebFrameTestProxy can return the WebWidgetTestProxy,
which matches the new world order of ownership relationships well. Yay.

Relatedly, we add a WebFrameTestProxy::Reset() so code doesn't need to
get the widget when it is dealing with frames. And we move some frame
reset code there from other places.

R=avi@chromium.org

Bug: 866140
Change-Id: I8c31983aa21d86c9cffcd876dff620180d9176f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2139358
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757235}