Hook up IsolationInfo.
Replace NetworkIsolationKey in URLLoaderFactoryParams with
IsolationInfo. IsolationInfo includes top frame origin,
frame origin, site for cookies, and a rule on how to update them on
redirects, and can be used to obtain the NetworkIsolationKey.

This is part of an effort to make SiteForCookies consistent with
NetworkIsolationKey and move SiteForCookies logic from the renderer
to the browser process to better protect against compromised
renderers.

Currently the SiteForCookies field of the IsolationInfo is not used,
as it still needs to be plumbed through ResourceRequest::TrustedParam,
and there are some cases where its SiteForCookies field is not
consistent with the value received from the renderer, which need to
fixed.

Bug: 1065152
Change-Id: I759124cc946b3643d9ffe5e1c67fdbc393bc3251
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2113153
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757242}