[Native Cast MRP] Launch a new session after the existing one is removed
When we launch a session on a sink with an existing session, we remove
the existing one. We make sure that we request to launch the new
one after OnSessionRemoved() is called for the existing one, because
otherwise the new one would get removed.

This code path is used when we do an auto-transition from mirroring to
flinging.

Bug: 1049636, b/149844005
Change-Id: Ie68603ebbfd3145e0f906bc4f13e41ee0474cbbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2113434
Commit-Queue: Takumi Fujimoto <takumif@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757245}