[flaky test] Wait for a frame to be produced in the test.
The Event.Latency.* metrics are generated from the gpu-process, and the
Graphics.Smoothness.* metrics are generated from the renderer process.
So by the time the smoothness metrics are available, the latency metrics
may still remain unavailable. So wait for a frame to be presented in the
gpu process before checking for the existence of the latency metric.

BUG=1067492

Change-Id: Ic6ac465f547de945321126b91d18e9b83e810b42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2140297
Reviewed-by: Mohsen Izadi <mohsen@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757261}