App List: Fix expand_arrow_view_ from being seen in shelf
Currently the expand_arrow_view_ has its opacity animated in two
different functions within contents_view. These functions are
UpdateExpandArrowOpacity() and AnimateToViewState().

The bug is caused because UpdateExpandArrowOpacity() does not check
whether the app_list is closed when animating, so the opacity ends up
being 1 when it should be 0. Also, the animation from
UpdateExpandArrowOpacity() is redundant since AnimateToViewState()
already takes care of it.

As a fix, this CL removes UpdateExpandArrowOpacity().

Bug: 1064258
Change-Id: Icda618df19199f35b617c99214d6f9db70a5e4e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134585
Reviewed-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757286}