[Suggested files] Only prefetch drive results when experiment enabled
We submitted a change that does a prefetch of Drive search results on
login, which is necessary for the suggested files experiment. However
that prefetch currently happens regardless of whether the experiment
is enabled or not.

The Drive ItemSuggest backend doesn't currently have capacity for the
expected QPS increase of this change, so we need to guard it behind the
experiment flag to control when the QPS increase happens.

Bug: 1034842
Change-Id: Ice8762b44b0e6279aaf5633941f1ad3b93b2eaf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2141471
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757324}