[devtools] Add request resource for issue reporting
This CL adds a request resource, identified by a devtools navigation
token, that is used to associate a cookie issue with a request.

This is crucial for the front-end, as some cookie issues are reported
during navigation, and without the associated devtools navigation token
the front-end cannot preserve these issues for after the navigation.
Such issues belong logically to the page that navigation navigates to.

Bug: chromium:1060628
Change-Id: If26dd6b78d409f0aba1f2badbefc6da76cb518ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2098632
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757382}