Remove Repair/Install Dialog when an extension is uninstalled
A crash occurs when user clicks on the Repair/Install dialog buttons
after the extension is uninstalled. The scenario is that the user
invokes both the "Repair" and "Remove" dialogs, but first clicks on
the "Remove" button to uninstall the extension. Now, when the user
clicks on "Repair" button, the browser crashes as the
CHECK(extension.get()) fails.
This CL adds Extension Registry Observer for extension install dialog,
which enables it to listen to extension uninstall event and close the
dialogs related to the extension.

Bug: 1026206
Change-Id: I98a5b5833c0bf5c120a8ee1c29110c2e9d051e52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2049444
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Commit-Queue: Shifona Garg <shgar@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#757406}