[User events] Fix and re-enable single-client RetryParallel test
This CL fixes a bug in a previously flaky and disabled test and
re-enables the test.

The issue was with the fake server simulating failure for the first
committed user event but it actually simulated failure for the first
committed sync entity which could belong to a completely different data
type.

Bug: 1006155
Change-Id: I335b8acba95086977713a9799efa696bcfca7046
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142292
Commit-Queue: Jan Krcal <jkrcal@chromium.org>
Auto-Submit: Jan Krcal <jkrcal@chromium.org>
Reviewed-by: Maksim Moskvitin <mmoskvitin@google.com>
Cr-Commit-Position: refs/heads/master@{#757468}