Don't close process handle early in ServiceProcessControlBrowserTest
In the Setup() function for ServiceProcessControlBrowserTest the process
handle is closed (on all platforms) in order to avoid unspecified
failures on MACOSX. However doing this means that the error code
checking in TearDown() is never run. This hides an apparent bug on Linux
and avoids useful error checking on other platforms.

This change makes it so that Close() is only called early on platforms
where it is necessary.

Bug: 1059446
Change-Id: I5a408e92a5587a8f52718b6c56660453abce5eef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2137856
Auto-Submit: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Wei Li <weili@chromium.org>
Reviewed-by: Wei Li <weili@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757479}