Deprecate and remove the ResourceReporter
This system was added in 2015 in the hopes that it will deliver
actionable results to detect and fix performance and memory
consumption regression.

The rappor metrics (which this system is using) was deprecated,
and the system failed to give us any useful stats. Therefore
the decision was made to remove it.

BUG=976423
TEST=Should compile, and tests should pass.

Change-Id: Ib57283ae6b8eda6d46ca1f3d67c68648de8c02c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2140594
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757482}