Deflake TLSDeprecationConfigTest.NoProto
TLSDeprecationConfigTest.NoProto is currently in the top 10 flakes on
the Flake Portal. The test is so flaky because it relies on global state
of a singleton but the test on trybots are executed in random order.
The test expects that the singleton is uninitialized but observes the
state left by other tests. This CL resets the state before performing
any validation of the behavior.

Bug: 1069016
Change-Id: If3f78c5809490166e233f67e60d0b9551b85c6cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2141099
Commit-Queue: Dominic Battré <battre@chromium.org>
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Auto-Submit: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757483}