Add a button color colorid
This cleans up the double mapping for kColorId_DialogBackground in the
color pipeline as well as adds a new kColorId explicitly for button.
The new colorid is hooked up as appropriate for button cases and returns
the same colors as it did with DialogBackground. Having it be separate
would be helpful for cases like high contrast where the button may need
greater differentiation from the background and for any downstream
consumers who use a different color even in regular modes.

Bug: 1067359
Change-Id: I966bb0446c85ba08d8fe9d844544ccd06aff7b24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2135385
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: John Smith <johnsm@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#757527}