CrOS: Add a feature and code to enable core scheduling for renderers
This change allows core scheduling to be turned on when renderers
start and before entering the sandbox. By doing it before entering the
sandbox it cannot be disabled. There is a flag we can use to disable
this if necessary via a finch kill switch should bugs in the kernel
be encountered.

BUG=b:152605392

Change-Id: I0f3b81a03eac2e3603bff8514e0c58e0d4e286c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2140066
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Brian Geffon <bgeffon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757541}