android: Disable gpu timeout for in-process gpu
In-process GPU has lower stability guarantees as it cannot be restarted.
It does not have a watchdog timer. Establish channel timeout is not
useful as it's only waiting for a thread to start and do work. So it can
really only trigger for false positives.

So disable the timeout for in-process GPU.

Bug: 1068670
Change-Id: I047e07c29d74e8b68a8bcf2ecdb333f066c377fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2140818
Reviewed-by: Khushal <khushalsagar@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757545}