Reland "Add field to SiteForCookies to indicate schemeful same-site of ancestors"
Fix and reland https://crrev.com/c/2133237 which was reverted in
https://crrev.com/c/2141207.

Add a new field to SiteforCookies which is used to indicate if |this|
would be the same when considered with schemeful same-site.

The field is updated in the browser and renderer when th computation for
a sub-frame's SiteForCookies is done.

This CL is groundwork for implementing schemeful same-site and thus
nothing is done with this new field yet.

Tbr:chlily@chromium.org,dcheng@chromium.org,morlovich@chromium.org

Bug: 1030938
Change-Id: Ie532e548a9842682170e8f97c7a44337013bbcbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142133
Commit-Queue: Steven Bingler <bingler@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757584}