If a distilled URL doesn't correspond to an article, show original URL.
In the omnibox we should show the original article's URL for distilled
pages, but if the distilled URL is invalid then we should continue
to show the dom-distiller:// URL rather than nothing.

This might happen if a user types in a URL beginning with
dom-distiller://. We do not want users to try to navigate directly to
dom-distiller pages, but we should handle the omnibox well when they
do.

Bug: 1066552
Change-Id: I7e569b74c49f9c5557647635f53815f724c8ce33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2129900
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757603}