crostini: reenable SubPagePortForwarding tests
I have a strong suspicion that the OSSettingsCrostiniPageTests were not
hanging, but instead were timing out because there were a lot of tests
under a single target, which only gets a 30 second window to run.

In crrev.com/c/2124281 the target was split up into 6 roughly equally
sized targets, of which one was disabled because we suspected that it
might be flaky. None of which have subsequently flaked.

Upon investigating the disabled test, I believe it isn't flaky, as all
of its asynchronous function invocations invoke flushAsync which cannot
timeout. Due to this, I believe it is safe to reenable, however I have
added some temporary logging, which will help us track down the cause if
it is in fact flakey.

If it doesn't flake within a week I will remove the new logging.

Bug: 1056268
Change-Id: I4d1869016642fcc26d94950aa0c138892d1df82f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2141874
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Julian Watson <juwa@google.com>
Cr-Commit-Position: refs/heads/master@{#757636}