DragDrop: Remove unused drag-over-remote-content test.
Dragging over cross-site sibling frames is already also tested in
DragAndDropBrowserTest.CrossSiteDrag[1], and more generally in
drag_and_drop_interactive_uitest.cc[2], so trying to do so from
web tests as well is no longer necessary.

This web test was previously disabled due to lack of OOPIF support in
eventSender, and as this support won't be added in the foreseeable
future, this test won't be usable in the foreseeable future anyways.

Note that DragAndDropBrowserTest.CrossSiteDrag[1] and its larger
test suite doesn't cover MacOS and Android, and many tests do not
cover Windows. For more information, see the linked bug and [3].

[1]: https://source.chromium.org/chromium/chromium/src/+/master:chrome/browser/ui/views/drag_and_drop_interactive_uitest.cc;drc=37470a6d05cfc50013acde00be27bb7723b1febf;l=1323
[2]: https://source.chromium.org/chromium/chromium/src/+/master:chrome/browser/ui/views/drag_and_drop_interactive_uitest.cc
[3]: https://chromium-review.googlesource.com/c/chromium/src/+/2142735/2#message-4a4534f4cc55ca581be8ed55a07cabfa0db7631f

Bug: 665058
Change-Id: Iec6b51af03b2edc72fa1fd3ee2c9c6097dc00b14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142735
Commit-Queue: Darwin Huang <huangdarwin@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757662}