Add ServiceWorkerResourceReader::ReadData()
This CL adds a method called ReadData() that returns a data pipe
consumer handle. The producer of the data pipe supplies the content of
a service worker script from storage. See [1] for the detail of the
proposal.

The internal implementation is almost the same as
ServiceWorkerInstalledScriptReader. Once ServiceWorkerStorage is
migrated to the storage service, we could stop using
ServiceWorkerInstalledScriptReader and pass the data pipe directly to
renderers.

This CL also moves ReadDataPipe() which was defined in
service_worker_installed_scripts_sender_unittest.cc to
service_worker_test_utils.cc as the function is useful to test
ServiceWorkerStorageControlImpl.

[1] https://docs.google.com/document/d/1UhdZnaeEP8GywZR0Gik6wF6jKmNY3je6aDhPfJ05YhU/edit#heading=h.z6754cqqpv4k

Bug: 1055677
Change-Id: I05aa5a27fabde65102ae3e34781ea1cf39e8b8f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2138677
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757670}