exo: Use interface versions directly for zcr_ interfaces
For zcr_ interfaces, protocol definitions are always updated in the
same CL of the corresponding exo change, and there's no risk of the
implementation being behind the definition xml. In fact, it is
error-prone that we always have to update versions in two places, so
let's just cosolidate them.

TEST=tast run eve-DUT arc.Boot
BUG=None

Change-Id: I802422292fa860dc7347e440fe9f0db2ecd046f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2138851
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Tetsui Ohkubo <tetsui@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757699}