Add support for CSS screen-spanning media feature
This change adds parsing and evaluation support for the `screen-spanning`
media feature, which currently accepts three values: none,
single-fold-vertical, and single-fold-horizontal. These values will
describe the state of a viewport that is split by a fold or segment of
the window.

Plumbing of the values to MediaFeatures will be added in an upcoming
change.

Intent-to-prototype:
https://groups.google.com/a/chromium.org/d/topic/blink-dev/KsPRO3hkOGw/discussion


Bug: 1039051
Change-Id: I4646faa49a139e1a0af55da676bdb460a8a6baf5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2129994
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#757705}