[css-flexbox] Migrate overflow-auto-resizes-correctly.html test to WPT
Partially migrate the contents of this test into the WPT-specific
directory, including only the checks that can be upstreamed and
leaving the parts that are Chromium-specific in the original test.

In addition, this CL also adapts the description of the original
Chromium-specific test to reflect what it actually does now after
having removed some parts. In the newly-added test, we also add
the relevant meta information i.e links to the relevant specs and
a test assertion describing its purpose.

Bug: 1063749
Change-Id: I57df80bf8a3c48dd94c70315a07d31715477a28b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2129537
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Reviewed-by: David Grogan <dgrogan@chromium.org>
Reviewed-by: Robert Ma <robertma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757780}