Refactor embedding token logic in DidCommitNavigation.
This is mostly a no-op. It reduces redundancy and hopefully makes it
simpler and clearer. It also skips calling SetEmbeddingToken if it's
already set.

There's a question as to whether this is OK to land during our stability
freeze. In terms of stability, this should be a no-op and fairly easy to
verify. In terms of cherry-picks, this is a very local refactor,
if something has a merge conflict with it, we could just CP this too.

Change-Id: Icb575f5fe3ff7a9bbea3d24cd7dec9c5ad884794
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2138935
Commit-Queue: Fergal Daly <fergal@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757798}