Refactor: Rename ShowManualPasswordGenerationPopup and expand comment
The documentation is expanded to reflect that this is the callback
executed at the end of the manual (user-triggered) generation flow. The
method is renamed to ManualGenerationResultAvailable to match its
counterpart AutomaticGenerationAvailable and avoid confusion with
ShowPasswordGenerationPopup. The renaming is also justified by the fact
that on some platforms a dialog gets shown instead of popup.

Bug: 1060131
Change-Id: I147daae8fe3cb4b87be576c9841f87475ae3c63b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2144154
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#757845}