Add a inclusion context getter to SameSiteCookieContext
Provide a function for consumers of the class to use to retrieve a
context value when they need to determine a cookie's inclusion status.

While this currently has no meaningful difference between accessing the
context member directly it will be helpful when the schemeful
context member is added so that callers will be returned the correct
value depending on the state of the schemeful same-site feature.

This CL also makes the member variables private.

Bug: 1030938
Change-Id: I29e30ab1266d1f8195483ae9fbe8e4e63b62e2d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2140032
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Commit-Queue: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757871}