WebLayer: Remove media stream notifications from prior app instances.
When the app crashes, it can leave behind media streaming notifications.
Track these in a pref and remove on next startup.

We could use getActiveNotifications instead of storing a pref, but it's
only available in API 23.

Testing is still a todo and may be somewhat difficult as full
integration testing requires crashing the app. For now this is manually
tested by removing the cancel() calls that execute on stream/tab
destruction and restarting the app.

Bug: 1025622
Change-Id: Icdd475707ac1af0789968264c202e00fd8964d33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142394
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757944}