Remove default case from switch statement in GetColorModelForMode().
The default case in the switch statement allows for a NOTREACHED() to be
invoked at runtime if a new ColorModel is added that isn't handled in
the switch statement. By removing the default case, the target will not
compile unless the new case is handled by the switch.

https://crrev.com/737599 already does this for IsColorModelSelected().
This CL refactors the common code used by IsColorModelSelected() and
GetColorModelForMode() into ColorModeToColorModel().

Bug: 1069266
Change-Id: I38287e828d48640d6a3eef9891bb7040bf7450b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142953
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757954}