[Live Caption] Fix regression that did not show captions settings subpage.
The captions settings subpage must be outside of the main content div.
This fixes a regression in which the caption settings subpage was not
linked to the a11y settings subpage. It also fixes a regression
discovered: "Regression: The spinner keep on spinning in
'chrome://settings' page."

Bug: 1069422, 1055150
Change-Id: Id40f612c09e97983be7b30dd26bbc0c09201a9bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2144315
Reviewed-by: Katie Dektar <katie@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Auto-Submit: Abigail Klein <abigailbklein@google.com>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757995}