Mark NSS smart card password dialog as blocking
The smart card password dialog is invoked by NSS when it needs to access
a smart card, and must not return control back to NSS until it has the
password. The dialog itself is configured as a global, but any NSS
operations that may interact with smart cards (such as signing messages
or verifying certificates) are invoked on worker task runners and with
the base::MayBlock() trait.

Bug: 1058070
Change-Id: Ib18c26e3d4c0b5063fe5337c6bf0da514ed4db73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142715
Auto-Submit: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758011}