Switched to using a groupping role for embedded objects with children
The Core API Mappings 1.1 specification is not 100% clear on how to expose
embedded and plugin objects in IA2 and UIA.
What is clear is that screen readers Jaws and NVDA don't work with PDF documents
at all when an IA2 role of embedded object is used.

In this patch, I tried to follow the spec, while being consistent between how we expose <embed> and <object> tags:
1. If an embedded object or a plugin has children in the accessibility tree
an MSAA role of system groupping and an IA2 role of section are used.
This is because Jaws and NVDA do not "see" any of the children if an IA2 role of embedded object is used.
2. If the embedded object or the plugin doesn't have children
an MSAA role of system client and an IA2 role of embedded object are used.
3. UIA mappings have not been modified, but it would be nice if we could decide
whether UIA_PaneTypeId could be used for both plugins and embedded objects.
4. Also, since all of our plugins are windowless, (we don't support ActiveX objects), I exposed
an MSAA state of SYSTEM_UNAVAILABLE as per the spec.

R=​dmazzoni@chromium.org, adam.ettenberger@microsoft.com

(cherry picked from commit b14dd4d9b29652c43eae84a0309cb186ff027309)

Bug: 1047856
Change-Id: I6a00eca551ca7aaf1310e2b539f69adbc4583611
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2031762
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Adam Ettenberger <adettenb@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#739676}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2144820
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/branch-heads/4044@{#915}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}