Disable autofill test on mlbshop as bug is investigated
Unrelated to autofill blink change causing a failure as
a checkbox click isn't being detected.

Disabling the offending site test. If blink change will
revert or change, can reenable.

If it is the proper site behavior, can change action 20 in
recipe to click the text of the element instead of the box and
update the file in the cipd payload.

Bug: 1069615
Change-Id: Idb3f10c5c3508b3cd75b665a3838098e46c16e52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2144826
Auto-Submit: Ian Struiksma <ianstruiksma@google.com>
Commit-Queue: Sven Zheng <svenzheng@chromium.org>
Reviewed-by: Sven Zheng <svenzheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758060}