overview: Fix crash in OverviewItemView::GetPreviewViewSize.
The stack points right at that line so it was pretty straightforward.
I tried reproing locally and writing a regression test but it was a
strange timing I couldn't get to work. The best I discovered was a
semi-repro steps which showed |overview_item_| was indeed a nullptr.

Test: manual
Bug: 1069533
Change-Id: Ia525a029e5afbdf863880cd888135c9f105d503c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2144677
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758091}