[OutOfBlinkSandbox] Move SandboxFlags blink=>network (reland)
This reland the following patch:
https://chromium-review.googlesource.com/c/chromium/src/+/2062336

The diff:
https://chromium-review.googlesource.com/c/chromium/src/+/2144100/1..5

The previous CL was reverted because if was not compiling reliably.
The target:
  //third_party/blink/common:common
was missing:
  //services/network/public/mojom:mojom_blink

However, adding it would creates a dependency cycle:

ERROR Dependency cycle:
  //third_party/blink/public/common:common ->
  //third_party/blink/common:common ->
  //services/network/public/mojom:mojom_blink ->
  //mojo/public/cpp/bindings:wtf_support ->
  //third_party/blink/public:blink_headers ->
  //third_party/blink/public/common:common

It means we can never #include a generated .mojom-blink.h from
//third_party/blink/common:common

Fortunately, there are already many .mojom-shared.h there. So
we can can include web_sandbox_flags.mojom-shared.h instead. The enum
definition is entirely contained in this file anyway.

The mojom-shared.h is already a dependency::
/third_party/blink/common:common ->
/services/network/public/cpp:cpp ->
/services/network/public/mojom:mojom [generated HERE]

R=kinuko@chromium.org
TBR=thakis@chromium.org,mfoltz@chromium.org,iclelland@chromium.org,mkwst@chromium.org,lfg@chromium.org,arthursonzogni@chromium.org

Bug: 1041376
Change-Id: I272d4971920d6d3c04a965f1aa9b90376353b446
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2144100
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758153}