[Android] Start to decouple InterceptNavigationDelegateImpl from Tab
We are targeting InterceptNavigationDelegateImpl for componentization
and sharing with WebLayer. The main blocker is its dependence on Tab.
This CL starts decoupling InterceptNavigationDelegateImpl from Tab:

- We add a //chrome-level InterceptNavigationDelegateTabHelper object
  to attach InterceptNavigationDelegateImpl to a Tab.
- We add an InterceptNavigationDelegateClient interface that the
  embedder implements to give the context that
  InterceptNavigationDelegateImpl is currently obtaining from Tab,
  and populate it with an initial method.

The implementation of the Client interface is currently separate from
the TabHelper object due to a thorny circular dependency issue in
InterceptNavigationDelegateTest.java. Once
InterceptNavigationDelegateImpl is cleaned of all dependencies on Tab I
suspect that this circular dependency will go away and we will be able
to merge the Client implementation into the TabHelper object.

Future CLs will complete the cleaning of //chrome-level dependencies
from InterceptNavigationDelegateImpl.

Bug: 1031465
Change-Id: I0f63b31d77efffe9d74ff68065c55e59ba163cd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134290
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758160}