[iOS][Credential-Provider] Set encrypted password
Adds to the password form the encrypted password when fetching from the
password store or passing through a change list.

On iOS this corresponds to the UUID used to retrieve the passwords
from the keychain services.

This is needed in order to fetch the passwords in Credential Provider
without duplicating the entries in the keychain.

Bug: 1066803
Change-Id: I921ad33cdc7579835059a4ff8768761ac5e6bf86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142112
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758173}