WebUI NTP: fix background image selection issues
The background iframe was not loading correctly on initial load and
after an image was selected following the customize dialog closure. To
fix this, only set the path of the background image iframe if the path
is not empty and is different. Also do not reset the background
selection unless the theme changes.

The logo was not updating when the customize dialog background selection
changed.

Bug: 1032328
Change-Id: I99304d80e7149439f57c02c9192af98af5deb4de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2145554
Commit-Queue: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758243}