WebUI NTP: update OneGoogleBar dark theme
This CL uses the same approach to update the OneGoogleBar theme as the
following CL.

https://chromium-review.googlesource.com/c/chromium/src/+/605951/

After the theme and OneGoogleBar are loaded, a message is posted to the
OneGoogleBar iframe to turn on/off dark theme based on the theme. When
a background is selected in the customize dialog, the OneGoogleBar is
also updated by setting dark theme when previewing an image.

Bug: 1039913
Change-Id: I7d420be98006185e0d55d5d41fdbe0f267c08875
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2145818
Commit-Queue: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758291}