And an HttpResponseHeaders::AddHeader overload that takes two arguments.
Also convert all of net/ to use it instead of the one-argument
constructor. The one argument constructor takes arguments in the form
of "foo: bar", which results in a bunch of code redundantly merging
keys and values into a single string to pass it in.

Also add a SetHeader convenience method which is a combination of
RemoveHeader and AddHeader.

Having AddHeader itself do this in only one place is safer and allows
it to validate the key and value arguments are reasonable. Also,
the single argument overload places limitations on the implementation of
AddHeader works, unless we want to have AddHeader potentially parse
the single string (that many consumers just merged from two separate
strings) into a name/value pair, so ideally we can just remove it, once
consumers are all converted to use the two-argument overload.

Bug: 1068194
Change-Id: I19734b9e597799234c0f50b2312540be559dc302
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2140553
Reviewed-by: Eric Roman <eroman@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758298}