android: Move WPT json results into storage bucket
We want to create a layout-test-results folder and put the json results
file in there. The bot will be changed so that it creates a storage
bucket where the results file can be retrieved by the rebaseline tool.

TEST: Look for the "archive results for weblayer_shell_wpt (with patch)"
step in trybot. Check for full_results.json file in downloadable archive.

TBR=robertma@chromium.org

Bug: 1050760, 1050754
Change-Id: If0912629053b4972e1426af6ea3bb3f061fbb128
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2140095
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Andrew Luo <aluo@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Luke Z <lpz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758303}