Storing extension Omaha attributes and performing actions.
Based on the custom attributes sent from
update client, if there is a key that
describes the extension as malware,
disable the extension. The custom attributes
are stored in an extension-specific
pref. An extension is disabled if a
"_malware" key is found otherwise it
checks if it needs to be re-enabled
if it was previously disabled remotely.

Bug: 1038797
Change-Id: I6aa72bac49417778df7b723b462752a7c0e6a560
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2073064
Commit-Queue: Bettina Dea <bdea@chromium.org>
Reviewed-by: Sean Topping <seantopping@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758323}