De-flaky testRenderIphDialog_Landscape
For some devices in landscape mode, the IPH entrance might not
be completely showing on the screen and thus Espresso cannot correctly
perform a click. This is causing the test to fail on x86 Pie trybots.
This CL fixes this issue by making it scroll to a position where the
view is completely showing for Espresso to click.

Bug: 1069809
Change-Id: I7ec8b5f43ebdfb318472ae923bce4080c917f7df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2144455
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758328}