Fix two-finger scrolls of Fullscreen Magnifier
In any screen rotation other than 0, the two-finger gesture scrolls
to move the focus region of the fullscreen magnifier are broken; i.e.
scrolling horizontally would move the focus region vertically or vice
versa, and also the scroll directions are not consistent (in some
rotations user needs to move fingers in the same direction where the
focus region is desired to move, and in other rotations in the opposite
direction.

This CL fixes these issues.

BUG=867537
TEST=Manually by enabling the fullscreen magnifier, rotating the
     screen in various orientations, and performing 2-finger scrolls.
     New test was added.

Change-Id: I17bd9dfe79d1c042f12fafdac11ede70e7b4697f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2143115
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758336}