[omnibox] Update Header to always show the toggle icon
This CL updates the Header to always show the toggle icon.

This CL fixes mouse events for the header, by deriving the control
from views::View correctly (instead of Label, which was a typo).

It also brings the layout closer to the final layout, with a
left-aligned label and the button on the right, by setting the flex
on the label to 1.

Bug: 1052522
Change-Id: I751dee3a3e73c1f1cba936e6a29b0ff7629a4b74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2145822
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Auto-Submit: Tommy Li <tommycli@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758346}