Group supported VAEntrypoints by CodecMode.
This CL fixes a problem where VaapiWrapper would detect a given codec as
being supported when in reality it wasn't. This occurs because before
this CL, VaapiWrapper::GetEntryPointsForProfile() ignored the CodecMode.
So, for example, in lulu, JPEG decoding is supported but encoding is
not. However, since GetEntryPointsForProfile() didn't care about the
decoding or encoding mode, it would return VAEntrypointVLD for both.
This would cause VaapiWrapper::IsJpegEncodeSupported() to return true.

This CL makes VaapiWrapper::GetEntryPointsForProfile() take the
CodecMode into consideration by grouping the VAEntrypoints according to
the corresponding CodecMode.

Additionally, a test is added to ensure that the default VAEntrypoint
(as determined by VaapiWrapper::GetDefaultVaEntryPoint()) is actually
supported. This is useful because we use the default entrypoint in
VaapiWrapper::Initialize() and we never check if it's actually
supported, so this test serves as an assertion on all boards.

Bug: 1068440
Test: ./vaapi_unittest --gtest_filter="VaapiTest.*" on lulu and grunt.
Change-Id: Ib4cceb75d67a4d26582874f810c04deacd1b51f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2144863
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758364}