bindings: Add missing [Exposed=Window] to IDL interfaces
This patch adds [Exposed=Window] IDL extended attribute to
the IDL interface definitions that are missing the extended
attribute.  The necessity of [Exposed] is determined by the
current behavior regardless of what each web spec specifies.

In a follow-up patch, the bindings generator will be fixed
so that it won't expose IDL interfaces unless [Exposed] is
specified.

Bug: 839389
Change-Id: Ib2cdb8801d1e3648c2e41d16617c41171fda0005
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2143173
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758387}