[serial/hid] Don't corrupt chooser state in Select()
As an optimization SerialChooserController and HidChooserController
were written to use std::move() to move the select device (or list
of devices) out of the list of options. This leaves the list in an
inconsistent state (a single null entry) which causes the crash
reported in issue 1069057.

These classes have been updated to follow the pattern from
UsbChooserContext of returning a clone of the selected option
instead.

Bug: 1069057
Change-Id: I016d12adad2977e49939dfb00044bfd7201767ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2145052
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758497}