[Extensions] ShowExtensionActionPopup -> ShowExtensionActionPopupForAPICall
Rename ShowToolbarActionPopup() and ShowExtensionActionPopup() to
ShowToolbarActionPopupForAPICall() and
ShowExtensionActionPopupForAPICall(). These methods were only used when
displaying a popup in response to an API call (outside of tests), and
this name makes it more obvious.

Additionally, this lets us get rid of the additional "grant_active_tab"
bool, which was always (outside of tests) set to false.

Bug: None
Change-Id: I02b4c489586f6cbec3ac99b02ce15413f2ceff40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2145858
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758549}