[scheduler] Field trial config for prioritzing rendering until begin main frame
Deleted the config for the old experiments, which are no longer active and were testing various prioritization schemes. Now that we have chosen a method of prioritizing rendering, this experiment is expanding on that by prioritizing rendering until we specifically see a frame as opposed to any rendering task.

Change-Id: I4da364bb32269c92c68730cc5be68c1c9c71c356
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2133270
Commit-Queue: Katie Dillon <kdillon@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758575}