[Switch Access] Add a MenuBubbleController
Currently, the Switch Access menu is drawn by the extension, but in
order to match the visual specification (see bug) the menu needs to be
drawn using native views.

This change creates a menu bubble controller for Switch Access and
adds the ability to show the menu. It also refactors the back button
to be a subsidiary of the menu. These functionalities are not yet used.

Relnotes: N/A
Bug: 973719
Change-Id: Iecfdda207873fccc9f1b3652fd424071684de449
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2140102
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758586}