Add pixel browser tests for focus rings
Focus rings are rendered differently on different platforms, and the
platform specific differences are currently untested because layout
tests force a platform-agnostic focus ring rendering code path. This
browser test is needed in order to run the platform specific code.

These tests need to be interactive_ui_tests because the window needs
dedicated focus in order to make the focus rings show up consistently.

A similar test was added in the content layer in
http://crrev.com/c/1969772

Bug: 1038425
Change-Id: I84ce398ab5d782e657cc13d0e073bf9997056bff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986972
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758587}