Adds browser side support for subframe translation behind a flag
This is a subset (2nd of 3) of a large change proven out in CL 2117459.
Its focus is to maintain the existing ContentTranslateDriver code as much
as possible for patch-ability while adding a PerFrameContentTranslateDriver
that is meant to replace it in the future.

[Update: there will also be a follow-up CL to move the CLD3 language
detection work into a sandboxed utility process.]

The overall change description:

This conditionally will perform language detection in the browser
process using Accessibility apis to get page contents in order to
determine language from subframes as well as main frame and then
supports translating subframes in addition to the main frame. This
new behavior is available when Feature "TranslateSubFrames" is
enabled; otherwise, the existing language detection (using
WebFrameContentDumper) and main frame translation support is preserved
(in the spirit of the current guidelines against refactors).

Bug 1064974 tracks follow-up clean-up tasks for the duplicated code
in this CL once the feature is launched (once refactors encouraged
again).

Bug: 797328, 1063520
Change-Id: I98b21af2d32285cb0ceb61672ad7140fc21e40d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2141217
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Megan Jablonski <megjablon@chromium.org>
Reviewed-by: anthonyvd <anthonyvd@chromium.org>
Commit-Queue: Doug Arnett <dougarnett@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758596}