Bring captions bubble in-line with spec, add title.
Updates colors, fonts, sizes to match new Live Caption spec.
Adds a title per spec which is shown when the caption is only
one line high.

Sets the max width of the bubble, but does not yet dynamically
resize.

Note that the bubble is still too tall because it shows a
close button. Per spec we will probably move this close button
out later, but leaving it in for functionality during testing.

Bug: 1055150
Change-Id: I5f4a3fa0b218f5c8a4b50fb0fedb8d4a1547f44e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2144813
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Commit-Queue: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758614}