Remove ssl_blocking_page_unittest.cc
The test in this file checks the SB reporting API is called when an
SSL interstitial is shown or clicked through, but checks the non
committed interstitials code (and has been disabled for a long time).
For committed interstitials this is better checked with a browser test
and one already exists (SSLUITest.TestExtensionEvents).

Bug: 1026557
Change-Id: I696e628a66341a377130bbc201db9cc25a98a064
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2145871
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758617}