Treat insufficient events as non-failure
When testing whether a system's status is clean or not it is possible to
not have enough events to be able to make a determination. In this case
we are in a state of not-success but not-failure. Given that this state
is known to happen we need to not fail the test.

The cause of insufficient events is well understood (running
browser_tests generates enough events to flush out all older events) and
Microsoft's recommendation is to ignore the DCOM events that cause this
flushing so that is what this effectively change does. Reference:

https://support.microsoft.com/en-us/help/4022522/dcom-event-id-10016-is-logged-in-windows

Bug: 968440,1070362
Change-Id: Ie5e2e10289da02b306419cdbcccc707774f55001
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2147795
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Jesse McKenna <jessemckenna@google.com>
Cr-Commit-Position: refs/heads/master@{#758623}