[Extensions] Replace type-specific action getters in CommandService
Replace GetBrowserActionCommand() and GetPageActionCommand() in
CommandService with GetExtensionActionCommand(), which takes an
ActionInfo::Type to specify the type to look for.

This reduces the complexity of many callsites (like
ExtensionActionViewController and ExtensionInstalledBubbleModel) and
reduces the dependency on specific types of actions, which will make it
easier to support a generic "action".

Bug: 893373
Change-Id: I9f566669f77256092676ccdea4e6aeb7b5054bab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2145157
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758644}