Change kSupervisedUserApprovedExtensions to SYNCABLE_PREF
The pref kSupervisedUserApprovedExtensions contains extension ids that
the parent has approved for the supervised user. We want these
approvals to sync across ChromeOS devices. As a SYNCABLE_OS_PREF,
kSupervisedUserApprovedExtensions does not sync from M83 to M84 due to
the SplitSettingsSync feature in M84. Changing this pref to a
SYNCABLE_PREF fixes this syncing issue from M83->M84.

Moreover, this pref should be a browser pref instead of an OS pref to
be more consistent with the rest of extensions.

Bug: 1068794
Change-Id: I8ae34dbd1e12d312ff72f062c9e5026bb8e6a525
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2147853
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Commit-Queue: Toby Huang <tobyhuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758654}