Fix accuracy of the kNewVersionApprovalGranted UMA metric
For supervised users, the kNewVersionApprovalGranted bucket in the
SupervisedUsers.Extensions UMA histogram is only supposed to increment
when an already-installed extension gets a version update.

Currently, the enum increments every time the supervised user enables
and disables the extension. This CL fixes the accuracy of this UMA
metric.

Moreover, SupervisedUserService::UpdateApprovedExtensions() currently
gets called every time the extension is re-enabled. This operation is
unnecessary and expensive because the function iterates over the
entire map of approved extensions. This CL prevents this unnecessary
operation from happening.

Bug: 1068438
Change-Id: I521fca2bbb3d0bc83744d74e8d9abeb427864bf6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2138637
Commit-Queue: Toby Huang <tobyhuang@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758707}