[Android] Add Drawable badge support to ProfileDataCache
Currently ProfileDataCache doesn't support vector graphics. This cl replaces
the use of Bitmap with Drawable in ProfileDataCache.BadgeConfige so that
vector graphics can be provided to it. Also this cl adds a new factory
method in ProfileDataCache to create an instance of this class with a
provided badge resource id.

Bug: 1064957
Change-Id: I23e5417611787f5d3d17f1a2b81c195bcea781ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2130246
Commit-Queue: Tanmoy Mollik <triploblastic@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758738}