[Android] Clean InterceptNavigationDelegateImpl of Tab deps
This CL removes the remaining deps on Tab from
InterceptNavigationDelegateImpl in order to make it suitable for
componentization and sharing with WebLayer. To do so, observance of the
Tab is moved to InterceptNavigationClientImpl. There is a slight
reordering of method ordering calls in order to so:
InterceptNavigationClientImpl starts observing the Tab immediately
after the construction of InterceptNavigationDelegateImpl whereas
that observance currently starts in the middle of the
InterceptNavigationDelegateImpl constructor. However, no observer
callbacks can happen as a result of the code running in the
InterceptNavigationDelegateImpl constructor (and the TabImpl#addObserver
call itself does not cause any observer callbacks), so this reordering
is harmless.

Bug: 1031465
Change-Id: Ic1f43d91076c414f095dd84f37bc24ca8a904e34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2135749
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758754}