Remove expectations on merge deltas from syncable services unit tests
A preparation for https://crrev.com/c/2087914. Expectations on the
values of SyncMergeResult::num_items_{added/deleted/...} are removed.
Coverage for the sync merge logic is still present, since the tests
already verified it by mean of other checks.

The CL complements such pre-existing checks with checks on the initial
and final size of the local model. One of the implementations of
GetAllSyncData that was removed in crrev.com/c/2083455 is re-added to
enable this. It is suffixed with ForTesting as with all other cases in
crrev.com/c/2083455.

Change-Id: If83c977b8e5dac6279b3fd7fc7517dc3d2dbc1f5
Bug: 1057577
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2091346
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758773}