[Passwords] Show empty state after unlocking account storage
This CL adds an empty state preventing the user from seeing the loading
state forever if no passwords are available after unlocking the account
storage.
An empty state is preferable over closing the dropdown since it provides
the feedback that the requested action (look in account storage) turned
up empty and provides the user with a path to settings.

Thus, the empty state completes the unlock flow:
1. Focus a password field
2. Click on the unlock button
3. Reauth
4. Wait for store to push an update (either new data or delete data)
 a. Use the newly available suggestions.  <-- Already implemented.
 b. See the empty state; go to settings.  <-- This CL


Screenshot for with empty state string:
https://storage.cloud.google.com/chromium-translation-screenshots/5a36eba8302b1aca1911d2a2873ecbc09363d290

See linked bug for a short video.

Bug: 1068555
Change-Id: I419d508e1faa2c230f3e4c85de71a9e08da1df54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142113
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Christos Froussios <cfroussios@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758831}