[CSP] Remove an outdated warning for `report-uri`.
We previously required `CSPRO` headers to contain `report-uri`. Since we
added a DOM interface (`SecurityPolicyViolation` events and Reporting
API integration), that requirement is no longer applied. We still
surface a warning, but that's more verbose than it needs to be, given
that the use case it warns about is completely legitimate.

Bug: 1070533
Change-Id: Ica89fa0c8d53d563c6510426e2f6f11d2e7400c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2145881
Commit-Queue: Mike West <mkwst@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758851}