cbuiv: remove IdleActionWarningDialogView Cancel override
This cancel override is essentially a no-op now. There is no cancel
button on this dialog so the only call path is via CloseDialog, which
always stops the timer before calling into Cancel, so Cancel was
always returning true. This change removes the Cancel override and
the explicit stop of the timer, which is no longer needed.

Bug: 1011446
Change-Id: Ifa6d42591ebcb4063aae05aa959199f407e50df0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2148678
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758962}