Commit tab closure for last tab when session restore is enabled
If the one and only tab in a session is an auto-closed tab, such as a
facebook login host, then subsequent launches will restore that tab and
immediately close. While this can be worked around by launching to
another tab, many users could get lost and frustrated and not know what
to do. To fix this, we can detect the situation and force commit the
tab close.

Bug: 1052096
Change-Id: I7372e2a0a13922abb7f2245b8c4a2edf7514fe95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2057627
Commit-Queue: Cliff Smolinsky <cliffsmo@microsoft.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758972}