Trust Tokens: Add a data removal interface
This is part of a sequence of CLs implementing data removal logic for
Trust Tokens [*], a prototype feature that enables annotating requests
with some lightweight cryptographic state. The internal storage stores
the results of these cryptographic operations, including some state
necessary for maintaining privacy invariants.

The first CL in the series expanded the Trust Tokens storage system to
satisfy deletion requests parameterized by a
network::mojom::ClearDataFilter. This CL extends NetworkContext with a
Trust Tokens deletion interface and adds a call to this interface in
BrowsingDataRemoverImpl.

The control flow is quite similar to that for clearing reporting data
(which is also scoped per-NetworkContext and deleted from
BrowsingDataRemoverImpl whenever cookies are cleared).

[*]
https://docs.google.com/document/u/1/d/1TNnya6B8pyomDK2F1R9CL3dY10OAmqWlnCxsWyOBDVQ/edit

R=csharrison

Bug: 1064747
Change-Id: I989790c932d1fff2f2fe6245b9c66b456f07b31b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120102
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758983}