Make Variations unittests use ResponseHeader::SetHeader()
We want to get rid of the single argument AddHeader() method as the two
argument AddHeader() and SetHeader() methods are safer and more
flexible.

In this case, switching to SetHeader() to make clear that there's only
one instance of each of the headers being set.

Bug: 1068194
Change-Id: Id17aba78bb54c47d4087324726b1c32920f29233
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2148021
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#758999}