SafeBrowsingEnabled pref needs to be on for Enhanced Protection
SafeBrowsingEnabled pref is turned on when Enhanced
or standard protection is turned on. It is always
turned off if no safe browsing is selected. Devices
can be out of sync or not on the most updated versions.
This means that enhanced protection may not be offered
on these devices. To make sure that all safe browsing
is kept consistent, SafeBrowsingEnhanced needs to
also check SafeBrowsingEnabled. If the user has
no safe browsing on a device and syncs user pref,
all devices should not have safe browsing on unless
they select a different safe browsing afterwards
on any device.

Bug: 1068777
Change-Id: I9a13c8347c30c965d8e0da4e704856ba9b054c30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2148433
Commit-Queue: Bettina Dea <bdea@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759008}