[webauthn] Inline biometric enrollment sheet
Add a dialog sheet to webauthn that will be used for inline biometric
enrollment. The sheet contains an image of a fingerprint and a circle
that fills progressively as the user taps the fingerprint scanner. This
is based on the existing UI in the settings page but implemented in C++.

The strings for the sheet are taken directly from the settings page.

This UI is not connected to anything at the moment, but it can be tried
by running the following command:

browser_tests --gtest_filter=BrowserUiTest.Invoke \
  --test-launcher-interactive \
  --ui=AuthenticatorDialogTest.InvokeUi_inline_bio_enrollment

Tbr'ing cpu@ as per chrome/app/OWNERS.

Tbr: cpu@chromium.org
Bug: 1056317
Change-Id: Ib18290c2eefedfa8a1b80a240153d8e881e3402b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2140832
Commit-Queue: Nina Satragno <nsatragno@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#759024}