Stub in Assistant search provider.
The Assistant search provider will be used to provide a single search
result from its internal cache of conversation starters. This will be
surfaced as a suggestion chip in the launcher.

This CL simply stubs in the provider (along with a single hardcoded
result) so as to unblock development of the logic which will allocate
Assistant a dedicated chip slot.

Changes are gated behind a new feature flag:
  AssistantLauncherChipIntegration

Bug: b:146673437, b:153165833
Change-Id: I54159c3ee17110c8fad46bbb3cea4a88b3250df9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134651
Commit-Queue: David Black <dmblack@google.com>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Tony Yeoman <tby@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Jia Meng <jiameng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759050}