Add entries to the main console for missing waterfall builders.
There are some configurations that we do not have the capacity to run
on CQ and/or do not catch enough errors to justify running on CQ but
still want to roll back breaking changes. This adds those builders back
to the main console.

Change-Id: I9c8059fdda5f181521c85037c2d4c57bc54aba94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2147873
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Aaron Gable <agable@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759089}