Update the aria label for the safety check reload button
The steady state button uses the aria label "Run safety check now".

Per feedback from the accessibility review, when we swap that button
with a reload button, we should also update the aria label.

We will use the label "Check again". This string already exists
in settings_strings.grdp, with a reasonably matching strings
description, so we will reuse it in order to be able to merge this
change into M83.

We will introduce a dedicated string resource in a follow-up CL
(that will not be merged into M83):
https://chromium-review.googlesource.com/c/chromium/src/+/2149418

Bug: 1070841
Change-Id: I9b9ee80a53ce0efc591d78fe384ce29748f183d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2149647
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Commit-Queue: Martin Šrámek <msramek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759112}