media/gpu/v4l2: use NOTREACHED when V4L2Buffer::GetVideoFrame() is called improperly
V4L2Buffer::GetVideoFrame() is an internal method that is never exposed
outside of V4L2Device. Change a regular error handling (which would be
the way to do if this was exposed) into a NOTREACHED() so we catch
potential bugs in our code.

BUG=None
TEST=VDAtest --use_vd passes on Kukui.

Change-Id: Ide5cf00447df154d39d1400d156633ea67b45e2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2143426
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Cr-Commit-Position: refs/heads/master@{#759137}