[Android] Componentize InterceptNavigationDelegateImpl
After preceding CLs this class has no remaining //chrome dependencies
other than resources, which this CL componentizes along with the class
itself. As this implementation of InterceptNavigationDelegate is
specific to the overall intent launching logic in
//components/external_intents, this CL componentizes it into that
component. Componentizing this class will enable it to be reused in
WebLayer.

TBR=jochen@chromium.org

Bug: 1031465
Change-Id: Iea7f68ebdee3fc4d766604c46ba5868429c14e65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2133984
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Mugdha Lakhani <nator@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759187}