[WebLayer] Reuse //chrome's InterceptNavigationDelegateImpl
This CL eliminates //weblayer's custom implementation of
InterceptNavigationDelegate in favor of using the //components-level
implementation that Chrome uses. As part of this change, we adapt
//chrome's InterceptNavigationDelegateClient implementation as suitable
for WebLayer. With this change, WebLayer now eliminates stale
navigation entries left on intent launches as Chrome does and also
detects cases where tabs should be closed on intent launches as Chrome
does. Actual closing of tabs in those cases is left as a followup.

As part of this change, we add an implementation of
ContentBrowserClient::IsHandledURL() to //weblayer's
ContentBrowserClientImpl. This implementation is needed for the shared
InterceptNavigationDelegateImpl to behave as expected. The
implementation is based on that in //chrome, with Chrome-specific items
deleted (notably, the kChromeSearchScheme is not handled in WebLayer).

The diff between PS6 and PS8 contains the differences between //chrome's
InterceptNavigationDelegateClient implementation and //weblayer's. One
of the notable differences is that WebLayer's Tab objects (a) cannot
be directly observed and (b) do not change their associated WebContents
during their lifetime. Thus, the //weblayer
InterceptNavigationDelegateClient implementation observes WebContents
directly, with //weblayer's TabImpl additionally calling directly into
the client implementation as necessary.

Bug: 1031465
Change-Id: I010c4c3ef066bfb2d718f597c654a3ce3df890db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2137441
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759230}