[ios] Fix DIR_SOURCE_ROOT for EarlGrey2 tests
On iOS, the application has no access to source_root directly as
all applications are bundled. Thus the test data are bundled as
asserts and need to be located in DIR_ASSETS.

For most of the tests, the directory containing the app binary
is the directory containing the assets, but this is not the case
for EarlGrey2 test that are part of a PlugIns bundle.

Using DIR_ASSETS instead loads the correct path since it is
relative to the code that is running, not to the main application.

Bug: none
Change-Id: I25b0c7b9ac8a7686332591f4488440ffb7f5d4c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2149427
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Mark Mentovai <mark@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759233}