Update required screen appears with no rendering of UI.
onBeforeShow() of checking-downloading-update element is not being
called due to which the oobe dialogs contained inside it are not
being rendered because these dialogs use lazy rendering in
OnBeforeShow() of the oobe dialog. This creates a situation with no
UI element on the screen but the update process running in the
background. Now, calling onbeforeShow() for this element from the
defaultControl() of update required screen.
This CL also adds a test to avoid such case in future. It tests
the update process along with the verification of the UI actually
being displayed.

Bug: 1048607
Change-Id: Ib1ac1c8d935ec059e62202095e7009f89e98a900
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2146901
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Cr-Commit-Position: refs/heads/master@{#759244}