[FlexNG] Don't consider replaced elements as having children.
Previously we considered any legacy layout root as potentially having
%-dependent descendants.

This typically wasn't an issue for when we first rolled out LayoutNG as
it was relatively common for *something* to stop the propagation of this
bit up the tree.

With FlexNG however, we run into far more cases due to the
"definiteness" potentially changing.

This patch treats images, svgs, etc has having no children for the
purposes of this calculation.

Bug: 845235
Change-Id: I98648b8182fd549d6e6bee1167cd8ae89d55687b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2148216
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: David Grogan <dgrogan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759271}