[b4p] Refactor: Offload client authentication APIs to dedicated helper
First step to remove the duplication for b4p reauth. The implementation
of methods TriggerReauthForAccount and TriggerSignin are moved to a new
helper class, as they are.

In the next CLs, the class we'll start to internally recover the primary
account id and set the opt in preference for account storage.

Bug: 1070944
Change-Id: Ic6c950d28551a6b624f2c88943f880a961d9e4d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2150426
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759301}