Remove SkiaOutputSurfaceImplTest parameter
The test was parameterized so it would run with an offscreen surface on
all platforms in one version and a real XWindow on Linux in another
version. There is a crash due to not calling XInitThreads() on Linux
when using a real XWindow.

The version of the test with the real window doesn't seem to add much
value in terms of testing the SkiaOutputSurfaceImpl API. GPU tests will
ensure that SkiaRenderer works properly with platform window already.

Bug: 1069332
Change-Id: Ibad1db5c49cca67121828f6f3f6ed0d7081ee7ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2149513
Reviewed-by: Jonathan Backer <backer@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759313}