device chooser: de-flake tests
Right now, the count of extant bubbles is used to keep track of
how many bubbles exist. This leads to flaky tests though: the bubble's
view is only deallocated after closing the bubble finishes, and since
closing the bubble is async, whether the view has been deallocated or
not at the end of the test is variable.

This change has the tests instead ensure that the close process has
*started* at the end of the test, which is always true.

Bug: 1069695
Change-Id: I29535149fd3f318c22b44d8f93d22c5b248816da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2149685
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759330}