[CrOS Settings] Fix battery Idle Setting
Recently an additional dropdown has been added to allow setting
battery idle behavior differently from AC idle behavior. This
setting is visible only when battery is present and is hidden
behind a dom-if template. Since nodes under dom-if are dynamically
created, they are not added to the this.$ hash. As a result battery
idle setting dropdown is not working. Instead use this.$$ for
accessing the dynamically created #batteryIdleSelect node. This
change also adds a test to catch the same.

Bug: b:148105990, 1065050, b:154051991
Test: Run browser_tests and manual tests.
Change-Id: I5970d6b362fc4c4c9685ddfe7d388e0a197007b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2147015
Commit-Queue: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Auto-Submit: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759340}