Remove WebTestDelegate
Convert pointers to WebTestDelegate into pointers to BlinkTestRunner.

Change a few places that hold a pointer to TestInterfaces in order to
get a (potentially changing over time...?) BlinkTestRunner to point to
a BlinkTestRunner directly instead. We still grab the BlinkTestRunner
off TestInterfaces for now though.

R=avi@chromium.org
TBR=

Bug: 866140
Change-Id: Id1f336a7a178a8ec0b0f8a0b792ef9040c788d5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2149506
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759345}