Creates public interface for AssistantController.
This CL:
- Creates public interface for AssistantController
- Renames existing AssistantController to AssistantControllerImpl
- Updates references

This new interface (which exposes a singleton accessor) will be used
moving forward to get a handle to the AssistantController from outside
of Ash. The most pressing need for this is the new Assistant search
provider.

Note that this CL doesn't yet add public APIs onto the interface, as I
attempted to only make the smallest possible change in this CL.

Follow up CLs will migrate public APIs onto the interface as well as
remove the existing accessor to AssistantController that is currently
exposed through Shell so that we can enforce consistent access to
AssistantController through this new public API moving forward.

Bug: b:146074655
Change-Id: Id9f414f66fbfc60c90eb5d0c8a9b86681c641da2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2148098
Commit-Queue: David Black <dmblack@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759399}