[ios] Checks if Bookmarks rootNode is not nullptr before caching row.
I've been trying to figure out why this is happening, my theory is that
this is somehow related to swipe down to dismiss (the crash is iOS13
only) and dismissWithURL being called twice (the second time _rootNode
might've been niled).

Though I'm not sure why this didn't happen on M80 where this behavior
existed. So I'm not sure what could've changed that caused this crash.

This is a speculative fix, but modifying the code and setting _rootNode
to nullptr reproduces the crash stack exactly. Even if we'd know the
root cause, I'm not sure we'd want to cherrypick something more complex
than this to M81.

Bug: 1061882
Change-Id: Ic47449388e57e17c8acc34e104cc297dc8c2638d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2147955
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759413}