Ensure time limit is reset at the specified time in PATL policy.
This Cl ensures that time limit is reset at the specified time in
PATL policy. When the computer wakes up it decides to reset time
limit based on the last reset time. If reset time boundary
has been crossed because either time has moved forward or backward,
chromeos resets the time and saves the current time as the last
reset time which will result in wrong reset times in the future.

The following is an example scenario where the reset time is 6 am.
1. User opens computer at 1 pm on day 1. There was no reset time before
so the computer saves 1 pm day 1 as the last reset time.
2. Computer schedules the next reset time to be on 6 am on day 2.
3. After a few hours of activity, the computer is closed before reaching
the correctly scheduled limit.
4. User opens computer at 9 am on day 2.
   User observes ERROR:
   	-> EXPECTED: time limits are reset because the reset time
		     has been crossed.
	-> ACTUAL: time limits are not reset until after 1 pm on day 2.

This cl corrects the issue seen in step 4.

Bug: 1069932
Change-Id: I9cd4ad4c887e2067292612cf1d32c9dd43bafa6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2149289
Commit-Queue: Yilkal Abe <yilkal@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759427}