[ntp] Remove doodle flower
On customized NTPs we don't show a doodle since the doodle may not jibe
well with the custom background. Before this CL, if a doodle was
available we showed a bouncing ball animation - the flower. And, upon
clicking the flower, we temporarily set the NTP to the default theme
showing the doodle. However, it turned out that the flower causes a
significant amount of CPU usage and it is unclear whether users
understand the flower lets them peek the doodle. Therefore, this CL
removes the flower.

Fixed: 1046786
Change-Id: If236a3fc4be4f7ef2cf9deebad1d579103902c59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151318
Commit-Queue: Tibor Goldschwendt <tiborg@chromium.org>
Auto-Submit: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759465}