DNR: Introduce 'id' key as part of Ruleset.
This helps the extension to uniquely identify a ruleset. It would be
helpful once we start allowing extensions to dynamically change the set
of enabled rulesets.

This is a breaking change to the API schema, however it's ok since the
API is only on Beta and doesn't have significant usage.

BUG=754526

Change-Id: I41edf43acdaa1c97c121f4cca0eda5204fd2f9ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2146339
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759475}