Remove hidden cursor from screen
Currently ui/resources/cursors/none.cur is used to hide the cursor,
which is a dark 1x1 bit bmp. This may cause DWM to do composition
with the underlay video at very frame. To solve it, the hidden
cursor should be further removed from the screen.

Bug: 1069698
Change-Id: I4122a7a2085dbde5a78053e69f29f988a1b20de4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2145234
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Jie A Chen <jie.a.chen@intel.com>
Cr-Commit-Position: refs/heads/master@{#759481}