Preserve KVO when the underlying scroll view changes.
Lack of this functionality was revealed in the Trybot failure of:
https://crrev.com/c/2137201

With this change, CRWKeyValueObserverForwarder is now created for each
call to -addObserver:, not for each pair of (observer, keyPath), to
allow restoration of all observations.

-startObservingScrollView: is now called in the initializer. KVO against
the placeholder UIScrollView is not necessary, but it simplifies the
implementation. Otherwise it needs to keep track of two KVO observation
states separately (one for observations by |self| and another for
external observations) which looks complicated.

Bug: 1023250
Change-Id: I2f553d38d319f7b064c8c76e5d13e914ebcef8eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2147433
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Hiroshi Ichikawa <ichikawa@chromium.org>
Auto-Submit: Hiroshi Ichikawa <ichikawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759503}