WebXR: Avoid unbound-Remote errors on shutdown
Depending on the shutdown sequence, we may get to OnExitPresent via
OnPipeConnectionError / NotifyError when the client is not bound.
On the assumption that this is due to being called twice, clear
the client set to avoid "Cannot issue Interface method calls on an
unbound Remote" errors.

Bug: 1070806
Change-Id: I1081512948ca175a99707b158e32243eb45d569e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2150033
Commit-Queue: Klaus Weidner <klausw@chromium.org>
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759509}