Make NavigateFrameToURL return a success value like NavigateToURL.
This is just a convenience. I've simplified one use of this. I'll be
adding another use in a different CL.

https://crrev.com/c/2148914 wraps all calls to this in EXPECT_TRUE but
of course there are failures. It would be a bad idea to submit this
right now with the code-freeze anyway. So I'll come back to it.

Change-Id: I16b170fe8e697d0eb06716e30b877efdaa523869
Bug: 1071297
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2148711
Commit-Queue: Fergal Daly <fergal@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759519}