Change return type of SyncableService::MergeDataAndStartSyncing
The caller of this method now only uses the error field of the returned
SyncMergeResult, and lives under the Unified Sync and Storage directory,
so this CL changes the return type to base::Optional<syncer::ModelError>
instead. The implementations of the interface are refactored to remove
the logic that previously filled the remaining fields of the merge
result.

Bug: 1057577
Change-Id: I9f2f5d32e3f3379d0cd9cd50714be248483ff9a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2087914
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Reviewed-by: Dan S <danan@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#759609}