bindings: Update code around IDL dictionaries
This CL updates IDL files and/or related C++ code to be consistent
with spec, as far as the changes do not change actual behaviors.

What this CL does are
- Add default values on some dictionary members
- Check if a dictionary member is missing before accessing it.
- Check if a dictionary member value is null, if its type is nullable.

After this CL, they would work as expected also with the new IDL
compiler which is under development.


Bug: 839389
Change-Id: Ic7740092fe28bc4a76c0e5f5e268c6d112612cda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120079
Commit-Queue: Hitoshi Yoshida <peria@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759631}