[signin] Extend lifetime of Signin.GetAccessToken* histograms
These histograms are still useful in particular because we are going
to migrate desktop platforms to multilogin, and these histograms will be
used to check the success of the migration.

We may consider keeping the histograms even after the migration is
complete, but we'll see at this time. If we want to keep them indefinitely
we should maybe consider renaming them however, because they are specific
to multilogin, and it's not really captured in their name.

Fixed: 1059815
Change-Id: Iff2860a961945e01abd965a51be9d90a1e121185
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151457
Commit-Queue: David Roger <droger@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759646}