Feedv2: Implement background refresh
Note that I've chosen to NOT use 'next_stream_state'.
Instead, we load the data similar to the way we
load data when doing the initial load. Doing it this way
is simpler, but will limit us to doing background refreshes
only when there are no attached surfaces.

Bug: 1044139
Change-Id: I304a7ba158e6a9c7e333606f63c50c947216638a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2145933
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Commit-Queue: Dan H <harringtond@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759693}