[iOS] Add InfoBarIOS::infobar_type()
InfobarUIDelegate will not be necessary once we move to the overlay-
based implementation.  The only functionality needed from the UI
delegate necessary with overlays is the InfobarType.  This CL adds
InfoBarIOS::infobar_type() to access this type, as well as a constructor
that does not take an InfobarUIDelegate.

This CL also updates the password infobar to be created with this new
constructor if the overlay feature is enabled.

Bug: 1033228
Change-Id: I3080bc2e5bb9e2e3db23dfe437ada90b70bb1661
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2150125
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759741}