cbuiv: remove CreateChromeApplicationShortcutView Accept & Cancel
This dialog's Accept method could return false under two circumstances:
if the dialog's app info had not yet been loaded, or if the accept
button was disabled. When the accept button is disabled there's no way
to hit the Accept() code path, so one of these was dead. For the other,
this dialog now disables the accept button when the app info is not
available.

Bug: 1011446
Change-Id: Ic6dd8bdd6fd2bba2c092004c0dde5f98bf98ea54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151309
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759747}