WebLayer: add NavigateParams to NavigationController.navigate()
This struct should more or less mirror LoadURLParams, or at least
as much of that struct as necessary. For now it only contains
mShouldReplaceCurrentEntry.

This makes replace() obsolete. Since this is designed to be merged
back to 83, replace() was added in 82, and 82 did not make it to
stable, replace() can be removed without the normal deprecation
process.

Bug: 1070851
Change-Id: I0845bab184a182e9a616b01b6115615275422fc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2150968
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759766}