Fix black window on DWM crash
Freeing DComp resources such as visuals and surfaces causes the
device to become 'dirty'. We must commit the changes to the device
in order for the objects to actually be destroyed.

Leaving the device in the dirty state for long periods of time means
that if DWM.exe crashes, the Chromium window will become black until
the next Commit.

Bug: 1071288,1011655
Change-Id: I7d118e5bc5f1e4fc6959661b1902939ba7dc83f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2152784
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Rafael Cintron <rafael.cintron@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#759773}