Prevent overlay scrollbars from processing events when transparent.
Mac offers overlay scrollbars that become transparent when not in use.
When transparent, the scrollbar still exists and takes click and hover
events. This  causes a visual bug for the ExtensionsToolbarMeunuView
where a button sits partially under the overlay scrollbar. This
change updates the scrollbar ability to process events with the subtree
when the scrollbar is transparent.

Bug: 1016578
Change-Id: I64c363943167a31f39308eacf7863977ab47d2ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134648
Commit-Queue: Caroline Rising <corising@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759774}