[omnibox] Fix tab to focus Suggestion Removal button, and add test.
The Tab to Focus Suggestion Removal button was regressed in this CL, which
I reviewed:
https://chromium-review.googlesource.com/c/chromium/src/+/2131021

The above CL is a great refactoring CL, which we don't want to revert.
It just broke a piece of functionality.

This CL fixes the functionality and adds a test to prevent regressions.

We will need to backmerge this into M83.

Bug: 1071542
Change-Id: Ib458c1404deccfb85f06015ccb2cae751db9744f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2153323
Auto-Submit: Tommy Li <tommycli@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759829}