blink/images: Ensure cached PaintImage invalidated on alpha change.
BitmapImage caches the last generated PaintImage which is updated only
if more data for the image is loaded. The PaintImage stores the alpha
state for the image which may change once it has been decoded. Ensure
that this cached PaintImage is invalidated if there is an alpha change.

R=wangxianzhu@chromium.org

Bug: 1063679,739339
Change-Id: Icbe2eda8861d71b6cf200e939982293635e520a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2114582
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Auto-Submit: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759832}