webview: Fix exception handling in evaluateJavascript.
This case was converted to use PostTask along with all the other cases
in the file, but because the PostTask queue is managed in native (once
native is initialized) this fails to achieve the original goal of
posting it described in the comment: to make sure there is no native
code on the stack, so that we can handle the app's callback throwing an
exception without crashing in native.

Introduce an AwThreadUtils class that contains specifically-documented
methods for posting tasks in the desired way, and use it in
evaluateJavascript and other methods which want this exception-handling
behaviour, to make it clear that using Handler directly is intentional
and necessary.

Fixed: 719396
Bug: 944437
Change-Id: I95cc029801e632a22987eb92e00cc28d4c292332
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2144461
Commit-Queue: Richard Coles <torne@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Auto-Submit: Richard Coles <torne@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759845}