Reland "Android: Fix a bug of text selection not working in SERP"
The previous CL was broken due to a bug not updating the currently
active tab ahead of triggering the init-web-contents observer
manually. This CL ensures the active tab is updated first.

The actual diff showing the bug fix is
https://chromium-review.googlesource.com/c/chromium/src/+/2151377/1..2

TBR for WebContentsUtils as there's no change in that class.

TBR=dtrainor@chromium.org
Bug: 1070516, 1067524, 1070542, 1067524

This reverts commit dc423f43d12a7a8b5df4208cafb3d1adfba54de6.

Change-Id: I8cdbe210256d49483e39ccf2f7060af0a23b1487
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151377
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759861}