Removed previous version of light estimation API
Will shortly be replaced by a more up-to-date variant of the API, but
due to some conflicting interface names it's going to be cleaner to
remove the current blink code (minus a couple of stubs that I know the
upcoming code will need). XRCubeMap specifically will be sticking around
as an internal utility, but won't be web exposed with the new API shape.

Bug: 1070819
Change-Id: I40f752c8caebf6de49d1abc1d8385857389a5eee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2152954
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759903}