chrome/browser: Add flag to enable/disable Zero-Copy Video Capture
Zeo-Copy Video Capture is controlled by ChromeOS overlays [1, 2].
This adds the chrome UI flag so as to switch the feature via UI.
The default behavior respects the overlay settings. The feature
is forced to enable and disable once the flag is switched in
chrome://flags.

[1] https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1835900
[2] https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1835900

Bug: 982201
Test: Disable/Enable the feature in chrome://flags
Change-Id: Ic329b3c01122a1ba3eac1de9363c3c2b3c0e139d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2143093
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759925}