Don't surface download service API background downloads in extensions API
This CL prevents downloads by the internal download service from being
shown in the extensions API. The motivation comes from running the
Plugin VM installer while using ChromeVox -- the download progress is
reported to the user with a guid as a name, which is a bit confusing
and clashes with updates from the installer's progress bar.

In general, it seems like these internal downloads shouldn't be exposed
to the extensions API. chrome://download-internals uses a different API
to access this data so it will continue to display this data.

Bug: 1021414
Change-Id: Idabbd4548a672fe9b773289a7809c32013dda56c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2143186
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759960}