system-web-apps: Adds ChromeOS specific properties to Web Apps
On most OSs these properties are stored in the OS itself, e.g. in
Windows an app is added to a folder so that it shows up in the
Applications Menu. For ChromeOS, we need to implement our own way of
storing this.

This CL adds a new property to WebApps: ChromeOSData, which
holds properties specific to Chrome OS and needed to integrate with
various UI surfaces.

Bug: 1054195
Change-Id: I23dab5aeb5f30c94d28db0062bb82b4f173047da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132900
Commit-Queue: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759962}