[iOS] Fix incorrect flag name for OmniboxOnFocusSuggestions
Because the name was incorrect here, when Finch sends a config for this
Feature, both the gws_id from Finch and the gws_id defined in
about_flags were sent back to gws. If an enabled and a disabled ID are
sent to gws, the disabled id is the one that takes affect, so the
feature ends up disabled.

Bug: 1071078
Change-Id: Ic67f1d5415ac5895a84ee41adc8b2c98d1c38456
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2153106
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#759983}