WebLayer: Close tabs on launching external intents when instructed
This CL augments WebLayer with closing of tabs when the external
intent launching code asks to do so (namely, when an initial navigation
in a Tab results in the launch of an external intent). To do so we
invoke BrowserImpl#destroyTab(). As clients < 84 are not necessarily
expecting the last tab to be closed by the WebLayer implementation, we
do this only when the client is 84+.

As part of this CL we looked at hardening WebLayer against the case
where the client calls Browser#destroyTab() from within
TabListCallback#onTabRemoved() (which itself is received after the Tab
destruction process has started in the implementation). It turns out
that WebLayer will already fail fast in this case:
Browser#onTabRemoved() sets the removed Tab's Browser to null, and
Browser#destroyTab() raises an exception if the passed-in Tab's Browser
is not equal to |this|.

This is a reland with a minor fix to update for changes in the parent CL
(fix shown between PS1, which contains the original CL, and PS4):
BrowserImpl#destroyTab() no longer throws RemoteException, so
InterceptNavigationDelegateClientImpl#closeTab() doesn't need to catch
it.

TBR=sky@chromium.org

Bug: 1031465
Change-Id: Id61177d1f4aa0ae6058322d5cb044c12b47f4d6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2152833
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#759996}