[throughput] Fix a OOM crash in FrameSequenceTracker
We are seeing OOM crash in FrameSequenceTracker::ReportSubmitFrame
and it is pointing at ignored_frame_tokens_ set being too big.

The only way this can happen is that a tracker is scheduled for
termination, but the last frame took very long time to get
its presentation feedback. During this period, there are a
lot of frames getting submitted, and these frame tokens will
be pushed into ignored_frame_tokens_ set and causing a OOM crash.

This CL fixes the problem and added a unit test.

Bug: 1070209
Change-Id: I0a9c132d8f9d45a5a89f1ff763b20d9ba12ea524
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2148164
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760119}