Fix inner viewport scroll animation update
Prior to https://crrev.com/2d6fe2bcbcb3f2, when a scroll animation was
created on the inner viewport, we would latch to the inner viewport
scroll node by setting it as the currently scrolling node. This wasn't
intentional as the inner viewport should never be latched; the outer
viewport represents all viewport scrolling so that scrolling code knows
to use the cc::Viewport class to correctly distribute between inner and
outer scroll nodes. This was part of the overloaded nature of
CurrentlyScrollingNode that the above CL and several follow-ups
addressed.

The above CL missed an important case: when a scroll updates an
in-progress inner viewport animation, it must use the animating
ScrollNode, the inner viewport scroll node. After the CL, an animating
inner viewport sets the CurrentlyScrollingNode to the outer viewport and
incorrectly uses that to update an in-progress curve.

In LayerTreeHostImpl::ScrollLatchedScroller, in the case of an existing
scroll animation, we call ScrollAnimationUpdateTarget on the
CurrentlyScrollingNode. This mostly doesn't cause an issue as the
animation knows which ScrollNode it's mutating. However, the passed in
ScrollNode is used to clamp the updated target offset which can lead to
incorrect scroll offsets.

In the linked bug, since outer viewport has no horizontal overflow, the
maximum scroll offset is 0 in the x-axis. Thus, when the inner viewport
animation is updated, any x component will be zeroed out due to
clamping, causing horizontal scrolling even if the update was only
vertical.

This CL fixes the issue by exposing and using the scroll node that the
animation is bound to when updating an animation curve.

Bug: 1070561
Change-Id: I7bf1993e1ffc4e06447dcb16e57f193c47fb2391
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151005
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760128}