All DumpAccessibilityTree tests expose the HTML element.
These were the final DumpAccessibilityTree tests
that needed to be migrated. They were previously
excluded because of an Android issue.

Move back the remaining tests to the correct place
in the file and get rid of the temporary subclass.

For context, this is an incremental step towards
eventually trying to include everything in the
accessibility tree, rather than having some
AXObjects that are "skipped".

Bug: 1063155
Change-Id: I6b9363f6cae4020a6263b9d9ada722834849ec94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2141219
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760149}