Update DNS.SB DoH entry
Split the entry into two, one for autoupgrade and one for UI selection,
and remove the no_ecs option from the user-selected one. That parameter
was originally added to maintain same-behavior on upgrade, but if it's
not an upgrade scenario, the parameter should only be used if added by
advanced users through manual URL entry.

Bug: 1065338
Change-Id: I52fd4d346595cff78681ccf6f29fd03013c815de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2154846
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: Eric Orth <ericorth@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760156}