Clean up PrintCompositorImpl.
PrintCompositorImpl ended up with 2 CompleteDocumentToPdf() methods,
which is confusing. The first is a required PrintCompositor interface
impl. The second is a private method that is only called by
CompleteDocumentRequest(). Since CompleteDocumentRequest() is just a
minimal wrapper around the private CompleteDocumentToPdf(), merge the
two and avoid the confusion.

Also remove some methods declared in the header but never defined.

Change-Id: I19832cb994a23a02dd7f18ad1757937576bd498c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2152919
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760163}