Move reconcileVisualState to be called when the TabModel is swapped
The current logic is setup in such a way that it's called along with
#setShowIconsWhenUrlFocused. The result of this is it's called more
often then it needs to be. Some calls are setup where status view icon
is the first focusable view. When the status view icon is set to GONE,
Android goes looking for the next focusable view and focused it. This
just happens to be the urlbar, which focuses the entire omnibox.

Moving this to the tab model changed callback guarantees that when
it's called, the status view icon won't have focus (either on the NTP or
the tab switcher).

Bug: 1063203
Change-Id: Ie9501da027b23133a4a4f2d78ed5c11225147d0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2138263
Commit-Queue: Brandon Wylie <wylieb@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760195}