arc-a11y: Drop selection/focus events from invisible nodes
There is an invisible list in PlayStore app, and when the loading
completed, a selection event is dispatched from the list.
This made ChromeVox move the focus to the list when opening PlayStore.

Before http://crrev/c/2129381, this was not observed because selection
event handling was incomplete, but now observable.

This CL prevents the behavior by dropping selection and focus events
from invisible nodes.

Bug: b:154293371
Test: unit_tests --gtest_filter="AXTreeSourceArcTest.*"
Test: manual. Open PlayStore, focus no longer moves to the invisible node.
Change-Id: Icfe4c931bef53b573878c3a55223e9e553b47307
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2154358
Commit-Queue: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Sara Kato <sarakato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760378}