[Android WebAPK] Add extra lightweight first run experience tests
This CL adds two new tests for the lightweight first run experience.
The motivation for the new tests is because the purpose of removing the
WebApkConstants.EXTRA_WEBAPK_PACKAGE_NAME in
WebappLauncherActivity#tryCreateWebappInfo() is to make the lightweight
FRE logic work correctly but the logic was not tested in the FRE
context.

BUG=None

Change-Id: I5958a7333722f10de1e3599b07d82be9e9e39446
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2147042
Commit-Queue: Peter Kotwicz <pkotwicz@chromium.org>
Reviewed-by: Glenn Hartmann <hartmanng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760389}