[base] Apply BindOnceRewriter in //ui
This change applies the BindOnceRewriter in //ui.
This rewriter replaces calls to base::Bind and base::BindRepeating with
calls to base::BindOnce, in case the result is immediately assigned to a
base::OnceCallback. Given that base::RepeatingCallback is implicitly
convertible to base::OnceCallback, there is no change in functionality.

Steps:
  1. run_tool.py --tool base_bind_rewriters \
                 --tool-arg='--rewriter=bind_to_bind_once'
  2. git cl format

This CL was uploaded by git cl split.

R=sadrul@chromium.org

Bug: 714018
Change-Id: I741a996064d7932749e6970992b333d0ae94cb5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132214
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760407}