[base] Apply BindOnceRewriter in //device
This change applies the BindOnceRewriter in //device.
This rewriter replaces calls to base::Bind and base::BindRepeating with
calls to base::BindOnce, in case the result is immediately assigned to a
base::OnceCallback. Given that base::RepeatingCallback is implicitly
convertible to base::OnceCallback, there is no change in functionality.

Steps:
  1. run_tool.py --tool base_bind_rewriters \
                 --tool-arg='--rewriter=bind_to_bind_once'
  2. git cl format

This CL was uploaded by git cl split.

R=reillyg@chromium.org

Bug: 714018
Change-Id: Ie566724abb5410a4d53751fa439972bf8db7bc65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132134
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760435}