[WebLayer] Create Java BrowserContextHandle in //components/embedder_support.
This CL introduces a BrowserContextHandle interface that provides a
pointer to a native BrowserContext, and can be used to by components to
take a BrowserContext reference within Java code. Clank's Profile class
implements this interface, which allows us to pass a Profile to
components that need a BrowserContext. In the future, WebLayer's Profile
representation will also implement this interface.

This change was motivated by the componentization efforts underway as
part of the WebLayer project, several of which would benefit from a Java
representation of the BrowserContext.

Bug: 1058600
Change-Id: I65778791731c8d1f1529f06eab543774a60155ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2130674
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760540}