PasswordManager: Use |typed_value| as a fallback if |value| is empty
There is a bug that the save prompt is not shown if the password value
is set as an empty value by JavaScript before form submitted. The bug is
a regression since crrev.com/c/2029514 (M81 stable), but it is enough
reasonable. So, this patch fixes FormParser instead of reverting the CL.

In the current implementation, FormParser is treated as parsing error if
|value| is empty. On the other hand, according to crrev.com/c/1404093,
if |typed_value| is present, we preferentially use it rather than
|value|. Similarly, this patch makes FormParser use |typed value|
preferentially even if |value| is empty.

Reproducible sites:
  - https://m.naver.com
  - https://zino.dev (This is a minimal test page I wrote)

Bug: 1042196
Change-Id: I6652822fd7fe268b54647434c23c68d41abf5010
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2154644
Reviewed-by: Vadym Doroshenko  <dvadym@chromium.org>
Commit-Queue: Vadym Doroshenko  <dvadym@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760543}