Bump theme pak version from 72 to 73 on M81.
This forces a theme rebuild to address bug 1070716.

The actual version 73 bump on trunk happened in
https://chromium-review.googlesource.com/c/chromium/src/+/2050827 ,
which was a followon change to
https://chromium-review.googlesource.com/c/chromium/src/+/2042376 ,
which is way too dangerous to merge to M81.  So in theory, this
introduces a discrepancy in the meaning of version 73.

However, the only way this could cause a failure is if someone ran this
build, then upgraded their profile to a build between r739661 and
r744105 inclusive.  Current and future beta/dev/canary builds are all
long past that range, so this seems like an extremely theoretical
concern.

Bug: 1070716
Change-Id: I18c62923a8a7b407b9a2eda62916e0e04450e057
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155827
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/branch-heads/4044@{#959}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}