[Fuchsia] Fix FuchsiaAudioRenderer to report decoder stats.
Previously FuchsiaAudioRenderer wasn't reporting any statistics to the
pipeline. As the result the pipeline was always reporting 0 for
audio_decoded_bytes. Fixed it to report this parameter after every
packet it sent to the AudioConsumer. Note that decoding itself is
handled by the AudioConsumer implementation, so the data may be decoded
a bit later than it's reported.

Bug: 1067068
Change-Id: Iecbce0d757b6c753360073faf5a073b71b33e1f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2148379
Auto-Submit: Sergey Ulanov <sergeyu@chromium.org>
Commit-Queue: Kevin Marshall <kmarshall@chromium.org>
Reviewed-by: Kevin Marshall <kmarshall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760561}