flags: rigorously define expiration milestone
There are two competing definitions of expiration milestone:
a) The last milestone in which a flag is present
b) The first milestone in which a flag is not present

The docs in //docs/flag_expiry.md and the code in //tools/flags
implemented (b), but the block comment in flag-metadata.md described
(a). Since that block comment has been much more widely read and
relied upon than either of the other two sources, this change makes
(a) the behavior everywhere. This will hopefully reduce confusion in
client teams about where their flags have gone :)

Bug: 1072433
Change-Id: I0f8fcbde2509b1bed05b98ac847c98595c74836f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155841
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760681}