[conversions] Add support for data deletion
This CL plumbs an Origin matcher through the ConversionManager and
ConversionStorage layers, and implements deletion using the matcher
in the SQL layer. The deletion follows the following algorithm:

1. Deletes all impressions within the time range. If any conversion is
   attributed to this impression it is also deleted.
2. Deletes all conversions within the time range. All impressions
   attributed to the conversion are also deleted.
3. Delete all conversions that are unattributed after (1) and (2).

In particular, time ranges that fall _between_ an attributed
impression/conversion pairs do not cause the pairs to be deleted.

A follow-up will wire this up with the deletion code in
StoragePartition.

Bug: 1045165
Change-Id: I1b666bc834782050d89abe47c4e616d37bf28d25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2150071
Commit-Queue: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760685}