Reduce external dependencies on //device/vr
In preparation for creating a content/public interface method to get
VRDeviceProvider objects, this moves that interface type to
device/vr/public/cpp, and adds a build target for it (since it is a
class).
Additionally, removes a bunch of includes of vr_device.h which were
no longer necessary (and should only be necessary within device/vr).

Bug: 1070807
Change-Id: If9e08a495d654821838d2cc57047e4c3c7cdc933
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155073
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760730}