Tracing sampling profiler: Output process priority in profiling packet
This used to be output to the ProcessDescriptor message which is being
phased out, and complicates the handling in the trace processor.
Instead, always just output this in the StreamingProfile message.

Requires https://android-review.googlesource.com/c/platform/external/perfetto/+/1287904
to roll in /third_party/perfetto first.

R=ssid@chromium.org

Bug: 1065077
Change-Id: I7a7764e09382e7584a82503267feb286d6a9df65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2153926
Commit-Queue: oysteine <oysteine@chromium.org>
Reviewed-by: ssid <ssid@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760753}