Windows: Hide MenuListPopup and MenuListOptions when collapsed
On Windows, a <select> element is replaced by a combobox that contains
an AXMenuListPopup parent of AXMenuListOptions. When the select dropdown
is collapsed, the subtree of that combobox needs be hidden and, when
expanded, it must be accessible in the tree.

Before this change, I worked for a while on a different approach where
the subtree of the combobox would be marked as ignored on Blink.
However, as the Try Bots showed me, this might not be an approach we
want on all platforms.

Prior to this CL, the subtree was always unignored and accessible and
that resulted in Narrator announcing all options of a select even when
it was collapsed. Now, the options are only announced when the
dropdown is expanded.


Bug: 928948
Change-Id: Iadfcfe8d14fc72de6941fc13b459070d6f6746de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142134
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760777}