WebUI NTP: OneGoogleBar, hide backdrop before updating z-level
When the OneGoogleBar OGB is layered on top of the NTP content, a
backdrop is shown. It is desirable to show the backdrop only after the
OGB is on top of the NTP content, otherwise the NTP will appear on top
of the backdrop. The backdrop should likewise be hidden before the NTP
content is layered on top of the OGB, which is what this CL fixes.

Bug: 1039913
Change-Id: I7883c47ef7dce49c4b927f62f254500d72307fcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2157993
Commit-Queue: Esmael Elmoslimany <aee@chromium.org>
Auto-Submit: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760779}