Android: Another fix for the preview tab on CCT
Preview tab initialization for CCT was not obtaining a properly
instantiated BottomSheetController, therefore it didn't work. This CL
passes the supplier instead of the controller itself to handle it.
This has an unintended but positive effect of making the Preview Tab
coordinator ctor lighter. An observer for the BottomSheetController
is added when necessary, and removed after use.

Bug: 1070490
Change-Id: I453b90f71c8766d6a7cb033c4184156c611742b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151770
Reviewed-by: Donn Denman <donnd@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760785}