[Kiosk] Reset A11y prefs upon entering kiosk session
Currently, kiosk session's preferences are not ephemeral and are
preserved during several kiosk app launches. Main usage of a11y
features in kiosks assumes that there is a different user each time the
session is started. It will be a mess, if one user could tweak a11y
prefs so badly, so the next user is not able to efficiently interact
with the device.

Bug: 1049566
Change-Id: I6751fb049cffb78d76c5e513c4fcf04e04e579d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134009
Commit-Queue: Anatoliy Potapchuk <apotapchuk@chromium.org>
Reviewed-by: proberge <proberge@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760790}