DNR: Resolve race between initial ruleset load & dynamic rule update.
As of now, a dynamic rule update can race with an initial ruleset load
in response to OnExtensionLoaded, and cause a DCHECK in
RulesetManager::AddRuleset. Resolve this by ensuring that we queue any
dynamic rule update requests to happen only after the initial ruleset
load is complete.

BUG=931967

Change-Id: Ib87aa660910478e7cb9fb7841ca59d6789dc9eb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151895
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760798}