[css-text] Ignore trailing spaces when rewinding overflow
In r734897 we added some changes in the RewindOverflow to avoid removing
the last space, which should hang instead, when 'pre-wrap' is used.

However, this change caused a few asserts violations that indicate an
incorrect execution flow.

Unfortunately, I couldn't define a proper reduced case to analyze with
more detail under which conditions these asserts are violated; the only
test cases available are the ones provided by the Fuzzinator, which are
impossible to reduce.

Hence, this CL is just a speculative fix.

Bug: 1064189, 1070082
Change-Id: I27f9b14afbd7f71ed4ff0575191fc84ea7acc094
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2157130
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760800}