[ios] Handle EditBookmarks pref in ui/bookmarks/
EditBookmarksEnabled is an enterprise policy that determines whether
the user can edit any bookmarks. The policy is mapped to a user pref.

This is part of a series of CLs that add support for this
enterprise policy on the iOS platform.

Specifically in this CL, the bookmarks UI is modified to disable
editing when EditBookmarksEnabled is false.
- Swipe-to-delete on bookmarks and folders is disabled
- Long press on bookmarks and folders disables some options
- New Folder and Edit toolbar buttons are disabled

A unit test will be added in a followup CL.

Bug: 1072326
Change-Id: I11e6f85fde5bdae642c576ae0e0a74651ead0b88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2156140
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760806}