[Launcher settings] Use settings app icon for search results.
We want to use the icon for the OS settings app for all results. This
CL uses the app service to grab the icon when it's available and adds
it to search results.

The logic is in two parts:

1. Make the search provider observe updates from the app registry
   cache. This lets us know when the settings app has been loaded
   and its icon is available.

2. Then, use the app service proxy to request the icon, and store it in
   the OsSettingsProvider.

Note the user won't see an icon for results returned before the app
service is ready. I don't think there's much we can do about this.
But, the app service loads so quickly on login that I don't think
a user could realistically search fast enough for this to be a problem.

Bug: 1068851
Change-Id: I6d2ad362db9b0d9ca0c8a3930511849a43da25e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151496
Reviewed-by: Jia Meng <jiameng@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760818}