[LocationDelegation] Add flag and store and get geolocation permission
This CL is part of the implementation of Location delegation:
1. Adds the feature flag kTrustedWebActivityLocationDelegation.
2. Adds storing geolocation permission and also getting the RuleIterator
via the InstalledWebappProvider for location permission, similar to
the notification permission.

This CL makes TrustedWebActivityPermissionManager and
TrustedWebActivityPermissionStore takes a @ContentSettingsType int
argument when setting and getting permission states, so we can reuse
most of the logic between Notifications and Geolocation.

Bug: 1069506
Change-Id: I86ed7b144c7041ca7d9dfa59d8a0037469df9d30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2130675
Commit-Queue: Ella Ge <eirage@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760856}