[iOS] Fix crash due to multiple completion attempts on interrupt.
On interrupt action for adding an account during the user sign-in flow
we currently perform sign-in completion (1) asynchronously in the add
account mediator during clean-up and (2) in the interrupt action method.

Ensure that we only process the sign-in completion once and that the
interrupt completion is called afterwards.

Bug: 1005509
Change-Id: Id96157b68e9173be07758949f1a36914f0e95906
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2154948
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760891}