[AF][Nickname] Add suggestion selection duration metrics for nickname
Add sub-histogram CreditCard.WithServerNickname.SelectionDuration to
log suggestion selection duration between the first time suggestion was
shown and the first card is selected when any server nickname is available
to see whether having a server nickname available helps user reduce the
time spent on selecting the card they want.

Note that same as the selection rate, we would like to compare selection
duration only for the same group of users who have server nicknames,
which means for the Control group we will also log selection duration
if the user has any masked server cards with a nickname, but we won’t
display the nickname in the UI. For Enabled group, we will log and show
nickname when nickname is available.

Add a timestamp and boolean to credit_card_form_event_logger to
record the first suggestion shown timestamp and to ensure the selection
duration is only logged once per form.

Metrics design (googlers only):
https://docs.google.com/document/d/1SfrBduc3gJrouT4zLmQHWtUPtcE-9JbkBGLHYgFLTFc/edit?ts=5e50549a#heading=h.olagc6jelr42

Bug: 1059087
Change-Id: I0ac4eb8e758d693a8b8240144655d53c7560a259
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151832
Commit-Queue: Sujie Zhu <sujiezhu@google.com>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Siyu An <siyua@chromium.org>
Reviewed-by: Jared Saul <jsaul@google.com>
Cr-Commit-Position: refs/heads/master@{#761019}