[bluetooth] Refactor BluetoothAdapterFactory interface
This change is preparing for a refactor of the BluetoothAdapterFactory
internals. Those changes will move the factory state out of globals and
into the singleton instance of BluetoothAdapterFactory.

As a result it is easier to make methods like GetAdapter() non-static
but this requires updating all callers.

The return value of BluetoothAdapterFactory::Get() is changed from a
reference to a pointer because the style guide recommends pointers
for non-constant values.

Bug: 1030549
Change-Id: Ie6da74fa5d17c586cbbd306f5a1e444b9d39260e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155465
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761047}