[WebLayer] Split SiteSettingsClient.launchHelpAndFeedbackActivity into two methods
This CL splits SiteSettingsClient.launchHelpAndFeedbackActivity into
launchSettingsHelpAndFeedbackActivity and
launchProtectedContentHelpAndFeedbackActivity methods. The method
originally took a "helpContext" string, which mapped to one of Chrome's
constants defined in //chrome/android/java/res/values/values.xml,
which were used to determine which Chrome help page to launch. These
constants are specific to Chrome, and shouldn't be part of a
//components API. Splitting the method into two allows the embedder
to choose which help page to show the user.

Bug: 1058595
Change-Id: I193dd6c0767fe53223afaa952f311d0fa045c6c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2158408
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761054}