Replace eventSender.mouseScrollBy with gpuBenchmarking.smoothScrollBy Part2
After we finish the scroll unification, the scrolls happens mainly on
the compositor thread, and the scroll code in the main thread will be
removed. eventSender sends the scroll events to main thread, so it
would not work after the scroll unification. We should replace
eventSender.mouseScrollBy with gpuBenchmarking.smoothScrollBy
in all the layout tests, this is part 2.

I remove two tests in this CL because they no longer work after scroll
unification that wheel scrolling happens on compositor.

Bug: 1047176
Change-Id: I954a97175fb4102a1d70b21fc7217c9400ba5671
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2131296
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761079}