[throughput] Fix a OOM crash in FrameSequenceTracker
There are OOM crashes in FrameSequenceTracker::ReportSubmitFrame,
where the ignored_frame_token_ is getting too big.

This could only happen if a tracker is somehow never terminated,
and thus not getting removed.

This CL is a temp fix that if a tracker is scheduled for termination,
and still receive 3 frame submissions, then we terminate this
tracker right away.

Following this, we will add debug strings to find out why is it
not terminated.

Bug: 1070209
Change-Id: I5a28cb0f52abf44e76211b87f8d8a6b3c1657fdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2154498
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761084}