[Extensions] Remove NOTIFICATION_EXTENSION_COMMAND_[ADDED|REMOVED]
The NotificationService is deprecated. Remove two notifications from it:
NOTIFICATION_EXTENSION_COMMAND_ADDED and
NOTIFICATION_EXTENSION_COMMAND_REMOVED. These (as the names imply) were
used to notify of when an extension command was added or removed, but
can be substituted with the CommandService::Observer.

This also fixes a bad cast bug.

Bug: 411569
Change-Id: I7b2f0a1cf006a80d62571d20cf8a20968f2d0ac0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155248
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761093}