Update and enable WebNavigationApiTest.Crash test.
This test was disabled on some bots due to flakiness. I am updating the
test to explicitly wait for process crash, which will likely eliminate
the source of flakiness and updating some of the code to take advantage
of newer test methods instead of using base::StringPrintf.

Bug: 974787
Change-Id: I68182c88905f187b972d93da3504e701a82b96d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159318
Reviewed-by: Aaron Colwell <acolwell@chromium.org>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761113}