Cros: Add feature flag for new drag spec
The new feature will enable the new app dragging functionality in the
launcher. When the user drags an app within the launcher, the apps grid
pages will be scaled down to 84% and will show a background card to show
them in a more modular way.

Bug: 996822
Change-Id: I993693f7ada613760384c981792b57e08ac8f3dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159756
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761125}