Update browser history on portal activation
When a portal contents is activated, we swap it with its predecessor,
so to the user it appears that the tab has navigated. However, we did
not notify the history service about this, so the page the user is now
viewing does not appear in chrome://history/.

Now upon portal activation, we have the history service add a page visit
for the portal's last committed entry.

Bug: 944198
Change-Id: Iebb0c66a2afddefc5172c84158ea2d01511c8309
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080832
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Lucas Gadani <lfg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761141}