Prevent event-bubbling for Quick Answers views
The child buttons inside the Quick Answers view, which is itself a
button, do not work properly because of the way events are bubbled down
the view hierarchy for Quick Answers views.

To prevent that, this change will modify the event-handling of the
related view by propagating events in a bottom-up fashion instead and
stopping at the first child to handle the event to avoid simultaneous
firing of container buttons.

Bug: b:152922056
Test: On Chrome OS VM.
Change-Id: I5e309cf9cecf88e79651a138375025b7120ab2eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2140995
Commit-Queue: Abhijeet Singh <siabhijeet@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761176}