Support spec changes to wrap media feed items with broadcast event.
The schema.org metadata should wrap the entire item in a broadcast
event now instead of embedding the broadcast event inside the item.
Our mojo structs stay the same and place the corresponding LiveDetails
in the item, or in the main episode if one exists.

Bug: 1071657, 1065986
Change-Id: Id973bed4574212763cf6741be4d4b747cef54d72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2158047
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Commit-Queue: Sam Bowen <sgbowen@google.com>
Cr-Commit-Position: refs/heads/master@{#761192}