Protect against crashes in TabHoverCardBubbleView::ClearPreviewImage()
It's not 100% clear from the stack trace, but the result of
GetThemeProvider() can be null and was not being checked for null.

In addition, we now have a DCHECK() that ensures the method is only
called when the thumbnail_image_ view is actually present.

Bug: 1070980
Change-Id: I206d22ad7687184887c890c8104edc82480a38fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159920
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761223}