Trust Tokens: Use the three "oldest" keys for issuance and redemption
This change adds a parameterizable number K of Trust Tokens keys
such that issuance and redemption operations use the K "oldest"
(soonest-to-expire-but-unexpired) keys available. This allows issuers
to provide more than K keys in their key commitment responses with the
knowledge that the client will be able to continue on to new keys once
old keys expire.

Example: if K=3, it's currently time T, and an issuer's commitment
result contains
  K1 expiring at T-1 minute
  K2, K3 expiring at T+1 minute
  K4 expiring at T+2 minutes
  K5, K6, K7 expiring at T+3 minutes

In order to ensure no more than K keys are concurrently active, we break
ties by keeping the key bodies earliest in lexicographic order. Assuming K1's
body is "K1" and so on, we'd get the following behavior:
  At time T, the keys we use for issuance and redemption are K2, K3, K4;
  at T+1: K4, K5, and K6;
  at T+2: K5, K6, and K7.

R=csharrison

Bug: 1063510
Change-Id: I709feac7c17e9d3621e5f7d2e7b134281dc9a96d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159306
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761252}