LocalDOMWindow::GetProbeSink() should go straight to the frame, rather than through Document.
LocalDOMWindow::GetProbeSink() can get called in the narrow window when
LocalDOMWindow is between Documents, when a feature policy header parse
error is sent to the console during Document construction. In that case,
LocalDOMWindow has a valid LocalFrame* but not a valid Document*, but it
should still be able to log to the console.

Bug: 1073037
Test: http/tests/inspector-protocol/log-feature-policy-parse-console-message.js
Change-Id: Ic47d148e1c8f6a1d15acf54702e47cc3e9a834fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159887
Auto-Submit: Nate Chapin <japhet@chromium.org>
Commit-Queue: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761254}