appcache: remove use of unset group token_expires
The group's token expires was removed in this patch:
  https://chromium-review.googlesource.com/c/chromium/src/+/2157602

...and so it's incorrect to set the cache's token_expires from the
group record (which is unset).

The SQL and token_expires is left on the records and in the tables as
risk avoidance, as the testing situation for SQL is not thorough here
and it is safer to leave them as-is.

Bug: 582750
Change-Id: I53b8eb08bffb7a30d42c34eafb983fe5b060dcae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2160085
Commit-Queue: enne <enne@chromium.org>
Commit-Queue: Victor Costan <pwnall@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761255}