Add end to end IsolatedPrenreder tests with a CONNECT proxy
Adds a test utils class which has the proxy tunnel behavior, which is
grafted on top of the embedded test server but is not a full server
in itself.

Adds an end to end test with prefetching through the proxy tunnel, as
well as an additional test that authentication challenges from an
origin (through the tunnel) are not shown to the user.

The TabHelper also gets an Observer which is only used for testing
right now, but is left as public prod code in case it is useful later.

Bug: 1067300
Change-Id: Ia3c50aad2b94ead31044770b91edb95c8a821910
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2153753
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761301}