[@property] Add PropertyRegistry::Version
The version is simply a number that increases monotonically for each
change to the registry. This is needed by the animations machinery to
handle cases where the type of a registered custom property changes
on the fly during an ongoing animation. (See InvalidatableInterpolation
::EnsureValidInterpolationTypes). When the version changes, we can
clear the conversion cache, which causes the keyframes to be
reinterpreted according to the new registration.

Also move some utils from at-property.html to utils.js, for reuse
in other test files.

Bug: 973830
Change-Id: I97d0a37f016ccde649fe1482084707edd123903b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159212
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761323}