Align generation of keyframes for a CSS animation with the spec
The patch brings the construction of keyframes for CSS animations more
in alignment with the spec. The initial keyframe generation process from
the keyframes rule remains largely unchanged, but pulled into a
separate method for readability. The remainder of the algorithm is in
fairly close agreement with the spec.  Remaining issues are the handling
of shorthand and logical properties.  In the process of aligning the
implementation, the handling of timing functions was addressed.

Bug: 1070627, 1068639
Change-Id: I4a324abcc30bd8085e0cbcb583ef9a860f58a6a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2152996
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761399}