Implement CopyImageBitmapToTexture part II : Enable GPU T2T copy path
This patch added the path that uploading imageBitmap contents to WebGPU texture if the backed resource
of imageBitmap is on GPU.

It leverage shared image system to share content from image bitmap to a dawn texture as intermediate resource
and than starts a encoder and submit a T2T copy.

But this path has lots of constraints and may failed. When it is not possible to using GPU uploading path, the API
will fallback to CPU path.

BUG=966582

Change-Id: I46da847c1632dcd2e39507ced71f7cf7bc4b60e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2094002
Commit-Queue: Shaobo Yan <shaobo.yan@intel.com>
Reviewed-by: Austin Eng <enga@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761403}