heap: Only query liveness through LivenessBroker
This CL removes all places that directly query mark bits of objects
and instead makes them go through LivenessBroker. Except for tests
this is compatible with the library design where the broker object is
created by the heap internally.

The tests that create the broker manually have to move to the library
eventually.

Bug: chromium:1056170
Change-Id: Ib76f6dfe2fc67300a953120d5f7497db6161b777
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2153114
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761419}