Fix LiveTab -> IOSLiveTab cast
The faulty code assumes that every LiveTab is an IOSLiveTab.
This is no longer true now that RestoreIOSLiveTab exists.
Rename IOSLiveTab to IOSWebStateLiveTab and create a new
abstract IOSLiveTab that will be inherited by both classes.

The result of the bug is a read after the limit of the object.
The result value is only used for a lookup in a table, so there should
be no memory corruption after this faulty read.

Bug: 1073038
Change-Id: I9db37a970c09e1f76f72118a44f165e3f4b749d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159477
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761447}