[identity] Remove gaia id from IdentityAPI at the old consent flow
This is a reland of https://crrev.com/c/2152795 with a modification to
clear the gaia id cache as soon as possible, at the beginning of the old
consent flow. This version is better because the user doesn't have to
complete the flow to clean the state.

Original description:

It may happen that the user switches back to the old consent flow after
have passed through the new flow. The old flow doesn't communicate back
to Chrome which account has been chosen by the user.

Gaia id cached by IdentityAPI may contain incorrect information at this
point, so it'd be better to clean the cache.

This CL also makes Chrome send the full version number in the "lib_ver"
parameter so Gaia can determine whether the current version contains
the fix.

Bug: 1026237
Change-Id: I90e96403cce68078dc18da52df3f9456e1d6afb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2160967
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761449}