Fix coordinate conversion for OOPIF compositing rect
ViewportIntersectionState.viewport_offset represents the offset of an
iframe from the top/left of the top-level document, but
RemoteFrameView::UpdateCompositingRect was treating it as the offset
of an iframe from the local root view. This is fixed by adding the
kApplyRemoteRootFrameOffset flag to AncestorToLocalRect, which fixes
the linked bug.

The above change exposed two additional issues:

- When FrameView::UpdateViewportIntersection computes the value of
viewport_offset, the result needs to be in the coordinate system of
the iframe contents, including any transforms applied above the
iframe. This patch switches from LocalToAbsolutePoint to
AbsoluteToLocalPoint for that reason.

- SitePerProcessBrowsertest.ScaledNestedIframeRasterSize was using
faulty arithmetic for its expected value. This was hidden by the
combination of both issues above.

BUG=1071666

Change-Id: I5f015d62044e81711814a3fb1892a64409b3ae99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159912
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761484}