[AF][Nickname] Add digit check for nickname validation.
In Gpay app, We don't allow user to use any digits within the nickname
to avoid user saving card number/cvc as card's nickname. We also add this
check in Chrome for consistency. If we see any digit in nickname, we
will treat it as invalid nickname.

Bug: 1059087
Change-Id: I1d830e549957a225519e59511cd8c581baeab298
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2160009
Commit-Queue: Sujie Zhu <sujiezhu@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Jared Saul <jsaul@google.com>
Cr-Commit-Position: refs/heads/master@{#761485}