Check Search and Assistant state through HTTP endpoint
We used to get the Search and Assistant (SAA) bit from the Assistant
Server, which means once the SAA is disabled we cannot know when it get
enabled again. Also the delay and unstability of the Assistant Server
brings up corner cases and timing issues.

This change replace the SAA bit check with request to the mothership
HTTP endpoint, which makes the state check faster and more stable.

Bug: b/144870198
Test: Manual Test
Change-Id: If938a2c811b85a6ec00e0e9e4ac09670f1f8583b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2150062
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Yue Li <updowndota@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761509}