Handling critical updates for MinimumChromeVersion Policy.
This CL adds the handling for critical updates (warning time equal
to zero) for the policy MinimumChromeVersionEnforced. In this case,
the policy handler firsts checks for end of life status of the device
to find out if it has been reached. In this case, we use
eol_warning_period instead of warning_period.
If warning period is zero, the user is logged out of session and/or
shown update required screen on the login screen. To handle the
external dependencies, the minimum_version_policy_handler_delegate
class is used to make the policy handler easier to test.
This CL also adds unit and browser tests for the critical updates case.

The CL does not fully implement the policy. It only handles the case for
critical updates (warning time is zero) in-session and if the login
screen is held for long enough for policy to be updated.
If user is in-session, he gets logged out and if is on the login
screen, update required screen is shown.
Subsequent CLs will handle other cases of this policy like network
limitations, triggering update process, showing in-session notifications
and showing the update required screen right after reboot if there is. no
warning time.

Bug: 1048607
Change-Id: Ie71a7b283cfb390b12f1377c0248ca3da2d0f87d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2144078
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761535}