Sort CSS computed style properties in lexicographical order, again
This is a follow-up of r760732, where I included logical properties when
iterating a computed style declaration. The problem was that I inserted
some of them at the wrong position, breaking the lexicographical order
achieved in r760690.

This time I add a test that checks that the order is right.

BUG=1072722,1072180

TEST=external/wpt/css/cssom/cssstyledeclaration-csstext.html
TEST=external/wpt/css/cssom/getComputedStyle-property-order.html

Change-Id: I3ea31024deed8b5a4c67a5a117c358cab6c9d7d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2160913
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/master@{#761556}