Make webgl_conformance_integration_test support using SwiftShader ICD
When setting VK_ICD_FILENAMES to point at the SwiftShader ICD, and
running webgl tests with --use-gl=angle --use-angle=vulkan, ANGLE will
return that the actual backend is 'swiftshader', which fails the backend
validation check in this script. Modified so that this case works as
expected.

Change-Id: Ic427a6737a7a239ff793322f54fcd69645c20464
Bug: none
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2160109
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Antonio Maiorano <amaiorano@google.com>
Cr-Commit-Position: refs/heads/master@{#761645}