[WebLayer] Reload navigations on transient errors
This copies more of the logic from chrome's NetErrorHelper that deals
with auto-reload on failed navigations. This should prevent the error
page from showing up when changing from wifi->cellular. This logic is
copied for now so we can merge to M83.

Deduplicating this logic is tracked in crbug.com/1073624.

Bug: 1072453
Change-Id: If623a563e8748a598b9430716d145c92447257be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159800
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761685}