Add quick answers logging for active impression.
Active impression is a Quick Answers result impression that fulfills a
user's intent.

For Quick Answers rendered along with the browser context menu, the
impression that the user doesn’t click on other menu items is considered
as "Active Impression".

This CL add a new API to |RenderViewContextMenuObserver| that is invoked
when a command is about to executed, including commands that are not
supported by the observer. The API is used to determine whether a quick
answers impression is a active impression.

Histogram recorded to understand a quick answer active impression and
impression duration:
* QuickAnswers.ActiveImpression (enum="QuickAnswersResultType")
* QuickAnswers.ActiveImpression.Duration - Aggregated duration across
  result types
* QuickAnswers.ActiveImpression.Duration.[result type] - Duration for
  each result type

The result type is something like "UnitConversion".

Bug: 1072907
Test: manual test
Change-Id: I5563468f38fad679e5175d65a6a2a8764716ed8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159927
Commit-Queue: Li Lin <llin@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761706}