Cros: Scale down apps when dragging.
This is part of the new drag spec in the launcher.
This change triggers the following actions when the user drags an app
inside the launcher:
- Define a new state in the apps grid view: Cardified.
- Scales down apps in the launcher to 84% of its size. The dragged
  view will remain with its current behaviour (scaled to 120% of its
  size)
- Reduce the padding between app tiles proportionally to the scale.
- Add 16px of padding to the left of each page according to the UI spec.

When the user ends the drag, the Layout of the apps grid will be
animated back to normal.

This functionality is hidden under the app_list_feature
kNewDragSpecInLauncher.

Bug: 996822
Change-Id: I04766849253d2f3689dc491417bf8269df515150
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2161694
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761726}