Introduce ui::TEXT_INPUT_TYPE_NULL.
The new text input type is introduced to represent the state that a text
input with rich editing like composition or candidates is not focused,
but the focused view wants to receive key events.
While this text input type is used, the virtual keyboard should show up
when the focused view wants and provide a way to generate KeyEvents.
It corresponds to Android's InputType.TYPE_NULL[1].

[1]: https://developer.android.com/reference/android/text/InputType#TYPE_NULL

Bug: 1065365
Change-Id: Ieacc448abf916b53e128b3f0dec8b363d10a8977
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2157442
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>
Reviewed-by: Keith Lee <keithlee@chromium.org>
Commit-Queue: Yuichiro Hanada <yhanada@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761773}