Ensure PermissionRequest::Cancelled() is called for dropped requests.
Prior to this CL, when a tab was closed, all in-flight and pending
permission requests were finalised and deleted. However, this did not
call the decision callback in the PermissionRequest, causing problems if
that decision callback eventually wrapped a Mojo callback - which DCHECK
fail if destroyed without being run.

This CL ensures that the decision callback for in-flight and pending
permission requests is run prior to the finalisation and deletion of
those requests at tab close. Minor fixes to tests are also made. This
(speculatively) fixes the DCHECK failure.

BUG=851030

Change-Id: I3109ca0fb026c40d20bc36da1632c31b2e234ca7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2153179
Commit-Queue: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761818}