Return detailed error from SafeManifestParser
A new struct ManifestParseFailure is created which contains detailed
error that occurred while parsing instead of just the error string.
We store the error in UpdateManifestResult in this CL so that it is
easy to group them by ids later where we can handle the errors and
successful parsing in a single list.
It is a pre work for reporting manifest invalid detailed errors.

Bug: 1063011
Change-Id: I2c87193c989488c3984f04e7224ccdd1488bba50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2126688
Commit-Queue: Swapnil Gupta <swapnilgupta@google.com>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761852}