[cronet] Disable remoting when building cronet
Stop compiling remoting app at the same time as cronet. This is
not needed (since remoting is build with Chrome bots, not cronet).

This will reduce the build time of cronet bots (less target to
build) and the complexity of remoting target (no need to support
fat build which is only enabled for cronet).

Bug: 1073377
Change-Id: I2e3213234c7dc97137ad857c802b542b278727ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159719
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761869}