Componentize ChannelsInitializer
ChannelDefinitions is made into an abstract class that exposes
functionality to ChannelsInitializer, and defines useful types like
PredefinedChannel. The channel definitions themselves live in the
Chrome subclass, ChromeChannelDefinitions. Some of these definitions
will be duplicated later for WebLayer.

ChannelsInitializerTest uses ChromeChannelDefinitions so it's left
in //chrome. (We could replace ChromeChannelDefinitions with a
TestChannelDefinitions, but that would reduce coverage of production
code.)

Both the channel IDs and the channel group IDs will have to be different
for WebLayer. It would be possible to account for this in code by
judiciously prepending "org.chromium.weblayer.", and then keeping
the channel definitions together, but that was both more complicated
and seemed more error prone, as it would mean that the
ChannelDefinitions.ChannelId type would no longer represent the Android
system notification channel identifier. For example, code that used
NotificationManagerProxy.getNotificationChannel() would have to be
careful *not* to pass ChannelDefinitions.ChannelId without adding a
prefix. In Chrome, which has no prefix, this would succeed, but in
to-be-shared code, such as NotificationCompatBuilder, this would succeed
for Chrome and fail for WebLayer.

Bug: 1069895
Change-Id: I55890466797aafc34ef8a821a1cb43de8e3d03ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2150066
Commit-Queue: Evan Stade <estade@chromium.org>
Auto-Submit: Evan Stade <estade@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761922}