[Android] Remove no longer needed punting of webapp interstitials to browser
Currently if a webapp shows a safe-browsing interstitial, the webapp is
closed and the page is launched in browser mode. The purpose of this
logic was to work around bug where Chrome could not display multiple
interstitials (e.g. interstitial in webapp and interstitial in Chrome
browser). crbug.com/293717 Based on testing, the underlying bug is fixed
and the workaround is no longer needed.

BUG=1073748

Change-Id: If1aedb38d7c6c5eed08e3de2c61ed72818ce1244
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2161453
Reviewed-by: Peter Conn <peconn@chromium.org>
Commit-Queue: Peter Kotwicz <pkotwicz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761956}