Fix double zoom bubble on Mac full screen
On Mac, the toolbar can be shown or hidden on full screen. The zoom
bubble changes its anchor position when the toolbar visibility
changes. We can not simply refresh the shown Zoom bubble if its anchor
position changed. So we need to allow Zoom bubble to be closed in that
case. Otherwise, double zoom bubbles may be shown as described in the
bug.

BUG=1072353

Change-Id: I5abc492e81b0e207556d528de3635e85f6bce9d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159315
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Wei Li <weili@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761972}