Guard some unit test code against an upcoming SkBitmap bug fix
This code was relying on a bug: That SkColors passed to erase were
treated as being in the bitmap's color space, rather than sRGB. (In
every other context, they are interpreted as sRGB). Fixing that bug
in Skia breaks this one unit test.

To keep the current expectations valid, rewrite the code that fills
the bitmap to clearly do what was happening before - draw rectangles
with the test colors, accompanied by the bitmap's color space, to
prevent any color space conversion. This version of the test passes
now, and continues to pass when the Skia bug is fixed.

Bug: skia:8663
Change-Id: I3a7e94847f08472081390b23fcfe0a12d8e5a38f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2161617
Reviewed-by: Mike Klein <mtklein@google.com>
Reviewed-by: Leon Scroggins <scroggo@google.com>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Commit-Queue: Brian Osman <brianosman@google.com>
Cr-Commit-Position: refs/heads/master@{#762007}