WebContentsObserver::OnServiceWorkerAccessed
Instead of passing WebContentsGetter to ContentBrowserClient, add explicit
callbacks to WebContentsObserver taking RenderFrameHost* or NavigationHandle*
as a parameter.

This allows to attribute cookie/javascript accesses to the correct
frame or navigation request, which is a prerequisite for attributing
content settings to the correct document and making them per-document.

R=falken@chromium.org,engedy@chromium.org
BUG=998171,1061899

Change-Id: Iec55a1db235b26bab44f1c1de76900b73ac6614d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151866
Commit-Queue: Alexander Timin <altimin@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762018}