edid_parser: Don't log parse failure.
This is not critical enough to log at the ERROR level for all Chromium
builds, and in practice it ends up appearing when a tip-of-tree Chromium
build is run under Chrome Remote Desktop.

There are already also UMA histograms that gather this in the field.

Change-Id: If272b274d6a57f514e11c9b81ba2a0217cc2bbfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2156602
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762050}