views: add DialogClientView button row layout test
This change adds a new test that covers the behavior of the button row,
as a prerequisite to a refactor of the button row's implementation. This
test does overlap with some of the other tests in this suite,
especially the button width linking tests.

Bug: 1011446
Change-Id: Ib7b60dbed5725f6e26a063bd851911d85cba53ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163394
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762069}