Remove browser-unpin-trace key.
As far as I can tell, it hasn't been used usefully in investigations
in years, and StackTrace members can cause problems on Android
(https://crbug.com/1056216 -- admittedly due to bugs elsewhere,
so removing this is papering over symptoms in a way. But if this
is unused, we should remove it for that reason anyways).

Bug: 1056216
Change-Id: I32e4f468a9f8dd5cb6ff4ffb2938f25c86af7530
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163261
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762154}