[ozone/x11] Enabled drag and drop.
This is the first step towards enabling drag and drop in Ozone/X11 using
the components extracted in previous patches.  Here the core part of the
functionality is enabled.

As Ozone allows selecting the platform at run time, it is necessary for
it to select the proper OS exchange data provider at run time too.
In this CL the new platform method is introduced that creates the
provider, so the provider factory now asks the platform to instantiate a
new provider.

The XDragDropClient is modified so it can now be aggregated; it is used
by the X11Window to handle the XDND events.

Subsequent CLs will add rich drag and drop features such as drag image,
mouse pointer shapes reflecting the current operation, etc.

Bug: 1014860
Change-Id: I1a32f880fb9ac4fe8748507142d227c6cbafddbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2033149
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Darwin Huang <huangdarwin@chromium.org>
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Cr-Commit-Position: refs/heads/master@{#762232}