[breadcrumb] Add click / key navigation tests for the main buttons
The main breadcrumb buttons respond to mouse click and Enter and Space
key events.

These events should make Files app 1) navigate to the path in the main
button that was activated by the event and 2) update the breadcrumb to
that path when navigation completes.

Add tests for this behavior. Click or Enter key a main button and test
that the breadcrumb is updated to that main button's path.

Tbr: adanilo
Bug: 1061388
Change-Id: Ifad382da48196476baef3794b98458b424fc82c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2164345
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762283}