[Autofill Assistant] Moved interaction creation to member method.
This is both for cleanup and as a prerequisite for an upcoming CL
where having this method as a member will actually be required.

This is a refactoring only and should have no user-facing changes.

Bug: b/145043394
Change-Id: I8746f5fa7e748b2727f42ef4f9b828b42fa73226
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2162815
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Marian Fechete <marianfe@google.com>
Auto-Submit: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Marian Fechete <marianfe@google.com>
Cr-Commit-Position: refs/heads/master@{#762319}