[Autofill Assistant] Send script url instead of deeplink url
Before this change the deeplink url (set once on |Start|) was used to
send along when asking for initial actions. This has now been changed
to send the more current script url.
This url does *not* get logged in the backend but does get stored in the
temporary memory to be used in messages to the user.

Bug: b/153959945
Change-Id: Iab10958296c9f21630318b9b5ef52851567d13f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2148747
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Reviewed-by: Marian Fechete <marianfe@google.com>
Cr-Commit-Position: refs/heads/master@{#762334}