[Passwords] Clone MultiStoreFormFetcher properly
Before this CL, calling Cline on a MultiStoreFormFetcher would create a
FormFetcherImpl with most of the properties of the cloned fetcher.
This CL ensures that the a cloned fetcher will retain multi store
properties (like blacklisting and ability to check whether the
credential can be moved to another store).

Change-Id: Id278d71d809bc41c12317b23cd78d1ec9dc5d00f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2162731
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762335}