OOBE - Fix a minor visual glitch on sign-in screen
Content shadows that are shown on the Gaia sign-in screen during OOBE
should not be visible when there is a popup open with a dark overlay.
This fixes an issue that caused the shadows not to be hidden sometimes.

The fix is to move the 'hideshadows' attribute from the webview to
the div that contains it.

Fixed: 1068251
Change-Id: I3afd42a82fa7adec43b66924996da9624629060f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163576
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Commit-Queue: Renato Silva <rrsilva@google.com>
Cr-Commit-Position: refs/heads/master@{#762337}