Improve performance of eliding long URL.
This CL is adding a binary search to reduce the amount of call to
ShapeRuns (called by ElideText(...)). The behavior should be
exactly the same than before but it is more efficient for
URL with many components.

The only behavior change is for long URL since they are cut using
the same algorithm now. The following code is removed:
  const size_t kMaxNumberOfUrlPathElementsAllowed = 1024;
  if (url_path_number_of_elements > kMaxNumberOfUrlPathElementsAllowed)
  {
    [...]
  }

R=pkasting@chromium.org

Bug: 1068603, 739975
Change-Id: Ia2ebc081eae283dbfa5034745db171037ca24509
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142578
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762356}