Avoid notifying timezone observers when the timezone didn't change
TimeZoneMonitor should not notify clients when the timezone stay
the same between update.

On windows, TimeZoneMonitorWin is watching for an WM_TIMECHANGE
windows message. This event is sent when system time settings
are modified. It seems this event is sent during suspend/resume
phases which is triggering a cascade of calls that should not
happen during suspend/resume.

Also, the network time update will trigger WM_TIMECHANGE messages
periodically even if the timezone stay the same.

Bug: 1074036
Change-Id: I1804e16ec4f191a22ffc60394192550cdaee9d14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2164225
Reviewed-by: Mark Mentovai <mark@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762373}