Display extensions as disabled after web time limit is reached.
Hosted extension apps and legacy packaged apps should be displayed as
disabled after the time limit is reached. They are opened in a browser
tab or browser window and are paused when Chrome time limit is reached.
The paused icon state should match that of Chrome's.

Bug: 1073541
Change-Id: Ic2c62362ce9cd8159095200a2ad340a066ff20e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2161666
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Commit-Queue: Yilkal Abe <yilkal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762425}