Bump Content Shell to "Chromium 999"
In r128926 (2012), Content Shell was started as “Chromium 19”
as that was “reasonably current” but quickly became outdated
as sites said, “that’s too old”.

So in r361414 (2015), it was bumped to give a version of
“Chromium 99” so that any UA sniffing would think it was
modern enough.

But given that Chromium 99 is actually coming soon (2022!)
let’s bump Content Shell to “Chromium 999”. This kicks the
can about 100 years down the road.

Bug: none
Change-Id: If0ae7f02fde953ee6dbd352e0ab3431e81ad953c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2161476
Auto-Submit: Avi Drissman <avi@chromium.org>
Reviewed-by: Rick Byers <rbyers@chromium.org>
Commit-Queue: Rick Byers <rbyers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762440}