ServiceWorkerContainerHost: Use ServiceWorkerClientType.
Replace the use of ServiceWorkerContainerType with an
optional ServiceWorkerClientType. This avoids the need
to convert from one type to another for GetClientType().

Additionally, this change simplifies a future CL where
ServiceWorkerContainerHost knows about the ID of the worker,
if the container host is a worker client.

Bug: 993029
Change-Id: I29a41d060545e9e88fcdb5e2f83af40c823887f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151465
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762441}