Mac: Make Find Bar activatable
Widgets of TYPE_CONTROL are not activatable by default. On Mac, this
means that if a user focuses the Find Bar in window B while window A
is key, window A does not lose key status, and key strokes are still
dispatched to it.

"Active" is a concept without a crisp cross-platform definition, so this
change is very narrowly scoped to Mac.

Bug: 937345

Change-Id: I4f366485eef000e5f66dbab6f8119a2a2ac93a6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163636
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Auto-Submit: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762447}