fido: disable caBLE if the BLE adapter does not support advertising
On Windows, some Bluetooth adapters do not support advertisements. When
discovering such an adapter, treat Bluetooth as unavailable during a
caBLEv1 request.

caBLEv1 relies on the client broadcasting an advertisement, and there
isn't much use in trying that if the adapter doesn't support it.

For requests where caBLE is the only permissible transport in the allow
list but Bluetooth is deemed unavailable, the WebAuthn UI automatically
advances to the "Your device doesn't support this type of security key
error screen".

Bug: 1069140
Change-Id: Ib7e45ec3c8dbc57f8bb6d1923678c7183e85d5f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2161904
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762484}