Exchange SandboxType::kInvalid for a CHECK.
No processes should be launched with an invalid combination of flags for
sandboxing. This eliminates SandboxType::kInvalid and adds a CHECK at
the point where the command line is parsed when child processes start.

Unit tests exist for this code and continue to pass. Tests that
verified that bad command lines were recognized are now removed as
Chrome will safely CHECK in these cases.

The following changes in default or behavior:-

* cloud print service (--type=service) now reports kNoSandbox
* NaCl loader process on non-Mac now reports kUtility
   - The NaCl loader defines its own sandbox but its command line
     is processed in some tests and needs a valid value.
* Fuchsia: sandbox types are now enumerated and the type must be
   specified when the low level sandbox policy is constructed.
* Linux: type added for Zygote to have before it grows up to get a
   real sandbox

Bug: 1065087
Change-Id: Ibe277153fa58771d12bae7e3c2f9c6b92b9370a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120049
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762525}