Fix never translate behaviors in Desktop UI
When clicking never translate, the current behavior is to dismiss the UI
and remove the chip. If the user unintentionally did this, this is very
detrimental. If they had already translated, there is no way to get back
to the UI. Instead, keep the translate chip around and show a check next
next to the never translate menu items. Also on clicking never
translate, revert the translation if it had happened.

Bug: 1066621
Change-Id: Ibd5c0fdc8a8542c123eaa4595a79f255b114542d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163038
Reviewed-by: anthonyvd <anthonyvd@chromium.org>
Reviewed-by: Scott Little <sclittle@chromium.org>
Commit-Queue: Megan Jablonski <megjablon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762566}