[WebLayer] Add BrowserContext parameter to WebsitePreferenceBridge.get/set*SettingForOrigin
The Site Settings UI is currently being componentized so it can be
reused by WebLayer, which will require it to be provided the
BrowserContext that should be used as WebLayer supports multiple
Profiles, with no Profile.getLastUsedRegularProfile() equivalent.

This CL adds a BrowserContext parameter to several of the JNI methods
defined in WebsitePreferenceBridge. Most of the other methods in the
class will need to have a BrowserContext parameter added in followup
CLs.

Bug: 1058600
Change-Id: I4f0fda26f1645d16f72bff5c4a6b39fa1ce5a4d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2140954
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Natalie Chouinard <chouinard@chromium.org>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762587}