[M81 merge] Chrome OS: Reparent always-on-top wallpapers
Showing or removing an always-on-top wallpaper causes the wallpaper to
only be reparented if device is locked due to https://crbug.com/1051375.
If the device is not locked the wallpaper is not reparented, causing
the always-on-top wallpaper to not show or remove correctly.

Currently, always-on-top wallpapers are only used in Demo Mode to
display the "Loading demo mode..." splash screen. This always-on-top
wallpaper is always displayed correctly (while device is locked), but
is sometimes removed while in an active session - causing all wallpapers
to display on top, rendering the Demo Mode session unusable.

This CL fixes this by always reparenting the wallpaper when
always-on-top wallpapers are shown or removed.

Additionally, in the existing AlwaysOnTopWallpaper test, the
GetWallpaperContainerId() checks were replaced by checking the child
count of containers.
GetWallpaperContainerId() is determined solely by the two booleans
is_always_on_top_wallpaper_ and locked, so checking it does not
accurately determine which container the wallpaper is in.

(cherry picked from commit 363f449436d2a1942580101034ab8878aeb9dac8)

Bug: b:149176883
Test: ash_unittests --gtest_filter="WallpaperControllerTest.AlwaysOnTopWallpaper*"
Change-Id: I94bb3ec59f28119e43fe3b31bdc7f4fd8ef8bfb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2051784
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Patti <patricialor@chromium.org>
Commit-Queue: Michael Cui <mlcui@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#741794}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2166203
Commit-Queue: dstockwell <dstockwell@chromium.org>
Reviewed-by: dstockwell <dstockwell@chromium.org>
Cr-Commit-Position: refs/branch-heads/4044@{#979}
Cr-Branched-From: a6d9daf149a473ceea37f629c41d4527bf2055bd-refs/heads/master@{#737173}