Add a PRESUBMIT check for editing external/wpt/interfaces/ files
These files are auto-generated by reffy, and we have seen cases where
Chromium developers edit them and then are surprised (or worse, never
notice) when reffy reverts their changes upstream.

The added PRESUBMIT is an upload error, but is not checked during commit
time as on occasion it may be ok to change these files. We just want
people to loop us in when they are doing so.

Bug: 1016354
Change-Id: Ie2c0c36add56140a90dade56db5783c96cc8a5a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2164569
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Robert Ma <robertma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762655}