[FragmentItem] Remove MoveToNextSibling
This patch removes |NGInlineCursor::MoveToNextSibling|
because it relies on |NGFragmentItem::IsSiblingOf| which does
not work well with culled inlines. |IsSiblingOf| checks
sibling of |LayoutObject|, but when culled inline exists,
the |LayoutObject| tree does not match the fragment tree.

Existing usages were removed in r761413 crrev.com/c/2160366
and r761415 crrev.com/c/2160448 , and this patch removes the
last two usages.

Currently we generate |NGFragmentItem| for culled inlines
since r729656 crrev.com/c/1774425 but this turned out to
consume too much time to allocate memory for them. It is
possible to fix it, but since replacement is easy, this
patch removes it.

Bug: 982194, 1035773
Change-Id: Ic98ca47dbbff53e53814430111a2a7a9699f2b49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2166202
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762697}