arc: bluetooth: Separate mojo callbacks for connection state changed
For the device connected state change, use a separate callback for
Android, and leave the existing callback for reflecting the state of
BluetoothGattConnection object we hold. Also move the code for this
from DeviceChanged() to DeviceConnectedStateChanged() which will be
much clearer than before.

This change is needed for the gatt connection management and
acl_stated_changed() callback (ag/10312919) on the Android side.

BUG=b:148355935
BUG=b:149658771
TEST=unit_tests --gtest_filter=ArcBluetoothBridgeTest.*

Change-Id: I8cdd734c88402f61547d8dc1d0b3178d3f8c7f0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2053748
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762706}