[Media Audio Element] Implement renderer error action.
This CL adds error reporting on the audio media element when playing
a local track or a WebRTC track. The reported error can be caught
via the onerror event handling and is of type Abort.

The change was tested on both local track and pc track using WebRTC
samples.

Fixed: 1003751
Change-Id: I20a2a194f07417048d39b520d85e68e585821699
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2153033
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Armando Miraglia <armax@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762738}