Stop using HttpResponseInfoIOBuffer in ServiceWorkerInstalledScriptReader
The main motivation of this CL is to modify
ServiceWorkerInstalledScriptReader::Client::OnStarted() so that it
doesn't depend on HttpResponseInfo.

ServiceWorkerResponseReader::ReadResponseHead() is added to move
HttpResponseInfo -> URLResponseHead conversion from Client
implementations (i.e., ServiceWorkerInstalledScriptLoader and
ServiceWorkerInstalledScriptsSender) to ServiceWorkerResponseReader.

ReadResponseHead() is added temporary to reduce code changes at one
time. ServiceWorkerResponseReader (and Writer) will be replaced
with storage::mojom::ServiceWorkerResourceReader in the future.

This CL also modifies ResourceResponseHeadAndMetadata to use
IOBufferWithSize for metadata instead of std::vector<uint8_t> so that
we don't allocate additional memory for metadata when it isn't needed
(ServiceWorkerInstalledScriptsSender doesn't need a copy).

Bug: 1060076
Change-Id: I177586ced3b2f2b283e81d7134c19f2aeca2fbdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2162435
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762763}