testing/buildbot: remove chrome_public_test_apk from swarming builder
This test is not run by production builder:
* https://ci.chromium.org/p/chromium/builders/ci/android-lollipop-arm-rel
And it is noisy to have test case in staging that does not exist in prod.

This also removes the test from
https://luci-milo-dev.appspot.com/p/chromium/builders/ci/android-marshmallow-arm64-rel-swarming/2019
but let me not take care that for configuration simplicity currently.

Bug: 1074759
Change-Id: I1a7643ebd9ae8feef972a504e8b14fafa7f412b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2166214
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@google.com>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@google.com>
Auto-Submit: Takuto Ikuta <tikuta@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762765}