base/bind: Check validity of nested callbacks in release builds
We are seeing a number of scheduler crashes caused by posted tasks that
contain null callbacks. Since these crashes occur much later and with a
different stack from where the problem originated, they are hard to
diagnose in the wild.

base::Bind() already checks that null callbacks can't be constructed.
However, there are some omissions in this logic:

1) If a callback is bound directly from another callback without binding
   any extra variables, the validity of the inner callback isn't
   checked. Because of this, it's possible to silently end up with a
   null callback, e.g., by binding a moved-from callback.

2) Callback validity is only checked in DCHECK-enabled builds. While
   this catches most issues at development, the linked bug is an example
   of when this safety net wasn't enough and buggy code ended up in the
   release.

This patch adds a CHECK() for case #1. Ideally we'd do the same for #2
too, but this turns out to regress APK size on Android by about 80KiB.
As a compromise, we only add the CHECK() for the case where a
{Once,Repeating}Callback is passed to Bind(). This only adds about
2.7KiB of code.

Bug: 1071475
Change-Id: Ia1671800f21eb1658b51e004218546ea933a8d97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159441
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Sami Kyöstilä <skyostil@chromium.org>
Auto-Submit: Sami Kyöstilä <skyostil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762889}