Simplify LayoutTreeBuilderTraversal::NextSibling and PreviousSibling
Just a small tweak after the refactoring of r718609. The parent_element
variable can only be null if node is not a pseudo-element. So we can
just check parent_element in the non-pseudo-element case instead of in
following pseudo-element cases.

This patch should have no effect in practice.

BUG=457718

Change-Id: Id4cb80da0e17d39931ab0188d6202149f41fb762
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2166182
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762927}