overview: Re-enable OverviewSessionTest.RoundedCornersVisibility.
Was not able to reproduce locally, or after running the test bot 5
times. Its possible that something else has fixed this in the meantime.
The old crash stack seems to suggest something was messing up with
the animation when we tear down:

#4 0x559a34efa56f ui::LayerAnimationElement::ProgressToEnd()
#5 0x559a34efd61c ui::LayerAnimationSequence::ProgressToEnd()
#6 0x559a34f0125a ui::LayerAnimator::FinishAnimation()
#7 0x559a34f02097 ui::LayerAnimator::StopAnimatingInternal()
#8 0x559a34ef786a ui::Layer::CompleteAllAnimations()
#9 0x559a34ec0d23 aura::Window::~Window()

So I added a waiter to make sure overview animations are finished
before TearDown.

Test: none
Bug: 1000730
Change-Id: Ia1768413c4e9b7b087a924f2ed63678dbbc5c339
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2165246
Reviewed-by: Malay Keshav <malaykeshav@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762933}