Convert EventType and TouchState to be a mojo type.
This removes the need for a typemapping to native type if we can
use the mojo type directly. There are aliases set on the classes as not
to break any existing code.

Removing the content native typemaps will allow us to move this mojo
definition into blink once they are all removed.

Change-Id: I09473d3582cd5ea19a47467291ba326f5dbe09c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2160523
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762941}