views: add WidgetDelegate::Params and setter for title
This change adds a Params struct to WidgetDelegate, analogous
to DialogDelegate, and adds a setter for the window title, which
is by far the most common customization of WidgetDelegate. Ideally
the title would instead be part of Widget::InitParams, but clients
of Widget often don't have direct control of their Widget's
InitParams.

Bug: 1075649
Change-Id: Ia2f88f4101e694f89ec6b7eb7162e290ba9ba597
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2168212
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762958}