De-flake IdentityDiscControllerTest
As far as I could tell, the main source of flakiness was that waiting
for a page to load didn't guarantee that the optional button had been
settled yet. Waiting for the button to appear (or go away/be replaced
with a button that has a different description) should hopefully be more
robust and compatible with the addition of new buttons.

Bug: 1072436
Change-Id: I92091fcccbb694cac2e74adaa5d7e4e44372b00c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2161306
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Reviewed-by: Pavel Yatsuk <pavely@chromium.org>
Commit-Queue: Patrick Noland <pnoland@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762974}