[iOS] Create SafeBrowsingUrlAllowList
This will track which unsafe navigations for a URL are allowed.
Since iOS is developing the safe browsing interstitials with only
the committed behavior, most of the UI manager functionality from
other platforms is stripped down to simply managing the url
allow list.  Classes that need to reference the whitelist can do so
directly through their WebStates rather than using the UI manager as
and intermediary object.

As an implementation detail to this CL, TestWebState::
DidChangeVisibleSecurityState() was implemented to notify observers
of the change.  This allows the changing visible security state due
to pending and completed unsafe navigation decisions to be tested.

Bug: 1064803
Change-Id: I38699dac16c53dd8472827dcccc7cbc9518d31d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163854
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Auto-Submit: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762984}