app_list: Changing container should not always deactivate window
App list view is shown in different window containers depending whether
it should be shown on top (in clamshell, or when showing assistant UI)
or behind app windows (in tablet mode). When handling the case where
the app list is moved from foreground to the background container in
tablet mode (when hiding the assistant UI),
AppListControllerImpl::UpdateLauncherContainer() deactivates the app
list window, but it does it when this behavior is not desired nor
necessary:
*  in clamshell, where the app list is only shown in foreground
*  even if the app list window is not currently active

This may cause issues if the container is changed while showing the app
list in clamshell mode - call to deactivate the window may change the
focused window and cause the AppListPresenter to think the app list has
lost focus and needs to be dismissed - note that UpdateLauncherContainer
is called just before the app list view is shown, so this may cause the
logical app list visibility state to be hidden while the app list is
being shown (in which case calls to dismiss the app list would be no-op).

This for example happens if the app list is shown in clamshell mode
while overview UI is active soon after exiting tablet mode
(UpdateLauncherContainer called from AppListPresenterImpl::Show updates
the container, changes the active window, which causes the overview to
exit and activate the app window, "dismissing" the app list just before
it actually gets shown).

The CL also moves the call to UpdateLauncherContainer during tablet mode
exit before dismissing the app list - otherwise, the container is not
updated at that time - UpdateLauncherContainer uses
presenter_.GetWindow() to get the app list window, which is null if the
target app list state is closed.

BUG=1073548

Change-Id: Iedc11e386ec44728921f25378d30cc82c8027bfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2166350
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763046}