CacheStorage: Attempt to make cached index size more reliable.
This CL adjusts the logic used to read the cached index size when
loading quota usage data.  Previously we were comparing the index
file size to the base path of the cache_storage directory in order
to determine if the index is stale.  Unfortunately, it appears that
some platforms will make the containing directory timestamp slightly
in the future compared to the last file updated within that directory.
So the index file was being considered almost always invalid.

The fix in this CL is to simply ignore the base path timestamp and
to only look at the timestamps on the cache directories.

Bug: 1047803,988001
Change-Id: If2ac6bc9313f8fc33ec004c0e2a56af4d363f034
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2165905
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763050}