[bluetooth] Clean up adapter initialization logic
This change rewrites BluetoothAdapterFactory and associated adapter
creation logic to use a more modern code style. The core goal is to fix
a flaky test failure caused by the previous logic not retaining a
reference to the BluetoothAdapter while it is being asynchronously
initialized. This meant that the WeakPtr held by the factory could be
invalidated the moment that initialization completed.

Bug: 1030549
Change-Id: I9e612c0c1745144f852aaf535b64f1fd5f81bd83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2154146
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Sean Topping <seantopping@chromium.org>
Reviewed-by: James Hollyer <jameshollyer@google.com>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763057}