Fix for multiple source devices trying to animate the scroll offset.
This CL fixes an issue where more than one source device tries to
animate the scroll offset on the compositor thread. For example: when
a scrollbar autoscroll (which is a linear animation) is in progress
and a mousewheel tick takes place, UpdateTarget gets called on the
linear animation. This is not allowed. The fix here is to momentarily
abort the ongoing autoscroll animation and allow the mousewheel tick
to animate. Once the mousewheel tick is done animating, the autoscroll
will resume on the next frame.

Bug: 1068305
Change-Id: Iafe41042c7b977a1c74027a2b8fe3349cc1b1fb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2158057
Commit-Queue: Rahul Arakeri <arakeri@microsoft.com>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763075}