[omnibox] Hide Page Action Icons for On-Focus Suggestions
Currently, when the user is typing in the omnibox, and the popup is open
showing suggestions, all the page action icons (content settings, zoom,
star icon) are hidden.

This should be true for on-focus suggestions too.

This CL makes that happen, by keying the hiding logic on whether the
popup is open (in addition to user input in progress).

Bug: 1030423
Change-Id: Ibf3b40171610e8245acd64a64f59f25d51c1b740
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2165878
Commit-Queue: Tommy Li <tommycli@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763087}