device/fido: keep HidConnection around to write Cancel.
It's important that we write a Cancel message when aborting an
operation: not doing so could result in a stray touch creating
credentials that the user doesn't expect.

However the FidoHidDevice (and thus the HidConnection that it owns)
are destructed quickly after the cancel message is written. On
Windows, this causes the writing containing the cancel to be aborted
and thus for the message never to be sent.

This change causes the HidConnection to be reference counted and for
the callback from writing the cancel message to hold a reference. This
means that the HidConnection will live long enough to ensure that its
written.

BUG=1029730

Change-Id: Ic9d2bf3a577f34047285f7bc519c5da74050e4da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155247
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#763156}