Only report FMP and FCP for main frame to MainThreadSchedulerImpl.
The change description introducing the signal in MainThreadSchedulerImpl
(crrev.com/2397753006) implies that it is intended to react only to top-level frame
FirstMeaningfulPaint and FirstContentfulPaint. However, there was no
check to see that the signal being sent was from a main frame, and this
could cause the signal to be reported too early (if it's from a
sub-frame).

Change-Id: I850eb5a8bbded30374fbbc6b989c94b1eeb761e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2141551
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Tal Pressman <talp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763169}