Use gpuBenchmarking in wheel event paging xy scrolling tests
After we finish the scroll unification, the scrolls happens mainly on
the compositor thread, and the scroll code in the main thread will be
removed. eventSender sends the scroll events to main thread, so it
would not work after the scroll unification. We should replace
eventSender.continuousMouseScrollBy with gpuBenchmarking.smoothScrollBy
in the wheel event paging horizontal and vertical scrolling tests. To
simulate paging scrolling, we can set scroll_by_page to be true in
smoothScroll function.

Bug: 1047176
Change-Id: If4a843798c27ca9683ea05e5812809dcf51c58ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2110521
Commit-Queue: Lan Wei <lanwei@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763179}