Prepare to flip the MediaApp feature.
This stops short of flipping the flag by default, but ensures that all
tests are happy when that happens.

Similar to component extensions (and component_loader.cc), the
SystemWebAppManager is configured to not install all system apps upon
browser startup in tests. So we need to deal with the situation where
the app is enabled, but not installed more generally.

We also need to explicitly disable the feature in FileTasksBrowserTests
to retain test full coverage of both flag states.

Bug: 996088
Change-Id: I9d419900bcc5ac79010c15896b6b5165fedac758
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2166239
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763184}