Update comment to reference correct manifest
This comment refers to chrome.exe.manifest which is an unused file.
This change updates the comment to refer to
chrome.VisualElementsManifest.xml which is the correct file, as seen
in the commit message for the change which added the comment
(crrev.com/c/1335075).

Change-Id: Iae0f3cb62cdc003d6903cda13c99a48196200ef3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2168017
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763186}