Move where we set the Help app prefs during OOBE.
Previously they were only set when the app launches. This means for
devices in tablet mode or users that were not type REGULAR, the prefs
were always false. By setting the prefs when we check if the app should
launch, these should be available to read for tablets and child accounts
on manual app opens.

Bug: b/155136046
Change-Id: I14bd9f6fbafaee151914ce04d4380d3addce7696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2169437
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Commit-Queue: Rachel Carpenter <carpenterr@google.com>
Cr-Commit-Position: refs/heads/master@{#763224}