[Autofill Assistant] TelephonyManager crash fix.
Before this patch, AA would crash when in airplane mode since the
telephonyManager is not available. This DCHECKS in the jni layer,
which is still something that should not happen even if AA is not
designed to work offline. This patch returns a fake country code
instead of null to prevent the DCHECK.

Bug: b/154321209
Change-Id: Iaa35b40eb0908fced41ba68b01c7c78393530870
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2167862
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Marian Fechete <marianfe@google.com>
Commit-Queue: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763231}