[WebAPK] Send shortcut icons with the update request.
This is needed to support the stalemanifest case, i.e when the
manifest is no longer available but an update is still required
to use a newer shell version.

This is also needed for intranet requests, since the server won't
be able to fetch the icons.

Bug: 1074351

Change-Id: I538b0ee054d14ed4c2881d28ca6c6c2f6636df8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2162733
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Commit-Queue: Rayan Kanso <rayankans@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763318}