Structured Headers: Dictionary members default to True
This implements the change in
https://github.com/httpwg/http-extensions/issues/1083, where dictionary
members are specified to default to boolean True if their value is
omitted from headers, and are required to be omitted from serialization
if their value is True, even if they have associated parameters.

This allows dictionary headers of the form

    Header: key;param=value

where 'key' will be assigned the value 'true' by default.

Bug: 1069784
Change-Id: I1d08ef812bf7316fd69b36094e4ea932ac367be2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2091712
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763330}