WebLayer: Implement ExternalNavigationDelegateImpl#closeTab()
In the initial bringup of ExternalNavigationDelegateImpl this method
was not implemented, and //weblayer's flow was configured to ensure
that closing tabs was never requested. However, WebLayer's flow now
properly detects cases where tabs should be closed, and hence
ExternalNavigationDelegateImpl can actually be requested to close a tab.
This CL implements that closing via the same flow as //weblayer's
InterceptNavigationDelegateClientImpl#closeTab(). The only case in which
this flow currently occurs is an extreme corner case that is unlikely
to occur in practice:

- The app using WebLayer does not have storage permissions
- The user goes to a file URL in a context where the tab should be
  closed if the navigation is overridden (essentially, the navigation
  is the initial navigation in the tab)
- The user denies the app storage permissions in the dialog that
  WebLayer pops up

I looked at testing this flow. However, it turns out that it would
require adding a private AIDL API to configure the WindowAndroid
instance to not regard the app as having the READ_EXTERNAL_STORAGE
permission.  This seemed like an unnecessary amount of overhead for a
flow that is basically non-existent in the product. When I implement
the warning dialog for launching intents from incognito, that will
also go through this closeTab() method. At that point I will look at
constructing a test for that more-realistic use case.

Bug: 1074301
Change-Id: Ic342877c569298b2a0e29cda10316c0cbe5074d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2162750
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763356}