[iOS] Using sign-in commands from GoogleServices
Sign-in UI cannot be dismiss without asynchronous calls. This was
previously ignored. The new architecture is more strict, don't allow
to ignore callbacks.
To prepare the transition to the new architecture, all sign-in UI has
to be managed by the SceneController.
The SceneController already supports asynchronous dismiss of the sign-in
UI.

Adding support for to asynchronously dismissing the sign-in UI on top
of the settings view.

=> crrev.com/c/2161007 GoogleServicesCoordinator
   crrev.com/c/2160911 AccountTableViewController
   crrev.com/c/2161027 SettingsTableViewController
   crrev.com/c/2165732 Fixing tests

Bug: 971989
Change-Id: I29718d306fd4b35e48eaadb90834cd0626ddb22f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2161007
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763450}