[LayoutNG] Skip hit-testing of culled inline children of columns.
The first idea was to allow |limit| in HitTestCulledInlineAncestors() to
be nullptr and just exit when we reach the first LayoutBox ancestor, but
that wasn't good enough, as that would still trigger DCHECK failures in
LayoutInline::HitTestCulledInline(), because neither
ContainingNGBlockFlow() nor NGInlineCursor work if the container is a
fragmentainer.

Therefore, just skip hit-testing of culled inlines that are contained by
a fragmentainer / column for now.

This is needed by CL:2134294

Bug: 829028, 1043787
Change-Id: I22348d69b5159de56614fe28ce5456b87643afe6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2170000
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763556}