[Switch Access] Add function to open the menu from the extension.
Part of re-implementing the Switch Access Menu in native views.

This change adds plumbing for the component extension to open the menu
at a specified location with specified actions available.

This code is currently not used, but will be in a follow-up change.

Relnotes: N/A
Bug: 973719
Change-Id: I0e5acacc0afd62d08fdd977ce42b1c935efa0756
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2153594
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/master@{#763574}