Make sure to issue Network.loadingFailed for navigation requests
We are missing loadingFailed when the request was successfull,
tried to commit in the renderer but failed to do so. In that case
NavigationRequest::state_ can be CANCELING or READY_TO_COMMIT.
Changing to HasCommitted() should do the trick.

One such scenario would be renderer-initiated navigation that was
also canceled in the renderer.

Bug: none
Change-Id: I819b1b487715c8c7aed5cd14fa4727b5fd94b3cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2161572
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Yury Semikhatsky <yurys@chromium.org>
Commit-Queue: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763576}