Move the close button to the Caption Bubble from the frame.
This allows us to draw it in exactly the place defined by
the spec. It will also make showing and hiding the close button
with animation easier later. In addition, when the expand button
is added it cannot be place in the BubbleFrameView, which
does not currently support buttons other than close.

Note: The spec called for the button to be 20x20, but here
I've got with 24x24 which is the default size for all
BubbleFrameView dialog bubbles. Tap targets should be as
large as possible for accessibility. The icon is still positioned
to spec with respect to the edge of the bubble view.

Bug: 1055150
Change-Id: I5949986ccc8682a778fdc215e1512b97f16345aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2165869
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763580}