Support 'sampling groups' for consistency between UKM events.
This allows a metric to declare its sampling to exactly match another,
not just in frequency but also which pages are sampled in/out.  This
will allow sampling for interdependent metrics which otherwise have
to be unsampled in order to be accurately compared.

It is activated by having a Finch parameter string give the name of
the control metric instead of an integer that is the sampling rate.

Some functionality was broken out into other functions for easier
testing.

Bug: 766909
Change-Id: I1d5f3a248722e44d4d5790f9581dc6d8b1eaa0e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151107
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Brian White <bcwhite@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763592}