content: adds special case to avoid reload when changing user-agent
For WebLayer, we want to allow changing the user-agent override during
DidStartNavigation(). Doing this currently triggers a reload (which
propmpty crashes). This patch adds the logic to detect the user agent
override and update the headers appropriately.

The test case shows what will happen in WebLayer.
. During DidStartNavigation() the WebLayer embedder will change the ua.
. WebLayer will call NavigationEntry::SetIsOverridingUserAgent(true);
. WebLayer will also call WebContents::SetUserAgentOverride(...);

In contrast to Clank, WebLayer will not set a ua-override on all
WebContents. The ua-override will only be set in the scenario described
above. As the override should only apply to the current entry, WebLayer
will change whether renderer initiated navigations should inherit the
ua override. Patch for that is here:
https://chromium-review.googlesource.com/c/chromium/src/+/2163148

NOTE: this doesn't yet handle the client hints related to the user-agent
override, that will be done in a follow on patch.

The patch showing usage of this API is here:
https://chromium-review.googlesource.com/c/chromium/src/+/2170409

BUG=1033029
TEST=WebContentsImplBrowserTest.SetUserAgentOverrideFromDidStartNavigation

Change-Id: Ic888ed39b52326016212c2b81387df12451a67f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163050
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763604}