[iOS] Create ConfirmationOverlayResponse
This is a general purpose response for use by overlays that are asking
for confirmation from the user.  AlertRequests that are asking for
confirmation can use GetConfirmationResponseConverter() to create a
response converter that creates a ConfirmationOverlayResponse for an
alert.

Bug: none
Change-Id: Ia0939996cd49cca1d736b923b6a6c266183430d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2165846
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Auto-Submit: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763649}