Exposes some text field related utility methods on AXNodeData
This patch is a precursor to a more substantial change that
would move hypertext computation to AXNode from AXPlatformNodeBase which will
in turn open the way for AXNodePosition to be merged with AXPosition and
for Views to be exposed as a tree of AXNodes.
See https://chromium-review.googlesource.com/2150294

Currently, we have two versions of some text field related utility methods,
one in BrowserAccessibility and another in AXPlatformNodeBase.
We want to centralize all platform independent utility methods in AXNode and AXNodeData.

RELNOTES: N/A

R=dmazzoni@chromium.org, aleventhal@chromium.org

Change-Id: I0518d675294737271fa21b6c2a5a106a95035f84
Bug: 1049261
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2170410
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763682}