Implement saving of passwords by smbfs
Passwords for smbfs are stored by smbfs in the daemon store directory
for the user. To restore, smbfs is told to read the store password from
the daemon store. Password storage is one-directional and there is no
means to extract a stored password. To prevent an instance of smbfs from
reading a password it is not entitled to, a salt is used by smbfs to
obfuscate both the filename and password when stored on disk.

Bug: 1054705

Change-Id: Id4a3187de4038f0556aeef7d8dc0d86dcb621e25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2167745
Commit-Queue: Anand Mistry <amistry@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Sergei Datsenko <dats@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763702}