Refuse to parse forms with invalid URL
FormDataParser turns FormData into PasswordForm. In particular,
FormData::url becomes PasswordForm::origin. If the URL is not valid,
the PasswordForm is useless, because it cannot be filled on any page.

There is no known source of invalid URLs for the parser as of now,
because the URL comes from the verified URL of the renderer process.
Nevertheless, should a future bug cause an invalid URL to be passed to
FormDataParser, the parser should abort parsing, instead of creating a
useless PasswordForm.

This CL adds such check.

Bug: 1075515
Change-Id: I7dc8e3ab2668506b789150a5cb4cf14d87739873
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2171096
Reviewed-by: Vadym Doroshenko  <dvadym@chromium.org>
Commit-Queue: Vaclav Brozek <vabr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763736}