Replace callback with direct call to client in chrome.passwordsPrivate
No behavior is changed, this CL replaces the callback that used to
abstract the account reauth flow with a direct call to the client API.
The main reason for the change is that it allows to replace calls to the
free functions in password_manager::features_util with their
corresponding methods in PasswordFeatureManager. The latter can be
mocked in tests.

This CL serves as a prefactoring to cleaning the "default password
store" pref upon opt-out, in order to allow better testing.

Bug: 1049141
Change-Id: I32d52aeb26cd17c6d6b6edf11df0b4df88b50ebb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2165897
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#763786}