Trust Tokens: no longer base64 encode the SRR header
When TrustTokenRequestSigningHelper was initially written, the
assumption was that an SRR was an opaque binary blob. It ended up as a
Structured Headers dictionary. There's no reason to base64-encode it,
and our server-side partners don't expect it to be base64 encoded.

R=csharrison

Bug: 1036483
Change-Id: I006bafb49c6a4e4fa2a94483ab593d67b4c8e354
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163402
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Auto-Submit: David Van Cleve <davidvc@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763788}