Clearly mark auto-open by user preferences as such.
This will make it easier to understand which value were set by the user
vs policy (when it's added).

The only functionality change is changing a forloop in download_prefs.cc
to be range based (as suggested by clang tidy). The rest of this cl
is just a naming change.

Bug: 92846
Change-Id: I5c14fee4f1f781b5afe65c142dec1dd6f1fd2b69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2160807
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Chris Sharp <csharp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763795}