Implement support for WPT manifest v8
This utilizes what is essentially a 'hack' to deal with the v8 manifest.
We decode the entire file to JSON (as always), and then flatten the v8
trie format into the legacy v7 list format.

This gives us none of the performance improvements that v8 can bring,
but it lets us keep rolling WPT without dealing with the complexity of a
trie datastructure. In the future we can properly handle the trie and
thus start gaining the performance improvements.

WPT tools are rolled to 41d7c3597e5d3128ff3c42bb128781e4c382ba7b.

Bug: 912496
Change-Id: Ibe7d28bd355cc7c4738bb3198e73f7171d964417
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155618
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Robert Ma <robertma@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763801}