Convert FrameMsg_SwapIn to Mojo
This CL migrates the legacy FrameMsg_SwapIn IPC message
to the new Mojo message in FrameNavigationControl interface.

In RenderFrameImpl class, the existing OnSwapIn() is renamed
to SwapIn() which implements the Mojo message, then SwapIn()
is renamed to SwapInInternal() to avoid duplicating the
implementation of the SwapIn of FrameNavigationControl
interface.

Bug: 1064342
Change-Id: If131fa3d91c73a7a71c609ca282ef437f16482a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2167737
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763840}