Add flag for interstitial warning on target embedding domains.
This CL creates a feature flag for the target embedding heuristic.
When enabled, navigating to a target embedding domain should result
in showing a lookalike interstial warning page. This CL is only for
defining the flag and integrating the flag with the interstitial will
be done in a future CL.

Bug: 1075659
Change-Id: I329a547ff147f3ccad181c2179d31dcf2ef8e675
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2171526
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Commit-Queue: Behnood Momenzadeh <behnoodm@google.com>
Cr-Commit-Position: refs/heads/master@{#763861}