Remove ChromeFullscreenManager.FullscreenListener#onUpdateViewportSize
Remove the FullscreenListener method in favor of a delegate. The
viewport isn't actually resized until ChromeFullscreenManager processes
the event (so it's the delegate).

There's more refactoring that could be done here to make this even
better, but this is a step in the right direction.

Also removes AssistantOverlayDrawable#onUpdateViewportSize since it
was determined that's not needed due to other viewport size polling
logic.

BUG=995914

Change-Id: I82b0c48863ab428abf5a4ae78f2e0654ad5c185b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2170560
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Theresa  <twellington@chromium.org>
Auto-Submit: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763880}