Remove extension dev tools infobar only when the user closes/cancels it.
Previously, the bar also disappeared if it was no longer applicable,
i.e. no more extensions were attached.  This is problematic since the
extension may have made persistent changes to the page, and the user
needs to be guaranteed an opportunity to see something telling them
that.

Bug: 1070066
Change-Id: I3bf573013037752be9654c2840b153b5ab8d85a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159967
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763940}