Don't reset computed 'float' for flex and grid items
Floated flex and grid items have their 'float' incorrectly computed to
"none". This change fixes this by adjusting IsFloating() for flex and
grid items inside LayoutBox::UpdateFromStyle() instead of inside
StyleAdjuster::AdjustStyleForDisplay().

Beyond this, legend elements are not allowed to be rendered legends
if they are floating. However, if legends are flex items, we
adjust IsFloating() to be false. This causes legends to be
used as rendered legends when they are flex items, even if they
were styled as a float.

This is fixed by checking ComputedStyle::IsFloating() instead
of LayoutObject::IsFloating() when finding the rendered legend in
order to get its non-adjusted floating value.

Bug: 875235,350505
Change-Id: Ia1e7f7c244cc0c443cd58be42854866884b7f7e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2157889
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#763988}