Add helper functions to set Connectors prefs
These helpers use the legacy policy values but map to connectors. The
idea is to use these in tests to keep their coverage for the new
connector policies. This CL just gets the boilerplate out of the way
and doesn't actually refactor any test.

Bug: 1076083
Change-Id: I7f35f2e8550458ade396dd7a4a40b88a9745ecb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173341
Reviewed-by: Roger Tawa <rogerta@chromium.org>
Commit-Queue: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763989}