Use BigString16 for JavaScript execution
When passing strings to the renderer for execution, use BigString16,
which has no strict limit on their length. This is due to the fact
that this API is sometimes used to pass JavaScript that is
user-provided and therefore we can't make any guarantees about its
length.

Bug: 1073673
Change-Id: I013312968161c56e0856226cd650095558c201f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2167577
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763999}