Fix mirroring animated video playout delay
In the mirroring service, currently the min and max playout delay values are
effectively ignored, and only the animated playout delay is used to
negotiate latency with the receiver.

My previous patch didn't update the video playout delay, only the min
and max, so the audio latency was configurable but not the video
latency. This patch resolves that issue by ensuring that min, max, and
animated playout delays are all set to the same value for both audio and
video. The now-dead constants for min, max, and animated playout delay
are also removed.

Change-Id: I20632d84d618ad75520f085233ae1d102339c6a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173660
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Commit-Queue: mark a. foltz <mfoltz@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764051}