Add eye dropper view for Windows.
This CL implements the custom eye dropper window for Windows.
The eye dropper is a magnifier control that magnifies the content of a
particular area of the screen and displays the content in a window.

The eye dropper follows the cursor and avoids stealing the focus from
the color picker which opened it.
Its position is updated by using SetTimer with a 16ms granularity which
is good enough for a 60HZ refresh rate.

A listener is introduced in order to notify the color picker about the
results from the eye dropper.
The update is sent by using the inter-process communication introduced
by a previous CL [1].

When the ColorChooserPopupUIController gets the information, a message
is posted to the color popup in order to update its state.

This feature is under the EyeDropper flag.
For now this functionality was manually tested, I am planning to send
future CLs with interactive_ui_tests and eye dropper window
implementation for other platforms.

The UI matches: https://eyedropper-test-10x.glitch.me/, but different
colors are used.

[1] http://crrev.com/c/2113706

Bug: 992297
Change-Id: I81cbc4834e410c64c0e625a05943b31fd868d850
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134938
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#764054}