Set initiator to stylesheet's URL for resources loaded from CSS
Currently, resources loaded from CSS (e.g. fonts) do not have their
initiator set appropriately. It is often set to the script that caused
a style recalculation. Instead, we fix these cases to use the referrer,
if present, as the initiator.

This does not fix the case where a stylesheet without a URL (e.g. inline
CSS) loads a resource. In these cases, the initiator falls back to the
document loading the resource. This also does not set the TextPosition
of the initiator as it is non-trivial to plumb that through Blink's CSS
logic.

Bug: 918196
Change-Id: I48120b7d91ebad439824e886b5b9d2b07bd480f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163039
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764056}