Simplifiying OffscreenCanvas-placeholder-createImageBitmap.html test
This test was supposed to be testing the commit method.
That method has been removed for a while, fixing the test so it actually
test what we would like to be tested.

The test is still flaky, that will be addressed in another CL.

Change-Id: I2f49b4e453c2e4f5acdc6381f8370f21ff91d850
Bug: 1071980
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155170
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Auto-Submit: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764059}