[fuchsia] Use cors_exempt_headers in WebEngine and CastRunner.
Adds the headers injected by CastRunner to the CORS header safelist,
so that OOR-CORS does not block the requests for servers which do
not support the "Access-Control-Allow-Headers" HTTP response header.

Defines a new interface CorsExemptHeaderProvider, which is used by
CastRunner to source an exemption list from the Agent.


Bug: 1039788, b/143307063
Change-Id: I614cb87c42e887564f34a60cd8200c7add22a824
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2150064
Commit-Queue: Kevin Marshall <kmarshall@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764099}