Select the largest video quad for overlay on Windows.
DCLayerOverlayProcessor used to select the first video quad for
overlay when multiple video quads are present. This CL chooses the
quad with the largest size instead of the first in the quad list.

This change does not apply to non-root videos. The current size
comparison can only be done in the same render pass. The code has to be
capable of comparing quads across render passes in order to apply the
change to non-root videos.

Bug:1074030

Change-Id: I21b8e3af0756e2fa2afce8ac04884cb2ef9eb366
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2170170
Commit-Queue: Maggie Chen <magchen@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764106}