Stop using HttpResponseInfoIOBuffer in ServiceWorkerCacheWriter
This CL adds ServiceWorkerCacheWriter::ReadResponseHead() which
replaces ServiceWorkerCacheWriter::ReadInfoHelper() to remove
HttpResponseInfoIOBuffer use in ServiceWorkerCacheWriter.

Since the signature of the callback of
ServiceWorkerResponseReader::ReadResponseHead() is different from
other IO methods (which use net::CompletionOnceCallback), this CL
adds a specialized version of AsyncOnlyCompletionCallbackAdaptor
to turn the method "always-async" to "async-or-result".

Bug: 1060076
Change-Id: I0ca9b8aa6d991d13373faf2e0a54150c77d79517
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2169453
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764169}