[cascade] Reanalyze when direction/writing-mode changes
The current way of dealing with surrogates (~aka css-logical properties)
is very complicated, and hard to understand and maintain.

This CL suggest a far simpler approach:

The direction and writing-mode properties are inherited properties.
When we Analyze the MatchResult, the inherited values of these
properties are present on the StyleResolverState's ComputedStyle, and
the StyleCascade can speculate that these inherited values will be the
final values, and proceed as if they will be.

With the above assumption, we can now resolve surrogates into their
original properties Analyze-time. In other words, no surrogates are
ever added to the CascadeMap.

Then, we apply cascade-affecting properties at the very beginning
of the Apply process, and check if the computed values of direction/
writing-mode *did* change. If the didn't, great. Our previous
assumption was correct, and we don't have to do anything.

If they did change, our entire CascadeMap might be wrong, so we clear
it, and Analyze again (aka Reanalyze), this time with the correct
direction/writing-mode present on ComputedStyle.

This approach has the following benefits:

 - Significantly easier to reason about, and makes especially the
   code related to 'revert' far more acceptable.
 - Can be used for other cascade-affecting cases, e.g.
   forced-color-adjust.

However, it did regress the perf test CSSLogicalDirection.html by
~10% in local testing, so there's a tradeoff.

Note: DCHECKs everywhere asserting !property.IsSurrogate() may seem
excessive, but I think it's helpful to signal to future devs where
a surrogate is definitely not going to appear.

Bug: 947004

Change-Id: If0cdcb653b4dbbfcc04ba26140e39512f6c587ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132417
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764184}