CrOS Settings: Hide shelf navigation setting on clamshell-only devices.
Currently the setting for showing shelf navigation buttons is shown in
settings for all devices as long as the
IsHideShelfControlsInTabletModeEnabled feature is enabled. This setting
should be hidden on devices that are not tablet capable and are
clamshell-only.

This cl hides the setting from showing up in settings when the device is
not tablet capable.

Bug: 1062213
Change-Id: I99574f7b340d41250165618f327815617300a128
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2152988
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764211}