Remove a useless RefCountedData in the Trust Tokens installer
The Trust Tokens component installer loads a file into a string on the
IO sequence and passes it back to the UI sequence. When I initially
implemented this, I was under the mistaken impression that
RefCountedData<T> was a ref-counted, thread-safe wrapper around a
type T, and I shared a RefCountedData<std::string> between the
sequences in order to pass the loaded string.

In fact, though, RefCountedData<std::string> is just a ref-counted
wrapper where the *ref count* is thread-safe; after talking it over with
some teammates, it seems like the easy way to pass data from one
sequence to another is with base::PostTaskAndReplyWithResult. I've gone
ahead and removed the useless RefCountedThreadSafe, which also
simplifies the code.

R=waffles

loaded with the change applied.

Test: Manually tested that the Trust Tokens keys are still successfully
Bug: None
Change-Id: I6e3dc8cd6b05ea01bbcabd88abf765fce2263e56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2172212
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Joshua Pawlicki <waffles@chromium.org>
Auto-Submit: David Van Cleve <davidvc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764227}