Remove CSSVariables2 runtime flag
This flag toggles CSS.registerProperty, which shipped in M78.

The current code lazily creates the registry when the feature is
enabled, but it does so also for the const version of the getter.
This seems a bit wasteful, as it probably means the registry will
always exist in practice, even when no properties are registered.

Hence this CL uses a more common Ensure/Get pattern, where the getter
can return nullptr because the registry hasn't been created yet.

Change-Id: Ic399ce7ac1e9b67a6337065d2cfabf9b829f984d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2172298
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764278}