Relax lifecycle DCHECK to account for non-lifecycle layout dirtying.
Rendering dirty bits can and are set via non-rendering script tasks.
During such tasks, the DocumentLifecycle is not in a well-defined
state.

It appears to be extremely uncommon for this to be possible for layout,
however, because basically any DOM API will mutate style and not layout
directly. For example, mutating the |width| attribute on an <img>
will cause style to be dirtied for the |width| presentation attribute,
which in turn dirties layout via StyleWillChange/StyleDidChange.

Animated SVG attributes are an exception, as they are implemented
via special code that reflects their attributes and dirties layout
directly. See SynchronizeAnimatedSVGAttribute, and for an example
of dirtying layout, SVGTextPositioningElement::SvgAttributeChanged.

Note that SynchronizeAnimatedSVGAttribute is only called outside
of the regular lifecycle or a forced style if one
of these attributes is mutated and then read back from script.

Bug: 1075977

Change-Id: If42186d40d9a953ed21ff2ea0e308a6f0e9a3f5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2171678
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764315}