Unexpire RTCVideo(Decoder|Encoder)Init*Success metrics.
These are health metrics that are necessary to evaluate WebRTC in
the field. They expired in the XML but were not removed from the
code. They will be useful in https://crbug.com/982799 and
https://crbug.com/709631.

Also fixes an issue with the recording of the profile histogram.

Bug: 982799, 709631
Change-Id: I491281ba07c84bc6675fb575bf81d0a892abf201
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2170984
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764367}