Fix DCHECK if _NET_CURRENT_DESKTOP is not set
If you request a nonexistant property, the X server doesn't send an
error, instead it just returns a reply with the format set to 0.  Also
check that the size of the data is at least 4 since we're reading 4
bytes (not sure if that's possible if the format is 32, but just to be
safe).

R=sky

Bug: 1076913
Change-Id: Ie7fdd7b22567f7d9664117ef655e4a8105841d5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174532
Auto-Submit: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764432}