[Fuchsia] Fix FuchsiaVideoDecoder to fail on fraims without PTS
FuchsiaVideoDecoder always sets timestamp for the packets it sends to
mediacodec, so it's expected to receive timestamp for all decoded
frames. Previously it was silently ignoring the case when timestamp is
not set. As results the corresponding frames were dropped by the
renderer. There are some cases when mediacodec doesn't return timestamp
and these bugs are hard to debug right now. This change fixes
FuchsiaVideoDecoder to fail decoding when it receives frame without
timestamp, since that normally indicates a bug in the decoder.

Bug: fuchsia:47591
Change-Id: I0704edb300d9155b02c15998ff51b4020b3bf44a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174795
Auto-Submit: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Kevin Marshall <kmarshall@chromium.org>
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764473}