Support iframe and window.open() inside a Web Bundle file
Currently Chrome doesn’t support iframes and window.open() inside a Web
Bundle file. To support them, this CL changes RenderFrameHostImpl to
check the existence of WebBundleHandle of the parent frame (or the
opener’s frame) in the beginning of the navigation. If the frame has a
WebBundleHandle, a WebBundleHandleTracker is created from the handle and
will be passed to the NavigationRequest of the navigation.
NavigationRequest::OnStartChecksComplete() will call
WebBundleHandleTracker::MaybeCreateWebBundleHandle() which creates a new
WebBundleHandle if the web bundle contains the page, and the page will
be loaded from the web bundle.

Bug: 1040800
Change-Id: I308d85d8192345e82dcf943c499fc15b0529d459
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2160633
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764486}