[ReaderMode] Adds click to dismiss behaviour for settings dialog
This patch implements the "click to dismiss" behaviour requested by UI.
When the settings dialog is open, if a user clicks anywhere outside of
the dialog, it is closed. This is a pretty common pattern in Chrome UI,
but is not supported by the native dialog.

This patch refactors the settings dialog into a class to better track
the open/close state of the dialog, and clean up the JS a bit.


Bug: 1027612
Change-Id: Ic8895642f9070e5dcf7a36cb81ef5185fdd2a986
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173918
Reviewed-by: Aran Gilman <gilmanmh@google.com>
Commit-Queue: Meredith Lane <meredithl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764526}