Add tooltip for Readonly indicator.
The tooltip is as info card style, not the ordinary tooltip.

As this tooltip needed to be aligned on the left side, it needs to a
separate calculation for right-to-left mode.

Having two separate styles for the card tooltip (this change) and normal
style needed explicit style cleanup after transition to switch between
different styles.

Multiline tooltip supported through <br> tag as tooltip content now set
as innerhtml not text with this CL.

Bug: 1058549
Change-Id: Ief23d0b2d260ffd600eaec7a7657c0b0b554a1a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2166210
Commit-Queue: Omid Tourzan <oto@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764528}