[Switch Access] Verify MenuManager instance exists before use
This is a hotfix for a bug that causes Switch Access to crash on
physical devices. It does not affect standard development workflows.

Root causes and additional test coverage are being investigated, but
this bug is blocking new UX team members from familiarizing themselves
with the feature.

TBR=katie@chromium.org

Relnotes: N/A
Bug: 1077120
Change-Id: I3824cc4d360c2212f0e4af64528710484210dae3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2175120
Reviewed-by: Anastasia Helfinstein <anastasi@google.com>
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/master@{#764535}