weblayer: Implement Tab.captureScreenShot
Asynchronous API that's implemented as a readback on
RenderWidgetHostView. Screenshot height is reduced by the the amount
of top browser control. Converting to java bitmap is done on a
background thread since it involves a copy.

Try to ensure the callback is never dropped by reusing old
CompositorImpl implementation that forces a draw before unsetting
Surface. It was recently removed in
https://chromium-review.googlesource.com/c/chromium/src/+/2171318
It ensures the callback is executed for edge cases like becoming
in visible, losing surface, or being destroyed.

Add a test that ensures callback executes even if fragment is
destroyed immediately calling capture. Also add a basic test to
ensure the bitmap indeed captures the contents of the page.

Bug: 1076100
Change-Id: I2d782d497349a0af8d528c715e62c8b0a9fbbb41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2170425
Commit-Queue: Bo <boliu@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764548}