Update usage of OTR profiles in global media controls test.
Profile API is being updated (issue 1033903) and the usage of
GetOffTheRecordProfile and HasOffTheRecordProfile without argument is
getting deprecated. Use cases of this call that ONLY refer to incognito
or guest mode can use [Get/Has]PrimaryOTRProfile() instead.

The call to this function in global_media_controls_in_product_help
unittest is updated.
The test name is also updated from Other profiles to Incognito profiles
to emphasize that we are not testing non-primary OffTheRecord profiles,
which are not used to create browser objects.

This CL does not make any behavioral changes.

Bug: 1060940
Change-Id: I8421353970307fb3cd21935d0d6d175786d24435
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174197
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764570}