Make call to ObserveGestureEventAndResult synchronous
This CL simplifies the call to ObserveGestureEventAndResult in the
InputHandlerProxy. There isn't a strong reason to have to call this
asynchronously. Besides, scroll events are rarely handled on the main
thread for the root scroller (and that path is soon going away). This
was added in https://codereview.chromium.org/713413002/diff/20001/content/renderer/input/input_handler_proxy.cc#newcode360

Bug: 1058071
Change-Id: I3c3739cbde95cfb27a69b38befd903bc05d3fe40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173555
Commit-Queue: Rahul Arakeri <arakeri@microsoft.com>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764589}