[Android] Rename RedirectHandlerImpl to RedirectHandler
We introduced the RedirectHandler interface while componentizing
ExternalNavigationHandler in order to work around the latter class's
dependence on the //chrome-level TabRedirectHandler. We've subsequently
componentized TabRedirectHandler as RedirectHandlerImpl, rendering the
interface unnecessary. A predecessor CL did a first step of eliminating
this artifact of the componentization by eliminating the RedirectHandler
interface and having the component directly talk to RedirectHandlerImpl.
This CL moves RedirectHandlerImpl to RedirectHandler. We did this in
its own followup CL because doing both in one CL resulted in Gerrit
showing the change as hundreds of lines of diff in RedirectHandler.java.

Bug: 1017390
Change-Id: I8a6503578cc2e725bc9978a277f5735f2b7076a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173365
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764591}