Remove //chrome's AuthenticatorNavigationInterceptor
This is no longer referenced downstream and hence can safely be removed
without breaking the downstream build. We change the AppHooks method
that is no longer being overridden downstream to talk in terms of the
//components-level interface; hence, I can follow up with a downstream
CL that implements that method and kick off the process of removing the
V2() version that I had to add during the componentization.

Bug: 1071390
Change-Id: Ic8cd8c5734635a4b3708a9ef40f68c0596a1a61f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174427
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764630}