[freezing] Deprecate the TabFeeezingEnabled enterprise policy (reland).
This is a reland of https://crrev.com/c/1965907. It was reverted because
tests were not updated. This also updates tests.

Latest experiment results suggest that we might not ship
background tab freezing to 100% of Stable.

This CL deprecates the TabFreezingEnabled enterprise policy, to
avoid confusing admins with a policy to disable a non-shipped feature.
Tab freezing is not enabled for any Beta or Stable user at this time.

No-Presubmit: true
Bug: 844426
Change-Id: I8af0efe77a163315fbb8296de9058c6ff1a0ba76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174350
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764631}