weblayer: renames TopControlsContainerView
I'm in the process of adding support for a view anchored to the
bottom of the browser. I plan to share what is TopControlsContainerView
between the two. As a precursor to that work I'm renaming
TopControlsContainerView to BrowserControlsContainerView. There
should be no functional change here. I realize some of the names
aren't updated. I was hoping to purely move the files in this patch,
but that's made hard by java class names (file name and class name must
match) as well as the jni code we generate expecting C++ class names
to match the Java name.

Patch that completes this is here: https://chromium-review.googlesource.com/c/chromium/src/+/2174941

BUG=1025608
TEST=none

Change-Id: I4ddedb703a2e3dd12fd581e0e320be7a01d2a402
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174862
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764654}