Print Preview: Move key computations out of destination_select
The PDF destination key is a constant; move it to destination.js and
use a constant rather than a function to compute.

Instead of using a boolean to indicate whether the drive destination is
ready and separately computing the key in destination_select.js, use the
key that is already computed in destination_settings.js and set it to
empty if the destination is not ready.

Bug: 1059607
Change-Id: I40f858d67008534dd1d36124247bb72c188066c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174341
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764671}