appcache: bump manifest version to force token read
Currently, there's a timing issue with adding Origin-Trial to a
manifest.  If a user hasn't upgraded to M84 and gets the new manifest,
the token_expires will not be correctly set on their cache.  This change
makes it so that so long as a user has visited a site with an origin
trial token in the manifest, even if they are on a previous version of
Chrome, we will reparse the manifest and look at the token even if the
remote manifest is unchanged.

Bug: 582750
Change-Id: I6d15aa75093c6594a8ad9e668955c82d9dafc437
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173594
Commit-Queue: enne <enne@chromium.org>
Reviewed-by: Chase Phillips <cmp@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764699}