Android: Dedupe clobberCurrentTab() in intent launching code
This CL dedupes the implementations of clobberCurrentTab() between the
//chrome and //weblayer embeddings of //components/external_intents. We
do this by moving these functions into the shared
ExternalNavigationHandler. This necessitates adding a new
ExternalNavigationDelegate method to load a URL.

One minor note: The //chrome version of this method guarded the call
to mTab.loadUrl() by a check that mTab hadn't been destroyed, while
the //weblayer version guarded it by a check that hasValidTab() is
true. We use the latter in both delegate impls of loadUrlIfPossible():
it's strictly more safe as //chrome's impl of hasValidTab() just checks
that mTab is not null and has not been destroyed. In practice it
shouldn't make any difference: if the //chrome delegate ever called
mTab.loadUrl() on a null Tab instance it would just blow up anyway.

Bug: 1071390
Change-Id: I5439e8c7f43d72729c7ab52388724a1ee3e98eb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174503
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764703}