crostini: Add a confirmation dialog when resizing a sparse disk.
Since this is an irreversible operation, the mocks specify a sparse disk
should not be resized without warning the user.

There is necessary complexity introduced here. In order to show the
confirmation dialog at the correct time, it is necessary to query
disk info when the crostini page first loads. The time
cost of starting a VM is too high to allow delaying the settings page
load, so a cheaper version of disk info is now available.

Bug: 1074188
Change-Id: I7ec8f8da34be0ddd9ab3eae6e30b8d8a786d151d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173921
Auto-Submit: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Cr-Commit-Position: refs/heads/master@{#764709}