Fix WebAppFrameToolbarInteractiveUITest when ExtensionsToolbarMenu is enabled, again.
After fixing WebAppFrameToolbarInteractiveUITest.CycleFocus in
crrev.com/c/2027788, we stopped showing the toolbar menu when no
extensions are installed, causing the test to skip it in the focus
order. To fix this, this patch installs an extension at the start of the
test, and then parameterizes the test to ensure coverage of both paths.

This patch also changes WebAppFrameToolbarTest from a test base to a
mixin, so the suite can use ExtensionBrowserTest to initialize the
extensions system.

Bug: 1071507
Change-Id: I04002aaa4c8319cac53b39d77c637c93aad6f19b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2160914
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764723}