Handle the WebUI tab strip when calculating top inset
A more complete fix would probably change the conditional to be more
representative of what it's testing: whether or not the window is a
normal browser or an app window. This is a fix with minimal side effects
intended to be mergeable.

Bug: 1076675
Change-Id: I6f7e5d6ce672446205749166f9d5f8276e4f161e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2175121
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764727}