Walk FrameTrees to register origins in all frames.
In addition to doing a global walk of the session history to find
existing instances of an origin that is opting in to isolation (see
https://chromium-review.googlesource.com/c/chromium/src/+/2018252), we
must also look for that origin in the FrameTreeNodes. This looks at
committed origins as well as navigations that are pending commit.

The FrameTree walk is necessary since some frames may be missing
FrameNavigationEntries, and also since pending navigations will contain
information not yet in the session history.

Bug: 1042415, 1062719
Change-Id: I802c1f391037471a698e7eb7795a367bd0604f5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2136703
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764730}