Make md5sum faster for remote Android device usage.
Changes are primarily implemented in catapult in
https://chromium-review.googlesource.com/c/catapult/+/2170402/

This changes the interface of md5sum to accept a gzipped list of files,
removing the recursive directory walking so that resulting md5 sums can
be printed without also having to print the corresponding filename. It
also truncates the MD5 sums to 16 characters (64 bits) as this should be
more than sufficient for detecting changes and speeds up the transfer
time significantly.

Bug: 1076580
Change-Id: I2af0fd2437d0ab6040041663fcbc3cbb8dd1afbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2171416
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764779}