[pgo] sparse disable flag for code_coverage merge scripts
Currently, the code_coverage merge scripts invoke llvm-profdata with
--sparse=True as a default and hardcoded setting.

PGO utilizes these code coverage merge scripts to merge profile data,
but needs to run it without the --sparse flag. Introducing a disable
argument to turn off sparse.

This is introduced as a disable flag because we want to retain the
default 'on' setting for branched tracks that run code coverage.

Bug: 1077310
Change-Id: I7fdd8fb93cb7d4985961529fc451e4299b7aeeb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2176154
Commit-Queue: Jeff Yoon <jeffyoon@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764780}