Use IsInitializedAndNotDead() to check if a process is still alive.
This CL brings the service workers and shared workers in line with how
the dedicated worker factory checks if the RenderProcessHost is valid.

While the current methods are correct, I think there's a benefit to
being consistent between the 3, and the sematics of
IsInitializedAndNotDead() more closely resemble what is needed for
our use case, which is to guarantee that OnRenderProcessExited() will
be invoked.

Bug: 1035555
Change-Id: Iecc208bd90422e464fa0b17d396fc4a1974c00c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124541
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764847}