Use MessagePortDescriptor instead of raw mojo::MessagePipeHandles.
This allows the usage of MessagePorts to be instrumented, as they are
created, passed around, used, detached, and finally closed. A follow-up
CL will build the actual instrumentation. The instrumentation will
eventually drive priority boosting logic hosted in the browser process,
ensuring that background (or frozen) endpoints are woken up if a high
priority foreground frame is attempting to communicate with them.

BUG=980533

Change-Id: I8f0858c33805b90d7db5247796302534450ef415
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1952124
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Chris Hamilton <chrisha@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764906}