Add PDFExtensionTest for opening context menu from touch selection menu
This CL adds a browser test for TouchSelectionMenu in PDF. There is no
utility class which could help in determining when touch selection menu
is opened. This CL adds TouchSelectionMenuWaiter() to determine when
touch selection menu is shown and helps in executing commands on the
touch selection menu.

Fix some nits in TouchSelectionMenuViews along the way.

Bug: 1027842
Change-Id: I9ce95ab2b697d474fedddec2777f75252f520d6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2082750
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Ankit Kumar 🌪️ <ankk@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#764914}