[menus] Add files-ng ripple to elider drop down
Adds a child <paper-ripple> element to the buttons in the elider drop
down menu.

Sets 'position: relative' on the buttons to contain the ripple to the
bounds of the button.

Enables the ripple to only be visible when '.pointer-active' is set on
the parent document root.

Tbr: noel
Bug: 1002396
Change-Id: I91eed9cfbb0687e4270fd40dc747bde0c15247a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2178331
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764998}