SyncTest: Disable pre-Commit GetUpdates
Before this CL, Sync integration tests sent a GetUpdates request before
every Commit. This was because in the past, invalidations weren't
working in tests, and so this workaround was required to make tests
pass.
Now, we do have working invalidations in integration tests, so this hack
isn't needed anymore.

Bug: 657130
Change-Id: I7c90c9a1e0ff0ac07e2bc41ac7df272f3cc26b45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2172972
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765022}