[Autofill Assistant] Moved element_area position comparison to native.
Before this CL, native would notify Java every 100ms about the showcast
positions, even if nothing changed. Java would then do a comparison and
only update the overlay if a change has occurred.

Moving the comparison to native allows covering it with unit tests, not
to mention that we don't want that kind of logic in Java to begin with.

This CL also adds VLOG(3) logging to position updates, which are
intended for debugging of hard-to-reproduce bugs.

Bug: b/154478864
Change-Id: I573bbb1cfe21e802bec6ec2ea9f21c003319c9ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2172871
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765031}