Set the referrer on ResourceRequests created by CSS @import rules
When a CSS stylesheet loads another stylesheet with an @import rule, the
current behavior never sets the referrer or referrer policy. We change
this so that the stylesheet sets these correctly by using the parent
stylesheet's CSSParserContext.

We also re-enable the Web Platform Tests for this behavior.

Bug: 1076481
Change-Id: I2091bfff347e3cc96038ca75a3a8a91555d8690c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2169875
Commit-Queue: Alex Turner <alexmt@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765093}