make computedStyleMap return deferred transform interpolations as the absolute-part matrix
When computedStyleMap encounters a kInterpolated TransformOperation
(a deferred, layout-dependent interpolation) it now serializes it as
the matrix() value obtained by applying it to a box-size of 0x0, which
returns a value only off from the true transformation by a relative
translation (i.e. the absolute portion). Previously, this inserted
'none' into the transform list, which is not valid CSS and will not
parse.

For use in animations code, which has access to layout and needs to
preserve visual data, ValueForTransformOperation now has a box_size
parameter which defaults to 0x0.

All possible transform lists now output valid CSS in computedStyleMap.

Bug: 1056197, 1057307
Change-Id: Id94d59592cef94837ba5df9d98b6ce668a373ca9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173744
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: George Steel <gtsteel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765138}