UpdateComputedHypertext to rely on ChildBaseIterator
This change introduces AXNode::ChildIteratorBase to AXPlatformNodeBase
and for AXPlatformNodeBase::UpdateComputedHypertext to utilize this
iterator for computing hypertext.

This change resolves a hang issue in UpdateComputedHypertext for a
large file with the following AXTree:
  RootWebArea
  ++genericContainer ignored
  ++++genericContainer1 ignored
  ++++++staticText "0;"
  ++++genericContainer2 ignored
  ++++++staticText "1;"
  ++++genericContainer3 ignored
  ++++++staticText "2;"
  ++++genericContainer4 ignored
  ++++++staticText "3;"
   ... 20,000x more ...
  ++++genericContainer20k ignored
  ++++++staticText "20,000;"

Previously, calling genericContainer1->UpdateComputedHypertext() would
return hypertext "0;1;2;3; .... 20,000;". This computed hypertext is
incorrect since we should not consider genericContainer1's siblings
for hypertext, and it also takes O(N) to compute where N is number of
sibling text nodes in ignored container. If we update all the nodes
for hypertext on Windows IA2 in BrowserAccessibilityManagerWin::
OnAtomicUpdateFinished, we would get O(N^2) and therefore the hang for
a document with large number of nodes in the above structure,
say 20,000.

In this new change, by using ChildIterBase in UpdateComputedHypertext,
calling genericContainer1->UpdateComputedHypertext() would give us
the correct hypertext "0;". It also mitigates the performance for
updating all the hypertext in the above AXTree, thereby fixing the
hang.

Bug: 1070783

Change-Id: Ifdb32202efac33d5eb6fad16fac5c187d4f9e657
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2160345
Commit-Queue: Victor Fei <vicfei@microsoft.com>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765147}