[CrOS Settings] Do not instantiate search v2 code without flag
Previously, the flag was used to update our settings search UI, but the
C++ code was still instantiated and running in the background even with
the flag off. This ends up using unnecessary resources and introduces
the potential for bugs in our updated search even when the flag is
disabled.

This CL prevents this possibility. Note that this CL will be present in
the M-84 build.

Change-Id: I5c3dee8755e3a06938d09f1fe08c65d00db1b9be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2180364
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Reviewed-by: Regan Hsu <hsuregan@chromium.org>
Auto-Submit: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765205}