Refactor enumeration of variable names
This CL simplifies how all known variables for a given ComputedStyle
are collected. ComputedStyle::GetVariableNames now collects the
names from all relevant sources, sparing call-sites of this burden.

Note that the (now-removed) comment in ComputedStyleCSSValueMapping::
GetVariables is no longer accurate: that traversal of the registry
was necessary before StyleInitialData existed, but it's not needed
anymore. It's better if the ComputedStyle itself is the single source
of truth as to which variables exist for that particular style.

It's probably not an author-visible problem, but imagine this:
if you have one ComputedStyle created at time A, then
ComputedStyleCSSValueMapping::GetVariables would report a certain
set of variables for that ComputedStyle. Then, later at time B,
a new property is registered with some initial value. If you were
to call ComputedStyleCSSValueMapping::GetVariables again with the
same ComputedStyle from time A, it would now report a different set
of variables (i.e. it would add the initial value from the registration
at time B), even though that didn't exist when ComputedStyle A was
created. This is not correct. The ComputedStyle needs to retain the
initial values according to the registered properties at the time it
was created (and we do, it would just be misrepresented if queried via
ComputedStyleCSSValueMapping::GetVariables).

As mentioned, it's probably not visible to authors, but it would be
nice to do the right thing anyway.

Change-Id: I947352f2fb33d60a6edfd95ad46e1939d8decb0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2171486
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765212}