remoting: Open PAM session in child process.
Previously, when the session relaunched itself after the user logged
out, it would do so using the same process that had managed the PAM
session. As a result, when the same process opened a new PAM session,
systemd wouldn't recognize it as a new log-in session, and would fail to
set some key variables (e.g., XDG_SESSION_DIR).

This commit modifies user-session to fork a subprocess to manage the PAM
session. That way, when the process relaunches and opens a new PAM
session, the new session is not a descendant of the old one, and systemd
properly creates a new login session and sets the appropriate variables.

Bug: 1077449
Change-Id: Iacea643e7f8d662ead1e7c87b6de05bfd170d36a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2176678
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Erik Jensen <rkjnsn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765237}