[Safety Tips] Streamline tests and remove race conditions.
This CL makes a few small quality-of-life improvements to the Safety Tip
tests to make them easier to write/maintain. In so doing, it removes a
race condition present in several of the tests wherein the tests would
check the reputation of the page prior to the check completing.

Bug: 1077484
Change-Id: I9898fe15ae7b4f7de2e529c5d83b57bcff2f7f0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2176986
Commit-Queue: Joe DeBlasio <jdeblasio@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765240}