[Autofill Auth] Implementing settings page opt-in on Clank
Unlike Desktop, where the opt-in request is sent immediately once the user
opts-in through the settings page, on Clank, the request is not sent until a
card is downstreamed.

This CL implements the opt-in request based on the user's intention from the
settings page.

Bug: 949269
Change-Id: I0380514670091641e312e00d042ca185c0b11f45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142150
Commit-Queue: Manas Verma <manasverma@google.com>
Reviewed-by: Jared Saul <jsaul@google.com>
Cr-Commit-Position: refs/heads/master@{#765264}