cc: Add heuristic to detect renderer frame rate with videos.
The display compositor can lower the overall composition rate if all
surfaces are known to drawn at the lower rate. While video surfaces can
provide an explicit playback rate, the same is not available for the
LayerTree compositor in the renderer. This prevents this optimization
from being used in a case like meets where the animations for the volume
bar are happening at a rate lower than the video streams, but there is
no explicit signal for it.

This change adds a heuristic to detect the renderer's frame rate during
video playback to allow enabling this optimization for this use-case.

R=sunnyps@chromium.org

Bug: 976583
Change-Id: Ia8664c6b3217eb124e79e09a11d39e9f8670bc48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2167515
Auto-Submit: Khushal <khushalsagar@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765276}