Refactor the metrics flow in IsolatedPrerender
Since many more metrics are being added to the after-SRP pageload UKM,
having a dedicated metrics class for it too is needed.

The AfterSRPMetrics class is set and fully populated on commit so that
the PLMObserver can simply copy it on commit.
AfterSRPMetrics is non-null only on the page load after the SRP.

Flow:
0. User goes to SRP.
1. If NavigationPredictor parses the SRP and updates the prediction then
   |srp_metrics.predicted_urls_count_| becomes > 0 and this is used as
   the source of truth for identifying pages where prefetching may
   occur.
2. Some amount of prefetching happens
3. User clicks on a SRP link and starts a new navigation
4. The URLLoaderInterceptor may intercept the navigation and use the
   cached response. When this happens, it notifies the TabHelper thru
   |OnPrefetchStatusUpdate|.
5. Commit happens. On commit, the TabHelper checks
   |page_.srp_metrics.predicted_urls_count_| to determine if
   |page_.after_srp_metrics_| should be created. When after_srp_metrics_
   is created, the last known status of the committed URL in
   |page_.prefetch_status_by_url_| is copied into the status member of
   AfterSRPMetrics. Then |page_| gets reset and |after_srp_metrics_| is
   moved into the new |page_|.
6. PLMObservers are notified of the new commit. The IsolatedPrerender
   PLMObserver copies the AfterSRPMetrics from the TabHelper and tada!

Many more status enums will be added in the next CL so that every
parsed SRP link will have a status that can be copied over to the next
pageload.

Then I'll send a UKM CL after all that.

Bug: 1074385
Change-Id: I785d778321a29a81bd58c2c27bf2c7a93188f90f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2175547
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765281}