Temporarily disable the EagerCacheStorageSetupForServiceWorkers feature.
Recently some changes landed for COOP/COEP that have made this long
standing feature unstable.  This CL temporarily disables the feature
until it can be stabilized again.

The feature can be safely disabled because its a performance
optimization.  It currently has somewhat marginal benefit, but we expect
it to be more important in the future and would like to keep it working.

Bug: 1077345,1077916
Change-Id: I88f4429a3845da329b761a2d86473b52f75905fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181302
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765286}