[Media History] Fix a race between reset and fetch
If we start fetching a feed, then reset and then return
the fetch we might store stale data. This adds the
concept of a reset token which is reset everytime the
feed is reset. When we store the fetch results we
check the reset token and if it does not match we
do not save the results.

BUG=1076058

Change-Id: I92cfbbdf114c2c8522d76f2fc7dde9f7f153a4fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173313
Commit-Queue: Becca Hughes <beccahughes@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765328}