Add the grey Focus to app icons when opening context menu
Content menus use the AppService mojom, so it is asynchronous. When
right click on an icon for context menu, the context menu callback is
not run yet, so the checking in Ash
(context_menu_ && context_menu_->IsShowingMenu()) can't return true to
draw the round rect for the icon.

Add waiting_for_context_menu_options_ to the checking condition, to draw
the round rect for the icon.

BUG=1076839

Change-Id: I0fbc2809dd11afacddde9cf21db4043cb5196fe0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174251
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765362}