[Offline indicator v2] Fix cc texture issue
The previous CL invalidates the cc texture and requests and update from the LayoutUpdateHost before hiding the Java view and starting the browser controls animation. However, we could still see the stale texture for a moment, though not as frequently.

With this CL, we will wait for the #didSwapBuffers() (the sequence is onCompositorLayout -> didSwapFrame -> didSwapBuffers) call to move on to the browser controls animation. This should mostly solve the issue.

One thing to note is that there is no guarantee the updated buffer will be the one currently displayed. If we still have issues, we can explore better options mentioned in the bug.

Bug: 1068346
Change-Id: I32d0de4359b3c086d7ba3e9fb744831b771ccfe9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174857
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Cr-Commit-Position: refs/heads/master@{#765372}