Fix DCHECK in BrowserList::MoveBrowsersInWorkspaceToFront
MoveBrowsersInWorkspaceToFront() should not be called with an empty
workspace argument.

Also, a recent change [1] would sometimes call
MoveBrowsersInWorkspaceToFront() with the workspace of the browser
window, not the current workspace.  This is incorrect, so a comment is
added to make usage clear.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2093438

BUG=1076913
R=sky

Change-Id: If3a0be8b40352315a8afea8741e3888e5790d9fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181576
Auto-Submit: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765374}