crostini: delay restart if vm is stopping
This change should be safe as it should just make CrostiniRestarter
appear to be slower. In situations where we do not want restarters to
run any more (e.g. removing crostini), we have code to abort all
restarters so the "delay restart" behavior will not happen.

Notes
=====

Restarter shutdown observer registration
----------------------------------------
CrostiniRestarter now registers vm shutdown observer in Restart()
instead of the ctor. This is to correctly handle the case when there are
multiple restarters for the same vm: all restarters will wait for the vm
to shutdown without being terminated, and only the first restarter,
whose Restart() has been triggered, will kick off the restarting process
in OnVmShutdown().

This will cause a subtle behavior difference in a rare situation where
1) there are multiple restarters, 2) the first restarter is aborted just
before it gets OnVmShutdown(). The old behavior is that all restarters
will be terminated (because all of them get OnVmShutdown()), the new
behavior is other restarters will survive and one of them will be called
on Restart() by CrostiniManager::OnAbortRestartCrostini(). Note that if
the first restarter is not aborted, all restarters will still be
terminated by the FinishRestart() in OnVmShutdown() of the first
restarter.

CrostiniManager::OnStopVm()
---------------------------
CrostiniManager::OnStopVm() does not call OnVmStoppedCleanup() any more.
This is because when the VM is stopped, CrostiniManager::OnVmStopped()
will be invoked, and it already calls OnVmStoppedCleanup(). Removing
this avoids CrostiniRestarter::OnVmShutdown() to be called twice for the
same shutdown.

Bug: 1075731
Test: unit tests and manually testing (including removing linux with multiple delayed restarters)
Change-Id: Ic34e5333d544a2ff849079a085f342e669ed9523
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2172187
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Jason Lin <lxj@google.com>
Cr-Commit-Position: refs/heads/master@{#765393}