Fix file renaming for chrome://media-app.
ArrayBuffer has no size so files that were renamed had 0 bytes.
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/ArrayBuffer

Also check the contents of the file in the RenameIPC test to prevent
this regression in the future. In doing so we update
FakeFileSystemHandle to be writable.

I don't think b/152832337 & b/152832025 are problems anymore with
this change and changes to loadFile in crrev/c/2174095.

Bug: b/155359417, b/152832337, b/152832025, 996088
Change-Id: I3ba74c5666d4b72f25fe23189fa2bd50f3d90e43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174019
Commit-Queue: David Lei <dlei@google.com>
Reviewed-by: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765409}