[iOS] Fixing add account button in account list view
In crrev.com/c/2160911, the SSO add account view is now opened with
a command send to the SceneController using the dispatcher.
Unfortunately the dispatcher was not set when the account list is
opened by the Google services coordinator (it is set in all other
situations).
The dispatcher is now set, and an EG test is added.

Bug introduced with: crrev.com/c/2160911

Fixed: 1076843
Change-Id: I4a1bb94f9ca6b0c78f8f85c59a707e955164c1d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2180062
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765442}