[Android] Use AppHooks#createAuthenticatorNavigationInterceptor()
As part of componentizing AuthenticatorNavigationInterceptor without
breaking downstream's usage of that interface, we needed to introduce
a variant AppHooks#createAuthenticatorNavigationInterceptorV2() method.
Downstream has now re-wired up an implementation of the new
AppHooks#createAuthenticatorNavigationInterceptor() that is identical
to its implementation of
AppHooks#createAuthenticatorNavigationInterceptorV2(), so we can
transition upstream back to using the former. We will then be able to
eliminate downstream's implementation of the latter and finally
eliminate the latter upstream, at which point all the scaffolding
introduced to enable this conversion will have been removed.

Bug: 1071390
Change-Id: Ib816019fa3e00e865e7b2cb78c3fd3f7800e9a9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173369
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765448}