Always output a row in our sql table.
chrometto expects no rows if the metric doesn't apply to this trace, so
we didn't output anything. So instead COALESCE the sum with zero so we
always output a row as long as there was a scroll in the trace.

Change-Id: Iaebc1e3efd4c8b9ac28774a0c83ae04b04f21df5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2182472
Commit-Queue: Stephen Nusko <nuskos@chromium.org>
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Reviewed-by: Mikhail Khokhlov <khokhlov@google.com>
Cr-Commit-Position: refs/heads/master@{#765548}