[omnibox] Extend OmniboxPopupModel to handle the header collapse button
OmniboxPopupModel encapsulates our exotic keyboard focus traversal in
the Views Omnibox.

This CL extends that to handle also the new Headers and the collapse
button.

This CL is one step towards completing support for that.

Further CLs will be required:
 1. Making the focus traversal skip hidden rows.
 2. Making Enter activate the chevron in this case.
 3. Highlighting the header when the chevron is focused.
 4. Drawing the focus ring around the chevron.

This CL is step 0, and we will tackle the rest after this.

Bug: 1078183, 1052522
Change-Id: Ia2f86b8df66f065a5ae3af565e06cad53b9edf33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2180855
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765580}