[Autofill Assistant] Introduce is_client_side_only flag to ui model.
This is mostly for upcoming support for payment methods and contact
profiles. Those values will be auto-created by the client and will have
this new flag set to true. There is no way to change this flag, which
makes it safe for the client to temporarily store certain data without
that data ever leaving the device.

Bug: b/145043394
Change-Id: Ide2dfa1ee07ba5bfdc4697de8c22b7867de82e2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2126930
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Marian Fechete <marianfe@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765586}