[signin] Add plumbing for WebSigninInterceptor
This CL adds the infrastructure for web signin interception.
The code is still incomplete, and the feature is disabled by default.

The WebSigninInterceptor is both a profile keyed service and a
webcontents observer, because it depends on the lifetime of both the
profile and the tab.

The interception flow is:
- DiceResponseHandler calls its delegate after an account was signed in
  on the web.
- The delegate ProcessDiceHeaderDelegateImpl simply forwards the call to
  the WebSigninInterceptor, along with the web_contents that initiated
  the signin (if it still exists).

Bug: 1076880

Change-Id: Ia3867eb535c8fd103d2db659d8889c1da5495aa7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2179202
Reviewed-by: Jan Krcal <jkrcal@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765589}