Add utility-sub-type command-line option
When debugging or profiling Chromium it can be helpful to know what type
of utility process is being debugged. Putting this on the command-line
can save developer time. This change uses the metrics_name_ for this
purpose and a --utility-sub-type option. Chromium ignores this option -
it is purely for developer convenience.

See bug 1075479 for an example of the value.

Bug: 1077885, 1075479
Change-Id: Ibbaaeb7a000bf35340835cf9d1dfe6e9d9cb2f40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181974
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765594}