[Paint Preview] Reduce copying in BuildResponse
There are a few crash stacks that mysteriously contain either just
paint_preview::BuildResponse or that + some V8 lines. Some of these are
OOMs. To attempt to reduce memory usage, this CL reduces the copying
needed when creating a mojo response for a capture. I also added some
nullptr checks as some of the stacks claim there is a segfault.
Although, I'm not sure how that would be possible.

Bug: 1077746
Change-Id: I04416d8b122f020121dd67440dd8bded184ab3a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2182524
Reviewed-by: Ian Vollick <vollick@chromium.org>
Reviewed-by: Mehran Mahmoudi <mahmoudi@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765595}