Coupling hover enter/exit AccessibilityEvents on Android
This CL implements a coupling mechanism for certain pairings
of AccessibilityEvents on Android.

The premise is that some event types have a sibling type, such as:

TYPE_VIEW_HOVER_ENTER
TYPE_VIEW_HOVER_EXIT

Until now we have handled each of these separately, and neither has
been part of our throttling performance improvements.

This CL makes two changes, first it throttles the rate at which
TYPE_VIEW_HOVER_ENTER events are fired. This leverages the existing
AccessibiltiyEventDispatcher, but with a new throttle delay of 50ms.

The second change is that HOVER_EXIT events are no longer treated
independently, and instead they are always fired in pair with a
HOVER_ENTER event. This saves computations and it accomplishes
the same throttling result for EXIT as it does for ENTER events
without risk of us becoming out-of-sync with what we think is
focused/hovered vs. what the Android system thinks.

This approach is general and can be applied to any set of sibling
events, although this CL only addresses HOVER for now.

Change-Id: I7fd2fbbc1ea62639999dda47a8521d7b57ce310d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2175751
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/master@{#765624}