wm: Add drag to snap maximize.
Adds functionality to snap a dragged window to the top of the
screen and snap it to maximized state. When the drag on caption
reaches within 32 dips of the top edge, or past the work area,
a phantom spanning the work area will be displayed. If the mouse
is released while this phantom is visible, the window will be
snapped to maximized state.

Test: ash_unittests WorkspaceWindowResizerTest.DragToSnapMaximize
Bug: 1066159
Change-Id: I6cbadfea764465bb8bd63fff8277e52a7dec884e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2168000
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765661}