Fix the only ShouldUseProcessPerSite() call not using a site URL.
All calls but one were passing a site URL to
ShouldUseProcessPerSite(). This change fixes that one call
so it passes in a site URL. A number of parameter names in
related functions were also updated to make it clear that
a site URL is expected to be passed to those functions.

Bug: 1015882
Change-Id: Ia344debe8237b635cc377bed8200349a4b2c24f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174875
Commit-Queue: Aaron Colwell <acolwell@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765693}