Fix testharness.js's assert_object_equals error message
When using assert_object_equals, if the two objects have the same
property but with different values, then the error message used to say
that the actual and expected values of the property were the root
objects themselves.

This patch fixes it so that the error message references the actual and
expected value of the property.

Change-Id: If9e6a5445db13444e22fd58729614c8c73046aea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181322
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765713}