[Offline indicator v2] Ensure visibility is updated immediately when not animating height change
After landing a fix for the compositor texture invalidation issue where
we changed the code to invalidate the texture then wait for the request
to be effective (a buffer swap with the newly rendered frame), the test
for showing/hiding on Start surface started flaking.

The reason was that even though we don't wait for any cc-side browser
controls animations on Start surface, we ended up waiting for a buffer
swap to proceed with hiding.

This CL uses the canAnimateNativeBrowserControls supplier to decide how
to proceed when the Java animation ends.

Bug: 1078252
Change-Id: I1b2b513b2f14c44e220a64d3dbabecce8bf988a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2182399
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765732}