chromeos: Fix input method sync on first login with SplitSettingsSync
Chrome OS only applies a user's preferred input methods once, on first
login. We presume that further configuration is per-device, and if a
user has multiple devices they don't want to sync IMEs back and forth.

This is handled via InputMethodSyncer, which checks its instance of
PrefServiceSyncable to check if the initial sync is complete.

SplitSettingsSync changed the timing of the initial sync of OS prefs,
deferring them until after the OOBE sync consent dialog is shown.
It also introduced 2 additional PrefModelAssociators and changed
the semantics of IsSyncing() and IsPrioritySyncing().

Instead, we need to provide a mechanism for observers to detect if
OS prefs are syncing. Add functions AreOsPrefsSyncing() and
AreOsPriorityPrefsSyncing() to PrefServiceSyncable to support that.
Change to the new functions for:

* Input methods (IME)
* User image
* Shelf position and autohide
* Shelf app item pinning

Also fix call sites that were testing for the wrong type of pref
(e.g. checking for regular prefs when they wanted to check for
priority prefs):

* Timezone resolution method
* "Natural scroll" default setting

      sign in, input methods are restored

Bug: 1076676
Test: see bug, configure some input methods, reset Chromebook and
Test: added to components_unittests
Change-Id: Idb158681d7b0e52dc56eddba5fc4045de27b903f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174860
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765740}