Adding an 'Echo' message and logic to Chromium FTL signaling classes
This CL adds a new message which can be used to check whether a CRD
endpoint is receiving and responding to FTL messages.  In order to make
this as simple as possible, I didn't want to reuse the existing message
payloads which may have side-effects or require special processing.

The main use case is for the client to be able to ping the host to see
if it is online before connecting.  I think we will also use it for
checking to see if a reconnect is possible after a connection drops.

Change-Id: I8825451bebc41c4219186bf7e54d79e3229d9717
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174966
Commit-Queue: Joe Downing <joedow@chromium.org>
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765750}