Fixed several visual bugs for the eye dropper.
This CL fixes the following visual bugs:

1.the magnified region doesn't match the cursor position: this
happens for a device scale factor != 1 because we were not correctly
matching widget's bounds to the captured frame; the fix here is to
convert the widget's bounds to screen coordinates

2. the circle seems to shift lower sometimes: because of scale factor
the view's width can be different than its height which contradicts
the current assumption when computing the padding; the fix here is to
change the padding to gfx::Size.

3. large black box appears on and off: this was caused by the widget
not being translucent; the fix here is to force this by using
force_software_compositing.

For now I have added some pixel tests to validate the UI, but I am
planning to add more tests in future CLs.

Bug: 1077817
Change-Id: I42c3b285ce0f65b2acfca752b9429db175fbcc9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181563
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765869}