Desktop PWAs: Open popup windows from browser tabs
When the source URL was in the scope of an installed app (including
a shortcut app), we were responding to NEW_POPUP requests by opening
app windows even when the source was open in a browser tab. We now
only open app windows if an app id was requested in NavigateParams,
or the popup is being opened from an app window.

Bug: 1068443
Change-Id: I19ef25dae2a9d7ae3ab953c036759d30d0e76b31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2178478
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Commit-Queue: Michael Wasserman <msw@chromium.org>
Auto-Submit: Eric Willigers <ericwilligers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765871}