Remove editing gesture logic
This is part of a series of changes improving ChromeVox with touch.

This change removes gesture logic to move the caret while on an editable. ChromeVox commands such as nextCharacter should be capable of setting selection, which then moves the caret.

R=akihiroota@chromium.org

AX-Relnotes: Makes navigation smoother when performing touch gestures with ChromeVox and moving across editable text fields.
Change-Id: I966eb79621296fe548adff0af27fc6867e46d545
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181656
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765893}