[iOS] Remove WaitForContextMenuItemDisappeared
This wait was added because EarlGrey was not properly synchronized with
the dismissal of the context menu. EG tests now appear to pass
correctly without this additional wait so it can be removed.

This CL is safe to revert if context menu failures or flake appear.

Fixed: 682871
Change-Id: Ied8bb2a510e5411e8658be1095fef02e378dae98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2175621
Auto-Submit: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Eric Noyau <noyau@chromium.org>
Reviewed-by: Eric Noyau <noyau@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765943}