[b4p] Re-enable password generation for opted in users in local mode
crrev.com/c/2157529 disabled password generation if the user was opted
in to passwords account storage but had the profile store as default.
This CL re-enables generation for those users. This makes the behavior
more consistent with password saving for example, where the user is not
prevented from saving to the account just because they *currently* are
in local mode. The message in the generation popup should be enough to
warn users that the password will be saved to their account and not
locally.

Bug: 1060131
Change-Id: I453ae66802651300d1faf5b52ba2cc627e7073ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2184152
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#765955}