[Android] Share getAvailableContext() logic in intent launching impls
This CL shares the core logic for getting the available Context between
the Chrome and WebLayer embeddings of //components/external_intents. To
do so we move this logic into ExternalNavigationHandler.java, with a
new API to the embedder to get the Context associated with the Activity
if present. As this method is invoked by the delegate impls we make it
a static method on ExternalNavigationHandler to have it be available to
those impls without them needing to have a reference to the handler
instance.

Followup work will build on this by sharing the implementations of
startActivity(IfNeeded).

Bug: 1071390
Change-Id: If3fb2cd8475df5cc874848946b2cc1a72ed7a4c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181385
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765960}