[css-pseudo] Support 'revert' in ::marker
The default UA styles for ::marker pseudo-elements used to be directly
assigned to the ComputedStyle, without participating in the cascade.
Then the 'revert' keyword didn't take them into account, and just
reverted to the initial styles.

This patch fixes the problem by adding these styles to the cascade.

There was an early return optimization in case there was no selector
targeting the ::marker, I'm removing this in order to run the normal
code that takes care of applying the cascades styles. The relevant perf
test, flexbox_with_list_item, doesn't seem negatively impacted.

This early return had the non-obvious side-effect of avoiding style
resolution for nested pseudo-elements, which was good since that's not
supported yet and would trigger DCHECK failures. I'm replacing this with
a more explicit conditional.

The change also affects list-color-change-no-layout.html, since now
updateStyleAndReturnAffectedElementCount() will also include a ::marker
if it exists, not just if it has author styles.

BUG=457718

TEST=third_party/blink/web_tests/external/wpt/css/css-pseudo/marker-reverted-styles.html

Change-Id: Ica9e438288b5a613863876a45947f3b66ffe5d88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2182427
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765965}