Store original CSSPropertyID when applying interpolations by lookup
Each CascadePriority has 32 bits (the lower 32 bits) reserved for the
"position" of the declaration. The position bits have two purposes:
1) making it possible to locate a given declaration with random
access, and 2) serving as the "order of appearance" criterion in
css-cascade.

For regular declarations, the position is the (composite) index into
the MatchResult.

For interpolations, there is no numeric declaration-level index to
store in the "position" bits, because interpolations are stored in
HashMaps indexed by the CSS property name. A numeric declaration-level
index is therefore also not necessary, since (obviously) only one
interpolation can exist in the HashMap for a given CSS property name.

So, if we want to apply the CSS property "color" with random access
(LookupAndApply), one of the following things happen:

 - If the winning declaration is a regular declaration (UA, user,
   author): we use the position bits to access the right location in
   the MatchResult.
 - Or, the winning declaration is an interpolation (animation,
   transition): we know from context which property we want to
   apply, so we just look up the interpolation using the CSS property
   name.

In a world without surrogates/css-logical properties, the above would
be perfectly fine, but of course we don't live in such a world.

As of recent changes to surrogate handling, an animation for
'inline-size' for example will be added to the CascadeMap as its
corresponding physical property (let's say 'width'). The property
'inline-size' is never added to the CascadeMap at all. So, if we want
to apply 'width' with random access, and the winning declaration is
an interpolation, we look up 'width' in the map of interpolations,
but we can't find it, because it really existed in the interpolation
map as 'inline-size'. Hence, we need a way to locate the original
interpolation.

This CL does that by adding the original CSSPropertyID to the position
bits, making it possible to locate the right interpolation, even when
the CascadeMap contains one thing, and the interpolations map contains
another. It would be nicer if active interpolations were stored in a
Vector rather than a HashMap, and long term we might consider switching
to that.

Change-Id: I2225166d16547408a22040b42f463999cc4612ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181288
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765985}