Add a policy setting for Get Image Descriptions
This provides a way for administrators to disable the
Get Image Descriptions from Google feature, which sends
data to Google servers. Note this policy setting in the
traffic annotation for image labeling.

Manually tested. Confirmed that if I set the policy to false,
then the feature is disabled, it no longer shows up in
the context menu and in Settings it has the icon
indicating it's controlled by policy.

Bug: 1077102, 1070505
Change-Id: I818f1b81710d9d37c06167fbe1f6fe2489e86c11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174865
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Reviewed-by: Katie Dektar <katie@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Auto-Submit: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766002}