[WebLayer] Use embedder name for Page Info UI.
This adds a way for the embedder to set embedder name to be used for
Page Info UI. This is currently used on the WebLayer side to set the
name of the WebLayer embedder as soon as the native library has been
loaded.

A simple browsertest verifies that this name is set to the expected
string.

For both Clank and WebLayer, IDS_PAGE_INFO_SECURE_IDENTITY_VERIFIED has
been updated to take the embedder name and IDS_PRODUCT_NAME is used for
non-WebLayer code which resolves to the content embedder's name
(Chromium, Chrome etc).

Bug: 1052375
Change-Id: I2b8ebc8f5f49e8eac2010bdafb2bfe1908f213dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2179524
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766006}