[iOS] Preliminary multi-session support for multi-window.
This is the basics of multi-session support. It adds a SessionID to
the session restoration agent, which is used as a subdirectory for
session storage. When multi-window isn't enabled, this is ignored and
the usual storage path is used.

With multi-window, the persistent UISceneSession ID is set as the
session ID for the session restoration agnet.

This CL doesn't implement many of the further features that are needed:
  - Clearing closed sessions.
  - Handling cases where a Chrome session (set of open tabs) changes
    which UISceneSession it's associated with (and thus changing its
    session ID). I suspect this can happen, although I haven't yet
    determined how.
  - downgrading from this model back to regular sessions.

(and so on).

This is primarily to give this system a shakedown in Canary without
disrupting stable.

Bug: 1070794
Change-Id: Iddc20fe7e7542d5f5adb826978c412b84bcc15cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2179165
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766010}