[iOS] Fix NavigationManager tests
This CL fixes NavigationManagerTest.Restore. It was failing because of
crrev.com/c/2172325 which only restores the UserAgent for iOS 13.
The change was cherry-picked but the updates of the test was done in
another CL which wasn't cherry-pick.

This CL is only updating the test to make it pass.

Fixed: 1078713
Change-Id: Ia7be2e3d9eafb054daf2498479cd73218417e3a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2184150
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/branch-heads/4103@{#467}
Cr-Branched-From: 8ad47e8d21f6866e4a37f47d83a860d41debf514-refs/heads/master@{#756066}