Reland "Desktop Reader Mode should not load insecure / mixed content."
This reverts commit 05cdf7552ded731b0156bc7c25ece2fbb88bf1ff after
fixing flaky tests by removing sanity-check of unblocked image.
It is likely that the image has not completed loading before the
DistilledPageObserver returns. A future change will add the
sanity-check image back in, but it is not necessary for testing
whether Reader Mode is loading insecure/mixed content, so relanding
with it in a TODO for now.

Original change's description:
> Desktop Reader Mode should not load insecure / mixed content.
>
> Reader Mode on Desktop will only be run on SECURE pages. To avoid
> any security level downgrade, no mixed content will be loaded on
> these pages, and resources with invalid certificates will also not
> be loaded.

--enable-features=NetworkServiceInProcess with --test-launcher-jobs=500
many times. Reproduced flake before change, and not after.
TBR=creis@chromium.org,carlosil@chromium.org

Test: Ran with --enable-features=ForceWebRequestProxyForTest and
Bug: 1069907,1078547
Change-Id: I00978f1a363bef08fd974c957fd7e8a19be79143
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183252
Reviewed-by: Katie Dektar <katie@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766125}