[PM] Add dedicated worker clients to the graph
This CL connects dedicated workers and to their client frames in the
performance manager graph.

To work around the possibly missing frame node for the client, the
WorkerWatcher now ignores client-worker connections when the frame node
could not be found.

A new histogram was added so that we know how often that happens.

Bug: 1078161
Change-Id: Ie54d2b84ddb1e958416cb06b0e8441eb6a3a113f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181712
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766235}