[WindowsElasticOverscroll] Allow ElasticOverscrollController creation.
This CL enables ElasticOverscrollController for the root scroller on
Windows to be instantiated. Turning on elastic overscrolling works
for PTP but not for touch scrolling, as it does not have the necessary
inertial_phase information that InputScrollElasticityController relies
on. This CL adds that information for touch scroll events.

Bug: 1058071
Change-Id: Ibcb1a373def9564a7b0ea81f62493d455055a8a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183709
Commit-Queue: Rahul Arakeri <arakeri@microsoft.com>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766317}