[Security][Coop] Parse COOP report headers
This adds the parsing of the report-to parameter in the
cross-origin-opener-policy header as well as the parsing of the
cross-origin-opener-policy-report-only header.
To that end this changes the mojom structure representing the COOP
value.

These changes are fairly similar to what was implemented by yhirano@ on
embedder policy, first in https://crrev.com/c/2060090. The same logic
applies regarding the introduced typemaping.

Bug: 1076456
Change-Id: Iaa331b4a177265e585b8d16da359bbd072aef09f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2172695
Commit-Queue: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Auto-Submit: Pâris Meuleman <pmeuleman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766330}