[Autofill Assistant] Added integration test for server cards.
For now, this is simply a smoke test that opens the edit dialog for a
server card and verifies that this does not break the client. b/155624806
tracks progress for adding a more elaborate test, this CL is merely to
get the effort started.

Bug: b/155624806
Bug: b/153498495
Change-Id: I02cdef5115d011748410b72de38913c2cb262ede
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2179883
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Reviewed-by: Marian Fechete <marianfe@google.com>
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#766337}