[css-pseudo] Fix legacy ::marker originated by pseudo-element multicol
This patch fixes a heap-use-after-free in legacy layout when having a
an inside ::marker originated by a ::before or ::after pseudo-element
which is dynamically converted into a multicol container.

The problem was caused by some code that was aiming to prevent the
::marker from being placed after the generated contents of the ::before
or ::after. To do so, it reinserted the ::marker as the first child of
the originating list item. The problem was that, in the multicol case,
the LayoutInsideListMarker is a descendant (not a child) of the
LayoutListItem, since there are a LayoutMultiColumnFlowThread and a
LayoutBlockFlow between them.

Therefore, this patch reinserts the ::marker as the first child of its
parent, instead of as the first child of the list item.

BUG=1078236

TEST=third_party/blink/web_tests/external/wpt/css/css-multicol/multicol-list-item-002.html

Change-Id: I580e4730d6e8ec6d41e423d0d027f67de4d48fa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2185030
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/master@{#766374}