[devtools] Add structured request cookies to requestWillBeSentExtraInfo
This allows the devtools front-end to provide better information and
to support more accurate filtering in the network view.

Previously, cookie information for requests was loaded separately
from the cookie store, and in some cases this caused the cookie
information to be out-of-date or inaccurate, because the headers
only mention the cookie name, but to identify a cookie in the cookie
store, the triple (name,domain,path) is required.

Additionally, this ensures that full request cookie information is
present in the request without consulting the back-end again.

Bug: chromium:1027752, chromium:1051163
Change-Id: I95e1ba7993eaeeebcfdec33e3ff571efb9466473
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183937
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766376}