[Autofill Assistant] Make client_side_only sticky.
This change ensures that values that derive from client_side_only values
are also client_side_only, such that neither the value itself nor values
derived from it can be sent to the backend.

Since the only way to derive values is via a ComputeValue interaction,
the additional complexity seems to be manageable. We may have to consider refactoring in the future if there are other ways of deriving values.

This CL also removes some unused code in value_util.

Bug: b/145043394
Change-Id: Ida9642903b3ef5bd938930163dd422250eb657bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2184270
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Marian Fechete <marianfe@google.com>
Cr-Commit-Position: refs/heads/master@{#766423}