[Switch Access] Enforce menu layout is refreshed
As part of the process of transitioning to using the new Switch Access
menu, it was discovered that the menu would often appear blank. Some
digging revealed that the root of the issue was that InvalidateLayout()
was never being called. This change fixes that.

AX-Relnotes: n/a
Bug: 973719
Change-Id: I24a2f1a033bc1f11b8bea6296f774f210770577a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2176258
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/master@{#766480}