[Switch Access] Add back button focus ring
The Switch Access UI specification (see bug) indicates that the back
button should have a perfectly round focus ring exactly around the back
button. The general accessibility focus ring infrastructure is not set
up to be round, or to draw exactly adjacent to a UI element, so the
preferred approach is to modify the back button's appearance to include
the focus ring, when the node should be focused.

AX-Relnotes: n/a
Bug: 973719
Change-Id: I80aaa036d13192ce8b60fb2abaeb573a3f8ac5a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181428
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766483}