viz: Don't check for DeJelly feature every frame
HandleDeJelly adds a trace and looks up if the feature is enable each
frame.

Looking up that feature flag 60 times a frame shows up in flame graph
as a tiny percentange of time, but overall taking ~4.8ms cpu time in
a 45 second CPU time trace:

https://pprof.corp.google.com/?id=0103069a7a440702d31d9d624b2b8bdb&flamesearch=DeJelly


This CL avoids the map lookup and avoids adding the trace when dejelly
is disabled.

Change-Id: I57201dfcedd11dbc26cfc3deb7ce8d1c0b636444
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2188514
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766608}