Invalidate browser frame on touch UI changed.
Calculation of browser window client view area happens on browser frame
layout. For the case outlined in the bug (browser starts in tablet mode,
then changes to standard mode) the top inset of the client area remained
equal to the height of the caption bar even after switching back, which
caused the top half of the tabs not to render properly.

This would persist until the entire browser was forced to re-layout,
which would happen on minimize/restore or other similar operation.

This CL invalidates the frame on touch mode changed rather than just the
(lower-level) non-client view, if a frame is present. This change forces
a re-layout of the frame view, which correctly recalculates the client
size based on the new value of tablet mode.

Bug: 1035959
Change-Id: I803301f3b960277c2a9941950b856a59156dbfdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2186739
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Connie Wan <connily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766614}