Files app: Dialog text style
Change all dialogs to mark the dialog frame when it's displayed without
the title, to be able to change the color of the text as per spec.

When there is a title, the text is a secondary color. When there is no
title the text gets the primary color.

Add 16px top margin for elements displayed after the dialog text in
DefaultTaskDialog and InstallLinuxPackageDialog.

Bug: 992827
Change-Id: Ia5d81e32de881e5145cb55f06e3bdca7039ebac0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2186834
Reviewed-by: Noel Gordon <noel@chromium.org>
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Auto-Submit: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766712}