Remove orphan git submodule entry third_party/pywebsocket/src
Git contains a node third_party/pywebsocket/src with mode 160000. This
should correspond to a git submodule, but we do not have a .gitmodules
file. This seems to be an inconsistency (that breaks, for example, the
command "git submodule status"). I guess this file was committed by
mistake and should be removed.

Change-Id: Ic74c2541edc289342b64afdcddf298f80f665e15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2187371
Reviewed-by: John Delaney <johnidel@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766724}