components/arc/vea: return more detailed error from Initialize()
The current Initialize() method can report that the VEA creation has
failed, but doesn't give any hint on the nature of the error. Add a new
Initialize() method and mark the previous one as deprecated to the ARC
client can react according to the specific failure.

BUG=b:149441744
TEST=VEA usable on ARC++

Change-Id: I9db8e382a0715120cc2699d716480b3003923bf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068234
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766742}