Files app: Fix context menu icon for archive type
It showed broken because the menu style to push the icon to the right
applies to background-image and the icon for archive type was using
-webkit-mask-image. The -webkit-mask-image doesn't work well when the
same element has text, it was removing the text.

To fix the issue, force the desired color for this icon in the SVG file
and use it as background-image. This icon is almost always displayed in
this grey color, only in the directory tree it shows as blue, but this
is still working because there it it uses -webkit-mask-image.

Bug: 1080504
Change-Id: If1231b4bc7c177d1b8f5bdb7c3c60e9f5a3c66fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189898
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Auto-Submit: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766752}