[Launcher settings] Guard use of search_handler_
The SearchHandler currently depends on a different chrome flag than the
one guarding the construction of the OsSettingsProvider, and will be
nullptr if that flag is disabled.

This CL adds a null check to handle this. The check may not be
necessary long-term if the depending feature is launched.

Bug: 1068851
Change-Id: Iaf769446d3f56d49419e9e90a88daaa31ff498f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189839
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766769}