Fix mojom deps in content/public/common.
The public_deps for the mojom targets in content/public/common are
mostly dependencies for the typemap traits, which are largely common
due to the use of the shared common_param_traits.h.

common_param_traits.h includes header files from:
  //base
  //ipc
  //services/network/public/cpp:cpp_base
  //services/network/public/mojom
  //third_party/blink/public:blink_headers
  //third_party/blink/public/mojom:mojom_platform
  //ui/accessibility:ax_enums_mojo
  //ui/base
  //ui/gfx/ipc
  //ui/surface

These targets are added to the public_deps for the typemap traits.

The direct mojom dependencies are reduced to those mojom targets that
are imported in the mojom files. We need to keep the dependency on
//third_party/blink/public/mojom:mojom_platform in the mojom targets,
since (as we are outside blink) we need to rewrite that dependency to
//third_party/blink/public/common to support the component build.

Change-Id: I65de458dcc831a29d165a895be491aa308bc8e3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2185049
Commit-Queue: Mark Brand <markbrand@google.com>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766779}