Speed up NGPaintFragment::AssociateWithLayoutObject() for DCHECK_IS_ON.
Per the report in the bug, the existing DCHECK_IS_ON block has high
cost.  I assume this is because it is O(n) in the length of the
|layout_object|'s fragment chain when the rest of the function body has
lower complexity than that.

The goal of the existing DCHECK_IS_ON block is to detect an attempt to
add the same fragment twice.  Since we've already DCHECKed that
|next_for_same_layout_object_| is null, we need not compare |this| to
every entry in the list of fragments for |layout_object|, but only to
the last entry -- since the last entry will be the first entry with a
null |next_for_same_layout_object_| and thus the first entry that can
possibly be |this|.

For the case when |add_result| is a new entry, this can be trivially
checked: if (!first_fragment), then no check is needed, and otherwise,
we can DCHECK_NE(this, last_fragment).

If it is safe to assume that a fragment in the middle of such a list
will not suddenly have its |next_for_same_layout_object_| field
cleared (note, I don't know that it's safe to assume this, since I
don't know this code; but that seems like a reasonable constraint to
expect given how this code works), then in the case where |add_result|
is not a new entry, the stored value on that result is the only element
we need to check.  Thus once again this reduces to
DCHECK_NE(this, last_fragment).

This CL implements the above transformation, along with some code motion
to combine the common tails of the two cases where FirstInlineFragment()
was non-null.  The result of this is that the check becomes an O(1)
additional cost to the rest of the function.

Bug: 1080406
Change-Id: Ia70f3216887f6f546585252eed157be42bdc5a54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2190030
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766793}