[Autofill]  UMA collection for address import requirements.
Before an address profile is successfully imported from a
submitted form into Chrome, the requirements are first
determined based on the country of the form and subsequently,
the filled values are tested against those requirements
and basic validity checks are performed.

With this change, the outcome of the individual steps of this
process is logged to UMA to get a better understanding of the
requirements in individual countries.

Change-Id: I8c2f1a6bae48d88d2294dd17af00de5c08ba0a5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2172304
Commit-Queue: Matthias Körber <koerber@google.com>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766833}