[OSL-RV] Converting timing metrics to AutoCloseable interface
This patch extract the repetitive code for measuring the time it took
to execute a method to a separate Autocloseable class. Unfortunately
due to AutoCloseable#close() offering a throw Exception declaration
it is not possible to use the type directly without adding a catch
clause. That is why SuggestionsMetrics.TimingMetric needs to be public.

Bug: 1075602
Change-Id: Iae99d2b4d5546493056274f2b66c5e2eca4fc961
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181032
Commit-Queue: Filip Gorski <fgorski@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766834}