Add a FeatureParam to control min threads allowed in ThreadPool.
In bug 960012 we are trying to move cache_storage operations from the
browser IO thread to a SequencedTaskRunner backed by a thread pool.  On
ChromeOS we surprisingly got significant performance regressions.  The
scheduling team noted that ChromeOS gets the desktop minimum number of
threads set to 16.  Given that ChromeOS can have low number of cores
this may be too large.

This CL adds a FeatureParam to allow us to control the minimum number
threads in the ThreadPool so we can experiment with different values.
The goal is to perhaps set a more appropriate thread minimum for
ChromeOS in the future.

Bug: 960012
Change-Id: Ibe6bec5f9cbd06008bb982d9ba71ad791b64024b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2188650
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766838}