Remove unused spinner theme
This theme does not affect the style of the only spinner in Settings
(Clear Browsing Data). I suspect it was used on KitKat before
Preferences was migrated to the Support Library, but has no effect on
current UI that I've been able to identify.

Bug: 1077375
Change-Id: I497d290f337ddd21f9eb82b3d291135dd35a468f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189690
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Natalie Chouinard <chouinard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766875}