MacPWAs: Fix double-open of non-open-in-window bookmark apps
If an app shim specifies a URL, but fails to launch a PWA, then, as a
fallback, Chrome opens that URL.

If the app shim is a open-in-tab shortcut, then this will result in
two copies of the shortcut being opened -- one for the launch (which
just launched a tab, not a PWA), and one for the fact that we didn't
launch a PWA.

Rename the status code for "not a PWA" from AppShimLaunchResult::
kNoHost to AppShimLaunchResult::kSuccessAndDisconnect, which is more
clear. Don't launch the fallback URL when that is the status.

Also, move the handling of that error condition to a slightly later
part of the connection process (in anticipation of a refactor).

Bug: 1070163
Change-Id: I2cc8b36fc739476d3623a3a74187f829b00891ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189956
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766894}