[CrOS Settings] Add InternetSection search tags
Generated by running a script to parse go/cros-settings-search-tags.
This CL also migrates InternetHandler to this section and updates the
related test.

Note that any search tags which point to a "network detail" page do not
work properly because the URL path needs to have the network GUID added.
No other URLs so far have required this, and it will take a refactor to
add this functionality. I'll add this in a follow-up CL.

Bug: 1076268
Change-Id: If8b5759c919da685058d6e87011226a287622de8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189254
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Regan Hsu <hsuregan@chromium.org>
Auto-Submit: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766909}