[Safety Tips] Reland: Streamline tests and remove race conditions.
This CL makes a few small quality-of-life improvements to the Safety Tip
tests to make them easier to write/maintain. In so doing, it removes a
race condition present in several of the tests wherein the tests would
check the reputation of the page prior to the check completing.

This is a reland of crrev.com/c/2176986 with a minor tweak that ensures
that the reputation-check-has-run latch is only flipped for navigations
where the reputation checks were actually run. On macOS, tabs were
loading with an about:blank tab prior to the test loads, which were
causing the latch to get flipped prematurely.

Bug: 1077484
Change-Id: Iaf0816a5681e6bfc133cf0337ee7eb022aa2e116
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2186208
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766913}