Fix bug where avatar toolbar button would be the wrong size.
Was causing a state where text in the avatar button was unnecessarily
elided, even when there was sufficient space to display it.

This is a workaround for a deeper set of issues around caching and
updating ToolbarButton/LabelButton preferred size and layout.

In general, after updating text of a ToolbarButton, it must be resized
as part of layout for it to get its correct preferred size, which then
must be laid out again. Normally, this simply manifests as inefficiency,
but inside an AnimatingLayoutManager it can actually cause the layout to
get stuck on the wrong size.

This CL applies a spot fix to AvatarToolbarButton where once the text is
updated the button is artificially resized once in order to force the
necessary updates, and the rest of the layout process then proceeds
normally with the label given its appropriate size.

A thorough fix of the underlying problem is probably not possible without a more thorough solution to the
LabelButton/ToolbarButton sizing and layout code; see crbug.com/1078221

(cherry picked from commit 6bd033efc02736127abc25d18cea63d41b03ef88)

Bug: 1050168
Change-Id: I8f94df6af190b1061f04cf6cd2f6cdc7a0de6def
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2180859
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#765716}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2190748
Reviewed-by: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/branch-heads/4103@{#491}
Cr-Branched-From: 8ad47e8d21f6866e4a37f47d83a860d41debf514-refs/heads/master@{#756066}