Add null check for webcontents in AwContents::DidFinishNavigation
There have been cases where SecurityInterstitialTabHelper::
FromWebContents was being called with a null contents and causing a
crash. This fixes that by adding a null check for web contents and
skipping the interstitial check if it is null, since an interstitial
can't be showing in that case.

There is no regression test because I wasn't able to trigger the crash
at all when testing by hand, so this fix is based on the stack trace
from the crashes.

Bug: 1075889
Change-Id: I9ea2e981dad6d25ff1c8ffeb4e6580a1741ffb8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181149
Reviewed-by: Tao Bai <michaelbai@chromium.org>
Commit-Queue: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766942}