[conversions] Gate API usage on a feature policy
This CL adds a new 'conversion-measurement' feature policy, that is
enabled for self by default. This feature policy controls impression
declaration and conversion registration for the API. It also allows
impression declaration in subframes with the policy enabled.

Console messages are logged on attempted declaration/registrations that
do not have the policy enabled.


Bug: 1077877
Change-Id: Icea98deb5b365eced905e47134fd3c9c0c073998
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2180763
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766950}