Fix test_buildbucket_api_gpu_use_case retval
Fixes test_buildbucket_api_gpu_use_case.py always returning 0 even if it
reports failures, which was causing LUCI to be unhappy when a failure
actually occurred.

Test is still failing, but the retry should work properly with this and
allow the CQ to absolve CLs of the failure.

TBR=martiniss@chromium.org

Bug: 1080749
Change-Id: I42109c306f0c5b224b34518beb8d18be1774095f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2191156
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766959}