Restore software GL override in GLSurfaceTestSupport on Fuchsia
This change is currently a no-op; the default GL implementation is
swiftshader-gl on Fuchsia regardless of whether GLSurfaceTestSupport is
used or not.

Once we set the default GL implementation to ANGLE, tests will start
using the real GPU without this patch. Restoring the override avoids
this behavior change.

Bug: 1075803

Change-Id: I5d18a05ff0a8e97c09f76229783652e3414722cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2191123
Commit-Queue: Michael Spang <spang@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767017}