Expose installer progress from update_client::UpdateClient.
The change allows an installer to provide installer progress, as a
percentage value, and then have update_client::UpdateClient surface
that value in the state update repeating callbacks, which callers
of update_client::UpdateClient could consume, for instance to display
progress in their UI.

This change is needed for the code in //chrome/updater. The
change is mechanical relative to the existing component updater
code. The component installers don't capture installer progress and
don't need/use this feature for the moment.

The progress is injected by invoking a callback provided to the
component installer when CrxInstaller::Install function is called.
The progress callback must be invoked directly instead of being posted,
and no invocation of this callback must occur after the completion
callback is posted.

Change-Id: I08eb653800bce0c5f79e0dab5985c1a5cebe8d7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2177492
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767033}