Add new metrics for schemeful same-site downgrade
New UMAs that count, on a per cookie basis, what kind of context
downgrades occur, along with the scheme of the origin causing the
downgrade, for a cookie which was either included or excluded
only due to insufficient same-site context.

New UKMs which capture similar information to the UMA.

New Use Counter which captures the UMA/UKM information on a
per pageload basis.

Bug: 1030938
Change-Id: I12147e400f5952b0d2a0290e99365e4fd460cbd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173486
Commit-Queue: Steven Bingler <bingler@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767066}