[Extensions UI] Introduce a Settings Overridden dialog
Introduce a dialog to inform the user that some of their settings were
overridden by an extension. This is an alternative UI to the existing
ExtensionMessageBubble classes, in dialog instead of bubble form.

This CL introduces the dialog view and a generic controller class, as
well as unit tests and a browser UI test for the view. It does not
currently leverage the class in any production code.

Bug: 1079364
Change-Id: Ie5d374cc17ac12392934bd1f6e785c9e48a97630
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183711
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767080}