Block notification only on displays with fullscreen window
Previously with multiple displays in Chrome OS, if any one of them
was fullscreen, then notifications were not shown on any displays.
This changes continues to block notifications on any fullscreen windows,
but show the notifications on the displays that are not fullscreen.

MessageCenter and the associated FullscreenNotificationBlocker are
singletons, but MessagePopupCollection has a separate instance for
each display.

ash::FullscreenNotificationBlocker will now only block notifications
when all displays are fullscreen.

MessagePopupCollection for each display checks if it is currently
fullscreen and blocks any notifications which should not be shown.

Bug: 1077165
Change-Id: I6e918b9f3fdff6e367fc38b307307c6f78483a8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2179862
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767190}