[FlexNG] Use fragment order as source of truth within simplified layout
Step 3/4 in getting flex to use the simplified layout pass.

We need to use the fragment order as the source or truth as we use the
fragment order as the correct paint order.

The downside of this approach is that we now iterate twice over the
children, once for the fragments, and again for the nodes.

Adding placeholder fragments for the static-positions for the OOF
will fix this.

Bug: 845235
Change-Id: Idb149a25975885cee389c3bae1dd9021dbb5a467
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2187854
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767208}