LoginDatabase metrics test: Cover the account store too
This CL adds some first, minimal test coverage for metrics from the
account store: It verifies that the account store doesn't write into
any of the same histograms as the profile store.
It achieves this by creating a second LoginDatabase instance in the
ReportMetricsTest, and instructing it to record metrics. No expectations
need to be changed due to that, so we're good!

Bug: 1063852
Change-Id: Ife4bdf22b1567ee53f1021f4ec067a793eb98410
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2187875
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767223}