Remove ServiceWorkerVersion::Observer::OnNoWork()
This callback was implemented only by ServiceWorkerRegistration
to trigger activation. This indirection seemed a bit obscure.
We can look up live registrations and directly trigger
activation in ServiceWorkerVersion::OnNoWorkInBrowser().

A downside of this change is making ServiceWorkerRegistration
and ServiceWorkerVersion tightly coupled but it simpifies
the code base so I think this change is worth doing.

This change doesn't fix the linked bug itself.

Bug: 1015692
Change-Id: I10f31fdb32609ba488ea00c690617b96be572658
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2192579
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767243}