Deduplicate the size for bubbles with vector icons in their headers.
There are only a few Chrome dialogs that return a vector icon from
GetWindowIcon, but those that do all use the same size (20 dips). We're
going to change PermissionPromptBubbleView to also have one, so this
patch moves the constant into ChromeLayoutProvider so it can be reused.

Bug: 1080651
Change-Id: I7efefeb1fa953e9c9e7cb5e4015982a2369d112c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2190476
Reviewed-by: Bettina Dea <bdea@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Bret Sepulveda <bsep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767283}