Add ExtensionInstallEventLogCollector and ExtensionInstallEventLogger
We want to use event based reporting for forced extensions. We have
classes AppInstallEventLogCollector and AppInstallEventLogger which
are used for ARC++ push-install event reporting. Two new classes
are implemented as follows:

1) ExtensionInstallEventLogCollector
2) ExtensionInstallEventLogger

ExtensionInstallEventLogger behaves as a delegate for
ExtensionInstallEventLogCollector.

The whole pipeline for reporting is not ready yet. This is a prework
for the event based reporting for forced extensions.

Bug: 1068942
Change-Id: Ic7c3e85e280675a11a099d7b1562ed5bf03a1bc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2141920
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Askar Aitzhan <askaraitzhan@google.com>
Commit-Queue: Swapnil Gupta <swapnilgupta@google.com>
Cr-Commit-Position: refs/heads/master@{#767284}