tracing: Use ScopedDeferTaskPosting when returning TraceWriter
We can destroy a ThreadLocalEventSink within a trace event macro.
Destroying the sink causes its writer to be returned, which in turn
posts a task to the PerfettoTaskRunner. We have to ensure that we
proxy this task via ScopedDeferTaskPosting to avoid deadlocks with
PostTask-related trace events.

Change-Id: I79430e5101822408b4791f6df06f64cf3c9f7830
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2193614
Reviewed-by: Stephen Nusko <nuskos@chromium.org>
Commit-Queue: Eric Seckler <eseckler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767294}