[Android] Move startActivityIfNeeded() to ExternalNavigationHandler
This CL dedupes //chrome's and //weblayer's implementations of
ExternalNavigationDelegate#startActivityIfNeeded(), moving the code to a
shared location in ExternalNavigationHandler.java. One complexity here
is that //chrome's CCT-specific subclass of its
ExternalNavigationDelegate implementation actually overrides
startActivityIfNeeded() with custom logic. To support this, we change
the delegate method to maybeHandleStartActivityIfNeeded() and adjust its
contract so that implementations return whether they handled the event
at all as well as indicating what the result of any such handling was.
When the delegate doesn't handle the event, ExternalNavigationHandler
handles it itself via the startActivityIfNeeded() logic that is now
being shared across embedders.

Bug: 1071390
Change-Id: If8da5548aef60dd7280d41d0d03d5fb0a15fcf8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189654
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767344}