[omnibox] Fix bug where users can tab to an non-existent Remove button
Since we migrated the fake focus logic to live exclusively in the
OmniboxPopupModel, there has been this edge case, where a match that
both has a keyword and a Remove button, would be able to tab to both,
even though only the keyword control is visible.

This CL fixes that, and is designed to be small enough to be merged.

Bug: 1081269
Change-Id: Ic942717836a26e065c975647e2de0eb3f6c01c0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2193933
Commit-Queue: Tommy Li <tommycli@chromium.org>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Auto-Submit: Tommy Li <tommycli@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767373}