Touch: Refactor touch device manager & reduce bleeding of its complexity
This patch refactors and limits the usage of TouchDeviceIdentifier to
within the class that uses it. TouchDeviceIdentifier was formulated to
be used as a unique identifier for touch devices that is consistent
across hotplugs and session restarts, while also being easy to use with
stl containers. This idea, however, restricts the validity of touch
devices and their unique identification to only work when the device
firmware provides enough information for the stable identification for
each of its input interfaces. This information is not always provided,
and the assumption of its presence in such scenarios can render the
touch device and display association unreliable, inconsistent, or in a
not working state. There is no way to identify such interfaces
persistently & uniquely.

For this reason, the unique touch device identifier usage is removed
from various places in the code and only limited to the class that needs
it to store touch devices & display association information
persistently. For devices that do not provide interfaces with enough
stable information, the associations are not persistently stored. For
such cases, the association for a given session would still be
unreliable when multiple displays are present, and unfortunately, I do
not see a way around it. But with this change, at least the device won't
be rendered broken or not working, only not working as intended in some
cases.

Bug: 1029824
Change-Id: I474a3a5fc372cd1582ab952d644e02002419dc6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189594
Reviewed-by: Michael Spang <spang@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Malay Keshav <malaykeshav@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767411}