Improve apps container animation in search results page
The animator passed to AppListPage::AnimateYPosition applies
transform that follows the search box bounds. This offset may cause
small jump in apps container page shown underneath search results UI
when app list drag ends.

The apps container is positioned relative to where the search box would
be in the apps page at the current app list height. That is generally
not the same as the search box position in search results page. which
means that the apps container view has different distance to pass during
animation to fullscreen / closed state than the search box.

This updates the AppListPage::AnimateYPosition to let the app list page
implementations override the default offset, and makes use of that
ability in the AppsContainer.

Bug: None
Change-Id: Iee69ae2ea94073da8eccab4942ecf947cb396d58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2191227
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767420}