[Payments][Desktop] Fix potential use-after-free of render_frame_host.
Cherry-pick to refs/branch-heads/4103.

PaymentRequest gets a RenderFrameHost* in its constructor that is later
used when PaymentRequest::Init() is triggered asynchronously by mojo.
This can lead to a use-after-free.

This CL converts the RenderFrameHost* to its GlobalFrameRoutingID
in the PaymentRequest constructor, when it is guaranteed to be live.
If the RenderFrameHost is freed when PaymentRequest::Init() is called,
the RenderFrameHost::FromID() will return nullptr, and the rest of the
code can handle it appropriately.

The real fix is to change all RenderFrameHost* to WeakPtr. This is tracked
in crbug.com/1058840.

TBR=rouslan@chromium.org

(cherry picked from commit 1584079d544f53ed6b45d80ca025afcc4b67a25a)

Bug: 1072772
Change-Id: I2c7b54aeea67aa39fd11248d35faecbf8c771594
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2186396
Commit-Queue: Danyao Wang <danyao@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#766641}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2194271
Reviewed-by: Danyao Wang <danyao@chromium.org>
Cr-Commit-Position: refs/branch-heads/4103@{#511}
Cr-Branched-From: 8ad47e8d21f6866e4a37f47d83a860d41debf514-refs/heads/master@{#756066}