Replace !=/== with !==/=== in chrome/test/data/webui/bookmarks
Mostly a portion of
https://chromium-review.googlesource.com/c/chromium/src/+/2003795
by dbeam@ generated using the same command line, but only updating the
bookmarks/ folder.

Additional non-automated changes:
- Fixing a comparison between strings and numbers in DND manager test
(compared a number enum value to the result of Object.keys)
- Fixing some flakiness in list_test.js; should not have been the
result of the changes above, but noticed it when testing locally;
flushTasks() didn't always wait long enough for the multiple async
steps that happen between the contextmenu event and opening the menu.
Instead, waiting on the recordInHistogram call that happens as soon
as the menu is actually opened.

Bug: 720034
Change-Id: Id3a66f013b71ccb330b6dde34dcd83d93d339218
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2190037
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767439}