quota: enum QuotaClient::ID -> enum class QuotaClientType.
This CL converts the QuotaClient::ID enum into an enum class, in
preparation for converting the entire QuotaClient abstract base class to
a mojo interface. The type is also lifted to the top level, to prepare
for the mojo move.

The QuotaClient::id() method is renamed to QuotaClient::type(), and the
enum class is named QuotaClientType. This prepares us for a future where
multiple QuotaClient implementations of the same type (in different
browser processes) may connect to the same QuotaManager (in the Storage
Service).

Chrome currently uses ints to represent sets of QuotaClient::ID values.
This is accomplished by having each ID value be a mask with exactly one
bit set. Set operations map to bit operations, and the code looks decent
because enum values are implicitly converted to int.

Enum class values don't implicitly convert to int (which is a feature),
so this CL switches the set representation from bit field ints to
base::flat_map<QuotaClientType>. The new set representation is more
expensive to copy than ints. This isn't expected to have a significant
impact on performance, because QuotaClientType sets are only used for
Clear Site Data-like functionality, which is invoked very rarely.

The new set type base::flat_map<QuotaClientType> is aliased as
QuotaClientTypes, to facilitate a potential future conversion to
std::bitset, or a similar type. We may pursue this conversion if we want
the performance benefits -- std::bitset is exactly as cheap to copy as
an int. This CL does not use std::bitset, because it has a more
cumbersome API.

Bug: 1016065
Change-Id: I46fd15bd4319bc34671d9f788b5d0f4fdf59c37a
Tbr: avi, creis, shimazu
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2136478
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Jarryd Goodman <jarrydg@chromium.org>
Commit-Queue: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767474}