Fix crash on entering VR from StartSurface
When entering VR with the StartSurface experiment enabled, one of two
crashes occur. The first is that the "new tab" button, which VrShell
uses to create a tab if no tab is present, is no longer there. The
second is that the StartSurface isn't really a tab, so the fullscreen
manager fails to get the appropriate controls position.

This fixes the first issue by using the TabCreator to open a new tab
rather than "clicking" the UI, and the second issue by treating not
having a tab as if the offset is not initialized.

Bug: 1079261
Change-Id: Ia529ab88e39f6c9e208eaa1d6421d08d46135c54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2190593
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767481}