Trust Tokens: Make key commitment batch sizes mandatory
It transpires that the key commitment batch size is an integral part of
the Trust Tokens protocol, not an optional value issuers can provide to
indicate a preference about the number of tokens they wish to issue.
(For instance, it is conceivably possible that BoringSSL could use a
completely different cryptographic algorithm for generating signed
tokens, based on the value of the batch size it receives).

As a consequence, it's necessary that clients and servers share the same
batch size value; in particular, this means we need to have a batch size
value provided in key commitment results. This change makes the batch
size field mandatory in TrustTokenKeyCommitmentResult.

A previous change already made the issuance and redemption helpers
reject if the field was absent.

R=csharrison

Fixed: 1073925
Change-Id: I167725fd2df7f921ff46989c83fb2e1838c30d5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2185402
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Auto-Submit: David Van Cleve <davidvc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767507}