Assert on use_ozone & ozone_platform only in target toolchains.
We set use_ozone to be unequivocally true for every toolchain in
a CrOS build today:
https://source.chromium.org/chromiumos/chromiumos/codesearch/+/master:src/third_party/chromiumos-overlay/chromeos-base/chromeos-chrome/chromeos-chrome-9999.ebuild;l=245

I'd like to remove that so it relies on the default of "is_chromeos",
but doing so as-is fails today due to this assert.

That's because we also set ozone_platform_* args true for every
toolchain like we do with use_ozone today. To use use_ozone's
default, we need this assert to only check the target_toolchain (and
ignore whatever we set for others like the host toolchain).

Bug: 937821
Change-Id: I9c92bf9d0f8a8fff1611f95b6ca4019bc19808c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183249
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Michael Spang <spang@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767574}