Add 1hr expiration for conditional tab strip session
Previously, we define the session of conditional tab strip as identical
to app lifecycle. This CL adds a 1hr expiration so that regardless of
whether the feature is enabled or disabled in previous session, the
feature state is reset after Chrome has been put in background or shut
down for at least one hour.

Bug: 1076295
Change-Id: Ic8afb01c6245792f4b1bdfbce1a8b9202a58fea8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2186453
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767590}