Shelf should not be affected by tray bubbles on a different display
Tray bubbles should force the shelf to be visible. Currently we tell
whether there is any tray bubble by value of
g_current_tray_bubble_showing_count_.
However, that variable is the sum of tray bubbles across all
displays. As a result, the shelf may be affected by tray bubbles on a
different display.

In this CL, the count of tray bubbles is stored in status area widget.
When a tray bubble is shown or destructed, only the status area widget
sharing the display with that tray bubble updates the tray bubble count.

Bug: 1079464
Change-Id: I1a648d707d61dc30bc7038edf8de927915b30f70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2191149
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767633}