Prepare for enabling the UKM demographics browser test on Android.
This preparation includes the following:
- placing if !defined(OS_ANDROID) around desktop-specific code
- placing if defs around platform-specific #includes
- updating BUILD files
- adding code to the demographics test to support getting Android's
  way of getting a browser, which is actually a tab

The code that gets a tab for Android and a browser for desktop will be
improved in a follow-up CL by creating an abstraction that encapsulates
tabs and browsers: crbug/1080661.

Additionally, this CL updates some test comments related to keeping
certain tests in sync with their iOS and Android counterparts.

Bug: 1047474
Change-Id: Id21c1c228e9c3d4d8d1cd1126d81745e9159335c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2190163
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767647}