Make PdfActionableHighlight have presentational children
During continuous reading, screen readers don't announce PDF highlights
when entering a PDF highlight. The reason for this is because when an
AT client is on the child text node of PDF highlight, the enclosing
element returned by |ITextRangeProvider| implementation does not return
PDF highlight as the enclosing element.

We had already faced similar issue with controls which should have
presentational children, for example the elements listed here:
http://www.w3.org/TR/core-aam-1.1/#exclude_elements2.
|kPdfActionableHighlight| role is not part of the list, because it is
a new role being introduced specifically for PDF. The issue has been
explained in details in the bug: https://crbug.com/1075988.

To fix it, this change makes |kPdfActionableHighlight| reports all its
children as presentation elements. In code, this is implemented by
adding |kPdfActionableHighlight| in
|AXPlatformNodeWin::ShouldHideChildrenForUIA()|

Bug:1081022

Change-Id: Idd49ebe2c9eeb9f452c1ee4e74f97a691f7c2993
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2191713
Commit-Queue: Virender Singh <virens@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#767703}