Remove obsolete WritableStream::IsLocked
blink::WritableStream::IsLocked() was an old API that supported throwing
an exception. Since with the C++ implementation it is no longer possible
for this operation to throw an exception, it is obsolete. Remove it.

Use WritableStream::locked() instead in the callers.

No functional changes.

Change-Id: I4272ca0040182477c50122ccbebaaf37ca32ea9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2195589
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767717}