[Security][Coop] Use COOP only if this is top level
COOP is used only in top level document, and COOP headers from iframes
are ignored. This led to an issue in the linked bug, where COOP prevents
a sandboxed iframe to load.

The spec change corresponding to this is under review here:
https://whatpr.org/html/5334/browsing-the-web.html

with this relevant extract:
```
Let navigationCOOP be "unsafe-none".

If browsingContext is a top-level browsing context, then:

Set navigationCOOP to the result of obtaining a cross-origin opener
policy given response and reservedEnvironment.

If sandboxFlags is not empty and navigationCOOP is not "unsafe-none",
then display the inline content with an appropriate error shown to the
user, with the newly created Document object's origin set to a new
opaque origin, run the environment discarding steps for
reservedEnvironment, and return.
```

Bug: 1081169
Change-Id: I2c0b59c84ca52f63436a2312529a4bb0351fff30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2193771
Commit-Queue: Pâris Meuleman <pmeuleman@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Auto-Submit: Pâris Meuleman <pmeuleman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767761}