Use gpuBenchmarking in common-wheel-event.js
The file common-wheel-event.js related to this CL is located in
virtual/form-controls-refresh-disabled/fast/forms/resources

After we finish the scroll unification, the scrolls happens mainly on
the compositor thread, and the scroll code in the main thread will be
removed. eventSender sends the scroll events to main thread, so it
would not work after the scroll unification. We should replace
eventSender.continuousMouseScrollBy with gpuBenchmarking.smoothScrollBy
in scroll tests in fast/scroll-behavior/smooth-scroll/.

Bug: 1047176
Change-Id: I5a17b46875da78234494a21d3be7a5d26e22bc65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2192861
Commit-Queue: Liviu Tinta <liviutinta@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767791}