[throughput] Fix a DCHECK in FrameSequenceMetrics
This CL fixes a mistake in a DCHECK. There was a mistake in the
DCHECK, because the DCHECK is doing DCHECK_LE(a, a) which is
always true.

Together with it, we found a bug where the aggregated_throughput_
may not be cleared after ReportMetrics. A unit test is added.

Bug: None
Change-Id: I89cbf012d2b1235a6179a80a5e98dc51f97a2aac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2185528
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767792}