Query Tiles : Added query editing experimental variation
This CL adds a variation for the query tile experiment. When a query
tile is tapped on NTP or omnibox, the omnibox will be focused showing
the query text. The next level tiles will not be displayed. User will
have a chance to edit or cancel the query before submitting. Added
the feature flag as well.

Bug: 1080766
Change-Id: Ice3b5652102a8b0aeb89650520a46726aa7c22e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2190221
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767793}