[iOS] Separate use of Scene API from multi-window.
This CL adds a new gn flag (ios_enable_scene_startup) which enables the
iOS13 Scene API for the application lifecycle without enabling multiple
windows on iPadOS. This will allow the new lifecycle code to be enabled
and tested without requiring multi-window support or features.

The IsMultiwindowSupported() is joined by IsSceneStartupSupported(). The
latter returns true whenever the former does, but also when scene startup
is enabled without multi-window.

Call sites for IsMultiWindowSupported() are changed to the new helper
in cases that don't involve the creation of multiple windows.

It is expected that enabling the ios_enable_scene_startup flag will
cause test failures and many other problems at this point; this CL
allows those issues to be resolved incrementally before the flag is
turned on by default.

Change-Id: Ie2ca8247e519e78b90063f15fffd3fdd2a6de3ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2193995
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767794}