[Android] Generalize naming in componentized external intents code
The external intents code was componentized for reuse in WebLayer, but
Chrome-specific naming stuck around several places. Now that the dust
has settled on the componentization, this CL generalizes all relevant
uses of "Chrome" in the //components-level code. One or two uses that
really are specific to the Chrome app (e.g., a mention of Chrome in the
context of CCT) have been left as-is.

Bug: 1071390
Change-Id: I09dcf2621a0655638563ed68f5c95fc21252b069
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2196344
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767842}