Fix the shelf icon load issue when the app is not pinned to the shelf.
If the unpinned app has been launched for the second user, when switch
to the second user, AppServiceAppIconLoader is called to load the icon
at very early stage for the launched apps due to the session restore
process. But ChromeLauncherController::ActiveUserChanged is not called
yet, so AppServiceAppIconLoader for the first user is still used to
load icons. Then ChromeLauncherController::ActiveUserChanged is called,
and AppServiceAppIconLoader for the first user is destroyed, so the load
icon callback can't be called to load icons.

Modify ChromeLauncherController's app_icon_loaders_ to use a per profile
based map. When loading icons, select the correct profile's
app_icon_loaders_ to load icons.

Modify ChromeLauncherController's OnAppInstalled and OnAppUpdated to
reload icons for both the pinned app and launched unpinned app.

Modify LauncherControllerHelper::GetAppTitle to use extensions to set
the title for extensions only.

BUG=1076540

Change-Id: I432047a4c965422f4721acb1fa6d6885f96faa01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189999
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767849}