Update RenderFrameHostImpl::LifecycleState to Active only after commit
Currently, when a page restores from back-forward cache we update the
LifecycleState to Active in RenderFrameHostImpl::LeaveBackForwardCache
even before the commit happens.

This CL updates the LifecycleState from kBackForwardCache to kActive
after the commit happens to align with the correct definition where
LifecycleState::kActive which corresponds to when a RenderFrameHost
is the current one in its RenderFrameHostManager and FrameTreeNode.

We also rename LeaveBackForwardCache with WillLeaveBackForwardCache
and EnterBackForwardCache with DidEnterBackForwardCache.

BUG=1076931

Change-Id: I6ad104dcc56bfa8e7a4201882584b1d6fc555b9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2184057
Commit-Queue: Sreeja Kamishetty <sreejakshetty@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767855}