Replace RFHI::is_active in tests
RenderFrameHostImpl::is_active is used in tests to check for the RFH
state when the navigation commits or when the RFH is in pending
deletion.

We would like to remove all the |is_active| occurrences to avoid
confusion with all the other methods.

We have more specific ones:

- RenderFrameHost::IsCurrent
- RenderFrameHostImpl::IsPendingDeletion
- RenderFrameHostImpl::LifecycleState

This patch updates the callers of |is_active| and make use of the two
methods above instead.

BUG=1073449

Change-Id: I59095b4e70a2b10bffbf95fcdd9f9f4d9097e568
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2182360
Commit-Queue: Sreeja Kamishetty <sreejakshetty@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767878}