Fix in-product help for WebUITabStrip
IPH was previously triggered on clicking toolbar new tab button. It was
since removed, but the IPH code was not updated. This CL triggers IPH on
any new tab opened and adds a regression test.

It additionally removes the defunct code for the toolbar new tab button.

Bug: 1045669
Change-Id: Iec42bba9de2ce39dc62124cc375ed15e35302dda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2188529
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Reviewed-by: Connie Wan <connily@google.com>
Reviewed-by: Connie Wan <connily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767879}