Switch avrt.dll path to use %SystemRoot%
The path expansion for avrt.dll currently uses %WINDIR% to expand the
Windows installation directory from an environment variable. WINRDIR is
just a normal user env variable, it can be modified and might not be set
in sandboxed processes. It looks like it has been superseded by
SystemRoot which is set on installation for Windows 7+ and is read-only.

Bug: 1074445
Change-Id: I425c1875a9d8f9ae7a3585daf0225dae75461cd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2196644
Reviewed-by: Henrik Andreasson <henrika@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767884}