[WebLayer] Add SiteSettingsClient.isCategoryVisible()
This method allows embedders to control which categories are visible in
the main Site Settings page, and hide ones that aren't supported by
their product. This only applies to the main page; launching the
SingleCategorySettings Fragment directly will still work, and
SingleWebsiteSettings will still show any category that has a value set
for a permission/category, regardless of isCategoryVisible()'s return
value. This is intended, as embedders that don't support a web feature
shouldn't be granting permission to it.

Bug: 1049683
Change-Id: Ifbd53ed875dbf6d51f3f1e59f7773654ad2eacaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2194588
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767946}