Remove WebContentsView::SizeContents()
The method is considered a hack for changing the size of the
RenderWidgetHostView independently of the actual window. We note
that RenderWidgetHostView::SetSize() now exists though too, and it
allows the same thing.

Replace for tests the use of SizeContents() with the Shell code
that does this exact thing for web tests, and supports headless
shell in web tests.

Remove use of it for GuestView which led to a code deadend and did
nothing.

R=avi@chromium.org

Bug: 866140
Change-Id: I7302190e01a3b1150aa1354ea1a8adb5d2f48d5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2188931
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767949}