Separate logic deciding how ready a page is for thumbnail capture
The thumbnail capture code considers several aspects of page state when
deciding whether to capture:
* Page navigation and loading state
* Page visibility (whether the page is in a foreground or
  background tab)
* Thumbnail visibility (whether the thumbnail is actively requested)
* RenderWidgetHostView availability (whether we are even able
  to capture)

The first corresponds to how ready a page is for capture. This is
independent from how the page is presented and whether we are able to
capture.

This CL separates out the logic tracking the page readiness into a
ThumbnailReadinessTracker.

This should be a pure refactor. No behavior change is intended.

Bug: 1073141
Change-Id: Iee3ac51890ca45ec6c70bbc143987d2e1a5cc610
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189591
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767972}