[Sampling profiler] Install Android unwinder DFM and gate profiling on availability
Requests installation of the stack unwinder dynamic feature module
on canary and dev if not already installed, so that it should be
available on the next run. Disables profiling on Android if not
installed.

This is supplemental to the the platform and channel support check, for
which Android profiling remains disabled until we have the end-to-end
profiling implementation in place.

Bug: 1004855
Change-Id: Ie2db9203694aa64ecffab2e852cd6e422c40449d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2190988
Commit-Queue: Mike Wittman <wittman@chromium.org>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768005}