PDF viewer: Exclude resources in Print Preview
Add a separate index.html for the PDF viewer in Print Preview vs the
normal viewer. This allows resources not used in Print Preview to be
excluded from the data source, and not imported at runtime, saving
~18ms (5%) of the time to load the viewer in an optimized build.

Also add a few tests for the viewer in Print Preview.

Change-Id: I6599bb61d061cf42b71ef26b48736bf7d69cdfaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183405
Reviewed-by: dpapad <dpapad@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768013}