Refresh open files from handles on navigation.
Files can be changed or renamed or moved while chrome://media-app is
running, and this currently breaks its ability to navigate. The DOM
File machinery checks modification timestamps and requires a File to be
reopened if it changes.

Rather than keeping an opened File in the file descriptors held in the
privileged context, we set them all to null whenever they are sent to
the guest context. Navigation reopens (but doesn't read) the files to
ensure recent handles are passed.

Bug: b/154550689
Change-Id: I0fa19efd03752069f168015a0210c412cd553640
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2186873
Reviewed-by: Bugs Nash <bugsnash@chromium.org>
Commit-Queue: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768021}