[CrOS Settings] Refactor OsSettingsManager
This CL creates two helper classes:
(1) OsSettingsSections, which contains all of the sections in settings.
(2) SearchTagRegistry, which is responsible for adding/removing tags
    from the LocalSearchService and for providing metadata about
    registered SearchConcepts.

It also adds tests for SearchTagRegistry, and adapts
OsSettingsManagerTest to verify the initialization flow instead of
search tag registration.

A follow-up CL will move SearchHandler into OsSettingsManager as well.

This is in preparation for building a settings hierarchy.

Bug: 1080892
Change-Id: Ib4762dc5608b107a08e783f5e6cbee72fa3e76f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2191451
Reviewed-by: Regan Hsu <hsuregan@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768041}