Add E2E browsertests for Conversion Measurement
This CL adds a new browser-test suite which tests the full functionality
of the Conversion Measurement API.

To make tests deterministic, a new command line switch is added
which turns off all report delays and noise. This is implemented as a
switch so that it can be re-used for web-platform-tests, where mocking
out clocks or other components is not feasible.

Change-Id: I380fe413b46464895e872451f835080a659979ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151700
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768045}