[SB Delayed Warnings] Handle downloads on delayed warnings
This CL cancels downloads when a SafeBrowsing warning is delayed on
a page. It introduces a new navigation throttle to be able to observe
and cancel downloads. The existing user interaction observer can
observe navigations but can't cancel them. Instead, it records a new
histogram entry when it detects a download.

This CL also changes UserInteractionObserver to handle
DidFinishNavigation() events instead of DidStartNavigation(). When a
navigation starts, the order of events is as follows:
1. NavigationThrottle::WillStartRequest()
2. WebContentsObserver::DidStartNavigation()
3. NavigationThrottle::WillProcessResponse()
4. WebContentsObserver::DidFinishNavigation()

The observer can't detect that a navigation is a download until
DidFinishNavigation(). Additionally, the throttle checks if a
UserInteractionObserver is attached to the page to determine if a SB
warning is delayed on the current page. It does this check in
WillProcessResponse(), so the observer must be present at that time.

The behavior introduced in this CL (not showing the warning on a
security moment) is different than existing handling of events: If the
observer detects user input or page attempting to enter fullscreen mode,
it immediately displays an interstitial. However, this CL silently
blocks downloads without the interstitial. This is done to simplify
the implementation. A full discussion of this can be found at
go/sb-delayed-warnings-downloads.

Bug: 1057157
Change-Id: I7da8bfe969fccfe7f12cd336098c8230fc799739
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2137819
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768086}