[remoting host] Tweak encoder quantizer settings.
This sets the target quantizer for topup to be equal to the min
quantizer. Frames are topped-up to the target quantizer. Prior to this
CL, the topup continued slowly every 2 seconds (when desktop image is
static and "empty" frames are sent), until the min quantizer is reached.
So the top-up to min-Q was occurring, but spread out over a long time.
Setting target-Q = min-Q speeds up this process and should greatly
improve the user experience.

This CL also removes the per-codec setting of min/max Q, as these had
no effect and were misleading. The quantizer is completely controlled
by the VideoFrameScheduler, so it is actually the same for each codec:
max Q = 50, min Q = 10. Some work needs to be done if we want these
settings to be per-codec.

Change-Id: I3eeee43dd8ce56679718796ba5608c7e5fe2de06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2191228
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Jamie Walch <jamiewalch@chromium.org>
Auto-Submit: Lambros Lambrou <lambroslambrou@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768087}