Add LoginDatabase.ReportAccountStoreMetricsTest
This test is very similar to the existing
LoginDatabase.ReportMetricsTest, it just swaps the roles of the profile
and the account store. It covers mostly the same metrics (minus a few
that don't make sense in the account store), just with ".AccountStore"
in their names.

Bug: 1063852
Change-Id: Ie5378820921b768c39b4e9b6c104d4bcee2e3b7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2193660
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768176}