[mpc] Introduce CSSProperty::ComputedValuesEqual
In order to handle dependencies in the MatchedPropertiesCache, we need
to be able to know if two computed values are the same. This CL adds
a new function CSSProperty::ComputedValuesEqual for this purpose.

Unfortunately we're not able to support this function on all
properties right away, as each of the ~500 properties needs to be
thought about and tested individually. Hence, there's a
computed_value_comparable flag which indicates which properties are
supported.

The reason we have to think about each property, is that some
properties have a more complicated storage situation on ComputedStyle,
since the computed value is split over multiple fields. Therefore the
simple "compare two getters" approach in this CL will not work properly
for all properties. The code generated for CSSProperty::
ComputedValuesEqual will become more advanced in later CLs,
as we add support for properties that need more advanced behavior.

But let's start simple by just comparing the two main getters. This
will be enough for a large portion of the properties.

Bug: 1057072
Change-Id: I12f0a7da6d476dafd0da05b8137f5a5a28638b43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2193011
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768215}