PPAPI:  Add NetworkIsolationKeys to proxy lookup requests.
This is needed to correctly isolate any DNS requests made by PPAPI in
the context of one page from those made in the context of other pages.

This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/2008050,
which was reverted due to the DNS cache expiring entries.  The fix
is to allow stale DNS cache results.

That was a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/2050743. The
original CL was reverted because the DNS layer auto-detects whether the
current network supports IPv6, and the detection value was changing in
the middle of tests. When the value changes, the cache key used for the
DNS layer changes, which was resulting in cache misses. To fix the
issue, the test now relies on IPv4 DNS lookups, which don't have this
issue.

Bug: 1021661
Change-Id: Ia0a09ce772e6c54a73dc308e47848391872337ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050743
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#740487}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2194412
Cr-Commit-Position: refs/heads/master@{#768232}