Move Brand Code writing to a more core part of MetricLog.
This allows it to be naturally reused by UKM which uses this part of MetricLog.

I think this should be reasonable to do - main concern is for early logs could this be not yet available on some platforms. I suspect it should be ok (Windows from registry, chromeOS from prefs, for example). I am planning on querying to verify the #logs with brand codes is similar before and after, but let me know if you feel that we should do more per-platform testing.



Bug: 1002210
Change-Id: Ibed1fa874388a7ef772318ae49c42b08e69a76ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2196502
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768249}