Add android/bottomsheet directory in components/browser_ui
This patch lays the groundwork for the new home for the bottom sheet.
'components/browser_ui' now has an 'android' directory for android-
specific widgets, which is where the 'bottomsheet' directory now is.
BottomSheetContent has been moved to that directory to start with since
it has no other dependencies. Build targets for dependent modules and
features have been updated to support this.

Bug: 100227
Change-Id: I998273d3e56df6015e86df3ce993e255b642a8c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2196854
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768254}