[WebAudio] Call SetFormat() for every sink Initialize() call.
If a media element is attached to WebAudio after it has begun
playing, a later suspend/resume cycle may result in the sink
being initialized with different audio parameters.

In the case of the linked bug, the cause of the difference is
because the first attachment is w/o WebAudio so the normal
playback pipeline initiates resampling and channel remixing
to the  hardware parameters. After the element resumes from
suspend the playback pipeline notices that WebAudio is
attached and skips resampling and channel mixing.

These differences lead to issues in the buffer sizes and
channel counts that are requested during Render() operations;
which can lead to garbled audio or CHECK() failures.

R=hongchan

Fixed: 1079232
Test: New unittest, original issue no longer reproduces.
Change-Id: If4d6575786ddbc02f0405f3c6fd7aa237983e0fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2197416
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768350}