Add EnabledExtensionNames when collecting Vulkan info
Add "VK_KHR_surface" and "VK_KHR_win32_surface" to
VkInstanceCreateInfo.ppEnabledExtensionNames in vkCreateInstance() when
collecting Vulkan info. These two extension are also enabled when Vulkan
is on.

EnabledExtensionNames are not required when creating a instance. But
since there are a lot of crashes in the Vulkan drivers when calling
vkCreateInstance(), adding the EnabledExtensionNames will change the
the Vulkan driver code path to a more popular one and it might help
reduce the number of crashes.

Bug: 863086
Change-Id: I132412d267b4a0f7916429c723cbe8937edb087a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2197260
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768362}