Run dawn_end2end_tests on just discrete and integrated GPUs
In https://dawn-review.googlesource.com/c/dawn/+/21202, Dawn tests
are expanding to run on all available GPU adapters with command-line
arguments to request which adapters to run on. This CL adds the
arguments to preferably run on the discrete and then the integrated
GPU. This matches the current devices the tests run on.

Separate tests will be added in the future to run on CPU adapters
which would run the tests using Swiftshader.

Bug: dawn:396
Change-Id: I970f03f21a33d0a23a998de19918fd558512e0da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2194845
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768373}