Improve comments and readability of input code
This CL adds explanatory comments to various parts of the input pipeline
and leaves TODOs for some future improvements that require more
investigation or could be breaking.

It also renames some variables and methods to be clearer and adds
DCHECKs to clarify invariants.

This CL has no intended behavioral differences.

The comments in InputHandlerProxy come from investigating some of the
history of these bits:

  |is_scroll_end_from_wheel| comes from  https://crrev.com/c/557463/
  with more explanation in https://crbug.com/704732

  |is_from_set_non_blocking_touch| comes from https://crbug.com/704732
  and https://codereview.chromium.org/2869823003

  |blocking_wheel_update| comes from https://crrev.com/ff58622d23a1de5

Bug: NONE
Change-Id: I05d1f9a19866fdb16eaf1c9bcb34ff6fe9a2bea4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2192673
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768437}