[XProto] Hide list size from user-visible structures
libxcb has a separate length field for lists because it uses C-style
arrays to implement lists.  We use std::vector which has a length field
built-in, so we can derive the length from that to send over the wire.

However, the length field can actually be a general expression in
xcbproto (eg. the sum of two other fields).  In cases like these where
it's tricky or impossible to derive fields from the length (eg. x + y
= length has many solutions for x and y), the fields are left in
the user-visible structures.

BUG=1066670
R=sky

Change-Id: I0db56e0b6582eddf903e7b8d8943533b2d7901f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189702
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768479}