Refactor mouse move hit testing in compositor
The compositor does some basic hit testing on mouse moves in order to
animate overlay scrollbars when the cursor is near the thumb. The way it
does this is by hit testing a LayerImpl and then using
FindScrollNodeForDeviceViewportPoint to find the associated scroller.

However, FindScrollNodeForDeviceViewportPoint isn't well suited to this
purpose. It's primarily used to target scrollers during a ScrollBegin
and it accounts for whether or not a ScrollNode can be scrolled on the
compositor. Scrollbar animations shouldn't be accounting for this. It
also has some complex fallback behavior that is also likely wrong for
scrollbars.

Instead, we replace its use with a much simpler and self-contained
HitTestScrollNode. This method simply returns a hit layer's ScrollNode,
encapsulating the hit scrollbar case and correctly accounting for failed
hit tests due to layer squashing. This is a small preparation for scroll
unification as this method will be used for scroll targeting in that
case.

This change revealed a bug in the scrollbar animating code. Because
FindScrollNodeForDeviceViewportPoint falls back to returning the
viewport node if no layer is hit, hit tests outside the viewport would
return the viewport scroll node. This is questionable for scroll
targeting but definitely wrong for scrollbar animations. When the change
was made, the ScrollBeforeMouseMove test started failing because it
sets up a scrollbar outside the viewport and sends mouse moves to
negative viewport coordinates.

FindScrollNodeForDeviceViewportPoint is renamed to
FindScrollNodeForCompositedScrolling to better reflect its behavior.

Bug: 1047182
Change-Id: Idb20285872c0da0ef33a404cffca4fbe8347ff72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2198598
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768482}