Cursor and tooltip not updating correctly when window bounds change
When window bounds changes (e.g. toggle between fullscreen mode and
windowed mode), the content underneath the cursor changes.

We found that the cursor didn't change correctly because this
situation failed to trigger SynthesizeMouseMoveEvent(). We changed
the condition of triggering PostSynthesizeMouseMove() and reset
|synthesize_mouse_move_| when |move_hold_count| going to 0.

We also found that the tooltip was wrong when toggling back to windowed
mode. We updated tooltip not only when it's visible but also scheduled
to show.


Bug: 1041396
Change-Id: I1a51c0b658e97b2b67079f7be80afb4dc393efb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2186362
Commit-Queue: Yichen Zhou <yichenz@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768494}