Enable ProfileSyncServiceUsesThreadPool by default.
After we previously run as an inconclusive beta trial by itself we
grouped it with the HistoryService ThreadPool trial to try and coax a larger
impact from the ThreadPool migration. This meant groups ran with either both
services on the ThreadPool, none or only HistoryService (status quo).

28 day aggregation trial in beta showed no likely negative impact. Crashes
seemed slightly elevated in the group when running both the HistoryService
and the ProfileSyncService on the ThreadPool but closer examination suggests
this is simply noise since all top-crashers are the same.

Bug: 1014464
Change-Id: I3a1604413fcf1f4b5350c3b9161251f46a449865
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2198932
Commit-Queue: Oliver Li <olivierli@google.com>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768548}