Introduce a function to format a DER-encoded SPKI for UI
This introduces x509_certificate_model::ProcessRawSubjectPublicKeyInfo
which takes a DER-encoded X.509 SubjectPublicKeyInfo as input, and
returns a string intended for displaying the public key represented by
it.

It will be used in the UI code for the Chrome OS built-in certificate
provisioning feature. When a certificate provisioning process is in
progress, only a key pair is on the device and not a certificate (yet).
This function will be used to display the public key of the key pair on
the UI.

Bug: 1045895, 1081396
Change-Id: I8428dec1a03ac1fdc3edffdaeeda304be0cdb9be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2036053
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768551}