Fix app list suggestion result removal dialog positioning
The dialog is shown when a zero state suggestion app list search
result is removed, and asks the user to confirm the suggestion removal.

It's positioned relative to the SearchBoxView, but was not properly
getting updated on the app list/search box bounds changes.
It was using ContentsView::SearchBoxUpdateObserver to detect search box
bounds updates, but that did not work reliably (and did not handle app
list contents animations at all, nor bounds updates due to app list view
state changes).

This CL removes this observer interface and lets search results page
view handle the dialog positioning, and hides the dialog when the app
list page changes away from search results page.

Introduces SearchResultPageAnchoredDialog to encapsulate
*   the dialog widget positioning relative to the search box bounds (in
    app list contents view coordinates)
*   coordinate conversions between the app list contents view (in app
    list window coordinates) and the dialog widget (screen coordinates)
*   detecting the dialog widget closure
*   detecting app list window bounds changes

Note that the helper observes the app list window bounds changes to
detect when the app list view bounds changes that do not require
contents view layout (and do not cause app list contents view bounds
change) - the removal dialog has to be re-positioned in this case as
well (its bounds are in the screen coordinates). An example of this
update is when the virtual keyboard gets hidden/shown and app list view
is shown.

BUG=942653

Change-Id: I1b12f6b518d952a6a0f492ab73c4d6afb24f1bea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2198357
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768558}