viz: Re-enable ignore_undamaged on platforms using GL buffer queue
SurfaceAggregator has an option to output only the quads that
intersect the damage rect computed by SA itself.

This optimization was disabled on devices with HW overlays or where
the renderer might expand the damage, since SA would not know the
final damage rect.

This CL re-enables the optimization on devices where we can estimate
a bounding rect of the renderer expanded damage.

This is achieved by asking to the renderer for a bounding rect that
represents the maximum area that can be expanded.
SurfaceAggregator will aggregate the boundingrect union the damage.

The damage on the root render pass will be unchanged, so that HW
overlays damage computation won't be affected.


Bug: 1077210
Test: viz_unittest (new test and Display tests)
Change-Id: Ie94d1f09923959e2e025845f89a9b3eae759a64a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2179763
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768572}