[mpc] Introduce MatchedPropertiesCache::Key
It's a bit strange that the MatchedPropertiesCache doesn't know
how to hash its own key, and that this instead happens in StyleResolver.
It looks like it's done this way to prevent calculation of the hash
value twice (first when checking if there's a cache entry, and then
once again when it's being added to the cache).

In order for CachedMatchedProperties to take ownership of the caching,
but also prevent calculating the hash twice, this CL introduces
CachedMatchedProperties::Key which can be instantiated at the call-
site.

There should be no functional change in this CL.

Bug: 1057072
Change-Id: I5ba0b8b59009f67318a610f3da51254f77dfecd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2193772
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768708}