[signin] DiceWebSigninInterceptor distinguishes sync singin from web signin
The DiceTabHelper is attached to the tab at the beginning of a signin
flow that are started from Chrome UI, and can be used to tell whether
a signin flow was started in a given tab.

This CL now also tracks the end of the signin flow: as the flow
progresses and completes, the DiceTabHelper is notified, and now is able
to tell if the sync signin flow is currently in progress.

This information is then passed to the DiceWebSigninInterceptor, which
uses it to decide if the signin is a sync signin or a web signin.
Note that this is slightly approximative, as the user can abort the sync
signin and do a web signin in the same tab instead.

Since the ProcessDiceDelegateImpl now needs to know about the
DiceTabHelper, the code can be simplified: parameters of
ProcessDiceDelegateImpl are removed and instead the
ProcessDiceDelegateImpl gets the information from the DiceTabHelper
directly.

Bug: 1076880
Change-Id: I0c91c1404f4198d8759b42577accc6b3b7600168
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2185050
Commit-Queue: David Roger <droger@chromium.org>
Reviewed-by: Jan Krcal <jkrcal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768717}