[Autofill Assistant] Reenable direct action tests.
We forgot to reenable these tests. They pass locally and were already
passing when we last looked into this. The reason for the flakes is
usually the underlying testing infra that has a hick-up. The last
instance being a "Power off" widget obscuring the button we needed to
click.

Bug: 1057806
Change-Id: Idbbd63afca3d8df0da20b228f863498677d40024
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2201556
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Commit-Queue: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768762}