[Autofill Assistant] Allow implicit login selection if no pwm logins.
Previously, implicit login selection was only enabled if the list of
available choices had degenerated to a single item. Now, it is also
enabled if a default option is set and there are no password manager
logins available.

This CL also fixes a number of issues where the action would be allowed to prematurely terminate because the check for OnlyLoginRequested was erroneous and did not account for some of the more recent section like date/time and additional prepended/appended input sections.

Bug: b/155971226
Change-Id: Ic7df9e8381265207debe6c1f17229285ade2112f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199115
Reviewed-by: Marian Fechete <marianfe@google.com>
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#768763}