indexeddb: add indirection to IndexedDBControl
Rather than passing around a remote, pass around a wrapper that owns the
remote.  This is a setup patch for a patch that will add watching
origins for changes to storage policy.  This wrapper is needed to
intercept origins that have factories created for them in order to watch
them, and to have a place (that's not StoragePartitionImpl) to implement
observers of the special storage policy.

This is a no-op on its own, but sets up future work.

Bug: 1015214
Change-Id: Icd99823cd46faf33a0bf6e4401f27f58b51af708
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199914
Commit-Queue: enne <enne@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768787}