Adds focus ring background alpha option for Chrome OS.
Users have requested a way to darken all of the screen except what
is being read by Select to Speak, in order to improve focus. This
draws a semi-transparent black color over the rest of the screen
except for where the focus ring and its contents are drawn.

Plumbs the focus ring background alpha to Select to Speak, although
it is not used yet. Setting the alpha in prefs will be a separate
change.

Ax-Relnotes: N/A
Bug: 1079424
Change-Id: I5c0355fd33b8f595c0e761a46871405a0437ec22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2188532
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Anastasia Helfinstein <anastasi@google.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768861}