Replace 'incognito' with 'OffTheRecord' in DatabaseTracker.
DatabaseTracker uses the term 'incognito' with the meaning of
'in-memory'. This is used for all profile types that are OffTheRecord
(including incognito, guest, and non-primary OTRs).
To fix the possible misinterpretation, 'incognito' is replaced with
'OffTheRecord'.

This CL does not make any behavioral change and is just for better code
readability.

Bug: 968028
Change-Id: Ifce346865d9d054883ab45679fd3a87144e88bd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2201377
Commit-Queue: Victor Costan <pwnall@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768893}