[ua-ch] Turn on feature flags
This CL turns on the feature flags for User-Agent Client Hints, as well
as the Client Hints Feature Policy infrastructure.

I2S: [1]
One test [2] was impacted by a combination of this change and an issue
related to the Client Hints infrastructure (but not to this change
in particular).

It is being tracked in [3] and fixed by [4], which we'll rebase and
land once this CL lands (and both change the test).


Bug: 928669
Change-Id: I0056827b739b0f6773a49fc8cffe456e5de21803

[1] https://groups.google.com/a/chromium.org/d/msg/blink-dev/A4wxFpvqUfA/g7iccl9ICgAJ
[2] https://chromium-review.googlesource.com/c/chromium/src/+/2199266/8/chrome/browser/client_hints/client_hints_browsertest.cc#1879
[3] https://bugs.chromium.org/p/chromium/issues/detail?id=1082674
[4] https://chromium-review.googlesource.com/c/chromium/src/+/2201976

Change-Id: I0056827b739b0f6773a49fc8cffe456e5de21803
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199266
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768910}