WebUI: Add JS type checking for cr-button tests.
Also adding placeholder js_type_check() targets for the remaining
cr_elements/ Mocha test files.

This is done to facilitate the parallelization of the effort to add JS
type checking for WebUI tests. It makes it easier to see which parts of
the code are not type checked yet.

Bug: 1000989
Change-Id: I5bd77fe43ef43b574d6631633800b25c7cfe0f63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199181
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768964}