Feed: Leave it up to Offline Pages to decide how to open Feed articles.
Previously, Feed articles would always open the offline version of a
page if it exists, which isn't ideal for users who are online.

This CL changes this behavior to open the article normally, which
leaves it up to Offline Pages to decide whether or not to open the
offline version of the page (e.g. when offline).

TBR=twellington@chromium.org,wychen@chromium.org

Bug: 1069906
Change-Id: I5723e418747867411511ede72f43c323b493b154
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151175
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Commit-Queue: Scott Little <sclittle@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768972}