Unqueue'ed Uninstall + Purge DBus methods
DBus method calls to |Uninstall()| and |Purge()| to dlcservice will not
be queue'ed and all dlcservice errors will be propaged to callers.

This change is to reduce complexity in the client and transition the
complexity into the platform dlcservice.

BUG=chromium:1071656
TEST=./unit_tests

Change-Id: Ic58432cd09dc8f41422ab6ff5667e5e36a4d6c3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199687
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Auto-Submit: Jae Hoon Kim <kimjae@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768981}