Batch accessibility event processing during page load
This change allows us to defer sending events to
improve performance during the initial page load.

EventScheduleMode determines if we schedule events
with a delay or not.

EventScheduleStatus replaces ack_pending_, and tracks
what state we are in with regard to event scheduling.

This is based on:
CL 1691523: Large batch accessibility event processing
during page load

Bug: 976849
Change-Id: I74cd5f31102199f8c5752e0ef5d0eb623fdbd9d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2123310
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Adithya Srinivasan <adithyas@chromium.org>
Commit-Queue: Jacques Newman <janewman@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#769010}