[Extensions UI] Add an NTP-specific settings overridden dialog
Add a specialization of the settings overridden dialog for extensions
that override the New Tab Page. This involves adding a generic
controller for the settings overridden dialog to handle extension-
controlled settings (which will be re-used for additional controlled
settings) and an NTP-specific provider to handle the NTP bits.

Add unit tests for both the NTP-specific provider and the generic
extension-controlled settings dialog controller, and a UI browser
test for the NTP overridden dialog.

Currently, this is only used in testing; it is not hooked up to any
production codepaths.

Bug: 1079364
Change-Id: I1f360bd603b6514fb492acc0b065d1b6d5236a4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2186642
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769024}