media/gpu/vaapi: Introduce BitrateControl to AcceleratedVideoEncoder
We want to be able to run a VA-API driver in Constant
Quantization Parameter mode for vp9 encoding.
VaapiVideoEncodeAccelerator needs to tell VP9Encoder about the
bitrate control.
This CL adds enum of BitrateControl in AcceleratedVideoEncoder
interface so that VaapiVEA can configure vp9 encoder with the
enum.

Bug: 1060775
Test: video.EncodeAccel.*
Change-Id: I16db7bafcee0642e71bc36ef3e2bce257e33c7b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159669
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769033}