[Cast MRP] Fix for crash in Chrome OS.
This CL is a workaround for the fact that when the user tries to cast
their desktop from the system tray, there is no WebContents available.

The main difficulty is that when there is no WebContents, there's no
obvious way to get the information DesktopStreamsRegistry uses to
authenticate requests for streams.  The solution here is to just leave
that information out when registering the stream on Chrome OS.  This
works because the request used to register the stream is the same one
used to retrieve it again later, which is ridiculous in itself, but an
existing comment in the code addresses why it's done that way.

Change-Id: Ic1243b123caead38526520cb49d6d29dd9d3dc06
Bug: 1078572 b/155807876
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202685
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Commit-Queue: John Williams <jrw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769108}