Remove PictureImageLayer
Now that directly composited images are supported by detecting valid
candidates within cc, we no longer need PictureImageLayer. The remaining
usage (outside of extraneous includes) were in cc unittests. Instead of
PictureImageLayer, these tests are converted to use a FakePictureLayer
that uses a FakeContentLayerClient. The FakeContentLayerClient is then
set up such that it has the same bounds as the layer and contains a
single drawImage, which is recorded to a DisplayItemList inside of
PaintContentsToDisplayList.


R=pdr@chromium.org

Bug: 875110
Change-Id: I58f029ae224de391e7d2a8a907d89f8a8a8beff7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2200041
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#769117}