Return -1 for renaming/deletion errors in chrome-untrusted://media-app.
This patch gracefully returns with a UNKNOWN_ERROR for renaming and
deletion failures due to file / directory handles being out of sync
checked in `assertFileAndDirectoryMutable()`. We should not reach
this state in normal usage, but if we do we should return something
useful i.e. UNKNOWN_ERROR so we can take the appropriate action in
our internal app.

This puts the onus of reporting errors from IPC such as Renaming on
the chromium code tightening the scope of what our internal app needs
to do to handle them i.e. cl/308950872 does not need to try catch.

Bug: b/308950872, 996088
Change-Id: I19e8dc955c586afa2e02fb8b33816ab8cc5c2adc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2193213
Commit-Queue: David Lei <dlei@google.com>
Reviewed-by: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769128}