[a11y] Add RetargetForEvents to BrowserAccessibilityManager for Android
This CL introduces RetargetForEvents to BrowserAccessibilityManager
for Android and moves the logic of updating the target node using
PlatformGetClosestPlatformObject at FireBlinkEvent,
FireGeneratedEvent, and HandleHoverEvent. RetargetForEvents is called
before CanFireEvents. So, Android doesn't need to have separate
implementation for CanFireEvents since PlatformGetClosestPlatformObject
is walking up a leaf node and it checks a leaf node with
PlatformIsLeafIncludingIgnored which makes the difference between
other platforms and Android.

AX-Relnotes: N/A
Bug: 1073900
Change-Id: I03ed93d785fa52f2cf65a14fc8d28c74a8d2a24d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174098
Commit-Queue: Julie Kim <jkim@igalia.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769155}