ozone/wayland: Fix nested menu windows.
Previously, we didn't assume that the there can be
more than 1 level of nested windows. That was wrong.

Basically, there can be an arbitarery number of nested
menu windows, and all of them must use the topmost
window in the stack of the same family.

Bug: 1076815
Change-Id: If10eb47c5c27787e2c56f2209646b352d87509dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2190694
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769181}