blink/heap: stop `std::move`ing a function in a loop
If this branch is taken more than once, we'll use `function` after it
was already moved. That's potentially bad, depending on the
implementation of `function`. If `std::move` is a nop for `function`,
then not `std::move`ing it should produce just as high-quality code.

Bug: None
Change-Id: I77c38567cbd2a16d1ef292dd815cd20c77a517c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202692
Commit-Queue: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769183}