Replace IsIncognitoProfile with IsOffTheRecord in HostContentSettingsMap
Usages of IsIncognitoProfile() in HostContentSettingsMapFactory are
updated to IsOffTheRecord() function.

Currently IsIncognitoProfile() covers incognito and non-primary OTR
profiles, but soon it returns false for non-primary OTRs. This change
ensures that HostContentSettingsMap for the original profile of all OTR
profiles exists, and passes the flag correctly to the newly created
HostContentSettingsMap.

Bug: 968028, 1074201
Change-Id: If40d1cb5958a35909c480c0a1baa14ad4ffd01d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202799
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769279}