[WebLayer] Reduce memory usage in WebView compat mode
If only WebLayer is loaded in WebView compatibility mode, we can save a
chunk of memory by not using the WebView class loader.

Note that this breaks the WebView crash key that tells if it is running
in the same process as WebLayer. I gated this change on M84 or higher
implementation, so we should still have the crash key for the experiment
in M83. It's a bit complicated to get the crash key to actually work
without the class loader available, so I don't think it's really worth
the effort (especially since it's hard to write tests to verify the
crash key is working correctly).

Bug: 1061512
Change-Id: Iedf569e13d61048d241edb6e743c9384091c75fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2200082
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769311}