Print Preview: Make NativeLayer an interface
Make the NativeLayer class an interface, and add NativeLayerImpl to
implement the class. Remove the custom getInstance/setInstance methods,
and leverage the shared addSingletonGetter method.

Update NativeLayerStub to implement NativeLayer and update several
tests to the the new methods. Remove calls to PolymerTest.clearBody()
along the way.

This is preparation work to allow for closure compilation of the
Print Preview JS tests.

Bug: 1000989
Change-Id: I17bd3a0c749a092b2dc8eb52be430b8f14628153
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2203028
Reviewed-by: dpapad <dpapad@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769327}