Use CQ results to update test expectations.
The purpose of this cl is add try server test results
to update test expectations.

This change filters cq results for which retry without patch
has passing tests.
If any test have passed it indicates the patch is unreliable,
and the results for that build can be ignored.

Bug: 1079451
Change-Id: I67527c2107ff40d88ac4ecfeca6cf7a7e70ee44c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067319
Reviewed-by: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Robert Ma <robertma@chromium.org>
Commit-Queue: Ilia Samsonov <isamsonov@google.com>
Cr-Commit-Position: refs/heads/master@{#769351}