Add finch feature AllowNonSecureOverlays
AllowNonSecureOverlays controls whether MediaCodecVideoDecoder will
try to promote non-L1 content to an overlay.  This is normally the
behavior we want; it allows power savings.  For the power A/B
experiment, though, we want to compare overlay vs non-overlay
performance.  L1 is excepted because it'll fail otherwise.

Bug: 1081346
Change-Id: I9d365c297e6919f5ac10a89b0b312d6d7fa0fb6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202968
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769368}