[ios] Create TranslateOverlayTabHelper
This change adds a TabHelper that listen for
TranslateInfobarDelegate::Observer callbacks, inserting a Translate
overlay banner request when Translate finishes and adding a placeholder
when Translate is initiated. It will indicate to its own observers
of these changes in states. PlaceholderRequestCancelHandler
will be one observer that, upon receiving that callback, will cancel
its own request now that the second Show Original banner is ready to be
displayed.

InfobarPlaceholderRequestConfig is added here for testing support,
but will be used by the Translate overlay interaction handlers
when Translate is initiated.

Bug: 1071914
Change-Id: I1ca35fd98c6c218f4b36700f43652336c9224802
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2180864
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769388}