Introduce accessibility focus highlight feature.
When enabled, this feature draws a highlight around the focused
object on a web page whenever focus changes. This can make it
easier to see what object has keyboard focus, especially since
many pages hide it.

There's also a similar feature on Chrome OS, but it uses a layer
above the entire system. This implementation for Chrome browser
uses a layer on top of a BrowserView.

This is behind a flag for experimentation. It won't be shipped
until UI review is complete.

a flag. Run with --enable-features=AccessibilityFocusHighlight
to try it.

Bug: 1021939
AX-Relnotes: introduces the new focus highlight feature behind
Change-Id: I6fa26be34711e2bfa45501b02fbda3c620ccd248
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2191975
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Alice Boxhall <aboxhall@chromium.org>
Reviewed-by: Chris Hall <chrishall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769434}