Only set background_color on cc:Layer where background is painted
This patch improves the checkerboarding color in composited scrolling
cases when the main layer's background color is not opaque. Previously,
when the scrolling contents layer checkerboarded, we would use a
non-opaque version of the background color on the scrolling element.
However, since the layer that actually paints the background color is
*not* checkerboarded, it's better to just leave the checkerboarding for
the scrolling contents layer transparent, which will reveal the
background painted behind it.

This also makes the pre-CAP code path behave more like CAP.

Change-Id: I8edb050bef84096b4929fc6a230344c28a511386
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181659
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769446}