HSTS Preload List edits.
This change updates the policy one one entry from "bulk-1-year" to
"public-suffix" as that domain is now on the PSL. It also adds an entry
with policy "bulk-1-year" for a domain that was recently removed from
the PSL (but the change to the PSL hasn't propagated to the go net
library). Both entries meet all other bulk entry criteria.

Change-Id: I71555122b53118894507912cee5bcc98cf2de251
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189355
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Auto-Submit: Nick Harper <nharper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769449}