sensors: Remove cross-origin iframe test from runGenericSensorTests().
This test is wrong, which causes it to be flaky: it just adds an iframe and
waits for it to load without properly taking care of window focus, so the
sensor in the main window continued running and we could end up receiving an
unexpected "reading" event if the iframe's "load" event took to long to be
delivered.

Remove the test altogether in favor of run_generic_sensor_iframe_tests(),
which checks for the same thing the right way (although it is also a bit
flaky...).

Bug: 731018
Change-Id: I568f032233be767a2d2a3d50bd3de8d8ce3db0ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204024
Auto-Submit: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769458}