[omnibox] [rich-autocompletion] Add complementary omnibox view.
*Rich autocompletion background*

Currently, autocompletion is limited to when the user types a prefix of
the URL. E.g. 'wiki' can autocomplete 'wiki.org/thecookinglist - top
awesome recipes'.

Rich autocompletion will allow autocompletion even when the input is
from the suggestion title or middle of the URL. E.g. 'thecoo' and
'recip' will also autocomplete 'wiki.org/thecookinglist - top awesome
recipes'.

Design doc with further details: go/richomniboxautocompletion

Previous CLs:
- crrev.com/c/2107796 added chrome://flags flags
- crrev.com/c/2108780 supported multiselection in textfield views.

*This CL*

This is the 3rd CL for rich autocompletion. It adds a label view to the
location bar to display the complementary omnibox text. I.e., while the
omnibox displays either the URL or title depending on which matches the
user input, this label displays the other field. This label is placed
either to the right of or below the omnibox, depending on the 'two line
omnibox' param.

This CL has a small user facing change not guarded by the rich
autocompletion flag. The location bar now redirects mouse events to the
omnibox in order to handle mouse events on the complementary label. This
has the beneficial, user facing side effect that clicking in the
leftmost or rightmost 20px's of the location bar will now correctly
focus the omnibox when the location bar decorations are hidden (i.e.,
when user input is in progress).

This CL has no other user facing changes when the yet-non-functional
rich autocompletion flag is disabled. Followup CLs will wire in rich
autocompletion to make use of these changes.

Bug: 1062446
Change-Id: Iedb681babdb03e05b80a8252444bf5882ab408cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199859
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769514}