Fix extension cache inconsistency for version and Id mismatch
When saving the CRX file to the cache, we have assumed(before this CL)
without any check that its extension id and version are the same as we
expect from the update manifest. If the time of update request on the
client is close to the extension version update, we may encounter the
issue of cache inconsistency: save cache entry marked with one version,
while in fact it's another.

This CL fixes this problem by inserting in the cache only when the crx
meets the expected requirements from the manifest. This insertion logic
has hence, been moved extension_downloader to the extension_updater
which sets |expectations_passed_callback_| in the crx_installer. This
callback is invoked when the crx passes the expectations and inserts
a new entry in the cache. It also holds the responsibility to delete
the crx file in case cache does not exist or insertion fails.

Browser tests are also included in the CL to verify this fix. A new
update manifest file has been added for testing which mentions an
incorrect version for the contained crx file.
Also fixed the existing tests that fail due to having incorrect
extension version for their crx file.

Bug: 1047637
Change-Id: I6bc291b3f5d21e536a5d8350d34d3dcc222c8f5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2179543
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Cr-Commit-Position: refs/heads/master@{#769585}