MacPWAs: Separate app focus from app reopen
This is towards allowing apps to run with no windows open.

Two different messages differentiated here:
- applicationWillBecomeActive (aka Focus)
  - happens when any window in the app is focused, or when the app
    is command-tabbed to be active
  - should focus existing windows
- applicationShouldHandleReopen (aka Reopen)
  - happens when the dock icon is clicked
  - should launch a new window if no window is open

It used to be that Focus was used by legacy apps and ignored by
PWAs. Also, Focus was used to re-open the app (inappropriately), and
also was used to close the app if it somehow was uninstalled while
still running.

Update the legacy app behavior to open new windows only on Reopen,
and remove the "close the app" behavior.

The next step here will be to handle Reopen for PWAs, and to allow
them to keep running even if no windows are open.

Bug: 1080729
Change-Id: I5bf86e7e7874b71c030d8dc9c5440ee7fec430d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2197231
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769602}