Handle file errors gracefully in chrome://media-app.
Handles errors that may result when trying to open (or "stat") files
upon launch or navigate.

Modifies the mediaApp API to include an 'error' field on AbstractFile,
which associates a file error name with files that could not be opened.
These errors are ignored unless that file is navigated to.

Translated UX for the more common errors needs to be built in the app.

"Unexpected" file errors are also dumped to console in the privileged
context because we don't try to send over the full error message. The
unprivileged context will also generate a crash report for unknown
error names, and do UMA for "known" error names.

Errors on the *launch* file are always dumped, and a placeholder file
with just the name (and the error) are sent to the guest so that some
UX can be built around it. Currently it just says "unable to decode."
in a toast.

Bug: b/149801176, b/134975059, b/156437880
Test: MediaAppUIBrowserTest.LaunchFile
Test: MediaAppUIBrowserTest.LaunchUnopenableFile
Test: MediaAppUIBrowserTest.LaunchWithUnopenableSibling
Test: MediaAppUIBrowserTest.NavigateWithUnopenableSibling
Test: MediaAppUIBrowserTest.FileThatBecomesDirectory
Change-Id: I0468e54b6f9fa24d19a49288b223b696f440e840
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2200822
Commit-Queue: Trent Apted <tapted@chromium.org>
Reviewed-by: Bugs Nash <bugsnash@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769621}