[AA direct actions] Direct actions disabled fix.
Before this patch we would pass an empty bundle to the unchecked
Consumer callback interface. The corresponding test got disabled a while
back due to trying to verify that. The Consumer interface for direct
actions expects a List<>, which is what this patch fixes.

This in turn exposed a bad test requirement for performing an action,
where the response should be an empty bundle for disable direct actions
and not fail as the test was written.

Bug: 1034712
Change-Id: I1d49d6ebb65dade9e1b0967babde9443e694e4e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2201777
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769629}