[Android] Photo Picker: Implement full screen.
This CL adds the Full Screen button to the bottom right
corner and implements the necessary logic to remove the
status bar and the navigation bar, freeing up more space
for the video player and giving a more immersive
experience.

This CL also hides the overlay controls until the size of
the video is known, to avoid having the controls briefly
appear in the wrong position.

Bug: 895776, 656015
Change-Id: I2a09bc74324389eb815e571e8fea923cc255770a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181211
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769653}