[content] RenderDocumentHostUserData - Fix for crashed renderer
While introducing RenderDocumentHostUserData, we have encountered a
corner case which happens when the current RenderFrameHost crashes
when (or before) navigating to a new URL.

In in this case the RenderFrameHost commits before the navigation
commit happens, therefore clearing the user data before the commit
in the speculative RenderFrameHost state.

This CL proposes a fix to this case along with adding couple of
test cases.

Bug: 1057611
Change-Id: I2b1f3685bd143710b3ad500cbaccb44cafd8dd95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2128076
Commit-Queue: Sreeja Kamishetty <sreejakshetty@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769677}