Sync: Start again (in transport-only mode) after StopAndClear
The implementation of ProfileSyncService::StopAndClear() has a special
case for when IsSyncRequested() is already false. It turns out this
special case failed to call TryStart() again after StopImpl(), which
could lead to the SyncService getting stuck in "Start deferred" state
when StopAndClear() was called while in transport mode.
This CL adds the missing TryStart() call and a test for this case.

Bug: 1082267
Change-Id: I8b3f172b74ca1945b680dd544c2ea96509e0e9f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199108
Reviewed-by: Jan Krcal <jkrcal@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769692}