[b4p/Settings] Introduce MultiStorePasswordUiEntry
This CL introduces a new class representing passwords after
deduplication across the two stores. It replaces the uses of
PasswordUiEntry in most of the code for chrome://settings/passwords.
For the moment, no actual deduplication occurs, the entries are simply
put in the new format, so no behavior is changed.

Bug: 1049141
Change-Id: I040910647a806e249e9a77e526299cb02e8ad64b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202237
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769731}