Trust Tokens: Treat SRR cache hits as request cancellations
net::ERR_TRUST_TOKEN_OPERATION_CACHE_HIT is semantically a "good thing":
it means that the purpose of this request was to update some state in
the network stack (with a response from the server), but that this
state was already present, so there was no need to send the request.

Currently, this case results in an unactionable "error" message getting
printed to the console, which we want to avoid. This change fixes this
behavior by making blink::ResourceError::IsCancelled return true for
Trust Tokens cache hits. This makes FrameConsole::DidFailLoading decline
to print an error message, as desired.

Even though it won't be visible through the console, the fact that we
abort the request for this reason will still be available to developers
programmatically: Trust Tokens operations "failing" due to a cache hit
throw NoModificationAllowedError. (For more, see the "Errors" section of
https://bit.ly/trust-token-prototype-user-guide.)

R=jbroman

Bug: 1080812
Change-Id: I1f982e70a9cc446d53463684cef246da4ff36712
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199678
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769735}