Reland "viz: Re-enable ignore_undamaged on platforms using GL buffer queue"
Previously the CL (crrev.com/c/2179763) got reverted since it enabled
the optimization in cases where it shouldn't have.
In particular, the optimization was enabled for SkiaOutputSurface where
only_invalidates_damage_rect used to be false and overlays are not
supported.
The diff between http://crrev.com/c/2205886/1..2 fixes that,
avoiding removing only_invalidates_damage_rect and fixing the logic
to enable the optimization in display.cc.

Original CL description:
"SurfaceAggregator has an option to output only the quads that
intersect the damage rect computed by SA itself.

This optimization was disabled on devices with HW overlays or where
the renderer might expand the damage, since SA would not know the
final damage rect.

This CL re-enables the optimization on devices where we can estimate
a bounding rect of the renderer expanded damage.

This is achieved by asking to the renderer for a bounding rect that
represents the maximum area that can be expanded.
SurfaceAggregator will aggregate the boundingrect union the damage.

The damage on the root render pass will be unchanged, so that HW
overlays damage computation won't be affected."


Bug: 1077210, 1082947
Test: viz_unittest (new test and Display tests)
Change-Id: Ib660320b03c0146c62e208332fa6eaa19632c84d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2205886
Reviewed-by: Peng Huang <penghuang@chromium.org>
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769741}