[Merge M84][WebLayer] Reduce memory usage in WebView compat mode
If only WebLayer is loaded in WebView compatibility mode, we can save a
chunk of memory by not using the WebView class loader.

Note that this breaks the WebView crash key that tells if it is running
in the same process as WebLayer. I gated this change on M84 or higher
implementation, so we should still have the crash key for the experiment
in M83. It's a bit complicated to get the crash key to actually work
without the class loader available, so I don't think it's really worth
the effort (especially since it's hard to write tests to verify the
crash key is working correctly).

TBR=torne@chromium.org

(cherry picked from commit 8632f4046c18eb580f86a935b7468c7042fe56fe)

Bug: 1061512
Change-Id: Iedf569e13d61048d241edb6e743c9384091c75fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2200082
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#769311}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2207518
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/branch-heads/4147@{#36}
Cr-Branched-From: 16307825352720ae04d898f37efa5449ad68b606-refs/heads/master@{#768962}