[webauthn] Fix GetAssertion crash with UV blocked
When getting a UV token, if the authenticator returned an error because
internal user verification is blocked, we try to dereference
`authenticator_` to see if fallback to PIN is supported, but it hasn't
been assigned yet. Fix the code to reference `authenticator` instead.

This also includes tests that cover this case for MakeCredential and
GetAssertion and repro the issue for the latter.

Fixed: 1082671
Change-Id: I0ceb03f47f8e88e048b08db6cbf8dcc19ae2fc77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204379
Auto-Submit: Nina Satragno <nsatragno@chromium.org>
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#769764}