DevTools: do not set devtools id for internal requests
Internal requests do not issue Network.requestWillBeSent
and similar events. However, we currently mark them with devtools id,
and corresponding Fetch events have a networkId reference. This brings
confusion to the protocol, because the request with the same networkId
never arrives in the Network domain.

This was regressed in https://crrev.com/757049.

Bug: 1053468
Change-Id: I66b8da720833c45b683004334afbea709c999223
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2203539
Commit-Queue: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769802}