[FlexNG] Skip relayout boundary optimisation if a flex-item.
Before this patch we'd consider a text-control a relayout root. If this
was also a NG flex-item it wouldn't have enough information to correctly
layout, (it wouldn't know what size to stretch to for example).

This patch disables the relayout boundary optimisation for text-controls
and SVG roots - which are also flex-items.

Bug: 1082447
Change-Id: I0e3b2186301fbf9488c0723c525a5e45e2ca80df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2207611
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769810}