Page name propagation.
The main motivation behind this change is that we may want to name the
first page (to give it a non-default orientation, for instance).

This requires two things:

1. Manually name the first entry in the page name mapper. We normally
insert named pages as part of inserting a forced break. However, there's
no break before the first page (or it wouldn't be the first page
anymore). Add a way to name the first page(s) manually, at the end of
LayoutView layout.

2. Correct start/end page name handling and propagation, as described in
the spec [1]. An author may want to set the first page name on the first
SECTION, DIV, etc., rather than specifying it on the root element (which
would be rather non-intuitive). Therefore, improve on our current,
rather simplistic, approach to how the layout engine deals with named
pages, and do it exactly as the spec describes.

Page name propagation is currently only supported for block containers.
It would also be appropriate to support this for table, flex, etc.

[1] https://drafts.csswg.org/css-page-3/#start-page-value

Bug: 1053768
Change-Id: I0f16cfb4325f6641b2c77355e03a6e439084d333
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2205937
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769840}