Simplify PaintController::FinishCycle()
We don't need to do anything in FinishCycle() for transient
PaintController, or if we haven't committed anything since the last
FinishCycle().

Remove code that resets paint property changed flags from
PaintArtifact::FinishCycle() for CAP, so that LocalFrameView can do that
in the same way for both CAP and pre-CAP.

This doesn't actually reduce any extra calls for crbug.com/1049159, but
prepares for that. This CL actually avoids the repeated VLOGs in
FinishCycle when it does nothing.

Bug: 1049159
Change-Id: I6f4f11acebc3194cf85ac26b909c663dba422a3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2205977
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769843}