Use the WebUI associated with RenderFrameHost directly.
The usage in RegisterWebUIControllerInterfaceBinder of
WebContents::GetWebUI() leads to potential race conditions. Navigations
between different WebUIs are cross-process, which means that two WebUIs
run in separate processes and their IPC messages to the browser process
can race against each other.

One particular case that I believe is hit in reality is as follows:
1 Start with WebUI1 in a tab (runs in process 1).
2 Navigate to WebUI2 in the same tab.
3 Send the CommitNavigation IPC to WebUI2 (in process 2).
4 WebUI2 sends DidCommitNavigation IPC to browser process from process 2
5 Process 1 sends interface bind request to browser process for WebUI1
** This is still possible, since WebUI1 is running in process 1 and has
   not been unloaded yet.

In this sequence of events, in step 4 the main RenderFrameHost for the
tab will be changed. In step 5, the code in
RegisterWebUIControllerInterfaceBinder gets the WebUI from WebContents,
which will return WebUI2, while the interface request is for WebUI1,
leading to mismatch.

This CL fixes this race by exposing the WebUI associated with a
RenderFrameHost directly, through the public API, so the code binding
incoming interface requests will use the WebUI from the RenderFrameHost
itself instead of the potential mismatch from WebContents::GetWebUI.

Bug: 1081078
Change-Id: I4998efb72573bcfb4166db7a99ca98014689eb3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204850
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769853}