Use std::unique_ptr overload of View::AddChildView.
This CL changes:

    AddChildView(webui_view_ = new WebUIOmniboxPopupView(
                 location_bar_view_->...));

into:

    webui_view_ = AddChildView(
        std::make_unique<WebUIOmniboxPopupView>(
            location_bar_view_->...));

This makes the code easier to handle for the |rewrite_raw_ptr_fields|
tool that changes raw pointer fields like |webui_view_| to use
CheckedPtr instead (this project is still in an investigation phase,
we don't know yet if we will proceed with such rewrite).

Additionally, the change is desirable on its own - the overload
of View::AddChildView that takes |std::unique_ptr| is preferable
over the one that takes a raw pointer (as pointed out in the
comment for that method).

Bug: 1080832
Change-Id: I9c371b99244396ffe89efd95cd3ea4496b712230
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2207474
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769866}