[iOS] Create PendingUnsafeResourceStorage
This helper class stores a copy of an UnsafeResource while its allow
list decision is pending.  When the pending decision is finished (either
by allowing or disallowing the threat), the resource is reset.

This CL also adds a new util function that returns the allow list for
an UnsafeResource.

Bug: 1083557
Change-Id: Iec9e32fdc68a59f43c98c33e2f04f1ee2f5f5bdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2205115
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769911}