Fix fallback task provider
The fallback task provider triggers and provides a task for a render
process if the render process has a service worker in it but no tabs.

Add the ability for the fallback task provider to fall back for a
collection of providers, not just one. Use that to fall back for both
the tabs and the workers.

Bug: 1083664
Change-Id: Ia7d9d0248fb40df3c25c43ec2bcd915f78c159ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204363
Commit-Queue: Avi Drissman <avi@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769955}