ozone/wayland: Share ScopedWlArray and SendConfigureEvent between tests
Just moving ScopedWlArray into a shared place and making the
SendConfigureEvent available for other tests.

It also automatically sends configure event to the window so that
tests wouldn't need to activate the window by themselves.
Basically, this is going to be needed for the CL I'm going to submit
next.

Bug: 1022750, 1083949
Change-Id: I4ce6f7d8720266b8ee20c9d8cf6ce3c50490791e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2207170
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#769991}