[WebLayer] Add ability to start shell / browsertests in incognito mode
For testing that translate works in incognito mode I'll need a
browsertest where the Profile is actually *in* incognito mode. This CL
adds the ability to start the WebLayer shell and browsertests in
incognito.

There was a TODO that mentioned adding an incognito Profile to the
regular Profile in WebLayer Shell, but this CL instead takes a more
straightforward approach of just having the existing Profile be created
in incognito mode if desired. That is sufficient for the upcoming use
case.

Bug: 1072334
Change-Id: Ib56c1177e2e6f6c953f6bb985d532961a2e01245
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2198970
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770037}