[content] Record UKM events without using WebContents::GetMainFrame()
This CL fixes the potential problem of wrong frame association for
sending UKM.

When we access WebContents::GetMainFrame(), it returns a current frame.
But for DisplayCutoutHostImpl::RecordUKMEvents(), we would like to send
the events to the frame which could have been navigated away.

With this CL, the frame association is included when we queue the events
so that we do not use WebContents::GetMainFrame() any more.

Bug: 1061899
Change-Id: Ib1a9d616812b38098332e86f6604ef837791f459
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2198329
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Sreeja Kamishetty <sreejakshetty@chromium.org>
Commit-Queue: Yuzu Saijo <yuzus@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770051}