Always enable QUIC draft27 when speaking QuicTransport
QuicTransport requires TLS 1.3 and message frames, but the version of
QUIC we currently use by default does not support that.  Making the
version used by QuicTransport depend randomly on current experiment
config seems to be counterproductive for the purposes of origin trial,
so let's pick draft27 and use that for the duration of the trial.

Bug: 1084289

Change-Id: I7deb72e742692cf08c8922756109b48ce64ec9fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208090
Commit-Queue: Victor Vasiliev <vasilvv@chromium.org>
Reviewed-by: Zhongyi Shi <zhongyi@chromium.org>
Reviewed-by: David Schinazi <dschinazi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770052}