[Android] Remove InvalidationGcmUpstreamSender implementation
InvalidationGcmUpstreamSender uses deprecated ChromeSigninController.
This class is also part of the old invalidation code and can be removed.
This cl removes the implementation of the class in order to reduce usage
of ChromeSigninController. An empty override of deliverMessage() is kept
to avoid crashes.

Possibly this whole class will be removed in the future.

Bug: 1074072
Change-Id: I4874a1582cbcb30d63884c7c764a8932b963bedf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202976
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Tanmoy Mollik <triploblastic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770059}