ozone/wayland: Do not grab if no seat available.
If seat is not available, and Chromium tries to create a popup
window (like in browser tests), Wayland will drop the connection
with a protocol error.

Thus, add additional check for that. Also, create "CanGrabPopup"
method so that the same condition is shared between xdg-unstable-v6
and xdg-stable.

Test: WaylandWindowTest.DoesNotGrabPopupIfNoSeat
Bug: 578890
Change-Id: If47d9044c45b37e149fae1d67b244c54a99ee4e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2207378
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#770086}