[Lint] Add Missing TargetApi annotation in MultiWindowUtils
While there was a NewApi lint outage a new method was added to
MultiWindowUtils that should have been annotated with TargetApi. This
CL adds the missing annotation and adds a check to make sure this
won't cause a crash if called from a pathway on a lower API version.

Bug: 1081242
Change-Id: Idd54a73d849daadd5ab915158d10be43f5dc6fde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199718
Reviewed-by: Peter Wen <wnwen@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770096}