Add trusted intent check before we handle with PaymentHandlerActivity.
Through IntentURI we can create a CCT via PaymentHandlerActivity which
opens the CCT halfway in front of the regular tab. This allows a
malicious CCT to represent itself as the content of the background tab.

This CL adds an extra check to ensure we fire PaymentHandlerActivity
only if the intent was sent from Chrome and not via external navigation.

This CL also adds a browser test.

Bug: 1083972
Change-Id: I136774e23a641fd5feef9e47aaa87ba2d2f94b48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2205933
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Rohit Agarwal <roagarwal@chromium.org>
Commit-Queue: Rohit Agarwal <roagarwal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770215}