Deflake geolocation-api/watchPosition-page-visibility.html
The test wants to updatePosition() after its not visible but it was
doing it before waiting for the visibility change to occur. Then it
wants to verify that it's not received but it was being received
flakily and racing with the next visibility change. So we use
roundTripToBrowserThen() to wait for the updatePosition() to take
effect if it would do so incorrectly.

R=nasko@chromium.org

Bug: 866140, 1069111, 1083452, 1083585
Change-Id: I0a51f0919b579456154b408e1455cd41d5416c95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204646
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770222}