[Sampling profiler] Pass Android native unwinder to StackSamplingProfiler
Instantiates the native unwinder provided by the stack_unwinder dynamic
feature module, and passes it to the StackSamplingProfiler on Android.
This is required for StackSamplingProfiler to unwind native and Java
stack frames on Android.

This implementation shares the same MemoryRegionMap over all native
unwinders. If we need to use separate MemoryRegionMaps between unwinders
(e.g. to account for new modules loaded) this should be possible with
local changes to create_native_unwinder_factory(). The unwinder factory,
and implicitly the Module and MemoryRegionMap, is leaked since it
needs to have full process lifetime anyway and that alternative of
having callers manage a shared resource for it would be cumbersome.

Bug: 1004855
Change-Id: I3d8c2981a720f564deb8774f39ec3e7adb60619e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202416
Commit-Queue: Mike Wittman <wittman@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770238}