AW UMA: add CHECKs for malformed UMA proto
This adds CHECKs to verify the chrome_user_metrics_extension proto is
correctly formed for WebView. We see records on the backend which are
missing system_profiles (and other fields). This adds CHECKs to catch
this scenario and provide more info for why this is happening.

Bug: 1081925
Test: None
Change-Id: I7a58edfc0679798bd50e1dac6da9defff76a0c93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2197481
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770290}