Hide hotseat when non-shelf window is activated
It is a partial fixing for the issue 1072043, where hotseat still shows
after launching an app from context menu. In this CL, when the activated
window is not a shelf window, reset |is_manually_extended_| of Hotseat-
Widget. Create an end-to-end browser test for this scenario.

However, this solution does not cover the edge case of launching an app
in a new tab instead of a window.

Bug:1072043

Change-Id: Ia94947baecc8d71690e82a7ba4def08f9a7bab7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2197343
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770356}