Extend MessageCardProvider component to support grouping suggestion
This CL modifies the TabSuggestionMessageCardViewModel to build a
PropertyModel based on the TabSuggestionAction.

In addition, this CL updates the relationship between the shown Message
and the available Message in MessageCardProviderMediator. Prior to this
change, each MessageType is allowed to live in either the shown Message
queue or the available Message queue, not both. After this CL, each
MessageType can live in both. The available Message for the same
MessageType can be queued up, and each MessageType will have one Shown
Message.

Change-Id: Icb57147823bbcf81679d38c600d894ed6e47317c
Bug: 1076538
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2196845
Commit-Queue: Mei Liang <meiliang@chromium.org>
Reviewed-by: Yue Zhang <yuezhanggg@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770370}