Mark <svg> and <img> child elements of cr-icon with role none
|cr-icon-button| contains icons, represented by |iron-icon| polymer
controls. The icons appear in the accessibility tree of the
|cr-icon-button|. The tree looks like this

   |BUTTON name=Download|
     - |IRON-ICON|
       - |SVG/IMG name=''|
   |BUTTON name=Print|
     - |IRON-ICON|
       - |SVG/IMG name=''|

As child <svg>/<img> are not interactable and lack name property, it
makes sense to remove the svg and image children of |iron-icon| from
a11y tree. This change explicitly marks the <svg> and <img> children
of |iron-icon| inaccessible by adding aria role as none.

Bug:1069531

Change-Id: Idcd7912c865ce08e8871d416b30959ab64ae3d09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2143183
Commit-Queue: Virender Singh <virens@microsoft.com>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770375}