[usb] Convert DCHECKs to CHECKs to narrow down crash
In issue 1084316 we see a crash in GotDescriptorFromNodeConnection()
which seems to point to a read or write outside of either the source or
destination of the memcpy. This shouldn't happen if the transfer length
given to us by the operating system is valid. These checks will narrow
down whether it is too large or too small.

Bug: 1084316
Change-Id: I1627ca80430de9d617f551de54bc81068ce20f1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208174
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770407}