chromeos: Cleanup ExternalPrefLoader::PrioritySyncReadyWaiter
This class exists as a first-login performance optimization for
Chrome OS. Wrap it in OS_CHROMEOS macros to make that clear.

Cleanup an unnecessary check for browser_defaults::kSyncAutoStarts.
This is always true on Chrome OS.

This is in anticipation of making some Chrome OS specific changes to
this class for the go/split-settings-sync project.

Bug: 1078479
Test: existing unit_tests, manually tested default apps on chromebook
Change-Id: I00ee353fe8c94834ef112d5841dd16fd958badb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204696
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770413}