[WebLayer] Implement translate in incognito mode
WebLayer currently does not support translate in incognito mode for the
following reason: In //chrome, TranslateRankerFactory supplies the
original Profile that is associated with the incognito Profile when the
service is requested for a Profile in incognito mode. However, WebLayer
doesn't have a concept of incognito profiles being associated with
regular profiles.

This CL adds support for translate in incognito mode in WebLayer by
having TranslateRankerFactory just have its own instance in incognito
mode. This seems reasonable based on the fact that in WebLayer
incognito Profiles really are standalone Profiles just like regular
Profiles.

Bug: 1072334
Change-Id: Ibb5a6ae3823316a6ce966245fff146846f262280
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2198786
Reviewed-by: anthonyvd <anthonyvd@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770481}