Convert content's non-'Routed' input test APIs to routed versions
The routed version of SimulateMouseClick and SimulateMouseEvent should
be preferred, since it exercises the real codepath taken when handling
input (and can target OOPIF). Remove the non-routed implementations and
replace existing implementations with the routed version.

Also remove a couple of 'Routed' methods that are not currently used.
- RouteMouseEvent is the same (now) as SimulateMouseEvent
- Taps are not aura specific, so new tests should use SimluateTap* APIs


Bug: 708651
Change-Id: Ic1c5f3702a81ec6f65195d03cb08a38efa81ee01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2200427
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#770497}