Handle negative zero when hashing FontDescription
FontDescription may hold negative zero float members. In this case,
it compares equal to positive zero, but the hash values are different.

This patch converts negative zero to zero before hashing, so that
the hash values are the same, and the common assumption that equal
values have equal hash is ensured.

Bug: 1081017
Change-Id: Ibc02993967d7638a065d045a138e4d966a20b2bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2194912
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770498}