[signin] Add backend for reauth webUI page
This CL contains backend changes in preparation to adding a new
chrome://signin-reauth webUI page. This page is going to be displayed in
a tab modal dialog as part of the account password storage opt-in flow.

This CL puts the reauth webUI page as a replacement for the fake reauth
dialog. A proper integration with Gaia Reauth will be implemented in
following CLs.

This CL should be submitted together with https://crrev.com/c/2193667,
which adds a proper frontend implementation of chrome://signin-reauth.

Bug: 1045515
Change-Id: I043f8c0bf1a460828e38663537ddef201ea1e1ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199146
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770511}