Fix GPU integration test tags
Both [no-]skia-renderer and [no-]use-vulkan tags in GPU integration
tests were based on the command line. Code assumed that if there was
nothing on the command line then the feature was disabled which was
incorrect.

Change the tags to be based on if the feature is enabled or not. Also
change use-vulkan to just be vulkan for brevity.

Bug: 1069551, 1084569
Change-Id: I33fdbfb4d8078127f066f4a86a267477e4e9e020
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2209348
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770579}