Move /deep/ and ::shadow tests to web components v0 folder
This is part of a multi-CL set that will move all "Web Components v0
only" tests into a separate folder. This will make landing the eventual
removal of WCv0 easier, because this entire folder can simply be skipped
rather than requiring file-by-file changes. There should not be any
tests in this folder that are needed for testing Web Components
VERSION 1 features.

While this CL does not change any behavior or testing now, care should
be taken in reviewing the list of tests, since when WCv0 is removed,
all of these tests will be deleted. So any tests being moved here that
are required for testing a non-WCv0 feature should *not* be included in
this move.

This set should only include tests of the (Shadow DOM v0 specific)
/deep/ and ::shadow pseudo element. Two tests were modified in-place
to remove the dependency on these concepts.

Bug: 937746
Change-Id: I082f92b47675db5537ffebad1ad273a6913a27e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2209620
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770594}