[Web Payment] PaymentHandlerHost for all payment apps.
Before this patch, PaymentHandlerHost was passed only to
ServiceWorkerPaymentApps, which are based on payment handlers.
This prevented introducing more types of payment apps which may be based
on payment handler, such as a JNI wrapper on top of C++ payment app
interface.

This patch sends the PaymentHandlerHost to all payment apps, so it's up
to the payment app implementation to determine whether the Host should
be used.

After this patch, it's easier to add a JNI wrapper on top of a C++
payment app, regardless of whether that app is backed by a payment
handler.

Design: https://bit.ly/cross-platform-pay-app-factory

Bug: 1083242
Change-Id: Ibf785fd665691a402500705c5b6d39061c904f73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2207290
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Danyao Wang <danyao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770605}