SplitSettingsSync: Fix shelf default apps when user opts out of sync
The function IsSafeToApplyDefaultPinLayout() checks to see if sync has
started so that ChromeLauncherController can decide when to apply the
default pinned apps.

We should only check for app sync and OS pref sync if the user has
enabled OS sync. We didn't need this before SplitSettingsSync because
sync was always on.

Bug: 1084189
Test: added to unit_tests
Change-Id: I02fef532d880cc30fa96d1c177c376f4f8ec9ab7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2209597
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770618}