xvfb: weston: Set size for virtual display
By default, the size of the virtual display that
Weston size is quite small and some browser_tests
(not part of the Wayland tester yet) fail because
of their expectations of windows' vs screen's bounds.

Thus, set the size to be 1920x1080, which adequate
enough and gives plenty of room to manage sizes of
windows.

Bug: 1022750
Change-Id: I4c90b9923431b533020fb63f8a9eb0153823a451
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2207199
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770628}