Remove ichikawa@ from the owners of //tools/licenses.py.
I'm leaving a project which touches Chromium, so I want to remove
myself from the owners. I was the only owner of //tools/licenses.py,
so this means that we need to rely on owners of //tools to review this
file. I hope it's OK since the file receives roughly 1 CL per month,
mostly with a trivial change to add directories to the whitelist which
is part of this file. But let me know if this is not preferred. I will
seek for an alternative owner in that case.

Change-Id: I2067fada5c3d0a629b0ff0065cc5be1c4c3dc347
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2210032
Auto-Submit: Hiroshi Ichikawa <ichikawa@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770636}