Let both minimum font-size settings apply to CSS px units.
Applying the minimum font-size to the device pixels would not give the
expected result, either for using zoom for DSF, or applying page zoom.

I added the documentation for the settings based on what I found on
stackoverflow:

https://stackoverflow.com/questions/52943595/cannot-understand-android-webview-documentation

The behavior changes for both the non-"logical" setting in that we will
still apply a minimum font-size even if the zoomed font-size would have
become larger than the minimum font-size in device-pixels. That is, if
we have a minimum font-size of 20px and two elements, one with a 10px
font-size and the other with a 20px font-size, they will both be
rendered at 40 device pixel size at 200% zoom.

This changes the behavior for the logical setting when we have a page
zoom factor which is lower than 100%, where we also now just adjust the
CSS px font-size before zooming is applied.

The logical setting is only exposed via the Android WebView API or via
a hand-edited config file. The minimum font-size setting the
chrome://settings is the non-"logical" one.

If the settings object is null, return the font-size without minimum
sizes applied instead of the not-so-useful 1px return we had.

Bug: 1082127
Change-Id: Idc03bf37fb7d39206ad4d4dc50eef01b56a69ddb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199134
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770661}