[Media Feeds] Add logging to media feeds converter
This adds logs for high-level problems when converting a media feed.
It explains when a feed is invalid but not necessarily why.

Converts the file to a class to have access to the log.

Will handling printing this result in the JS in a follow-up.

Bug: 1081515
Change-Id: I06e37ae7d9cf32066fc1d13754ec218e4ed29bcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2209596
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Commit-Queue: Sam Bowen <sgbowen@google.com>
Cr-Commit-Position: refs/heads/master@{#770670}