goma_link: add slash to -thinlto-prefix-replace
Prior to LLVM D77223, a slash was automatically appended to the
replacement prefix, so that -thinlto-prefix-replace=;lto.x resulted in
paths of the form "lto.x/...". Since that LLVM revision, this is no
longer true, so goma_link now needs to append it explicitly.  The new
goma_link.py works with LLVM versions before and since the change.

Bug: 1082626
Change-Id: Ia99a580c8b27eb22eabacab60c37eed1d39a3b08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2207587
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770676}