Simplify logic for shouldSwapLogicalResolutionText
The former logic was susceptible to rounding errors when the aspect
ratio of the current display mode did not perfectly match the aspect
ratio of the display's native mode. The former logic was unnecessarily
strict because all we need to determine the order of the dimensions
must be flipped is to check whether the longer edge of the current mode
and native mode are the same edge or not.

Bug: 1085064
Change-Id: I86d4f4fb77098f1a034e0bc62671a7aa1cd62e14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2210008
Commit-Queue: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Auto-Submit: Bailey Berro <baileyberro@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770749}