[Extensions SW] Fix ServiceWorkerTest.WebAccessibleResourcesFetch
For the page that performs fetch, ensure it is controlled by the
service worker before initiating fetch. This is ensured by waiting
for oncontrollerchange event if navigator.serviceWorker.controller
is null. Note that the worker claiming itself and messaging the
page isn't enough, and causes flakiness.

Bug: 1006129
Change-Id: I313bbc79e04bd162c8dbd302b61c9ced446be39f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2207861
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Commit-Queue: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770773}