Migrate FrameMsg_GetSerializedHtmlWithLocalLinks to mojom::Frame
This CL moves this message out of frame_messages.h and adds a new
method in content::mojom::Frame to allow the browser request from
the renderer the serialization of a RenderFrame's contents, regardless
of whether such contents include same-site or cross-site resources.

Since this functionality also requires an asynchronous response from
the renderer once the serialization is complete, this CL does also
migrate FrameHostMsg_SerializedHtmlWithLocalLinksResponse, for which
some extra bit have been added to WebFrameSerializer[Client], so that
we can appropriately reply to the browser process when the serialized
data is available.

In a future CL, this functionality will probably get moved down into
Blink itself, but at the moment there's way too much functionality
in use in Blink's public API, so this CL does not tackle that part
yet, and focused on the migration from legacy IPC to mojo only.

Bug: 1044088
Change-Id: I4ce5b3cfe795e5019bb30a278577ec2be3ee2854
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2207234
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Aaron Colwell <acolwell@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770794}