Do not SetNeedsAnimate() in LayerTreeHost::QueueSwapPromise()
RenderWidget::QueueMessage() used to request a main frame after queuing
a swap promise. This request was moved into the LTH::QueueSwapPromise()
to be able to check whether we are already inside a main frame or not
(see r498671). This means that other calls to QueueSwapPromise() will
have the same side effect which is not necessary. For example, it might
cause reporting of EventLatency metrics when the event has not caused
any actual update, just because a swap promise was queued in the event
handling code. This CL moves the call to SetNeedsAnimate() back out of
QueueSwapPromise(), by calling a new function called
SetNeedsAnimateIfNeeded() which will only sets needs-animate flag if a
main frame is not in progress.

Bug: 1085074
Change-Id: I2fd29cb013f13fad0c86c624b5f57a0d62b9fb2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202529
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Mohsen Izadi <mohsen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770815}