Update and add UMA metrics for assistive features.
- Record coverage metrics only when we show the suggestions
- Change current coverage metrics to match metrics.
- Add disabled metrics

Previously we were recording coverage metrics for personal info in a
wrong place. Now we moved it to the correct place and added a new metric
for all the match scenarios.

Add disabled metrics for cases where we could show suggestion but the
feature is turned off.

For personal info suggester, we only record coverage metrics when we
first show a suggestion so that we won't record dup metrics when the
user keeps typing with suggestion shown.

Test: tested with Chrome on Linux
Bug: 1042084
Change-Id: I6559d81f8d3fc5fa2f9385a4df617f0ac06c5848
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208577
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>
Commit-Queue: Jing Wang <jiwan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770857}