Metal: Defeat caching when testing shader link
There exists a literal in the test shader's binary. Add a list of
valid values that can be inserted at that location in the shader,
and choose one at random when testing shader link time. This
will defeat caching that Metal might be doing (at least for
a while).

Remove DumpWithoutCrashings on failure (we just skip recording
a UMA on failure).

Randomly choose between testing shader compile or shader link
at runtime, so we get distributions for both.

Bug: 974219
Change-Id: I61b067596a69884fcdf967b9774397d0112a6a0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2209309
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770908}