Fix CreateApplicationWindow to create windows on display_id
Launching terminal via shelf context menu, or launcher is not currently
using the same display as the shelf.  The display_id is available
through most of the call chain, but is needed in a few extra places.

CreateApplicationWindow receives display_id in AppLaunchParams, but was
never setting Screen::SetDisplayForNewWindows which is required to
ensure that WindowSizer will put the new window on the specified
display.

Same for WebAppLaunchManager::OpenApplication.

Updated LaunchSystemWebApp for all callers to pass display_id where it
is available.

Bug: 1083825
Change-Id: I5e7046af772b09a3c5340083ded33e42642acd6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208799
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Ben Wells <benwells@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770914}