[breadcrumbs] Limit menu width to 40vw
The drop-down menu hanging off the elider button has a width range
between 192px and 288px depending on the content inside it. Wide menus
can get clipped on the right hand side if the window width is reduced to
the smallest size possible, obscuring any scroll bar in the drop-down.

Changes the max-width selector to take the minimum of 288px or 40vw
which guarantees the right side of the menu doesn't get clipped out.

Bug: 1083046
Change-Id: Ia64dc5befe62e5d4370044fe2545c1c720efee40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2210177
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770930}