Set <body> background color for chrome://media-app.
Currently the color is specified in the manifest and on the
<backlight-app> element loaded in the iframe. But the element only loads
after translations are ready, so there's a period where the background
color is white.

Setting it on <body> in the main frame (and not also in the iframe)
appears to be sufficient to avoid any white flash, since everything
else is transparent.

Bug: 996088
Change-Id: If9dbff68fb07a535113c26133721c5c8b28ea6fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2210337
Reviewed-by: Bugs Nash <bugsnash@chromium.org>
Commit-Queue: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770945}