WebApp: Add WebAppInstallFinalizer is started check.
Return runtime error if finalizer is not started.

Unfortunately, WebAppInstallFinalizer::FinalizeInstall is
used directly by PendingAppManager (for now).

This is a speculative fix for the crash in the bug:
Do not even initiate the writing operation if web app provider
is not yet ready.
In next CLs: Add more started/shutdown checks in
WebAppInstallManager, ManifestUpdateManager. We should return
WebAppProviderNotReady error as early as possible (in
WebAppInstallManager API).

Bug: 1084939
Change-Id: I3f7a27f587da4d2c8676f9863f94650f7357d5ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208889
Commit-Queue: Alexey Baskakov <loyso@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770958}