desktop-pwa: Introduce a new hide_from_use option for default apps
This CL helps fix two issues:

1. For default apps, `add_to_applications` is set based on
`create_shortcuts`, which according to its documentation controls
whether the app is pinned to the shelf. Currently, no default app sets
`create_shortcuts` to true, so `add_to_applications_menu` is always
false. This wasn't an issue with Extensions because the field was
ignored, but it is an issue with BMO where we use that field to decide
whether the app should show in the app launcher or not. To address
this issue, we could simply set `add_to_applications_menu` to true,
but that would be a problem for default apps that don't want to be
shown in the launcher.

2. Some default apps don't want to be shown in the launcher, but
currently there is no way for them to specify this.

To address both issues, we introduce a new `hide_from_user`
field for default apps. The value defaults to false, so it fixes issue
1. above, and can be set to true by apps that don't want to appear to
users, which addresses issue 2.

Also fixes some test files that were missing the user_type field.

Bug: 1084887
Change-Id: Id77a4f50eae3fb4517bca26b76aeb737161df431
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2210179
Commit-Queue: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#770981}