Reset text zoom in between web tests.
This moves the text/page zoom functions from eventSender to testRunner
so that we can reset them from TestRunner in a way consistent with how
we reset properties that can be changed on all frames/widgets/views
now.

We also move the text zoom API from WebView over to WebLocalFrame, where
it can be called for any local root. The only production usage is
Android WebView where it calls it for the main frame. In order to
support site isolation, it would need to do so on other local roots as
well. Since the renderer can't reliably tell what's a local root and is
the one setting/resetting the zoom level, it only allows it on the main
frame for now also.

We add a reset call for the zoom level and the text zoom there, which
have gone missing somehow in the refactoring process. We move other
calls from WebWidgetTestProxy that are resetting things testRunner
modifies over to the TestRunner class as well.

For example, without this, running the following tests in order causes
a failure:

[1/2] paint/invalidation/svg/zoom-coords-viewattr-01-b.svg passed
[2/2] paint/invalidation/svg/absolute-sized-content-with-resources.xhtml failed unexpectedly (text diff, image diff)

R=avi@chromium.org

Bug: 1083790, 866140
Change-Id: Idec021ba7a5d877a02b55b77d280454d6f8fb078
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208543
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771074}