[blink] Split freezing and pausing states
At the moment plumbing around freezing and pausing is highly convoluted
which leads to a few unfortunate consequences, namely:
- visibilitystatechanged events are not dispatched when bringing the
frozen page to the foreground.
- if bfcached page tries to execute JS, the events are silently ignored
instead of triggering eviction.

One factor which makes the plumbing complicated is that there is a
per-frame (instead of per-page) renderer-driven (instead of
browser-driven) freezing codepath — iframe.freezing.

Given that iframe.freeze API hasn't shipped and not under active
development, the suggested fix for the problems above is to unship it,
and reuse the modern plumbing instead when we decide to ship it.

This patch splits pausing and freezing into two independent states in
in Page (so the freezing-pausing-unpausing will still leave the page
frozen) and makes ExecutionContext::IsContextPaused return false
when the page is paused but not frozen.

R=dtapuska@chromium.org
CC=yuzus@chromium.org,haraken@chromium.org
BUG=1018601

Change-Id: Ibc1cb8aa5f2c0b1f221229d0e22855ddeb8e46bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202977
Commit-Queue: Alexander Timin <altimin@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771114}