[Web Payment] Move details update to payment app.
Before this patch, the logic for notifying the payment app of updated
pricing on desktop was in the desktop specific controller, which
prevented sharing this logic across platforms and between different
types of apps.

This patch moves the PaymentHandlerHost.UpdateWith() and
PaymentHandlerHost.OnPaymentDetailsNotUpdated() calls from the desktop
PaymentRequest controller to the ServiceWorkerPaymentApp. The PaymentApp
interface moves to //components/payments/content to enable depending on
mojom::PaymentDetails.

After this patch, it would be easier to make the C++
ServiceWorkerPaymentApp cross-platform.

Design: https://bit.ly/cross-platform-pay-app-factory

Bug: 1083242
Change-Id: Ib02cc981eb8ad955e33e593deb4c57cdebe93962
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208094
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Liquan (Max) Gu <maxlg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771141}