indexeddb: refactor special storage policy
Special storage policy is a dynamic, complicated, browser-only data
structure that IndexedDB needs to query on shutdown to determine whether
or not to delete particular session-only directories (or not delete
protected directories).  Because shutdown is time-sensitive, this
chooses to push updates to policies based on active origins (i.e.
origins that IDBFactories have been created for).

This is the IndexedDB equivalent of this local storage change:
https://chromium-review.googlesource.com/c/chromium/src/+/1945940

Bug: 1015214
Change-Id: I18f65a6416be508b298e8046d4d6c20c2e56b1f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2203146
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771144}