Implement a process snapshot server
We found that a process snapshot was being built multiple
times individually while the task manager is running by
vm and arc code. Since each build of the process snapshot
is full walk of /proc fs, this is wasteful and unnecessary.

This CL introduces the ProcessSnapshotServer to serve
a process snapshot to its vm and arc clients that is
refreshed only once every certain delay that matches
the smallest one requested by all clients.

BUG=chromium:1083062, b/154362057
TEST=Manually, added some tests.

Change-Id: I68ea1cfad6b024dc47276e4681a408f052ec3240
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183473
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771162}