Add skeletons without separators to top domains' skeletons.
For target embedding matching, top domains could be matched when they
are embedded without token separators (e.g. googlecom). In order to be
able to match them, we need to generate skeletons of top domains without
token separators. This change adds these skeletons to skeletons list as
well as adding a flag in |TopDomainEntry| to show if the found match was
with a without-separator skeleton or a regular skeleton.

Change-Id: Id1387ef438f1ae0c73190c0b1cad913f6ad51b4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2205517
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Behnood Momenzadeh <behnoodm@google.com>
Cr-Commit-Position: refs/heads/master@{#771185}