Fixing a crash in the It2Me host after disconnect
This crash is occurring when the user clicks 'stop sharing' on the
remote session.  The website will send a disconnect message which cleans
up the It2Me host and friends.  The website will then create a new
instance of the native message host (which it uses to determine if the
host is installed).  Soon after this, the host website receives a
session-terminate message from the client.  It helpfully sends this to
the It2Me host.  Since signaling is not connected for the new host
instance, it tries to run the incoming_message_callback_ (which is
invalid) and crashes.

This CL adds a check around incoming_message_callback_ to ensure it is
valid before calling it (as well as some logging to indicate the
condition occurred).  It's possible that we could address this issue in
the client w/o requiring a host fix but I think it is a good idea to
handle this scenario in case it is too difficult to fix client-side or a
regression sneaks in later.

(cherry picked from commit e7563c06eb450ca358b6e410b78adb7f55b9c071)

Bug: 1084670
Change-Id: I3714fc860339b12cfcab688652e5e8a83118a4f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2210732
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#770721}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2212558
Reviewed-by: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/branch-heads/4147@{#170}
Cr-Branched-From: 16307825352720ae04d898f37efa5449ad68b606-refs/heads/master@{#768962}