Use Widget's new SetContentsView API for unit tests
For most unit tests, they can use Widget's new API which takes a
unique_ptr instead of a raw pointer. Therefore, the ownership of that
View is clearer.

This CL also change a few non-test cases that are straightforward to use
the new API.

BUG=none

Change-Id: I51b5fb6e767831cfc346cd6f0afc3446210246e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208258
Reviewed-by: Justin DeWitt <dewittj@chromium.org>
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Commit-Queue: Wei Li <weili@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771224}