Record metadata for networks which don't have a NetworkState.
When hidden networks are first configured, they don't have a matching
NetworkState for the newly generated guid, but we still need to record
which user is the owner.  When retrieving the prefs, now check the
default profile for metadata about the guid if none exists on the user
profile, there are a few reasons this could happen now.

Bug: 966270
Change-Id: Id6368fce766907384aa4f41c1f03d458a2113a51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2211743
Commit-Queue: Jon Mann <jonmann@chromium.org>
Reviewed-by: Azeem Arshad <azeemarshad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771225}