[Live Caption] Truncate final_text when it gets long.
The caption bubble controller receives a stream of transcriptions from
the speech service to be set in the caption bubble. For long sessions,
this text could be very long, which would slow down the rendering of
the text in the caption bubble label. This CL removes characters from
that string.

We don't want to remove some arbitrary N characters, because then
the user will be able to see the string re-wrap. We want to remove
characters up to a line break.

Suppose CaptionBubbleControllerViews::OnTranscription receives the
string abcdefghijklmnopqrstuvwxyz. And suppose the caption bubble only
shows 9 letters per line. The layout manager ensures that only the
last two lines (starting at q) are visible.
abcdefgh
ijklmnop
qrstuvwx
yz

In order for the user to not notice any line wrapping, we need to
remove the characters a-p (index 0-17). To get the right number of
characters, CaptionBubbleControllerViews gets the location of the line
breaks from RenderText and cuts the stored text string up to the index
where there is a line break that is not visible to the user.

Bug: 1055150
Change-Id: Ifb59f5bcba39d0ce3de097a9f86bfbb11cbf16db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2187137
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Reviewed-by: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771254}