DistillerRenderFrameObserver determines if page is distilled from URL.
The DistillerRenderFrameObserver::SetIsDistillerPage was sometimes
being called after DistillerRenderFrameObserver::DidCreateScriptContext,
which caused the Distiller Javascript object never to be created
and injected into the page. SetIsDistillerPage happened via a mojom
call from the Browser process when
WebContentsObserver::DidFinishNavigation occurred. It wasn't guaranteed
to happen in any particular order. This was making a test flaky;
relanding the test will happen in a separate change.

It is cleaner to simply check if the page is distilled by checking that
the URL is correctly formatted as a distilled page.

AX-Relnotes: n/a
Bug: 1016615, 1075439
Change-Id: I12cb5bb9e2975aca25a79333f251d1a28546f32b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2209281
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771264}