Add volatile support
We want to qualify everything with const and volatile "just in case", so
that everything compiles if the underlying type T is also qualified.
In practice it doesn't matter, because the qualifiers are dropped when
casting to uintptr_t, and brought back when casting to T only if T has
those qualifiers.

Bug: 1073933
Change-Id: Ic054182924ce43f8e753390c0cbc7c225991eced
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204001
Commit-Queue: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Auto-Submit: Bartek Nowierski <bartekn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771279}