[Omnibox] Copy to clipboard work around for Android O
On Android, apps can give read permission to system clipboard to share
images between apps. But on Android O, this is not working. Apps need to
give permission to individual app to share image through clipboard.

This CL is a work around for Android O. When Chrome shares an image
through the system clipboard, it grants permission to all of the
installed apps. The permission is revoked when the clipboard is updated.

Did not add tests because this is for granting permission between apps,
our test infrastructure is only running in the same app.

Bug: 1081987
Change-Id: Ie1a2226cbb086872683dbfa71d58fb29b69fc563
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2203233
Commit-Queue: Gang Wu <gangwu@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Adrian Taylor <adetaylor@google.com>
Reviewed-by: Filip Gorski <fgorski@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771289}