[XProto] Add serialization/deserialization for a few special types
These are needed for [1].  Normally, events are only read, but
there is a special case for client messages where we want to write
an event that will be sent directly to another client.

Also the connection setup data doesn't have a corresponding reply,
so a special reader needs to be added for this too.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2212976

BUG=1066670
R=msisov
CC=​sky

Change-Id: Ibded85ee3451a891c41ada62f32167bb779ed924
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2209961
Auto-Submit: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#771310}