Add contents of popup note associated with a highlight in PDF plugin
This CL populates the data of a popup note associated with a highlight.
The populated data is sent from plugin to mimehandler. The pepper
interfaces are already structured to pass along the popup note text
data. The CL also modifies the structures related to highlight in the
plugin layer to store the note text, which is then passed on to the
pepper layer while sending highlight related data.

Unit test has been modified to validate the popup note text value.

Bug: 1008775
Change-Id: I0e83b7d04d23922daa20ccd3e5c8e602c6ac5077
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208318
Commit-Queue: Ankit Kumar 🌪️ <ankk@microsoft.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#771323}