[css-grid] Modify use counter kGridRowTrackPercentIndefiniteHeight
The use counter was counting all the cases in which we have
a percentage row track in a grid container with indefinite height.
However some of these cases render the same independently if
we resolve the percentages against the intrinsic grid container height
(like it happens right now for grid layout row gutters and tracks)
or not (like it happens for flexbox row gutters).

This patch modifies the use counter so we count only the cases
that will change the output depending on how we resolve the percentages.
So the cases in which we have a single row with 100% percent
are not counted anymore.

This is similar to what we do for kGridRowGapPercentIndefinite,
where we only count cases when there are more than 1 row.

This will give us a better understanding of the web compat risk
if we decide to change this behavior in the future.

Change-Id: I07ff2c914b39b2e27084afad39c949a95072ce19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2212128
Reviewed-by: Oriol Brufau <obrufau@igalia.com>
Reviewed-by: Javier Fernandez <jfernandez@igalia.com>
Commit-Queue: Manuel Rego <rego@igalia.com>
Cr-Commit-Position: refs/heads/master@{#771350}