Clean up and remove capturePixelsAsyncThen.
Tests that want to run the document lifecycle are changed to use
updateAllLifecyclePhasesAndCompositeThen().

The test that wants to do printing in the renderer and grab the pixels
is changed to capturePrintingPixelsThen(). This path explicitly prints
the current frame, rather than going into the end-of-test-find-a-frame
(non-OOPIF-compatible) path. This makes it possible to use the API from
any frame.

Remove all the main-frame-attached post/invoke v8 callback helpers in
TestRunnerForSpecificView, as they are no longer used.

R=nasko@chromium.org

Bug: 866140, 1069111, 952399
Change-Id: I8271f093bbc0ad9c68b2c05acc24d2062d090f72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204408
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771363}