[Launcher settings] Improve ranking of settings results.
This CL sets up the MVP version of settings result ranking. We:

 - Show at most two results
 - Which are glued to the top of the launcher
 - But show no results for single-character queries.

To avoid confusion, I've removed the max results constant from the
search controller factory, and replaced it with a 'generic max results'
version. This generic version is necessary for now because a value for
max results is mandatory, but we're moving away from needing it.

Bug: 1068851
Change-Id: I188324413dd120d6f8abbf1b0a8791d4fcaac0d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2214583
Commit-Queue: Tony Yeoman <tby@chromium.org>
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771501}