Receive powerwash after update from update engine
Currently, chrome deals with powerwash after an update (e.g. on channel
downgrade) itself, from computing whether it should happen, to actually
issuing a powerwash which is never executed because the update engine
takes care of it.

After this change and the corresponding changes in the update engine,
chrome will rely on the update engine sending a flag indicating whether
a powerwash will happen after the update.

The system notifications will be addressed in a separate CL because
adapting the code there also requires properly dealing with the
rollback flag.

Depends on https://crrev.com/c/2187672

Bug: 2187671
Change-Id: I2aa7024fb7978c21e6396e366f9876d42a72e873
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2187635
Commit-Queue: Miriam Polzer <mpolzer@google.com>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771503}