[OutOfBlinkSandbox] Compute sandbox flags in RenderFrameHost
Compute the sandbox policy of a document in the RenderFrameHost.

For now, this is simply used to double check the value computed from the
renderer. If this works correctly, the browser process will use it and
stop depending on the renderer process.

Ideally, this patch must land for M84, in order to get useful reports.

This patch is a step in the following direction:
~~
Stop parsing the HTTP CSP sandbox header in blink. Parse it in the
network service CSP parser. This allows the browser process to compute
the Origin of the new document before loading it.
~~
https://docs.google.com/document/d/1PechV73KKMF8leh7uTlyGkR32kD5qILvu3ZeG8TYGfk/edit?usp=sharing

Bug: 1041376
Change-Id: I26d0e3744bc1b5b2d893fd01b6e6ecea35cf47e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2190680
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771551}