[b4p/Settings] Refactor: introduce RemovePasswordBehavior
No behavior is changed. This CL introduces a new Polymer behavior
responsible for removing a password entry in Settings. In a multi
store situation, the behavior would delete all existing versions of the
deduplicated entry. The API introduced in crrev.com/c/2199148 for
removing batches of passwords is now called, so that undo-ing restores
every removed version of the password.

<password-list-item> is the only element currently implementing the new
RemovePasswordBehavior. Future CLs will extend this behavior to allow
choosing which versions to remove, and introduce a new dialog that
implements it.

Bug: 1049141
Change-Id: Id62eee3dcce3e3a78e4451371a90e324b8095f46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2209082
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#771568}