Fixing failing test in Mac by removing expectations
There were some specific web_test expectations that were giving a false
error, while the code was behaving as it should have been behaving in
all the other cases (not mac).

It probably means that at some point a false expectation was uploaded
as rebase, and therefore the rebase was broken, not the actual result.

Bug: 995142
Change-Id: I5bc2cd80a4c75f5daa10e61f55808c5aba1118ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2171596
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Reviewed-by: Yi Xu <yiyix@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771583}