system-web-apps: handle install_duration edge case of 0
A low resolution clock could result in an install_duration of 0. This
trips the DCHECK that |install_duration| is positive.

This CL changes to this DCHECK to test for non-negative, so it won't be
triggered when chrome is running with a low resolution clock.

Fixed: 1085403
Change-Id: I63bf90ee6d287e09e0cbb955568f2ae3c2ffa0f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2212237
Commit-Queue: Jiewei Qian  <qjw@chromium.org>
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771683}