Speculative fix for HotseatShelfAppBrowserTest.EnableChromeVox
When moving accessibility focus to home button, instead of just
tapping the button, also send a kHover accessibility event. The hover
event is generally sent after tap, but with a delay. ChromeVox expects
the kHover event to properly recognize home button as the currently
selected node, and to properly handle Search + R command (if the hover
event is not received in time, the command will fail, and ChromeVox
will report that no elements are focused - just as is the case in the
reported test failures).

BUG=1085875

Change-Id: I89de078ce623ad4b03ed8c6b32465f633dd6a4a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2214606
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771696}