Fix regression in chrome://settings/manageProfile.
When 'Allow Chrome sign-in' is turned off, navigating to
chrome://settings/manageProfile' by typing it in the omnibox leads to an
empty page. The reason for that is a failure assert in
|getEditPersonAssocControl_|. |profile-row| is inside a 'dom-if'
template with the condition |!shouldShowSyncAccountControl_|. In polymer
v3, this function gets called when at least when of its arguments is
defined. In the time where the manage profile page is loaded,
|syncStatus.syncSystemEnabled| is not defined and |profile-row| is null.

This CL ensures that |getEditPersonAssocControl_| is not called before
|syncStatus.syncSystemEnabled| gets defined.

(cherry picked from commit 27189b4fe4bf0fb0e712ebb8a586547375fd2d9a)

Bug: 1083879
Change-Id: Iff6a0f3fc0ebd474c2cbf0022a524358c3f2d0a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2207533
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Commit-Queue: Monica Basta <msalama@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#770506}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2215801
Reviewed-by: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/branch-heads/4147@{#242}
Cr-Branched-From: 16307825352720ae04d898f37efa5449ad68b606-refs/heads/master@{#768962}