[iOS] (Merge of) Speculative fix for out-of-bounds tab grid cell.
This is a speculative fix for the crash in the linked bug, where
-collectionView:cellForItemAtIndexPath: is called with an indexPath
item value that's beyond the bounds of the |items| index in the
GridViewController.

That shouldn't happen, of course, because the number of items in the
section is always the number of items in the |items| array, so my
theory is that this is a race condition under some circumstances when
a tab is closed in the tab grid -- the collection view asks for cells
based on the old tab count, but the |items| array has shrunk already.

My assumption is that the layout that's requested is transient, and
(if it didn't crash), another layout, with the correct item count, would
be done by the collection view shortly afterwards.

To mitigate this crash, this CL (a) adds a DCHECK for this overflow, and
then (b) corrects the index of the cell being configured if it's out of
bounds. This means that instead of crashing in this case, the collection
view will (hopefully briefly) show a duplicate of the last cell. Since
all of the model layer mutations are done by item ID and not by index,
this shouldn't cause further incorrect state.

(cherry picked from commit 71a6738032eeb8e74bfb00a646a41877b2807825)

Bug: 1068136
Change-Id: I18d767816ef6faf2813465d1fe8a0536b843a836
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2172889
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#764603}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2214900
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/branch-heads/4103@{#606}
Cr-Branched-From: 8ad47e8d21f6866e4a37f47d83a860d41debf514-refs/heads/master@{#756066}