[Android] Create Personalized Sync Promo for Bookmark page
This cl creates personalized sync promos for the bookmark page when the
user is signed in but not syncing. This cl also creates a new vector
image and changes the base personalized_signin_promo_view_body.xml
file that will be used to make promos in other places, e.g., Recent
Tabs, Settings etc.
The hooks to show the promo is only added in ReorderBookmarkItemsAdapter
as BookmarkItemsAdapter class is scheduled to be deprecated.

Screenshot: https://crbug.com/1057603#c8

Bug: 1057603
Change-Id: I148395e8ee1e910b117f64b1b7031254c6146043
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173057
Commit-Queue: Tanmoy Mollik <triploblastic@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771754}