[iOS] Fix crash in GetLastCommittedItemInCurrentOrRestoredSession
This CL fixes a crash that happens when we are checking if the last
committed item has the same URL as the current/back/forward item and
then updates its UserAgent.
This fix wasn't fully working. The correct fix is crrev.com/c/2193994.
Removing the code allows to cleanup the code for the correct fix, and
remove the crash.
As it wasn't really fixing the issue in the first place, the regression
isn't an issue.

Fixed: 1085348
Change-Id: I7c45ab5709e70f31edfc5eb3989f2c235ddbbac1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2215846
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Eugene But <eugenebut@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771828}