[synthesized scroll] Fix waiting for synthesized scrolls.
When wheel-events are dispatched to the renderer to perform a scroll, it
can start a compositor-driven animation to perform the scroll. The
animation can continue even after the event-dispatch completes. As a
result, for a synthesized scroll, when the request for
WaitForInputProcessed() returns to the browser, not all of the desired
scrolls may have taken effect. To fix this, introduce a notification API
in LayerTreeHost for when any ongoing scroll animation ends. Use this
API to ensure all scroll-animations drive by the compositor are complete
before returning from a WaitForInputProcessed() request.

BUG=1085891

Change-Id: Ie4d67df69786cd30da7d830fa3f952967f1399be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2213401
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771837}