Fix reported times for batched instrumentation tests
Changes the test runner to add a status report to each test, reporting
the duration in ms of the test.

I've chosen to account for the overhead of starting instrumentation,
etc. by adding the overhead duration to the runtime of the first test
in the batch. This means that one unlucky test in the batch will seem
really slow from looking at reported durations, but means the overall
duration for the batch will be correct.

See bot results for android-marshmallow-arm64-rel,
chrome_public_test_apk,
org.chromium.components.external_intents.ExternalNavigationHandlerTest
to see what this looks like in practice.

Bug: 989569
Change-Id: I93aae5e52ea51bb635bb13ce6b6db6e4e8bb55ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2212491
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771855}