Cleanup early long press activation in GestureDetector
This patch cleans up the early long press activation
that so far can only be caused by a stylus button on
CrOS with clear conditions. It adds the according
checks to the logic to only run it for stylus pointers.
So it should have no functional change.

This helps to add a similar mechanism for touch
pointers on Android in the later patches.

Bug: 1086263
Change-Id: I0a4c5a802505ac22b6186b88d343143472850d7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216310
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Commit-Queue: Navid Zolghadr <nzolghadr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771870}