Fix layer/phase violation of custom scrollbar painting
TL;DR, with this CL:
- We no longer update custom scrollbar part position during paint;
- We avoid unnecessary ScheduleAnimation when painting custom
  scrollbars;
- CustomScrollbar::*Rect() methods no longer have side effects;
- Some FIXMEs are fixed.

Details:

Previously LayoutCustomScrollbarPart::FrameRect() stored different
values at different document lifecycle stages:

- Before paint, its location was always zero, and the size was not the
actual size for some parts, but the minimum size calculated from the
style, not based on layout, though updated in UpdateLayout().
UpdateLayout() was called not during layout (as a scrollbar part is
not in the layout tree), but anytime we wanted the minimum size from
const methods of CustomScrollbar.

- During paint, as the FrameRect() calculated during UpdateLayout was
not the real frame rect, we updated the frame rect to make sure they
would be painted correctly. This caused unnecessary
SetShouldCheckForPaintInvalidation when the FrameRect changed and
unnecessary ScheduleAnimation().

This CL contains the following changes:
- Removes LayoutCustomScrollbarPart::UpdateLayout().
  Adds ComputeThickness() and ComputeLength(), and lets Margin*() get
  style directly, to replace the original UpdateLayout() (which called
  SetSize()/SetMargin*()) + Size()/Margin*() in the const methods of
  CustomScrollbar. This removes the dependency of the const methods
  on the temporary incorrect FrameRect().

- Adds CustomScrollbar::PositionScrollbarParts() to set the correct
  FrameRect() for scrollbar parts. It's called from several places that
  position of scrollbar parts can change. This replaces the Set* called
  from CustomScrollbarTheme::Paint*.

Bug: 856802
Change-Id: I5b120294a31a7f91831441cb361adfd27aefa44e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2167602
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771890}