Replace Document::context_document_ with an ExecutionContext pointer.
Currently, Documents that are never attached (created by
DOMParser, XHR.responseDocument, etc.) have a context_document_, which
is the attached Document that created them. If a Document is attached,
it is its own context Document.

The main use of context_document_ is to get at state that is shared
across all Documents in a window. However, we can't have non-attached
Documents hold a LocalDOMWindow, because there's already a dom_window_
member of Document for the attached window only. Therefore, hold an
ExecutionContext, which is initialized for *all* Documents, in place
of context_document_. For attached Documents, dom_window_ will always
equal execution_context_.

Bug: 1080735
Change-Id: I02d7280f6ceebdb18412dfc14165e4f611d38181
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2118850
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771929}