[Offline indicator v2] Fix text fade in animation in fullscreen
When the status indicator is shown in fullscreen mode, the text wouldn't
fade-in. The animation would run, but the frames wouldn't be presented
to the screen even though they are drawn.

To work around that, this CL adds a #requestLayout() call at the start
of the animation.

Bug: 1083461
Change-Id: If816b2882b77e41a7b00d8b863548e8629aff248
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216281
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Cr-Commit-Position: refs/heads/master@{#771944}