[webhid] Add WebHID permissions to Page Info
Adds the WebHID permission state to Page Info, including information
about devices that the current page has been granted permission to
access. Devices are listed with a button that removes the device
permission. If the guard permission is set to Block, a control is
provided to change it to Ask.

BUG=1030493

Change-Id: I2ca80fd137c9fb6898fa937a06c64567f5ead831
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1918681
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771960}