Disable texture upload through PBO in GLImageMemory on Windows.
First, this path is emulated in ANGLE on top of D3D on Windows and
is super slow, so it should be disabled.

Second, this path still has a bug that it messes up certain GL states
and therefore causes Meet present fullscreen to trigger rendering
issues on Windows.

TEST=manual

Change-Id: Icdb2e506979cc6bc07788b71a1668d393c67fbed
Bug: 1064245,1079184
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216361
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Zhenyao Mo <zmo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771961}