Unfork weblayer ContentView
WebLayer forked ContentView for expediency, and now we are unforking it.
Features that are added to ContentView in components:
* support for using EventOffsetHandler with ContentView
* ability to set and change WebContents later
* setIsObscuredForAccessibility to disable accessibility

Features kept in weblayer:
* autofill

Note also ContentView is now a FrameLayout, which does not matter to
weblayer since weblayer does not add child views to ContentView anymore.

Change-Id: I995aad7a3cd9c4a6d8f819d9fd9080548e14cd75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2213064
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771976}