Exposing textSelectionStart/End in AccessibilityNodeInfo on Android
This CL adds the indices of a user's text selection start and end
positions to the AccessibilityNodeInfo object for that node on
Android.

Until now this information has only been conveyed through the
ToIndex and FromIndex indices of AccessibilityRecord.

As per ARC++ request, we will now communicate the indices in both
the AccessibilityRecord and the AccessibilityNodeInfo object
for the given node.

This CL also cleans up some of the native code implementation of
building the AccessibilityNodeInfo objects. We have refactored the
code to no longer have KitKat specific methods since we have dropped
support for that OS.

Bug: 1059680
AX-Relnotes: N/A
Change-Id: I3ce1629271e976a9ef323097ee0bf86233d44aad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101306
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/master@{#771980}