display: Modify display name error handling
In cases where a single character in a display name is non-printable, we
use an empty string for the display name. This change instead replaces a
non-printable char with a space.

Added test case for a string with non-printable char. Note that the
display_id is generated from the display name, so this is also updated
in the test.

Change-Id: I8092bf10bdf5c93b450a0bf84b18c5aaae157452
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2212701
Commit-Queue: Sasha McIntosh <sashamcintosh@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771996}