[DNR] Unflake DeclarativeNetRequestAPItest.OnRulesMatchedDebug
For one of the subtests, an xhr request which matches a rule was sent
before the event listener for OnRuleMatchedDebug was registered. The fix
involves waiting for a round trip between the extension and browser to
complete after the listener has been added for the request to be sent,
which prevents the race condition causing the flake.

Bug: 1029233
Change-Id: I1e1c011159dfa9c463612b4f837ce9ce4316a234
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2209601
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Commit-Queue: Kelvin Jiang <kelvinjiang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771998}