CrOS: Change core-scheduling prctl value to a placeholder-value
The prctl in upstream kernel keeps changing as new prctls are added.
Numbers 57 and 58 are already taken. Let us choose 0x200 (512) to be
safe.

While the core-scheduling patchset is making great upstream progress,
the prctl(2) interface to use it will take time upstream and we might
have to carry it in product kernels for a while. Let us change it to
0x200 to keep products working. We can always change it to the final one
later.

BUG=b:152605392

Signed-off-by: Joel Fernandes <joelaf@google.com>
Change-Id: I15d7a534cab35678b017991ac67bae54128b43a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2213330
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772005}