Picture-in-Picture: move all logic to the PictureInPictureWindowControllerImpl.
This removes the duality between PictureInPictureServiceImpl and
PictureInPictureWindowControllerImpl that made handling the
PictureInPictureSession object more complicated than it should. Now, the service
is simply a shell passing the requests to the controller which then handles the
current session at the WebContents level.

This CL also introduces a content_browsertest test in order to allow for
integration tests that do not include the //chrome layer. Follow-ups may include
moving some browser_tests to it as they tend to be flaky and they are not
testing the actual UI.

Also adds a check for surface_id being null when the session creation is
requested.

Bug: 1067152, 1069858
FIXES: 1067152, 1069858
Change-Id: I1c389c513cc227e6d0f9d5e447926186600cdc2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2197996
Commit-Queue: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Reviewed-by: François Beaufort <beaufort.francois@gmail.com>
Cr-Commit-Position: refs/heads/master@{#772029}