Revise sharing permissions for DataPack::LoadFromPath.
When loading a .pak file, allow other entities on the machine to open
the file for deleting but not for writing. This is closer to the way an
executable is mapped into memory for process creation. With any luck,
this will reduce startup crashes resulting from an inability to map in
the locale's pak file.

This change also adds DataPack.Load to the Chromium metrics description
files (it has historically been in the Google internal one).

BUG=1076423
R=asvitkine@chromium.org

Change-Id: I3039922d0edab047e48cf03699f93bcb2c2f0673
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2215241
Commit-Queue: Greg Thompson <grt@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772043}