Prevent "WEB_APP" tasks being offered for non-native files on ChromeOS.
The NativeFileSystem API currently only supports "real" files, backed
by inodes. Attempting to open a non-native file results it it not being
found because there are no attempts to "uncrack" virtual file paths.

This CL causes file_manager::file_tasks::FindWebTasks() to early-exit
if any entries in the file selection are non-native files.

A testing chrome app is included that uses the file system provider
API as a basis for providing files that can be used for end-to-end
testing.

Bug: 1079065
Change-Id: I5ed5649eb85dcfb25cdfd4965770ae39c3c56dad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2213652
Commit-Queue: Trent Apted <tapted@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772108}