Add EduCoexistence.FlowResult histogram for Beaker
Send AccountManager.EduCoexistence.FlowResult to record the last
completed step in the flow.
The value is updated when active screen is changed in the flow, and
stored in InlineLoginHandlerDialogChromeOS. When dialog is getting
closed, we send UMA with last set flow result. It allows us to track on
which screens users close the dialog.

See go/cros-beaker-metrics.

Bug: 1082791
Change-Id: I422ea8c751f569b716f79dd3fe85a34c054c7df1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208982
Commit-Queue: Anastasiia N <anastasiian@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Reviewed-by: Kush Sinha <sinhak@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772200}