Spellcheck: Ignore words in script not used by enabled dictionaries
Some languages (Chinese, Japanese, ...) have no Windows or Hunspell
dictionary support, but a user will often have English spellchecking
enabled. In this case, when Windows spellchecking is enabled, all
editable content composed in the unsupported spellcheck
language will be red squiggled as misspelled, since the words aren't
recognized by the Windows spellchecker. This is a pretty ugly
experience and differs from what was seen when there was only Hunspell.

The fix is to check if a word that is flagged as misspelled by the
Windows spellchecker is written in a script associated with any of
the enabled spellchecker dictionaries (either Windows or Hunspell).
If not, ignore the word as a misspelling.

Bug: 1076677
Change-Id: I072a783d84a981e8f5f0316e8957d7ab93979ca8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2210796
Commit-Queue: Bruce Long <brlong@microsoft.com>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#772226}