cros: Fix ExternalPrefLoader tests when SplitSettingsSync is enabled
ExternalPrefLoaderTest.PrefReadInitiatesCorrectly is only relevant
pre-SplitSettingsSync, so disable the feature for that test.

ExternalProviderImplChromeOSTest.PriorityCompleted needs to be updated
to mark OOBE completed and simulate the correct type of priority pref
sync.

This is in preparation to enable SplitSettingsSync by default. See
failing tests in tryjobs from:
https://chromium-review.googlesource.com/c/chromium/src/+/2210623

Bug: 1060289
Change-Id: I3a3807c304334fa5a22df7e102f85ee1c629c312
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2215249
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772263}