Make ExpectCTReporter take and use a NetworkIsolationKey.
The reporting API it calls into will soon take NetworkIsolationKeys, and
we want to be able to DCHECK all requests have non-empty NIKs fairly
soon, so it makes sense to start wiring this up now.

The ExpectCTReporter is not currently passed non-empty
NetworkIsolationKeys by net/. That will be taken care of in followup
CLs.

Bug: 969893
Change-Id: I69c4c1615ac3001ad596fd0c97283a087a90a674
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2212462
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772296}