Add nickname field to local card editor and new instrumentation tests.
The field is flag protected and this is only a UI change, so setting the
nickname won't set the nickname in the database. Will follow up with a CL
that updates it in the database.

Screencast: https://drive.google.com/file/d/1aIMedpnwlWtJ1YKNT0cpRCowuhDOnNRq/view?usp=sharing
(Googlers only)

Bug: 1082013
Change-Id: Ib26a67c858da4388563e4ba241604517a5e7fde8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2213341
Reviewed-by: Jared Saul <jsaul@google.com>
Reviewed-by: sebsg <sebsg@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Siddharth Shah <siashah@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772310}