Fix a bug in undo group operation
http://crrev.com/c/2197417 assumes in TabListMediator, adding a tab to
a target index in TabListModel when the PropertyModel at that index has
the exact same ID indicates that this is a redundant addition. However,
this is not true when we undo a group by multi-selection. When undoing a
group operation, it is possible that we add the moved-out tab to the
index of the group PropertyModel, and the group PropertyModel has the
same ID as the tab.

Since this assumption doesn't hold for tab group, this CL skips the
fix for tab group and adds a regression unit test and e2e test to cover
user journey.

(cherry picked from commit 047c9c43ecca848612be34adc377c310c042f425)

Bug: 1083895
Change-Id: Ia1545a7b137c3089126498860e17f88308506158
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2209810
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#771106}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218495
Reviewed-by: Yue Zhang <yuezhanggg@chromium.org>
Cr-Commit-Position: refs/branch-heads/4147@{#297}
Cr-Branched-From: 16307825352720ae04d898f37efa5449ad68b606-refs/heads/master@{#768962}