Parse jdeps class graphs into package graphs
Added a new class that takes a JavaClassDependencyGraph as input and
groups the classes into their packages, forming a less granular
package-level dependency graph. Unit tests are included.

To allow for this change, functionality was added to the Graph class
to fetch a list of nodes and edges. Unit tests are included for this as
well.

Bug: 1081889
Change-Id: I6ff39a68945d7907afbbf6d58a6cb885fafcaba4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2213408
Commit-Queue: James Long <yjlong@google.com>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Mohamed Heikal <mheikal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772368}