wm: Fix drag to maximize for touches on bezel.
I forget why i put the extra check in the first place, I think did it
in case the displays are stacked vertically and dragging i thought maybe
dragging from the bottom to top display would trigger a maximize. But
we pass the display as well, so that is not a problem. Removing the
extra check allows for touch dragging on bezel.

Test: manual
Fixed: 1081985
Change-Id: I5a58eb7874b109f7403a9fe383f27e5bf9fa87e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218506
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772405}