Add a throbber to the WebUI Tab Counter
See attached bug and inline comments for details. The throbber is used to visually indicate that a background tab has been created, to ensure the user is aware of the activity. It is not meant to exactly represent the loading state of the background tabs.

Bug: 1085560
Change-Id: Id51096620a487eb549fe1395f3199f59fe49a806
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2212812
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Taylor Bergquist <tbergquist@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772422}