Removed use of set_owned_by_client() from FocusRing
Currently FocusRing calls Install() on a view and returns an
unique_ptr that is owned by the client and not the view to which it
is added.

This ends up with client classes holding a unique_ptr to the
FocusRing without any need for independent ownership.

This patch updates the view to own the FocusRing with classes holding
non-owning pointers to the FocusRing if required.

AX-Relnotes: n/a.

Bug: 1044687
Change-Id: Ib3c8a0f47df20249a750dca476ab2f73a992b53a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2209402
Reviewed-by: Justin DeWitt <dewittj@chromium.org>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772423}