Include image data when reporting UKM from canvas readbacks.
This cl computes digests of the underlying bytes of canvas images when
HTMLCanvasElement.toBlob, HTMLCanvasElement.ToDataURL, or
HTMLCanvasElement.convertToBlob are called.

OffscreenCanvas.convertToBlob is not recorded yet, since we do not yet have a
UKM source or Recorder in that context.

CanvasRenderingContext2D.getImageData and WebGLRenderingContext.readPixels are also not recorded yet.

Bug: 973801
Change-Id: Ic2de567b16c98b613c945afa823a078a4740a7a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2207979
Commit-Queue: Chris Fredrickson <cfredric@google.com>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772448}