Add additional info to GPU Gold data
Adds "os", "os_version", and "os_version_detail_string" to the the JSON
data that is uploaded to Gold when an image is compared. This will make
it easier to tell what bot produced an image.

Also adds "driver_version".

Example output is "win", "win10", "10.0.18362", and "26.21.14.3102"
respectively.

Bug: 1086751
Change-Id: I8a722db102808502a8100d700942a8348d2b5c54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218479
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772495}