NetworkService: remove "network-service-in-process" from chrome://flags
Now that the in-process configuration is not supported on desktops,
and default on Android, we do not provide a flag at chrome://flags.

The configuration can be modified through the base::Feature
to run tests for Android on Linux, but it isn't supported configuration
as well.

Bug: 1087170
Change-Id: I5b0d41a8dec783524895d57dfee56ab80c24711d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2215154
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Auto-Submit: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772516}