Add application locale and use that instead of language.
The language that is currently used in loadtimedata has had the region
code stripped out. This means that we can't tell the difference between
en-GB and en-US, or zh-CN and zh-TW. That means they fall back to
english for our app.

Using application locale solves this.

Bug: b/157427039
Change-Id: Id83d420752aa9da0808652693194ecf6bc9b5702
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2215158
Reviewed-by: dstockwell <dstockwell@chromium.org>
Commit-Queue: Rachel Carpenter <carpenterr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772523}