desktop-pwas: Check for nullptr return from CreateWebApp()
This CL adds a null check for the return value of CreateWebApp().
It's valid for this function to return nullptr in the event of DB
corruption so we must check for it.

Bug: 1086775
Change-Id: Idd491a136537540f8a65cd466d65772fbdc66300
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2217990
Commit-Queue: Alan Cutter <alancutter@chromium.org>
Commit-Queue: Alexey Baskakov <loyso@chromium.org>
Auto-Submit: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772531}