[WebSocket] Reassemble small messages
The quota-based WebSocket implementation would never fragment outgoing
messages of 64KB or less as long as quota had time to be updated.

Emulate this behaviour with the new datapipe-based implementation, to
avoid the risk of breaking applications dependent on the old behaviour.

The behaviour is controlled by the feature flag
"WebSocketReassembleShortMessages" which defaults to enabled. This can
be disabled for experiments to determine if the feature is really
needed.

No extra bytes are copied and no extra allocations are done. The main
overhead of the feature is in extra code complexity.

Add a web test, expect-unfragmented.html. This has been tested against
the old quota-based implementation (passes) and with
--disable-feature=WebSocketReassembleShortMessages (fails).

BUG=1086273

Change-Id: I958dd5d99931e3023ee339481f101ffa56ed274e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2215383
Commit-Queue: Adam Rice <ricea@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772575}