[signin] Combine Gaia Reauth with reauth the confirmation dialog
This CL adds Gaia reauth as a second step after the
chrome://signin-reauth dialog.

A new SigninReauthViewController class is created to manage the whole
reauth flow. This class is responsible for:
- opening and closing dialogs
- loading the Gaia reauth in background, while the confirmation dialog
is shown
- presenting the Gaia reauth after the user clicks Next in the
confirmation dialog either in the same dialog or in a new tab
- calling the reauth callback after the flow is completed

Bug: 1045515
Change-Id: I7504c860af600cd0dba37616af09c894ded82d91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2210464
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772643}