RenderFrameDeleted should not remove InputEventObserver
WebContentController is a RenderWidgetHost::InputEventObserver in order
to receive acks and generate gestures. RenderFrameCreated and
RenderFrameDeleted have been used to manage this, deriving the
RenderWidgetHost from the given RenderFrameHost. However, deletion
of a RenderFrame does not actually imply deletion of the associated
RenderWidget, so that should not be cause for unregistering the
InputEventObserver.

This does mean that we don't definitely know when a RenderWidgetHost
goes away, but this is safe because ~WebContentController already
checks its validity, and a deleted RenderWidgetHost certainly won't
be calling any observers.

Bug: b/151325287
Test: manual on device
Change-Id: I5085d048dfd2f4ec240e7035d2c2e9d06e598f54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218764
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Commit-Queue: Kevin Schoedel <kpschoedel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772677}