[Passwords] Show store indicator for manage passwords bubble
Similar to the dropdown, this CL adds a G icon to account-stored
passwords in the Manage Passwords Bubble. Since the store might be
relevant for a11y users when they intend to delete a password, an
accessibility text informs screen-reader users.
Flag-guarding the icon row is not necessary since it is only shown if at
least one password is account-stored. Without the flag enabled, no
password can be account-stored.

Note: the only test for this surface is currently disabled on bots due
to flakiness. The test passes locally but debugging the flakiness
exceeds the scope of this CL.

Screenshot in the linked bug is identical to
IDS_MANAGE_PASSWORDS_ACCOUNT_STORE_ICON_DESCRIPTION:
https://storage.cloud.google.com/chromium-translation-screenshots/fa9d35ac7a358c49def0764514a910c300ccf368

Bug: 1086888
Change-Id: Ica42938001219391d49dd339f3f414f45b109d38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218071
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772687}