[omnibox] Delete the Remove Suggestion feature flag
We've launched this feature and it seems to have "stuck" the landing.

This CL just deletes the feature flag now on ToT.

This CL also deletes the Confirm Delete popup flag for Remove
Suggestion, because we never liked it or seriously experimented with it.
We launched without it.

Finally, it also deletes a browsertest that is out of date. We have a
more comprehensive unit test in OmniboxPopupModelTest now, and this
old test was apparently not working Mac anyways.

In general, we want to get rid of these end-to-end browser tests, which
are flaky and slow, and replace them with integration tests that span
across two to three classes.

I don't think we will have to sacrifice coverage, but even a bit of
coverage loss seems okay to me, as the end-to-end ones have been flaky
for years, and we've never been able to deflake them entirely on all
platforms.

Bug: 1205
Change-Id: I2a3514a15e6de5dbe942ec1fd2a323867133571c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218719
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772753}