Remove old and likely stale comment.
The comment removed in this CL is old and it is unlikely at this time
that FilterURL will be changed to terminate a renderer process. Removing
the comment for now and we can always adjust the code if we end up
changing the behavior.

Bug: 172694, 1015882
Change-Id: I5c0d56144a4af25929d0b43696d6fcf04432d4c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2219043
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772794}