[Passwords] Log Password Manager UKM enums to history table
This change annotates the Password Manager UKMs backed by an enum in
such a way that they are logged to the history table, enabling their
usage in the UKM dashboard. In addition, it performs some minor clean
ups in the PasswordFormMetricsRecorder, such as using scoped enums and
explicitly linking to enums defined in enums.xml.

Bug: 918846
Change-Id: If038400084fbb7ee1b83fb1db0e01934b5677cfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2214893
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772803}