[css-flexbox] Tables as flex items should obey the flex container sizing
For most of the boxes, "width:auto" means use all the available space from your container in the inline direction. This means that a flex container does not need to do anything in particular to stretch them in the inline axis. However that is not true for tables because their width mostly depend on the sum of the sizes of their columns (whichever algorithm is used).

The current code delegates that computation in LayoutBox generic code. However that code does not properly handle all the peculiarities of tables. In this CL we check whether or not it has an override for the content logical width which is the way flexbox uses to stretch flex items (and
use that override width).

Bug: 782948
Change-Id: Ia8725ab89414c084803aa8ac973594fcc12cf8b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2214895
Reviewed-by: Javier Fernandez <jfernandez@igalia.com>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: Aleks Totic <atotic@chromium.org>
Commit-Queue: Sergio Villar <svillar@igalia.com>
Cr-Commit-Position: refs/heads/master@{#772823}