Removes CookieList parameter from NetworkDelegate::CanGetCookies
This parameter is not used by any of the implementations of
NetworkDelegateImpl. The method is used to check for cookies
blocked according to user preferences, which depends on the
request only, and not the content of the cookies themselves,
so there is no need to pass it the CookieList.

Bug: 1080782
Change-Id: I5cd9831823d6425adeaac7a61de4c7814bb18f43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2212123
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772838}