Notify ash when locale changes during user login
Notifies LocaleUpdateController observers in ash if the locale changes
when a user logs in, even if the preferred user locale has remained
unchanged (which would be the case if the user has different preferred
locale than the device owner).

Splits LocaleUpdateController::OnLocaleChanged into two methods
*   OnLocaleChanged, which only notifies observers that the locale has
    changed, and
*   ConfirmLocaleChange, which also shows a notification to the user
    giving them an option to revert or accept the locale change.

The OnLocaleChanged was used in both cases, changing the behavior
depending on the current session state. With this change, the intent
will be more apparent.

BUG=1076321, 1076318, 1076310, 1076302, 1076301,
BUG=1077147

Change-Id: Id7e490564f7c816df458b5339f039b6a42072e65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216850
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772850}