Fix no-startup-window to not open browsers in session restore.
Removes check for process_startup when deciding not to open browser
windows when no-startup-window is specified on the command line.

Adds support for restoring last opened profiles to
ProcessCommandLineAlreadyRunning, if no browser windows are open.

Originally submitted as crrev.com/c/2134594, which was reverted
because of test failure on debug Mac bot.

Bug: 850919
Change-Id: I0597729b17e853f9d685a78822c144be1337d180
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2144931
Commit-Queue: David Bienvenu <davidbienvenu@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772852}