media: Refine CDM process display name
Currently we use the site URL directly as the CDM process name. In cases
where the iframe is hosted by an extension, the site URL will be
chrome-extension://<extension_id>/#<original_site_url>, which is harder
to read.

This CL adds ContentBrowserClient::GetSiteDisplayName() to get a site
display name. When the site URL is prefixed with "chrome-extension://",
the extension name will be used.

Bug: b/157373695
Change-Id: Ic774415703a1049f78c6851e3d247849bf341a42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218698
Reviewed-by: Aaron Colwell <acolwell@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772923}