Add ServiceWorkerContextObserver::OnControlleeNavigationCommitted
This CL adds the ability for observers to know when a navigation
for a window controllee commits to a RenderFrameHost.

This is needed in the performance manager so that it is possible
to connect service worker nodes to the frames it controls. The
current OnControlleeAdded() notification does not include which
render frame host this navigation is associated to because it is
not known at the time OnControlleeAdded() is invoked.

Bug: 993029
Change-Id: I91032a3ad606aceb98ae346e5f76bfec01cd4cfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2212808
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772955}