EventGenerator for autotest support nested loops
EventGenerator uses WindowEventDispatcherObserver to throttle
the mouse events. This does not work with nested loops in window
moving/resizing or drag-n-drop. The mouse move event that triggers
the nest loop does not finish until the nest loop quits. However,
without further mouse events, the operation stalls and the nested
loop would not quit.

This CL revises the throttling part. It assumes that hitting the
message pump implies the processing of the current event finishes
and schedules the next event when that happens.

Bug: 1032766
Change-Id: Ib32f06aa6d93b36372ea2b21d941e04293dbde2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220655
Reviewed-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772962}