Enable chrome_launcher_controller_unittest.cc with BMO.
Call TestWebAppProvider::Start() in the tests to stop the App Service
waiting for web apps to be published. Since one subclass test suite
creates multiple profiles, we have to use a virtual method to call
Start(), and override it to do nothing in that test suite so it can
independently call Start() for each new profile. Additionally, if
TestWebAppProvider::Start() is called with no extensions system active
and BMO inactive, a call to CountUserInstalledBookmarkApps() is made
which crashes. Thus, call TestWebAppProvider::Start() only when BMO is
active.

BUG=1082875

Change-Id: Ib1489d7f7dece702f1ef603bfd8c0e2e71ff1a76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216100
Commit-Queue: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772998}