WebXR - blink - use mojo enum and union type directly
Instead of creating a distinct enum and union type blink-side, use the
mojo types directly.

Changes:
- blink::XRReferenceSpace::Type is equivalent to mojom's
XRReferenceSpaceCategory and was removed
- mojom's XRReferenceSpaceCategory was reordered to match the blink's
enum values since reference space type enum was used in UMA histogram
- blink::XRNativeOriginInformation is roughly equivalent to mojom's
XRNativeOriginInformation and was converted into a static class
(it is impossible to create instances of
blink::XRNativeOriginInformation, but the class was still kept for
namespacing purposes & provides factory methods for mojo type)
- the factory methods exposed via blink::XRNativeOriginInformation no
longer return base::Optional<...> since they are guaranteed to succeed
(but the XRSpace::NativeOrigin() still returns optionals to allow for
the situation where an XRSpace does not have a native origin known to
the device side)
- XRSpace::NativeOrigin() is now purely virtual w/o implementation

There should be no functional change introduced by this CL.

As a follow-up, mojom's XRReferenceSpaceCategory enum will be renamed
into XRReferenceSpaceType - it was originally named differently to avoid
confusion with blink::XRReferenceSpace::Type, which this CL removes.

Bug: 1071231
Change-Id: Icb2d713b60a99395a8943556acc62e3208fb3702
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2203277
Commit-Queue: Piotr Bialecki <bialpio@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773016}