Stop feedback panel text clipping
When text is displayed on a feedback panel, it sits inside a child span
with a fixed width. In the case when its parent container has its width
reduced to fit wide buttons, the parent can end up smaller than the
span, causing the end to be clipped.

Set a minimum width for the parent container to 'fit-content' so it
won't shrink and clip its child content.

Changes the name of the growable <div> that's used to help the flex
layout look correct and to make its meaning a bit more obvious.

Bug: 1087210
Change-Id: Iae9f6a7d46503c81e8cfd9bfbbe1389ba828ed2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2217997
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773029}