[Fetch] Prevent cloning body with no execution context
Attempting to Tee() a BodyStreamBuffer object when the execution context
was destroyed would crash the render process. Instead, explicitly throw
an exception.

This makes Request and Response clone() throw when there is a body.
clone() is still permitted when there is no body, which is convenient
for Request.

Bug: 1082688
Change-Id: I2d18f7a13b74caa89a9bd6c56eb9241bb2f5513d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2217753
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773037}