[css-pseudo] Check the marker argument in ListMarker methods
The ListMarker class contains common logic for LayoutNGOutsideListMarker
and LayoutNGInsideListMarker. But it's not a superclass, since multiple
inheritance would be slow. Instead, the layout marker classes have a
ListMarker member. But the ListMarker methods need to reference back
their LayoutObject owner, this could have been done with a pointer, but
it would take more memory. Instead, the ListMarker methods accept a
LayoutObject& parameter, which is supposed to be the marker that owns
that ListMarker instance.

The problem was that there was no check that the LayoutObject argument
was actually the owner of that ListMarker, or even that it was a marker.
So if by mistake some caller passes another LayoutObject, something will
go wrong but it may go unnoticed, or the cause of the misbehavior may
not be obvious.

This patch adds some DCHECKs to address this.

Bug: 457718
Change-Id: I6cd34fceaf763f8d8ce8032c288a8066eaf90a54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218311
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773059}