[gamepad] Keep track of which buttons and axes are used.
Often not all the buttons and axes between 0 and buttons_length /
axes_length are actually used. This change adds some fields to keep
track of which ones are used. This is necessary in some cases to
determine which mapping is appropriate.

Also fixes a bug where special buttons might not be counted towards
the buttons_length on Windows.

BUG=1073130

Change-Id: I9f69b7bc9f68e124b5682588af22a135ed8964a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216510
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Commit-Queue: Andrew Walbran <qwandor@google.com>
Cr-Commit-Position: refs/heads/master@{#773106}