Add LUID to GPUInfo and about:gpu in Windows
When initializing the GPUInfo of each GPU on the computer, also include
the LUID. A LUID is needed to uniquely identify a GPU adapter. The
vendor ID and device ID is not sufficient because device IDs are unique
only relative to the vendor ID, not to each other. If a computer has
two of the same exact GPU, they will have the same vendor ID and
device ID, but different LUIDs.

This change is one part of a series of changes to add multi-GPU support
for VR. This code was originally reviewed as part of a larger CL here:
https://chromium-review.googlesource.com/c/chromium/src/+/2096778

However, that change is being split into smaller pieces to facilitate
code review and landing. The other parts of the split will be put up
within a few days.

Bug: 792657
Change-Id: I4821e2d1317caf335add61b16a9adc7e0d08a41b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216166
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Patrick To <patrto@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#773118}