Translates selection offsets from ignored to unignored and vice versa
Anchor and focus offsets in accessibility selection may refer to children that are
ignored but are included in the accessibility tree.
We need to translate between ignored and unignored offsets at the platform boundary
both when setting as well as when retrieving the selection.

AX-Relnotes: Reading the current selection as well as setting the selection using Jaws and NVDA no longer reads
the wrong selection or fails to set the selection.

R=dmazzoni@chromium.org, aleventhal@chromium.org

Bug: 1082814
Change-Id: I4530f3079eaa282d6db9778f7d6af221345841f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2205400
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773159}