[WebLayer] Set type of navigations from embedder appropriately
In WebLayer, navigations coming from the embedder currently get the
type TYPED | FROM_ADDRESS_BAR. This type has implications on how intent
launching is handled for these navigations: for user-typed navigations,

(1) Intents are blocked from launching without a redirect
(2) Intents to URLs that the app can handle stay within the app even
    if (a) there was a redirect and (b) an external app has registered
    as a default handler for such URLs

Fact 2 in particular is problematic for intended production use cases
of WebLayer. This CL changes the navigation type for navigations coming
from the embedder to be reflective of these use cases: LINK with a user
gesture. Setting the navigation params this way ensures that intents
will be launched in the cases (1) and (2) described above.

Bug: 1087434
Change-Id: Ib43d2ea2cbbfd0352dcde3f5895ac82292b3415f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2219892
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773161}