[b4p/Settings] Keep only one array of passwords in PasswordsSection
In crrev.com/c/2202237, PasswordsSection started owning an array of
passwords deduplicated across stores, multiStoreSavedPasswords. The
original array, savedPasswords was kept around at the time, due to
code that aims to optimize updates to the password list. This code
relies on having unique identifiers for list elements, which can be
easily obtained for the elements of the original array but are not
immediate for the deduplicated one.

This CL removes the array of non-deduplicated entries, while still
keeping the optimized list updates. This is achieved by constructing a
combined id from the ids of each duplicate. The remaining array is named
savedPasswords, since information about deduplication is already present
in the variable type and documentation.

Bug: 1049141
Change-Id: Id91d71d7b074801195ae3de7ef46af204be00463
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220088
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773177}