[Autofill Assistant] Wrong proto field fix.
Before this patch the client and server protos were not in sync. The
reason why this worked is that the proto messages used had the same
fields and the same ids. This patch introduces a proper Result message
that also contains the navigation_ended field used to signal to the
backend that the prompt action was terminated due to a reload or
navigation of the underlying website.

Bug: b/152724265
Change-Id: I8247dddc058fda95d5a1948d5b9e5405eceaaa95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2215055
Commit-Queue: Mathias Carlen <mcarlen@chromium.org>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#773179}