Changed ordering of (de)registering accelerators in Views
Currently all pending accelerators are registered and unregistered
in postorder. This patch changes it so that pending accelerators are
registered in preorder and deregistered in postorder.

This is an issue as accelerators registered later take priority over
accelerators registered earlier.

This change gives child views the chance to take priority over parents
when registering their accelerators rather than having an ancestor
potentially override an accelerator added in the child.

Bug: None
Change-Id: I06fd3db9dccec3bae61c148b2f18753080246117
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2219166
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773198}