Add uma metrics for the permission chip
Added two metrics:
Permissions.Prompt.TimeToDecision to record how long
does it take for the user to make a decision about a permission request
(from the permission requested to user clicking on a button in the
permission bubble view).
Permissions.Chip.TimeToInteraction to record how long does it take for
the user to interact (click) on the permission chip after the permission
was requested.

Bug: 1019129
Change-Id: Id1a8e1e9bb0f3dc51bd5e927509caf23cf59b05a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2214392
Commit-Queue: Olesia Marukhno <olesiamarukhno@google.com>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Bret Sepulveda <bsep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773199}