Trust Tokens: Fix two (!) browser test bugs from a recent CL
https://crrev.com/c/2217855 added browser tests confirming that Trust
Tokens operations can only be executed in secure contexts; these tests
had two bugs.
1. To check if the XHR interface exposed the `setTrustToken` method, the
initial CL used `(new XMLHttpRequest).hasOwnProperty('setTrustToken')`,
but this always returns false [*]; the fix is to use `in` instead.
2. The iframe test case neglected to provide keys necessary for
executing an issuance operation, with the result that the operation
would have failed in a secure context, too. This CL changes the iframe
test to inspect the associated request for the presence of Trust Tokens
parameters. This makes the test slightly more robust because it allows
distinguishing between requests that failed because of the secure
context check (this results in not executing a Trust Tokens operation,
but still sending the request) and requests that passed this check
(these requests will receive TrustTokenParams structs en route to the
network service).

R=csharrison

[*] https://developers.google.com/web/updates/2015/04/DOM-attributes-now-on-the-prototype-chain#using_hasownproperty_on_a_dom_object_instance_will_now_return_false

Bug: 1087200
Test: Verified all three EXPECTs fail in a secure context.
Change-Id: If663e9190321bf5e245f0b042f4442d4de69f10c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2221415
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Auto-Submit: David Van Cleve <davidvc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773215}