Remove NetworkIsolationKey::Todo() in test files.
This is used to track areas where we need to fill in empty
NetworkIsolationKeys with the correct value. While we certainly should
add tests when we do this, keeping Todo()s in tests makes it harder to
audit the production code where we need to wire up NIKs.

Bug: None
Change-Id: I99fcca16a2b2d9a266444512c19bddaba2cc77a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222524
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773241}