ozone/drm: wire and react to new GLSurface ColorSpaces
When there are HDR quads to be presented, BufferQueue calls
Reshape() which in turn calls its GLSurfaces' Resize(). This
CL extends the implementation of GLSurface, GbmSurfaceless,
to catch the ColorSpace and send it to its DrmWindowProxy.

Said Proxy then tells DrmThread which sends the ColorSpace
to the DrmGpuDisplayManager that tells all its DrmDisplays.
If the DrmDisplay supports HDR, configures the linear gamma
ramp to reach full scale (full brightness) or just a fraction
of it, depending on the ColorSpace.

A diagram of the classes and their relationships can be found
in https://i.imgur.com/MqW3Gb8.png (https://imgur.com/a/KSxSln7).

Bug: 958166
Change-Id: I095ac9804bb6499b03f6b1179635827fcff07d06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2210637
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773248}