[ShareButtonInToolbar] add explicit end animation state to explicit show
The share button was getting stuck in a intermediate step where it was
being marked as VISIBLE, but with an alpha of zero and an offset.
This makes explicit the assumption that setting the button to visible
will find it in its default state.

Bug: 1084488
Change-Id: I2ac9797cfdc254c255e31d41bc82a96603bb3f93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218707
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Jeffrey Cohen <jeffreycohen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773252}