media: Add CdmAuxiliaryHelper::GetCdmOrigin()
CdmAdapter uses the CDM origin in crash reporting. Currently the origin
comes from the render process during CDM creation. This CL adds a new
CdmAuxiliaryHelper::GetCdmOrigin() where the implementation is in the
browser process that is more secure. This allows us to remove the
|security_origin| from various CreateCdm() calls which will be done in
the next CL.

Bug: 1065264
Change-Id: I8db09de9586f814c4af0a43e62b0e12eb08e1d41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220214
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773269}