[webhid] Populate physical device ID on Mac
On Mac, the physical device ID is the location ID rendered as a
string. The location ID is a system-supplied device identifier
that describes how the device is connected to the system. HID
devices that are exposed by the same physical device have the
same location ID.

BUG=1011499

Change-Id: I04284c62e22b057f2ccaf256acb03c655043308f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1972329
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773280}