Do not map MatrixID::SMPTE240M to Rec601 YUV space
In ColorSpace::ToSkYUVColorSpace(), do not map MatrixID::SMPTE240M to
kRec601_SkYUVColorSpace. In ITU-T H.273, the matrix coefficients of
SMPTE 240M are KR = 0.212 and KB = 0.087, which are quite different from
the matrix coefficients of BT.470 B, G and SMPTE 240M (KR = 0.299 and
KB = 0.114).

If we just need a close match, then the matrix coefficients of BT.709
(KR = 0.2126 and KB = 0.0722) would be a closer match.

Change-Id: I280df566a031437115f4bbb284ea8603cfb5136d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2188630
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Commit-Queue: Wan-Teh Chang <wtc@google.com>
Cr-Commit-Position: refs/heads/master@{#773287}