blinkpy: Move disabled WPT tests to AndroidWPTNeverFixTests
This CL will make a change that makes expectations for unexpected skips
in the AndroidNeverFixTests file. This CL also changes the Android tags
produced by the test harness to include the Android WPT product
argument.This CL also removes the --android-apk command line argument
adds the --android-product command line argument to the
build_wpt_metadata.py script. The product argument for the wpt
runner will be passed in as the value for this argument.

Bug: 1050760, 1050754
Change-Id: I32a8df293ed1442cb810976ec643e4c63b0414ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204543
Reviewed-by: Robert Ma <robertma@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Cr-Commit-Position: refs/heads/master@{#773302}