Remove custom XR Consent Flow
Starting in M83, WebXr uses the permissions api. As such, the custom
consent flow that it initially shipped with is no longer needed and can
now be removed.

Since the change to permissions is launched via finch for M83 (and not
enabled by default in code until 84), the feature itself (and the finch
configuration) will be left until M86; though it will not change any
behavior. The feature is not exposed in about_flags.

Bug: 1064384
Change-Id: I7fda1e70bc57feb9977bef7bcd07b47cb93f38ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2219161
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773335}