Remove obsolete overlaypanel logic from BottomSheetController
This patch removes logic that would prevent the bottom sheet from
showing when contextual search was visible. The code in question was
relevant when the bottom sheet could trigger without a user action.
Since this is no longer the case, the code is unneeded. This block of
code should have been removed with https://crrev.com/c/1808180, which
contained the rest of this logic.

Bug: 100227
Change-Id: I6c1896c9c8907ec01542568162291e85aa2892cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204298
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773337}