Properly sort the list of sources for browser_tests.
Move the comment, which adds a break to the sources list, and its
associated files into a separate section. As is, the break in the
sources list often confuses people, and that leads them to add files to
the wrong part of the sources list.

Change-Id: I1c969cec45ee678fc2d0c90ab76809d44d98e9f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2219605
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773375}