[AW] Remove HistogramRecordList proto message
MetricsBridgeService interface receives and sends serialized byte data
and writes and reads HistogramRecord proto to file directly without
HistogramRecordList. This makes HistogramRecordList a utility class
to serialize and deserialize byte data which we can easily do without
it.

This CL removes HistogramRecordList proto and replaces it with
a List<byte[]> where the service stores the byte representation of
HistogramRecord objects. This also changes the service AIDL interface to
return a List<byte[]> instead of byte[]. However, AIDL syntax doesn't
support this syntax so the return type is just List.

This also optimize file writes by keeping an open FileOutputStream
throughout the lifetime of the service instead of opening a new one for
each write and then closing it again.

Fixed: 1086678
Test: bin/run_webview_instrumentation_test_apk -f "*MetricsBridgeServiceTest*"
Change-Id: Ic8ccbcf1d714d521c27038c5802c1376e612edea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202276
Commit-Queue: Hazem Ashmawy <hazems@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773400}