[Web Payments][Android] Delegation support for intent based apps [2/3]
This cl introduces IPaymentDetailsUpdateService.aidl and
IPaymentDetailsUpdateServiceCallback.aidl interfaces so that the invoked
native app can call changePaymentMethod, changeShippingOption, and
changeShippingAddress methods to notify the browser about the user
choosing a different payment method, shipping address, or shipping
option.

This cl implements the plumbing for sending the change events to the
merchant, but the plumbing for sending the updated payment details from
the merchant to the invoked app will be landed in a follow up cl.

To test the feature enable "Experimental Web Payments API features" from
chrome:/flags or use the following commandline flag:
enable-features=AndroidAppPaymentUpdateEvents

Bug: 1026667
Change-Id: I799c9dfced15fe2fa916b8ca3634b91366cfc32f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2168009
Commit-Queue: Sahel Sharify <sahel@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Liquan (Max) Gu <maxlg@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773401}