Show that JS content settings are not respected in PDFs with tests
JavaScript content settings in PDFs can be verified by testing whether
the JavaScript in a PDF can successfully "beep." Add tests that are able
to record a "beep" regardless of the content settings. In doing so,
re-enable PDFExtensionJSTest.NoBeep as
PDFExtensionContentSettingJSTest.NoBeep with the incorrect behavior.

Create parameterized tests that are able to enable/disable the feature
flag that controls whether JS content settings are honored in PDFs.
The current behavior of the tests should not be impacted by the
parameter.

Two of the tests try to alter the content settings between the
initialization of two same-origin PDFs.

Bug: 696650
Change-Id: Ie62abc66cc2ba671f250f98bde7aa9b2fb132980
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202972
Commit-Queue: Daniel Hosseinian <dhoss@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773486}