NetworkFetcherWinHTTP must close its File object on a blocking sequence.
NetworkFetcherWinHTTP manages a base::File object as a destination for
network fetches when the DownloadFile is called. This file object must
be closed on a blocking sequence (if the file is opened) when the
fetch_complete_callback is posted and the instance of
NetworkFetcherWinHTTP is eventually destroyed.


Bug: 1088301
Change-Id: I4e0c40fa9093d3ece9961b6c783e08f7002af647
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2224312
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773566}