[Launcher settings] Simplify ranking and deduplication
Much of our search-provider-side logic for ranking and deduplicating
settings results is about to be implemented in the SearchHandler
backend. To avoid doing the work twice, this CL simplifies our ranking.

Note the SearchHandler change (crrev.com/c/2220866) isn't submitted yet,
but this CL blocks some related stuff in the search provider so there's
a case to submit it early.

Bug: 1068851
Change-Id: Ic5d64cd2f478980c1792c7d13740131767e5fc3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222082
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773569}