[WebSocket] Send all data before closing handshake
Changing message sending to use data pipe in
https://chromium-review.googlesource.com/c/chromium/src/+/2206556
introduced the possibility of a race condition between the last chunk of
data in the data pipe becoming readable and the StartClosingHandshake
message arriving. This could result in failing to send the final data
frame to the server before the close frame.

To fix this, check in network::WebSocket::StartClosingHandshake() if
there is still data pending to be sent and postpone the call to
net::WebSocketChannel::StartClosingHandshake() if there is.

The race only appears to manifest on Windows 7 currently, but it could
in principle happen on other platforms.

BUG=1086353

Change-Id: I757525c94dee2f3cb974e69836b9510091efdca9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2214654
Commit-Queue: Adam Rice <ricea@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773633}