cc: Calculate Layer DrawableContentRect with Visible Rects
Layer contributes its entire bounds as drawable content to its target
render surface even when it is not fully visible. Change the drawable
content rect calculation to use layer's visible rect instead of its
bounds. This is to fix a bug where a layer outside of the viewport is
increasing the render target's drawable content rect so much that the
render target is clipped by max texture size, which makes another layer
that should be fully visible clipped.

Bug: 1050748
Change-Id: If1603973c158eee285eaf44bb892b3fc96027378
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2153601
Reviewed-by: Khushal <khushalsagar@chromium.org>
Commit-Queue: weiliangc <weiliangc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773686}