Fix StartupTabPreloader#shouldLoadTab() if TabCreatorManager not ready
This usage of TabCreatorManager#getTabCreator() relies on it returning
a non-ChromeTabCreator TabCreator to know it is not initialized. This
CL changes it to expect an exception instead, which is what it really
does.

An alternative would be adding a method to the TabCreatorManager
interface to check if it is initialized, but no other usages need this.

Another option would be to pass in a Supplier<TabCreatorManager> that
does not provide it until it's ready, but this is a larger change in
ChromeActivity.

Bug: 1085818
Change-Id: I6e9d4bee309825c111c93cb42eae2fc3a24ac752
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222802
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773693}