fido: remove FidoRequestHandler::AuthenticatorState
A timer instantiated in the location where we want to measure request
duration is simpler than using timer for the entire lifetime of the
entire authenticator, and allows us to eliminate the additional type.

It's probably also more precise, because we previously included
GetInfo requests in the timer, which are irrelevant to the thing the
timer is actually supposed to measure.

Change-Id: I0669bcedc4dec517a6dc82922f836335cfc50855
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204551
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773716}