Deflake and re-enable CrElementsInputTest.All.
Previously, when testing the focus() and blur() cases, the  test
was not waiting for both opacity and width transitions to finish,
resulting in transitions from focus() finishing while the test was
making assertions related to blur().

Properly waiting for all transitions to finish before moving on
to the blur() case should fix the issue.

Re-enabling the test on Mac, where it was disabled because of this
reason.

Bug: 997943
Change-Id: I6badb546fba4f093c2fa45490e2a52e5dae84982
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2223082
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773729}