chrome: makes SwipableOverlayView only attach listener when necessary
SwipableOverlayView makes use of GestureListenerManager, which is
currently rather broke. This is because GestureListenerManager needs
to see cc::RenderFrameMetadata on every frame, which it currently isn't.
Making GestureListenerManager get render-frame-metadata on every frame
is quite costly. I'm in the process of making GestureListenerManager
get every frame only when listeners are attached. This patch changes
SwipableOverlayView to only attach a listener when it actually needs
the information.

Patch that makes GestureListenerManager get cc::RenderFrameMetadata is
here: https://chromium-review.googlesource.com/c/chromium/src/+/2222623
(it's a work in progress that I'll be doing right after this).

BUG=1087480
TEST=none

Change-Id: Ib22f1d1ac1caefece5fb1c6059cdd37da72b4618
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225139
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773745}