Some bug fixes for macOS updater
There were a few issues with end-to-end testing macOS updates with the
new updater.

This CL fixes the following issues:

1) When there was a space in the existence checker path, we did not add
the arg correctly. We also weren't passing version to the updater script.

2) Network fetcher was not passing the post data from the server. It was
instead passing the response description which did not have the response
json. Instead we should pass the _downloadedData that we aggregate.

3) Network fetcher was giving status code back to the downloader. When
the status code was 200 (Ok), we were still error-ing out as the expected
value passed to _downloadToFileCompleteCallback when we are successful is
0.

Change-Id: I9e620d4701697403f2e8899cf14c41ce36a653ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220618
Commit-Queue: Michael Chang <donchan@microsoft.com>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773793}