Make non-SUCCESS deep scan uploads fail-open
The issue was caused by DlpTriggeredRulesOK returning false for any
non-SUCCESS status. This likely slipped our attention for a while
because the function's name doesn't suggest it reads the entire verdict.
This CL renames it to DlpVerdictAllowsDataUse so its purpose is clearer
and fixes tests so fail-open is accounted for.

Bug: 1085241
Change-Id: I7a81a3635da4fffbf2b888253083aa9a438fb81f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222551
Reviewed-by: Roger Tawa <rogerta@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773806}