Remove special handling for attaching iframe root graphics layer
The iframe's root graphics layer should not be treated specially; it
should be incorporated into the regular GraphicsLayerTreeBuilder
flow. This has the added benefit that it's not necessary for the
iframe element in the parent document to be composited. A future CL
will stop compositing iframe elements by default.

Change-Id: I9bcfdc2384b072601c87786b269a32d86d7973e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220868
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773820}