TYPE_APP_POPUP uses 'legacy app' features not TYPE_POPUP features
With my recent crrev.com/c/2189893 to create popups from apps as
TYPE_APP_POPUP, rather than TYPE_APP, we are now showing the location
bar on these popups which is not expected.

TYPE_APP_POPUP should match the set of features from legacy apps, rather
than those of popups.  The legacy apps are already 'inheriting' from
popups, with 2 differences:

FEATURE_TITLEBAR:
Popups will show this for non-trusted sites which are not in fullscreen.
Legacy apps show this for     all     sites which are not in fullscreen.

FEATURE_LOCATIONBAR:
Popups will show this for non-trusted sites which are not in fullscreen.
Legacy apps never show this. However they will display the
CustomTabBarView when the URL navigates out of scope of the app.

This CL renames 'LegacyApp...' function to 'AppPopup...', and modifies
the TODO suggestion of having TYPE_WEB_APP and TYPE_LEGACY_APP which
predates the creating of TYPE_APP_POPUP.  It seems that the solution now
would be to create legacy apps using TYPE_APP_POPUP rather than
TYPE_APP.

Bug: 1088300
Change-Id: I77044b62cc5ff88208d02df74af42b59783b78e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2224473
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773824}