Cros: Add background card to apps grid
This is part of the new drag spec in the apps grid.
When the user drags an app, the apps grid enters a new cardified state.
In this state we scale down items to 84% of its size and add a
background card to the apps grid, giving a more modular look to the app
pages.
We define an array of Layers for the background cards so we are able to
manipulate the animations during the cardified state in every step
(start, translation across pages, end).
When we end the cardified state, we also delete the cards.

This functionality is hidden under the app_list_feature
kNewDragSpecInLauncher.

Bug: 996822
Change-Id: Ic8dcca8cfa2f11ceb52634dbf1206321d924752f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2171453
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773854}