media: Ignore GetCdmOrigin() failure
GetCdmOrigin() makes a sync call to the browser during CDM creation in
the CDM service in the utility process. Since CDM creation is an async
process, by the time GetCdmOrigin() is called, the render frame host
hosting that call may already be gone, causing GetCdmOrigin() failure.
In this case we should just ignore the error instead of the CHECK which
would cause a crash in the CDM process.

Bug: 1088352
Change-Id: I0298ed3a0421a0baa9c6d11fc2921f4446fd2d7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225347
Reviewed-by: John Rummell <jrummell@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773897}