Speculative fix for browsertest flakes with conversion feature enabled
A number of flaky timeouts are seen in browsertests, both in test
harness tear down and during test execution when waiting on
asynchronous output from the storage layer.

This CL bumps the task priority for the conversion storage task
runner from BEST_EFFORT to USER_VISIBLE. It also adds an empty test
body to a test with the feature enabled, to verify that the issue
is indeed in test harness setup if this fix does not work.

This feature is not enabled currently, so this should have no
performance impact.

Bug: 1087775,1058199,1080764,1058199,1087406,1084201
Change-Id: I7322f9d8139ab14bccbf8dd71365cfda7514cdcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225739
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773913}