Add the account id for OnAppUpdate.
When a component observes AppRegistryCaches for the multiple users
case (e.g. Ash), OnAppUpdate could be called for multiple
AppRegistryCaches. Add the account id in the AppRegistryCaches and
AppUpdate class, so that the AppRegistryCache observer can know the
current AppUpdate is for which account in the OnAppUpdate interface.

Unit tests are modified to test the account id in the OnAppUpdate
interface.

BUG=1068884

Change-Id: Id07b2e8480330667c45ff11fe1cf21d3ebe048ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222244
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773991}