media/gpu/video_encode_accelerator_unittest: Enable features in TEST_P
We formerly settle enabled features in the test initialization.
However, if base::TestSuite::Initialize() is called, it turns out
that these features are overwritten and thus no effect in the
test. There is no way to stop the overwrite. Therefore, this
workarounds the behavior by re-enabling features in TEST_P.

Bug: None
Test: video_encode_accelerator_unittest on eve
Change-Id: I26ec9afcbb4d88786fd18d4701a6dde3938ecc99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2210186
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774007}