ozone/wayland: Don't mark previous feedbacks as failed if one discarded
Previously, we assumed that if we are waiting for f0, f1 and f2
feedbacks and we get f2, the f0 and f1 are failed (because of
destroyed buffer or anything like that). However, there is one
case we missed - f2 can be discarded by the Wayland server, but
f0 and f1 still can be valid and come later.

Thus, pass discarded bool variable to OnPresentation if the
feedback was discarded.

Bug: 1087839
Change-Id: I1f60c282d4871b7aedbae901f8a79b76727e2a80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218779
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Cr-Commit-Position: refs/heads/master@{#774042}