Fix "blocked-iframe-are-cross-origin.html".
Reported by antoniosartori@, there was an error in the test:
"blocked_iframe-are-cross-origin.html"

In Javascript, lambda capture is done by reference. The reference was
the loop 'variable'. As a result the second test case was run twice.

The first test case couldn't work, because embedded enforcement do not
apply to same-origin iframes.

The patch fixes the test.
TBR=mkwst@chromium.org

Bug: 1041376
Change-Id: Id5f223aa138470cb263eea5b0af9f616a314a374
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218049
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774065}