Accessible image descriptions: fix bug in language computation.
In http://crrev.com/c/2148248 we added support for computing the
preferred language for a request, which isn't necessarily the same
as the user's primary language or the primary language of the web page
being visited.

However, the request key we were expecting to get from the server was
being generated based on the original language tag - while the request
key computed from the server response was based on the computed language.

As a result, if the computed language differed from the page language they
didn't match and getting results would fail.

The fix is just to compute the preferred language before we generate the
request key.

Bug: 1090234, 1070505
AX-Relnotes: N/A
Change-Id: Ia2b4038d9e634c6908478333aae6c00781fba49e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2226300
Auto-Submit: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Commit-Queue: Andrew Moylan <amoylan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774084}