[NTP][Realbox] Use AutocompleteMatch's |transition| for navigation
Autocomplete system provides the correct transition type to use when the
user navigates to a match. This is ui::PAGE_TRANSITION_TYPED for URL
matches and ui::PAGE_TRANSITION_GENERATED for other types which do not
look like a URL.

This allows for URL navigations from the NTP realbox to be correctly
recorded in the history backend and for the most visited tiles to be
updated accordingly.

Bug: 1090004
Change-Id: I1af862bfe8b926ccc729dc340fe8384e5759101b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2224020
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774191}