Wayland: linux-explict-sync on with drm-atomic
The linux-explicit-sync Wayland interface allows to exchange two
different kind of fences: release and acquire.
The release fence comes from DrmDevice, but only when drm-atomic is
available. Similar, when drm-atomic is present, the Acquire fence can be
used in DrmAtomic. And technically, the Acquire-fence can be
waited-on outside DrmDevice, but practice shows that when this happens
and if drm-atomic is not present, some flickers migtht appear.

So, this CL disables the Wayland linux-explicit-sync interface when
drm-atomic is not present.

This CL should only land *AFTER* these two CLs from ARC are uprevved:
* ARC-P: http://ag/11624164
* ARC-R: http://ag/11661407

      Tested on ARC Pi and ARC Rvc with the CLs mentioned above.

Bug: chromium:1083569
Test: Manually disabled linux-explicit-sync on Chrome side.
Change-Id: Iff78eebc5567f96cc9d775d9fce3f851104b04f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2223090
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>
Auto-Submit: Ricardo Quesada <ricardoq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774220}