Avoid accumulated error in TransformNode::to_parent for snapping
We snap scroll offsets etc. to whole pixels. When we update a
cc::TransformNode which has snapping enabled, we do the following:
  to_parent.Translate(-old_snap_amount); (in TransformTree::UndoSnapping)
  ...
  to_parent.Translate(new_snap_amount); (in TransformTree::UpdateSnapping)
In most cases (when there is no non-translation transform to screen),
we want to_parent to be integral after the snapping, but the above
operation can cause floating-point errors causing a non-integral scroll
offset. Also repeated operations accumulate errors.

To avoid the problems, add
gfx::Transform::IsApproximatelyIdentityOrIntegerTranslation(),
and use it in TransformTree::UpdateSnapping to check for error, and
reset the error.

Bug: 1070802
Change-Id: Ic05ab6fa8393aaa5446d6892a85090d36c78b800
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225463
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774227}