Record when a network has a known bad-password.
This saves additional metadata when a wi-fi network has a bad password.
Once a network has had a successful connection this will be cleared and
further bad-password errors will be ignored unless the saved password
is changed.  This is accessible from a new public method in order to be
used to avoid syncing networks with a bad password.

Bug: 966270
Change-Id: I084bc46da7302e93644484bf34cebf39ec8b1de2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220658
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: James Vecore <vecore@google.com>
Commit-Queue: Jon Mann <jonmann@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774250}