Fix shutdown crash due to COM re-entrancy from UIA call
At window close, we make an outbound call to UI Automation to signal it
that objects associated with that window are no longer valid and can be
cleaned up. If that cleanup call happens to be the first call into UIA
during the lifetime of the browser process - because UIA has not
otherwise been used - it can trigger UIA to initialize itself, which
can result in an outbound COM call from the browser main thread,
which can trigger pumping of messages back into our window. That can
lead to bad behavior since objects are already partially torn down at
that point.

The fix is for each message handler to keep track of whether it has
returned any UIA objects and only make the cleanup call if so.

Bug: 1087553
Change-Id: Ibf55df4d2e531fb8c4e404ede80e17833ff61543
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220473
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774264}