[omnibox] Add accessibility messages for popup headers
This CL adds the accessibility messages for popup headers.

It also moves the logic out of OmniboxResultView into OmniboxPopupModel.
We had to do this, since we now have headers that are not encapsulated
within OmniboxResultView.

Moreover, we've had crashes with one View reaching into other Views.
Our new design going forward is to have Views be relatively independent
while coordinating shared business logic though the OmniboxPopupModel.

One more thing: This CL will also make VoiceOver enunciate the
accessibility labels for all focused-button SelectionStates, not just
BUTTON_FOCUSED. We need this because we now have a sundry of focused
button states.

This CL is not a complete solution, we still have a few more things
to do, that will have to be in separate CLs:
 - Don't update the temporary text when tabbing INTO a Header.
 - Update the temporary text when tabbing OUT of a Header.
 - Provide the proper 'Selected' accessibility event when focusing
   the header button. We don't do this yet.

Bug: 1078183, 1052522
Change-Id: I94108ca5a527754c0087f9a06f1aeb5d53dc57e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220348
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774269}