blink/bindings: Fix DCHECK() with DiskDataAllocator.
With some configurations, DiskDataAllocator triggers DCHECK()s:
- Serializing an empty file is supported by Mojo, but discouraged
- Background writes need to be explicitly annotated in blink

This also adds a threading restriction exception for reads on the main
thread. This wasn't hit because a write always happen before a read, but
is required as well. This is expected, and is covered by metrics to
ensure this is not an issue in practice.

Bug: 1087862, 1029320
Change-Id: I41f5fca2b9b371a6ad7b9988ba3914543bd44d6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222580
Commit-Queue: Benoit L <lizeb@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774278}