[Web Payment][Desktop] Reduce dependency on Views.
Before this patch, the service worker payment app factory in C++
depended on the payment request delegate interface, whose concrete
implementation depended on the desktop Views framework, which should not
be used on Android.

This patch moves the references to the delegate from the service worker
payment app factory into the payment request state object, which is not
used on Android at this time.

After this patch, it's possible to use the service worker payment app
factory on Android.

Design: https://bit.ly/cross-platform-pay-app-factory

Bug: 1022512
Change-Id: I649673ffdfe419708d796249c68998080efaa00f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2217849
Commit-Queue: Danyao Wang <danyao@chromium.org>
Reviewed-by: Danyao Wang <danyao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774300}