[Instant Start] StartSurface will hide after navigating from the Omnibox
Currently, typing an URL in the fake Omnibox doesn't trigger the Start
Surface to hide and navigate to a new Tab when instant start is
enabled. This is because the TabModelObserver is registered before the
Tab State is initialized. Thus, the observer is registered to an empty
TabModel, rather than the correct one. This CL caches the pending
request, and registers the observer after native/Tab State is
initialized.

Bug: 1086608
Change-Id: I734cd9b3622f37cf672ba54a639ab8f7085ce0f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218468
Commit-Queue: Xi Han <hanxi@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774305}