Do not send OnLayerTransformed when the transform doesn't change
On some occasions, ui::Layer gets a transform which is identical
to the original. Setting and handling transform causes some
tasks, and that doesn't make much sense to do so when the
transform doesn't change actually.

It seems a few of unittests expect to get invoked, so this CL
also changes those expectations. This wouldn't change the actual
system behavior.

Bug: 1082562
Test: compositor_unittests, aura_unittests, views_unittests
Change-Id: Ia83b3b2ba87aa9825da6e2862aee3db9c665c72b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225227
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774331}