Include portals in the frame count limit
With this CL, we count the portal's web contents as a single frame in
its parent, ensuring that we do not exceed NodeRareData's connected
frame count bits. However, we do not yet include frames within the
portal in the total.

If limit is exceeded, HTMLPortalElement::InsertedInto will early out.
Successful insertion and removal of portal elements will impact the
subframe count (and, naturally, the consistency check on subframe count
must also account for portals as well as frames).

In order for the test to run in a reasonable amount of time, I've
also made it possible to reduce the frame limit for testing.

Bug: 970038
Change-Id: I64d458b640c193c96662b0dd1e2101b6438e2e38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2187200
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Ian Vollick <vollick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774358}