Remove BlobMemoryController intervention to critical memory pressure
Under critical memory pressure the system is probably already swapping
out memory and making heavy use of IO. Adding to that is not desirable.
Furthermore, scheduling a task to write files to disk risks paging-in
memory that was already committed to disk which compounds the problem.

Bug: 1087530
Change-Id: I2f5ba6e703eef0d804e6f4d99dd504c7eadf7740
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2227029
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Sébastien Marchand <sebmarchand@chromium.org>
Commit-Queue: Oliver Li <olivierli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774376}