Inject FakeArcWindowDelegate to the constructor of ArcImeService.
ArcImeService provides the method, SetArcWindowDelegateForTesting() to
inject a fake instance of ArcWindowDelegate for testing.
However, the instance of ArcWindowDelegate was generated in the
constructor and used in the constructor, so that usage cannot be faked.
This CL adds a new constructor for testing to ArcImeService to fix it.

Bug: b:148193316
Test: component_unittests
Change-Id: Ia3b6c68bbc66fd4d5bc640ecdabbb707143cfff6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2224943
Commit-Queue: Yuichiro Hanada <yhanada@chromium.org>
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774433}