Setup Conversions database in memory for browsertests
This is a speculative fix for Conversions browsertest flakes caused
by the database initialization task presumably hanging. This is
a followup of https://crrev.com/c/2225739, as tests are still flaking.

Sets up the Conversions storage in memory rather than on disk for
browsertests. If the database initialization task is just slow and
running into timeouts, running in memory should vastly speed it up.

Bug: 1087775,1058199,1080764,1058199,1087406,1084201
Change-Id: I740acb32ce835553b29709bd854944eee7da5122
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225286
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774466}