[FragmentItem] Re-enable position cache on Android
This patch re-enables position cache on Android when
|NGFragmentItem| is enabled.

It was disabled by r660251 crrev.com/c/1614133 to reduce
memory usage on Android, but hinders the layout performance
in some cases.

As |NGFragmentItem| saves memory on text-heavy pages, we
should have a room to re-enable the position cache.

Bug: 982194, 1042604
Change-Id: Ib4b0543f181024f0b825d4a378e0a3aa92109368
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2226579
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774479}