Make providers report their type on update, and add hook to controller
This is CL #2 for integration testing. It uses the new ResultType method
to allow search providers to report _which_ provider is returning
results on an update to the SearchController.

The SearchController has also had a callback hook added, which is passed
the updated result type on each call. This hook can be used by a browser
test to control a run loop that stops when the desired set of providers
have returned.

See the CL description in crrev.com/c/2214736 for more details on the
implementation plan for integration tests.

Bug: 1086713
Change-Id: Iedd8e7ce6824dc6c90623837342456db26494069
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2217945
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774536}