[geolocation] Hold queries to the mock until position is set
This is an alternate to crrev.com/c/2216317 to fix test flake in
geolocation-api/watchPosition-page-visibility.html now that page
visibility changes are asynchronous.

The original mock required a call to setGeolocationPosition() for every
call by Blink to queryNextPosition(). This is a difficult invariant to
maintain because Blink will always call queryNextPosition() after
calling the success callback. Most tests avoided the error by marking
the test passing during the watch position callback and so the error
generated by the next "tick" was ignored.

The page visibility test is sensitive to this issue because the call to
queryNextPosition() after state 2 races with the change in visibility.
If it wins then there is no position available and the test fails.

This version of the mock allows queryNextPosition() and
setGeolocationPosition() to be called in any order, matching up the
request from Blink to the position update when it is available. In this
case if queryNextPosition() is called before the visibility state change
then it will wait for the position to be updated after the page is
hidden. This position update will still be ignored by Blink because it
has, at that point, already closed the Mojo connection to the
geolocation service.

Bug: 1083585, 866140
Change-Id: Ida2469738b19d0862c7af3870360c755c71f7e78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2219148
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774664}