Fix the crash issue when the browser is removed.
Sometimes the browser is removed first, and the tab close callback
OnTabClosing is not called (e.g. launch media apps in browser, then
shutdown), so the tab windows are not removed from
browser_window_to_tab_window_.

Modify that code, when the browser is removed, go through the list of
browser_window_to_tab_window_ to remove the tab windows in the browser.

BUG=1090615

Change-Id: I83346b648110fd4f831661cfd168a69676cb39e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2228018
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774675}