Add RespectMacLCDTextSetting field trial
Based on early results in dev/canary[1], this feature is looking
promising and we would like to continue this experiment in Beta. To
unblock that we are enabling this experiment for the perf bots now to
get more data on the impact.

[1] Canary/dev UMA data:
https://uma.googleplex.com/p/chrome/variations/?sid=1aeb80b2fff1543c20033559d40b2917

Bug: 1079418
Change-Id: I09573cb89e969970bd172c82ffcf83cbbecc6791
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2228999
Auto-Submit: Philip Rogers <pdr@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774686}