Windows Spellcheck: Add flag to delay spellcheck service initialization
As a step towards improving spellcheck initialization performance,
with the ultimate goal of not loading the Windows spellcheck platform
DLLs until needed (e.g. when actually performing a spellcheck), this
CL adds a new feature flag allowing the initialization of the
spellcheck service on browser startup to be prevented. This should
allow addressing the rest of the performance issue in stages.

Since Chromium already by default initializes the spellcheck service
on startup for custom dictionary synchronization, the command line
for launching the browser with Windows hybrid spellchecking enabled
but no initialization of the spellcheck service is:

chrome --enable-features=
WinUseBrowserSpellChecker,WinDelaySpellcheckServiceInit
--disable-sync-types="Dictionary"

Downstream Edge has Windows spellcheck enabled by default and does
not perform a custom dictionary sync, so the corresponding command
line for Edge is:

msEdge --enable-features=WinDelaySpellcheckServiceInit

Bug: 1064351
Change-Id: I91559193918cce6895133a9e9c440513b55fdd65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2227450
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Commit-Queue: Bruce Long <brlong@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#774710}