[SuperSize] Add single Container support.
TriChrome consists of {Library, Chrome, WebView}. Each of these units
can be an input to SuperSize. To get SuperSize to aggregate these into
one .size file with a common pool of symbols, we'd like to to abstract
these units as "containers", represented by the Container class, which
relates to other data structure by the following:
* A SizeInfo has 1 or more Containers.
* A Container consists of {metadata, section_sizes} (formerly in
  SizeInfo), and has a distinct name.
* Each Symbol belongs to a Container.

This CL adds the Container class to SuperSize, and rewire former usage
of {metadata, section_sizes} to go through it. Only the *single*
Container case (with name='') is handled, to simplify review and
commits.

The format of .size files is kept invariant. And as seen in .golden
file, very few changes to SuperSize-console output. A number of TODOs
are added to indicate upcoming code changes for multi-Container support.
Other changes:
* describe.py: Use itertools.chain.from_iterable() so that chained
  generators can be appended into array, instead of needing distinct
  local variables.
* diff.py: Add _DiffObj() to generalize diffing.

Bug: 1040645
Change-Id: I799c960d0510157f55876016726aa2eda25d55a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2228169
Commit-Queue: Samuel Huang <huangs@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774728}