Make SkiaRenderer/Dawn/D3D12 run within GPU sandbox
D3D12 requires that we use flip model swap chains. Flip swap chains
require that the swap chain be connected with DWM. DWM requires that
the rendering windows are owned by the process that's currently doing
the rendering.

To workaround this problem, we use gl::ChildWindowWin to create a window
that is reparented by the browser to be a child of its window.

Bug: 1071458
Change-Id: I212df12ec21f3fbc00e08b63a741b05b14771c04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2223371
Commit-Queue: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: kylechar <kylechar@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774742}