chromeos: Clean up sync consent screen browser tests
Several tests use a test helper method that contains both logic and
expectations, as well as conditionals on the feature SplitSettingsSync.

Eliminate the test helper and put the expectations directly in the
tests. This involves a bit of copy/paste, but follows the spirit of
DAMP not DRY testing. go/tott/598

Only expect conditions relevant to the test (e.g. don't check for
metrics histograms again in the language switch tests).

Force branded builds once in the test suite, which makes the tests
more concise. The existing tests that un-force branded builds
continue to work fine due to AutoReset.

Bug: 1074436
Change-Id: I41865cf2e8df15fed0c78a6b5b8ea761d29a9495
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2228354
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774761}