MSE: Remux noise-xhe-aac.mp4 to correctly mark nonkeyframes
Previous fragmented mux of this test file marked nonkeyframes in input
as keyframes in output. Remux fixes that, though the MOOV's duration and
creation date are cleared, so test is updated to expect the stream to
appear to be "live".

The test is also updated to expect the correct sequence of emitted key
and nonkeyframes.

TEST=MP4StreamParserTest.MPEG_XHE_AAC
BUG=1079034

Change-Id: I9c7fd981fe4cea92e1ff0a90058ea1f10905d996
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225244
Commit-Queue: Matthew Wolenetz <wolenetz@chromium.org>
Auto-Submit: Matthew Wolenetz <wolenetz@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774769}