Origin isolation: add mismatch console warnings and fix use counter
This adds console warnings when header-based origin isolation is
requested, but not granted, or when it is not requested, but is imposed.
Along the way, a bug with the use counter code for the Origin-Isolation
header was discovered. Fixing this is best done via similar mechanisms
as adding the warnings, namely saving some state early in the
navigation lifecycle to be reported later.

Bug: 1087562, 1066931
Change-Id: I5dce9f48c64f4dd62dcdcea60be94e6492ab9085
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220358
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Domenic Denicola <domenic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774800}