FontAccess: FontMetadata::blob returns full file bytes
Add a new method, FontMetadata::blob, which returns the contents
of a valid font file as a Blob.

FontMetadata::getTables returns the tables from a font file.
However, developers have to synthesize the tables into a full font file
before using the data.

Returning the full font bytes saves developer effort, computational
resources for merging the tables and increases the font byte throughput
by a median of 4+X on a benchmark on my machine.

This CL also fragments the tests for fonts for easier management.
Large font table tests are skipped for now because they are flaky and
sometimes cause crashes of content_shell on older versions of MacOS.

Bug: 1087669
Fixed: 1076467, 1087669

Change-Id: Ia45037778aff3942335d7e4575d2d4174b5b00c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199987
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Commit-Queue: Olivier Yiptong <oyiptong@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774855}