Demo Mode: Install sample photos to My Files instead of Downloads
This works around the ARC++ Download folder bug in Managed Guest
Sessions, where the <shared storage>/Download mount in Android isn't
working correctly. This bug causes any reads of the Download directory
in ARC++ to fail, resulting in the sample photos not being readable from
Android apps - including the media scanner.

This CL moves the sample photos from Downloads to My Files, which allows
the Android media scanner to pick up on those photos. Most notably, this
allows Google Photos in Demo Mode to display the sample photos bundled
in Demo Mode, which was broken before this CL.

Note that we would still prefer the sample photos to be stored in the
Downloads directory, as most apps expect photos to be there instead of
My Files (such as the CrOS camera app and the "open file" dialog).
We should revert this CL once the aforementioned bug is fixed.

Bug: b/158053270
Change-Id: Id4f39c23ec3321bca424bed405281e42045a136d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2228272
Commit-Queue: Michael Cui <mlcui@google.com>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Patti <patricialor@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774899}