Failed navigations may commit origin incompatible with the origin lock.
RenderFrameHostImpl::GetExpectedMainWorldOriginForUrlLoaderFactory takes
into account pending navigations, so that URLLoaderFactory objects
pushed to the renderer are compatible with the most recent Commit IPC
that was sent to the renderer process.
GetExpectedMainWorldOriginForUrlLoaderFactory relies on
GetOriginForURLLoaderFactory helper functions which asserts (via CHECK)
that the computed |request_initiator_site_lock| is compatible with the
origin lock of the renderer process.
GetExpectedMainWorldOriginForUrlLoaderFactory looks for
Commit-send-but-DidCommit-not-yet-received navigations in the
RenderFrameHostImpl::navigation_requests_ container.

The problem is that:
1. RenderFrameHostImpl::navigation_requests_ container may include
   failed navigations.
2. RenderFrameHostImpl::ValidateDidCommitParams explicitly relaxes
   security checks for failed navigations (see
   |bypass_checks_for_error_page|).
3. GetOriginForURLLoaderFactory unconditionally CHECKs
   CanAccessDataForOrigin without accounting for failed navigations.

The CL here teaches GetOriginForURLLoaderFactory to relax security
checks for failed navigations.

Bug: 1058111
Change-Id: I59dbe5b92728fcbc28da35a1cf0226a8464b0e48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225242
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774901}