Move PopupBlockerTabHelper and popup_blocker.h to components/blocked_content
A delegate interface was created to wrap chrome's NavigateParams, and
provide chrome specific logic on blocked popups. Things are a bit tricky
since in the original code MaybeBlockPopup may or may not take ownership
of the NavigateParams (by calling std::move(*params)). The delegate is
now passed as a unique_ptr to MaybeBlockPopup, and the unique_ptr is
returned if the navigation should continue (nullptr if the popup should
be blocked).

Bug: 1084013
Change-Id: Ic05d7539c6f53927721158dfffd88b1aba7f3125
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2224824
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774905}