Fixed crash while setting the VK visibility request.
While updating the text input state, the VK visibility request was being set
unconditionally even if the VK policy was not defined. Added checks to ensure
that the VK visibility request is only set when the ExecutionContext is valid
and the last request set it to value other than none.

Bug: 1088364

Change-Id: Ibb62f906de4df42881028af6791b84b53b34d3c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2224652
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Anupam Snigdha <snianu@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#774941}