Use ServiceWorkerStorageControl::GetNewRegistrationId()
This CL replaces ServiceWorkerStorage::GetNewRegistrationId() call
with the corresponding mojo method as a preparation for the Storage
Service migration.

This CL increases the android binary sizes by ~30kb. Most of the
increase (~29kb) comes from mojo generated code and there isn't much we
can do to reduce them. Introducing the mojo interface is necessary for
migrating ServiceWorkerStroage to the storage service. See the design
doc [1] for the motivation of the migration and plans.

[1] https://docs.google.com/document/d/1hO0WEuoEOgEBlf5nDE3fDjuQY2bV_d5hLtk25iBo0LM/edit?usp=sharing

Bug: 1055677
Binary-Size: Size increase is unavoidable (see above)
Change-Id: I33e1e72b1da94ae02dfbe2a929895fc36e216e66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2217935
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774947}