Crostini: Attempt the upgrade up to 3 times
Upgrading is a long and potentially flaky process. If the flakes are
due to remote repos being unavailable, retrying the upgrade is useful,
since some apt commands in the upgrade can continue where they left off.
This should increase the number of upgrade successes in the wild.

Bug: 1084825
Change-Id: Ib0adf945fb60eb2e682204d949cf7aa7a273d0a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2228590
Commit-Queue: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Daniel Ng <danielng@google.com>
Cr-Commit-Position: refs/heads/master@{#774970}