Refactor FrameAncestorThrottle in order to extend it for CSPEE
In this change I refactored the ProcessResponseImpl function of the
FrameAncestorThrottle so that it can process a list of different
checks, and I moved into a subfunction the checks for the
X-Frame-Options logic. In this way it will be easier to add additional
checks to the FrameAncestorThrottle, that will be required for moving
Content Security Policy - Embedder Enforcement out of blink.

Change-Id: I6b5ff5a7c767e1c1c000d193742059aa47f1cae3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218076
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775008}