Rename HasExplicitlyInheritedProperties to ChildHasExplicitInheritance
The name is misleading, because it's actually a flag set on the parent
style when some child uses explicit inheritance.

This CL renames the flag such that name makes sense, and also because
we need to free up the current (approximate) name for a new flag
(in a subsequent CL).

Change-Id: Ie960dccf961b49b716824c44de9d3233522f96d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2228879
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775066}