Use last committed origin to track active content with cert errors
This is a speculative fix for a test flake that I can't reproduce
locally or on trybots.

Previously, when a page loaded an active subresource with a cert error,
we were using the *visible* navigation entry to compute the
origin to mark as insecure. My theory is that the visible navigation
entry might not correspond to the document that actually loaded the
subresource. In the flaky test, we load a page with a cert error, and
then navigate back. The page with a cert error triggers a favicon
request (we treat favicons as active subresources even though they are
just images), and I think (?) it's possible that this insecure favicon
resource will be processed by SSLManager after the back navigation has
changed the visible entry, therefore marking the page we're going back
to as insecure even though it shouldn't be affected.

This speculative fix drops the code that computes an origin from the
visible navigation entry and passes that into SSLManager (and also
drops a couple obsolete TODOs alongside it). Instead, we look at the
committed entry's origin and mark that origin as insecure.

Bug: 638576
Change-Id: Ief60adf121261a7021bf5be6c196d5c065b06663
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218545
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Commit-Queue: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775122}