Don't use @media with feature expressions in UA style.
We don't support MediaValues based feature evaluation in UA stylesheet,
only matching media types. There was an instance of
(-webkit-min-device-pixel-ratio: 2) in mediaControls.css which would
always evaluate to true in the UA sheet, but not in the context of the
evaluator in the StyleEngine for dpr=1. That caused us to constantly
update active stylesheets for frame resizes because we thought the media
result would change.

This also contributed to hide a bug on tweetdeck.twitter.com before we
optimized the ruleset invalidation for media query changes.

This should not be a behavior change for the mediaControls stylesheet
since the removed query would always evaluate to true so that we would
always use the high resolution image.

Bug: 1088340
Change-Id: I23f582893b21db6f14903836d58608caa82e5577
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2228848
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775144}