De-flake webaudio-autoplay-iframe-with-gesture due to compositing race
chrome.gpuBenchmarking.pointerActionSequence() has synchronization
problems, as the hit testing is done against compositor submissions
but the gpuBenchmarking methods do not cause or wait for compositing
to occur. So tests must explicitly do this themselves to avoid flaky
behaviour.

The webaudio-autoplay-iframe-with-gesture.html test on its own works
fine, somehow. But the compositing of the iframe fails to beat the
gpuBenchmarking synthetic events if it runs after the
document-user-activation-cross-origin-feature-policy-header.html
test which also involves gpuBenchmarking synthetic events.

This adds waits in the test's js code to make up for the
gpuBenchmarking harness being flaky.

R=yigu@chromium.org

Bug: 872952, 1091299
Change-Id: I6c4cb1ffc51286584ef9c7079bfff8677da1e1a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2230768
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Yi Gu <yigu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775195}