Offline installation of force installed extension from cache
If the device is offline, we fail to download and install the extensions
which are listed in ForceList policy. We only look in the cache if the
manifest fetch was successful. If manifest fetch failed, then we report
the failure. Here, we try to fetch the extension from cache in case we
failed to fetch the manifest and try to install the extension. In case
we received MANIFEST_FETCH_FAILED or MANIFEST_INVALID error, instead of
reporting the failure, try fetching the extension from cache.

For force installed extensions,
In case of no internet connection:
We don't retry fetching the manifest now, and try installing directly
from the cache. If the extension is not present in the cache, we retry
downloading with backoff delay.

In other failures related to Manifest fetch like server shut down etc.,
we make retry attempts and then in case of failure try installing it
from the cache.

A new histogram is added:
Extensions.ForceInstalledCacheStatus : It reports the cache status for
force installed extensions whenever a CRX file is looked up in the
cache. It reports the cache status irrespective of whether the extension
failed to install or installed successfully.

Bug: 1051870
Change-Id: Iab8d95a7ffbf144ae415686b6a80d501a19823e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2054244
Commit-Queue: Swapnil Gupta <swapnilgupta@google.com>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775200}