Fix fallback favicons in Chrome site settings
crrev.com/c/2171450 introduced a bug in Chrome's Site Settings UI that
caused fallback favicons to not load. That CL used the iconUrl parameter
of the onFaviconAvailable callback method to create the fallback favicon
when a real one wasn't available, but that parameter actually stores the
original url of the saved favicon, not the url of the page we're looking
up a favicon for, so it was empty when one wasn't found. This CL makes
the callback use the original url when generating the fallback favicon.

Bug: 1077716
Change-Id: I5e9d047994f0ec71a83930e97d410636d9e3134e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2226297
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775217}