MacPWAs: Use RemoteCocoa for some Google apps
Change YouTube, GMail, and Google Drive (non-bookmark) apps to use
RemoteCocoa. Do this just by inspecting their app id.

These apps exist in a strange corner where they are not bookmark apps,
but largely behave like bookmark apps. It's an open question if they
should be multi-profile or not. In this patch I err on the side of
caution and treat them as old-style non-multi-profile apps, because much
of the multi-profile handling code is now PWA-only (post-BMO).

In working on this, I found a bug wherein we assume when generating app
shims, that any app that has a start URL is a bookmark or PWA app. This
is not a valid assumption. Add a new field to web_app::ShortcutInfo to
track if an app should be multi-profile, and populate it accordingly.

There will be one bug here that I won't be able to fix without more
effort and risk than it's worth. If someone created a multi-profile app
shim for any non-PWA app that happened to have a valid URL, that
shortcut will no longer "work" in that it will open the app in a new
Chrome window instead of an app. Launching through chrome://apps will
repair the situation. This is documented in crbug.com/1091318.

R=dmurph
TBR=rdevlin.cronin

Bug: 1086824
Change-Id: Ibd829b07ed806c3e3c45b67075c5e25e368b9341
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2229663
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775221}