[Shelf] Refactor on layout strategy calculation
When calculating the target layout strategy, only two parameters are
required: the scroll distance on the main axis and the length of
available space to layout shelf icons. In other words, we don't need
the scrollable shelf's view bounds which is the third parameter of
the current function.

This CL should not introduce any visible difference.

Bug: 1042911
Change-Id: I57431969cb0bdaee71e5ba2b82faa4fdcc02428a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231090
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775253}