[Extensions UI] Customize the search overridden prompt
If the default search provider was overridden by an extension, we show aprompt to the user confirming it was intentional. This customizes that
prompt in different scenarios:
- If Google was the old search provider, use "Back to Google" language.
- If another search provider from the default list was used, use
  "Back to <other search provider>" language.
- In all other cases, use the generic "Did you mean to change your
  search provider" language.

This requires some fun plumbing in order to determine what the search
engine would be, if the extension weren't enabled. In particular,
finding the "secondary" search engine from the TemplateURLService.
For this, introduce TemplateURLService::GetDefaultSearchIgnoringSource()
and DefaultSearchManager::GetDefaultSearchIgnoringSource(), which allow
querying for the default engine ignoring any from a specified source.
Using this and ignoring the extension source, we should be able to find
the "secondary" search engine.

Add browser tests and UI tests for each of the dialog options (back to
google, back to <other>, and the generic version).

Bug: 1079364
Change-Id: I00fec35a7b3e3442fbe89c276513023e9ca12a3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218805
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775262}