[mpc] Set ChildHasExplicitInheritance also for cache hits
When a style resolve is a cache hit, we don't go through the normal
process of applying declarations via StyleBuilder::ApplyProperty.
We then also don't set the ChildHasExplicitInheritance flag on the
parent style for cache hits. This wasn't a problem a before, since
any style with a parent with this flag set could not enter the MPC.

However, with MPCDependencies enabled, such styles can now enter the
cache, and cache hits for those entries will then fail to set the flag
on the parent style.

To fix this, this CL adds a new flag HasExplicitInheritance, which
pertains to the _current_ ComputedStyle being computed (not the
parent). This way we can get the information we need into the MPC,
and perform the necessary propagation to the parent style when we
need to.

TEST=external/wpt/css/css-transitions/inherit-height-transition.html

Bug: 1057072
Change-Id: Ic03cdc6377ab5e58d6fe08a6fef2bd7505064f16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2228535
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775263}