Keep hotseat shown after window drag from shelf to split view
When flinging a window in split view from shelf, the side of the split
view with the flinged window transitions to overview (though, split view
remains active). With existing logic, the hotseat was not getting marked
as extended for this gesture, which created a weird interaction with the
hotseat during window drag/fling from shelf. The hotseat would be
extended from the bottom of the shelf as pointer was moving up from the
shelf, then immediately hidden as the gesture ended. This CL changes the
logic to keep the hotseat in extended state when the gesture ends - the
hotseat is marked as manually extended, and logic that hid the hotseat
on visibility updates with split view with overview is removed.

Additionally, fixes an issue where window drag controller was not
exiting overview when drag window bounds were getting restored for
gestures from split view. Also, makes window activation changes while
dragged window bounds are being restored not mark hotseat as not
manually extended. These two were caught by a test added in this cl.

BUG=1025301

Change-Id: I494d1f924bdc6f88294cd0610b9dcc929fa0591f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2226066
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775306}