Don't reset install_progress_observer_ipc_ when update is complete.
There is no need to reset install_progress_observer_ipc_, and in fact,
it is harmful to reset it.

When a sequence of progress callback, followed by an install complete
callback occurs, it is possible that the UI thread processes the
progress notification after the complete callback has run, and
thus install_progress_observer_ipc_ was reset.

That is because the WTL UI runs on its own thread. The execution of
these two threads is only synchronized when the UI is dismissed,
therefore, install_progress_observer_ipc_ can't be reset while
messages are posted to the UI, and they are in flight on the UI thread.

Bug: 1091452
Change-Id: I5c0975eab31c3683a16d2524fd15a708218cfe2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231744
Reviewed-by: S. Ganesh <ganesh@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775322}