Shelf: Refactor some code in preparation for per-desk shelf
We used to have first and last visible indices, with the
assumption that ShelfAppButton views in the range:

  [first_visible_index_, last_visible_index_]

are all visible, and the indices are contiguous.

This assumption won't stand when we add the per-desk shelf behavior,
i.e ShelfAppButton views that correspond to apps running in inactive
desks won't be visible, and they may be positioned in the middle of
currently visible app buttons.

Therefore, instead of tracking first and last visible indices, this
CL introduces an std::vector |visible_views_indices| which will
contain the indices of views/shelf items that are currently visible.

Currently, all views are visible. Hiding the inactive desks app views
will be done in a follow-up CL.

BUG=997102
TEST=All tests must pass

Change-Id: I3c1242cff8f136932538256a77e1297c64f37c19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2227615
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775333}