Move V8PerFrameMemoryDecorator out of the chrome layer.
The decorator implementation is now in components/performance_manager
and the renderer side mojom implementation is now in content/renderer.
This will allow the decorator to be used from other embedders.

This removes the host_id lookup for extensions since this is not
available from content. A followup patch will expose this in an
embedder-neutral way.

R=siggi

Bug: 1080672
Change-Id: Ie5832219e72d350647424359632178ebd050f038
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2230903
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Commit-Queue: Dmitry Gozman <dgozman@chromium.org>
Auto-Submit: Joe Mason <joenotcharles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775334}