Fix relevant mutations behavior for lazy loaded images
The CL crrev.com/c/2163889 fixed a bug whereby when
ImageLoader::DoUpdateFromElement was invoked for a deferred image, the
image would be requested, even when the loading attribute was still in
the kLazy state, instead of continue to be deferred. However the fix was
very specific, and only let a specific type of re-invocation continue
to be deferred. This CL implements a more-general fix, whereby all
re-invocation of ImageLoader::DoUpdateFromElement continue to defer the
image request, if the deferral conditions are still met.

R=fs@opera.com,sclittle@chromium.org

Bug: 1082961, 999466
Change-Id: I644c36166efae0815be095ecdf99c86a36241233
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204619
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Scott Little <sclittle@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775419}