BackForwardCache: Fix the crash at handling first-paint events after back-forward cache restore
When the navigation for back-forward cache happens on a main frame and
then a regular navigation happens in an iframe, the new mojom::
BackForwardCacheTiming doesn't include the first-paint event after
back-forward cache restore. This conflicts the assumption that the new
BackForwardCacheTiming should include this information at
DispatchObserverTimingCallbacks.

Bug: 1090761
Change-Id: Ie5de298425dbb5d5028b395179d131e24377501e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2230580
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775428}