crostini: installer supports closing from system UI
We are adding a shelf item for the installer, the side effect is that we
are going to have a "close" item in the shelf context menu, which do not
do anything before this CL because the installer blocks closing from
system UI.

This CL adds the closing support by adding Page::RequestClose() to
crostin_installer.mojom. CrostiniInstallerDialog will trigger it
(indirectly) when there is a close request (e.g. from shelf context menu
or gesture in the overview mode). When the web page receives the
request, it treats it as if the "cancel" button is clicked. Eventually,
when the page has done the closing, it will called
`PageHandler::OnPageClosed()` (renamed from `Close` for clarity) to let
the dialog close itself.

There are also some other renamings in this CL to make the code clearer.

Bug: 1026827,970196
Change-Id: I9ae9c4c708a82ce3fb642b0080880439c668c016
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2230383
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Jason Lin <lxj@google.com>
Cr-Commit-Position: refs/heads/master@{#775473}