chromeos: Add metrics to sync consent screen for SplitSettingsSync
For the new SplitSettingsSync version, which has both an accept and a
decline button, track which button the user clicked.

For both the old and the new version, track whether the screen was
shown, skipped, or skipped with sync force-enabled.

Bug: 1090989
Test: added to browser_tests
Change-Id: I1697c82728b8f183bd807f910255d3820bd06587
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2229720
Auto-Submit: James Cook <jamescook@chromium.org>
Commit-Queue: Roman Aleksandrov <raleksandrov@google.com>
Reviewed-by: Roman Aleksandrov <raleksandrov@google.com>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775508}