[Infobars] Componentize almost all of //chrome's InfoBarService
We will shortly be bringing up infobars in WebLayer. //chrome's
InfoBarService inherits from the
//components/infobars/core-level InfoBarManager to glue the latter to
both //content and //chrome. Almost all of InfoBarService is just
glueing InfoBarManager to WebContents (e.g., so that infobars disappear
correctly on navigation). All of this logic should be reused by
WebLayer. A small part of InfoBarService is the //chrome-level
implementation of InfoBarManager::CreateConfirmInfoBar(), which is
actually implemented in a platform-specific fashion in various files
throughout //chrome. We add a stub implementation of
CreateConfirmInfoBar() ContentInfoBarManager, which embedders must
override if they want to actually create confirm infobars (WebLayer will
not need this functionality at least initially). //chrome's
InfoBarService does exactly that. (An alternative would be to just have
//chrome sites directly implement
ContentInfoBarManager::CreateConfirmInfoBar(), but it would be
conceptually wrong and confusing to have implementations of a
//components-level ContentInfoBarManager::CreateConfirmInfoBar() method
scattered throughout //chrome).

As a side-effect of leaving InfoBarService in //chrome,
ContentInfoBarManager is not itself a WebContentsUserData: it only
makes sense for a leaf class to be a WCUD, as a WCUD object is created
via a static method at the level of the class that is the WCUD.

Finally, WebLayer will potentially want to alter the behavior of opening
URLs so that they are opened in the current tab rather than a new tab.
However, we will deal with generalizing this when we encounter it; to
date, we have not seen a need for links in the limited use cases for
WebLayer's infobars.

Bug: 1025620
Change-Id: Id21c616862357dccc909914cc85db8af29ba50e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225903
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775555}