CorsUrlLoader: Update SiteForCookies and IsolationInfo correctly.
When CorsUrlLoader intercepts a redirect, it needs to update the
site_for_cookies and TrustedParams::isolation_info fields correctly,
which it wasn't doing. These only change for frame requests, so this
may not have been an actual bug, but we shouldn't implicitly assume
that's the case in the CORS logic.

Fixed: 1091382
Change-Id: Ia04f94b315517a1e3d68935c3c72fc9a616039c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231578
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775566}