Use CommitPendingWrite to expedite flushing of prefs file in updater.
This change is intended to reduce the lag between scheduling the
prefs writes and the writes landing on persistent storage.

This CL fixes the app install case. The service case is going to
be addressed in a future CL, as part as the on-demand work.

Bug: 1091532
Change-Id: I08d0929a97183b4fe3ffc588b8db8c28dcf2fe1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231806
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775570}