[WebLayer] Add InfoBarService
This is an analogue of //chrome's InfoBarService and serves to tie
ContentInfoBarManager to a WebContents. Unlike //chrome, //weblayer
has no need to create confirm infobars.

Relatively shortly I will do a refactoring that estade@ suggested that
will enable ContentInfoBarManager to be a WebContentsUserData itself
and eliminate the need for //weblayer to have a subclass. However, I'm
getting this in now as this is on the critical path for bringing up the
translate infobar in WebLayer and I won't be able to get to that
somewhat-involved refactoring immediately.

Bug: 1025620
Change-Id: Iaa2a443f78605ba00c36f4d51ffc808030e673b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2230477
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Mugdha Lakhani <nator@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775599}