[Viz, capture] Merge to M84: Fix corruption bug: Invalidate marked buffer on reuse and remark frames on size change
The marked buffer must be invalidated, because otherwise it might be used as if it had a correct latest
version of the image, while it was overwritten by a smaller image.

The second part of the CL is an optimization: If the capture size was reduced due to pool overuse, but
content is static, previously it was impossible to cache the result and unnecessary CopyOutputRequests
were constantly made.

The drawback of the change is that if captured size is changed very rapidly, more copyOutputRequests
would be made than before. In all other cases it is optimization.

(cherry picked from commit 3f24544d13525a109e1af8f6568e44abfdc4e4ac)

Bug: chromium:1090301
Change-Id: Ib32e3908114ff2d1a276e807f5bb707893b6ae8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225766
Commit-Queue: Ilya Nikolaevskiy <ilnik@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Cr-Original-Original-Commit-Position: refs/heads/master@{#774757}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231121
Cr-Commit-Position: refs/branch-heads/4147@{#488}
Cr-Branched-From: 16307825352720ae04d898f37efa5449ad68b606-refs/heads/master@{#768962}