Rework find-from-selection so it's synchronous -- fixes flaky tests
CL that introduced the flakiness: https://crrev.com/c/2145247

That CL addressed the original problem by collapsing the selection via
AdjustSelectionByCharacterOffset. However, that call is asynchronous,
so there's a race between the selection being collapsed and the
find-in-page running.

This CL addresses the original problem in a better way -- rather than
messing with the selection, it passes a flag in the find options to
control the behavior directly.

I explored an alternative approach of adding a callback to
AdjustSelectionByCharacterOffset, but that gets a lot messier, and
results in FindBarController::Show being async and needing to provide
a callback as well.

Bug: 1072464
Change-Id: I3789d24bbe1dd515e636363ca50912e17030abc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181570
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775618}