Force persistence of ChromeOS dump-without-crashing success metrics
The metric
CrashReport.DumpWithoutCrashingHandler.FromInitSharedMemoryIfNeeded is
getting no samples on eve, even though there are crashes showing up on
the server from versions with the metric, and they show that the code is
getting past the point where the metric is recorded. This is probably
because InitSharedMemoryIfNeeded() is called early before metrics
persistence is set up, and the conditions that lead to
InitSharedMemoryIfNeeded() failing also prevent metric persistence.

Go around the normal metric collection system and write to the file that
ChromeOS uses to communicate metrics with Chrome. This is a
ChromeOS-specific feature, but that's OK, I only really need the metric
on ChromeOS.

Bug: chromium:1074115
Change-Id: I095ce3c55fd5d514fd11afcc337fcfa3d471147d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2229301
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775653}