Reland "WebUI Tab Strip: Animate tabs when moved"
This CL adds animation when an ungrouped, unpinned tab is moved, to
mostly facilitate the drag and drop experience. When the TabList's
placeTabElement is called, the method first determines if the
placement is causing a TabElement to move, then identifies which
elements the TabElement is being dragged across, and finally
animates the necessary distances for each TabElement.

Bug: 1082344
Change-Id: Ie7192c21e4feb0e9e3141ef38128840360566840
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231293
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775656}