Allow transformed rasterization for opaque layers
Because transformed rasterization may cause pixels along the edges
of a layer to become translucent, previously we disallowed transformed
rasterization for opaque layers, which caused blurriness in opaque
layers with fractional transforms. This was contradict to the guidance
of achieving LCD text by specifying an opaque background: that would
cause even blurrier text with fractional transform.

Now allow transformed rasterization for opaque layers, and clear the
contents_opaque flag of the layer. Occlusion tracking will be less
effective, but fractional transforms are not common, and it's more
important to avoid excessive blurriness in the case.

Also allow LCD text for fractional transform / offset if transformed
rasterization is enabled for the layer.

Bug: 1077071
Change-Id: I01edd307537a08f581a902c7087781dfd92307cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2229243
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775658}