[Reland][ExpandablePaymentHandler] Test UI appearance and user interaction
original CL: crrev.com/c/2227512
revert CL: crrev.com/c/2231122
revert reason:
showPaymentHandlerUi was flaky. It's very likely because the animation
took so long to finish that that it hits the timeout. Many screenshots
prove that because the payment sheet still had a thin scrim on it.

change:
Compared to the original CL, the new one removes the animation.

Bug: 1042892, 1084705, 1084707, 1084708, 1084711, 1084721, 1068741
Change-Id: Ib3401e3e422e37a733f85eb56e6d08598d3ade1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2232679
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775685}