[Spellcheck] Move some spell check histograms to open-source repo
In preparation for the Windows native spell checker integration launch
experiment, there are certain important histograms that need to be
extended for the duration of the experiment. We want to make sure that
the Windows native spell checker integration does not regress spell
check quality for users, and the ratios of misspelled words, suggestion
hit and word replacement are great indicators of that.

Since these histograms were still in the internal list for historical
reasons, I am moving them to the open source repo to extend them.

See the launch bug for more info:
crbug.com/978460

Bug: 978460
Change-Id: Iced9122820b1404ec0f442ef1bd9d4e997817961
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2232768
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775720}