[serial] Implement persistent permissions on Windows
This change implements support for persistent permissions in
SerialChooserContext and assigns devices a persistent ID based on the
Windows "device instance ID" path. This is the path used internally by
the Windows Plug-and-Play infrastructure to track the identity of
devices.

Unlike the Web Bluetooth or WebUSB APIs we cannot only use properties
provided by the device because there is not a consistent transport for
serial devices. UARTs that are part of the hardware platform especially
do not have useful identifiers. The operating system is aware of the
overall device tree and can provide something that is stable enough
for our purposes.

Bug: 908836
Change-Id: I98628278f16f57efa58c0e019687660c33a22c0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231981
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775739}