Replace the Context with Activity.
Since Context is a base class for Activity, we can replace it with the
activity itself, and pass it as an argument whenever the context is
needed.

We need to do this because we require the activity for some UI operations.
This is in preparation for getting the bottom sheet manager, which is
only available on the ChromeActivity object.

Bug: 1091529
Change-Id: I26791f3a8536eb52fab36c4e01c115fe589e432a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2232822
Reviewed-by: Dan H <harringtond@chromium.org>
Commit-Queue: Peter Williamson <petewil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775752}