Animate status area widget when entering/exiting overview
Currently the bounds animation of status area widget is disabled if
the target bounds have the different size from the current bounds.
Based on https://crrev.com/c/2158911, it aims to disable the animation
when tray items are added/removed. However. we do have a scenario where
the status area widget's bounds change meanwhile it does not involve
the tray item's movement: when entering/exiting overview mode in tablet,
the status area widget's height changes.

In this CL, StatusAreaWidget::LayoutInputs has a new attribute to
indicate whether animation is allowed. The animation is banned if the
status area widget's length on the main axis (that is the width for
horizontal shelf or the height for vertical shelf) changes.

Bug: 1079347
Change-Id: I9190168b40af6e88bf774e1f98c7a0de0467be7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2227099
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Tim Song <tengs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775800}