[Android] PhotoPicker: Make sure playback pauses while seeking.
Videos that are playing should pause while the user is
interacting with the SeekBar.

The best way to do this is via touch tracking. Trying to use
OnSeekCompleteListener inside onProgressChanged is unreliable,
because it won't fire if and while the thumb is moving too
quickly (pausing would then happen only when seeking is slowing
down, and it looks jerky).

Bug: 895776, 656015
Change-Id: Ie044b5cac30741bba0f5ee4a59650ef514ae8a6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231359
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775818}