Enable full testing of CSSReducedFontLoadingLayoutInvalidations
This is preparation for enabling finch experiments of the feature on M85

Note: though M85 hasn't reached Beta, we'd like to have it covered by
test bots first to catch any breakage. We'd also like it to be tested
by perf bots, since this "feature" is a performance optimization.

Bug: 441925
Change-Id: I3ef98ffcfa960381efd034cd1cd63e55fbda4b01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231269
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775847}