Add back NetworkServiceNetworkDelegate referrer DumpWithoutCrashing
TL;DR: We need to add back the DumpWithoutCrashing() call to
NetworkServiceNetworkDelegate::OnCancelURLRequestWithPolicyViolatingReferrerHeader.

In crrev.com/c/2229205 I moved the DumpWithoutCrashing() call from the
NetworkServiceNetworkDelegate to the mojo serialization code, to help
debug the DWCs in crbug.com/1090391. This opened up crbug.com/1091711.
I was able to reproduce a request that hits the DWC in the mojo
serialization code, so I thought all was well. I then moved the DWC back
to the network service, only to find that it was not being hit.

Therefore I found a separate referrer-related bug that was likely not
(always) contributing to the original crbug.com/1090391. This means
crbug.com/1090391 and crbug.com/1091711 are exclusive.

The repro _does_ cause a DWC in the mojo serialization code, because at
this point, the network::ResourceRequest has the following properties:
 - referrer: "about:blank"
 - referrer_policy: kNever

Clearly the request's referrer violates its policy, so the DWC is hit.

The repro _does not_ cause the DWC in NetworkServiceNetworkDelegate,
because by the time the request gets to the place where the referrer is
checked / validated [1], the net::URLRequest's referrer is the empty
string, and therefore does not violate the policy. The reason the
referrer is the empty string at this point, is because the
net::URLRequest's referrer is set (in URLLoader::URLLoader) by calling
GURL::GetAsReferrer() on the input network::ResourceRequest::referrer.
GetAsReferrer() strips "about:" URLs to the empty string.

So, I will fix the repro case I found which should greatly reduce or fix
the DWCs in crbug.com/1091711, but at the same time we should add back
the DWC to the NetworkServiceNetworkDelegate to verify that there are
other causes of this DWC.

[1]: https://source.chromium.org/chromium/chromium/src/+/master:net/url_request/url_request.cc;l=633-637;drc=c77ccde5b729af314595e528d89837d2c890f94f

TBR=kinuko@chromium.org

Bug: 1090391
Change-Id: I11ff955755dd2d9ed47a15d3faaaabb6fbdcca84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2234121
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775879}