Fix the top border colour on Windows >=1809
Windows changed the way themes work in 1809. The window border is now an
optional setting, but Chromium still creates its own. This results in an
out-of-place accent-coloured border when the Windows setting is
unchecked. This patch changes the top border to not use the accent
colour on 1809 and newer, if the checkbox for displaying accent borders
is not set in the Settings app.

Microsoft fixed this in Edge, and it behaves very similarily.

Bug: 950362
Change-Id: I64325ccae4fe1f4b8d13900836ff68637a1ded1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231049
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775927}