[b4p/Settings] Add link to 'device passwords' page
This CL adds a button in the passwords settings page linking to the
new 'device passwords' page. The button only appears for users eligible
for b4p, opted in and with at least one password on the device.

Screenshot of page before the CL:
https://screenshot.googleplex.com/5gSRLRXgUBr

Screenshot of page after the CL:
https://screenshot.googleplex.com/DMPPYKWLKuq

Screenshot of mock:
https://screenshot.googleplex.com/tqGBcM4PNXy

Bug: 1049141
Change-Id: I7bc79c2a16d166ac6a3392939abe341441c6d6ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231058
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775968}