[WebLayer] Bring in fork of Chrome for Android's infobars UX
WebLayer will be adding infobars, but its initial implementation will
be experimental. After discussion, we have decided that the best path
is to fork rather than componentize Chrome for Android's infobars UX
code for this initial experimental implementation. This approach will
avoid burdening chrome-team@ with componentization before we see that
it is necessary and will enable the WebLayer team to iterate on this
implementation as necessary quickly and without impact on Chrome.

This CL brings in the initial support that will be necessary to enable
compact infobars to be rendered in WebLayer. This initial change does
not actually do any rendering of infobars but simply brings in the
support and get it to build. Followup changes will bring in the UX for
the translate infobar and have that render. Those followup changes have
been locally tested and have informed what has been brought in in this
CL.

The way to examine this CL is to look at the diff between PS4, which
contains just a straight copy of the classes from //chrome, and the
latest PS. Note that we did not try to strip down the code at this
point beyond (a) making the minimal changes necessary to get the
code to build in //weblayer, and (b) eliminating code that would
definitely not build in //weblayer and which is either not relevant
or for which we will find //weblayer-specific solutions. Throughout, we
have left TODOs for followup work to enable getting this initial support
in and building on it ASAP.

Bug: 1025620
Change-Id: If859a9687e348728084148d4f924895b71340ca3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2224869
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775988}