[Autofill Assistant] Moved view management into dedicated class.
This CL moves ownership and management of the map containing all views
with proto-defined identifiers to a dedicated class. This is in
preparation for an upcoming CL, where retrieving views by their ID will
become a bit more complicated.

This is a refactoring only and should have no user-facing changes.

Bug: b/145043394
Change-Id: I1c16182bee7036a153a13c54f2c718b45e73e851
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2232718
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/master@{#775998}