[iOS][tabmodel] TabModel Deprecation (without create/disconnect).
This CL reduces a dependence on TabModel by updating Browser,
BrowserImpl, BrowserInterfaceProvider, BrowserViewWrangler,
StubBrowserInterface, TabOpening, TestBrowser to get the values of
WebStateList and BrowserState without a use of TabModel.

The follow-up CL is going to include the migration of creating and
disconnecting of TabModel in Browser.

Bug: 1045575,1049335,1049844,1049839,1050124,1050127,1050138
Change-Id: I39ae7e09a713efd30418d951cf0f2ecc6bfd0447
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2228843
Commit-Queue: Nazerke Kalidolda <nazerke@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776002}