[identity] Record getAuthToken() result in a histogram
This CL adds a Signin.Extensions.GetAuthTokenResult histogram that
records the result of chrome.identity.getAuthToken() call.

I modified the error handling in IdenityGetAuthTokenFunction class to
report an IdentityGetAuthError object instead of a string. This object
is then converted to an enum for the histogram recording and to a string
for returning to the the extension.

Some error strings got mapped to several error codes so we can uniquely
identity a code path that returned an error.

Bug: 1091423
Change-Id: I4ab1116305a0aa3777d829b002165b4c4c1a2cd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2234752
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776035}