Uses all nodes, including ones that are platform ignored, when finding the next valid position
In AXPosition in Blink, when we are unable to convert in a straight forward manner from a DOM to an AX position or
vice versa, we do the next best thing by moving to the next or previous valid position in the
corresponding tree.
However, the method that finds the next unignored / valid position in the AX tree skips over
nodes that are included in the AX tree but are accessibility ignored, i.e. they don't appear
in the browser's platform tree.
Similarly, the method that computes the next unignored node in the accessibility tree given
a DOM node, only looks at the next / previous unignored sibling of the ignored node
without looking at the node's descendants. However, there are cases when an ignored DOM node
could easily be represented by one of its children. For example:
<span>hello</span>
Even though the span node is ignored in the accessibility tree, its text child is not.

AX-Relnotes: Screen readers more reliably read the currently selected text.

R=dmazzoni@chromium.org

Change-Id: I687785cdc00ec739b61c943bfc9a3db030c5ae28
Bug: 1082814
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220035
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776049}