[ExpandablePaymentHandler] UI destroys on activity destroyed
Bug cause:
The UI didn't listen to the activity destroy event. When it happens,
the UI didn't notify the PaymentAppProvider (via
ServiceWorkerPaymentProviderBridge). As a result, the
PaymentAppProvider was still alive after the UI has closed. In this
situation, if the merchant page try to show() the UI, it would hit the
error "Already called show() once"[1] (because accept_resolver_ != null)

[1] https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/renderer/modules/payments/payment_request.cc;l=763

Fix:
Let the UI listens to the activity's destroy event. When that happens,
notify PaymentAppProvider that the UI has closed.

Change:
* the fix
* centralized the calls of ServiceWorkerPaymentAppBridge and call it
just before the WebContents is destroyed. This is to ensure that only
one "onClosingPaymentAppWindow*" would be invoked.
* added scrimCoordinator null check. The null is triggered when a dying
activity triggers the UI to close (e.g. theme change as in
crbug.com/1092309).


Bug: 1091137, 1087410, 1092309

Change-Id: I78bfafadd98278765170540ae369d3d497dcbf0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231978
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776064}