[blinkpy] Always consider WPT idlharness tests to be slow
These tests consistently take 5-6x longer to run when DCHECK is enabled
versus when it is now. This takes a 0.5-1s test into flaky timeout
category, and we have seem many, many flakes with idlharness tests on
the CQ bots (which are DCHECK-enabled, versus the waterfall bots that
aren't).

Investigations into this slowdown were inconclusive. It appears that the
heavy use of JavaScript in idlharness tests means they get a much worse
slowdown than most web_tests (which are rarely JavaScript-bound), but
even with the v8 team's help we were unable to identify any particular
DCHECKs that could be modified or disabled to improve the situation. As
such, our final option is to just hack them into always being considered
slow tests.

Note that since we can't detect whether or not DCHECK is on, we consider
them slow for all setups.

Bug: 1047818
Change-Id: I33b27017f6c9641dbe9881a78a33563fb8b2df67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2232698
Reviewed-by: Robert Ma <robertma@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776069}