Simplify LayoutBox::LayerTypeRequired
This patch removes redundant checks from LayoutBox::LayerTypeRequired:
1) LayoutObject::HasTransformRelatedProperty has been removed because a
call to ComputedStyle::HasTransformRelatedProperty is already included
in IsStackingContext. The LayoutObject and ComputedStyle functions are
not equal (e.g., for LayoutInline), but they are equal for LayoutBox.
2) LayoutObject::HasReflection has been removed because it is included
in LayoutObject::CreatesGroup.
3) ComputedStyle::ShouldCompositeForCurrentAnimations has been removed
because it is included in LayoutObject::IsStackingContext.

A stale comment from 2015 about hasAutoZIndex has also been removed,
since this is now in IsStackingContext.

"IsPositioned() || IsStackingContext" has been replaced with a call to
IsStacked().

Bug: 1068596
Change-Id: Ifb05c4ff5536011ccff6271277d6eaba6a230554
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2234123
Reviewed-by: vmpstr <vmpstr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776076}