[Live Caption] Stop speech recognition when caption bubble is closed.
Introduce a reply to the OnTranscription function of the live caption
mojom interface. The reply is a boolean of success which denotes whether
the browser successfully received, routed, and set the transcription
result. If there was not success, because the caption bubble was closed,
the browser was closed, or the associated web contents did not match an
existing browser with a caption bubble, then the OnTranscription mojo
function returns false, causing speech recognition to stop.

Bug: 1055150
Change-Id: I8c8e0b443c8c90afa78e85e81e3ca7cda8a49920
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225878
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Katie Dektar <katie@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Cr-Commit-Position: refs/heads/master@{#776086}