Add test for image_annotator language fallback.
This test is a follow-up to this patch: http://crrev.com/c/2226300

It ensures that if the server request uses a different language than the
language passed with the original request, the annotation still succeeds.

Also adds a log to make it more clear if we were ever to hit this case
in the future. Previously it was a silent failure.

AX-relnotes: N/A
Bug: 1090234, 1070505
Change-Id: I6c728a859bfa4c9a182ea4dd446b3b54c1836014
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233482
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776131}