Fix a crash in TabGroupUiMediator#onPageLoadStarted
http://crrev.com/c/2227890 spotted that the TabModelSelectorTabObserver
can be leaked across sessions during theme change, and tried to fix
this by removing the observer when tab is detached. However, there are
still crashes with similar stack, and the suspect is that there are
still leaked observers. This CL adds a band-aid fix to the problem for
M84 experiment.

Bug: 1087826
Change-Id: I34c149e0651e30b08a19872c6332d70c4028fd5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2236098
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776133}