Clearly distinguishes between methods that work on ignored vs unignored nodes in Blink
In AXObject, there are a lot of methods that traverse the accessibility tree.
However, there is a lot of confusion as to which methods work on unignored nodes,
which on ignored nodes that are included in tree, and
which on nodes that are totally excluded from the tree.

This confusion is the cause behind a bug in AXPosition, whereby
the wrong method was called, unignored AXObject::NextInPreOrder(), instead of the ignored version
AXObject::NextInTreeObject.
This bug was fixed in https://crrev.com/c/2220035/

This patch renames all the tree traversal methods to give them more explicit names.
In the process one more bug, having the same cause, was discovered in AXPosition
(see ax_Position_test:PreviousPositionAfterARIAHiddenShouldNotSkipARIAHidden).

This patch also harmonizes somewhat the method names between Blink and
those used on the browser.
For example, took the opportunity to introduce UnignoredChildCount() and UnignoredChildAt(index)
methods which are especially useful in tests.

AX-Relnotes: n/a.

Bug: 1082814
R=dmazzoni@chromium.org

Change-Id: I6c092dda12406ea7fb0ce9c2ade0410470f28125
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2205399
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776201}