[omnibox] OmniboxPopupModel: Fold SetSelectedLine into SetSelection
Currently, there's all three methods:
 1. SetSelection
 2. SetSelectedLine
 3. SetSelectedLineState

SetSelection calls both SetSelectedLine and SetSelectedLineState.
That has worked well so far, but this will cause a problem for section
headers.

The problem is: SetSelectedLine is what updates the textbox inline
autocomplete. With Headers, we may need to update the inline
autocomplete when the line state changes (from header to match), rather
than when the line number changes.

Philosophically, I think this is also a good change, because in the
omnibox code, there should be one, and preferably only one, obvious way
to do something.

This CL is not intended to change any user facing behavior but it does
change some of the logic of what happens in some circumstances or in
what order. So it's possible that some unanticipated changes may arise.
If so, they will be addressed in subsequent CLs.

I'll fix the actual inline autocomplete behavior for headers in a
followup CL.

Bug: 1078183, 1052522
Change-Id: I3c086de080acf03eeaba72a6ac358408ca3462f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2227453
Commit-Queue: Tommy Li <tommycli@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776216}