[Settings] Refactor: change createExceptionEntry(); set ids in tests
This CL is simply the analogous of crrev.com/c/2218055 for password
exceptions. This means it changes the signature of this helper function
to: a) make it easier to write tests in a multi-store scenario; b) allow
the deduplication functionality to be introduced without breaking the
existing tests.

Besides that, the CL fixes a comment typo in validatePasswordList() and
also exposes chrome.passwordsPrivate.removeExceptions() API in the
PasswordManagerProxy, so it can later be used at PasswordsSection when
exceptions are deduplicated.

TBR=fhorschig@chromium.org

Bug: 1049141
Change-Id: I3a501283c964904b981a6e2463336f4f147816bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2234755
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776221}