Fixed download progress infobar for night mode
We cache an old object in the DownloadProgressInfoBar native while we
keep updating the java object. When activity restarts after a theme
change, we show the UI from the old object in native. This CL updates
the java object as well to fix it.

Bug: 1092000
Change-Id: Ic5f12e5e38db2f20a3dd2ce5ec6ce061f88bfa20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233828
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776245}