Remove |AssistantState::READY| flag
This flag was flawed as it only signals the Assistant has started,
but not that it is ready to handle queries.
The NEW_READY flag does signal Assistant is ready to handle queries,
and will be used going forward (and will be renamed to READY in a
follow up CL).

Bug: b/142428671
Change-Id: I1343d9b36c1164d95e55e14f38f8bef57033b84f
Tests: ash_unittest, chromeos_unittest, and manually launched.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2236284
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Cr-Commit-Position: refs/heads/master@{#776289}