Fix ButtonController::OnGestureEvent() when button is disabled
ButtonController::OnGestureEvent() should behave in the same
way as ButtonController::OnKeyPressed() and
ButtonController::OnMousePressed(); it should not change the
button state if the button is disabled, and just early out.

This bug caused the new desk button to show as enabled, even
though it is disabled, only when gesture tapped. This didn't
happen when mouse clicked.

BUG=1084241
TEST=Added a new test that fails without the fix.

Change-Id: Ifd007a8b123a26c7109b13e0a7929c0f77f4725b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233362
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776290}