tablet: Speculative reenable of Tablet*ScreenshotVisibility*.
It's failing on asan/lsan builders. It wasn't reproducible locally
after 1000 tries, so uploading a speculative fix. According to the
stack in the builders, it seems there's no animation running after
waiting for tablet mode to enter. It seems plausible the animation
has been scheduled but not started yet, so add a waiter if the
animation is not running.

Test: ash_unittests TabletModeControllerScreenshotTest.ScreenshotVisibility
Bug: 1091085
Change-Id: I8bf796016122b0368edf860fe93299d8718f6f86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231270
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776298}