Always enable ClientSideDetection.
This change allows a PhishingClassifierDelegate
object to always exist and removes the dependency
on the Safe Browsing pref on the renderer
side. We are checking the Safe Browsing pref
on the browser side which should make this safe.
Reordered methods in chrome_render_frame_observer.cc
to match the declaration position in the header.

Bug: 1078964
Change-Id: If7388a1c038d2558abcadf3b50e41f7748dfae7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225819
Commit-Queue: Bettina Dea <bdea@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776300}