PDF Viewer: Move some functions and typedefs to a utils file
This reduces the size of pdf_viewer.js, in preparation for making it a
Polymer element that defines the UI of the PDF viewer. There will need
to be two versions of this element, one for Print Preview, so moving
shared code out of pdf_viewer.js itself will reduce the amount of
duplication.

Change-Id: I40e2a0565159bc028cab52fffb0eb8de8984da16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233487
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776320}