Query LUID from OpenXR runtime and send back to browser process
When initializing the OpenXR runtime, send the LUID of the GPU adapter
the VR headset is plugged into back to the browser process. This change
is one part of a series of changes to add multi-GPU support for VR.
crrev.com/c/2216166 added the LUID of the GPU that the GPU process is
using. A subsequent CL will compare these two LUIDs to determine
whether the GPU process needs to be restarted and initialized on the
correct GPU to be XR compatible.

This change adds a XRDeviceData struct that wraps around a LUID so
that the LUID is only enabled on Windows with minimal
"#if defined(OS_WIN)" wrappers. A subsequent change specified by
crbug.com/1090029 will consolidate all information about an XR device
into this XRDeviceData struct.

Bug: 792657
Change-Id: If38fd6d0b16c39b36d529d4aaa02c22fe0bcff0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225366
Commit-Queue: Patrick To <patrto@microsoft.com>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776324}