[DeviceSync v2] Force enrollment after SyncMetadata 404 errors
This is a mitigation put in place to handle 404 errors returned by the
CryptAuth server during SyncMetadata requests. A 404 error is returned
when the server cannot find an enrolled device with the same Instance
ID.

Theoretically, this should never occur assuming that 1) we never run a
v2 DeviceSync until the device has successfully enrolled using v2
Enrollment and 2) the Instance ID does not rotate. Without this
mitigation, the user will unsuccessfully try to sync until the next
enrollment, which could be up to 30 days.

This fix will be removed when the root cause is uncovered.

Bug: 1092113, b/155642672
Change-Id: Ibc766ad21d83d1ace26c5e8c6c45f3eed835da74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2234126
Reviewed-by: James Vecore <vecore@google.com>
Commit-Queue: Josh Nohle <nohle@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776330}