Create chrome_enterprise_url_lookup_service and its factory.
This CL creates classes for enterprise real time URL lookup.
chrome_content_browser_client injects an enterprise or consumer real
time URL object based on the type of the user.

A feature flag SafeBrowsingRealTimeUrlLookupEnabledForEnterprise is
also added in this CL. It is disabled by default.

The implementation for StartLookup will be added in a future CL. The
enterprise policy for real time URL check will also be added in a
future CL.

Bug: 1085261
Change-Id: I3741bbcb326c6f17e3bc95b45f1366874e97a98e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2234794
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776344}