Fixed behavior for virtual keyboard
Virtual keyboard and Search box are expected to be dismissed at one
tap/click when there is no text in the search box.
Add a test case that will tap/click somewhere outside the search box
when there is no text in the search box, which is different from the
original one that search box is not empty.

R=tbarzic@google.com

Bug: 1021276
Change-Id: I8358eb6b4eb754e0fa40c99e24ce0a0f6f170a9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233687
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Wen-Chien Wang <wcwang@google.com>
Cr-Commit-Position: refs/heads/master@{#776361}