Download later: Plumb download schedule to DownloadTargetInfo.
DownloadSchedule carries application data for download later feature,
this CL plumbs it from DownloadDialogBridge to DownloadTargetInfo.

This is mostly plumbing a base::nullopt through the pipeline. There
will be a few more CLs to plumb it to DownloadItemImpl and the in
progress download db.

Bug: 1078454
Change-Id: I187269c340d775997b6840ba1f16d67815a1b845
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2234129
Reviewed-by: Hesen Zhang <hesen@chromium.org>
Commit-Queue: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776366}