Extend Renderer4.GpuImageDecodeState* lives
The Renderer4.GpuImageDecodeState* histograms will be expiring soon.
These track the cache hit rate for the GPU decode cache.

This CL changes the expiry to 1 year from now, as suggested in [1]
for a histogram which will be useful indefinitely.

It also adds me as an owner of the histogram since I plan to be doing
a lot of image decoder work.

[1] https://chromium.googlesource.com/chromium/src/+/master/tools/metrics/histograms/README.md#how-to-choose-expiry-for-histograms

Bug: 1089344
Change-Id: I8ab8839e8c2433f95b5f3539c607f72f7c003a35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233926
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: Chris Blume <cblume@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776403}