chromeos: Extend metrics for the sync consent screen
For the new SplitSettingsSync version, which has both an accept and a
decline button, track which button the user clicked.

For both the old and the new version, record whether the screen was
shown or skipped, including the skip reason. Also record whether sync
ended up enabled or disabled.

This is a better version of:
https://chromium-review.googlesource.com/c/chromium/src/+/2229720

Bug: 1090989
Test: added to browser_tests
Change-Id: Id2fcc1deaf06a39b591a7f85023640459f7f80ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233486
Commit-Queue: Roman Sorokin [CET] <rsorokin@chromium.org>
Auto-Submit: James Cook <jamescook@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776441}