[content] Add WebContentsImpl::RenderFrameHostStateChanged
This CL adds a new method WebContentsImpl::
RenderFrameHostStateChanged. This is called whenever the
RenderFrameHostImpl::LifecycleState changes.

One pager doc explaining RenderFrameHostStateChanged usage:
https://docs.google.com/document/d/1vH_41CU61disTGWB2aFcP8-weHwfzHjEH7fbqdf3Ff8/edit?usp=sharing

This is useful to make sure all the UI elements like popups/dialogs
are closed when the RenderFrameHostImpl leave the kActive state.

One of the reasons why we want to use RenderFrameHostStateChanged
instead of WebContentsObserver::DidFinishNavigation is because a
committed navigation changes state of not just the navigating frame
but all of its subframes as well.  So, determining whether a particular
frame's UI element needs to be closed would require listening to
DidFinishNavigation in that frame as well as all of its ancestor frames.

This CL closes the color chooser popup which rely on a signal coming
from the renderer triggered by the document destruction which don't
happen in the case of BackForwardCache and would be problematic,
instead we close it when the LifecycleState changes from kActive state.

Bug: 1073144
Change-Id: I44eab243d5581e6e639406054d79da9efc444b60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159444
Commit-Queue: Sreeja Kamishetty <sreejakshetty@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776476}