Notifications in offline device for minimum chrome version policy
This CL introduces the handling of showing in-session notifications
when update is required as per the policy but the device is not
connected to a network for the updates to download. The notification
is to be shown when the policy is received and on the last day of the
deadline. A new WallClockTimer is introduced to keep track of the last
day for showing the notification. A new test is also included to
verify this behavior.

The subsequent CLs will introduce the handling for notifications in
case of end-of-life devices and metered network.

Bug: 1048607
Change-Id: Idecb0f976ad96af537e2ffe4faf9936b7ddfc407
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218020
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776504}