ui: fill in ChromeOS HDR DisplayColorSpaces
This CL is a refry of crrev.com/c/2072621, where only the HDR
DisplayColorSpaces are filled in and the SDR ones are skipped
and left for a later CL.

The original landed-and-reverted CL is kept in PS0 to ease
review.

Bug: 958166, b:158126931

Original CL description -----------------------------------------------
This CL adds code for populating the necessary gfx::DisplayColorSpaces
in CreateManagedDisplayInfo for supporting SDR and HDR target color
spaces (depending on the monitor capabilities and the big hdr flag).

It needs a sibling CL to allow AR/B30 FBs on CrOs crrev.com/c/2078655,
Tested with that CL and by playing an HDR video: the primary
framebuffer format in /sys/kernel/debug/dri/0//state correctly changes
to XB30, whereas when there are not HDR quads in the BufferQueue, it's
XR24.

Bug: 958166
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2072621
Reviewed-by: ccameron <ccameron@chromium.org>

Cr-Original-Commit-Position: refs/heads/master@{#745461}
Change-Id: I2700d4d22a6b7e3a91433e6ca5bd042760cf165b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2235977
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776543}