[NativeFS] Refactor NFSManager.
For drag&drop we'll need to factor out
the method responsible for setting permissions
for a NFSHandle. We'll also need to expose the method
responsible for getting a FSURL from a path. This is done
to deduplicate code responsible for
and choosing permissions for SharedHandleStates,
and to facilitate using the same logic when
turning a NFSTransferToken into a Handle.

Bug: 1080811
Change-Id: I783b6f1959e0fa03e0178466be4c07ee9d8776d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2232879
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Hazim Mohamed <hazimmohamed@google.com>
Cr-Commit-Position: refs/heads/master@{#776597}