Ignore CanvasResourceProvider::Clear in OOPR Mode
In OOPR Canvas2D we no longer have access to Skia in the client process.
We need to update CanvasResourceProvider::Clear to not access SkCanvas
in the case of OOPR. A more ideal fix is tracked in
https://crbug.com/1090081 but simply ignoring it is fine for now as
OOPR automatically handles clearing the canvas during initialization.

Bug: 1090083
Change-Id: I5fb75e4fac312ec00b8e5215d3756219672461ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2236193
Commit-Queue: Nathan Zabriskie <nazabris@microsoft.com>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776618}