Portals: DCHECK instead of CHECK that the embedder supports activation
In the case of a bad renderer trying to activate a portal when portals
are disabled, the renderer kill in RenderFrameHostImpl::CreatePortal
would have already prevented this. So the only way this could fail is
due to a programming error within the content embedder. A DCHECK is
more suitable for that.

Bug: 942305
Change-Id: I471bd408ec4969bd7c247390932681388ed4477d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2237651
Reviewed-by: Lucas Gadani <lfg@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776658}