[UI] Menu item view should respect model icon size
Before this CL:
The size of the icon in the menu was hardcoded.

After this CL:
The size in the icon is respected.

In addition it does two fixed.
1- It makes sure the icon_view is set upon calling
SetIcon(const ui::ThemedVectorIcon& icon)
This is to make sure the proper size is taken
into account when computing the menu size and also consistent with
SetIcon(const gfx::ImageSkia& icon)

2- When UpdateIconViewFromVectorIconAndTheme(), it makes sure the
ImageView is populated first before setting using SetIconView().
That's because SetIconView() does invoke
InvalidateLayout() and SchedulePaint()
and hence needs to have the actual icon_view set.

Bug: 1044038
Change-Id: Ic24b53f309a57585121a6bb1c900598d73e1d891
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2232978
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776660}