Remove obsolete ScrollbarAnimationController code
DidScrollBegin and DidScrollEnd are not currently called. Given that
these were supposed to be used for a minor optimization to avoid posting
overlay scrollbar "fade out" tasks while a scroll was in progress, and
there are no behavioral or performance issues to be address, this CL
just removes the dead code and bools that are no longer needed.

There are two non-trivial unittest updates (i.e. ones where we don't
simply remove the calls to the deleted methods):

- AwakenByScrollingGesture starts the fade after a scroll update without
a previous scroll begin.

- MainThreadScrollQueuesFade removes the test case for validating state
when a scroll begin comes in before a scroll update.

R=bokan@chromium.org

Bug: 755576
Change-Id: Idb4f3f67ed5fa52a63f38539b9ca8cbab58e51c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2236037
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#776662}