[NativeFS] Remove most of per-tab usage tracking.
This was only used by the per-tab permission model, so with that gone
this code can be deleted as well. This leaves the most basic "is this
tab using the API" since that is used by page info as well, and it's
not easy to change that code to get usage information some other way.

Bug: 1087205
Change-Id: Ic7ae3335159cecc806e9969a59659a3808fba055
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2237431
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776710}