DPWA: Refactor and rename WebApplicationShortcutInfo
This CL moves shortcut_icon_bitmaps out of WebApplicationShortcutInfo.
This data now lives in a separate
std::vector<std::map<SquareSizePx, SkBitmap>>
shortcuts_menu_icons_bitmaps;

When detecting updates to the manifest, this enables direct comparison
between the WebApplicationShortcutsMenuItemInfo created from the
manifest and the shortcuts menu info persisted in AppRegistrar. This is

The separation of icon bitmaps into shortcuts_menu_icons_bitmaps allows
us to use WebApplicationShortcutsMenuItemInfo in the WebApp class. So
this change also removes the WebApp::WebAppShortcutsMenuItemInfo struct.

Finally, as part of this change, WebApplicationShortcutInfo has been
renamed to WebApplicationShortcutsMenuItemInfo. This clarifies the
feature powered by this struct. It also helps disambiguate it from the
preexisting ShortcutInfo struct in the web_app namespace.

This CL is part 1 of 2 to enable shortcuts menu updates when the Web App
Manifest is updated.

The follow up CL will contain code to detect changes to the manifest's
shortcuts member and to unregister and register the Shortcuts Menu as
part of the PWA update process.

Bug: 1087212
Change-Id: Ib0d0401b0454ecae462acd0bdba2bda6735326d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2229879
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Commit-Queue: Rahul Singh <rahsin@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#776716}