Android: Stop ContentView from using EventForwarder after destruction
When WebContents is destroyed, its ContentView is detached from View
hierarchy to stop receiving UI events from framework. But it doesn't
seem to stop it from processing some of events that were passed before
it was detached. It uses EventForwarder to pass them to WebContents but
the EventForwarder was destroyed as well, therefore can cause UAF
exception.

This CL updates ContentView to check EventForwarder before using it.

Bug: 1080552
Change-Id: I7ab4099461d5906393e4305e4f5a88cf86302fd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2235259
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776717}