Do not decode the first frame when decoding size
Chromium creates some image decoder instances just to get image
metadata, so decoding the first frame when decoding metadata will result
in decoding the first frame of an image multiple times.

It turns out that the AVIF spec says the container "box" or property
with the the image width, height, or bit depth is mandatory. Therefore,
if decoder_->containerWidth/containerHeight/containerDepth is 0 after
avifDecoderParse() returns, we can consider the image as invalid. Then,
the only metadata we need that libavif doesn't expose yet is whether the
image is monochrome (grayscale). This will be addressed in a new version
of libavif.

Partially revert the changes to avif_image_decoder_test.cc in commit
11746b040c64d5c8dcdea1e011dffddfb55274a7 because now Parse (decoding the
metadata only) and Decode are distinct phases of decoding again.

Test:
1. blink_platform_unittests --gtest_filter=*AVIF*
2. Run chrome with --enable-features=AVIF and visit
   https://resources.link-u.co.jp/avif/images.html.

Bug: 1092694
Change-Id: If4d4ea7aa02591ba9269c810931bac7e4ff85895
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2236888
Commit-Queue: Wan-Teh Chang <wtc@google.com>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776719}