OOBE - Extend marketing email country list
Extends the marketing email opt-in in OOBE to support the following
additional countries: France, Germany, Netherlands, Finland, Sweden,
Norway, Denmark, Spain, Italy, Japan, and Australia. All the countries
are protected behind a flag, with a special flag just for double
opt-in countries such as Germany.

Added suffixes to the metrics that are recorded on that screen. Opt-in
rates, and backend request responses will now be based on the country.

Added additional metrics to track the success rate of the current
geolocation approach.

Fixed: 1090724
Change-Id: Icf71e8d07d1b97465649dbd1e5c02c93b50c003b
Bug: 1090724
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2232604
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Commit-Queue: Renato Silva <rrsilva@google.com>
Cr-Commit-Position: refs/heads/master@{#776723}