[AW] Record timediff for non-embedded histograms
This records the timediff for histograms since they are first recorded
using UmaRecorder in a non-embedded WebView process till they are
actually sent to UMA in an embedded WebView.

This required adding a Metadata field to the HistogramRecord proto.

A small refactor is made to allow injecting a delegate to
AwNonembeddedRecorder to allow mocking actions like adding metadata
and service name in testing.

Fixed: 1084757
Test: bin/run_webview_instrumentation_test_apk -f "*AwNonembeddedUmaRecorderTest*"
Change-Id: Id96a3eb9a7d0831a0b8db0a181baf313881f2844
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2226737
Commit-Queue: Hazem Ashmawy <hazems@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Tao Bai <michaelbai@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776737}