Fix the single app folder issue during first time user signs in.
This is how the single app folder is created:
The single app item folder happens when a new user sign on a device
for the first time, if a folder has two apps in it, one of the app
takes less time to install than the other. After the first app is
installed, app list item will be created for the app in app list model
and user will see a single app folder on launcher page. If user removes
it from the folder, the sync item of the app will be updated to be
moved out of the folder, the folder will disappear on launcher page.
Since the second app is not installed yet, it is not in app list model
at this time. Later, when the second app is installed, it is still
associated with the original folder, therefore, it shows up in a single
app folder. On the other devices, when the update of the first app
(moving out of the folder) is synced across, it will also show a single
app folder.

Our fix:
When user removes the first app from chrome app list model and update
its sync item, we check the sync data to see if the original folder
will become a single item folder after that. If so, we update the sync
item of the remaining app by moving it out of the folder and place it
at the same location of its original folder. Later, when the second app
is installed, it will show up at the top level at same spot as its
original folder. Also the change will be synced to other devices and
the second app will be moved to top level as well.

Bug: 1082530
Change-Id: I764fc6c4111f5f76b028b53209ad8ce2abc8ad86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2236993
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Jenny Zhang <jennyz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776738}