[AW] Keep the MetricsBridgeService running while having active tasks
Although MetricsBridgeService has a sync AIDL interface where all its
calls block till they finish, recordMetrics does it job
asynchronously behind the scenes where it posts the actual recording
task to a background TaskRunner. This may result in some rare cases
that the service may be killed by the system (because all clients had
unbound the service) while still having some writing tasks in the
background that haven't finished yet. To solve this:

- Make SequencedTaskRunner static so that all service instances
  created by the system will share the same task runner avoiding,
  possible race conditions writing to file.
- Make the service both start and bound service. The service will
  start itself with the first client binding and it will manage its
  own lifecycle to ensure that it stays running until it has finished
  its writing tasks. Service will not be destroyed by the system until
  all clients unbind and the service calls stopSelf()
  See https://developer.android.com/guide/components/bound-services.

There is no evidence that this actually happens and there is no
reliable way to reproduce this in tests but this is still a potential
bug that should be fixed.

Fixed: 1082572
Test: /bin/run_webview_instrumentation_test_apk -f "*MetricsBridgeServiceTest*"
Change-Id: I5e8e9226379f4b2b38f8b720405f78603ae2b3c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2235597
Commit-Queue: Hazem Ashmawy <hazems@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776763}