[ios] Enables Messages to 100% in Canary, Dev and Beta.
Finch is not being used since the metrics team doesn't think its an
appropriate use of Finch (Which is more aimed at A/B testing) and not
a release tool.

Metrics on NonStable Channels are good since the top crashes are barely
occurring. We hope that this will help catch any of these elusive crashes
(Hopefully with breadcrumbs) before the Stable release.

Bug: 1065385
Change-Id: I49707077afd577b612d739ade4c9b3edaa9b5ecd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238377
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776771}