[M84] Trust Tokens: Require secure contexts for the Fetch interface
M84 note: This is a combination of
1. 30eaece5d10def94069109b1ea689838d1401eb2 (crrev.com/c/2221415)
2. 0211bf0ed86e5c9a9e248e441d029a802fcb28a8 (crrev.com/c/2217855)

To be in line with other new security-relevant Web Platform features, we
only allow callers to execute Trust Tokens operations in secure
contexts. Usually, WebIDL restricts functionality to secure contexts
with the [SecureContext] extended attribute. This is how we initially
implemented the restrictions on the three Trust Tokens interfaces
(fetch, XHR, and iframe). However, this doesn't work for the fetch
interface, because the [SecureContext] extended attribute doesn't work
on dictionary members (like the `trustToken` member in
request_init.idl); instead, it silently no-ops! This change:

1. removes the useless SecureContext attribute from RequestInit;
2. adds an explicit check that a Fetch request bearing Trust Tokens
parameters is coming from a secure context, throwing a TypeError if not;
and
3. for good measure, adds tests ensuring the XHR and iframe interfaces
are only available in secure contexts, too.

R=csharrison, yhirano

Fixed: 1087200
Change-Id: I833c210834142eac260e401a32272b3873e68a12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238489
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/branch-heads/4147@{#569}
Cr-Branched-From: 16307825352720ae04d898f37efa5449ad68b606-refs/heads/master@{#768962}