Propagate delegated ink metadata into viz
Continuing to propagate the required metadata for a delegated ink trail
further down the pipeline into viz via the compositor frame metadata
and storing it on the display, for now.

This CL also transforms the metadata (specifically, the point and
presentation area) into the root target space. This is done so that they
are in the correct coordinate space for when they are eventually drawn
onto the buffer prior to being swapped onto the screen.

Bug: 1052145
Change-Id: I18544ee0f8b1d808669ab7ea5aa4e97792e43e3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183328
Commit-Queue: Mario Bianucci <mabian@microsoft.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#776830}