[WebLayer] Fix weblayer browsertests which use dynamic attributes
WebLayer browsertests load the WebLayer implementation from the same APK
the client is running in. This confuses Android and it no longer
resolves dynamic attributes correctly, since a shared library APK is not
expected to both be run as a normal APK and loaded as a shared lib. To
fix this, switch to using app_as_shared_lib instead of shared_resources,
which is what Monochrome uses to run both as Chrome and WebView.

This was exposed by Colin's infobar CL which must be the first
browsertest to hit a piece of UI using a dynamic attribute:
http://crrev.com/c/2232555.

Note that passing shared_resources to the "test" gn rule was added
earlier for WebLayer, so switching to app_as_shared_lib shouldn't affect
anyone else: http://crrev.com/c/1864517.

Bug: 1025620
Change-Id: I95ec5500cdcf52167dfedffc9dbe79071895dc37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238706
Auto-Submit: Clark DuVall <cduvall@chromium.org>
Commit-Queue: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776843}