web-ui: Add kChromeUIUntrustedScheme to HasWebUIScheme()
As precursor to adding support for chrome-untrusted:// to
WebUIController we add chrome-untrusted:// to HasWebUIScheme().

After auditing uses of HasWebUIScheme() it seems fine to add
chrome-untrusted:// to it.

After calling HasWebUIScheme(), ChromeWebUIControllerFactory only
compares hosts when looking for the WebUIController for the URL. So
this change causes ChromeWebUIControllerFactory to return
WebUIControllers for chrome-untrusted:// URLs with the same hosts as
chrome:// URLs. To avoid this we early return if the URL's scheme is
chrome-untrusted.


Change-Id: Ic63ccd9fc1c9066b79957af55ffcaedb27be9cc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2223709
Commit-Queue: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776847}