[Extensions] Re-enable a declarativeContent API for on Windows.
The UninstallWhileActivePageAction test was disabled for flakiness
on Windows. It turns out this test was flaky on multiple platforms,
because it would occasionally try to execute a JavaScript function
defined in the extension's script before the script was finished
initializing. The simple fix is to use TestExtensionMessageListener
instances to wait for the extension script to finish initializing.

Bug: 515717
Change-Id: Icc8e3accbbd6bf5e47524b1b34501eba2fe0fd41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2229416
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776851}