[files-progress] Background implementation of transfer speed details.
It adds a Speedometer to Task and get updated when new progress update
comes to the task.

The average window is set to 20, based on some experiment update rate
is not consistent but it showed about per sec, so the average is for
about 20 seconds. It would be updated through more experiments.

It only works with CopyTask as the task types like Zip or Move don't
report processed bytes.

The following new fields dispatched to FE as part of task status:
 - currentSpeed
 - averageSpeed
 - remainingTime

All the times unit to pass to FE is in second and speed is in B/s.
FE supposed to polish it how to arrange magnitute.

Bug: 953308
Change-Id: Idccb0385dc10ff5a6cee83c55f49b1cc0b39da5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2217921
Commit-Queue: Omid Tourzan <oto@chromium.org>
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776863}