Introduce placeholder target for absl component.
This placeholder target (that will become a "component" in a follow up
CL) will make it possible to switch both WebRTC and OpenScreen to
depend on //third_party/abseil-cpp:absl (see for example [1]).

This will then allow to create a CL to transform
//third_party/abseil-cpp:absl into a "component" without having to
change anything in WebRTC and OpenScreen.

[1] - https://chromium-review.googlesource.com/c/openscreen/+/2173190

Bug: 1046390
Change-Id: I7870826edd2a34765bcdbd6a9adb0fe9c20c1dd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2234258
Commit-Queue: Mirko Bonadei <mbonadei@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776882}