[chooser] Blocklist FileChooser for BackForwardCache
Currently, the FileChooser window is not explicitly closed when
navigating away.

This CL disables BackForwardCache when RunFileChooser is invoked
to avoid any renderer interactions when the page is in
BackForwardCache and user seeing any unexpected behaviour.

To make FileChooser bfcache compatible is is the next step.

BUG=1071232

Change-Id: I219143190425107a7604f2f23dd25c5e462d7106
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2193932
Commit-Queue: Sreeja Kamishetty <sreejakshetty@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776908}