[content] Remove RenderFrameHostImpl::is_active
RenderFrameHostImpl::is_active is used either when the RFH is in
pending deletion or in back forward cache.

We would like to remove all the |is_active| occurrences to avoid
confusion with all the other methods and use specific methods like:

- RenderFrameHost::IsCurrent()
- RenderFrameHostImpl::IsPendingDeletion()
- RenderFrameHostImpl::IsInBackForwardCache()
- RenderFrameHostImpl::IsInActiveAndDisallowReactivation()

This patch updates the callers of |is_active| and make use of above
mentioned methods along with removing RenderFrameHostImpl::is_active().

Detailed list of investigations for each is_active usage is
presented in this spreadsheet:
https://docs.google.com/spreadsheets/d/1TTzZ5r0m7nHBS_TkrEAdcFGs0M2iXO_tS0-lJmbQ3EY/edit?usp=sharing

BUG=1073449

Change-Id: I3be34f9ac66389edb36f53a73fc806f2d9d942a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2186641
Commit-Queue: Sreeja Kamishetty <sreejakshetty@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776924}