Fix to use noised intervention type for Heavy Ad reporting.
Currently if we would trigger a network intervention but it falls below
the noise threshold for an intervention, and then later we do trigger
a different type of intervention, the intervention is reported as a
network intervention.  This change fixes the reporting to use the
noised intervention type for reporting, rather than the non-noised.

Bug: 1092962
Change-Id: Ia88838e980fab369b3eb9f5af2204a6f94eecd72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238455
Commit-Queue: Eric Robinson <ericrobinson@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776934}