[privacy_budget] Introduce IdentifiabilityStudySettings.
On the renderer there needs to be a method for determining whether the
identifiability study is active. In addition to being a performance
tweak -- no need to generate samples that would never be used -- this
facility can also be used to turn down misbehaving samples.

Bug: 973801
Change-Id: I717f3abc4fbda1c97f5d6518c4f1f7d1ca174c66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225537
Commit-Queue: Asanka Herath <asanka@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776944}