[fuchsia] Teardown web Components if the top-frame Renderer crashes.
fuchsia.web.Frame instances will now transition to the PageType::ERROR
state if their top-level renderer process crashes.  This allows the
caller to choose whether to reload the Frame contents, navigate to a
different URL, or to teardown the Frame.

WebComponent listens for navigation events from its Frame, and signals
component exit with return-code ZX_ERR_INTERNAL if the Frame enters the
ERROR state. This results in the component destroying itself if the
web content running in it crashes, rather than being left live but
non-functional.

The WebEngine's NavigationController is also simplified with a single
function which calculates the current NavigationState based on the state
of the visible navigation entry, and related fields.

Bug: b/158339587, 1092506
Change-Id: Ic6c0ddab9088382637872e02afe291217cd3b2d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233442
Commit-Queue: David Dorwin <ddorwin@chromium.org>
Reviewed-by: David Dorwin <ddorwin@chromium.org>
Auto-Submit: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#776962}