[Logging] Move bubble type change to OnBubbleClosed()
crbug.com/1090771 is because of a corner case where the offer bubble is
accepted, but the card saving fails. At this time if a navigation
happens, the previous_bubble_type_ is still BubbleType::LOCAL_CARD but
the bubble is now actually BubbleType::FAILURE, so the callback is null
already. This is a new corner case introduced by my previous change.

I did not simplly do a null check in my fix but remove the local var
|previous_bubble_type_| as I found it confusing and hard to track.
And thus some related changes are also required.

Bug: 1090771, 1070799
Change-Id: Id789f5a749009c26d2165b912c9f673676e5dacd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2236958
Reviewed-by: Jared Saul <jsaul@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Siyu An <siyua@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777015}