Adding device policy MinimumChromeVersionEolMessage
This CL adds a new device policy MinimumChromeVersionEolMessage for
configuring return instructions from the device administrator if the
device has reached end of life and is blocked for further use by policy
MinimumChromeVersionEnforced. These return instructions are shown on
the EOL dialog of the update required screen which blocks the user
from signing in.

Bug: 1048607
Change-Id: I6fb8fa39ddb974f39c6289c04ea80868429fea3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2235818
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Cr-Commit-Position: refs/heads/master@{#777022}