Android: Output lint errors when not failing build
Lint error messages are stored in the exception object that is raised.
During default local builds lint errors do not fail the build, and so
the exception is not re-raised. This caused the error messages to be
silently dropped. Thus we now explicitly print the exception object if
it is not re-raised.

Bug: 1093345
Change-Id: I789e1f61627bf598dee8ccb24ced43719a60db91
Fixed: 1093345
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2239680
Commit-Queue: Peter Wen <wnwen@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Auto-Submit: Peter Wen <wnwen@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777040}