[LargestContentfulPaint] Implement LCP-including-removal
This CL implements experimental candidates for largest image and largest
text where a candidate is only updated when something larger is seen.
This means that the LCP computed off of those candidates will be the
LCP-including removals. For a largest image that is still loading, the
candidate is immediately replaced when the larger image is found, but we
still need to update the paint time whenever it first becomes available.
This is consistent with how we treat a largest-loading-image in the
current LCP implementation. The experimental values will be plumbed to
the browser process, where they will need to be processed similarly to
the current LCP to obtain an experimental LCP which will be logged in
UMA/UKM.

Bug: 1045640
Change-Id: Ie6d8ed34cb1ea565c0443a0a628086a25742ae3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2236594
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777055}