TabStripModel: Fix UAF by preventing openers with dangling reference.
TabStripModel::FixOpeners(index) is called to ensure that no
WebContentsData has its opener set to the WebContents at the given
|index|. It does this by setting the opener of any tabs that reference
the tab at |index| to that tab's opener. However this can cause a tab to
be its own opener (example consider two tabs which are the openers for
each other).

If such a tab (a tab which is its own opener) is removed, the same
FixOpeners logic would cause the opener of any tabs referencing it to
point to a dangling pointer. This can lead to use-after-free errors.

Fix this by ensuring that a tab can't be its own opener. Also add some
DCHECKs checking the same.

BUG=1092308
TEST=Ensure the test extension in issue 1092308 doesn't crash under
ASAN.

Change-Id: I90d8b5342de96400dee2d502868273c6965937c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238557
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777127}