[ReaderMode] Update settings header and styles to h1
It changes the <h2> title of the settings dialog to a <h1>. Although the
article title is marked up as a <h1>, this is not accessible to users
once the settings dialog is open. This means that when the dialog is
open, the only heading exposed is a <h2> which breaks semantic ordering.

Note: This patch has no visual change.

AX-Relnotes: N/A

Bug: 1087939
Change-Id: Idce691938e5ad8a71e897c679dd75ada5f60dc3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2237497
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Commit-Queue: Meredith Lane <meredithl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777170}