Only hit test caption buttons in browser layout if they in top container
This fixes a bug where in full screen mode it was not possible to
interact with elements in the extreme upper right of the window, because
the caption button container was still there but hidden and the browser
view layout did not know to exclude the hit test for them despite not
actually owning the caption buttons (they were still on the frame).

Now the browser layout only hit tests caption buttons if they are part
of the top container rather than the frame.

This fixes a specific issue where it was not possible to close Dev Tools
using the mouse when the browser was in fullscreen mode (see attached
bug).

Bug: 1093205
Change-Id: I8798db65be56bad1fd9a32f4a54867d0a15482ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2240036
Reviewed-by: Connie Wan <connily@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777254}