Restrict sanitization of visible input values
This patch makes it so we only sanitize the text in input elements when
the value attribute (defaultValue) when the default value is actually
being used.

Before this patch, the visible text entered by the user is sanitized
in-place every time the value attribute (for the default value) is
modified, even though we aren't using the default value anymore.

This patch will check the dirty flag, which signals if we are using
the default value attribute or not, before sanitizing the visible value.
This way we will still apply sanitization to default values provided
programmatically when they are actually used, but won't change text a
user entered.

Bug: 1067117
Change-Id: Ic185d6d173d40c1b0726c0c3fb67a6937fc8534f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2133229
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777263}