Call Compositor.BeginFrame() in AdTracker test to add stylesheet.
With BlockHTMLParserOnStyleSheets enabled, adding the completed
stylesheet would happen on parsing finished without a BeginFrame() which
made this test fail with that flag on. Adding BeginFrame() here shows
that a @font-face rule in a stylesheet will notify the AdTracker about
the src url being fetched before it's actually fetched. It's quite
likely that is a bug, but BeginFrame() is anyway correct to make sure
the loaded sheet is active.

Bug: 891767
Change-Id: I206c468dbde97a0a25bfcba5d169954a333124f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2239569
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777290}