fixup: ResumeParsingAfterPause's PumpTokenizer behaviour
Several possibilities entering this function:
1) We're parsing with kForceSynchronousParsing (no background parser,
   tokenizer_ defined)
2) We're parsing with kAllowAsynchronousParsing without a background
   parser started (no tokenizer_)
3) We're parsing with kAllowAsynchronousParsing _with_ a background
   parser started (no tokenizer_)
4) We're parsing with kAllowDeferredParsing (with tokenizer_)

The previous code called PumpTokenizerIfPossible _only_ if tokenizer_,
wasn't defined, so the implication is that this operation should be
done synchronously and only if the background parser isn't being used.
The new code either pumped the tokenizer immediately, or deferred
the pump. This CL restores the previous behaviour.

Bug: 901056, 1087032, 1087325, 1087753
Change-Id: I7ed8ee0eea91eb7786e07a82b399f40f0c651dbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2235836
Commit-Queue: Richard Townsend <richard.townsend@arm.com>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777296}