[CrOS Settings] Don't add parent result with same name as child
Before this change, it was possible to show two identical entries (e.g.,
"About Chrome OS", "About Chrome OS"). This CL compares the search
result text and opts not to show a parent result if the strings are
identical.

This CL also updates the search tag for "Languages and input" so that
it does not include "settings". This matches the behavior for other
search tags, so it was an oversight that this was included in the first
place.

Fixed: 1093136
Change-Id: I535662d2eb1c5e94ea7eef66e218932ad7901c25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2240406
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777387}