Update action for Assistant timer notifications.
In v1, clicking a timer notification causes removal of the timer.
In v2, clicking a timer notification is a no-op.

Previously, an empty action for an Assistant notification was used as a
signal to indicate that notification payload should be retrieved from
the server.

Now, this is only the case for server notifications. An empty action for
a client created notification will simply result in a no-op on click.

Bug: b:149570650
Change-Id: I48258edb97c891e302e1085f5ed3bc019cfa9340
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238953
Commit-Queue: David Black <dmblack@google.com>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777410}