[WebLayer] Add site isolation on password sites for WebLayer
This change adds logic to WebLayer to detect when a user is typing a
password into a site so we can start isolating that site. This also
handles saving those sites to a pref and the restore/delete logic
related to that pref.

In chrome, the dynamic detection is done by
ContentPasswordManagerDriver, which also does many other things to
support chrome's password autofill. WebLayer uses Android Autofill
on Android, so really only needs the one callback notifying the
user has typed into a password form. To do this, WebLayer has a minimal
implementation of PasswordManagerDriver, which only handles that event.

The browser tests added for this change are heavily based on the site
isolation related tests in chrome_navigation_browsertest.cc.

Note that there is another bug to design a long term solution for site
isolation that will ideally include sharing this logic in a better way
with chrome: http://crbug.com/1088446.

Bug: 1058364
Change-Id: I62615af273b37843450366d3e05676ce228ed8fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2239030
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Maxim Kolosovskiy <kolos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777420}