[serial] Combine dialin and callout devices on macOS
This change modifies the serial port enumerator on macOS to generate a
single entry for each serial device so that the user is not presented
with duplicate options in the Serial API permission prompt.

Unlike other platforms macOS provides up to two devices paths for each
serial device. These paths represent the "dialin" and "callout" behavior
for the port. For most devices the "callout" behavior is preferred,
while the "dialin" behavior will cause the system to block while opening
the device until the "Data Carrier Detect" signal is detected. This
behavior was useful for writing applications that wait for a modem to
receive a phone call (hence "dialin") but is largely irrelevant today.

The ability to select the path for the "dialin" device is retained in
the chrome.serial API and the design of the change allows for selecting
this behavior in the Serial API if developers demonstrate a need. This
will require implementing support for this behavior on Linux and
Windows, which do not provide separate device paths.

Bug: 1084285
Change-Id: I290ffa0b58a096fe0a957db8f7d8c1210fa664e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2227690
Reviewed-by: James Hollyer <jameshollyer@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777429}