wm: Reland improve cross fade animation.
Reland crrev.com/c/2224990, including a fix for failing builders and
a fix with tablet screenshot related.

Speculative fix crrev.com/c/2218557 was not enough. I was able to repro
locally and confirm this fix this time. It seems with certain build args
we can get into a state where the CrossFadeObserver which observes window
can be not animating, which is strange as the observer observes window
and deletes itself on animation end.

It seems in that test and bot, it can happen perhaps OnWindowDestroying
is called before the window starts animating, or some other edge case.
In this case, make sure the window is deleted if not animating.

Test: none
Bug: 1088169, 1088039
Change-Id: Ibe035209e83400921f16cff2bb272a8e186934dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2224023
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777446}