Strip ANGLE revision off driver version
Strips the revision that's at the end of some ANGLE driver versions, as
it changes frequently and can mess with inexact matching since each
revision is treated as a separate trace in Gold.

Also reports the driver vendor in the data uploaded to Gold.

Bug: 1078914
Change-Id: I745993d8e7318c991cac2625718d5fd24b277a79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241996
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777489}