[TTS] Fix select-all after panel opened
The recent change to clear the selection after the user has opened
and closed the panel needs to check if the selection is invalid.
In that case there's no point in clearing the selection and might
clear the user's selection that we cannot handle (like long selections).

Also mark the change reason generated by closing for an invalid
selection.

BUG=1093271

Change-Id: I3c244227b2346b92e936807e5aea7f8f0bdb0d98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2242118
Commit-Queue: Theresa  <twellington@chromium.org>
Auto-Submit: Donn Denman <donnd@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777540}