PDF Viewer: Separate stand-alone and Print Preview strings.
A large portion of the UI is only present in the stand-alone PDF Viewer,
and not in Print Preview. This CL makes it such that only strings that
are used in each case are registered for $i18n{} replacements, as
opposed to always registering all of them.

Note that chrome.resourcesPrivate.getStrings() still returns
all of them, since there is no way currently to differentiate between
the two contexts.

Bug: 1091516
Change-Id: I6d12ef310dd74112cd9cc68216b260f0e2c0652e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2240017
Commit-Queue: dpapad <dpapad@chromium.org>
Auto-Submit: dpapad <dpapad@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777548}