Replace CreateDXGIFactory() with CreateDXGIFactory1() on Windows
We saw a big number of crashes in CreateDXGIFactory() on Windows 7. Try
CreateDXGIFactory1() and see if the number of crashes will drop.

According to the note from MSDN
https://docs.microsoft.com/en-us/windows/win32/api/dxgi/nf-dxgi-createdxgifactory1
Do not mix the use of DXGI 1.0 (IDXGIFactory) and DXGI 1.1
(IDXGIFactory1) in an application. Use IDXGIFactory or IDXGIFactory1,
but not both in an application.

Bug: 1048743
Change-Id: Id1fb76ceb8268751b8791718d318e093277f47cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241950
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777584}