Separate pipeline for window activation and frame activation.
Prior to changing how the "always paint as active" widget logic worked,
Widgets could respond to a Windows WM_NCACTIVATE message by changing the
visual state of their non-client area (i.e. window frame) independent of
the window's activation state. This allowed for the window frame to
flash when the window needed attention.

This behavior was broken by
https://chromium.googlesource.com/chromium/src/+/35976c71059e388d212cef4f83cf8dc9c2b6f530
which consolidated a lot of the logic but cut out the WM_NCACTIVATE
handling.

This CL reintroduces the concept of "frame-only activation" through a
separate call path in the event handling functions on
DesktopWindowTreeHost*, NativeWidget, and Widget. Currently, the path is
only used on Windows to correctly respond to WM_NCACTIVATE with a
visual change, allowing window flashing when we need to notify the user
of something (as is platform standard).

The standard "Activate" event also changes non-client activation, as it is
expected to on Windows as well (and must on other platforms that do not
have a separate non-client activation message).

Bug: 1088416
Change-Id: Ic95c6fa878f42a919fbc6ff2e6ce9980c9f6fb94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2239979
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777589}