concierge_client: Refactor Observer methods
This class has somewhat confusing behaviour currently. The
ConciergeServiceStopped is called only when the NameOwnerChanged
signal indicates that the concierge service has stopped without
replacement. ConciergeServiceRestarted is called under any other
circumstance.

This means that if DBus sends a NameOwnerChanged signal with both old
and new owners set, it won't trigger a ConciergeServiceStopped call,
even though the service did stop. The crostini stability metrics work
needs a way to be notified when the current concierge instance stops
regardless of how soon afterwards it comes back, so this should
change.

In practice, DBus doesn't appear to actually send the problematic
signal, so this should be a no-op, but it's good to clarify the exact
expectations.

This CL also renames ConciergeServiceRestarted to
ConciergeServiceStarted to better reflect its semantic.

Bug: 977201
Change-Id: I86ee9ad6d6640e66cbab42b1efc7e16179f1b91e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241196
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>
Cr-Commit-Position: refs/heads/master@{#777602}