Rebaseline some perf stories broken by a soon-to-land change
We're enabling the content::features::kReducedReferrerGranularity flag
by default shortly. It changes the value returned by a commonly used
Web API (document.referrer) and a corresponding HTTP request header.
Landing a field trial testing config triggered a number of pinpoint
"regressions" (and "improvements"!) that erikchen@ hypothesized were due
to test framework interactions with the existing WPR recordings. This CL
rebaselines the affected stories:
- browse:news:nytimes:2018 (desktop)
- browse:news:cnn:2018 (desktop)
- load:games:bubbles:2019 (desktop and mobile)

The first of these uses a CSS selector that includes the current year,
so it wasn't possible to rerecord it in its current form; rather, the
change updates it to a 2020 analogue. For the second and third stories,
recording again in their present form worked.

Bug: 1056329
Change-Id: Ic02b84584b3dcd177ce4db8ed12d25194ccb2b85
Fixed: 1058915, 1048550, 1050493, 1051294, 1050491, 1051451, 1056641
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2227109
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777612}