Preserve CurrentlyScrollingNode in unthreaded mode
The CurrentlyScrollingNode is tracked only on the active tree of the
compositor thread and specifies which node is currently latched to and
which node scroll (from GestureScrollUpdate) will be applied to.

This node designation is stored on the PropertyTree but it should be
preserved across property tree updates. Indeed, when a property tree is
pushed from the pending to the active tree, there is existing code in
LayerTreeImpl::PushPropertyTreesTo that ensures the currently scrolling
node isn't reset when the tree is overwritten.

However, in single threaded mode (web tests only if I understand
correctly), the main thread LayerTreeHost commits directly to the active
tree so the active tree is written from
LayerTreeHost::PushPropertyTreesTo which doesn't have the above code to
preserve the currently scrolling node. This was ok until now because
scrolling is never handled by LayerTreeImpls in single threaded mode,
it's always forwarded to the main thread.

However, this changes in scroll unification, scrolls are expected to
flow through the compositor gesture handling code even when a repaint is
required on scrolling so the main thread handlers are going away. This
necessitates handling the scrolls in the compositor even in single
threaded mode. This requires us to always preserve the currently
scrolling node when overwriting the active property tree.

Bug: 915926
Change-Id: Ie85f48ea74d30a47eebdef6ad666246131161af7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2240951
Reviewed-by: Xida Chen <xidachen@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777629}