[b4p/Settings] Add confirmation dialog for moving a password
Instead of directly moving the password to the account when the
"Move to your Google Account" option in the overflow menu is clicked,
a dialog is now shown to the user that allows them to confirm or cancel
the action. We also use the occasion to remove some unused dependencies
from PasswordsListHandler and fix a typo in the name of a property.

Adding the icons/avatar to the dialogue is left for future CLs.

Screenshot of implementation:
https://screenshot.googleplex.com/JigAPMx7jw4

Screenshot of mock:
https://screenshot.googleplex.com/iNV6JAtE9t9

Bug: 1090372
Change-Id: Ieb588aec2a000ae7a9d9bac8fcb22ccbb65eadeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2235850
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777719}