Switch camera app's "camera roll" to chrome://media-app with a fallback to Gallery.
Currently code exists to substitute chrome://media-app when Gallery is
requested by the camera app, so the change is just to make that request
explicit and more robust. (Actual behaviour is not changing).

Context: The Gallery app is going away. Its replacement is not a Chrome
App. There are some gaps in the "Web" app infrastructure (e.g.
https://crbug.com/1079065), but they do not affect the camera app.

This CL switches the openGallery() private API call over to MediaApp.
This was already achieved in effect, but the approach made it hard to
roll out the new infrastructure gradually due to special cases.

By having the Camera App ask for the Media App instead of the Gallery,
the only special case that needs to be dealt with at this layer is
"MediaApp not installed". Files App task determination deals with the
other special cases already.

Bug: 1030935
Change-Id: I7e643c0cf6f7f79c909a689e67b8d14db758520a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2215764
Commit-Queue: Trent Apted <tapted@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Wei Lee <wtlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777758}