[css-grid] Rename SmallestTrackStart to ExplicitGridStart
The SmallestTrackStart method used to return a signed int which was
never positive, representing the smallest untranslated start among all
grid items, clamped by 0 as a maximum.

In practice, though, what we usually want is the index of the first
explicit track/line among all tracks/lines, or equivalently, the number
of leading implicit tracks/lines.

That number is precisely minus the return value of SmallestTrackStart.
Thus, in grid layout there were various std::abs() to invert the value.

But it's a bit confusing to have an API which returns a non-positive
integer, which then needs to be made positive. Therefore, this patch
renames SmallestTrackStart to ExplicitGridStart, and makes it return
a positive size_t directly.

This patch should have no effect in practice.

TEST=GridTest.EmptyGrid
TEST=GridTest.ExplicitlyPositionedChild
TEST=GridTest.IntrinsicGrid
TEST=GridTest.SingleChild

Change-Id: Ie59a187ae29e41d5cf1030c39510aed0d6d27694
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241813
Reviewed-by: Javier Fernandez <jfernandez@igalia.com>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/master@{#777768}