chromeos: optimize for size by default
This is the second part of the overarching go/cr-os effort, and the
minimal change required to confidently optimize Chrome for CrOS for size
without regressing performance.

All benchmarks (CWP, shortlist @
docs/speed/benchmark/benchmark_short_list.md, CrOS critical user
journeys) indicate that this change is largely performance neutral.
Speedometer2 sees a regression of up to 0.7% (with some individual
metrics improved, depending on arch); the largest individual regression
I see is on a CWP benchmark, which is ~2%.

In return, Chrome for Chrome OS gets around 14% smaller, which is
multiple release cycles worth of growth budget.

This change was discussed on
https://groups.google.com/a/google.com/g/chromeos-performance/c/nGRhybO2Kxk.
Everyone seems on board with it.

Bug: 908661, 1018282

Change-Id: Ife644db24b0c105df4b7f8faec2235c2a745fb79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2211001
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777860}