Update more WCv0 tests to WCv1
This CL contains changes only to the fast/dom/shadow folder, and it is
mostly the result of a mechanical replacement of:
  '.createShadowRoot()' --> '.attachShadow({mode: 'open'})'
plus a manual replacement of 'content' with 'slot'. Some manual
tweaking was involved, and a few tests (which turned out to be V0 only
tests) were moved to the web-components-v0-only folder, which will
eventually be deleted.

Bug: 937746
Change-Id: I68c81e0e04211fcefbf09b8c5ac7e4669b58cfe2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2240791
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777862}