MacPWAs: Fix popup windows
At some point, popup RenderWidgetHostViews stopped being told
which process they should appear in. Fix this.

Note that popups are strange in that the RenderWidgetHostView
doesn't have a parent WebContentsView, rather, it is just
the content view for the NSWindow.

Add a sentinel remote_cocoa::kInvalidNSViewId to specify to
RenderWidgetHostViewMac::MigrateNSViewBridge to indicate
that we shouldn't mess with the app-side NSView hierarchy.

Bug: 1091179
Change-Id: I905877e7b306c9f648734e4a433859f7d636356c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2242851
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777880}