Remove unacceleration fallback for hidden canvases
The workaround for crbug.com/600386 (WontFix) can be removed, along
with removing software_rendering_while_hidden and logic around the
fallback mechanism from accelerated gpu to cpu.

The software_rendering_while_hidden field is only set to true when the
fallback mechanism is enabled and when a hidden canvas wants
acceleration. Since the mechanism is disabled, the field is never true.
The field is only used in conditional statements to check the truth of
the value, therefore it's safe to remove all instances of it.

CANVAS2D_BACKGROUND_RENDER_SWITCH_TO_CPU was used to disable the
fallback from gpu to cpu for canvases in the background as it was set
to always be 0. After the removing the fallback mechanism, it is safe
to remove this constant as well.

Bug: 1090180
Change-Id: I27f0819a00e63d6ffc9cc0791e1f11dbbbe50837
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238364
Commit-Queue: Teresa Kang <teresakang@google.com>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777881}