webauthn: don't use invalid algorithms to trigger a timeout in tests.
Several tests are setting unknown algorithms in order to trigger a
timeout for testing. This presumably dates to back when the virtual
authenticator couldn't produce valid responses. However, unknown
algorithms shouldn't produce timeouts because the virtual authenticator
acts like button presses happen immediately, so timeouts mean that a
request is getting dropped. (Which is what is happening and is the
motivation for this patch series.)

This change removes all but one instance where this timeout trick is
being used and deduplicates some tests too.

(Test-only change.)

BUG=1094066

Change-Id: I54ac9dbe649deb6d78f1bfd3e98b105e4145899b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2242531
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#777897}