Convert TabGridDialog to use new ScrimCoordinator
Now that http://crrev.com/c/2119188 has made TabGridDialog use the
ScrimView from RootUiCoordinator, this CL migrates the component to use
the new ScrimCoordinator to control the scrim view. To achieve that, we
expose a getter of ScrimCoordinator in RootUiCoordinator and pass it
down through constructor.

Bug: 1064140
Change-Id: I8f629fd2cab6f3fa8a6236fe58f982e637a994a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2223458
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777908}