[Nickname Management] Wrap long identifier in the Upstream bubble
Changed only the Upstream bubble in this CL to get suggestions on the
approach. If this is okay I will apply it to other places.

So after this change, the long identifier will wrap to two lines, and
the long nickname will get cut off.

Uploaded screenshots in bug comment #27

Bug: 1082013
Change-Id: I5677b51d61aa54fcf8c01b088d4ac2e5f80dda69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2240312
Reviewed-by: Jared Saul <jsaul@google.com>
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Siyu An <siyua@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777915}