[Android][Test] Destroy mSetActivity in ChromeActivityTestRule
This CL destroys mSetActivity in ChromeActivityTestRule. Prior to this
CL, the teardown of ChromeActivityTestRule only ends the tasks in
ActivityManager. As mSetActivity.finish() is not called before the
teardown, no Activity.finish() task is submitted to ActivityManager.
Thus we still have the ChromeActivity unfinished after the teardown.

Bug: 1083138
Change-Id: If0ee66770d66ef451fcb0188907ba759d501291f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2234517
Commit-Queue: Alice Wang <aliceywang@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777920}