CrOs: enable HDR by default
This CL flips the "enable-use-hdr-transfer-function" so that it's
enabled by default. This only affects ChromeOS. It also affects the
"standard" brightness of the screen in HDR-supporting models, making
it dimmer so that the transition SDR-HDR is not noticeablee. It
also takes the chance to tweak a bit said brightness.

Test: kohaku and codepen.io/miguelao/full/PooMbbx
Bug: 958166
Change-Id: I56f8f8f81b90ac58f139f7a54085ca11325ce0a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241946
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777944}