Adding the FAST_HOST_HINTS optimization type.
This is a bloom filter over fast hosts.

Also updating PerformanceHintsObserver to check this type if no hint was
found in link hints or page hints.

Apologies for the large change count in the unittests. This is mostly
due to rearranging tests and adding a using statement for ::testing:_.

Bug: 1087123
Change-Id: Ie25145f28354a17955b16b87ebb6c36601b23ba4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238357
Commit-Queue: Josh Simmons <jds@google.com>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Megan Jablonski <megjablon@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777948}