Set framework for visualizing Android dependencies
Rough code, mostly copied from wnwen's visualization, as a framework for
future visualization work. It is forecast that this code will be
restructured heavily in the future, hence the lack of more thorough documentation.

npm is used for this project, please read the README on how to run
locally. The dependencies so far are:
- ESLint
- d3 (however, as we continue to identify what is required, I plan to
break this up so we only import the v4 microlibraries we actually need)

The page is currently served the JSON file via a small Python script.
This is easy to use for development, but will hopefully be replaced with
either a user file upload or being served from somewhere else.

Bug: 1093962
Change-Id: I8d35ae2cde64581ff7d22fb06a29cd850e42cc86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2242171
Commit-Queue: James Long <yjlong@google.com>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777955}