[Offline indicator v2] Create Animators using current colors
Since we started caching the animations to clean-up on destroy, we've
been keeping the Animator instances once they're created. This ignores
the colors passed in #show and #updateContent, also keeps using an old
status bar color.

The fix is to simply remove the if null checks to create the Animators.

Bug: 1094373
Change-Id: Ibf6de0c66626f3a9fe7f58214d765d18696530f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2243493
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Cr-Commit-Position: refs/heads/master@{#777977}