[WebLayer] Fix Site Settings dialogs on newest AndroidX version
Site Settings dialogs crash on open with newer versions of the AndroidX
fragment library. This is due to them expecting the LayoutInflater
returned by the FragmentHostCallback to not have a Factory2 associated
with it yet. To get around this, this CL makes FragmentHostCallback's
onGetLayoutInflater method clone the Context's LayoutInflater.

This CL also prevents onStart from calling dispatchActivityCreated more
that once in case the Fragment gets stopped and restarted.

Bug: 1085553
Change-Id: I8b392fdd6b6c8231470698e76d90eae07734a818
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2244094
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778040}