[serial] Convert web_tests to standalone window tests
By using the Web Platform Tests framework to autogenerate the HTML
boilerplate which must surround these tests they now become pure
Javascript files. This is nice because they can be formatted with
clang-format and are easy to convert into multi-global tests when we
want to exercise support for this API in workers.

Bug: 884928
Change-Id: I5ea402835afe0af043e1676d7c70ff4ec7231e64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2244117
Commit-Queue: Victor Costan <pwnall@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778059}