chrome-apps: Remove expired "Apps.Window.Type" histogram
The histogram has two values "default" and "panel". "panel" is
deprecated and code seems to indicate that it's not being used.

Additionally 99.99% of the calls are being done with "default"
already, so there is no more value in measuring the number of
calls with the deprecated value.

Bug: 1052902
Change-Id: I74a7ffbd0d5a24cdc5096e62fa066caa03862342
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2237503
Commit-Queue: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Ben Wells <benwells@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778160}