Fix ShelfItemDelegate::ItemSelected for app shortcuts and windows
When an item with multiple windows was being selected,
AppShortcutLauncherItemController would do nothing, expecting that a
context menu with available app windows would be shown to the user, so
they can select the window to activate. Though, this would happen only
when launching the item from shelf. For other sources, the launch would
silently fail. This fixes the logic to activate an active window when
launching the item from sources other than shelf.

On the other hand, AppWindowLauncherItemController would always activate
an existing window (if one exists). This updates the logic to match the
logic for other delegates - when launching from shelf, do nothing, and
let the shlef show a context menu letting the user select a window to
activate.

BUG=1091575

Change-Id: Ib89d186f72d7b0e9926ece9caa6a77e25fedbe39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2244118
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778163}