Sync Keyboard config with VirtualKeyboardFeatures policy in Web Kiosk
We want to control the virtual keyboard features by a specific policy in
the Web Kiosk. For the other Kiosk types and sessions, the current
behavior still remains.

If the policy is not specified, all virtual keyboard features are
disabled by default in the Web Kiosk.

Bug: 1083893
Change-Id: I408cbf1b5463df043445a7ed9936c7dbb5bbe601
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2232319
Reviewed-by: Anatoliy Potapchuk <apotapchuk@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Anqing Zhao <anqing@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778191}