Query Tiles : Tile width adjustment
This CL tries to fix the tile width. If the tile is too small, it cuts
the words into multiline, and if too big, it might push out the feed.
With this CL, we try to keep the tile width close to the desired width.

Bug: 1092953
Change-Id: Iaf4f7c163d5b74b0cf83481f4958354e23de8687
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238999
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778226}