[WebLayer] Bring in fork of Chrome for Android's translate infobar UX
This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/2232322.
Compared to the original CL, the difference in this CL is that
resources added to //weblayer have had a "weblayer_" prefix applied to
avoid the ID collisions that result in crbug.com/1093293.

Original CL description:

WebLayer will be adding infobars, but its initial implementation will
be experimental. After discussion, we have decided that the best path
is to fork rather than componentize Chrome for Android's infobars UX
code for this initial experimental implementation. This approach will
avoid burdening chrome-team@ with componentization before we see that
it is necessary and will enable the WebLayer team to iterate on this
implementation as necessary quickly and without impact on Chrome.

This CL brings in the initial support that will be necessary to enable
the translate infobar to be rendered in WebLayer (building on a
predecessor patch that brought in support for compact infobars UX).
This change does not actually hook up the translate infobars but simply
brings in the support and get it to build. A followup change will have
it render. That followup change has been locally tested and have
informed what has been brought in in this CL.

The way to examine this CL is to look at the diff between the initial
PS, which contains just a straight copy of the classes from //chrome,
and the latest PS. The most substantive change is that at least for
this initial implementation WebLayer is not supporting snackbars,
and hence all of the snackbar usage has been taken out of
TranslateCompactInfoBar.java in favor of having it directly take the
actions that would be taken on snackbar expiration. Together with
megjablon@ we are adjusting the translate infobar UX in WebLayer to
account for this constraint (e.g., not closing the infobar if the
user clicks "Never translate" so that they can undo the action if it
was a mistake). We will make those adjustments in followup work.

Bug: 1025620
Change-Id: I9bd36bba9152f40ac27c928b84363ff277519d09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241676
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778236}