X11 and Ozone: add IsUsingOzonePlatform feature.
This is a first patch in the series of patches that
will add a switch that will allow to switch between
non-Ozone and Ozone path on Linux builds.

Please note that this is temporary and will be removed
as soon as USE_X11 is removed and USE_OZONE is default
on Linux.

In other CLs, I will add a bunch of changes that will
permit Linux builds to be built entirely without ozone
or with ozone where the use of the ozone vs non ozone
code paths can be controlled via runtime flag.

Bug: 1085700
Change-Id: I111eda4fb5e7e1b34d50ea0c39e388a195d4b122
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241685
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778247}