pref_hash_browsertest: use different regkey paths in parallel tests
The pref_hash_browsertests have been very flaky lately. Effectively, the
tests always fail on the first run and always succeed on the re-try.

When the tests run normally, the user data directory has format
"%TMP%\12345_6789012345\user_data". During the retry or with
--single-process-tests, the user data directory has format
%TMP%\scoped_dir12345_6789012345.

pref_hash_browsertest::GetRegistryPathForTestProfile was built with the
second format in mind and uses the directory's BaseName to generate a
registry path which can be used for the test. This assumption no longer
holds after
https://chromium-review.googlesource.com/c/chromium/src/+/1721429.
With the first path format, the directory's BaseName is always
"user_data" hence the registry keys used is the same between parallel
tests. The issue wasn't detected at the time since the tests succeeds
on retry.

This change adds some hacky logic to support the first directory format
and a DCHECK to attempt to detect future issues.

Bug: 1091409
Change-Id: I2e99f952d2dda78b277c3fc15213063b9e337382
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2240246
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: proberge <proberge@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778298}