Always enable filtering hint if available
Do not gate setting this hint on whether SwiftShader is being used. If
chromium_texture_filtering_hint is true, the driver has already
validated that it's available.

This will fix the problem of this hint not working when ANGLE uses
Vulkan with the ICD pointing at SwiftShader, in which case
chromium_texture_filtering_hint will be true, but is_swiftshader would
be false.

Bug: b/146423360
Change-Id: Ia248835aff434d9efb7e96f35e8cd0e25a366220
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2243711
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Jamie Madill <jmadill@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778325}