Implement |TextInputClient::ShouldDoLearning| for search bar.
This CL contains two parts. First part is to add a method in |TextField|
to set |should_do_learning_| flag. The flag indicates whether the owner
of current text field is in incognito mode or guest mode. Second part is
to let search bar implement |TextInputClient::ShouldDoLearning| so that
IMEs don't do text intelligence in search bar if the browser is in
incognito mode or guest mode.

Bug: 311180
Change-Id: I8e90fccef4a5d69e703c66464b1bdab4e5dc678b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2242036
Reviewed-by: Siye Liu <siliu@microsoft.com>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Commit-Queue: Siye Liu <siliu@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#778354}