Added trace events to Toolbar UI components
The toolbar in Chrome for Android is the primary piece of native UI in
the app. It is used to achieve most basic tasks (performing searches
or switching tabs). Therefore its performance has a large impact in
user experience and it should be measured precisely.

Currently there are no trace events or histograms that measure UI
performance on toolbar components, so it's hard to know how are these
components performing on the wild.

This CL adds trace events to some toolbar components to be able to get
performance data from the wild with chrometto.


Bug: 1094402
Change-Id: Iea618e5d08e853955c7f70802068c29a9b9ff9f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2236578
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Salvador Guerrero <salg@google.com>
Cr-Commit-Position: refs/heads/master@{#778376}