[Switch Access] Properly position back button on menu open
Currently, when the menu opens, the back button appears 8 pixels too
far from the menu, and it jumps closer when the back button is initially
focused. Upon investigation, this is because the resting bounds
calculated when the menu opens include 8 pixels of padding. This change
insets by the padding size so the position is correct upon opening.

AX-Relnotes: n/a.
Bug: 1094069
Change-Id: Ie10525bb03dc0d884feb0281bcc1e2f81e5e928c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2242082
Reviewed-by: Yoshiki Iguchi <yoshiki@chromium.org>
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/master@{#778419}