Update more WCv0 tests to WCv1
This CL contains manual modifications to a handful of tests that use
Shadow DOM v0, to attempt to upgrade them to SDv1. This CL also includes
the movement of several tests (which turned out to be V0 only tests)
to the web-components-v0-only folder, which will eventually be deleted.

Of particular concern for review:
 fast/css/cache/matched-properties-cache-no-flat-tree-parent.html
 fast/css/invalidation/shadow-add-sheet-content.html
 fast/css/boundary-crossing-scopes-null-resolver-crash.html

In this test, I simply deleted the V0-specific part of the test:
 shadow-dom/window-event.html

And since this test can't be moved easily, and it only tests the
(v0 specific) ::content selector, I just deleted it:
 inspector-protocol/css/css-shadow-host-content-selector.js

Bug: 937746
Change-Id: I74967e65b1a0ec52e2014b40d25b84107b4c46d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2243692
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778435}