MacPWAs BMO: Fix app doesn't launch when loading profile bug
It is possible to launch an app on a not-yet-loaded profile in the
following circumstances:
- Launching an app's shim
- Selecting a profile from the People menu
In this case, we may attempt to launch the app prior to the
WebAppManager having been loaded for the profile. Fix this by
adding (yet) another callback to the sequence used by
AppShimManager::LoadProfileAndApp.

Update the following functions to have the same structure:
- AppShimManager::LoadProfileAndApp
- AppShimManager::OnProfileLoaded
- AppShimManager::OnProfileAppRegistryReady
Wherein the create a base::Closure, and then either call Run
on it or pass it to something to asynchronously execute it.

Update tests to take this into account.

Bug: 1094419
Change-Id: I39fb11dea25e9b4769687210a12d13faa116b426
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2244035
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778448}