display: Privacy screen support criteria update
This CL modifies IsPrivacyScreenSupportedOnInternalDisplay() to return
true only when the current internal display is not disabled (i.e.
returns a valid current_mode()).

This is done so the UI can update accordingly and remove the privacy
screen pod when there's a the internal screen is disabled (e.g. when
an external monitor is connected and the lid is closed; AKA docked
device).

Bug: b:157775487, 1090385
Change-Id: Ia51d8e5b6cad48259e1114c4e61b194f97b1ff28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246186
Commit-Queue: Gil Dekel <gildekel@chromium.org>
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778481}