device/fido: don't route requests with no common algorithms.
Currently, creation requests that specify public-key algorithms that an
authenticator doesn't support can still be routed to that authenticator.
When that happens, it fails immediately and we ignore it, so it never
blinks.

Instead, don't route such requests to the authenticators in the first
place and replace them with dummy touches.

BUG=1094066

Change-Id: Ie89c387887339f4dfc970a9f2205c6c354d141f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2214100
Commit-Queue: Avi Drissman <avi@chromium.org>
Auto-Submit: Adam Langley <agl@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#778513}