[signin] Make reauth webUI dialog take 100% of visible viewport
chrome://signin-reauth is displayed within a native dialog. The content
should occupy all available space in the dialog.

It used to be done by setting body.height css property to the expected
height of the dialog. This logic has a flaw because the actual size
might be different from the expected one.

This CL defines the body.height parameter as 100% of viewport height,
which we actually want.

Bug: 1045515
Change-Id: I4b1200cb44882a5d9b22869a947a7bfeb6e554a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246171
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778520}