Add unit tests for path color in path-hiding experiments
The main goal of this CL is to add a regression test for
https://crbug.com/1093748. To do that, it adds an
OmniboxViewViews::SetPathColor() method that is overridden by
TestingOmniboxView to check that the path is not set to transparent
when it shouldn't be. The same method is used to test that the path is
set to transparent when the path-hiding field trials are enabled.

This approach waas complicated by the fact that the unit tests don't
actually call the real EmphasizeURLComponents(), which is where the
bug was located. TestingOmniboxView had an EmphasizeURLComponents()
override that did a subset of what the real version does. The real
EmphasizeURLComponents() was not immediately usable from unit tests
because it had an early-return when there was a null
location_bar_view_, which is the case in tests. However, I couldn't
find any reason why a non-null location_bar_view_ was needed, so I
simply removed that check and made the unit tests use the real
EmphasizeURLComponents().

The last wrinkle in this plan is that using the real
EmphasizeURLComponents() caused OmniboxViewViewsTest.OnBlur to start
failing. AFAICT this is due to either a real bug related to how RTL
URLs are scrolled in focus mode, or an incorrect assertion. I
commented out the assertion for now and filed
https://bugs.chromium.org/p/chromium/issues/detail?id=1094386.

Bug: 1092651,1093748
Change-Id: I8b4690251d5d82fc28f4141bb37087ca730726fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2243491
Commit-Queue: Emily Stark <estark@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778523}