Progress and error handling for the Windows on-demand server.
This improves the backward compatibility of the on-demand legacy
server. It feeds install progress to the update client in Chrome
and handles the installer errors.

There is a patch which can be used to fix some eager DCHECKS in
the update client in Chrome to facilitate debugging end to end.

Don't patch the production update client in Chrome.

Bug: 1067348

Change-Id: I9ea2e732c9a2ebc3102a2696179ca186aa2fd72a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233858
Reviewed-by: S. Ganesh <ganesh@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778541}