Mac jump on click.
This CL enables composited scrollers on Mac to jump when clicked on the
scrollbar. On Mac, the "jump to the spot that's clicked" setting can be
dynamically set via System Preferences. When enabled, the expectation
is that regular clicks on the scrollbar should make the scroller "jump"
to the clicked location rather than animated scrolling. Additionally,
when this is enabled and the user does an Option + click on the
scrollbar, the scroller should *not* jump to that spot (i.e it should
be treated as a regular track click). When this setting is disabled on
the Mac, Option + click should make the scroller jump and a regular
click should animate the scroll offset. On all other platforms, the
"jump on click" option is available (via Shift + click) but is not
configurable.

Bug: 1080590
Change-Id: Ibf5bf2c12184189b136f8b96acc20585406c168c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2244972
Commit-Queue: Rahul Arakeri <arakeri@microsoft.com>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778551}