Files app: Style background image in CWS error dialog
The error dialog in in the CWS (Chrome Web Store) shows an icon as
background image. Currently this icon is cut off by the dialog title and
it's off center because the image has empty space at the right hand
side.

This CL positions the image in the center with margin-inline-start and
sets the height to push the dialog title below the icon.

See crbug.com/992827#c47 for screenshot with this changes.

Bug: 992827
Change-Id: I70b871cd7b3a54e5043ef624f6861c3dc5f43c63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246086
Auto-Submit: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778569}