arc-a11y: Improve focusability computation
This CL tries to make linear navigation order of ChromeVox on ARC++
similar to TalkBack.

For example, imagine a11y tree in Android is like this:
++ nodeA container (focusable clickable) (name is empty)
++++ nodeB static text (not focusable, not clickable) (name X)
++++ nodeC button (focusable clickable) (name Y)

TalkBack's navigation order is as below:
1. nodeA, announce a name of nodeB.
2. nodeC, announce a name of nodeC.

Currently ChromeVox skips nodeA. With this CL, ChromeVox firstly focus
on nodeA and nodeB is skipped because it's ignored.

This CL introcues 'screen reader mode' in ARC accessibility, which is
enabled when SwitchAccess or ChromeVox is enabled. The focusability
change added in this CL is applied only when the screen reader mode is
enabled so that we don't break text location API for Select-to-Speak.

Note that this change adds a few TODOs, which will be resolved in follow
up CLs.

AX-Relnotes: n/a.
Bug: b:157441336
Bug: b:150344900
Bug: b:150344310
Test: unit_tests --gtest_filter="AXTreeSourceArcTest.*"
Test: manually confirmed with PlayStore and Android Settings.
Change-Id: I700cbbd348c9eb7bf71a7995994c926b1c839979
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222068
Commit-Queue: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Sara Kato <sarakato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778595}