[fuchsia] Only create a single channel for ApplicationContext
A channel for chromium.cast.ApplicationContext was created in both
PendingCastComponent and CastComponent. This sometimes caused issues
in agents that make use of ScopedServiceBinding to handle the
server-end of this channel. These agents could get into a state where
they never got a signal on the second client disconnect.

Bug: 1095217, b/158684780
Change-Id: I06e8e1dd35516bb88ea4929e9f43fb4cd862cd1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246794
Commit-Queue: Fabrice de Gans-Riberi <fdegans@chromium.org>
Reviewed-by: Kevin Marshall <kmarshall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778597}