arc-a11y: Clean up computations related to clickables
CL:2222068 improved focusability computation in ARC++, which introduced
a focusability computation similar to what TalkBack does. We no longer
need to see a clickable node as a leaf.
This change removes most of changes that was done in CL:1243670,
CL:1316095, and CL:1849548.

AX-Relnotes: n/a.
Bug: b:152930082
Test: unit_tests --gtest_filter="AXTreeSourceArcTest.*"
Test: browser_tests --gtest_filter="ChromeVox*.*"
Change-Id: I7b70116e534d555ab6f8c5c66d5af1cbbb37be9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241393
Commit-Queue: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Sara Kato <sarakato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778625}