Migrate ImageDataUrl, TextLength, AccessKey to AXNodeObject
Part of the ongoing effort to consolidate AXNodeObject and
AXLayoutObject, with the eventual goal to remove AXLayoutObject
entirely. This patch migrates the following methods to AXNodeObject:
  - AccessKey
  - ImageDataURL
  - TextLength

The methods in this patch are not thematically related or dependant in
any way, but were identified as being trivially easy to move across.

This patch should have no functional impact - it is purely a refactor.

Written with aboxhall and chrishall

AX-Relnotes: N/A
Bug: 1093651
Change-Id: Iebd7e71dbb10727e62f21e1c13abcad1f3f02e76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2237450
Reviewed-by: Alice Boxhall <aboxhall@chromium.org>
Reviewed-by: Chris Hall <chrishall@chromium.org>
Commit-Queue: Meredith Lane <meredithl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778627}