[AW] Reschedule the seed download job if it fails to download a seed
This CL makes the seed download job request that it be rescheduled via
the jobFinished call if there was an IOException while performing the
download. The retry will have an initial delay of 30 minutes, with
exponential backoff for subsequent retries. If more than 24 hours has
passed since the initial download, the next seed request made to the
service will reschedule the job with a fresh one.

Since we can't get metrics out of WebView's service, we can't know how
often seed downloads fail, but on Clank they fail ~7% of the time,
about 4% of which are SocketTimeoutExceptions, and 2.5% of which are
UnknownHostExceptions (http://shortn/_tT3Yc2QkFT). If we assume Clank
and WebView have similar failure rates, then we should expect a
similarly large number of seed download failures on WebView.

Test: out/aw/bin/run_webview_instrumentation_test_apk -f \
Test: 'AwVariationsSeedFetcherTest#test*Fetch'
Bug: 979075
Change-Id: Ic69df63be169aafbfa19c1edd0a4cd017a876040
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2008620
Reviewed-by: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778640}