[Extensions] Re-enable a Declarative Content API test.
This CL fixes flakiness with the RulesPersistence test on all
platforms except Windows release builds. See the referenced
bug for more details on those failures.

The test was flaky for multiple reasons:

1. There was a race condition between the ready listener for
the message the extension script sends. Removing this listener
fixed this issue. This test should be re-written to use
chrome.runtime.onInstalled to install rules, since that is what
we recommend to extension developers.

2. The C++ should wait for changes in page action icon visibility
before checking the state. This was fixed by using a
ChromeExtensionTestNotificationObserver instance to track these
changes.

Bug: 512431
Change-Id: Iebde17d85ee635467b983ca3120d5e9b551382d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2228054
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778678}