Attach ContentSetting state to the page not the tab
Store state per page instead of per tab. What content is blocked is
really page dependent not tab dependent. The tab should display the
settings for the currently active page. There can be multiple pages
around for one single tab:

- If back forward cache is active a page in the
cache is still there, and when we navigate back we should restore the
state as it was before it was put into the cache instead of resetting it
entirely as happens today (remember a back navigation in this case
triggers no load, so no cookie access no javascript etc..)

- Navigations that end up with a 204 error no not really happen, no page
changes, but the current code will reset all state.

This patch stores all this state in a per page structure and swaps it
as needed during navigations. Also state that happens during a navigation
is only really committed to the tab if the navigation completes
successfully.

Bug: 998171

Change-Id: I334ba080a9cb810b1431df3c9c27d2b536f33626
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220033
Commit-Queue: Carlos Caballero <carlscab@google.com>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Sreeja Kamishetty <sreejakshetty@chromium.org>
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778761}