Refactor AppInstallEventLogUploader
AppInstallEventLogUploader class is an adapter between the system that
captures and stores app push-install event logs and the policy system
which uploads them to the management server. Extract the common code
with ARC++ unspecific code into common class. This is a prework for
adding similar class for extension events.

Bug: 1090250
Change-Id: I878edb75e617adbba110bcf750710b8d303fac3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225399
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Swapnil Gupta <swapnilgupta@google.com>
Cr-Commit-Position: refs/heads/master@{#778820}