Fix blocking webRequest handler causing CORS related requests to hang
CorsURLLoader::FollowRedirect may decide to restart a redirected request
instead of calling FollowRedirect on the network loader if the original
request does not have the fetch CORS flag set, but the redirected
request does, or vice versa[1]. WebRequestProxyingURLLoaderFactory
receives the OnLoaderCreated event from the new loader, which then
rebinds the receiver stored in InProgressRequest. This receiver may have
been paused if webRequest is waiting on an extension to handle a
blocking event. Once the receiver is re-bound, it was no longer paused,
which caused bad things to happen. This change makes sure we save the
paused state so the rebound receiver will be in the same state as the
previous one.

I tried to write a browser test for this, but was not able to recreate
the situation reliably without adding the equivalent of a sleep in the
onBeforeRequest javascript handler. I'll look into it more and possibly
add one in a follow up. I have verified that my hacky sleep test fails
without this change, and passes after the change.

1.
https://cs.chromium.org/chromium/src/services/network/cors/cors_url_loader.cc?l=251-260&rcl=c9082ad0bfa317463ed695773a68542656352e42

Bug: 938560
Change-Id: Ieb8b0b00e4fddd3c04f254a796ab40cf8f9cb15c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2243933
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778835}