Split PasswordCheck Config
Due to various launch/survey requirements the original single 5 feature launch from PasswordCheck.gcl to 3 launches.

PasswordCheck now only contains the ImprovedCookieControl launch, this odd naming is to prevent user
shuffling for ImprovedCookieControls.

PasswordCheckLaunch is the launch config for PasswordCheck and is being set to launched state.

FriendlySettingsHats is the config to do a 98/2 split on stable for PrivacySettingsRedesign,
SyncSetupFriendlySettings, and SafeBrowsingEnhancedProtection for HaTS surveys.

Bug: 1032584
Change-Id: Ia982d6438ae099e449e855848b562ad404d383e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238168
Commit-Queue: Sean Harrison <harrisonsean@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778841}