Peak Gpu Memory Ablation Tracing Categories
Split off from master:
https://chromium-review.googlesource.com/c/chromium/src/+/2187203

We have previously added new traces for the ablation experiment under a
new category "gpu.memory". This was done to avoid trying to record and
parse all of "gpu".

This patch adds the new category to the loading and rendering stories.

The catapult patch to parse the results is:
https://chromium-review.googlesource.com/c/catapult/+/2187206

Bug: 908462

Change-Id: I548bd2a10a859f706342a438906bbee35f396970
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246931
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778870}