content/android: Reduce & move power UMA collection onto thread pool.
With the addition of per-thread UMAs, the collection became a little bit
more expensive. This patch moves it to a background thread and reduces
the collection interval for the browser and GPU processes, which
(i)  are persistent processes (so data loss at termination is less of
     an issue) and
(ii) tend to execute many small main thread tasks.

Bug: 1081760,1094434
Change-Id: I847179ae78c024bb506772c1c42804e9d896f762
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2245906
Commit-Queue: Eric Seckler <eseckler@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778878}