Enable JsonPrefStoreTest.RemoveClearsEmptyParent on iOS 12 device
This was disabled due to test crash happening on iOS 12 devices. The
crash doesn't happen any more when we switched to running XCTest based
unit tests on iOS 12 devices, but not launching tests with
idevice-app-runner (bug: crbug/1087078, CL: crrev/c/2219108). So this
test can be re-enabled to close the bug.

Bug: 1083299
Change-Id: Id1182e132fba2fad64cdc1777440d758b4dcd099
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2243836
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Zhaoyang Li <zhaoyangli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778889}