[WebLayer] Bring in fork of Chrome ConfirmInfoBar for WebLayer
The popup blocked infobar uses the ConfirmInfoBar, so we will need the
equivalent functionality in WebLayer. Since ConfirmInfoBar inherits from
InfoBar which was forked in http://crrev.com/c/2224869, we also need to
fork ConfirmInfoBar. The layout for non-compact infobars was
componentized in http://crrev.com/c/2243579 and can be used between both
Chrome and WebLayer.

The initial patchset here has a clean copy of the Chrome files, and the
subsequent patchsets have changes for WebLayer.

Bug: 1019922
Change-Id: I76e093e6113c26990dd1041ac4c901690489bad3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248538
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778903}