Create new pref to store input gain
This change creates a new pref, prefs::kAudioDevicesGainPercent, to
store the gain values for each of a user's input devices. A new pref is
created because the default value for input gain was changed to 50
rather than 75, and any users of mic gain before the switch will have
the old value stored in prefs rather than the new one.

Bug: b:158677629
Change-Id: I536b375dba02c00fa021688be09881eb59a81adf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246231
Commit-Queue: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778913}