Initialize OOPIF screen_info from root RenderWidgetHost().
Prior to this CL, when a CrossProcessFrameConnector is created for an
OOPIF, SetView() has not yet been called, so when the constructor calls
RenderWidgetHost::GetScreenInfo(), it defaults to calling
DisplayUtil::GetDefaultScreenInfo(), which will retrieve the information
for the primary display. If the primary display has a different
device_scale_factor than the display on which the OOPIF is rendering,
then (initially) the OOPIF will render with the wrong scale.

As noted in the associated issue, the OOPIF renders correctly after
OnSynchronizeVisualProperties() has been called ... this is because
it propagates the screen_info from the root view down to all the OOPIFs.

This CL makes a simple modification in forcing the constructor to
get its initial screen_info from the root RenderWidgetHost, which
presumably has a correctly attached view by the time OOPIFs are being
created.

Bug: 1076450
Change-Id: Ic85f0b099e2a8fb7aa1e8b579dbcddcdcac3e757
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247749
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778935}