ScrollAnchoring: Avoid adjusting contain layout paint element size.
This patch ensures that when we increase the size of elements to
determine whether we should descend into them to find a deep anchor,
we avoid descending into elements with layout and paint containment
that are otherwise off-screen. In other words, we don't adjust their
size because any overflow would not be visible.

R=chrishtr@chromium.org

Change-Id: Id06e4db3625ded3ad43c5e57aeb5b2e40c7ed68f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2245186
Commit-Queue: vmpstr <vmpstr@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#778952}