Add contents of popup note associated with a highlight in PDF accessibility tree
This CL adds the note text data associated with a highlight to the
accessibility tree. The CL makes the content of the note text accessible
to users via screen readers.

Tests have been modified to validate the addition of note text data to
the accessibility tree.

Bug: 1008775
Change-Id: I086d039b9296d0b8ac1b13926835375a7943f8c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208832
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Ankit Kumar 🌪️ <ankk@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#779008}