Cache Callbacks in PersistedTabData
PersistedTabData provides a mechanism for acquiring serialized UserData
objects in storage and attaching them to a Tab. The caching mechanism
provides the capability to cache callbacks acquiring the same data
which yields a performance improvement. It also means we avoid a bug
of the following:
1 FooTabData.from(...)
2 FooTabData.from(...)
3 setUserData changing what is stored on FooTabData
4 2) returning after 3) and changing what is stored on FooTabData

Bug: 1059602
Change-Id: Ib3534922ace1a653068caa17a78a7eb46d307dd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2236045
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779097}