Switch maps pixel test back to expected colors
Switches the GPU maps pixel tests back to expected colors instead of
using Skia Gold. The test is currently too noisy (~1500 unique images in
Gold), and while inexact matching is handling the vast majority of these
images, it's causing the triage log to be filled with nothing but
results for this one test. Additionally, the large number of images is
making the diff server (what is used to calculate the diff shown between
a particular image and the next closest) quite slow.

So, keep this test as an expected color test until some improvements can
be made in Gold to address these issues.

TBR=kbr@chromium.org

Bug: 1078914
Change-Id: I11e5102e90c8df2e294a1898991443b040422ac0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248492
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779104}