Quota: Grant persistence to installed apps.
In the current state, persistence (durable storage) is granted
only to ImportantSites. A list of origins is populated with the
following, in order:
 1. Origins with high engagement (installed adds to engagement score)
 2. Origins with notification permission
 3. Origins that have persistence already
 4. Origins that are bookmarked

The list is then sorted by a calculated score, and then the list is
chopped to a max length of 10.

This change splits out installed origins from this logic. Now,
DurableStoragePermissionContext::DecidePermission() will query
ImportantSitesUtil for a set of installed origins and check if the
requesting origin is in that set before querying for ImportantSites.
If the origin is found to be installed, we can grant the permission
and return early.

Bug: 933259
Change-Id: I2636f8eb5d4ecc5e58c343ec0ab7ed15ab0ee6fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2236727
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Jarryd Goodman <jarrydg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779111}