[consent toggle] Modify playstore search conditions.
After some discussion, we have decided that if the suggested content
toggle is not present, then playstore recommendations will still be
enabled.
This change modifies the logic so that we disable playstore search only
if both the following conditions are met:
(a) The suggested content toggle feature is enabled,
(b) The suggested content pref is disabled.

Bug: 1079169
Change-Id: I15d542bec594b6ce3bdc8449fc09db1da43e9472
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249280
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Rachel Wong <wrong@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779193}