Refactor transform update in LayoutSVGBlock subclasses
The transform is already stored in LayoutSVGBlock, so move the
|needs_transform_update_| flag there as well as the setter. Move the
actual update of the transform to a new helper method on LayoutSVGBlock
named UpdateTransformAfterLayout(). This should make it easier to fix
the referenced bug - which affects both of the subclasses of
LayoutSVGBlock.

Add a TODO about the early transform update in LayoutSVGForeignObject.

Bug: 1094020
Change-Id: Iaa351e06aaafa9dddba868d36ea7841c4d4653d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247788
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#779221}