Fix empty Location header in Authenticator
Avoid assert'ing on the "location" header being non-empty, and restore
the original handling (before r746703/r745762) that was treating this as
not having the "choose what to sync" option enabled.

This header happens to be empty in some browser tests, and also it might
potentially be empty in real-world usages.

Bug: 1033936, 1054394
Change-Id: I65aaa1a71cd86b5cf80be57a9407bcea41c80bb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248192
Reviewed-by: Anastasiia N <anastasiian@chromium.org>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779253}