A new policy for online re-auth on the lock screen.
We introduce a new policy SamlLockScreenReauthenticationEnabled that
will enable online re-authentication of SAML users on the lock screen.
With the policy set to true a user can be set through the online flow
on the lock screen in order to verify and sync their password. If the
policy is set to false or unset we will continue to use only local
credentials on lock.

Bug: 1090341
Change-Id: Ia24ac44f16abf7d140675a5cdfc0f9c349b85e4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233319
Commit-Queue: Maciek Slusarczyk <mslus@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779290}