[Adaptation] Make RTCPeerConnectionHandler aware of DeviceThermalState.
This CL plumbs base::PowerObserver::OnThermalStateChange with mojo from
PeerConnectionTrackerHost in content/browser/ to
PeerConnectionTracker in blink/renderer/modules/peerconnection/ which
informs the RTCPeerConnectionHandler of the event.

A mojom typemap for the enum is added to allow passing it and DEPS and
audit_non_blink_usage.py is updated to allow the usage of
base::PowerObserver::DeviceThermalState from blink.

Tests verify that the signal reach the handler, but nothing is done with
this signal yet. In a future CL, we will implement a webrtc::Resource to
make use of it.

Bug: chromium:1094844
Change-Id: I026e04238d89ced2d77672e015f58e7a497336f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246162
Commit-Queue: Henrik Boström <hbos@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779297}