[Sync] Commit nigori always separate from user types
This CL stops combining nigori commits with priority user type commits
because nigori cannot get combined with encrypted commits and because
we plan to break the guarantee that priority types are not encryptable.

Impact of this change on traffic / latencies / etc. is negligible
because nigori gets committed very rarely.

Bug: 1084969
Change-Id: I2a7bb840a30def87813bb45394881e4a871f69b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247878
Commit-Queue: Jan Krcal <jkrcal@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779300}