Notificafions: don't register notification click in all browsers
With current implementation of notifications on mac clicking on
notification created by any active chromium browser will be
registered in all the browsers. Additionally clicking in notification's
button Settings will open settings page in all browsers. This commit
adds new field to the notification: creator's pid which is checked
with the pid of current process when the information about notification
being clicked is send to the browser process from the NSUserNotificationCenter.
If there is no match we do not process the notification. It doesn't solve
the problem with the focus (all browsers gets it and the last one which got
it will have it) but at least click on the notification is only registered
in the proper browser. It also fixes the problem with button Settings.

Bug: 1086899
Change-Id: Ia484c03c076f841ba3695b029ebced3c1191e6ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218827
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779307}