PasswordManagerClientHelper: don't prompt to move in incognito
This CL adds a check in PasswordManagerClientHelper so that it doesn't
call PromptUserToMovePasswordToAccount() in incognito mode.
Note that this does not actually change the overall behavior: The move
prompt was already suppressed in incognito, but in a somewhat
convoluted (and potentially fragile) way (see bug).

Bug: 1094918, 1093310
Change-Id: Iccfaaae7095d585c5ea39d3bf409c75c288ab723
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247828
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779310}