[b4p/Settings] Remove exceptions from 'passwords on this device' page
This CL is essentially a revert of cc694e8, the reason being that UX
has decided not to have exceptions on this page anymore. The only two
differences regarding the original commit are: hasDeviceOnlyPasswords_
and hasDeviceAndAccountPasswords_ are still replaced with a shared use
of isNonEmpty_(); and the file documentation is updated to remove any
mention to exceptions.

Bug: 1049141
Change-Id: Ic95bf0042d6f4214ec40dd9dd0fedc492de56421
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248683
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779319}