Fix array-vs-object bug in chrome.test.checkDeepEq
This makes the chrome.test.checkDeepEq() and the chrome.test.assertEq()
functions more robust when handling the values of different types, in
particular in the following scenarios:

  chrome.test.checkDeepEq([], {});
  chrome.test.checkDeepEq([1], {0: 1});

Before this CL, both of these expressions evaluated to |true|, despite
that the values are logically different. The bug was caused by the fact
that iterating using a "for" loop may result in the equal key-value
pairs for an object and an array, so it's not a sufficient criteria.

The proposed fix is to additionally compare the results of the
"Array.isArray" check.

Also fix the "AppWindowApiTest.Get" test that was (accidentally?)
relying on this bug.

Bug: 1078761
Change-Id: I00f24a567ba8589e5445313d9b49dd7cf394f1f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2245913
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Omar Morsi <omorsi@google.com>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779346}