[iOS] Move the UserAgent to the WebState
This CL changes the way the UserAgent is stored and used.
Now the UserAgent to be used when loading a page is stored in the
WebState. Each NavigationItem is storing the UserAgent that was used
when the item was loaded. This is the UserAgent that is used to display
the option (request Mobile/Desktop) and to reload the page when
navigating back/forward.

As the UserAgent is now in the WebState, it has to be restored as part
of the session restoration. This also allows the UserAgent to be kept
when the preloader is used, as the session is restored.

When the mobile/desktop version is requested, the NavigationManager is
updating the UserAgent of the WebState.

Bug: 1049094, 1076855, 1078343
Change-Id: I459ec728f01da814ba23bafc0aa7fb1e185f95e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2193994
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779359}