Adjust collapsed tab group header width for visual balance.
Tab group header title chips are not centered in the tab group header.
Instead, they "tuck into" the stroke on the left side of the first tab
in the group. This looks good when the group is expanded, but bad when
the group is collapsed, because the title (unlike the underline stroke)
looks off-center.

We considered two approaches to fix this:
1. center the title in the existing header size
2. expand the header so the margin on the right side matches the left

We chose (2) because it prevents visual jitter in the title chip when
expanding and collapsing, as well as automatically aligning the title
with the stroke.

Bug: 1094040
Change-Id: Ie5e6252b56f7fdf612c580d403bd1b198bb2710c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249055
Commit-Queue: Charlene Yan <cyan@chromium.org>
Reviewed-by: Charlene Yan <cyan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779390}