Restart the GPU process if necessary to be XR compatible on Windows
When XR compatibility is requested, determine whether the GPU process
is using the same adapter that the active XR runtime is attached to.
If not, restart the GPU process and notify it to initialize and use
the XR runtime adapter.

This is the last series of changes to add multi-GPU support for XR
behind a flag. This change was initially reviewed as one large CL here:
crrev.com/c/2096778

That CL was split into this CL and the following:
crrev.com/c/2226071
crrev.com/c/2219780
crrev.com/c/2216166
crrev.com/c/2225366

After these land, there is followup work before enabling by default:
- crbug.com/1086697: Return a WebGL context if no XR runtimes available
- crbug.com/1090951: Migrate back to the default GPU on navigation
- crbug.com/1090955: Remove flicker when GPU process is restarting
- crbug.com/1087356: WebGL2 context

Bug: 792657
Change-Id: Ie3342afd630b624b78a4f24a70fedf2cae5eace4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2228544
Commit-Queue: Patrick To <patrto@microsoft.com>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779435}