Delete net::URLRequestSlowDownloadJob.
It doesn't work when the network service is enabled. Two tests are
actually using it. One uses a URLLoaderInterceptor when the network
service is enabled, but that also works when the network service is
disabled, so it's not necessary to use it at all. The other test has
been disabled since before the network service shipped, and does not
work with the network service enabled at all.

Bug: none
Change-Id: I70535435f60c9f93a70ab5cdd8d0df3972e5fb3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248841
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779450}