gpu: Remove GPU.TextureR16Ext_LuminanceF16 histogram
The histogram was added so that we could remove one of LUMINANCE_F16 or
R16_EXT if the other one had universal support, but it looks like that's
not the case. Latest numbers for the histogram across all platforms are:

HAVE_NONE 32.34%
HAVE_R16 09.96%
HAVE_L16 32.46%
HAVE_R16_AND_L16 25.25%

This suggests that we can remove neither LUMINANCE_F16 nor R16_EXT, and
the histogram can be removed since it's expiring in a few days.

Bug: 1081297
Change-Id: I46615e9b70a72d27f3952ec90c3c0193777fb866
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248909
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779466}