[Tab Groups Collapse] Make dragging past a group header feel better.
The collapsed tabs right now are to the right of the group header so
dragging left is lot more prominent than dragging right past the header.

When finding the insertion index, if the tab is in a collapsed group,
use the ideal bounds of the header rather than the tab.

Bug: 1018230
Change-Id: I4bbec1e76da9415a1701b7b09ba35dff98039667
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250298
Reviewed-by: Connie Wan <connily@chromium.org>
Commit-Queue: Charlene Yan <cyan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779482}