Revamp logic for clearing prefs for variations first run seed.
In additional to simplifying the code to not need to use extra
callback logic, this also makes the pref clearing happen when a
seed is imported initially, rather than later. This may resolve
an issue where the first run seed is not getting applied - if
the cause is that the pref clearing happens before the seed
gets used.

Bug: 1090968
Change-Id: Icd5d6a693ef2c0f1a5fdf2cffec2182a9a6fd5fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246179
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779483}