optimize_webui.py: Update 'unpak' folder detection to work for NTP.
This is a temporary fix which will be replaced with an assertion in a
follow up. Currently there is an assumption in the code that if
unpack_pak.py is used the output of that script is placed in a folder
that is suffixed with '.unpak', which is not true for NTP which uses a
folder named 'unpak' without the dot.

This fixes the contents of
out/<out_folder>/gen/chrome/browser/resources/new_tab_page/optimized_js.d

to not hold unexpected entries for files already in the unpak folder.

Bug: 1093405
Change-Id: Ib84b5b0890d1a98e938dcdf691e621b397b8de91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250660
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779545}