tablet: Make sure screenshot is deleted for popups type window.
See linked bug for why the types of windows are undesirable and should
also be fixed at the callsites.

But in the meantime, they can exist so we need to protect. For these
windows, they are part of the WindowCycleList. Since we group up
transient related windows in overview/alt+tab, those two surfaces are
seemingly unaffected. For tablet mode transition, a screenshot is
first created and then deleted either if no animation is run, or at
the end of TRANSFORM animation.

Since these windows cannot be maximized, they are actually given a
BOUNDS animation. The check to see if a window is animating should be
specifically check for TRANSFORM animation.

Test: added regression test
Fixed: 1096128
Change-Id: Ide585fae26f6ed7a96d3a181b399385f2fa461f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250459
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779547}