Remove non-compositing descendant DisplayItemClient invalidation
This is currently only used for scrollbar parts, which do not have
descendant layers. Instead just set needs paint on the painting layer
and invalidate the DisplayItemClient (the same LayoutObject the
ObjectPaintInvalidator is initialized with) via
SlowSetPaintingLayerNeedsRepaintAndInvalidateDisplayItemClient.

There was not any unit test coverage for
InvalidatePaintIncludingNonCompositingDescendants which uses the same
traversal algorithm, so convert tests of the deleted method to use this
one instead, and validate only the paint invalidation side-effects.

Bug: 1007989

Change-Id: Ic9849f889958300941bc9a48d97cf7f04a91c786
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247157
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#779554}