Add aria-label to small form quiet webview interstitial icon
Adds an aria-label to the interstitial icon when no text is being
rendered. This reuses the interstitial heading data as the label.

For non-small form frames, no aria-label is needed because the
interstitial heading already describes the icon. However, when no text
is rendered there is currently nothing that can be picked up by screen
readers.

Bug: 1095754
Change-Id: I51062e1028c74b61bc3e80922d05ad021f0e36e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249082
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Reviewed-by: Edward Jung (EMEA) <edwardjung@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779624}