[Launcher settings] Control search filtering parameters via Finch.
We have three parameters we want finch control over: minimum query
length, minimum query length for alternate matches, and minimum score
for alternate matches.

This CL is mostly a trivial move of the FilterResults function out of
the anon namespace and into the OsSettingsProvider class, except for
the addition of lines 141-151 in the ctor.

Bug: 1068851
Change-Id: Iefd09e85d5677c376ec3ae23d28808416088b098
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249258
Commit-Queue: Tony Yeoman <tby@chromium.org>
Reviewed-by: Rachel Wong <wrong@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779640}