Add policies for specifying allowed external protocols
This change adds two new group policies. The first policy accepts
a list of protocols that may be launched by the browser without
prompting the user. The second policy, if used, accepts a list of
origin patterns that restrict the behavior of the first policy to
just those origin patterns.

The rules for origin pattern matching are described publicly here:
http://www.chromium.org/administrators/url-blacklist-filter-format

"/path" and "@query" patterns cannot be used as part of origin
patterns, and this is noted in the policy description.

Bug: 911605
Change-Id: Icb3aa769d5f0010225cb21097e3bb635085ad7e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225522
Commit-Queue: Todd Sahl <toddsahl@microsoft.com>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779701}