[mpc] Use CSSProperty* in CascadeResolver's cycle detection stack
A subsequent CL needs to get the currently applying CSSProperty* from
the CascadeResolver, and it's more efficient to get the pointer
directly, rather than looking it up via the CSSPropertyName.

This is a refactor: there is no behavior change in this CL. (Except
one addition unit test is added).

Bug: 1057072
Change-Id: I7532f2c23f9da915c132cb42abef7b804d3a5732
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247821
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779735}