Refactor: Decouple PasswordListItem from embedder
This CL addresses a few design issues with PasswordListItem that predate
the introduction of PasswordsListHandler.

- PasswordListItem would refer to a "menu" even though it knows nothing
about such menu, which is actually provided by the embedder. This CL
replaces mentions to passwordMenu in the component with what it actually
represents: the "more actions" button.
- The event fired by clicking the button was not documented in the file
overview. Moreover, the typedef for the event lived in the embedder. Now
the file exports the typedef.
- Tests that simulated clicks on the button would use the $$('#id')
syntax instead of the $.id one. This is now fixed.

Bug: None
Change-Id: Ie815237134236468c639e73831bdb3eb7db36f98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248561
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779777}