[Autofill Assistant] Fixed EndAction for ShowGenericUi.
Before this CL, it was possible that an action ended before it was
initialized. In particular, this could happen if view inflation failed.
This is problematic, because any calls immediately after SetGenericUi
assumed that the action still existed.

With this CL, most of the action is only initialized after a successful
view inflation.

This CL also includes a small change to how EndAction works. Before, the action was allowed to send data to backend, as long as view inflation succeeded. Now, this is only possible if EndAction is called with an ok() status.

Bug: b/145043394
Change-Id: I8d7dc3d143d37ca4220b9d8f9c40a02964609d3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249766
Reviewed-by: Luca Hunkeler <hluca@google.com>
Reviewed-by: Marian Fechete <marianfe@google.com>
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#779794}