[throughput] Correctly compute expected frames for aggregated throughput
Right now, the |frames_expected| for the aggregated throughput is
simply the same as the |frames_expected| of the impl throughput.
However, there are some edge cases where this could be wrong, such
as the newly added test. In this case, we add jank in the script,
the jank is long enough such that exactly half of the expected
main frames gets presented on screen. In this case, half of the
impl frames reports no damage, and thus the |frames_expected| for
the aggregated throughput is only half of what it should be.

This CL fixes the problem and add a test case. It could causes
some changes to UMA for the:
Graphics.Smoothness.PercentDroppedFrame.*

TBR=sadrul@chromium.org

Bug: None
Change-Id: I3040a2db8a90e7d35ea2785204039b7c764e5435
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2234063
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779912}