[Nickname Management] Wrap long card identifier in migration dialog
Here in the migration dialog the expiration date has the format:
"Exp: 06/21" so it may lead to different length when translated.
So here to be safe I just limited the label width directly, and
hopefully will leave enough space for expiration date.

Uploaded screenshot in bug comment #32

Bug: 1082013
Change-Id: I63759412571ac4f163dee779bd3eb98e83952de3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241647
Reviewed-by: Dana Fried <dfried@chromium.org>
Reviewed-by: Jared Saul <jsaul@google.com>
Commit-Queue: Siyu An <siyua@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779933}