[LayoutNG] Detect break before a float with margin-block-start.
If the block-start edge of a float was above the fragmentainer boundary,
and its margin-block-start would take us past the boundary, we'd fail to
detect the need for a break. The bug was that MovePastBreakpoint()
expects a border edge block-start offset, but we were passing the margin
edge from the float code.

We also need to introduce an additional check before deferring to
MovePastBreakpoint(), because in our implementation, float margins are
unbreakable and also don't collapse with the fragmentainer boundaries
(which MovePastBreakpoint() assumes that we can).

Bug: 829028
Change-Id: If7336055df4e90b7cbf313e843d697f5e4c796a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250063
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779948}