Print Preview: Move PDF resources to a separate grd
Move Print Preview specific PDF resources to a separate grd file from
other resources for Print Preview.

This is in preparation for adding an assert in unpack_pak.py to prevent
unpacking any files in a pak file that are not under the root
directory.

Change-Id: I8145621439e5814383a33e837c6503d0f040eb4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2251100
Reviewed-by: dpapad <dpapad@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779973}