media/gpu/vaapi: tag ColorSpace of decoded VideoFrames
This CL uses the (H264|VP8|VP9)VaapiVideoDecoderDelegate's ColorSpace
for the decoded VideoFrames, if the former is valid. This ColorSpace
comes ultimately from the codec parsers e.g. [1,2] (VP8 doesn't seem to
support it in the VP8 frame itself, so it'll fall back to the container
provide one).

[1] https://source.chromium.org/chromium/chromium/src/+/master:media/filters/vp9_parser.cc;l=296;drc=24d1cda83eb280883bf7950900b7948b4b3f102a
[2] https://source.chromium.org/chromium/chromium/src/+/master:media/gpu/h264_decoder.cc;l=644;drc=fa9094c73631d4b167dc422d62ac40915462423a

Bug: b:147021705
Change-Id: I19c629f3c60a383e842206e18052057affa466ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247330
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780042}