Set flag on browser process to begin supporting delegated ink trails
In order to start having the browser process forward ink points to
viz for delegated ink trails, add a flag onto the RenderFrameMetadata
that lets the browser process know it should start.
RenderFrameMetadataObserver will pick it up and ship it over to the
browser process, where it will be picked up by RenderWidgetHostViewBase
and saved for a future CL, where RWHVEventHandler will use it.

Bug: 1052145
Change-Id: I8305292de6289dcec3fb4083666898161ae1dc41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2240356
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Mario Bianucci <mabian@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#780091}