[Suggested files] Add a full set of circular GM2 file icons
App icons in the suggested chips are circular, but our current set of
file icons are square. Medium-term this will be addressed by the move to
SVG from PNG icons, which will let us change this in-code. Until support
for SVGs is in place, we need a temporary fix for experiments that
add files to the suggestion chips.

This CL adds a full set of circular GM2 file icons, replacing the
handful of chip-customized file icons we had already.

Bug: 1093669
Change-Id: I25b628e6b8c57dfdaf70fc976f44499100bc92c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247988
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780108}