[remoting host] Use control channel for setting min/max bitrates
Previously the bitrates are set from host experiment configs, which can
only be set once before the session is connected, so any changes to the
bitrates require restarting the session.

This CL introduces a new control message for the client to alter
bitrates after the session is connected.

Bug: 1096197
Change-Id: I03eeb3a7951a1cc15a0e994ea995b6f7bc82511c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2253241
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780137}