Migrate content.mojom.HitTestResponse to use frame tokens
Instead of using a routingID for the |hit_frame_routing_id| field,
let's use a frame token instead, which uniquely identifies a frame
as well in a similar fashion to the routing ID, paving the way for
future Onion Souping of content/renderer/accessibility.

This also comes with a shorter term benefit, since we are in the
process of removing AXContentNodeData and, along with it, the
|child_routing_id| attribute that it's currently used to refer
to a child frame being hit on RenderAccessibilityImpl::HitTest().
By using a frame token instead we can remove that dependency on
|child_routing_id|, and therefore helpìng with the process of
removing AXContentNodeData along with other AXContent* classes.

Bug: 1094150
Change-Id: I2091abf6a5195ffd0128eeebc24c214eb5d5d254
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2251827
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Cr-Commit-Position: refs/heads/master@{#780229}