[files-banner] Add max 2-line limit for banner text.
It clamps the banner text in case of passing the limit and adds ellipsis
at the end, so the banner text doesn't overflow over/under the sibling
elements.

There is line-clamp property made for this purpose to do it simply by
css, and it's built on -webkit-box display. As webkit-box deprecated,
the css presubmit fails.

https://www.chromestatus.com/feature/5680142707851264
https://gist.github.com/bfgeek/b1dc3cb741681523ae043045411eb526

So the other solution is setting the property through the js code while
presubmit process updated.

Comparing patchset 2 and 5 shows the difference.

Tested RTL manually and it's fine.

There was a change in behavior on line clamp recently supposed to be
released in M85, as tested it works as expected.

Bug: 1082539
Change-Id: Ie677b367eb9fa28704d89eda7ea54af9ba5a81bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2251859
Commit-Queue: Omid Tourzan <oto@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780230}