media/gpu/vaapi: disable hw VP8 encoding on all Intel Gen 9.5 devices
Hw VP8 Encoding on Intel Gen9.5 devices is disabled for Celeron/Pentium
SKUs but sadly it seems that the same GPU hang is observed in Core
devices (b/139695266, b/158696760).

This CL extends the current blacklisting to all devices with a Gen9.5
GPU, so that said hang can be _temporarily_ avoided.

It also refactors a bit the code to identify if a given SoC has a GPU
of either Gen9 or Gen9.5.

Bug: b:158655609, b:139695266
Change-Id: Id4ebf705e2949fe0715480be0428215d9eae4c75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2253158
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780299}