Do not use transforms when bounds size changes
It is an issue exposed by the app scaling feature. Before app scaling,
shelf icons re-layout usually animates a icon to the target bounds with
the same size. It does not hold in app scaling where a shelf icon's
size may shrink after shelf button addition. What's more, when a
shelf button shrinks from the normal size (56*56) to the dense size
(48*48), the button's icon image changes from 44*44 to 36*36. As a
result, when scaling the shelf button, the icon image is not in the
ideal bounds. It causes bad animation.

In this CL, BoundsAnimator does not use transforms when the animated
view's target bounds have a different size.

Bug: 1019397
Change-Id: I57d9fbaf376e4a304ec3c6d0eedb33a8bb4e0c0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2242791
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780345}