Audio should continue playing when the browser switching to background
Previously,we judge whether the current media contains a video track to
decide to pause the current playback after switching to the background.
Because the media played through MediaPlayer are assumed to contain
both audio and video tracks, they will all pause. In this scenario,
we should judge whether the original size of the current media meta
video is empty to determine whether it contains a video track.

Bug: 1093697
Change-Id: I702465f677d9dd01b827051b21f77c20633a80c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241053
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780359}