Integrate wptrunner screenshots into results viewer.
The wptrunner output file contains a "screenshots" artifact which is a
list of files that were screenshotted, and the base64-encoded png of the
screenshot that was taken. We extract and decode these pngs and write
them to disk. The redundant data is then scrubbed from the output file
and replaced with the path to the newly-created file.

Bug: 937369
Change-Id: Ifeff993023219833e65af39548f556b54ac65b60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2252682
Commit-Queue: Luke Z <lpz@chromium.org>
Reviewed-by: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Robert Ma <robertma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780415}