Override AutofillClock in Captured Sites Tests
These tests record traffic from a particular moment in time and
override the time in the browser on replay to the recorded time.

This change also has the replay override the AutofillClock so that tests
will continue to work even if credit card info or other info is past the
real life time.

Bug: 1090042
Change-Id: I8964ae1818b9a64c9a7417a7aec628d053a8bfa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248346
Commit-Queue: Ian Struiksma <ianstruiksma@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780429}