Remove PositionFixedWithCompositedDescendants compositing reason.
The original CL that added it was:
https://codereview.chromium.org/1413363003

which has to do with the ways in which the pre-BGPT system explained
scrolling and fixed-position elements to the compositor. Since then
BGPT shipped, which no longer requires communicating this via the
graphics layer tree.

Bug: 1096732,950418

Change-Id: I9d6fe98298d69244ad4ca298e934d5c5765cf2da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2253282
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Auto-Submit: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780481}