[Instant Start] Add histograms of FeedContentFirstLoadedTime.
In this CL, we add histogram:
"Startup.Android.FeedContentFirstLoadedTime.Instant(.NoInstant)".
It records the time duration from a cold start till the Feeds articles
are first loaded on the StartSurface. The histograms are recorded as
deferred tasks after startup only if StartSurface is shown when Chrome
is launched from a cold start.

A new interface method ContentChangedListener#onAddFinished() is added.
It notifies FeedSurfaceMediator when the Feeds content is added to the
RecyclerView.

Bug: 1092931
Change-Id: Ice5fb669505f87e17af30daeecc6a2567205d504
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2240131
Commit-Queue: Xi Han <hanxi@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780503}