Inline/rename non-default Settings styles
The PreferenceScreenLayout and PreferenceMediumText styles are
non-standard and not used in the default settings screens, so inlining/
renaming them to discourage reuse.

Fixing the Clear Browsing Data text style is tracked in crbug.com/1081933.
Cleanup of the homepage editor change is not tracked because it will
eventually be replaced through theHomepageSettingsUIConversion.

Bug: 1077375
Change-Id: I955c4364770e68eb84c14fa3832450ccf1f56303
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254714
Commit-Queue: Natalie Chouinard <chouinard@chromium.org>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780519}