Fix the paging issue caused by overflow app.
This fixes the issue reproduced with steps in crbug.com/938098#c15.
On device A, when user moves an item from one page(P1) onto a page with
a full page of apps(P2), the last item of P2 will be overflowed to its
next page(P3) and inserted at the first position of P3 and merged with
the rest of the items of P3. The changes will be synced to device 2,
and should make apps on launcher page look identical to the ones on
device 2. However, the sync change for removing the page break after
the overflowed item is not populated to model updater, therefore, on
device 2, the overflowed item will be placed on a separate page instead
of P3. On device 2, it show one more page.
The following is an example of the bug:

Original layout:
A1 A2
B1 B2 B3
C1

Device 1 after moving A1:
A2
B1 A1 B2
B3 C1

Device 2 after syncing changes from device 1:
A2
B1 A1 B2
B3
C1

With the fix, device 2 should looks the same as device 1 after syncing.

Bug: 938098
Change-Id: I95b2023c56a5a160f8287e7ba28a1bfd2fdbda9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2253726
Commit-Queue: Jenny Zhang <jennyz@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780547}