Remove/replace unnecessary includes of logging.h (chromecast/)
If the file just needs the CHECK/CHECK_OP/NOTREACHED
macros, use the appropriate header for that instead.
Or if logging.h is not needed at all, remove it.

This is both a nice cleanup (logging.h is a big header,
and including it unnecessarily has compile-time costs),
and part of the final step towards making logging.h no
longer include check.h and the others.

TBR=ajgo

Bug: 1031540
Change-Id: I69bc2454eec768047ea61bcae7ca5d44f59eadf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254659
Commit-Queue: Hans Wennborg <hans@chromium.org>
Reviewed-by: Sergey Volk <servolk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780578}