Do not cross-origin prefetch from opaque requestor
One of the security checks the browser process performs for cross-origin
prefetches is a same-origin check between the cross-origin prefetch
request's requestor origin, and the last committed origin associated
with the RenderFrameHost associated with the renderer that made the
request.

However, if the requestor origin of the cross-origin prefetch is an
opaque origin (i.e., a data: URL), it is guaranteed to be cross-origin
with every other origin, and the security check will fail, killing the
renderer. This CL makes a renderer detect if the requestor origin of
a cross-origin prefetch is opaque, and if so, ignore the `as=document`
attribute value.

R=kinuko@chromium.org

Bug: 1096338
Change-Id: Ia8a16067c3cc49ed33102077c81635e029b96f22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2256001
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780606}