[FragmentItem] Check BreakToken when reusing lines
This patch checks |NGInlineBreakToken| when reusing lines,
and stops reusing if it does not have valid |ItemIndex()|
or |TextOffset()|.

The crashes logs are in |NGLineBreaker| where the the line
start offset is beyond the current |NGInlineItem|. If that
happens, |NGLineBreaker| should hit several DCHECK and it
will eventually crash.

This signature has been there for a long time with small
numbers, but it looks like there is a spike for FragmentItem-
enabled builds. From this, a speculation is that we may be
reusing lines even when items are changed.

This patch is speculative. Tested several sites with this
patch but I could not hit the DCHECK added in this patch.

Bug: 1097042
Change-Id: Ie9ebc329ad9ce2a73f8042e7da44d80e997101dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255943
Commit-Queue: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780647}