[FragmentItem] Fix MoveToNextForSameLayoutObject when range is limtied
This patch fixes |MoveToNextForSameLayoutObject| when
|NGInlineCursor| has limit range.

When |NGInlineCursor| is for an inline formatting context,
it can loop until |DeltaToNextForSameLayoutObject| is zero.
However, when |NGInlineCursor| is limited to subtree of an
inline formatting context, the current code fails to stop
at the end of the subtree. This patch fixes such case.

Bug: 1097553
Change-Id: If19b2a1f5651557b38f1cde6f6f073b4a3027c79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2256103
Commit-Queue: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780649}