Initial commit of attestation client for chrome os
This CL enables the communication between the browser and the
attestation service daemon (attestationd). This will be used to replace
all attestation-related APIs that go to cryptohome client at this point.

To test the further setup is correct, we use dbus-monitor to see if
sender is not cryptohomed after the switch in the child CLs.

BUG=b:158955123

Change-Id: I48706f287853ef19465ac92e7e6054ed64c9337a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249273
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Leo Lai <cylai@google.com>
Cr-Commit-Position: refs/heads/master@{#780653}