Fix invalid iterator usage in crostini_manager
CrostiniManager::OnAbortRestartCrostini gets an iterator into
|restarters_by_id_| but dereferences it before checking if it is the
past-the-end iterator, potentially causing a crash. Change this method
to bail out early if the restarter being aborted has already stopped.

This was not caught by the test for this case
(CrostiniManagerRestartTest.DoubleAbortIsSafe) because it incorrectly
terminates as soon as the first abort finishes, instead of waiting for
the second. Fixing this causes the original code to crash under MSAN,
and passes with the new code.

Bug: 1096979
Change-Id: I67322c5a8823dea4e517ba04d5fbf110b200e8d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2251644
Reviewed-by: Jason Lin <lxj@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Cr-Commit-Position: refs/heads/master@{#780664}