X11 and Ozone: //chrome: choose code path with IsUsingOzonePlatform
This patch adds usage of ozone platform in //chrome.

It's not possible to use linux wrapper [1] for everything as it
may introduce circular deps and other problems. Plus, there
are many places that are guarded either by single USE_X11
or USE_OZONE. Thus, use the IsUsingOzonePlatform feature
flag to choose the correct path.

[1] https://crrev.com/c/2250106

-----

Please note that this ugliness is added temporarily and
will be removed as soon as use_x11 is removed (hopefully
by Q1 2021 depending on how the finch trial goes).

Please also note that it's impossible to build use_x11 && use_ozone
without some hacks in PlatformCursor code. The changes to that are
on their way to upstream.

----

Bug: 1085700
Change-Id: I01653d42a448e708bc772ac2ec4aa64dd3f178b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250066
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780730}