[testing] Fix webtest arguments for V8 integration builder
The builder is setting release gn args, but uses V8 in debug mode.
After args.gn was included in the web tests isolate, a check started
failing, flagging release GN args on a debug builder.

This change attempts to run webtests as in release mode but with
a debug out directory. We'll rename the builder in a follow up to
be less confusing.

No-Try: true
Bug: v8:10626
Change-Id: Id79d8717917ca6a23e391f16cc40efd3544a915f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255078
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780750}