Remove/replace unnecessary includes of logging.h (ios/)
If the file just needs the CHECK/CHECK_OP/NOTREACHED
macros, use the appropriate header for that instead.
Or if logging.h is not needed at all, remove it.

This is both a nice cleanup (logging.h is a big header,
and including it unnecessarily has compile-time costs),
and part of the final step towards making logging.h no
longer include check.h and the others.

Bug: 1031540
Change-Id: I161278f3bf0b77eef2d98b2819ba0c46676b94f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254125
Auto-Submit: Hans Wennborg <hans@chromium.org>
Commit-Queue: Eugene But <eugenebut@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780849}