Fix hit test region for tab group headers.
Previous to this CL, the hit test area for a tab group header would
extend past the separators on either side of the header to the edge of
the view, which is not the case for other tabs (which use their stroke
shape to clip the event targeting area).

Since group headers are only highlighted during drag, there's no need to
use the curvilinear tab border path; we can just use the rectangle
between the tab separators to mask mouse events.

Bug: 1097303
Change-Id: I910e464df6323d3c2915a2a06c1d10d70a84a66f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255742
Reviewed-by: Charlene Yan <cyan@chromium.org>
Commit-Queue: Charlene Yan <cyan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780862}