Add histograms for detected network printers
This adds histograms Printing.CUPS.TotalNetworkPrintersCount
and Printing.CUPS.NearbyNetworkPrintersCount which will
record the number of detected network printers in a user's network
and the number of them that are not already saved, respectively.

Histograms will be recorded when the user navigates to the OS
Settings Printing page.

It extends the CupsPrintersManager API to emit the relevant
histograms with public API method: RecordNearbyNetworkPrinterCounts.

Bug: chromium:1012515
Test: unit tests
Change-Id: Ifa5f111c3815feb81b1a07b9b6a719a8ada3c8cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241029
Reviewed-by: Sean Kau <skau@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Commit-Queue: Jesse Schettler <jschettler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780884}