ContentVisibility: Fix two flaky tests which check layout state.
During the start of the test, we do some preliminary checks in the
tests to ensure layout is clean, which sometimes fails.

The checks are not actually necessary for the test, since the first
several states set up the correct dirty-bit state anyway. So this patch
removes the initial checks since they don't have any bearing on the test.

R=chrishtr@chromium.org

Fixed: 1096321
Change-Id: I390608b98f3ce39217a8b0fddca16b1d343b158f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2256420
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780963}