[sync] Fix BackendMigrator handling directory types only
UserShare is currently deprecated because it belongs the pre-USS
implementation of sync. Its use outside components/sync/engine* is an
indicator the bugs may exist, as is the case for BackendMigrator, which
prior to this patch consumed the legacy Directory state to determine
which datatypes are purged.

Instead, DataTypeManager interface is extended to explicitly represent
this state, although it's not possible to implement the desired logic
reliably (TODO added).

To avoid similar issues, the exposure of UserShare is removed from the
main sync APIs.

Bug: 923287
Change-Id: Iba0bc421416ae6a11787cecc4ecc4e45c9eff034
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2253848
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780972}