CrOS Shelf: Fix view targeting for shelf navigation home button
When the home button is tapped as the only button in the shelf
navigation widget, ViewTargeterDelegate::TargetForRect() would
sometimes return the incorrect view, because the center point of the
widget is closer to the tap than the center point of the home button.

The solution to this is to ensure that equal spacing is added on both
sides of the home button within the shelf navigation widget. This
ensures that any tap is equidistant from both the root view center
point and the home button center point. A test to check for this
case has been added.

This solution standardized the spacing added to the shelf navigation
widget so that the problem is solved in both horizontal and vertical
shelf.

Bug: 1083713
Change-Id: Iaeeb1bb1e945b507190927c2483fb733a26c57f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249047
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780977}