[Switch Access] Improve recovery from erroneous conditions
In cases where the function failing will not break other code, change
from throwing the error to logging it (so the rest of the execution
continues). In cases where the code cannot continue if the function
fails, request that the code recover to a valid state after throwing the
error (by setting a timeout of 0ms).

AX-Relnotes: n/a.
Bug: None.
Change-Id: I5be1f24d0a9c399975d21d3f332933014aaaa376
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2253239
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/master@{#780991}