[M84][Extensions] Fix unsafe variable access in set_icon.js
set_icon.js checked the `self` variable to determine whether the script
was running in a ServiceWorker (such as a ServiceWorker-backed
extension). However, this is unsafe because, unlike `window`, `self` can
be overwritten by other script. If an extension specified a value for
`self`, that would be used instead - and would subsequently crash if the
value didn't support the `in` operator (i.e., wasn't an Object).

When this happened, the custom bindings would fail to register any
hooks. This, then, would result in API calls to browserAction.setIcon()
in passing the arguments directly to the browser-side handling. When
this happened, the browser would classify it as a bad message (because
it didn't undergo the expected argument modifications), and would kill
the renderer.

Fix this by adding a native binding to return whether the script is
running in a ServiceWorker context. Add a regression test.

TBR=rdevlin.cronin@chromium.org

(cherry picked from commit bb686c7922639d306cf948d70e6ee9e35cb306e8)

Bug: 1087948
Change-Id: I1bb593af735dec6d7a2a25050e62677f23c23dce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2253284
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#780315}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2258414
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/branch-heads/4147@{#744}
Cr-Branched-From: 16307825352720ae04d898f37efa5449ad68b606-refs/heads/master@{#768962}