Ad metrics: Add heuristic to estimate proportion of ad frames throttled.
Our CreativeOriginStatus metric is recorded as unknown for approximately
70% of ads.  We want to see how much of this indeterminacy is due to
frame throttling, so we add another histogram that splits by whether
there is any unthrottled frame in the ad frame tree.

The CreativeOriginStatus metric relies on FCP, which for cross-origin
frames can be very large. This is partially due to the flow where: a
frames starts outside the viewport and is throttled; frame is later
scrolled into the viewport, becomes unthrottled; frame receives FCP.

Because paint-timings are relative to navigation start for the frame,
the values are skewed. This change propagates some additional info, a
best-effort timestamp of when a frame is first not throttled, up to the
browser process to address this.

Continuation of
https://chromium-review.googlesource.com/c/chromium/src/+/2088037
which had inexplicable trybot failures.  This is the same code ported
into a fresh CL.

Tests:

third_party/blink/renderer/core:unit_tests
out/Default/blink_unittests
--gtest_filter=*PaintEligibility*

chrome/test:unit_tests
--gtest_filter=*CreativeOrigin*Throttling

chrome/test:browser_tests
--gtest_filter=*CreativeOrigin*

Bug: 1049125
Change-Id: I8c9ddd41870ce9797ff0f01de853f7576efd4606
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2161606
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781029}