WebUI Tab Strip: Update animations to use indices
This CL updates the animation function to use the previous and new
indices of the moved tab. This allows for more efficient animation of
pinned tabs (CL coming soon) as indices are needed to determine the
vertical and horizontal movements of the pinned tabs within a grid.

Bug: 1082344
Change-Id: I9835baa597fac941df83c3113112bd8c18d44bd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2258333
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781057}