[Shelf] Apply App Scaling to hidden/extended shelf
Currently the hidden/extended shelf does not update the hotseat density
after screen rotation. This CL fixes the issue by calculating the
hotseat density based on the shelf target bounds in home launcher state.
In order to achieve it, code to calculate the hotseat target bounds is
refactored and the function to return the target bounds size for the
given hotseat state is created.

Bug: 1097504
Change-Id: If1bf3c677d997e2aabc1e7333fff101dbf625dd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2256155
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781072}