Fix shelf app button drag with context menus
To support app item drag operation for items for which a context
menu has been show, the shelf app button context menu is set up to
redirect gesture events to the menu owner - i.e. the hotseat widget.
While this works in the cases where the user keeps the touch over the
app icon while the context menu is shown, in certain situations the
gesture event would not reach the app button view, leaving the button in
drag state even after the gesture completion:
1.  it the touch moved just outside the button bounds as the context
    menu was being shown, the redirected gestures could get consumed
    by the shelf view
2.  if another touch point was added, the combined gesture may end
    up falling outside the hotseat widget, in which case the redirected
    gesture event would not reach the shelf at all.

To address 1. the HotseatWidgetDelegate view overrides the view targeter
logic to return the app button for which a context menu is shown (if
such button exists).
To address 2. the HotseatWidget clears the drag state for the app button
for which the a context menu is shown when it receives a gesture end
event directly (without relying on Widget::OnGestureEvent to deliver the
event to the app button view).

BUG= 1087539

Change-Id: I4a4b8b7cad16e0b00daedd8f8e2073cea6f1a82d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255315
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781089}