[Mac] Handle a child frame for TextInputClientMac
This CL fixes the problem that GetCharacterIndexAtPoint or
GetFirstRectForRange from TextInputClientMac couldn't handle a child
widget. When TextInputClientMac requests mojo calls for them, it
tries to find the focused frame but it couldn't handle a child widget
since it uses RenderViewHostImpl and RenderViewHostImpl is nullptr
when the widget's owner_delegate is empty. So, it introduces
GetFrameTree to RenderWidgetHostDelegate, gets FrameTree* through
the method, and finds the focused RenderFrameHostImpl using FrameTree.

Bug: 1094841, 1007365
Change-Id: I8da969886ef8d99347c09ae226bcbdeefc84ec0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2245450
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Julie Kim <jkim@igalia.com>
Cr-Commit-Position: refs/heads/master@{#781099}