Add feature flag to gate Quick Answers setting sub toggle.
Based on UI review feedback, we want to expand the scope of related info
setting toggle and remove the Quick Answers sub toggle.

We are going to proceed with two steps:
1. Gate the sub toggle with feature flag that is disabled by default in
case we want to turn it on.
2. Clean up the sub toggle code after launch and confirmed that it
doesn't significantly impact related info opt in rate.

This CL is the first step.

Bug: b/159166868
Test: unit tests
Change-Id: Ia0e87112c69251914b3bae9ae3f3a0c0d1bc96ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250475
Commit-Queue: Li Lin <llin@chromium.org>
Reviewed-by: Yue Li <updowndota@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781151}