[Adaptation] Make use of ThermalResource in WebRTC.
If we receive a thermal signal we lazily instantiate the
ThermalResource and inject it into the webrtc::PeerConnection.
The ThermalResource is wired to respond to new thermal measurements.

This will make the WebRTC video pipeline be able to react to overuse
or underuse due to thermal signals.

The ThermalResource is only used if the corresponding base::Feature is
enabled, which is DISABLED by default. Both code paths are tested.

Bug: chromium:1094844
Change-Id: I9f47c1f4f9bc98349b328d71f0eb71cc4c398444
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247837
Commit-Queue: Henrik Boström <hbos@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781223}