Add a use counter for overlay popup
We used to measure overlay-popup-ad. This CL adds another use counter to
measure the overlay-popup in general, including both ads & non-ads.

This is achieved by adding extra flags to OverlayInterstitialAdDetector
to denote whether the current candidate is an ad, as well as whether we
have encountered ad and/or general interstitials.

Bug: 1048136
Change-Id: I6dc809ceed51aa9b9747af44d8db6ed38721bc2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250957
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Bryan McQuade <bmcquade@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781231}