Hide ASK settings from PageInfo in incognito mode
When a user grants a permission in regular mode, we replace the ALLOW
content settings with an ASK setting in incognito mode. These settings
should not be shown in the UI if ASK is the default setting.

The current behavior also fails an assert in
PermissionParamsListBuilder::createPermissionParams, which is fixed
by this CL.

Bug: 1092809
Change-Id: Ic74f77cca11b6d3ff0839e6f908002cf76838c21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2235822
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Commit-Queue: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781249}