Fix: Avoid embedding cr-toast-manager twice
PasswordsListHandler contains a cr-toast-manager. PasswordsSection and
PasswordsDeviceSection both embed PasswordsListHandler. This means when
one navigates from one page to the other, Chrome attempts to create two
instances of cr-toast-manager, which is supposed to be a singleton. This
causes error messages to be displayed on Dev Tools.

This CL replaces the toast manager with a <cr-toast>, solving the
problem. We also use the occasion to fix the arity of the callback
passed to the requestPlaintextPassword() call in the file.

Fixed: 1096040
Change-Id: I07159913b6a8492118f1eefa09faae6d26874ff5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254281
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781291}