[gamepad] Add mappings for ELECOM pads in "D" mode
This CL adds mappings for the JC-U4013SBK (the wired version) and
JC-U4113SBK (wireless) when they're in "D" (DirectInput) mode.
(https://crrev.com/c/2252469 adds XInput support for the JC-U4113SBK,
and XInput support for the JC-U4013SBK is still to come.)

Bug: chromium:1097173, chromium:1096754, b:159466838, b:159368181
Test: Connect gamepads in "D" mode, load gamepadviewer.com, and check
that pressing buttons and moving axes is correctly reflected on-screen.

Change-Id: I30dae3bd586b5b4ceeccb0a4ae3c02bb8a91447a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2259012
Commit-Queue: Harry Cutts <hcutts@chromium.org>
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781466}