webauthn: don't send hmac-secret when not supported.
Previously we would send the hmac-secret extension to authenticators
that didn't advertise support for it. That's fine because it'll be
ignored but, in the future when we have more support for hmac-secret, it
gets expensive.

This change adds the basic infrastructure for that and stops sending the
extension unless the authenticator advertises support.

Change-Id: I0dc2806c91ab0e032598c1f9ce54899c0b47d269
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255210
Commit-Queue: Adam Langley <agl@chromium.org>
Auto-Submit: Adam Langley <agl@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#781479}