Create pdf/ppapi_migration directory
Introduces a new directory for temporary code related to the PDF
viewer's PPAPI migration. Code in this directory should abstract away
usage of PPAPI, so it may easily be removed after the migration is done.

Also moves the existing pdf/geometry_conversions.{cc, h} files, as an
early example of the kind of code that should be in this directory.

Bug: 1061182
Change-Id: I01372843f4149a607ef1b6dab80277986c177d4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2261413
Commit-Queue: K Moon <kmoon@chromium.org>
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781536}