Implement IItemContainerProvider & UIA unique id property
The motivation behind this change is for ATs to be able to switch from
an IA2 element to its corresponding UIA element. In order to achieve
the above, this CL implemented the following:
- IItemContainerProvider::FindItemByProperty for finding an
UIA element through its unique id.
https://docs.microsoft.com/en-us/windows/win32/api/uiautomationcore/nf-uiautomationcore-iitemcontainerprovider-finditembyproperty
- A custom UIA property id that maps to an element's unique id.
- Introduced ui::UiaRegistrarWin class for managing custom UIA
  property/event.

AX-Relnotes: ATs will be able to lookup an UIA element from and IA2
element.

Bug: 1084630
Change-Id: Ief9664ff04224972af6607304c6893c25772f95d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2242046
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Victor Fei <vicfei@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#781539}