Extensions: Add Metrics for Cast Extension usage
This CL adds a UKM and UMA metric to be logged when the Media Router
Component Extension (the internal extension used to support cast
scenarios) loads a file. This scenario is done as part of loading the
cast sender SDK, and this change will be used to support a migration
away from this hosting strategy for the SDK and an eventual migration
away from the extension.

Bug: 1084875, b/157107974
Change-Id: I5cba444a4534dcbdbd54a57408f3c4030f0aaff8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249054
Commit-Queue: Ryan Keane <rwkeane@google.com>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781570}