MediaLauncherActivity: Don't crash on SecurityException
When a user tries to open a media file with Chrome that Chrome doesn't
have access to, we currently crash with a SecurityException. Instead,
this CL will just catch and log the error.

This CL also adds a new histogram to track whether the launch was
successful.

Bug: 901292
Change-Id: I46dd6faf5be0a18410fb7dcfa89bcf662c9a46a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255133
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781587}