Simplify SinkDocument construction for plugins forbidden by sandbox
Historically, the sandbox flags for a Document were finalized in the
Document constructor. This meant that when a PluginDocument should be
blocked by plugin sandboxing, we constructed the PluginDocument, then
immediately replaced it with a SinkDocument once we detected that we
should have blocked the PluginDocument.

Now that the sandbox flags are finalized in DocumentInit, we can
inspect them before creating the PluginDocument and instead go
straight to constructing the SinkDocument.

Bug: 1029822
Change-Id: Id859fe817cddf81f0e83b8ded53279527c86bcb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2261094
Commit-Queue: Kentaro Hara <haraken@chromium.org>
Auto-Submit: Nate Chapin <japhet@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781623}