SharedImageBackingGLImage: Simplify bind and copy
SharedImageBackingGLImage doesn't correctly bind and copy GLImages to
GL textures. There are some rebinds and copies sprinkled here and there
in haphazard ways that only happen to be right most of the time.

Update SharedImageBackingGLImage to have a boolean indicating if the
GLImage has been changed and a re-bind or copy is needed. This variable
- starts true
- is set to false at a re-bind or copy
- is set back to true in the Update method

Do a bind at initialization because various other bugs require it for
now (I'll get to them later). Do a re-bind or copy lazily whenever Skia
or GL access is requested.

Skip the re-bind or copy when the mode
GL_SHARED_IMAGE_ACCESS_MODE_OVERLAY_CHROMIUM is specified, because
that type of GL access will access the GLImage directly. When
GLRenderer is removed, this mode can be removed.

Bug: 1092155
Change-Id: Ia5eb4ec87497ee14b06eb08404af312542926a8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255884
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781644}