vulkan: Add VulkanSwapChain::PostSubBufferAsync()
The bug is because X11 will present hided Window at 1Hz. So if there are
two windows on screen. One is hided, vkAcquireNextImageKHR() will be
blocked for 1 second, so the GPU main thread will be blocked with it.
Fix this problem by add PostSubBufferAsync(). It will call
vkAcquireNextImageKHR() off the GPU main thread.

Bug: 1097014
Change-Id: I2dc7988d5c6320d167b3280b3f92ac9979644887
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2256589
Commit-Queue: Peng Huang <penghuang@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781654}