Ensure no frames have been decoded for video context loss tests.
Once we have a valid video frame back from the decoder, we assume the
decoder is good and toss the encoded data we hold onto for fallback.

It's unclear to me how the current tests have ever worked. Since we
only fire canplaythrough after 4 frames have been decoded.

The fix for this is to ensure we don't decode any frames before
triggering the gpu crash. This is fine for the purpose of this test
which is for testing the rendering and not decoding.

R=kbr

Bug: 1082761
Change-Id: I7e798ff219ded36582b1576919b6e5bf9541f5fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2258772
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781665}