[Streams] Deflake transferable/readable-stream.html
The "cancel should abort a pending read()" subtest of
http/tests/streams/transferable/readable-stream.html was flaky because
it expected a delay(0) call to complete sending all pending messages,
which occasionally didn't happen on the bots.

Switch to waiting explicitly for the call to cancel() to happen to
ensure that the test is complete.

BUG=1073431

Change-Id: I66c2f8b726a5bb6c56596eb9b7327c1458423a04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2260094
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781699}