Add CHECK to WidgetObserver subclasses
WidgetObservers need to be removed from all ObserverLists before
destruction otherwise it will CHECK fail when iterating over the list.
The CHECK is added to all subclasses to retain readable stack traces in
release builds.

The goal is to find all instances where the observer is not removed as
that causes a hard to reproduce failure.

AX-Relnotes: n/a.
Bug: 1055483
Change-Id: I481ee30e681220ff697c7fbb2380470f17cfd465
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254598
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Richard Knoll <knollr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781737}