[Passwords][Leak Check] Add more tolerance to warning icon halo
On Windows, the existing warning halo might not be enough (on some fonts
at least) to contain the entire warning triangle. To mitigate that, make
the halo a bit bigger and the triangle a tiny bit smaller so that the
rounding errors that cause the misalignment aren't visible anymore.

Additionally, the dark mode still used pixel measurements which did not
scale with the text size. It's using the same height units now.

(Tested on Windows and on Linux; Screenshot in linked bug)

Bug: 1098238
Change-Id: I037a82403eb1acfbe9173500d9fb85cfd34a4e08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2260639
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781764}