X11 and Ozone: //components: choose code path with IsUsingOzonePlatform
This patch adds for usage of ozone platform in //components.
And changes the following patterns:

#ifdef(USE_X11)
DoXxx();
#endif

to

#ifdef(USE_X11)
if (!IsUsingOzonePlatform())
  DoXxx();
#endif

and

#ifdef(USE_OZONE)
DoXxx();
#endif

to

#ifdef(USE_OZONE)
if (IsUsingOzonePlatform())
  DoXxx();
#endif

-----

Please note that this ugliness is added temporarily and
will be removed as soon as use_x11 is removed (hopefully
by Q1 2021 depending on how the finch trial goes).

Please also note that it's impossible to build use_x11 && use_ozone
without some hacks in PlatformCursor code. The changes to that are
on their way to upstream.

----

Bug: 1085700
Change-Id: Icce90019c12db275a796dcab993abffa158023af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250006
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781766}