Fix include to use shared mojo definitions
The file ads_page_load_metrics_observer.cc included a
mojo definition file from blink, instead of the shared
version. This resulted in flaky build failures, as the
blink version of the mojo is not a GN dependency, and
was sometimes not yet generated.

TBR=jkarlin@chromium.org

Bug: chromium:1095617
Change-Id: Id6930a4f7af08c6ff0344f853dea7b79ce2d3748
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2262613
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Johann Koenig <johannkoenig@google.com>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781777}