[Telemetry SWX] Introduce probe service converters
Introduce probe service converters.

It's a good practice to use mojo::StructTraits or it's preferred to
use mojo:TypeConverter even when it's deprecated, but in our case we
may need to change converter in the runtime. For example, depending
on some policy value we may/may not provide some fields:
https://chromium.googlesource.com/chromium/src/+/refs/heads/master/docs/security/mojo.md#use-structtraits

Bug: b:158658869
Change-Id: Ibfe1ef41d71e4dcb90581354135b1a9d44d28a9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2252038
Commit-Queue: Oleh Lamzin <lamzin@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Mahmoud Gawad <mgawad@google.com>
Cr-Commit-Position: refs/heads/master@{#781783}