[LayoutNG] Clean up block-size calculation for fragmentation.
And even (unintentionally) fix a bug in nested multicol!

Avoid duplicated work and duplicated code. We used to have separate code
paths for setting the block-size, depending on whether we were in a
block fragmentation context or not. In the case of the block layout
algorithm, we even calculated the block-size twice when we were
block-fragmented: First at the common place (in FinishLayout()), then
we'd just discard that, and calculate it over again, but this time
correctly for block fragmentation (FinalizeForFragmentation()).

Let's just do it correctly the first time instead.

Cleaning up the code allows us to call FinishFragmentation() even when
the fragmentation block-size is unknown (initial column balancing pass).
This actually fixed a bug, which I discovered because a unit test
started to fail, but that was just because of the fact that the test
assumed that there was no column balancing, even though column-fill was
'balance' (initial value).  Set it to 'auto' instead to get expected
behavior. Also added a web test that demonstrates what got fixed.

Introduce two setters and getters for block-size in
NGBoxFragmentBuilder. Regular layout will set the CSS box size (as if
all fragments were stitched together into one tall fragment), then pass
it on to the block fragmentation machinery, which will set the actual
fragment size.

I also realized that intrinsic block-size makes no sense when a node
fragments (nobody uses it in such cases), so we'll now DCHECK that
there's only one fragment.

Bug: 829028

Change-Id: I820e1b281ba19bd8e39b1b55f52288000d8b802c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255365
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781784}