[Py3] Make build.py python3 compatible
Couple of fixes here include:
- change `file` to `open`
- add `universal_newlines=True` to subprocess calls, since the output is
later compared to a string
- change buffer to line buffer, this is the only "breaking" change that
might cause a different behavior.

For the buffer behavior, unbuffered output is only allowed in binary
mode, but we are always printing strings. So the compromise here is to
use line buffering (we could also omit it).

Bug: v8:9871,chromium:941669
Change-Id: I5c1e69da022d13f4234921fe5fe717d53c96f058
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2252867
Commit-Queue: Zhi An Ng <zhin@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/master@{#781911}