mac: minor tweaks to gcapi_mac target.
No behavior change.

The 11 SDK linker warns that building for both 10.6 and 10.7 is
deprecated and we build with -fatal_warnings, to that's a build error.

Ideally we get a flag to disable this warning. Else we could not pass
-fatal_warnings for this target and have a linker wrapper again to
filter out the warning ( :/ ), or we could bump up the deployment
target (which after this change requires changing only a single
number instead of 2). But that's for a future CL.

Bug: 1098738
Change-Id: I7952345f1333be023fab997cbf81baf394dfcc40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264592
Auto-Submit: Nico Weber <thakis@chromium.org>
Commit-Queue: Mark Mentovai <mark@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781913}