[iOS] Fix NTP animation when opening tab via keyboard
This CL changes the order in which the BVC/the omnibox is being focused
when opening a new tab from the TabGrid using keyboard shortcut.
Previously the omnibox was focused then the BVC, leading to an
animation glitch where the omnibox and the doodle disappeared.

This CL fixes it by focusing first the BVC then runs the completion of
the animation (which contains the focus of the omnibox), letting the
animation follow its natural course.

This should be relatively safe as the only completion block passed in
the TabGrid animation is the focusing of the omnibox.
So inverting the order of the calls should only have an effect in that
case.

The launchMask isn't used when BVC is contained.

Fixed: 1096517
Change-Id: I6a53e99e5c1a598401c668390c177e4bef2fc06a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264622
Reviewed-by: Bindu Suvarna <bindusuvarna@google.com>
Commit-Queue: Bindu Suvarna <bindusuvarna@google.com>
Cr-Commit-Position: refs/branch-heads/4181@{#3}
Cr-Branched-From: c3466f05bc9969c4018f3e12c96373eaf95c0a3e-refs/heads/master@{#781470}