WebLayer: share directly to WebApk as appropriate
This matches the external intent handling in Chrome, where a WebApk is
opened without a chooser dialog when it's the only non-browser handler
for an intent.

WebApkValidator, WebApkVerifySignature, and some constants are moved
to //components/webapk in a directory and build target structure that
mirrors //chrome/android/webapk, i.e. separate "client" and "common"
libraries.

Bug: 1063874
Change-Id: I55fb0968489a4a919bb383b182413e913091a020
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254350
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781934}