[GPU] Explicitly initialize the base::ThreadPoolInstance in time for angle::InitializePlatform
This is equivalent to the ThreadPoolInstance configuration that the
GPU process was implicitly getting from ChildProcess::ChildProcess
(constructed from GPUProcess::GPUProcess). Modulo the added ability to
explicitly name this ThreadPool and have independent metrics :).

The ThreadPoolInstance is still started at virtually the same order as
before so semantics should be equivalent for the GPU process.

This is necessary otherwise etienneb's fix for crbug.com/1091259
is a no-op in production due to the early-bail if there's no
ThreadPoolInstance on init (added for tests, breaks prod).

R=chromium-metrics-reviews@google.com, sunnyps@chromium.org

Bug: 1091259
Change-Id: I3ce25ac70bb4d1980e3d38ae6f90851041632bd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255655
Commit-Queue: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Auto-Submit: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781944}