content: adds RenderFrameMetadataObserverClient.OnRootScrollOffsetChanged
As part of making GestureStateListener.onScrollOffsetOrExtentChange
work again (
https://chromium-review.googlesource.com/c/chromium/src/+/2222623 )
I made it so that the browser gets any change in RenderFrameMetaData
when an infobar is present and the root-scroll-offset changes.
Unfortunately this regressed some perf tests. The test are triggering
an infobar to show, and then scrolling through the whole page without
dismissing. Uma stats indicate this isn't that uncommon, so it seems
worth fixing.

This patch creates a one off message (OnRootScrollOffsetChanged)
that is sent if root scrolls have been requested and the frame would
not normally be sent to the client.

https://pinpoint-dot-chromeperf.appspot.com/job/139b3471120000
is a pinpoint run that shows this gives back some portion of the
loss. In order to get back all the performance, a more complex
patch is likely needed. Given branch is Thursday, I'm landing this
patch now to give back some of the performance.


BUG=1091451

Change-Id: I54cc55a6dab1dc716978eb7d25cb932bfb662ee5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2257407
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781951}