Don't show invert bubble on Mac in increased contrast mode
http://crrev.com/c/2118693 caused the invert bubble to start showing up
on Mac in increased contrast mode when a new incognito window was
opened. http://crrev.com/c/2247407 addressed this temporarily by
disabling the bubble for incognito windows.

This change is a follow-up to http://crrev.com/c/2247407 as a more
permanent solution to the original problem. The invert bubble should
show up in the High Contrast dark theme on Windows. However,
GetHighContrastColorScheme() was calling into UsesHighContrastColors()
to detect this.

The problem with this is that on Mac, UsesHighContrastColors() has a
different meaning than on Windows. To fix this, make
GetHighContrastColorScheme() a virtual function and only return the
High Contrast theme on Windows.

Bug: 1095855
Change-Id: I24a7ed2dd5b9194a795d1560ba1e5f4be85afd98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250224
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#782052}