Add "New" badging for menu items as a user education feature.
Behind a flag (nothing happens if flag is not set).
This is part 1. It introduces the flag which enables the badging, as
well as the ability for menu items to be badged.

However, the pipeline to hook up badging is not included. Follow-up CLs
will include:
1. hook-up from SimpleMenuModel
2. specific items to be badged for release

See: 
https: //docs.google.com/document/d/1vczPpUCnYOKvSQtSWMYX8K1mK-vhrjwsAyRKFv22eYk
Change-Id: If8900433015e71802b5deb0e32bebb9477ea8d62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208794
Reviewed-by: Michael Wasserman <msw@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782090}