[omnibox] Fix Section Headers text color in dark mode
Before this CL, we always set the Headers text color to GoogleGray700.

This is not correct, and this CL fixes it so that it's dynamically
calculated by the theme just like the dimmed results text.

In practice, the result of this calculation is still GoogleGray700 in
the standard light-mode configurations, but is enhanced to be a better
contrast ratio in dark mode.

Bug: 1095205
Change-Id: Ieee2bd1db87b8806cb8cac8dcf0e30cc9859d6df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264765
Auto-Submit: Tommy Li <tommycli@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782093}