Change chrome.fileSystem.chooseEntry() behavior for extensions.
For extensions, instead of having a per-extension last-saved location,
use the browser downloads location.

Adjust some file_manager tests to pass with this change. In production,
file_manager is not affected by this because it is an app and not an
extension. However, some of its tests use a test extension, so the tests
are affected.

Bug: 1011587
Change-Id: I62d08a9b1855ee8e5454098d37ca5a8a685f9d22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2245686
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782110}