ContentVisibility: Ensure to keep auto f-i-p target unlocked for 2 frames.
This patch ensures that when we have a find-in-page target, we keep
it unlocked for two frames, so that the visibility detection can take
place and if needed keep the context unlocked.

The root problem here is that sometimes our contain-intrinsic-size is
too small, so that when we find a match and scroll it into view,
we actually scroll past the locked size by enough distance that
the container remains locked, so we never see the match.

With this patch, the container is unlocked for two frames which is
the least amount needed for visibility to kick in and start being
unlocked.

This patch also cleans up the previous focus-clearing bug to use
the same mechanism of unlocking-until-lifecycle.

R=chrishtr@chromium.org

Change-Id: I75a2d8b2f41335d5d6fb7a4ec0f42d93438807e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264756
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782115}