weblayer: Scroll direction change notification
There is no change on the renderer side. Just hook into the existing
RenderFrameMetadata::new_vertical_scroll_direction and bubble up that to
the embedder. Compute the scroll ratio in RWHVA.

The API is specifically made to be extensible, and new types such as
scroll start or end may be added in the future.

Bug: 1076473
Change-Id: I8c539588468d39da9095b0eb23bb61249d06b7bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264760
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782133}