Opt in to the Windows "segment heap"
Adding a SegmentHeap entry to the chrome.exe manifest will tell
recent-enough versions of Windows (20-04 and beyond) to opt chrome.exe
into using the segment heap instead of the legacy heap. Details are in
the bug but it appears that the default Windows heap is tuned for server
workloads where throughput is what matters most, and Chromium
(especially due to its multi-process architecture) also has to care
about memory footprint. Experiments with per-machine opting-in to the
segment heap for chrome.exe suggests that this could save hundreds of MB
in the browser and Network Service utility processes, among others, on
some machines. Actual results will vary widely, with the greatest
savings coming on many-core machines.

In order to avoid build warnings this change depends on switching to
building with the Windows 10.0.19041.0 (20-04) SDK.

The initial investigation of this is in crbug.com/982452.

Bug: 1014701
Change-Id: I217d045ee0d365ac57f7a73b0f5ed456469214c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163163
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#782134}