[fuchsia] Fix bug in runner_logs.IsEnabled().
The runner_logs.py IsEnabled() function should check both that _instance
is not None and that the instance._log_dir directory is not None instead
of only checking _instance.

This is because if the test runner scripts --runner-logs-dir argument is
not set, the runner_logs.py _instance is created with its _log_dir
member initialized as None.

Bug: 1067356
Change-Id: Id9fe98e0e31d8ce2256ef3bd3b24f895983a14b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2261635
Reviewed-by: Kevin Marshall <kmarshall@chromium.org>
Commit-Queue: Stephen Roe <steveroe@google.com>
Cr-Commit-Position: refs/heads/master@{#782171}