ui/base/clipboard: Add logging for read and write of various formats.
Add logging for read and write with the platform clipboard. This logs
all reads and writes, as well as their formats, that go through Chrome
and hit the clipboard, from all endpoints (context menu, shortcut, web
APIs, etc). This may be useful for determining usage of various formats
and their change over time.

An enumeration is used to understand which format is used for each read
or write.

Function calls for NOTIMPLEMENTED() reads or writes are not logged. The
following clipboards are also omitted, as these don't really reflect
use on "real" platforms:
- TestClipboard[1], which is used for automated web / WPT tests, and
  the MockClipboardHost.
- HeadlessClipboard[2], which is used for Selenium, Puppeteer, and other
  automated, non-UI platforms[3].

[1]: https://cs.chromium.org/chromium/src/ui/base/clipboard/test/test_clipboard.h
[2]: https://cs.chromium.org/chromium/src/headless/lib/browser/headless_clipboard.h
[3]: https://developers.google.com/web/updates/2017/04/headless-chrome

Bug: 1079036
Change-Id: Ia90c447018221a839842497896cdc1b8f8d13e09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255741
Commit-Queue: Darwin Huang <huangdarwin@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782216}