Fix PDF table attributes.
When I originally added these attributes, I conflated "name"
and "string", but they're not the same thing - a name is
alphanumeric but part of the spec with a specific meaning,
and a string is where the possible values aren't predetermined
and need to be more arbitrary.

Furthermore, node IDs are ints in Skia's tagged PDF APIs, but
stored as strings in PDF files, and the logic for converting
ints to strings shouldn't be duplicated.

This change fixes table attributes to use name and string types
correctly according to the spec, and uses a new Skia API for
adding node ID attributes where Skia converts to a string.

Bug: 1039816
Change-Id: Id1bbd84c38b0ef4f9ee6cbf91c209f5f28d2f510
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2251058
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Ben Wagner <bungeman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782265}