[Suggested files] Remove setting/subpage results when showing parent.
We want to deduplicate results in this case: a setting or subpage result
is present, and its parent subpage result is present too. In this case
we should only display the parent subpage. This requires traversing the
settings hierarchy to discover parent/child settings.

A follow-up CL will handle the case of setting/subpage duplicates, and
add tests after that.

Bug: 1097166
Change-Id: Iff8a0556d7ad153ff227570f1661e2e55c479d18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2262556
Commit-Queue: Tony Yeoman <tby@chromium.org>
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782317}