USE_X11 and USE_OZONE: //ui: choose between ozone and x11 path
This patch resolves mostly to the following pattern:

#if defined(USE_X11)
DoThis();
#elif defined(USE_OZONE)
OzonePlatform::DoThat();
#endif

Two ifdefs are combined to a single ifdef now with
if (IsUsingOzonePlatform()) condition is added.

-----

PS: Please note that this is a temp solution that will help to choose
between ozone and non-ozone X11 build. The switch that will be used
to choose the path is --enable-features=UseOzonePlatform. Once
non-Ozone X11 path is removed (hopefully by Q1 2021 depending on how
th finch trial goes), the wrapper will be removed.

Please also note that it's impossible to build use_x11 && use_ozone
without some hacks in PlatformCursor code. The changes to that are
on their way to upstream.

----

Bug: 1085700
Change-Id: I38d7302511ef6311515aa63a4839be0f7c9a5faf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2251620
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782318}