[iOS] Fix NTP animation when opening tab via keyboard
This CL changes the order in which the BVC/the omnibox is being focused
when opening a new tab from the TabGrid using keyboard shortcut.
Previously the omnibox was focused then the BVC, leading to an
animation glitch where the omnibox and the doodle disappeared.

This CL fixes it by focusing first the BVC then runs the completion of
the animation (which contains the focus of the omnibox), letting the
animation follow its natural course.

This should be relatively safe as the only completion block passed in
the TabGrid animation is the focusing of the omnibox.
So inverting the order of the calls should only have an effect in that
case.

The launchMask isn't used when BVC is contained.

(cherry picked from commit 6f260e6c99e7a453efcc38d126b7506691998b48)

Fixed: 1096517
Change-Id: I6a53e99e5c1a598401c668390c177e4bef2fc06a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264052
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#782309}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264436
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/branch-heads/4147@{#784}
Cr-Branched-From: 16307825352720ae04d898f37efa5449ad68b606-refs/heads/master@{#768962}