[css-pseudo] Fix DCHECK failure with ::marker and text-combine-upright
The 'text-combine-upright' property is not supported in LayoutNG, so it
forces the element to use legacy layout. However, when used on ::marker,
this could leave the list item using LayoutNG and only force the marker
to legacy. But markers depend on their originating list item, so this
would make some DCHECKs fail.

This patch makes DefinitelyNewFormattingContext always return false for
::marker, even if it's an outside marker with 'display: inline-block'.
This way ForceLegacyLayoutInFormattingContext will also force the list
item to use legacy layout when the marker requires it.

Bug: 457718

TEST=external/wpt/css/css-pseudo/marker-text-combine-upright.html

The test fails because 'text-combine-upright' doesn't seem to work well
on pseudo-elements, but now it doesn't crash.

Change-Id: I7f9f26914af39a9d30dbcff7fcd1ae1308981989
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264129
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/master@{#782375}