Fix installed but not loaded metrics for forced extensions
Currently, only the extensions that had status as LOADED were excluded
from the missing_forced_extensions list. This caused the extensions
with status READY to be included in the missing_forced_extensions
list and thus the metrics are reported wrong for the case when
extensions are installed but not LOADED. Now, we add the extensions
which have status as PENDING or FAILED to missing_forced_extensions
list to remove the false positives.

Bug: 1063024
Change-Id: I58594137a996c7f5a312507f3b0a3974fe51d61f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2260775
Commit-Queue: Swapnil Gupta <swapnilgupta@google.com>
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782378}