Clicking notification in action center does not trigger onclick handler
Previously, the WinRT dismissed event was used as the web
notifications close event in notification_platform_bridge_win.cc.
But the WinRT dismissed event fires before the notification is
closed. In particular, the WinRT toast dismissed event fires when
the notification enters the action center. As result we observe the
following behaviors:
  1- Clicking on the toast when it's in the action center does not
     work.
  2- For persistent notifications, the getNotifications() JavaScript
     does not return notifications in displayed in the action center
     they've already been deleted by the database.

This change update Chromium to stop using the WinRT toast dismissed
event as the web notifications close event in
NotificationPlatformBridgeWin.
Therefore, the notifications will remain stored in the browser when
it enters in the action center. But then we don't have any way to
tell when notifications are removed from the action center. Because
of this, we need to simulate the closed event using a timer.

NotificationPlatformBridgeWin will store locally in memory the
expected displayed notifications. The notifications synchronization
task will dispatch close event for the notification that is no longer
displayed.

The synchronization task run every 10 minutes.

Bug: 925892, 973776
Change-Id: Idd31815f5cd8619aca00bd655bedab9b736a5f83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159911
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Richard Knoll <knollr@chromium.org>
Commit-Queue: Daniel Soromou <fosoromo@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#782420}