[FragmentItem] Copy instead of move when reusing NGFragmentItem
When we changed the array of |scoped_refptr<NGFragmentItem>|
to the array of |NGFragmentItem| in r774510
<crrev.com/c/2208806>, we chose to move |NGFragmentItem|
because copying can involve memory allocations for
|NGInkOverflow|.

This patch changes it to copy. |NGFlexLayoutAlgorithm| can
run simplified layout of the same child multiple times when,
for instance, in |ApplyStretchAlignmentToChild|, and that
the same previous result may be reused multiple times.

There are some perf hits according to pinpoint:
https://pinpoint-dot-chromeperf.appspot.com/job/17cbb123120000
but is better than when we tried this before, because this
time we copy |NGInkOverflow| only for text items. This works
because for other items, pre-paint needs to re-compute ink
overflow anyway, because ink overflow may be affected by
changes in positioned objects.

Bug: 1099151
Change-Id: I649bffdcc0daac83f978b53a53efaef1f563613e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2266052
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782445}