[Instant Start] Fix the first tab fail to load with InstantStart.
LayoutManager#setTabModelSelector() will trigger the active Layout
shown. If the active layout is a StaticLayout, a LayoutTab is created in
the StaticLayout#setStaticTab(). However, when InstantStart is enabled,
LayoutManager#setTabModelSelector() is called too earlier and before a
tab is created, thus an invalid tab id is passed to
StaticLayout#setStaticTab() and fails to create the LayoutTab. To fix
it, we call the LayoutManager#setTabModelSelector() only if the
StartSurface overview is showing, but not when a regular tab is showing.

Bug: 1098794
Change-Id: I4b36ca92febb869fd084faa388864ff8e750a3bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264467
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Xi Han <hanxi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782448}