Speculative fix for ImpressionTagNavigatesExistingRemoteFrame_...
Moves the ImpressionObserver instantiation before the link click.
Declaring the observer after the click causes a race between
registering the observer and the navigation finishing.

ImpressionObserver::WaitForImpression() was also being used in a racy
way. If the impression was received prior to wait call, we would wait
for an additional impression, resulting in a timeout. This is fixed in
a few other tests in this suite as well, by tracking the number of
received impressions per waiter.

On https://crrev.com/c/2207175,
ImpressionTagNavigatesRemoteFrame_ImpressionReceived was disabled by
accident instead of this test and is also re-enabled.

The flake was not reproducible locally, hence the speculative tag for
this fix.

The whole test suite is currently disabled on Windows and will be
re-enabled on a separate change.

Bug: 1084201,1077216
Change-Id: I19a9d394831d4624cb8c62732136843317868f1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2209403
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782481}