WebUI Tab Strip: Disable pinned state change animations
When a tab changes pinned state, it can send two events: one for
moving to the very beginning of the tab strip if it already isn't,
and another for changing its pinned value. By sending the pinned state
for its move, the TabList is able to correctly move the pinned tab to
the pinned tabs container and prevent multiple animations and moves
for the same move.

The animations for this are disabled as these tend to be moves
across a large span of tabs and may be re-enabled upon better UX
treatment.

Bug: 1082344
Change-Id: I77bc6530838dece2ad407cf6b7ae5567515e4e04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2265244
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782483}