Handle ThroughputTracker outlive Compositor
It could happen when disconnecting an external display after
a tracker is started. This CL makes ThroughputTracker to use
a WeakPtr for its host and handles such situation in Stop/Cancel.
Start with a destructed host is a DCHECK failure and crash in
production.

Bug: 1021774
Change-Id: Ic181f1f3dba5da06dba63eda4277ace5d16fd768
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264932
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782513}