Fix high rate of CLASSIFIER_NOT_READY in CSD classification
We currently fail to distribute the model to the renderer in a large
number of cases, because we often get notified that a new
RenderProcessHost was created before any RenderFrameHosts are created.
This CL changes the model distribution so that it is distributed on
every RenderFrameHost creation, so we don't skip classifications.

Bug: 1068613, 1098064
Change-Id: Ie388ec472057fbcbd8f0df1f26df0f4ea01e76a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2261895
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Bettina Dea <bdea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782514}