Remove HwndMessageHandler DCHECKs
The DCHECKs that |exposed_pixels_| is smaller than the client area
doesn't hold when restoring a minimized content_shell window. On restore
the height of the window gets larger but the height of the client area
shrinks a bit. Since the height of the window was almost zero before
being restored exposed_pixels_.height() > cr.bottom and the DCHECK
fails.

I believe this is because content_shell lets Windows draw the title bar,
so it is non-client area, and the title bar wasn't drawn when the window
was minimized.

It doesn't seem problematic that area to be filled is potentially larger
than the client area, since FillRect() is already clipped by ps.rcPaint.
HwndMessageHandler also tells the delegate to repaint everything in
ps.rcPaint immediately the fill. As such, removing the DCHECKs seems
safe. Trying to account for window frame size changes would add more
complexity with no clear benefit.

Bug: 1084625
Change-Id: Ibb7966e5780d4d221d2b4e8fbb82eb5e37aa93b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2267501
Commit-Queue: kylechar <kylechar@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782534}