weblayer, wpt: Remove setting of actual results to layout test results
This CL is causing the wpt_expectations_updater.py script to have an
exception when reading results. The script does not expect the results
to be the layout test specific results. After integrating typ into
results and expectations handling in blinkpy the results need to be
set to FAIL for test failures. The results viewer already has code
which transforms the generic FAIL result to the layout test specific
result based on the artifacts dictionary contents.

Bug: 937369, 1050760, 1050754
Change-Id: I52c5daeacf58ba32339ecfbc5638ac63c1e841ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264021
Reviewed-by: Luke Z <lpz@chromium.org>
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Cr-Commit-Position: refs/heads/master@{#782556}