[conversions] Allow the embedder to disable
This CL adds a ContentBrowserClient API to allow the embedder to
disable impression / conversion registration wholesale.

The initial implementation of this in the //chrome layer is to
disallow when the third party cookie blocking setting is enabled,
though we expect this implementation to change in the future.

Bug: 1058018
Change-Id: I152a95a9f800f53d3be984ddd9f6586b3539b53c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155291
Commit-Queue: Charlie Harrison <csharrison@chromium.org>
Auto-Submit: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782585}