Use wheelTick in forms/resources/common-wheel-event.js
After we finish the scroll unification, the scrolls happens mainly on
the compositor thread, and the scroll code in the main thread will be
removed. eventSender sends the scroll events to main thread, so it
would not work after the scroll unification. This CL we should replace
eventSender.mouseScrollBy with wheelTick which uses
gpuBenchmarking.smoothScrollByXY in
forms/resources/common-wheel-event.js.

Bug: 1047176
Change-Id: I7416727dfcfbe2d88a283907ee132c838952747e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2171955
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782599}