Add ProduceOverlay/Dawn/Skia support to SharedImageBackingGLImage
On macOS, any SharedImageBacking that has a GLImage may be used as
an overlay, shared with Dawn, or shared with Metal-backed Skia.

There exists support for these things over in
SharedImageBackingFactoryIOSurface, along with a copy-paste of a lot of
the stuff found in SharedImageBackingGLImage.

To avoid doing more copy-pasting, add some small ifdef spaghetti that
calls the needed code in SharedImageBackingFactoryIOSurface.

After this patch, the only user of SharedImageBackingFactoryIOSurface
will be WebGPU. The next patch will remove that usage as well, and
turn SharedImageBackingFactoryIOSurface into a set of helper functions.

Bug: 1092155
Change-Id: Ia28ad13dd407d7de7bb3c5a5e16af95659238d6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264858
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782605}