Use pyfakefs in generate_buildbot_json_unittest.
Creating a fake so that read_file and write_file operate on a dictionary
means that many operations do not perform the same between the test and
production (e.g. raising a KeyError when attempting to read a
non-existent file instead of an IOError or finding the actual file on
the filesystem for a file that should not exist when using glob).

The fake filesystem provided by pyfakefs means that the test will only
see files that have actually been created by the test (modulo some
default files such as /tmp) and no mocking is necessary without paying
the cost of writing out to disk.

Change-Id: Ib3401b62cdac93289e9ba327d9a8f39d47bb45c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2267502
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782609}