[Text-Fragment] Fix ForceLoadAtTop anchor navigation
ForceLoadAtTop is a policy which pages can apply that blocks any kind of
scroll-on-load behavior including: fragments, text-fragments, history
scroll restoration.

However, we had a bug where by it blocked all fragment navigations,
including same page navigations (e.g. clicking a link that changes only
the fragment in the current URL). Since this isn't a new load, the
policy shouldn't apply in this case.

In addition to the small fix to this check, this CL also cleans up and
makes the scrolling to fragment condition more readable.

Bug: 1091247
Change-Id: Ic523b5b331720a822a5530698ab6560f580d2b97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2265758
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Nick Burris <nburris@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782667}