Create new AutofillOfferManager class.
This change will introduce the interface for a new AutofillOfferManager
class. The AutofillOfferManager will be responsible for retrieving offer
data via the payments client; it will do this through GetOfferData() and
the subsequent callback function OnDidGetOfferData(). This change will
be followed by another change to implement the interface and connect the
AutofillOfferManager to both the AutofillManager and the payments client.

Change-Id: I3d8824123330696f88c07d9729892a2c1d006a6e
Bug: 1093057
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248219
Commit-Queue: Ben Blake <fiorito@google.com>
Reviewed-by: Jared Saul <jsaul@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Siyu An <siyua@chromium.org>
Reviewed-by: Anne Lim <annelim@google.com>
Cr-Commit-Position: refs/heads/master@{#782772}