[Web Payment][Chrome OS] Check if in TWA.
Before this patch, requesting the unavailable payment method
https://play.google.com/billing method would return an error message
that mentioned that payment method is available only in a Trusted Web
Activity (TWA). This message was printed even inside of a TWA.

This patch adds a method to check whether Chrome is currently in TWA
mode and makes sure that the error message about requiring a TWA is
not printed in that case.

After this patch, requesting payment via https://play.google.com/billing
inside of TWA does not result in an error message about being outside of
TWA.

Bug: 1095808
Change-Id: I7d24532f871daf7ba3b2e690d52bbe5615393c3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2265413
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Liquan (Max) Gu <maxlg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782786}