VideoDecoderBroker: Connect WebCodecs to media::VideoDecoder(s)
"Broker" acts as a simple decoder, but under the hood manages creation
and selection of of underlying media::VideoDecoders, and thread-hopping
required to use them.

This enables WebCodecs to use hw accelerated decoders (depending on
platform/gpu) when not in a Worker context. For Workers, we still need
to expose the relevant Mojo interfaces (currently tied to the
RenderFrame). The code in webcodecs/ will then support hw-decode in
workers automatically.

Bug: 1094103

Change-Id: I9d1abcae842b16cbb095e2e24dc095c284075473
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2152106
Commit-Queue: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782793}