[XProto] Avoid memcpy()s for large requests/replies/events
The goal of this CL is to avoid memcpy's during {Get,Change}Property
and {Put,Get}Image.  The former can copy large data if an image is on
the clipboard, and the latter is used to draw to and read from the
screen.

Strategy for writes: instead of copying data to a linear buffer, break
up the data into chunks where each chunk is a base::RefCountedMemory.
If a RefCountedMemory needs to be written, just copy the refptr rather
than the data.

Strategy for reads: XCB gives us malloc'ed data for events and
replies, so we can wrap this in a MallocedRefCountedMemory.  If a
member is a RefCountedMemory when reading, create an
OffsetRefCountedMemory to wrap the MallocedRefCountedMemory.

BUG=1066670
R=sky
CC=msisov,nickdiego

Change-Id: I49f6e7d7ac9dde419e1afa6efe0081127c5cd31b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2265393
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782795}